Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  3.620460  3.620460  3.545138  3.545138  4.334574  4.334574  3.071149
#>   [8]  3.071149  4.335691  4.335691  2.220846  2.220846  3.799551  3.799551
#>  [15]  2.547941  2.547941  3.512355  3.512355  2.302417  2.302417  4.601114
#>  [22]  4.601114  3.703563  3.703563  3.217944  3.217944  4.347146  4.347146
#>  [29]  2.863746  2.863746  2.272419  2.272419  4.624312  4.624312  4.150934
#>  [36]  4.150934  3.444894  3.444894  4.396090  4.396090  3.769288  3.769288
#>  [43]  2.126469  2.126469  2.244059  2.244059  4.261198  4.261198  2.683885
#>  [50]  2.683885  4.705406  4.705406  3.326706  3.326706  3.212850  3.212850
#>  [57]  4.020639  4.020639  3.297076  3.297076  2.431304  2.431304  4.141991
#>  [64]  4.141991  2.786964  2.786964  5.858755  5.858755  4.608562  4.608562
#>  [71]  4.820753  4.820753  3.515174  3.515174  3.751056  3.751056  4.555651
#>  [78]  4.555651  3.325250  3.325250  3.601245  3.601245  5.993380  5.993380
#>  [85]  3.790881  3.790881  4.226591  4.226591  5.346611  5.346611  3.339710
#>  [92]  3.339710  7.709614  7.709614  4.346246  4.346246  5.679792  5.679792
#>  [99]  7.626197  7.626197  5.989119  5.989119  3.775974  3.775974  5.923973
#> [106]  5.923973  5.781643  5.781643  7.445446  7.445446  8.209210  8.209210
#> [113]  4.002014  4.002014  6.949869  6.949869  7.703841  7.703841  7.255943
#> [120]  7.255943  5.451285  5.451285  9.153329  9.153329  8.546051  8.546051
#> [127]  6.455901  6.455901  4.885849  4.885849  7.621275  7.621275  9.892884
#> [134]  9.892884  6.130043  6.130043  6.166666  6.166666  6.326812  6.326812
#> [141]  9.181991  9.181991  6.659265  6.659265  6.099167  6.099167 10.384341
#> [148] 10.384341  8.544587  8.544587 10.813507 10.813507  8.362423  8.362423
#> [155]  8.036241  8.036241  7.550035  7.550035  8.596939  8.596939  9.227368
#> [162]  9.227368  9.282448  9.282448  9.791161  9.791161  8.567417  8.567417
#> [169] 12.413693 12.413693 12.017416 12.017416  9.549510  9.549510 10.088558
#> [176] 10.088558 10.981515 10.981515 10.283630 10.283630 11.382927 11.382927
#> [183] 12.383840 12.383840 12.304344 12.304344 10.219727 10.219727  7.693798
#> [190]  7.693798 12.353684 12.353684  8.702196  8.702196  8.846648  8.846648
#> [197] 11.578243 11.578243 10.922247 10.922247 11.339912 11.339912 13.221862
#> [204] 13.221862 10.192110 10.192110 11.728311 11.728311 10.712973 10.712973
#> [211] 14.031571 14.031571  8.237337  8.237337 11.076493 11.076493 12.742950
#> [218] 12.742950 10.877053 10.877053 12.340875 12.340875 11.197482 11.197482
#> [225] 11.347134 11.347134 13.773263 13.773263 10.007393 10.007393 12.955053
#> [232] 12.955053  8.764170  8.764170 10.251119 10.251119 11.739450 11.739450
#> [239] 12.121269 12.121269 14.389759 14.389759 10.824614 10.824614 10.338792
#> [246] 10.338792 15.343509 15.343509 15.808889 15.808889 12.125557 12.125557
#> [253] 11.623812 11.623812  9.619067  9.619067 10.132465 10.132465 12.315351
#> [260] 12.315351 17.325619 17.325619 16.453681 16.453681 16.717260 16.717260
#> [267] 10.702347 10.702347 18.040922 18.040922 14.534982 14.534982 18.035424
#> [274] 18.035424 14.773591 14.773591 14.403798 14.403798 14.197690 14.197690
#> [281] 17.917089 17.917089 13.389746 13.389746 14.801091 14.801091 11.082022
#> [288] 11.082022 11.172977 11.172977 16.936801 16.936801 13.057873 13.057873
#> [295] 12.845954 12.845954 13.094685 13.094685 11.269405 11.269405 10.245218
#> [302] 10.245218 18.328384 18.328384 16.888676 16.888676 15.659705 15.659705
#> [309] 13.842181 13.842181 10.841037 10.841037 17.964976 17.964976 17.034333
#> [316] 17.034333 12.572320 12.572320 11.988811 11.988811 18.354520 18.354520
#> [323] 18.859953 18.859953 18.555232 18.555232 17.570141 17.570141 13.211056
#> [330] 13.211056 14.563268 14.563268 11.017313 11.017313 18.081153 18.081153
#> [337] 16.545229 16.545229 14.847115 14.847115 18.375558 18.375558 15.303889
#> [344] 15.303889 15.354856 15.354856 14.017696 14.017696 13.608018 13.608018
#> [351] 15.890228 15.890228 14.775848 14.775848 12.563931 12.563931 13.368610
#> [358] 13.368610 13.825930 13.825930 13.276596 13.276596 12.101189 12.101189
#> [365] 19.046988 19.046988 18.268688 18.268688 16.583635 16.583635 17.332590
#> [372] 17.332590 17.872369 17.872369 14.795679 14.795679 13.315279 13.315279
#> [379] 14.765271 14.765271 11.949050 11.949050 12.684663 12.684663 17.425414
#> [386] 17.425414 15.925028 15.925028 14.437570 14.437570 12.431760 12.431760
#> [393] 12.297353 12.297353 11.037526 11.037526 13.386293 13.386293  9.903085
#> [400]  9.903085 11.601964 11.601964 16.531356 16.531356  9.933941  9.933941
#> [407] 16.683424 16.683424 15.844943 15.844943  9.780873  9.780873 16.344261
#> [414] 16.344261 10.025723 10.025723 11.798666 11.798666 16.301282 16.301282
#> [421] 16.982395 16.982395  9.474777  9.474777 16.568049 16.568049 16.776531
#> [428] 16.776531  9.582976  9.582976  9.994210  9.994210 16.811952 16.811952
#> [435] 15.780456 15.780456 13.248501 13.248501 15.588096 15.588096 16.690829
#> [442] 16.690829 11.796375 11.796375  9.513813  9.513813  9.983021  9.983021
#> [449]  9.295168  9.295168 11.919089 11.919089 11.937430 11.937430 15.985157
#> [456] 15.985157 14.943855 14.943855 10.751980 10.751980 14.028121 14.028121
#> [463]  9.581689  9.581689  8.356926  8.356926  9.781802  9.781802 13.496809
#> [470] 13.496809  9.994468  9.994468 13.977624 13.977624 11.765696 11.765696
#> [477] 12.971115 12.971115  9.462969  9.462969 11.015826 11.015826  9.507375
#> [484]  9.507375 10.414668 10.414668 14.244289 14.244289 11.434907 11.434907
#> [491] 10.060194 10.060194 10.241493 10.241493 10.367107 10.367107 13.813789
#> [498] 13.813789 12.384543 12.384543 12.520086 12.520086 10.081315 10.081315
#> [505] 10.526625 10.526625 11.325461 11.325461  6.515961  6.515961 13.143567
#> [512] 13.143567  6.460901  6.460901 10.526968 10.526968 12.258361 12.258361
#> [519]  9.551775  9.551775  6.931078  6.931078  8.042869  8.042869  6.834879
#> [526]  6.834879 11.066497 11.066497  9.967537  9.967537  8.780235  8.780235
#> [533]  7.907958  7.907958  9.902367  9.902367  9.968563  9.968563  7.991286
#> [540]  7.991286  8.684145  8.684145  7.297285  7.297285  6.516728  6.516728
#> [547]  9.579698  9.579698  6.697124  6.697124  6.521686  6.521686  8.224714
#> [554]  8.224714  8.947936  8.947936  8.379427  8.379427  8.370165  8.370165
#> [561]  7.717355  7.717355  8.215280  8.215280  5.531139  5.531139  4.584110
#> [568]  4.584110  5.181602  5.181602  4.934186  4.934186  5.770371  5.770371
#> [575]  8.175556  8.175556  8.000609  8.000609  4.405743  4.405743  6.794047
#> [582]  6.794047  5.401071  5.401071  3.749679  3.749679  5.886974  5.886974
#> [589]  4.553666  4.553666  5.666434  5.666434  3.943889  3.943889  6.049192
#> [596]  6.049192  3.722931  3.722931  4.205127  4.205127  5.554375  5.554375
#> [603]  6.465466  6.465466  6.537709  6.537709  4.286013  4.286013  3.843896
#> [610]  3.843896  3.959019  3.959019  4.964317  4.964317  4.969529  4.969529
#> [617]  6.553917  6.553917  4.374680  4.374680  5.776857  5.776857  3.026060
#> [624]  3.026060  6.106928  6.106928  5.520524  5.520524  3.505219  3.505219
#> [631]  2.815206  2.815206  2.687717  2.687717  4.161327  4.161327  5.475677
#> [638]  5.475677  2.443150  2.443150  4.908809  4.908809  2.955404  2.955404
#> [645]  3.471159  3.471159  2.578003  2.578003  3.116722  3.116722  3.636289
#> [652]  3.636289  3.803344  3.803344  4.855391  4.855391  1.889503  1.889503
#> [659]  4.188360  4.188360  4.718618  4.718618  3.583248  3.583248  2.242591
#> [666]  2.242591  4.338670  4.338670  3.620946  3.620946  1.983066  1.983066
#> [673]  1.995733  1.995733  4.348988  4.348988  4.164227  4.164227  1.827188
#> [680]  1.827188  1.924729  1.924729  2.638154  2.638154  1.970007  1.970007
#> [687]  4.232048  4.232048  3.662251  3.662251  2.537643  2.537643  3.898163
#> [694]  3.898163  2.090779  2.090779  3.843095  3.843095  1.649609  1.649609
#> [701]  2.129299  2.129299  3.150557  3.150557  2.095387  2.095387  3.340877
#> [708]  3.340877  2.101405  2.101405  3.161395  3.161395  4.146553  4.146553
#> [715]  2.456277  2.456277  3.698530  3.698530  3.682438  3.682438  4.039458
#> [722]  4.039458  1.980057  1.980057  1.502355  1.502355  2.075649  2.075649
#> [729]  2.654261  2.654261