Calculate the dates of budburst and beginning of leaf fall

Wrapper for vegperiod

calc_vegperiod(
  budburst_method,
  leaffall_method,
  dates = NULL,
  tavg = NULL,
  out_yrs = NULL,
  budburstdoy.fixed = 121,
  leaffalldoy.fixed = 279,
  ...
)

Arguments

budburst_method

name of model for estimating budburst day of year. Either 'fixed' or one of the values accepted by the 'start.method'-argument of the function vegperiod.

leaffall_method

name of model for estimating day of year when leaffall begin. Either 'fixed' or one of the values accepted by the 'end.method'-argument of the function vegperiod.

dates

date vector passed to vegperiod, ignored if both leaffall_method and budburst_method = 'fixed'

tavg

vector of daily mean air temperature (deg C) passed to vegperiod, ignored if leaffall_method = 'fixed' and budburst_method = 'fixed'.

out_yrs

integer vector of the years to be returned. If not specified, values for the years in dates will be returned.

budburstdoy.fixed

vector of values to be returned if budburst_method = 'fixed'.

leaffalldoy.fixed

vector of values to be returned if leaffall_method = 'fixed'.

...

additional argument passed to vegperiod.

Value

a data.frame with columns year, start, end. If budburst_method = 'fixed' or leaffall_method = 'fixed', start and end contain the values specified in budburstdoy.fixed and leaffalldoy.fixed respectively.

Examples

# fixed budburst and leaffall doy
calc_vegperiod(out_yrs = 2001:2010,
               budburst_method = "fixed",
               leaffall_method = "fixed",
               budburstdoy.fixed = floor(runif(10, 120,130)),
               leaffalldoy.fixed = floor(runif(2, 260,280)))
#>    year start end
#> 1  2001   128 271
#> 2  2002   120 275
#> 3  2003   120 271
#> 4  2004   127 275
#> 5  2005   122 271
#> 6  2006   120 275
#> 7  2007   128 271
#> 8  2008   120 275
#> 9  2009   124 271
#> 10 2010   129 275

# dynamic budburst and leaffall using air temperature
data(slb1_meteo)

calc_vegperiod(budburst_method = "Menzel",
               leaffall_method = "fixed",
               leaffalldoy.fixed = 280,
               dates = slb1_meteo$dates,
               tavg = slb1_meteo$tmean,
               species = "Fagus sylvatica",
               est.prev = 3)
#>    year start end
#> 1  1960   136 280
#> 2  1961   128 280
#> 3  1962   141 280
#> 4  1963   137 280
#> 5  1964   135 280
#> 6  1965   145 280
#> 7  1966   132 280
#> 8  1967   136 280
#> 9  1968   124 280
#> 10 1969   133 280
#> 11 1970   142 280
#> 12 1971   134 280
#> 13 1972   141 280
#> 14 1973   143 280
#> 15 1974   136 280
#> 16 1975   139 280
#> 17 1976   135 280
#> 18 1977   140 280
#> 19 1978   141 280
#> 20 1979   140 280
#> 21 1980   144 280
#> 22 1981   131 280
#> 23 1982   141 280
#> 24 1983   137 280
#> 25 1984   141 280
#> 26 1985   135 280
#> 27 1986   136 280
#> 28 1987   140 280
#> 29 1988   134 280
#> 30 1989   134 280
#> 31 1990   126 280
#> 32 1991   142 280
#> 33 1992   137 280
#> 34 1993   120 280
#> 35 1994   132 280
#> 36 1995   137 280
#> 37 1996   137 280
#> 38 1997   135 280
#> 39 1998   129 280
#> 40 1999   133 280
#> 41 2000   126 280
#> 42 2001   134 280
#> 43 2002   134 280
#> 44 2003   127 280
#> 45 2004   132 280
#> 46 2005   135 280
#> 47 2006   131 280
#> 48 2007   122 280
#> 49 2008   134 280
#> 50 2009   125 280
#> 51 2010   143 280
#> 52 2011   127 280
#> 53 2012   132 280
#> 54 2013   130 280

calc_vegperiod(budburst_method = "Menzel",
               leaffall_method = "ETCCDI",
               dates = slb1_meteo$dates,
               tavg = slb1_meteo$tmean,
               species = "Quercus robur",
               est.prev = 3)
#>    year start end
#> 1  1960   137 338
#> 2  1961   125 312
#> 3  1962   146 305
#> 4  1963   138 303
#> 5  1964   138 297
#> 6  1965   145 295
#> 7  1966   131 305
#> 8  1967   137 313
#> 9  1968   129 314
#> 10 1969   135 314
#> 11 1970   148 329
#> 12 1971   136 316
#> 13 1972   143 296
#> 14 1973   147 295
#> 15 1974   136 289
#> 16 1975   141 287
#> 17 1976   139 322
#> 18 1977   141 322
#> 19 1978   139 315
#> 20 1979   145 300
#> 21 1980   148 309
#> 22 1981   131 289
#> 23 1982   143 335
#> 24 1983   138 320
#> 25 1984   146 322
#> 26 1985   138 299
#> 27 1986   138 336
#> 28 1987   143 313
#> 29 1988   136 308
#> 30 1989   134 331
#> 31 1990   124 311
#> 32 1991   140 296
#> 33 1992   136 291
#> 34 1993   122 293
#> 35 1994   132 353
#> 36 1995   138 311
#> 37 1996   140 323
#> 38 1997   135 299
#> 39 1998   127 308
#> 40 1999   132 294
#> 41 2000   127 325
#> 42 2001   135 314
#> 43 2002   133 312
#> 44 2003   128 291
#> 45 2004   130 316
#> 46 2005   134 323
#> 47 2006   133 354
#> 48 2007   122 297
#> 49 2008   136 327
#> 50 2009   127 291
#> 51 2010   143 294
#> 52 2011   127 319
#> 53 2012   132 305
#> 54 2013   135 319