Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2        var3        var4        var5
#>  1: 2002  1 -0.42733377 -0.68985386 -0.77378239 -0.92113446 -0.22041033
#>  2: 2002  2 -0.46711604 -0.99261935 -0.70253610 -0.38240983 -0.93973193
#>  3: 2002  3 -0.51914693 -0.26762467 -0.87301239 -0.11391017 -0.21203568
#>  4: 2002  4 -0.74959275 -0.61345675 -0.82420432 -0.79688664 -0.64160596
#>  5: 2002  5 -0.74613238 -0.26980908 -0.60398027 -0.40174897 -0.78822439
#>  6: 2002  6 -0.85680485 -0.15542003 -0.15990173 -0.28665214 -0.15892891
#>  7: 2002  7 -0.80733254 -0.67543729 -0.56243542 -0.66914815 -0.34706452
#>  8: 2002  8 -0.80855819 -0.12875116 -0.17404086 -0.49420833 -0.90539546
#>  9: 2002  9 -0.44718745 -0.07187861 -0.43122500 -0.32478227 -0.12542321
#> 10: 2002 10 -0.72057901 -0.35596874 -0.85453405 -0.51294518 -0.57048403
#> 11: 2002 11 -0.11576391 -0.62964936 -0.17742680 -0.98033947 -0.11971363
#> 12: 2002 12 -0.62332337 -0.23614605 -0.32304472 -0.13891607 -0.04606892
#> 13: 2003  1 -0.90877553 -0.86568589 -0.67111887 -0.17182442 -0.82688152
#> 14: 2003  2 -0.67236357 -0.19757968 -0.73992973 -0.47940334 -0.59431611
#> 15: 2003  3 -0.38986102 -0.99538940 -0.58204092 -0.64095546 -0.29732518
#> 16: 2003  4 -0.83466333 -0.67854713 -0.80727119 -0.50171503 -0.41488352
#> 17: 2003  5 -0.86186952 -0.87049767 -0.94147037 -0.62418752 -0.61167291
#> 18: 2003  6 -0.03146045 -0.17912003 -0.05630437 -0.98333821 -0.98948675
#> 19: 2003  7 -0.25838896 -0.78696784 -0.21301366 -0.19701496 -0.09876466
#> 20: 2003  8 -0.21519677 -0.09981102 -0.61689381 -0.72869816 -0.36340902
#> 21: 2003  9 -0.17662476 -0.22804749 -0.10847552 -0.41426520 -0.48395332
#> 22: 2003 10 -0.40708358 -0.23169091 -0.95028616 -0.34662939 -0.25852708
#> 23: 2003 11 -0.98074465 -0.33101772 -0.53373374 -0.44147011 -0.56026005
#> 24: 2003 12 -0.26895352 -0.33777266 -0.95987212 -0.08084575 -0.06801245
#>       yr mo        var1        var2        var3        var4        var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2       var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.68985386 -0.77378239 -0.92113446 1.743916 1.952090 1.149707
#>  2: 2002  2 -0.99261935 -0.70253610 -0.38240983 1.998485 1.830379 1.305649
#>  3: 2002  3 -0.26762467 -0.87301239 -0.11391017 1.892653 1.625725 1.986225
#>  4: 2002  4 -0.61345675 -0.82420432 -0.79688664 1.851144 1.333153 1.440539
#>  5: 2002  5 -0.26980908 -0.60398027 -0.40174897 1.597246 1.290402 1.943578
#>  6: 2002  6 -0.15542003 -0.15990173 -0.28665214 1.357672 1.980697 1.363167
#>  7: 2002  7 -0.67543729 -0.56243542 -0.66914815 1.857173 1.315628 1.902011
#>  8: 2002  8 -0.12875116 -0.17404086 -0.49420833 1.424069 1.604296 1.154441
#>  9: 2002  9 -0.07187861 -0.43122500 -0.32478227 1.168742 1.666442 1.489111
#> 10: 2002 10 -0.35596874 -0.85453405 -0.51294518 1.876219 1.086684 1.510149
#> 11: 2002 11 -0.62964936 -0.17742680 -0.98033947 1.574843 1.246824 1.401569
#> 12: 2002 12 -0.23614605 -0.32304472 -0.13891607 1.942615 1.639702 1.019027
#> 13: 2003  1 -0.86568589 -0.67111887 -0.17182442 1.173645 1.354115 1.912561
#> 14: 2003  2 -0.19757968 -0.73992973 -0.47940334 1.679159 1.279318 1.584792
#> 15: 2003  3 -0.99538940 -0.58204092 -0.64095546 1.819469 1.219170 1.761858
#> 16: 2003  4 -0.67854713 -0.80727119 -0.50171503 1.276576 1.700921 1.271095
#> 17: 2003  5 -0.87049767 -0.94147037 -0.62418752 1.546221 1.751472 1.477793
#> 18: 2003  6 -0.17912003 -0.05630437 -0.98333821 1.392878 1.454651 1.480817
#> 19: 2003  7 -0.78696784 -0.21301366 -0.19701496 1.957088 1.013532 1.165414
#> 20: 2003  8 -0.09981102 -0.61689381 -0.72869816 1.625525 1.196211 1.399932
#> 21: 2003  9 -0.22804749 -0.10847552 -0.41426520 1.420898 1.428105 1.936443
#> 22: 2003 10 -0.23169091 -0.95028616 -0.34662939 1.381646 1.929364 1.986578
#> 23: 2003 11 -0.33101772 -0.53373374 -0.44147011 1.771963 1.501785 1.647925
#> 24: 2003 12 -0.33777266 -0.95987212 -0.08084575 1.133769 1.261125 1.256867
#>       yr mo       var_2       var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.957832 2.208930 2.839440
#>  2: 2.878850 2.667885 2.512915
#>  3: 2.035025 2.402653 2.423549
#>  4: 2.906558 2.177081 2.859898
#>  5: 2.672744 2.354148 2.297847
#>  6: 2.655651 2.816389 2.817555
#>  7: 2.655847 2.080713 2.719197
#>  8: 2.236939 2.543305 2.581686
#>  9: 2.582778 2.587375 2.842282
#> 10: 2.254998 2.378625 2.188859
#> 11: 2.774527 2.968821 2.123404
#> 12: 2.325454 2.043106 2.913653
#> 13: 2.615318 2.895137 2.474962
#> 14: 2.270811 2.968620 2.768308
#> 15: 2.849244 2.242314 2.578766
#> 16: 2.444304 2.246384 2.406636
#> 17: 2.770959 2.931610 2.904837
#> 18: 2.392430 2.883461 2.428010
#> 19: 2.695172 2.681170 2.909716
#> 20: 2.235547 2.215542 2.998971
#> 21: 2.976219 2.513915 2.029538
#> 22: 2.562582 2.101512 2.875356
#> 23: 2.392357 2.515645 2.430816
#> 24: 2.343204 2.643460 2.144502
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.743916 1.952090 1.149707 2.957832 2.208930 2.839440
#>  2: 2002  2 1.998485 1.830379 1.305649 2.878850 2.667885 2.512915
#>  3: 2002  3 1.892653 1.625725 1.986225 2.035025 2.402653 2.423549
#>  4: 2002  4 1.851144 1.333153 1.440539 2.906558 2.177081 2.859898
#>  5: 2002  5 1.597246 1.290402 1.943578 2.672744 2.354148 2.297847
#>  6: 2002  6 1.357672 1.980697 1.363167 2.655651 2.816389 2.817555
#>  7: 2002  7 1.857173 1.315628 1.902011 2.655847 2.080713 2.719197
#>  8: 2002  8 1.424069 1.604296 1.154441 2.236939 2.543305 2.581686
#>  9: 2002  9 1.168742 1.666442 1.489111 2.582778 2.587375 2.842282
#> 10: 2002 10 1.876219 1.086684 1.510149 2.254998 2.378625 2.188859
#> 11: 2002 11 1.574843 1.246824 1.401569 2.774527 2.968821 2.123404
#> 12: 2002 12 1.942615 1.639702 1.019027 2.325454 2.043106 2.913653
#> 13: 2003  1 1.173645 1.354115 1.912561 2.615318 2.895137 2.474962
#> 14: 2003  2 1.679159 1.279318 1.584792 2.270811 2.968620 2.768308
#> 15: 2003  3 1.819469 1.219170 1.761858 2.849244 2.242314 2.578766
#> 16: 2003  4 1.276576 1.700921 1.271095 2.444304 2.246384 2.406636
#> 17: 2003  5 1.546221 1.751472 1.477793 2.770959 2.931610 2.904837
#> 18: 2003  6 1.392878 1.454651 1.480817 2.392430 2.883461 2.428010
#> 19: 2003  7 1.957088 1.013532 1.165414 2.695172 2.681170 2.909716
#> 20: 2003  8 1.625525 1.196211 1.399932 2.235547 2.215542 2.998971
#> 21: 2003  9 1.420898 1.428105 1.936443 2.976219 2.513915 2.029538
#> 22: 2003 10 1.381646 1.929364 1.986578 2.562582 2.101512 2.875356
#> 23: 2003 11 1.771963 1.501785 1.647925 2.392357 2.515645 2.430816
#> 24: 2003 12 1.133769 1.261125 1.256867 2.343204 2.643460 2.144502
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4