Aggregate and group model outputs similar to ancient LWFB90 textfile outputs (.ASC-files)

Returns selected groups of variables in the chosen temporal aggregation

process_outputs_LWFB90(x, selection = set_outputLWFB90(), prec_interval = NULL)

Arguments

x

Named list with items x$output and/or x$layer_output (e.g. as returned by run_LWFB90)

selection

A [7,5]-matrix with row and column names, flagging the desired groups of variables at specified time intervals (see set_outputLWFB90).

prec_interval

The precipitation interval of the simulation that produced x. If available, the value x$model_input$options_b90$prec_interval is used.

Value

A named list containing the selected groups of variables in the desired temporal resolution. The names are constructed from selection's row names and column names, suffixed by '.ASC' as a reminiscence to the former text file output of LWF-Brook90.

Examples

data("slb1_soil")
data("slb1_meteo")
opts <- set_optionsLWFB90(startdate = as.Date("2002-06-01"), enddate = as.Date("2002-06-05"))
parms <- set_paramLWFB90()
soil <- cbind(slb1_soil, hydpar_wessolek_tab(texture = slb1_soil$texture))

outsel <- set_outputLWFB90()
outsel[,] <- 1L

res <- run_LWFB90(options_b90 = opts,
           param_b90 = parms,
           climate = slb1_meteo,
           soil = soil)

# Calculate output-aggregations using the returned object
process_outputs_LWFB90(res, selection = outsel)
#> $BUDGPRE.ASC
#>       yr    mo    da   doy  prec  flow  evap  seep  snow  swat  gwat  intr
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.2   2.9     0     0 418.0     0     0
#> 2:  2002     6     2   153   0.0   0.2   3.7     0     0 414.1     0     0
#> 3:  2002     6     3   154   0.0   0.3   3.6     0     0 410.2     0     0
#> 4:  2002     6     4   155   2.9   0.4   4.0     0     0 408.6     0     0
#> 5:  2002     6     5   156   2.4   0.6   4.1     0     0 406.3     0     0
#>     ints
#>    <num>
#> 1:     0
#> 2:     0
#> 3:     0
#> 4:     0
#> 5:     0
#> 
#> $BUDGDAY.ASC
#>       yr    mo    da   doy  prec  flow  evap  seep  snow  swat  gwat  intr
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.2   2.9     0     0 418.0     0     0
#> 2:  2002     6     2   153   0.0   0.2   3.7     0     0 414.1     0     0
#> 3:  2002     6     3   154   0.0   0.3   3.6     0     0 410.2     0     0
#> 4:  2002     6     4   155   2.9   0.4   4.0     0     0 408.6     0     0
#> 5:  2002     6     5   156   2.4   0.6   4.1     0     0 406.3     0     0
#>     ints
#>    <num>
#> 1:     0
#> 2:     0
#> 3:     0
#> 4:     0
#> 5:     0
#> 
#> $BUDGMON.ASC
#>       yr    mo  prec  flow  evap  seep  snow  swat  gwat  intr  ints
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   5.3   1.8  18.3     0     0 406.3     0     0     0
#> 
#> $BUDGANN.ASC
#>       yr  prec  flow  evap  seep  snow  swat  gwat  intr  ints
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   5.3   1.8  18.3     0     0 406.3     0     0     0
#> 
#> $FLOWPRE.ASC
#>       yr    mo    da   doy  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2     0     0   0.0     0     0   0.2   0.2
#> 2:  2002     6     2   153   0.2     0     0   0.0     0     0   0.2   0.2
#> 3:  2002     6     3   154   0.3     0     0   0.0     0     0   0.3   0.3
#> 4:  2002     6     4   155   0.4     0     0   1.8     0     0   0.4   0.4
#> 5:  2002     6     5   156   0.6     0     0   1.6     0     0   0.6   0.6
#> 
#> $FLOWDAY.ASC
#>       yr    mo    da   doy  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2     0     0   0.0     0     0   0.2   0.2
#> 2:  2002     6     2   153   0.2     0     0   0.0     0     0   0.2   0.2
#> 3:  2002     6     3   154   0.3     0     0   0.0     0     0   0.3   0.3
#> 4:  2002     6     4   155   0.4     0     0   1.8     0     0   0.4   0.4
#> 5:  2002     6     5   156   0.6     0     0   1.6     0     0   0.6   0.6
#> 
#> $FLOWMON.ASC
#>       yr    mo  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   1.8     0     0   3.4     0     0   1.8   1.8
#> 
#> $FLOWANN.ASC
#>       yr  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   1.8     0     0   3.4     0     0   1.8   1.8
#> 
#> $EVAPPRE.ASC
#>       yr    mo    da   doy  flow  evap  tran  irvp  isvp  slvp  snvp  pint
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2   2.9   2.6   0.0     0   0.3     0   5.1
#> 2:  2002     6     2   153   0.2   3.7   3.3   0.0     0   0.4     0   7.6
#> 3:  2002     6     3   154   0.3   3.6   3.1   0.0     0   0.5     0   8.8
#> 4:  2002     6     4   155   0.4   4.0   2.6   1.0     0   0.3     0   6.1
#> 5:  2002     6     5   156   0.6   4.1   2.6   0.9     0   0.6     0  10.4
#>    ptran pslvp
#>    <num> <num>
#> 1:   2.6   0.3
#> 2:   3.3   0.5
#> 3:   3.1   0.6
#> 4:   2.6   0.4
#> 5:   2.6   0.8
#> 
#> $EVAPDAY.ASC
#>       yr    mo    da   doy  flow  evap  tran  irvp  isvp  slvp  snvp  pint
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2   2.9   2.6   0.0     0   0.3     0   5.1
#> 2:  2002     6     2   153   0.2   3.7   3.3   0.0     0   0.4     0   7.6
#> 3:  2002     6     3   154   0.3   3.6   3.1   0.0     0   0.5     0   8.8
#> 4:  2002     6     4   155   0.4   4.0   2.6   1.0     0   0.3     0   6.1
#> 5:  2002     6     5   156   0.6   4.1   2.6   0.9     0   0.6     0  10.4
#>    ptran pslvp
#>    <num> <num>
#> 1:   2.6   0.3
#> 2:   3.3   0.5
#> 3:   3.1   0.6
#> 4:   2.6   0.4
#> 5:   2.6   0.8
#> 
#> $EVAPMON.ASC
#>       yr    mo  flow  evap  tran  irvp  isvp  slvp  snvp  pint ptran pslvp
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   1.8  18.3  14.2   1.9     0   2.2     0    38  14.2   2.7
#> 
#> $EVAPANN.ASC
#>       yr  flow  evap  tran  irvp  isvp  slvp  snvp  pint ptran pslvp
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   1.8  18.3  14.2   1.9     0   2.2     0    38  14.2   2.7
#> 
#> $ABOVPRE.ASC
#>       yr    mo    da   doy  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.0     0     0   0.0     0     0   0.0
#> 2:  2002     6     2   153   0.0   0.0     0     0   0.0     0     0   0.0
#> 3:  2002     6     3   154   0.0   0.0     0     0   0.0     0     0   0.0
#> 4:  2002     6     4   155   2.9   1.0     0     0   1.8     0     0   1.8
#> 5:  2002     6     5   156   2.4   0.9     0     0   1.6     0     0   1.6
#>     smlt  slfl  srfl
#>    <num> <num> <num>
#> 1:     0   0.0     0
#> 2:     0   0.0     0
#> 3:     0   0.0     0
#> 4:     0   1.8     0
#> 5:     0   1.6     0
#> 
#> $ABOVDAY.ASC
#>       yr    mo    da   doy  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.0     0     0   0.0     0     0   0.0
#> 2:  2002     6     2   153   0.0   0.0     0     0   0.0     0     0   0.0
#> 3:  2002     6     3   154   0.0   0.0     0     0   0.0     0     0   0.0
#> 4:  2002     6     4   155   2.9   1.0     0     0   1.8     0     0   1.8
#> 5:  2002     6     5   156   2.4   0.9     0     0   1.6     0     0   1.6
#>     smlt  slfl  srfl
#>    <num> <num> <num>
#> 1:     0   0.0     0
#> 2:     0   0.0     0
#> 3:     0   0.0     0
#> 4:     0   1.8     0
#> 5:     0   1.6     0
#> 
#> $ABOVMON.ASC
#>       yr    mo  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet  smlt  slfl
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   5.3   1.9     0     0   3.4     0     0   3.4     0   3.4
#>     srfl
#>    <num>
#> 1:     0
#> 
#> $ABOVANN.ASC
#>       yr  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet  smlt  slfl  srfl
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   5.3   1.9     0     0   3.4     0     0   3.4     0   3.4     0
#> 
#> $BELOPRE.ASC
#>         yr    mo    da   doy    nl  infl  byfl  tran  vrfl  dsfl  ntfl
#>      <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#>   1:  2002     6     1   152     1   0.0     0   0.1  -0.2     0  -0.2
#>   2:  2002     6     2   153     1   0.0     0   0.1  -0.4     0  -0.1
#>   3:  2002     6     3   154     1   0.0     0   0.1  -0.5     0  -0.1
#>   4:  2002     6     4   155     1   1.8     0   0.1   1.3     0   0.1
#>   5:  2002     6     5   156     1   1.6     0   0.1   0.9     0   0.0
#>  ---                                                                  
#> 101:  2002     6     1   152    21   0.0     0   0.0   0.2     0   0.0
#> 102:  2002     6     2   153    21   0.0     0   0.0   0.2     0   0.1
#> 103:  2002     6     3   154    21   0.0     0   0.0   0.3     0   0.2
#> 104:  2002     6     4   155    21   0.0     0   0.0   0.4     0   0.2
#> 105:  2002     6     5   156    21   0.0     0   0.0   0.6     0   0.1
#> 
#> $BELODAY.ASC
#>         yr    mo    da   doy    nl  infl  byfl  tran  vrfl  dsfl  ntfl
#>      <num> <num> <num> <num> <int> <num> <num> <num> <num> <num> <num>
#>   1:  2002     6     1   152     1   0.0     0   0.1  -0.2     0  -0.2
#>   2:  2002     6     2   153     1   0.0     0   0.1  -0.4     0  -0.1
#>   3:  2002     6     3   154     1   0.0     0   0.1  -0.5     0  -0.1
#>   4:  2002     6     4   155     1   1.8     0   0.1   1.3     0   0.1
#>   5:  2002     6     5   156     1   1.6     0   0.1   0.9     0   0.0
#>  ---                                                                  
#> 101:  2002     6     1   152    21   0.0     0   0.0   0.2     0   0.0
#> 102:  2002     6     2   153    21   0.0     0   0.0   0.2     0   0.1
#> 103:  2002     6     3   154    21   0.0     0   0.0   0.3     0   0.2
#> 104:  2002     6     4   155    21   0.0     0   0.0   0.4     0   0.2
#> 105:  2002     6     5   156    21   0.0     0   0.0   0.6     0   0.1
#> 
#> $BELOMON.ASC
#>        yr    mo    nl    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#>     <num> <num> <int> <num> <num> <num> <num> <num> <num> <num> <num>
#>  1:  2002     6     1    15   770   3.4     0   0.4   1.1     0  -0.4
#>  2:  2002     6     2    15   770   0.0     0   0.9   0.9     0  -0.7
#>  3:  2002     6     3    15   770   0.0     0   0.8   0.8     0  -0.7
#>  4:  2002     6     4    15   770   0.0     0   1.1   0.8     0  -1.1
#>  5:  2002     6     5    15   770   0.0     0   1.3   0.8     0  -1.4
#>  6:  2002     6     6    15   770   0.0     0   1.2   1.0     0  -1.4
#>  7:  2002     6     7    15   770   0.0     0   1.5   1.4     0  -2.0
#>  8:  2002     6     8    15   770   0.0     0   1.3   2.0     0  -1.9
#>  9:  2002     6     9    15   770   0.0     0   1.4   3.0     0  -2.3
#> 10:  2002     6    10    15   770   0.0     0   1.1   3.9     0  -2.1
#> 11:  2002     6    11    15   770   0.0     0   0.8   4.9     0  -1.8
#> 12:  2002     6    12    15   770   0.0     0   0.7   5.8     0  -1.5
#> 13:  2002     6    13    15   770   0.0     0   0.5   6.7     0  -1.4
#> 14:  2002     6    14    15   770   0.0     0   0.5   7.6     0  -1.4
#> 15:  2002     6    15    15   770   0.0     0   0.3   7.7     0  -0.4
#> 16:  2002     6    16    15   770   0.0     0   0.2   7.0     0   0.5
#> 17:  2002     6    17    15   770   0.0     0   0.1   4.9     0   2.0
#> 18:  2002     6    18    15   770   0.0     0   0.0   3.9     0   1.0
#> 19:  2002     6    19    15   770   0.0     0   0.0   3.1     0   0.7
#> 20:  2002     6    20    15   770   0.0     0   0.0   2.4     0   0.7
#> 21:  2002     6    21    15   770   0.0     0   0.0   1.8     0   0.7
#>        yr    mo    nl    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#> 
#> $BELOANN.ASC
#>        yr    nl    mo    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#>     <num> <int> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#>  1:  2002     1    30    15   770   3.4     0   0.4   1.1     0  -0.4
#>  2:  2002     2    30    15   770   0.0     0   0.9   0.9     0  -0.7
#>  3:  2002     3    30    15   770   0.0     0   0.8   0.8     0  -0.7
#>  4:  2002     4    30    15   770   0.0     0   1.1   0.8     0  -1.1
#>  5:  2002     5    30    15   770   0.0     0   1.3   0.8     0  -1.4
#>  6:  2002     6    30    15   770   0.0     0   1.2   1.0     0  -1.4
#>  7:  2002     7    30    15   770   0.0     0   1.5   1.4     0  -2.0
#>  8:  2002     8    30    15   770   0.0     0   1.3   2.0     0  -1.9
#>  9:  2002     9    30    15   770   0.0     0   1.4   3.0     0  -2.3
#> 10:  2002    10    30    15   770   0.0     0   1.1   3.9     0  -2.1
#> 11:  2002    11    30    15   770   0.0     0   0.8   4.9     0  -1.8
#> 12:  2002    12    30    15   770   0.0     0   0.7   5.8     0  -1.5
#> 13:  2002    13    30    15   770   0.0     0   0.5   6.7     0  -1.4
#> 14:  2002    14    30    15   770   0.0     0   0.5   7.6     0  -1.4
#> 15:  2002    15    30    15   770   0.0     0   0.3   7.7     0  -0.4
#> 16:  2002    16    30    15   770   0.0     0   0.2   7.0     0   0.5
#> 17:  2002    17    30    15   770   0.0     0   0.1   4.9     0   2.0
#> 18:  2002    18    30    15   770   0.0     0   0.0   3.9     0   1.0
#> 19:  2002    19    30    15   770   0.0     0   0.0   3.1     0   0.7
#> 20:  2002    20    30    15   770   0.0     0   0.0   2.4     0   0.7
#> 21:  2002    21    30    15   770   0.0     0   0.0   1.8     0   0.7
#>        yr    nl    mo    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#> 
#> $SWATPRE.ASC
#>         yr    mo    da   doy    nl swati theta wetnes   psimi
#>      <num> <num> <num> <num> <num> <num> <num>  <num>   <num>
#>   1:  2002     6     1   152     1 3.219 0.335  0.830 -10.282
#>   2:  2002     6     2   153     1 3.108 0.324  0.801 -13.167
#>   3:  2002     6     3   154     1 3.024 0.315  0.779 -15.758
#>   4:  2002     6     4   155     1 3.096 0.323  0.797 -13.512
#>   5:  2002     6     5   156     1 3.063 0.319  0.789 -14.502
#>  ---                                                         
#> 101:  2002     6     1   152    21 6.316 0.316  0.723  -6.133
#> 102:  2002     6     2   153    21 6.425 0.321  0.739  -5.395
#> 103:  2002     6     3   154    21 6.614 0.331  0.767  -4.310
#> 104:  2002     6     4   155    21 6.818 0.341  0.797  -3.362
#> 105:  2002     6     5   156    21 6.957 0.348  0.818  -2.823
#> 
#> $SWATDAY.ASC
#>         yr    mo    da   doy    nl swati theta wetnes   psimi
#>      <num> <num> <num> <num> <int> <num> <num>  <num>   <num>
#>   1:  2002     6     1   152     1 3.219 0.335  0.830 -10.282
#>   2:  2002     6     2   153     1 3.108 0.324  0.801 -13.167
#>   3:  2002     6     3   154     1 3.024 0.315  0.779 -15.758
#>   4:  2002     6     4   155     1 3.096 0.323  0.797 -13.512
#>   5:  2002     6     5   156     1 3.063 0.319  0.789 -14.502
#>  ---                                                         
#> 101:  2002     6     1   152    21 6.316 0.316  0.723  -6.133
#> 102:  2002     6     2   153    21 6.425 0.321  0.739  -5.395
#> 103:  2002     6     3   154    21 6.614 0.331  0.767  -4.310
#> 104:  2002     6     4   155    21 6.818 0.341  0.797  -3.362
#> 105:  2002     6     5   156    21 6.957 0.348  0.818  -2.823
#> 
#> $SWATMON.ASC
#>        yr    mo    nl  swati theta wetnes   psimi
#>     <num> <num> <int>  <num> <num>  <num>   <num>
#>  1:  2002     6     1  3.102 0.323  0.799 -13.444
#>  2:  2002     6     2  6.211 0.323  0.800 -13.347
#>  3:  2002     6     3  6.223 0.324  0.801 -13.178
#>  4:  2002     6     4  9.357 0.325  0.803 -12.968
#>  5:  2002     6     5 12.523 0.326  0.806 -12.652
#>  6:  2002     6     6 12.580 0.328  0.810 -12.270
#>  7:  2002     6     7 18.982 0.330  0.815 -11.779
#>  8:  2002     6     8 19.120 0.332  0.821 -11.188
#>  9:  2002     6     9 25.709 0.335  0.828 -10.519
#> 10:  2002     6    10 25.948 0.338  0.836  -9.806
#> 11:  2002     6    11 26.176 0.341  0.843  -9.156
#> 12:  2002     6    12 26.392 0.344  0.851  -8.564
#> 13:  2002     6    13 27.021 0.330  0.823  -7.961
#> 14:  2002     6    14 41.061 0.334  0.834  -7.200
#> 15:  2002     6    15 31.487 0.339  0.846  -6.429
#> 16:  2002     6    16 42.672 0.344  0.860  -5.614
#> 17:  2002     6    17 43.824 0.353  0.883  -4.421
#> 18:  2002     6    18 12.847 0.367  0.754  -3.573
#> 19:  2002     6    19  6.866 0.343  0.804  -3.240
#> 20:  2002     6    20  6.728 0.336  0.784  -3.887
#> 21:  2002     6    21  6.626 0.331  0.769  -4.405
#>        yr    mo    nl  swati theta wetnes   psimi
#> 
#> $SWATANN.ASC
#>        yr    nl  swati theta wetnes   psimi
#>     <num> <int>  <num> <num>  <num>   <num>
#>  1:  2002     1  3.102 0.323  0.799 -13.444
#>  2:  2002     2  6.211 0.323  0.800 -13.347
#>  3:  2002     3  6.223 0.324  0.801 -13.178
#>  4:  2002     4  9.357 0.325  0.803 -12.968
#>  5:  2002     5 12.523 0.326  0.806 -12.652
#>  6:  2002     6 12.580 0.328  0.810 -12.270
#>  7:  2002     7 18.982 0.330  0.815 -11.779
#>  8:  2002     8 19.120 0.332  0.821 -11.188
#>  9:  2002     9 25.709 0.335  0.828 -10.519
#> 10:  2002    10 25.948 0.338  0.836  -9.806
#> 11:  2002    11 26.176 0.341  0.843  -9.156
#> 12:  2002    12 26.392 0.344  0.851  -8.564
#> 13:  2002    13 27.021 0.330  0.823  -7.961
#> 14:  2002    14 41.061 0.334  0.834  -7.200
#> 15:  2002    15 31.487 0.339  0.846  -6.429
#> 16:  2002    16 42.672 0.344  0.860  -5.614
#> 17:  2002    17 43.824 0.353  0.883  -4.421
#> 18:  2002    18 12.847 0.367  0.754  -3.573
#> 19:  2002    19  6.866 0.343  0.804  -3.240
#> 20:  2002    20  6.728 0.336  0.784  -3.887
#> 21:  2002    21  6.626 0.331  0.769  -4.405
#>        yr    nl  swati theta wetnes   psimi
#> 
#> $MISCPRE.ASC
#>       yr    mo    da   doy  pint vrfln safrac  adef  awat relawat  nits
#>    <num> <num> <num> <num> <int> <num>  <num> <num> <num>   <num> <num>
#> 1:  2002     6     1   152     1   0.2      0 0.045 254.8   0.983   233
#> 2:  2002     6     2   153     1   0.2      0 0.048 250.4   0.966   168
#> 3:  2002     6     3   154     1   0.3      0 0.051 245.9   0.949   134
#> 4:  2002     6     4   155     1   0.4      0 0.052 243.9   0.941   105
#> 5:  2002     6     5   156     1   0.6      0 0.054 241.4   0.931    80
#> 
#> $MISCDAY.ASC
#>       yr    mo    da   doy vrfln safrac stres  adef  awat relawat  nits balerr
#>    <num> <num> <num> <num> <num>  <num> <num> <num> <num>   <num> <num>  <num>
#> 1:  2002     6     1   152   0.2      0     1 0.045 254.8   0.983   233      0
#> 2:  2002     6     2   153   0.2      0     1 0.048 250.4   0.966   168      0
#> 3:  2002     6     3   154   0.3      0     1 0.051 245.9   0.949   134      0
#> 4:  2002     6     4   155   0.4      0     1 0.052 243.9   0.941   105      0
#> 5:  2002     6     5   156   0.6      0     1 0.054 241.4   0.931    80      0
#> 
#> $MISCMON.ASC
#>       yr    mo vrfln safrac stres  adef  awat relawat  nits balerr
#>    <num> <num> <num>  <num> <num> <num> <num>   <num> <num>  <num>
#> 1:  2002     6   1.8      0     1  0.05 247.3   0.954   720      0
#> 

# or calculate aggregations at run time by passing the function via output_fun-arg
run_LWFB90(options_b90 = opts,
           param_b90 = parms,
           climate = slb1_meteo,
           soil = soil,
           rtrn_input = FALSE,
           output_fun = process_outputs_LWFB90,
           selection = outsel)$output_fun
#> [[1]]
#> [[1]]$BUDGPRE.ASC
#>       yr    mo    da   doy  prec  flow  evap  seep  snow  swat  gwat  intr
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.2   2.9     0     0 418.0     0     0
#> 2:  2002     6     2   153   0.0   0.2   3.7     0     0 414.1     0     0
#> 3:  2002     6     3   154   0.0   0.3   3.6     0     0 410.2     0     0
#> 4:  2002     6     4   155   2.9   0.4   4.0     0     0 408.6     0     0
#> 5:  2002     6     5   156   2.4   0.6   4.1     0     0 406.3     0     0
#>     ints
#>    <num>
#> 1:     0
#> 2:     0
#> 3:     0
#> 4:     0
#> 5:     0
#> 
#> [[1]]$BUDGDAY.ASC
#>       yr    mo    da   doy  prec  flow  evap  seep  snow  swat  gwat  intr
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.2   2.9     0     0 418.0     0     0
#> 2:  2002     6     2   153   0.0   0.2   3.7     0     0 414.1     0     0
#> 3:  2002     6     3   154   0.0   0.3   3.6     0     0 410.2     0     0
#> 4:  2002     6     4   155   2.9   0.4   4.0     0     0 408.6     0     0
#> 5:  2002     6     5   156   2.4   0.6   4.1     0     0 406.3     0     0
#>     ints
#>    <num>
#> 1:     0
#> 2:     0
#> 3:     0
#> 4:     0
#> 5:     0
#> 
#> [[1]]$BUDGMON.ASC
#>       yr    mo  prec  flow  evap  seep  snow  swat  gwat  intr  ints
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   5.3   1.8  18.3     0     0 406.3     0     0     0
#> 
#> [[1]]$BUDGANN.ASC
#>       yr  prec  flow  evap  seep  snow  swat  gwat  intr  ints
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   5.3   1.8  18.3     0     0 406.3     0     0     0
#> 
#> [[1]]$FLOWPRE.ASC
#>       yr    mo    da   doy  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2     0     0   0.0     0     0   0.2   0.2
#> 2:  2002     6     2   153   0.2     0     0   0.0     0     0   0.2   0.2
#> 3:  2002     6     3   154   0.3     0     0   0.0     0     0   0.3   0.3
#> 4:  2002     6     4   155   0.4     0     0   1.8     0     0   0.4   0.4
#> 5:  2002     6     5   156   0.6     0     0   1.6     0     0   0.6   0.6
#> 
#> [[1]]$FLOWDAY.ASC
#>       yr    mo    da   doy  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2     0     0   0.0     0     0   0.2   0.2
#> 2:  2002     6     2   153   0.2     0     0   0.0     0     0   0.2   0.2
#> 3:  2002     6     3   154   0.3     0     0   0.0     0     0   0.3   0.3
#> 4:  2002     6     4   155   0.4     0     0   1.8     0     0   0.4   0.4
#> 5:  2002     6     5   156   0.6     0     0   1.6     0     0   0.6   0.6
#> 
#> [[1]]$FLOWMON.ASC
#>       yr    mo  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   1.8     0     0   3.4     0     0   1.8   1.8
#> 
#> [[1]]$FLOWANN.ASC
#>       yr  flow  seep  srfl  slfl  byfl  dsfl  gwfl vrfln
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   1.8     0     0   3.4     0     0   1.8   1.8
#> 
#> [[1]]$EVAPPRE.ASC
#>       yr    mo    da   doy  flow  evap  tran  irvp  isvp  slvp  snvp  pint
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2   2.9   2.6   0.0     0   0.3     0   5.1
#> 2:  2002     6     2   153   0.2   3.7   3.3   0.0     0   0.4     0   7.6
#> 3:  2002     6     3   154   0.3   3.6   3.1   0.0     0   0.5     0   8.8
#> 4:  2002     6     4   155   0.4   4.0   2.6   1.0     0   0.3     0   6.1
#> 5:  2002     6     5   156   0.6   4.1   2.6   0.9     0   0.6     0  10.4
#>    ptran pslvp
#>    <num> <num>
#> 1:   2.6   0.3
#> 2:   3.3   0.5
#> 3:   3.1   0.6
#> 4:   2.6   0.4
#> 5:   2.6   0.8
#> 
#> [[1]]$EVAPDAY.ASC
#>       yr    mo    da   doy  flow  evap  tran  irvp  isvp  slvp  snvp  pint
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.2   2.9   2.6   0.0     0   0.3     0   5.1
#> 2:  2002     6     2   153   0.2   3.7   3.3   0.0     0   0.4     0   7.6
#> 3:  2002     6     3   154   0.3   3.6   3.1   0.0     0   0.5     0   8.8
#> 4:  2002     6     4   155   0.4   4.0   2.6   1.0     0   0.3     0   6.1
#> 5:  2002     6     5   156   0.6   4.1   2.6   0.9     0   0.6     0  10.4
#>    ptran pslvp
#>    <num> <num>
#> 1:   2.6   0.3
#> 2:   3.3   0.5
#> 3:   3.1   0.6
#> 4:   2.6   0.4
#> 5:   2.6   0.8
#> 
#> [[1]]$EVAPMON.ASC
#>       yr    mo  flow  evap  tran  irvp  isvp  slvp  snvp  pint ptran pslvp
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   1.8  18.3  14.2   1.9     0   2.2     0    38  14.2   2.7
#> 
#> [[1]]$EVAPANN.ASC
#>       yr  flow  evap  tran  irvp  isvp  slvp  snvp  pint ptran pslvp
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   1.8  18.3  14.2   1.9     0   2.2     0    38  14.2   2.7
#> 
#> [[1]]$ABOVPRE.ASC
#>       yr    mo    da   doy  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.0     0     0   0.0     0     0   0.0
#> 2:  2002     6     2   153   0.0   0.0     0     0   0.0     0     0   0.0
#> 3:  2002     6     3   154   0.0   0.0     0     0   0.0     0     0   0.0
#> 4:  2002     6     4   155   2.9   1.0     0     0   1.8     0     0   1.8
#> 5:  2002     6     5   156   2.4   0.9     0     0   1.6     0     0   1.6
#>     smlt  slfl  srfl
#>    <num> <num> <num>
#> 1:     0   0.0     0
#> 2:     0   0.0     0
#> 3:     0   0.0     0
#> 4:     0   1.8     0
#> 5:     0   1.6     0
#> 
#> [[1]]$ABOVDAY.ASC
#>       yr    mo    da   doy  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6     1   152   0.0   0.0     0     0   0.0     0     0   0.0
#> 2:  2002     6     2   153   0.0   0.0     0     0   0.0     0     0   0.0
#> 3:  2002     6     3   154   0.0   0.0     0     0   0.0     0     0   0.0
#> 4:  2002     6     4   155   2.9   1.0     0     0   1.8     0     0   1.8
#> 5:  2002     6     5   156   2.4   0.9     0     0   1.6     0     0   1.6
#>     smlt  slfl  srfl
#>    <num> <num> <num>
#> 1:     0   0.0     0
#> 2:     0   0.0     0
#> 3:     0   0.0     0
#> 4:     0   1.8     0
#> 5:     0   1.6     0
#> 
#> [[1]]$ABOVMON.ASC
#>       yr    mo  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet  smlt  slfl
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002     6   5.3   1.9     0     0   3.4     0     0   3.4     0   3.4
#>     srfl
#>    <num>
#> 1:     0
#> 
#> [[1]]$ABOVANN.ASC
#>       yr  rfal  rint  sfal  sint  rthr  sthr  rsno  rnet  smlt  slfl  srfl
#>    <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#> 1:  2002   5.3   1.9     0     0   3.4     0     0   3.4     0   3.4     0
#> 
#> [[1]]$BELOPRE.ASC
#>         yr    mo    da   doy    nl  infl  byfl  tran  vrfl  dsfl  ntfl
#>      <num> <num> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#>   1:  2002     6     1   152     1   0.0     0   0.1  -0.2     0  -0.2
#>   2:  2002     6     2   153     1   0.0     0   0.1  -0.4     0  -0.1
#>   3:  2002     6     3   154     1   0.0     0   0.1  -0.5     0  -0.1
#>   4:  2002     6     4   155     1   1.8     0   0.1   1.3     0   0.1
#>   5:  2002     6     5   156     1   1.6     0   0.1   0.9     0   0.0
#>  ---                                                                  
#> 101:  2002     6     1   152    21   0.0     0   0.0   0.2     0   0.0
#> 102:  2002     6     2   153    21   0.0     0   0.0   0.2     0   0.1
#> 103:  2002     6     3   154    21   0.0     0   0.0   0.3     0   0.2
#> 104:  2002     6     4   155    21   0.0     0   0.0   0.4     0   0.2
#> 105:  2002     6     5   156    21   0.0     0   0.0   0.6     0   0.1
#> 
#> [[1]]$BELODAY.ASC
#>         yr    mo    da   doy    nl  infl  byfl  tran  vrfl  dsfl  ntfl
#>      <num> <num> <num> <num> <int> <num> <num> <num> <num> <num> <num>
#>   1:  2002     6     1   152     1   0.0     0   0.1  -0.2     0  -0.2
#>   2:  2002     6     2   153     1   0.0     0   0.1  -0.4     0  -0.1
#>   3:  2002     6     3   154     1   0.0     0   0.1  -0.5     0  -0.1
#>   4:  2002     6     4   155     1   1.8     0   0.1   1.3     0   0.1
#>   5:  2002     6     5   156     1   1.6     0   0.1   0.9     0   0.0
#>  ---                                                                  
#> 101:  2002     6     1   152    21   0.0     0   0.0   0.2     0   0.0
#> 102:  2002     6     2   153    21   0.0     0   0.0   0.2     0   0.1
#> 103:  2002     6     3   154    21   0.0     0   0.0   0.3     0   0.2
#> 104:  2002     6     4   155    21   0.0     0   0.0   0.4     0   0.2
#> 105:  2002     6     5   156    21   0.0     0   0.0   0.6     0   0.1
#> 
#> [[1]]$BELOMON.ASC
#>        yr    mo    nl    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#>     <num> <num> <int> <num> <num> <num> <num> <num> <num> <num> <num>
#>  1:  2002     6     1    15   770   3.4     0   0.4   1.1     0  -0.4
#>  2:  2002     6     2    15   770   0.0     0   0.9   0.9     0  -0.7
#>  3:  2002     6     3    15   770   0.0     0   0.8   0.8     0  -0.7
#>  4:  2002     6     4    15   770   0.0     0   1.1   0.8     0  -1.1
#>  5:  2002     6     5    15   770   0.0     0   1.3   0.8     0  -1.4
#>  6:  2002     6     6    15   770   0.0     0   1.2   1.0     0  -1.4
#>  7:  2002     6     7    15   770   0.0     0   1.5   1.4     0  -2.0
#>  8:  2002     6     8    15   770   0.0     0   1.3   2.0     0  -1.9
#>  9:  2002     6     9    15   770   0.0     0   1.4   3.0     0  -2.3
#> 10:  2002     6    10    15   770   0.0     0   1.1   3.9     0  -2.1
#> 11:  2002     6    11    15   770   0.0     0   0.8   4.9     0  -1.8
#> 12:  2002     6    12    15   770   0.0     0   0.7   5.8     0  -1.5
#> 13:  2002     6    13    15   770   0.0     0   0.5   6.7     0  -1.4
#> 14:  2002     6    14    15   770   0.0     0   0.5   7.6     0  -1.4
#> 15:  2002     6    15    15   770   0.0     0   0.3   7.7     0  -0.4
#> 16:  2002     6    16    15   770   0.0     0   0.2   7.0     0   0.5
#> 17:  2002     6    17    15   770   0.0     0   0.1   4.9     0   2.0
#> 18:  2002     6    18    15   770   0.0     0   0.0   3.9     0   1.0
#> 19:  2002     6    19    15   770   0.0     0   0.0   3.1     0   0.7
#> 20:  2002     6    20    15   770   0.0     0   0.0   2.4     0   0.7
#> 21:  2002     6    21    15   770   0.0     0   0.0   1.8     0   0.7
#>        yr    mo    nl    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#> 
#> [[1]]$BELOANN.ASC
#>        yr    nl    mo    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#>     <num> <int> <num> <num> <num> <num> <num> <num> <num> <num> <num>
#>  1:  2002     1    30    15   770   3.4     0   0.4   1.1     0  -0.4
#>  2:  2002     2    30    15   770   0.0     0   0.9   0.9     0  -0.7
#>  3:  2002     3    30    15   770   0.0     0   0.8   0.8     0  -0.7
#>  4:  2002     4    30    15   770   0.0     0   1.1   0.8     0  -1.1
#>  5:  2002     5    30    15   770   0.0     0   1.3   0.8     0  -1.4
#>  6:  2002     6    30    15   770   0.0     0   1.2   1.0     0  -1.4
#>  7:  2002     7    30    15   770   0.0     0   1.5   1.4     0  -2.0
#>  8:  2002     8    30    15   770   0.0     0   1.3   2.0     0  -1.9
#>  9:  2002     9    30    15   770   0.0     0   1.4   3.0     0  -2.3
#> 10:  2002    10    30    15   770   0.0     0   1.1   3.9     0  -2.1
#> 11:  2002    11    30    15   770   0.0     0   0.8   4.9     0  -1.8
#> 12:  2002    12    30    15   770   0.0     0   0.7   5.8     0  -1.5
#> 13:  2002    13    30    15   770   0.0     0   0.5   6.7     0  -1.4
#> 14:  2002    14    30    15   770   0.0     0   0.5   7.6     0  -1.4
#> 15:  2002    15    30    15   770   0.0     0   0.3   7.7     0  -0.4
#> 16:  2002    16    30    15   770   0.0     0   0.2   7.0     0   0.5
#> 17:  2002    17    30    15   770   0.0     0   0.1   4.9     0   2.0
#> 18:  2002    18    30    15   770   0.0     0   0.0   3.9     0   1.0
#> 19:  2002    19    30    15   770   0.0     0   0.0   3.1     0   0.7
#> 20:  2002    20    30    15   770   0.0     0   0.0   2.4     0   0.7
#> 21:  2002    21    30    15   770   0.0     0   0.0   1.8     0   0.7
#>        yr    nl    mo    da   doy  infl  byfl  tran  vrfl  dsfl  ntfl
#> 
#> [[1]]$SWATPRE.ASC
#>         yr    mo    da   doy    nl swati theta wetnes   psimi
#>      <num> <num> <num> <num> <num> <num> <num>  <num>   <num>
#>   1:  2002     6     1   152     1 3.219 0.335  0.830 -10.282
#>   2:  2002     6     2   153     1 3.108 0.324  0.801 -13.167
#>   3:  2002     6     3   154     1 3.024 0.315  0.779 -15.758
#>   4:  2002     6     4   155     1 3.096 0.323  0.797 -13.512
#>   5:  2002     6     5   156     1 3.063 0.319  0.789 -14.502
#>  ---                                                         
#> 101:  2002     6     1   152    21 6.316 0.316  0.723  -6.133
#> 102:  2002     6     2   153    21 6.425 0.321  0.739  -5.395
#> 103:  2002     6     3   154    21 6.614 0.331  0.767  -4.310
#> 104:  2002     6     4   155    21 6.818 0.341  0.797  -3.362
#> 105:  2002     6     5   156    21 6.957 0.348  0.818  -2.823
#> 
#> [[1]]$SWATDAY.ASC
#>         yr    mo    da   doy    nl swati theta wetnes   psimi
#>      <num> <num> <num> <num> <int> <num> <num>  <num>   <num>
#>   1:  2002     6     1   152     1 3.219 0.335  0.830 -10.282
#>   2:  2002     6     2   153     1 3.108 0.324  0.801 -13.167
#>   3:  2002     6     3   154     1 3.024 0.315  0.779 -15.758
#>   4:  2002     6     4   155     1 3.096 0.323  0.797 -13.512
#>   5:  2002     6     5   156     1 3.063 0.319  0.789 -14.502
#>  ---                                                         
#> 101:  2002     6     1   152    21 6.316 0.316  0.723  -6.133
#> 102:  2002     6     2   153    21 6.425 0.321  0.739  -5.395
#> 103:  2002     6     3   154    21 6.614 0.331  0.767  -4.310
#> 104:  2002     6     4   155    21 6.818 0.341  0.797  -3.362
#> 105:  2002     6     5   156    21 6.957 0.348  0.818  -2.823
#> 
#> [[1]]$SWATMON.ASC
#>        yr    mo    nl  swati theta wetnes   psimi
#>     <num> <num> <int>  <num> <num>  <num>   <num>
#>  1:  2002     6     1  3.102 0.323  0.799 -13.444
#>  2:  2002     6     2  6.211 0.323  0.800 -13.347
#>  3:  2002     6     3  6.223 0.324  0.801 -13.178
#>  4:  2002     6     4  9.357 0.325  0.803 -12.968
#>  5:  2002     6     5 12.523 0.326  0.806 -12.652
#>  6:  2002     6     6 12.580 0.328  0.810 -12.270
#>  7:  2002     6     7 18.982 0.330  0.815 -11.779
#>  8:  2002     6     8 19.120 0.332  0.821 -11.188
#>  9:  2002     6     9 25.709 0.335  0.828 -10.519
#> 10:  2002     6    10 25.948 0.338  0.836  -9.806
#> 11:  2002     6    11 26.176 0.341  0.843  -9.156
#> 12:  2002     6    12 26.392 0.344  0.851  -8.564
#> 13:  2002     6    13 27.021 0.330  0.823  -7.961
#> 14:  2002     6    14 41.061 0.334  0.834  -7.200
#> 15:  2002     6    15 31.487 0.339  0.846  -6.429
#> 16:  2002     6    16 42.672 0.344  0.860  -5.614
#> 17:  2002     6    17 43.824 0.353  0.883  -4.421
#> 18:  2002     6    18 12.847 0.367  0.754  -3.573
#> 19:  2002     6    19  6.866 0.343  0.804  -3.240
#> 20:  2002     6    20  6.728 0.336  0.784  -3.887
#> 21:  2002     6    21  6.626 0.331  0.769  -4.405
#>        yr    mo    nl  swati theta wetnes   psimi
#> 
#> [[1]]$SWATANN.ASC
#>        yr    nl  swati theta wetnes   psimi
#>     <num> <int>  <num> <num>  <num>   <num>
#>  1:  2002     1  3.102 0.323  0.799 -13.444
#>  2:  2002     2  6.211 0.323  0.800 -13.347
#>  3:  2002     3  6.223 0.324  0.801 -13.178
#>  4:  2002     4  9.357 0.325  0.803 -12.968
#>  5:  2002     5 12.523 0.326  0.806 -12.652
#>  6:  2002     6 12.580 0.328  0.810 -12.270
#>  7:  2002     7 18.982 0.330  0.815 -11.779
#>  8:  2002     8 19.120 0.332  0.821 -11.188
#>  9:  2002     9 25.709 0.335  0.828 -10.519
#> 10:  2002    10 25.948 0.338  0.836  -9.806
#> 11:  2002    11 26.176 0.341  0.843  -9.156
#> 12:  2002    12 26.392 0.344  0.851  -8.564
#> 13:  2002    13 27.021 0.330  0.823  -7.961
#> 14:  2002    14 41.061 0.334  0.834  -7.200
#> 15:  2002    15 31.487 0.339  0.846  -6.429
#> 16:  2002    16 42.672 0.344  0.860  -5.614
#> 17:  2002    17 43.824 0.353  0.883  -4.421
#> 18:  2002    18 12.847 0.367  0.754  -3.573
#> 19:  2002    19  6.866 0.343  0.804  -3.240
#> 20:  2002    20  6.728 0.336  0.784  -3.887
#> 21:  2002    21  6.626 0.331  0.769  -4.405
#>        yr    nl  swati theta wetnes   psimi
#> 
#> [[1]]$MISCPRE.ASC
#>       yr    mo    da   doy  pint vrfln safrac  adef  awat relawat  nits
#>    <num> <num> <num> <num> <int> <num>  <num> <num> <num>   <num> <num>
#> 1:  2002     6     1   152     1   0.2      0 0.045 254.8   0.983   233
#> 2:  2002     6     2   153     1   0.2      0 0.048 250.4   0.966   168
#> 3:  2002     6     3   154     1   0.3      0 0.051 245.9   0.949   134
#> 4:  2002     6     4   155     1   0.4      0 0.052 243.9   0.941   105
#> 5:  2002     6     5   156     1   0.6      0 0.054 241.4   0.931    80
#> 
#> [[1]]$MISCDAY.ASC
#>       yr    mo    da   doy vrfln safrac stres  adef  awat relawat  nits balerr
#>    <num> <num> <num> <num> <num>  <num> <num> <num> <num>   <num> <num>  <num>
#> 1:  2002     6     1   152   0.2      0     1 0.045 254.8   0.983   233      0
#> 2:  2002     6     2   153   0.2      0     1 0.048 250.4   0.966   168      0
#> 3:  2002     6     3   154   0.3      0     1 0.051 245.9   0.949   134      0
#> 4:  2002     6     4   155   0.4      0     1 0.052 243.9   0.941   105      0
#> 5:  2002     6     5   156   0.6      0     1 0.054 241.4   0.931    80      0
#> 
#> [[1]]$MISCMON.ASC
#>       yr    mo vrfln safrac stres  adef  awat relawat  nits balerr
#>    <num> <num> <num>  <num> <num> <num> <num>   <num> <num>  <num>
#> 1:  2002     6   1.8      0     1  0.05 247.3   0.954   720      0
#> 
#>