Functions to derive soil hydraulic properties from soil properties

A set of pedotransfer functions for deriving Mualem - van Genuchten parameters from soil physical properties of soil horizons, such as soil texture, bulk density and carbon content.

Usage

hydpar_puh2(clay, silt, sand, bd, oc.pct = 0.5)

hydpar_hypres(clay, silt, bd, oc.pct = 0.1, topsoil = TRUE, humconv = 1.72)

hydpar_hypres_tab(texture, topsoil)

hydpar_wessolek_tab(texture)

hydpar_ff_b90(n = 1)

Arguments

clay, silt, sand

Numeric vectors of clay, silt, sand in mass %. Particle size ranges for clay, silt and sand correspond to <2, 2-63, and 63-2000 \(\mu m\). For hydpar_hypres, the particle size limit between silt and sand should be 50 \(\mu m\).

bd

Numeric vector of bulk density in g cm-3.

oc.pct

Numeric vector of organic carbon content in mass %.

topsoil

Logical: Is the sample from the topsoil? Used in hydpar_hypres_tab.

humconv

Conversion factor from oc.pct to organic matter percent. Default: 1.72. Only for hydpar_hypres_tab.

texture

Character vector of soil texture classes. For hydpar_wessolek_tab classes according to KA5 (AG Boden 2005) have to be provided. When using hydpar_hypres_tab, texture classes according to FAO (1990) have to provided.

n

An integer value specifying the number of rows of the returned data.frame (i.e. the number of repetitions of the MvG-Parameter set, only for hydpar_ff_hamken).

Value

A data.frame with the following variables:

ths

Saturation water content fraction

thr

Residual water content fraction

npar

N parameter of the van Genuchten water retention function

mpar

M parameter of the van Genuchten water retention function

alpha

Alpha parameter of the van Genuchten water retention function (1/m)

ksat

Saturated hyraulic conductivity parameter of Mualem hydraulic conductivity function (mm/d)

tort

Tortuosity parameter of Mualem hydraulic conductivity function

Details

Function hydpar_puh2 derives Mualem - van Genuchten (MvG) parameters using the regression functions developed by Puhlmann & von Wilpert (2011). The equations of Wösten et al. (1999) are available via hydpar_hypres, and their tabulated values for soil texture classes can be derived using the function hydpar_hypres_tab. The table of MvG parameters from Wesselok et al. (2009; Tab. 10) is accessible by hydpar_wessolek_tab. For this function, soil texture classes after the German texture classification system (KA5, AG Boden 2005) have to be provided. To derive hydraulic parameters of forest floor horizons, the function hydpar_ff_b90 can be used. It returns the single MvG parameter set for forest floor horizons reported by Hammel & Kennel (2001) in their original LWF-Brook90 publication.

References

AG Boden (2005) Bodenkundliche Kartieranleitung Schweizerbart'sche Verlagsbuchhandlung, Stuttgart

Food and Agriculture Organisation (FAO) (1990) Guidelines for soil description FAO/ISRIC, Rome, 3rd edition

Hammel K & Kennel M (2001) Charakterisierung und Analyse der Wasserverfügbarkeit und des Wasserhaushalts von Waldstandorten in Bayern mit dem Simulationsmodell BROOK90. Forstliche Forschungsberichte München 185

Puhlmann H, von Wilpert K (2011) Testing and development of pedotransfer functions for water retention and hydraulic conductivity of forest soils. Waldökologie, Landschaftsforschung und Naturschutz 12, pp. 61-71

Wessolek G, Kaupenjohann M and Renger H (2009) Bodenphysikalische Kennwerte und Berechnungsverfahren für die Praxis. Bodenökologie und Bodengenese 40, Berlin, Germany

Woesten JHM, Lilly A, Nemes A, Le Bas C (1999) Development and use of a database of hydraulic properties of European soils. Geoderma 90, pp. 169-185

Examples

hydpar_puh2(clay = c(10,20), silt = c(40,20), sand = c(50,60), bd = c(1.6, 1.4))
#>         ths   thr    alpha     npar      mpar      ksat       tort
#> 1 0.3700561 0.069 2.063483 1.245316 0.1969907  35.34026 -0.1221964
#> 2 0.3992240 0.069 8.156158 1.278638 0.2179176 108.53398 -1.4644501

hydpar_hypres(20,20,1.5,2)
#>        ths thr    alpha     npar      mpar     ksat     tort
#> 1 0.394632   0 5.462451 1.181339 0.1535029 364.6788 -3.61904

hydpar_hypres_tab(texture = c("C","MF"), topsoil = c(TRUE,FALSE))
#>     ths   thr alpha  npar  mpar ksat tort
#> 1 0.403 0.025  3.83 1.377 0.274  600 1.25
#> 2 0.412 0.010  0.82 1.218 0.179   40 0.50

hydpar_wessolek_tab(c("Us", "Ls2", "mSfS"))
#>      ths    thr alpha    npar      mpar   ksat   tort
#> 2 0.3946 0.0000 2.747 1.22393 0.1829598 355.26 -2.728
#> 1 0.4148 0.1406 4.052 1.32416 0.2448043 384.30 -2.067
#> 3     NA     NA    NA      NA        NA     NA     NA

hydpar_ff_b90(n = 5)
#>       ths thr alpha  npar      mpar  ksat tort
#> 1   0.848   0    98 1.191 0.1603694 98000  0.5
#> 1.1 0.848   0    98 1.191 0.1603694 98000  0.5
#> 1.2 0.848   0    98 1.191 0.1603694 98000  0.5
#> 1.3 0.848   0    98 1.191 0.1603694 98000  0.5
#> 1.4 0.848   0    98 1.191 0.1603694 98000  0.5