Create a list of model control options

Create a list of model control options

set_optionsLWFB90(...)

Arguments

...

Named values to be included in return value.

Value

A list of model control options for use as

options_b90-argument in run_LWFB90.

Details

startdate

start date of the simulation.

enddate

end date of the simulation.

fornetrad

use global solar radiation (='globrad') or sunshine duration hours (='sunhours') for net radiation calculation?

prec_interval

number of precipitation intervals per day (default is 1). If prec_interval > 1, the precip-argument has to be provided to run_LWFB90

correct_prec

correct precipitation data for wind and evaporation losses using correct_prec?

budburst_method

name of method for budburst calculation. If 'constant' or 'fixed', budburst day of year from parameters is used. All other methods calculate budburst day of year dynamically from airtemperatures, and the method name is passed to the start.method-argument of vegperiod.

leaffall_method

name of method for leaffall calculation. If 'constant' or 'fixed', beginning of leaffall (day of year) from parameters is used. All other methods calculate budburst day of year dynamically from temperatures, and the method name is passed to the end.method-argument of vegperiod.

standprop_input

name of input for longterm (interannual) plant development. standprop_input = 'parameters': yearly values of stand properties height, sai, densef, lai are taken from individual parameters, standprop_input = 'table': values from standprop_table provided in parameters are used.

standprop_interp

interpolation method for aboveground stand properties. 'linear' or 'constant', see approx.method-argument of approx_standprop.

use_growthperiod

Should yearly changes of stand properties (growth) only take place during the growth period? If TRUE, linear interpolation of height, sai, densef and age are made from budburst until leaffall. During winter values are constant. Beginning and end of the growth period are taken from parameters budburstdoy and leaffalldoy. See use_growthperiod-argument of approx_standprop.

lai_method

name of method for constructing seasonal course leaf area index development from parameters. Passed to method-argument of make_seasLAI.

imodel

name of retention & conductivity model: "CH" for Clapp/Hornberger, "MvG" for Mualem/van Genuchten

root_method

method name of the root length density depth distribution function. Any of the names accepted by make_rootden are allowed. Additionally, 'soilvar' can be used if the root length density depth distribution is specified in column 'rootden' in the soil-data.frame

Examples

# Default options
options_b90 <- set_optionsLWFB90()
# Include specific options
options_b90_dynamic_phenology <- set_optionsLWFB90(budburst_method = 'Menzel',
leaffall_method ='vonWilpert')