Create the list of model parameters

Create the list of model parameters

set_paramLWFB90(...)

Arguments

...

Named arguments to be included in return value.

Value

A list with model parameters for use as param_b90-argument in run_LWFB90.

List of input parameters

Name	Description	Unit	Group

czr	Ratio of roughness length to mean height for smooth closed canopies for heights greater than HR when LAI>LPC. Default: 0.05	-	Canopy

czs	Ratio of roughness length to mean height for smooth closed canopies for heights less than HS when LAI>LPC. Default: 0.13	m	Canopy

hr	Smallest height to which CZR applies. Default: 10	m	Canopy

hs	Largest height to which CZS applies. Default: 1	m	Canopy

lpc	Minimum leaf area index defining a closed canopy. Default: 4	-	Canopy

lwidth	Average leaf width. Default: 0.1	m	Canopy

nn	Eddy diffusivity extinction coefficient within canopy. Default: 2.5	-	Canopy

rhotp	Ratio of total leaf area to projected area. Default: 2	-	Canopy

zminh	Reference height for weather data above the canopy top height. Default: 2	m	Canopy

dslope	Slope for downslope-flow. Default: 0	deg	Flow

bypar	Switch to allow (1) or prevent (0) bypass flow in deeper layers. Default: 0	-	Flow

drain	Switch for lower boundary condition to be free drainage (1) or no flow (0). Default: 1	-	Flow

slopelen	Slope length for downslope-flow. Default: 200	m	Flow

gsc	Rate constant for ground water discharge (remember that a first order groundwater reservoir is placed below the soil profile), for value 0 there is no discharge. Default: 0	d-1	Flow

gsp	Seepage fraction of groundwater discharge. Default:	-	Flow

ilayer	Number of layers from top to which infiltration is distributed. Default: 0	-	Flow

imperv	Fraction of area which has an impermeable surface (like roads). Default: 0	-	Flow

infexp	Shape parameter for distribution of infiltration in first ILayer, for value 0 infiltration is in top layer only. Default: 0	-	Flow

qffc	Quickflow fraction of infiltrating water at field capacity, for value 0 there is no quickflow (bypass or surface) unless soil profile surface becomes saturated. Default: 0	-	Flow

qfpar	Quickflow shape parameter. Default: 1	-	Flow

qlayer	Number of layers which are considered for generation of surface or source area flow. Default: 0	-	Flow

water_table_depth	Depth of the water table for a constant head boundary. -9999 means gravitational flow boundary. Default: -9999	m	Flow

gwatini	Initial value of groundwater storage. Default: 0	mm	Initial

snowini	Initial value of water content of snow pack. Default: 0	mm	Initial

snowlqini	Initial value of liquid water content of snow pack. Default: 0	mm	Initial

snowccini	Initial value of cold content of snow pack (positive). Default: 0	MJ m-2	Initial

intrainini	Initial value of intercepted rain. Default: 0	mm	Initial

intsnowini	Initial value of intercepted snow. Default: 0	-	Initial

psiini	Initial pressure head of soil layers. May have the same length as row.names(soil). Default: -6.3	kPa	Initial

cintrl	Maximum interception storage of rain per unit LAI. Default: 0.15	mm	Interception

cintrs	Maximum interception storage of rain per unit SAI. Default: 0.15	mm	Interception

cintsl	Maximum interception storage of snow per unit LAI. Default: 0.6	mm	Interception

cintss	Maximum interception storage of snow per unit SAI. Default: 0.6	mm	Interception

frintlai	Intercepted fraction of rain per unit LAI. Default: 0.06	-	Interception

frintsai	Intercepted fraction of rain per unit SAI. Default: 0.06	-	Interception

fsintlai	Intercepted fraction of snow per unit LAI. Default: 0.04	-	Interception

fsintsai	Intercepted fraction of snow per unit SAI. Default: 0.04	-	Interception

pdur	Average duration of precipitation events for each month of the year. Default: rep(4,12)	hours	Interception

alb	Albedo of soil/vegetation surface without snow. Default: 0.2	-	Meteo

albsn	Albedo of soil/vegetation surface with snow. Default: 0.5	-	Meteo

c1	Intercept of relation of solar radiation to sunshine duration. Default: 0.25	-	Meteo

c2	Intercept of relation of solar radiation to sunshine duration. Default: 0.5	-	Meteo

c3	Constant between 0 and 1 that determines the cloud correction to net longwave radiation from sunshine duration. Default: 0.2	-	Meteo

fetch	Fetch upwind of the weather station at which wind speed was measured. Default: 5000	m	Meteo

ksnvp	Correction factor for snow evaporation. Default: 0.3	-	Meteo

wndrat	Average ratio of nighttime to daytime wind speed. Default: 0.3	-	Meteo

z0s	Surface roughness of snow cover. Default: 0.001	m	Meteo

z0w	Roughness length at the weather station at which wind speed was measured. Default: 0.005 (Grass)	m	Meteo

coords_x	Longitude value (decimal degrees) of the simulation location (has no effect on simulation results). Default: 9.91	m	Meteo

coords_y	Latitude value (decimal degrees) of the simulation location. Default: 51.54	m	Meteo

zw	Height at which wind speed was measured. Default: 2	m	Meteo

eslope	Slope for evapotranspiration and snowmelt calculation. Default: 0	deg	Meteo

aspect	Mean exposition of soil surface at soil profile (north: 0, east: 90, south: 180, west: 270). Default: 0	deg	Meteo

obsheight	Mean height of obstacles on soil surface (grass, furrows etc.), used to calculate soil surface roughness. Default: 0.025	m	Meteo

prec_corr_statexp	Station exposure situation of prec measurements (passed to correct_prec Default: 'mg'		Meteo

dpsimax	Maximum potential difference considered equal. Default: 5e-04	kPa	Numerical

dswmax	Maximum change allowed in SWATI. Default: 0.05	percent of SWATMX	Numerical

dtimax	Maximum iteration time step. Default: 0.5	d	Numerical

budburst_species	Name of tree species for estimating budburst doy using Menzel-model (passed to vegperiod) Default: 'Fagus sylvatica'	-	Plant

budburstdoy	Budburst day of year - passed to make_seasLAI. Default: 121	doy	Plant

emergedur	Leaf growth duration until maxlai is reached.. Default: 28	d	Plant

height	Plant height. Default: 25	m	Plant

height_ini	Initial plant height at the beginning of the simulation. Used when options_b90$standprop_interp = 'linear' (see approx_standprop). Default: 25	m	Plant

leaffalldoy	Day of year when leaf fall begins - passed to make_seasLAI Default: 279	doy	Plant

leaffalldur	Number of days until minimum lai is reached - passed to make_seasLAI Default: 58	d	Plant

sai	Steam area index. Default: 1	-	Plant

sai_ini	Initial stem area index at the beginning of the simulation. Used when options_b90$standprop_interp = 'linear' (see approx_standprop), Default: 1	-	Plant

shp_leaffall	Shape parameter for leaf fall phase - passed to make_seasLAI Default: 0.3	-	Plant

shp_budburst	Shape parameter for leaf growth phase - passed to make_seasLAI Default: 3	-	Plant

shp_optdoy	Day of year when optimum value is reached - passed to make_seasLAI Default: 210	doy	Plant

lai_doy	Day of year values for lai-interpolation - passed to make_seasLAI	doy	Plant

lai_frac	Fractional lai values for lai interpolation, corresponding to lai_doy - passed to make_seasLAI Default: 210	doy	Plant

winlaifrac	Minimum LAI as a fraction of maxlai. Default: 0	-	Plant

standprop_table	Data.frame with yearly values of vegetation properties with columns 'year','age', 'height', 'maxlai', 'sai', 'densef'		Plant

cs	Ratio of projected stem area index to canopy height. Default: 0.035	m-1	Plant

densef	Density factor for MaxLAI, CS, RtLen, RPlant, not <.001, 1 for typical stand. Default: 1	-	Plant

densef_ini	Initial density factor at the beginning of the simulation. Used when options_b90$standprop_interp = 'linear' (see approx_standprop). Default: 1	-	Plant

maxlai	Maximum projected leaf area index - passed to make_seasLAI Default: 5	-	Plant

radex	Extinction coefficient for solar radiation and net radiation in the canopy. Default: 0.5	-	Potential Transpiration

cvpd	Vapour pressure deficit at which leaf conductance is halved. Default: 2	kPa	Potential Transpiration

glmax	Maximum leaf vapour conductance when stomata are fully open. Default: 0.0053	m s-1	Potential Transpiration

glmin	Minimum leaf vapour conductance when stomata are closed. Default: 0.0003	m s-1	Potential Transpiration

r5	Solar radiation level at which leaf conductance is half of its value at RM. Default: 100	W m-2	Potential Transpiration

rm	Nominal maximum solar shortwave radiation possible on a leaf (to reach glmax). Default: 1000	W m-2	Potential Transpiration

t1	Lower suboptimal temperature threshold for stomata opening - temperature relation. Default: 10	deg C	Potential Transpiration

t2	Upper suboptimal temperature threshold for stomata opening - temperature relation. Default: 30	deg C	Potential Transpiration

th	Upper temperature threshold for stomata closure. Default: 40	deg C	Potential Transpiration

tl	Lower temperature threshold for stomata closure. Default: 0	deg C	Potential Transpiration

betaroot	Shape parameter for rootlength density depth distribution. Default: 0.97	-	Roots

maxrootdepth	Maximum root depth (positive downward) - passed to make_rootden. Default: -1.5	m	Roots

rootden_table	Data.frame of relative root density depth distribution with columns 'depth' and 'rootden'		Roots

rstemp	Base temperature for snow-rain transition. Default: -0.5	deg C	Snow

ccfac	Cold content factor. Default: 0.3	MJ m-2 d-1 K-1	Snow

grdmlt	Rate of groundmelt of snowpack. Default: 0.35	mm d-1	Snow

laimlt	Parameter for snowmelt dependence on LAI. Default: 0.2	-	Snow

maxlqf	Maximum liquid water fraction of Snow. Default: 0.05	-	Snow

melfac	Degree day melt factor for open. Default: 1.5	MJ m-2 d-1 K-1	Snow

saimlt	Parameter for snowmelt dependence on SAI. Default: 0.5	-	Snow

snoden	Snow density. Default: 0.3	mm mm-1	Snow

rssa	Soil evaporation resistance at field capacity. Default: 100	s m-1	Soilevap

rssb	Exponent in relation of RSS to water potential. Default: 1	-	Soilevap

soil_nodes	Data.frame with soil nodes discretization passed to LWF-Brook90	-	Soil

soil_materials	Data.frame with soil materials (hydrualic parameters) passed to LWF-Brook90	-	Soil

age_ini	Age of stand (for root development). Default: 100	a	Water supply

initrdep	Initial root depth. Default: 0.25	m	Water supply

initrlen	Initial water-absorbing root length per unit area. Default: 12	m/m-2	Water supply

rgroper	Period of net root growth. A value of 0 prevents root growth. Default: 0	a	Water supply

rgrorate	Vertical root growth rate. Default: 0.03	m a-1	Water supply

fxylem	Fraction of internal plant resistance to water flow that is in the Xylem. Default: 0.5	-	Water supply

maxrlen	Total length of fine roots per unit ground area. Default: 3000	m m-2	Water supply

mxkpl	Maximum internal conductivity for water flow through the plants. Default: 8	mm d-1 MPa-1	Water supply

nooutf	Switch that prevents outflow from the root to the soil when the soil is dry. Default: 1	-	Water supply

psicr	Critical leaf water potential at which stomates close. Default: -2	MPa	Water supply

rrad	Average radius of the fine or water-absorbing roots. Default: 0.35	mm	Water supply

Examples

# Default parameter
parms <- set_paramLWFB90()
# Include specific parameters
parms_maxlai <- set_paramLWFB90(maxlai = c(4,6,5), height =20)