Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.949929  3.160149  3.864925  4.117430  4.196927  2.846339  2.754912
#>   [8]  4.223555  2.759978  4.023956  2.237192  4.094841  3.827938  4.200300
#>  [15]  2.307877  2.877355  2.317801  4.954047  4.645044  2.508653  3.867610
#>  [22]  2.540065  2.764332  4.386065  4.323606  2.623247  4.508754  4.037239
#>  [29]  3.725443  4.722862  3.320008  2.767657  4.113992  3.224455  6.354493
#>  [36]  3.322844  3.615032  6.145182  6.660478  6.266642  4.144332  5.201136
#>  [43]  5.980053  4.911398  3.578712  6.088426  4.776076  4.761533  5.658740
#>  [50]  3.591142  3.884379  8.322915  5.329451  5.846179  5.704937  7.327436
#>  [57]  6.395286  5.805344  7.877417  4.838125  4.955988  6.870236  6.455399
#>  [64]  6.609875  5.459103  5.362930  9.629447  5.790393  5.312646 10.247441
#>  [71]  5.416451  6.618900  8.524185  6.274485 11.556208  7.614022  6.912946
#>  [78]  8.548634 11.687776  9.899655  9.920316  9.255322 11.889397 10.147112
#>  [85] 10.368246  8.599226  6.436943  6.319213  6.784087  9.629809 13.360880
#>  [92] 13.783801 10.111133  6.860303 12.107349  8.409338 13.622789  9.648879
#>  [99]  9.332711 10.103264 11.184777  9.422957 14.688624  7.914835 15.407176
#> [106] 12.180591 10.039561  9.416196  9.027337 11.980354 11.598880 12.194440
#> [113] 15.711830 12.339302 13.137621  9.948194 11.097579 10.780838  9.244142
#> [120] 16.207556  9.940424 15.876770 10.785857 15.383036 13.921960 10.236088
#> [127]  9.145136 11.790159 17.072272 12.703521 16.716861 15.280629 12.307860
#> [134]  9.594780 17.437885 15.921125 16.822916 13.895525 12.456693 16.247553
#> [141] 10.300891 17.742972 15.347464 12.111737 12.990465 18.474317 10.381493
#> [148] 18.172223 17.854850 11.287535 18.791658 17.758246 15.055321 11.302575
#> [155] 17.693201 11.840965 11.005761 16.097997 11.735155 14.746226 14.498016
#> [162] 19.059071 18.109064 13.363791 16.559198 16.545120 15.987518 10.671233
#> [169] 15.553485 12.683458 11.972179 18.254524 18.737358 16.849751 10.504196
#> [176] 15.730930 12.564334 18.170727 17.926342 18.597782 15.407251 11.077004
#> [183] 15.469907 14.914142 11.354488 10.320871 12.114825 11.894026 17.946864
#> [190] 16.414674 11.802685 14.333704 16.251074 13.792988 16.030592 13.503357
#> [197] 18.047597 14.646577 15.657361 13.176252 16.948226 11.960818  9.978166
#> [204] 13.413346 12.955426 11.666725 10.255540  9.740682 16.138954 15.165024
#> [211] 11.906148 13.313431  9.746595 10.023932 15.569274 10.022735 12.947351
#> [218]  9.360842 10.391112 14.804400 10.934543 13.696979 10.634390 11.053969
#> [225] 13.644616 16.379138 14.415822 14.527489 13.600344 15.544264  8.731440
#> [232] 13.356485 14.164019 15.617461 14.512073 11.193246 12.972015  9.799323
#> [239] 13.444077 13.555883  8.268719 11.879626  9.282951  8.716397 12.105585
#> [246]  7.629572  9.838925 13.760212 12.297047 12.901369 11.978322  9.968793
#> [253]  7.235487 10.805203  9.278890 10.244311  6.450023  9.433056  7.981023
#> [260] 11.249895 10.003197  6.819307  9.535028  6.393128 11.460183  7.421632
#> [267] 10.659349  8.810262  7.686590 10.136378  9.591375  8.271658  9.114935
#> [274]  7.942149  7.658505  4.829787  6.513592  8.593585  8.074668  7.273042
#> [281]  9.303983  4.367734  9.380274  9.342039  4.604116  7.994961  7.450401
#> [288]  5.241916  6.336445  6.859104  6.355226  3.540680  4.537342  4.695072
#> [295]  4.837240  4.718394  6.198352  4.426772  4.544258  6.962380  5.329112
#> [302]  6.233746  2.998379  4.934854  5.785694  6.651847  4.659434  5.713188
#> [309]  5.465921  5.695823  4.016158  5.488227  2.875601  2.790834  3.423297
#> [316]  2.906997  2.512233  4.322508  5.485785  5.431475  5.048404  2.938764
#> [323]  3.685515  2.016642  3.763052  3.894234  3.963491  4.896173  2.063216
#> [330]  2.988634  2.074504  4.037288  2.162432  2.905130  4.188923  2.970191
#> [337]  2.088989  3.455701  2.546762  4.328073  3.807376  4.069019  2.911013
#> [344]  3.530972  2.000380  3.973761  1.674571  4.017139  3.930790  2.596399
#> [351]  3.558934  3.583859  3.705287  1.724276  3.691388  2.715982  1.534719
#> [358]  1.830842  2.836170  3.154832  4.049917  2.165706  1.712955  2.548428
#> [365]  3.013555  2.949929  3.160149  3.864925  4.117430  4.196927  2.846339
#> [372]  2.754912  4.223555  2.759978  4.023956  2.237192  4.094841  3.827938
#> [379]  4.200300  2.307877  2.877355  2.317801  4.954047  4.645044  2.508653
#> [386]  3.867610  2.540065  2.764332  4.386065  4.323606  2.623247  4.508754
#> [393]  4.037239  3.725443  4.722862  3.320008  2.767657  4.113992  3.224455
#> [400]  6.354493  3.322844  3.615032  6.145182  6.660478  6.266642  4.144332
#> [407]  5.201136  5.980053  4.911398  3.578712  6.088426  4.776076  4.761533
#> [414]  5.658740  3.591142  3.884379  8.322915  5.329451  5.846179  5.704937
#> [421]  7.327436  6.395286  5.805344  7.877417  4.838125  4.955988  6.870236
#> [428]  6.455399  6.609875  5.459103  5.362930  9.629447  5.790393  5.312646
#> [435] 10.247441  5.416451  6.618900  8.524185  6.274485 11.556208  7.614022
#> [442]  6.912946  8.548634 11.687776  9.899655  9.920316  9.255322 11.889397
#> [449] 10.147112 10.368246  8.599226  6.436943  6.319213  6.784087  9.629809
#> [456] 13.360880 13.783801 10.111133  6.860303 12.107349  8.409338 13.622789
#> [463]  9.648879  9.332711 10.103264 11.184777  9.422957 14.688624  7.914835
#> [470] 15.407176 12.180591 10.039561  9.416196  9.027337 11.980354 11.598880
#> [477] 12.194440 15.711830 12.339302 13.137621  9.948194 11.097579 10.780838
#> [484]  9.244142 16.207556  9.940424 15.876770 10.785857 15.383036 13.921960
#> [491] 10.236088  9.145136 11.790159 17.072272 12.703521 16.716861 15.280629
#> [498] 12.307860  9.594780 17.437885 15.921125 16.822916 13.895525 12.456693
#> [505] 16.247553 10.300891 17.742972 15.347464 12.111737 12.990465 18.474317
#> [512] 10.381493 18.172223 17.854850 11.287535 18.791658 17.758246 15.055321
#> [519] 11.302575 17.693201 11.840965 11.005761 16.097997 11.735155 14.746226
#> [526] 14.498016 19.059071 18.109064 13.363791 16.559198 16.545120 15.987518
#> [533] 10.671233 15.553485 12.683458 11.972179 18.254524 18.737358 16.849751
#> [540] 10.504196 15.730930 12.564334 18.170727 17.926342 18.597782 15.407251
#> [547] 11.077004 15.469907 14.914142 11.354488 10.320871 12.114825 11.894026
#> [554] 17.946864 16.414674 11.802685 14.333704 16.251074 13.792988 16.030592
#> [561] 13.503357 18.047597 14.646577 15.657361 13.176252 16.948226 11.960818
#> [568]  9.978166 13.413346 12.955426 11.666725 10.255540  9.740682 16.138954
#> [575] 15.165024 11.906148 13.313431  9.746595 10.023932 15.569274 10.022735
#> [582] 12.947351  9.360842 10.391112 14.804400 10.934543 13.696979 10.634390
#> [589] 11.053969 13.644616 16.379138 14.415822 14.527489 13.600344 15.544264
#> [596]  8.731440 13.356485 14.164019 15.617461 14.512073 11.193246 12.972015
#> [603]  9.799323 13.444077 13.555883  8.268719 11.879626  9.282951  8.716397
#> [610] 12.105585  7.629572  9.838925 13.760212 12.297047 12.901369 11.978322
#> [617]  9.968793  7.235487 10.805203  9.278890 10.244311  6.450023  9.433056
#> [624]  7.981023 11.249895 10.003197  6.819307  9.535028  6.393128 11.460183
#> [631]  7.421632 10.659349  8.810262  7.686590 10.136378  9.591375  8.271658
#> [638]  9.114935  7.942149  7.658505  4.829787  6.513592  8.593585  8.074668
#> [645]  7.273042  9.303983  4.367734  9.380274  9.342039  4.604116  7.994961
#> [652]  7.450401  5.241916  6.336445  6.859104  6.355226  3.540680  4.537342
#> [659]  4.695072  4.837240  4.718394  6.198352  4.426772  4.544258  6.962380
#> [666]  5.329112  6.233746  2.998379  4.934854  5.785694  6.651847  4.659434
#> [673]  5.713188  5.465921  5.695823  4.016158  5.488227  2.875601  2.790834
#> [680]  3.423297  2.906997  2.512233  4.322508  5.485785  5.431475  5.048404
#> [687]  2.938764  3.685515  2.016642  3.763052  3.894234  3.963491  4.896173
#> [694]  2.063216  2.988634  2.074504  4.037288  2.162432  2.905130  4.188923
#> [701]  2.970191  2.088989  3.455701  2.546762  4.328073  3.807376  4.069019
#> [708]  2.911013  3.530972  2.000380  3.973761  1.674571  4.017139  3.930790
#> [715]  2.596399  3.558934  3.583859  3.705287  1.724276  3.691388  2.715982
#> [722]  1.534719  1.830842  2.836170  3.154832  4.049917  2.165706  1.712955
#> [729]  2.548428  3.013555
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 4.78243644  7.463176  6.099033  3.478902
#> 3     2 2002-01-02 0.89203519  7.483435  6.149345  1.903841
#> 6     3 2002-01-03 3.77842804  7.505323  6.203934  3.112616
#> 7     4 2002-01-04 6.68242564  7.528820  6.262814  4.345076
#> 9     5 2002-01-05 1.43904420  7.553900  6.325996  2.184061
#> 11    6 2002-01-06 2.25736128  7.580540  6.393496  2.550313
#> 14    7 2002-01-07 2.77450993  7.608714  6.465324  2.795118
#> 15    8 2002-01-08 2.18509511  7.638396  6.541496  2.571028
#> 17    9 2002-01-09 3.61549381  7.669557  6.622023  3.216344
#> 19   10 2002-01-10 5.93482589  7.702170  6.706920  4.260703
#> 22   11 2002-01-11 6.71106347  7.736204  6.796198  4.646859
#> 23   12 2002-01-12 3.99086073  7.771631  6.889871  3.491499
#> 25   13 2002-01-13 0.43206859  7.808419  6.987950  1.940322
#> 27   14 2002-01-14 1.13643665  7.846537  7.090446  2.286077
#> 30   15 2002-01-15 4.64238116  7.885954  7.197371  3.917852
#> 32   16 2002-01-16 1.59895018  7.926639  7.308732  2.564336
#> 33   17 2002-01-17 6.60380416  7.968559  7.424539  4.932614
#> 35   18 2002-01-18 3.30920624  8.011681  7.544799  3.444381
#> 37   19 2002-01-19 2.54800394  8.055974  7.669519  3.130266
#> 40   20 2002-01-20 1.62173526  8.101405  7.798703  2.730246
#> 41   21 2002-01-21 3.52851562  8.147942  7.932354  3.700665
#> 43   22 2002-01-22 6.85827741  8.195553  8.070474  5.394422
#> 45   23 2002-01-23 4.48812421  8.244205  8.213063  4.288851
#> 48   24 2002-01-24 0.04511114  8.293867  8.360119  2.112766
#> 49   25 2002-01-25 2.53133094  8.344507  8.511639  3.418925
#> 51   26 2002-01-26 0.45773218  8.396093  8.667618  2.403172
#> 53   27 2002-01-27 3.91833553  8.448595  8.828046  4.254171
#> 56   28 2002-01-28 5.09824388  8.501983  8.992915  4.944546
#> 58   29 2002-01-29 6.56144231  8.556225  9.162213  5.803629
#> 59   30 2002-01-30 1.25508458  8.611293  9.335925  3.014331
#> 61   31 2002-01-31 2.70514831  8.667157  9.514035  3.863245
#> 63   32 2002-02-01 0.16188100  8.723789  9.696523  2.514096
#> 66   33 2002-02-02 0.53970992  8.781160  9.883368  2.774569
#> 67   34 2002-02-03 5.55954346  8.839243 10.074546  5.686887
#> 69   35 2002-02-04 5.15968597  8.898011 10.270030  5.545146
#> 71   36 2002-02-05 0.38152680  8.957438 10.469790  2.840419
#> 74   37 2002-02-06 6.30695104  9.017498 10.673795  6.401142
#> 75   38 2002-02-07 6.23941168  9.078165 10.882010  6.460099
#> 77   39 2002-02-08 4.61963813  9.139416 11.094398  5.577504
#> 79   40 2002-02-09 6.23699215  9.201226 11.310918  6.661246
#> 82   41 2002-02-10 1.82650214  9.263572 11.531528  4.019720
#> 83   42 2002-02-11 3.39694515  9.326432 11.756182  5.080009
#> 86   43 2002-02-12 0.38379917  9.389784 11.984831  3.241143
#> 87   44 2002-02-13 6.96036806  9.453606 12.217425  7.551993
#> 89   45 2002-02-14 2.49204538  9.517878 12.453910  4.743868
#> 91   46 2002-02-15 2.92169693  9.582579 12.694229  5.108772
#> 94   47 2002-02-16 6.91357373  9.647690 12.938322  7.870408
#> 95   48 2002-02-17 4.03053527  9.713192 13.186128  6.032356
#> 97   49 2002-02-18 2.91232417  9.779067 13.437583  5.360333
#> 99   50 2002-02-19 5.05836575  9.845297 13.692618  6.940685
#> 102  51 2002-02-20 3.44315246  9.911864 13.951164  5.910947
#> 103  52 2002-02-21 2.31511722  9.978754 14.213148  5.202045
#> 105  53 2002-02-22 0.84424160 10.045948 14.478496  4.227996
#> 107  54 2002-02-23 4.82451909 10.113433 14.747130  7.204273
#> 110  55 2002-02-24 1.41069872 10.181192 15.018971  4.795252
#> 112  56 2002-02-25 0.17832094 10.249211 15.293936  3.956530
#> 113  57 2002-02-26 1.56129228 10.317477 15.571940  5.071197
#> 115  58 2002-02-27 4.60452173 10.385975 15.852898  7.477339
#> 117  59 2002-02-28 6.67846246 10.454693 16.136720  9.188252
#> 120  60 2002-03-01 1.44413148 10.523618 16.423316  5.232695
#> 121  61 2002-03-02 0.93468702 10.592738 16.712592  4.915495
#> 123  62 2002-03-03 5.32155886 10.662040 17.004454  8.494682
#> 125  63 2002-03-04 2.90182687 10.731513 17.298805  6.663520
#> 128  64 2002-03-05 3.58537143 10.801147 17.595546  7.319251
#> 129  65 2002-03-06 0.75489098 10.870930 17.894578  5.094955
#> 131  66 2002-03-07 1.72220244 10.940852 18.195799  5.981052
#> 133  67 2002-03-08 1.92468428 11.010902 18.499104  6.241580
#> 136  68 2002-03-09 0.46493428 11.081072 18.804391  5.095591
#> 138  69 2002-03-10 5.14811540 11.151350 19.111552  9.189393
#> 139  70 2002-03-11 1.59798673 11.221729 19.420480  6.237869
#> 141  71 2002-03-12 4.61477433 11.292198 19.731067  8.964507
#> 143  72 2002-03-13 6.30864709 11.362749 20.043203 10.574837
#> 146  73 2002-03-14 6.49015356 11.433374 20.356778 10.866956
#> 147  74 2002-03-15 2.99205064 11.504062 20.671679  7.856132
#> 149  75 2002-03-16 0.55927976 11.574807 20.987796  5.754000
#> 151  76 2002-03-17 3.83724600 11.645600 21.305014  8.836273
#> 154  77 2002-03-18 3.97850147 11.716433 21.623220  9.077059
#> 155  78 2002-03-19 3.18077514 11.787297 21.942300  8.446115
#> 157  79 2002-03-20 5.43663057 11.858185 22.262140 10.668805
#> 159  80 2002-03-21 0.19544562 11.929088 22.582624  5.830652
#> 162  81 2002-03-22 4.46517439 12.000000 22.903638  9.987107
#> 163  82 2002-03-23 5.52424949 12.070912 23.225065 11.120739
#> 166  83 2002-03-24 3.05922746 12.141815 23.546791  8.853099
#> 167  84 2002-03-25 2.93971475 12.212703 23.868700  8.839888
#> 169  85 2002-03-26 2.89595638 12.283567 24.190677  8.899249
#> 171  86 2002-03-27 2.22253057 12.354400 24.512606  8.333035
#> 174  87 2002-03-28 3.95797902 12.425193 24.834373 10.164022
#> 175  88 2002-03-29 4.08295992 12.495938 25.155863 10.398716
#> 177  89 2002-03-30 3.27746223 12.566626 25.476961  9.691523
#> 179  90 2002-03-31 1.82906558 12.637251 25.797553  8.316306
#> 182  91 2002-04-01 2.07840260 12.707802 26.117528  8.665186
#> 183  92 2002-04-02 0.92712181 12.778271 26.436771  7.568247
#> 185  93 2002-04-03 2.06685201 12.848650 26.755170  8.840730
#> 187  94 2002-04-04 0.96944200 12.918928 27.072615  7.783924
#> 190  95 2002-04-05 4.93852647 12.989098 27.388994 12.053972
#> 192  96 2002-04-06 5.26926060 13.059148 27.704199 12.515260
#> 193  97 2002-04-07 2.61039758 13.129070 28.018120  9.789892
#> 195  98 2002-04-08 4.60823432 13.198853 28.330649 12.028329
#> 197  99 2002-04-09 1.36057654 13.268487 28.641680  8.628907
#> 200 100 2002-04-10 6.04419841 13.337960 28.951106 13.797484
#> 201 101 2002-04-11 0.03327972 13.407262 29.258822  7.351019
#> 203 102 2002-04-12 5.48293341 13.476382 29.564726 13.405460
#> 205 103 2002-04-13 1.16081536 13.545307 29.868714  8.747034
#> 208 104 2002-04-14 0.59028127 13.614025 30.170685  8.196746
#> 209 105 2002-04-15 0.27742785 13.682523 30.470539  7.926546
#> 211 106 2002-04-16 4.64959638 13.750789 30.768176 12.893913
#> 213 107 2002-04-17 6.82271338 13.818808 31.063500 15.434313
#> 216 108 2002-04-18 3.67896070 13.886567 31.356413 11.992722
#> 218 109 2002-04-19 1.16024711 13.954052 31.646820  9.227385
#> 219 110 2002-04-20 2.82844326 14.021246 31.934628 11.204672
#> 221 111 2002-04-21 4.25687742 14.088136 32.219743 12.922702
#> 223 112 2002-04-22 6.75452235 14.154703 32.502075 15.880397
#> 226 113 2002-04-23 2.73765919 14.220933 32.781534 11.350755
#> 227 114 2002-04-24 2.96055935 14.286808 33.058030 11.689704
#> 229 115 2002-04-25 3.96082152 14.352310 33.331477 12.932130
#> 231 116 2002-04-26 4.47769139 14.417421 33.601789 13.618386
#> 234 117 2002-04-27 2.72770323 14.482122 33.868881 11.656817
#> 235 118 2002-04-28 1.30924549 14.546394 34.132670 10.069220
#> 237 119 2002-04-29 5.30976949 14.610216 34.393074 14.847981
#> 239 120 2002-04-30 5.36649800 14.673568 34.650012 14.998700
#> 242 121 2002-05-01 5.25189915 14.736428 34.903405 14.945444
#> 243 122 2002-05-02 0.18834569 14.798774 35.153175  9.011993
#> 245 123 2002-05-03 5.50328728 14.860584 35.399245 15.404473
#> 248 124 2002-05-04 6.48548825 14.921835 35.641540 16.655840
#> 249 125 2002-05-05 4.59490120 14.982502 35.879987 14.471914
#> 251 126 2002-05-06 6.83722122 15.042562 36.114511 17.236102
#> 253 127 2002-05-07 4.82576563 15.101989 36.345041 14.893199
#> 256 128 2002-05-08 3.43419619 15.160757 36.571508 13.284944
#> 257 129 2002-05-09 2.74477950 15.218840 36.793843 12.516420
#> 259 130 2002-05-10 0.25446727 15.276211 37.011977  9.561262
#> 261 131 2002-05-11 5.71841609 15.332843 37.225844 16.248189
#> 264 132 2002-05-12 0.45649496 15.388707 37.435378  9.914091
#> 265 133 2002-05-13 1.49073577 15.443775 37.640517 11.226786
#> 267 134 2002-05-14 6.92310461 15.498017 37.841197 17.912302
#> 269 135 2002-05-15 5.58916423 15.551405 38.037357 16.344639
#> 272 136 2002-05-16 2.24034753 15.603907 38.228936 12.301614
#> 274 137 2002-05-17 5.23028944 15.655493 38.415875 16.021082
#> 275 138 2002-05-18 0.10900643 15.706133 38.598118  9.783472
#> 277 139 2002-05-19 6.61721866 15.755795 38.775606 17.836514
#> 279 140 2002-05-20 3.74474888 15.804447 38.948286 14.351328
#> 282 141 2002-05-21 6.09334570 15.852058 39.116103 17.296912
#> 283 142 2002-05-22 6.79731400 15.898595 39.279005 18.216460
#> 285 143 2002-05-23 0.86095695 15.944026 39.436940 10.924007
#> 287 144 2002-05-24 0.26412330 15.988319 39.589859 10.224472
#> 290 145 2002-05-25 4.55280001 16.031441 39.737713 15.577023
#> 291 146 2002-05-26 2.62093579 16.073361 39.880455 13.221584
#> 293 147 2002-05-27 6.35086363 16.114046 40.018039 17.890459
#> 295 148 2002-05-28 3.28117093 16.153463 40.150421 14.115381
#> 298 149 2002-05-29 2.18226831 16.191581 40.277558 12.783653
#> 300 150 2002-05-30 3.21891807 16.228369 40.399408 14.106490
#> 301 151 2002-05-31 3.05975274 16.263796 40.515933 13.940170
#> 303 152 2002-06-01 3.74320057 16.297830 40.627093 14.822283
#> 305 153 2002-06-02 1.84189827 16.330443 40.732853 12.480327
#> 308 154 2002-06-03 5.60571998 16.361604 40.833176 17.203310
#> 309 155 2002-06-04 5.24092609 16.391286 40.928031 16.775143
#> 311 156 2002-06-05 1.58417672 16.419460 41.017385 12.233059
#> 313 157 2002-06-06 2.52651384 16.446100 41.101209 13.432366
#> 316 158 2002-06-07 4.72907396 16.471180 41.179475 16.206430
#> 317 159 2002-06-08 4.20478996 16.494677 41.252156 15.570998
#> 319 160 2002-06-09 6.37120663 16.516565 41.319229 18.299181
#> 321 161 2002-06-10 3.49110008 16.536824 41.380671 14.713118
#> 324 162 2002-06-11 2.14119344 16.555433 41.436461 13.038703
#> 325 163 2002-06-12 2.64781371 16.572373 41.486582 13.685858
#> 328 164 2002-06-13 0.26950691 16.587625 41.531016 10.720141
#> 329 165 2002-06-14 6.04444964 16.601175 41.569750 17.960163
#> 331 166 2002-06-15 0.02633117 16.613007 41.602770 10.433662
#> 333 167 2002-06-16 2.28377795 16.623108 41.630068 13.267206
#> 336 168 2002-06-17 5.90587288 16.631469 41.651634 17.808204
#> 337 169 2002-06-18 0.26068193 16.638080 41.667464 10.743284
#> 339 170 2002-06-19 0.99630709 16.642934 41.677552 11.666873
#> 341 171 2002-06-20 1.58409059 16.646025 41.681897 12.403767
#> 344 172 2002-06-21 5.53765015 16.647351 41.680500 17.352520
#> 345 173 2002-06-22 0.89391277 16.646909 41.673363 11.537238
#> 347 174 2002-06-23 1.13441664 16.644700 41.660491 11.834804
#> 349 175 2002-06-24 3.82575857 16.640727 41.641891 15.197277
#> 352 176 2002-06-25 5.43057936 16.634994 41.617572 17.197528
#> 354 177 2002-06-26 2.30578815 16.627507 41.587545 13.280423
#> 355 178 2002-06-27 5.72083942 16.618275 41.551823 17.540061
#> 357 179 2002-06-28 5.68815800 16.607306 41.510423 17.486460
#> 359 180 2002-06-29 1.38579983 16.594614 41.463360 12.097122
#> 362 181 2002-06-30 4.45627309 16.580211 41.410655 15.917647
#> 363 182 2002-07-01 5.61185680 16.564113 41.352328 17.343086
#> 365 183 2002-07-02 4.36734597 16.546336 41.288405 15.771064
#> 367 184 2002-07-03 4.91070434 16.526900 41.218909 16.428497
#> 370 185 2002-07-04 1.77451563 16.505823 41.143868 12.497624
#> 371 186 2002-07-05 3.03301012 16.483128 41.063311 14.043795
#> 373 187 2002-07-06 5.65953092 16.458837 40.977269 17.289533
#> 375 188 2002-07-07 4.64397791 16.432973 40.885775 15.998628
#> 378 189 2002-07-08 4.48081895 16.405563 40.788864 15.767507
#> 380 190 2002-07-09 3.53329229 16.376632 40.686570 14.560749
#> 381 191 2002-07-10 4.58175192 16.346207 40.578932 15.831759
#> 383 192 2002-07-11 2.67784522 16.314316 40.465990 13.437558
#> 385 193 2002-07-12 0.29667598 16.280989 40.347784 10.454559
#> 388 194 2002-07-13 6.35859845 16.246255 40.224356 17.927766
#> 389 195 2002-07-14 6.75855481 16.210144 40.095751 18.382572
#> 391 196 2002-07-15 3.95323426 16.172686 39.962013 14.874639
#> 393 197 2002-07-16 3.51517705 16.133915 39.823189 14.294036
#> 396 198 2002-07-17 4.61300417 16.093860 39.679327 15.606501
#> 397 199 2002-07-18 2.64753821 16.052554 39.530476 13.142488
#> 399 200 2002-07-19 5.94161229 16.010029 39.376687 17.150874
#> 401 201 2002-07-20 1.45167534 15.966317 39.218012 11.587376
#> 404 202 2002-07-21 4.31875333 15.921451 39.054503 15.060466
#> 405 203 2002-07-22 5.03190477 15.875462 38.886214 15.884263
#> 408 204 2002-07-23 1.36308281 15.828384 38.713200 11.345220
#> 409 205 2002-07-24 0.78741276 15.780249 38.535519 10.595314
#> 411 206 2002-07-25 1.53436961 15.731088 38.353226 11.458744
#> 413 207 2002-07-26 1.37915684 15.680933 38.166382 11.219985
#> 416 208 2002-07-27 4.14111701 15.629816 37.975044 14.524502
#> 417 209 2002-07-28 5.73725665 15.577768 37.779274 16.401829
#> 419 210 2002-07-29 0.40655363 15.524820 37.579134  9.886832
#> 421 211 2002-07-30 2.23245127 15.471001 37.374685 12.040238
#> 424 212 2002-07-31 3.90763620 15.416342 37.165992 14.001797
#> 425 213 2002-08-01 1.33381630 15.360873 36.953119 10.842638
#> 427 214 2002-08-02 1.20118370 15.304621 36.736132 10.625651
#> 429 215 2002-08-03 2.42597221 15.247616 36.515097 12.033642
#> 432 216 2002-08-04 4.13098980 15.189886 36.290082 14.007184
#> 434 217 2002-08-05 5.03660802 15.131457 36.061156 15.016889
#> 435 218 2002-08-06 2.65735266 15.072356 35.828388 12.115484
#> 437 219 2002-08-07 6.73210440 15.012610 35.591848 16.878188
#> 439 220 2002-08-08 6.12410854 14.952243 35.351609 16.077522
#> 442 221 2002-08-09 6.86008482 14.891281 35.107744 16.863618
#> 443 222 2002-08-10 6.34629986 14.829748 34.860325 16.174213
#> 445 223 2002-08-11 2.45175216 14.767666 34.609429 11.525314
#> 447 224 2002-08-12 5.56140558 14.705060 34.355129 15.085281
#> 450 225 2002-08-13 2.32551712 14.641952 34.097505 11.232155
#> 451 226 2002-08-14 1.51077376 14.578362 33.836632 10.212424
#> 453 227 2002-08-15 6.47005287 14.514313 33.572591 15.875984
#> 455 228 2002-08-16 0.85450179 14.449824 33.305462  9.311138
#> 458 229 2002-08-17 4.18819656 14.384916 33.035326 13.067981
#> 460 230 2002-08-18 4.22708101 14.319607 32.762264 13.026200
#> 461 231 2002-08-19 1.90020732 14.253916 32.486361 10.286989
#> 463 232 2002-08-20 6.51827523 14.187862 32.207701 15.450456
#> 465 233 2002-08-21 6.37336062 14.121461 31.926370 15.186169
#> 468 234 2002-08-22 3.66217831 14.054730 31.642455 12.033080
#> 469 235 2002-08-23 6.99608548 13.987686 31.356043 15.680535
#> 471 236 2002-08-24 1.71596657 13.920345 31.067223  9.681640
#> 473 237 2002-08-25 4.46168030 13.852721 30.776085 12.650197
#> 476 238 2002-08-26 0.37548046 13.784830 30.482721  8.035835
#> 477 239 2002-08-27 6.11772270 13.716686 30.187223 14.278646
#> 479 240 2002-08-28 0.53635900 13.648302 29.889684  8.059732
#> 481 241 2002-08-29 0.34017142 13.579692 29.590199  7.768167
#> 484 242 2002-08-30 5.76594598 13.510869 29.288864 13.571925
#> 485 243 2002-08-31 4.00397607 13.441846 28.985774 11.563497
#> 487 244 2002-09-01 3.48143471 13.372633 28.681027 10.903670
#> 490 245 2002-09-02 6.33935304 13.303244 28.374722 13.854339
#> 491 246 2002-09-03 5.90701229 13.233689 28.066958 13.280747
#> 493 247 2002-09-04 0.74546177 13.163980 27.757836  7.725407
#> 495 248 2002-09-05 5.37367929 13.094126 27.447457 12.493925
#> 498 249 2002-09-06 4.65158140 13.024138 27.135924 11.629789
#> 499 250 2002-09-07 5.49230513 12.954027 26.823338 12.392173
#> 501 251 2002-09-08 5.93292036 12.883802 26.509805 12.731261
#> 503 252 2002-09-09 1.55410129 12.813472 26.195428  8.137433
#> 506 253 2002-09-10 3.67695078 12.743047 25.880314 10.203904
#> 507 254 2002-09-11 2.16604839 12.672536 25.564567  8.575949
#> 509 255 2002-09-12 3.95942812 12.601947 25.248295 10.278477
#> 511 256 2002-09-13 6.70460679 12.531289 24.931604 12.902470
#> 514 257 2002-09-14 0.02961344 12.460572 24.614602  6.182900
#> 516 258 2002-09-15 6.41907870 12.389802 24.297397 12.368514
#> 517 259 2002-09-16 0.13455934 12.318988 23.980097  6.125991
#> 519 260 2002-09-17 2.55531643 12.248139 23.662812  8.384077
#> 521 261 2002-09-18 1.11656655 12.177262 23.345649  6.906726
#> 524 262 2002-09-19 4.01297725 12.106365 23.028718  9.573921
#> 525 263 2002-09-20 5.55728936 12.035456 22.712128 10.921617
#> 527 264 2002-09-21 0.41843246 11.964544 22.395989  5.990621
#> 529 265 2002-09-22 5.73655348 11.893635 22.080409 10.845028
#> 532 266 2002-09-23 3.14142659 11.822738 21.765497  8.333036
#> 533 267 2002-09-24 2.42681313 11.751861 21.451363  7.577743
#> 535 268 2002-09-25 6.81579673 11.681012 21.138114 11.451506
#> 537 269 2002-09-26 6.34638306 11.610198 20.825859 10.898395
#> 540 270 2002-09-27 6.09525700 11.539428 20.514705 10.546726
#> 542 271 2002-09-28 5.91963724 11.468711 20.204761 10.265589
#> 543 272 2002-09-29 0.96488188 11.398053 19.896131  5.816169
#> 545 273 2002-09-30 3.56467578 11.327464 19.588923  7.979481
#> 547 274 2002-10-01 3.28126617 11.256953 19.283241  7.631226
#> 550 275 2002-10-02 2.24911219 11.186528 18.979189  6.652732
#> 551 276 2002-10-03 2.19631028 11.116198 18.676871  6.514282
#> 553 277 2002-10-04 2.96391179 11.045973 18.376389  7.059521
#> 555 278 2002-10-05 4.84625891 10.975862 18.077844  8.510488
#> 558 279 2002-10-06 2.02655584 10.905874 17.781337  6.097420
#> 559 280 2002-10-07 0.11408734 10.836020 17.486965  4.463797
#> 561 281 2002-10-08 6.52858270 10.766311 17.194826  9.512091
#> 563 282 2002-10-09 3.65203997 10.696756 16.905016  7.112073
#> 566 283 2002-10-10 6.86264154 10.627367 16.617629  9.519840
#> 567 284 2002-10-11 1.26686721 10.558154 16.332759  5.063069
#> 570 285 2002-10-12 3.02438697 10.489131 16.050495  6.326586
#> 571 286 2002-10-13 2.59860492 10.420308 15.770928  5.909200
#> 573 287 2002-10-14 3.81286887 10.351698 15.494145  6.727036
#> 575 288 2002-10-15 2.58141072 10.283314 15.220231  5.715418
#> 578 289 2002-10-16 3.60399391 10.215170 14.949271  6.374429
#> 579 290 2002-10-17 1.07790845 10.147279 14.681345  4.450109
#> 581 291 2002-10-18 3.62219272 10.079655 14.416534  6.194473
#> 583 292 2002-10-19 6.33565059 10.012314 14.154915  8.017244
#> 586 293 2002-10-20 6.57856582  9.945270 13.896562  8.070268
#> 587 294 2002-10-21 2.61653916  9.878539 13.641550  5.217013
#> 589 295 2002-10-22 0.22207111  9.812138 13.389948  3.499010
#> 591 296 2002-10-23 6.71359426  9.746084 13.141825  7.811832
#> 594 297 2002-10-24 3.18165044  9.680393 12.897247  5.343778
#> 596 298 2002-10-25 3.80491884  9.615084 12.656277  5.668265
#> 597 299 2002-10-26 2.60333071  9.550176 12.418977  4.797420
#> 599 300 2002-10-27 2.54636642  9.485687 12.185404  4.681894
#> 601 301 2002-10-28 3.36116021  9.421638 11.955615  5.121481
#> 604 302 2002-10-29 2.34835655  9.358048 11.729663  4.404167
#> 605 303 2002-10-30 3.01838006  9.294940 11.507600  4.745353
#> 607 304 2002-10-31 0.14053240  9.232334 11.289473  2.908291
#> 609 305 2002-11-01 0.74228898  9.170252 11.075328  3.217080
#> 612 306 2002-11-02 3.04942520  9.108719 10.865209  4.535035
#> 613 307 2002-11-03 4.57963947  9.047757 10.659156  5.362424
#> 615 308 2002-11-04 0.28417218  8.987390 10.457208  2.779625
#> 617 309 2002-11-05 2.66438421  8.927644 10.259398  4.095767
#> 620 310 2002-11-06 0.45372824  8.868543 10.065762  2.773930
#> 622 311 2002-11-07 6.56347120  8.810114  9.876328  6.147979
#> 623 312 2002-11-08 2.23218383  8.752384  9.691126  3.658580
#> 625 313 2002-11-09 6.70909810  8.695379  9.510180  6.046432
#> 627 314 2002-11-10 6.45136059  8.639127  9.333513  5.818329
#> 630 315 2002-11-11 4.91567277  8.583658  9.161147  4.913481
#> 631 316 2002-11-12 0.80679438  8.528999  8.993100  2.673623
#> 633 317 2002-11-13 6.42173045  8.475180  8.829388  5.552406
#> 635 318 2002-11-14 3.36844820  8.422232  8.670025  3.901283
#> 638 319 2002-11-15 0.57584718  8.370184  8.515023  2.421662
#> 639 320 2002-11-16 0.57582465  8.319067  8.364392  2.380579
#> 641 321 2002-11-17 5.12308561  8.268912  8.218139  4.600349
#> 643 322 2002-11-18 3.08236046  8.219751  8.076271  3.533346
#> 646 323 2002-11-19 2.62861177  8.171616  7.938791  3.261556
#> 647 324 2002-11-20 2.50691055  8.124538  7.805702  3.155691
#> 650 325 2002-11-21 4.61442066  8.078549  7.677004  4.111781
#> 651 326 2002-11-22 2.39720048  8.033683  7.552698  3.015013
#> 653 327 2002-11-23 4.97416863  7.989971  7.432780  4.171839
#> 655 328 2002-11-24 4.54118304  7.947446  7.317248  3.919855
#> 658 329 2002-11-25 1.27529618  7.906140  7.206095  2.382712
#> 659 330 2002-11-26 0.91303388  7.866085  7.099317  2.186846
#> 661 331 2002-11-27 2.24861792  7.827314  6.996907  2.754257
#> 663 332 2002-11-28 0.96690447  7.789856  6.898857  2.152869
#> 666 333 2002-11-29 1.12084267  7.753745  6.805157  2.193149
#> 667 334 2002-11-30 5.56246621  7.719011  6.715800  4.098716
#> 669 335 2002-12-01 5.57758316  7.685684  6.630775  4.063706
#> 671 336 2002-12-02 3.77202692  7.653793  6.550072  3.251557
#> 674 337 2002-12-03 1.65135700  7.623368  6.473680  2.319577
#> 676 338 2002-12-04 4.24168108  7.594437  6.401590  3.388120
#> 677 339 2002-12-05 6.75803277  7.567027  6.333789  4.411768
#> 679 340 2002-12-06 3.05943303  7.541163  6.270268  2.839484
#> 681 341 2002-12-07 3.63436761  7.516872  6.211016  3.054251
#> 684 342 2002-12-08 4.62064936  7.494177  6.156023  3.436801
#> 685 343 2002-12-09 3.42678189  7.473100  6.105278  2.926104
#> 687 344 2002-12-10 1.61689626  7.453664  6.058772  2.171847
#> 689 345 2002-12-11 3.98650858  7.435887  6.016497  3.116898
#> 692 346 2002-12-12 2.29919736  7.419789  5.978443  2.420892
#> 693 347 2002-12-13 3.74399330  7.405386  5.944603  2.988878
#> 695 348 2002-12-14 5.81509372  7.392694  5.914970  3.805101
#> 697 349 2002-12-15 0.90544423  7.381725  5.889539  1.833591
#> 700 350 2002-12-16 6.27752501  7.372493  5.868303  3.965445
#> 702 351 2002-12-17 4.58131275  7.365006  5.851259  3.282667
#> 703 352 2002-12-18 6.84293157  7.359273  5.838404  4.173986
#> 705 353 2002-12-19 3.71000017  7.355300  5.829736  2.927688
#> 707 354 2002-12-20 2.88076558  7.353091  5.825253  2.597411
#> 710 355 2002-12-21 4.64315021  7.352649  5.824956  3.295450
#> 711 356 2002-12-22 4.13353949  7.353975  5.828845  3.095357
#> 713 357 2002-12-23 4.59271725  7.357066  5.836924  3.281108
#> 715 358 2002-12-24 6.92420787  7.361920  5.849194  4.213010
#> 718 359 2002-12-25 3.15566638  7.368531  5.865662  2.722437
#> 719 360 2002-12-26 1.34574356  7.376892  5.886331  2.008495
#> 721 361 2002-12-27 0.30257437  7.386993  5.911209  1.598865
#> 723 362 2002-12-28 3.66404329  7.398825  5.940303  2.955953
#> 726 363 2002-12-29 1.74545936  7.412375  5.973622  2.196737
#> 727 364 2002-12-30 3.23058336  7.427627  6.011174  2.810048
#> 729 365 2002-12-31 4.40325273  7.444567  6.052971  3.303324
#> 2     1 2003-01-01 4.78243644  7.463176  6.099033  3.478902
#> 4     2 2003-01-02 0.89203519  7.483435  6.149345  1.903841
#> 5     3 2003-01-03 3.77842804  7.505323  6.203934  3.112616
#> 8     4 2003-01-04 6.68242564  7.528820  6.262814  4.345076
#> 10    5 2003-01-05 1.43904420  7.553900  6.325996  2.184061
#> 12    6 2003-01-06 2.25736128  7.580540  6.393496  2.550313
#> 13    7 2003-01-07 2.77450993  7.608714  6.465324  2.795118
#> 16    8 2003-01-08 2.18509511  7.638396  6.541496  2.571028
#> 18    9 2003-01-09 3.61549381  7.669557  6.622023  3.216344
#> 20   10 2003-01-10 5.93482589  7.702170  6.706920  4.260703
#> 21   11 2003-01-11 6.71106347  7.736204  6.796198  4.646859
#> 24   12 2003-01-12 3.99086073  7.771631  6.889871  3.491499
#> 26   13 2003-01-13 0.43206859  7.808419  6.987950  1.940322
#> 28   14 2003-01-14 1.13643665  7.846537  7.090446  2.286077
#> 29   15 2003-01-15 4.64238116  7.885954  7.197371  3.917852
#> 31   16 2003-01-16 1.59895018  7.926639  7.308732  2.564336
#> 34   17 2003-01-17 6.60380416  7.968559  7.424539  4.932614
#> 36   18 2003-01-18 3.30920624  8.011681  7.544799  3.444381
#> 38   19 2003-01-19 2.54800394  8.055974  7.669519  3.130266
#> 39   20 2003-01-20 1.62173526  8.101405  7.798703  2.730246
#> 42   21 2003-01-21 3.52851562  8.147942  7.932354  3.700665
#> 44   22 2003-01-22 6.85827741  8.195553  8.070474  5.394422
#> 46   23 2003-01-23 4.48812421  8.244205  8.213063  4.288851
#> 47   24 2003-01-24 0.04511114  8.293867  8.360119  2.112766
#> 50   25 2003-01-25 2.53133094  8.344507  8.511639  3.418925
#> 52   26 2003-01-26 0.45773218  8.396093  8.667618  2.403172
#> 54   27 2003-01-27 3.91833553  8.448595  8.828046  4.254171
#> 55   28 2003-01-28 5.09824388  8.501983  8.992915  4.944546
#> 57   29 2003-01-29 6.56144231  8.556225  9.162213  5.803629
#> 60   30 2003-01-30 1.25508458  8.611293  9.335925  3.014331
#> 62   31 2003-01-31 2.70514831  8.667157  9.514035  3.863245
#> 64   32 2003-02-01 0.16188100  8.723789  9.696523  2.514096
#> 65   33 2003-02-02 0.53970992  8.781160  9.883368  2.774569
#> 68   34 2003-02-03 5.55954346  8.839243 10.074546  5.686887
#> 70   35 2003-02-04 5.15968597  8.898011 10.270030  5.545146
#> 72   36 2003-02-05 0.38152680  8.957438 10.469790  2.840419
#> 73   37 2003-02-06 6.30695104  9.017498 10.673795  6.401142
#> 76   38 2003-02-07 6.23941168  9.078165 10.882010  6.460099
#> 78   39 2003-02-08 4.61963813  9.139416 11.094398  5.577504
#> 80   40 2003-02-09 6.23699215  9.201226 11.310918  6.661246
#> 81   41 2003-02-10 1.82650214  9.263572 11.531528  4.019720
#> 84   42 2003-02-11 3.39694515  9.326432 11.756182  5.080009
#> 85   43 2003-02-12 0.38379917  9.389784 11.984831  3.241143
#> 88   44 2003-02-13 6.96036806  9.453606 12.217425  7.551993
#> 90   45 2003-02-14 2.49204538  9.517878 12.453910  4.743868
#> 92   46 2003-02-15 2.92169693  9.582579 12.694229  5.108772
#> 93   47 2003-02-16 6.91357373  9.647690 12.938322  7.870408
#> 96   48 2003-02-17 4.03053527  9.713192 13.186128  6.032356
#> 98   49 2003-02-18 2.91232417  9.779067 13.437583  5.360333
#> 100  50 2003-02-19 5.05836575  9.845297 13.692618  6.940685
#> 101  51 2003-02-20 3.44315246  9.911864 13.951164  5.910947
#> 104  52 2003-02-21 2.31511722  9.978754 14.213148  5.202045
#> 106  53 2003-02-22 0.84424160 10.045948 14.478496  4.227996
#> 108  54 2003-02-23 4.82451909 10.113433 14.747130  7.204273
#> 109  55 2003-02-24 1.41069872 10.181192 15.018971  4.795252
#> 111  56 2003-02-25 0.17832094 10.249211 15.293936  3.956530
#> 114  57 2003-02-26 1.56129228 10.317477 15.571940  5.071197
#> 116  58 2003-02-27 4.60452173 10.385975 15.852898  7.477339
#> 118  59 2003-02-28 6.67846246 10.454693 16.136720  9.188252
#> 119  60 2003-03-01 1.44413148 10.523618 16.423316  5.232695
#> 122  61 2003-03-02 0.93468702 10.592738 16.712592  4.915495
#> 124  62 2003-03-03 5.32155886 10.662040 17.004454  8.494682
#> 126  63 2003-03-04 2.90182687 10.731513 17.298805  6.663520
#> 127  64 2003-03-05 3.58537143 10.801147 17.595546  7.319251
#> 130  65 2003-03-06 0.75489098 10.870930 17.894578  5.094955
#> 132  66 2003-03-07 1.72220244 10.940852 18.195799  5.981052
#> 134  67 2003-03-08 1.92468428 11.010902 18.499104  6.241580
#> 135  68 2003-03-09 0.46493428 11.081072 18.804391  5.095591
#> 137  69 2003-03-10 5.14811540 11.151350 19.111552  9.189393
#> 140  70 2003-03-11 1.59798673 11.221729 19.420480  6.237869
#> 142  71 2003-03-12 4.61477433 11.292198 19.731067  8.964507
#> 144  72 2003-03-13 6.30864709 11.362749 20.043203 10.574837
#> 145  73 2003-03-14 6.49015356 11.433374 20.356778 10.866956
#> 148  74 2003-03-15 2.99205064 11.504062 20.671679  7.856132
#> 150  75 2003-03-16 0.55927976 11.574807 20.987796  5.754000
#> 152  76 2003-03-17 3.83724600 11.645600 21.305014  8.836273
#> 153  77 2003-03-18 3.97850147 11.716433 21.623220  9.077059
#> 156  78 2003-03-19 3.18077514 11.787297 21.942300  8.446115
#> 158  79 2003-03-20 5.43663057 11.858185 22.262140 10.668805
#> 160  80 2003-03-21 0.19544562 11.929088 22.582624  5.830652
#> 161  81 2003-03-22 4.46517439 12.000000 22.903638  9.987107
#> 164  82 2003-03-23 5.52424949 12.070912 23.225065 11.120739
#> 165  83 2003-03-24 3.05922746 12.141815 23.546791  8.853099
#> 168  84 2003-03-25 2.93971475 12.212703 23.868700  8.839888
#> 170  85 2003-03-26 2.89595638 12.283567 24.190677  8.899249
#> 172  86 2003-03-27 2.22253057 12.354400 24.512606  8.333035
#> 173  87 2003-03-28 3.95797902 12.425193 24.834373 10.164022
#> 176  88 2003-03-29 4.08295992 12.495938 25.155863 10.398716
#> 178  89 2003-03-30 3.27746223 12.566626 25.476961  9.691523
#> 180  90 2003-03-31 1.82906558 12.637251 25.797553  8.316306
#> 181  91 2003-04-01 2.07840260 12.707802 26.117528  8.665186
#> 184  92 2003-04-02 0.92712181 12.778271 26.436771  7.568247
#> 186  93 2003-04-03 2.06685201 12.848650 26.755170  8.840730
#> 188  94 2003-04-04 0.96944200 12.918928 27.072615  7.783924
#> 189  95 2003-04-05 4.93852647 12.989098 27.388994 12.053972
#> 191  96 2003-04-06 5.26926060 13.059148 27.704199 12.515260
#> 194  97 2003-04-07 2.61039758 13.129070 28.018120  9.789892
#> 196  98 2003-04-08 4.60823432 13.198853 28.330649 12.028329
#> 198  99 2003-04-09 1.36057654 13.268487 28.641680  8.628907
#> 199 100 2003-04-10 6.04419841 13.337960 28.951106 13.797484
#> 202 101 2003-04-11 0.03327972 13.407262 29.258822  7.351019
#> 204 102 2003-04-12 5.48293341 13.476382 29.564726 13.405460
#> 206 103 2003-04-13 1.16081536 13.545307 29.868714  8.747034
#> 207 104 2003-04-14 0.59028127 13.614025 30.170685  8.196746
#> 210 105 2003-04-15 0.27742785 13.682523 30.470539  7.926546
#> 212 106 2003-04-16 4.64959638 13.750789 30.768176 12.893913
#> 214 107 2003-04-17 6.82271338 13.818808 31.063500 15.434313
#> 215 108 2003-04-18 3.67896070 13.886567 31.356413 11.992722
#> 217 109 2003-04-19 1.16024711 13.954052 31.646820  9.227385
#> 220 110 2003-04-20 2.82844326 14.021246 31.934628 11.204672
#> 222 111 2003-04-21 4.25687742 14.088136 32.219743 12.922702
#> 224 112 2003-04-22 6.75452235 14.154703 32.502075 15.880397
#> 225 113 2003-04-23 2.73765919 14.220933 32.781534 11.350755
#> 228 114 2003-04-24 2.96055935 14.286808 33.058030 11.689704
#> 230 115 2003-04-25 3.96082152 14.352310 33.331477 12.932130
#> 232 116 2003-04-26 4.47769139 14.417421 33.601789 13.618386
#> 233 117 2003-04-27 2.72770323 14.482122 33.868881 11.656817
#> 236 118 2003-04-28 1.30924549 14.546394 34.132670 10.069220
#> 238 119 2003-04-29 5.30976949 14.610216 34.393074 14.847981
#> 240 120 2003-04-30 5.36649800 14.673568 34.650012 14.998700
#> 241 121 2003-05-01 5.25189915 14.736428 34.903405 14.945444
#> 244 122 2003-05-02 0.18834569 14.798774 35.153175  9.011993
#> 246 123 2003-05-03 5.50328728 14.860584 35.399245 15.404473
#> 247 124 2003-05-04 6.48548825 14.921835 35.641540 16.655840
#> 250 125 2003-05-05 4.59490120 14.982502 35.879987 14.471914
#> 252 126 2003-05-06 6.83722122 15.042562 36.114511 17.236102
#> 254 127 2003-05-07 4.82576563 15.101989 36.345041 14.893199
#> 255 128 2003-05-08 3.43419619 15.160757 36.571508 13.284944
#> 258 129 2003-05-09 2.74477950 15.218840 36.793843 12.516420
#> 260 130 2003-05-10 0.25446727 15.276211 37.011977  9.561262
#> 262 131 2003-05-11 5.71841609 15.332843 37.225844 16.248189
#> 263 132 2003-05-12 0.45649496 15.388707 37.435378  9.914091
#> 266 133 2003-05-13 1.49073577 15.443775 37.640517 11.226786
#> 268 134 2003-05-14 6.92310461 15.498017 37.841197 17.912302
#> 270 135 2003-05-15 5.58916423 15.551405 38.037357 16.344639
#> 271 136 2003-05-16 2.24034753 15.603907 38.228936 12.301614
#> 273 137 2003-05-17 5.23028944 15.655493 38.415875 16.021082
#> 276 138 2003-05-18 0.10900643 15.706133 38.598118  9.783472
#> 278 139 2003-05-19 6.61721866 15.755795 38.775606 17.836514
#> 280 140 2003-05-20 3.74474888 15.804447 38.948286 14.351328
#> 281 141 2003-05-21 6.09334570 15.852058 39.116103 17.296912
#> 284 142 2003-05-22 6.79731400 15.898595 39.279005 18.216460
#> 286 143 2003-05-23 0.86095695 15.944026 39.436940 10.924007
#> 288 144 2003-05-24 0.26412330 15.988319 39.589859 10.224472
#> 289 145 2003-05-25 4.55280001 16.031441 39.737713 15.577023
#> 292 146 2003-05-26 2.62093579 16.073361 39.880455 13.221584
#> 294 147 2003-05-27 6.35086363 16.114046 40.018039 17.890459
#> 296 148 2003-05-28 3.28117093 16.153463 40.150421 14.115381
#> 297 149 2003-05-29 2.18226831 16.191581 40.277558 12.783653
#> 299 150 2003-05-30 3.21891807 16.228369 40.399408 14.106490
#> 302 151 2003-05-31 3.05975274 16.263796 40.515933 13.940170
#> 304 152 2003-06-01 3.74320057 16.297830 40.627093 14.822283
#> 306 153 2003-06-02 1.84189827 16.330443 40.732853 12.480327
#> 307 154 2003-06-03 5.60571998 16.361604 40.833176 17.203310
#> 310 155 2003-06-04 5.24092609 16.391286 40.928031 16.775143
#> 312 156 2003-06-05 1.58417672 16.419460 41.017385 12.233059
#> 314 157 2003-06-06 2.52651384 16.446100 41.101209 13.432366
#> 315 158 2003-06-07 4.72907396 16.471180 41.179475 16.206430
#> 318 159 2003-06-08 4.20478996 16.494677 41.252156 15.570998
#> 320 160 2003-06-09 6.37120663 16.516565 41.319229 18.299181
#> 322 161 2003-06-10 3.49110008 16.536824 41.380671 14.713118
#> 323 162 2003-06-11 2.14119344 16.555433 41.436461 13.038703
#> 326 163 2003-06-12 2.64781371 16.572373 41.486582 13.685858
#> 327 164 2003-06-13 0.26950691 16.587625 41.531016 10.720141
#> 330 165 2003-06-14 6.04444964 16.601175 41.569750 17.960163
#> 332 166 2003-06-15 0.02633117 16.613007 41.602770 10.433662
#> 334 167 2003-06-16 2.28377795 16.623108 41.630068 13.267206
#> 335 168 2003-06-17 5.90587288 16.631469 41.651634 17.808204
#> 338 169 2003-06-18 0.26068193 16.638080 41.667464 10.743284
#> 340 170 2003-06-19 0.99630709 16.642934 41.677552 11.666873
#> 342 171 2003-06-20 1.58409059 16.646025 41.681897 12.403767
#> 343 172 2003-06-21 5.53765015 16.647351 41.680500 17.352520
#> 346 173 2003-06-22 0.89391277 16.646909 41.673363 11.537238
#> 348 174 2003-06-23 1.13441664 16.644700 41.660491 11.834804
#> 350 175 2003-06-24 3.82575857 16.640727 41.641891 15.197277
#> 351 176 2003-06-25 5.43057936 16.634994 41.617572 17.197528
#> 353 177 2003-06-26 2.30578815 16.627507 41.587545 13.280423
#> 356 178 2003-06-27 5.72083942 16.618275 41.551823 17.540061
#> 358 179 2003-06-28 5.68815800 16.607306 41.510423 17.486460
#> 360 180 2003-06-29 1.38579983 16.594614 41.463360 12.097122
#> 361 181 2003-06-30 4.45627309 16.580211 41.410655 15.917647
#> 364 182 2003-07-01 5.61185680 16.564113 41.352328 17.343086
#> 366 183 2003-07-02 4.36734597 16.546336 41.288405 15.771064
#> 368 184 2003-07-03 4.91070434 16.526900 41.218909 16.428497
#> 369 185 2003-07-04 1.77451563 16.505823 41.143868 12.497624
#> 372 186 2003-07-05 3.03301012 16.483128 41.063311 14.043795
#> 374 187 2003-07-06 5.65953092 16.458837 40.977269 17.289533
#> 376 188 2003-07-07 4.64397791 16.432973 40.885775 15.998628
#> 377 189 2003-07-08 4.48081895 16.405563 40.788864 15.767507
#> 379 190 2003-07-09 3.53329229 16.376632 40.686570 14.560749
#> 382 191 2003-07-10 4.58175192 16.346207 40.578932 15.831759
#> 384 192 2003-07-11 2.67784522 16.314316 40.465990 13.437558
#> 386 193 2003-07-12 0.29667598 16.280989 40.347784 10.454559
#> 387 194 2003-07-13 6.35859845 16.246255 40.224356 17.927766
#> 390 195 2003-07-14 6.75855481 16.210144 40.095751 18.382572
#> 392 196 2003-07-15 3.95323426 16.172686 39.962013 14.874639
#> 394 197 2003-07-16 3.51517705 16.133915 39.823189 14.294036
#> 395 198 2003-07-17 4.61300417 16.093860 39.679327 15.606501
#> 398 199 2003-07-18 2.64753821 16.052554 39.530476 13.142488
#> 400 200 2003-07-19 5.94161229 16.010029 39.376687 17.150874
#> 402 201 2003-07-20 1.45167534 15.966317 39.218012 11.587376
#> 403 202 2003-07-21 4.31875333 15.921451 39.054503 15.060466
#> 406 203 2003-07-22 5.03190477 15.875462 38.886214 15.884263
#> 407 204 2003-07-23 1.36308281 15.828384 38.713200 11.345220
#> 410 205 2003-07-24 0.78741276 15.780249 38.535519 10.595314
#> 412 206 2003-07-25 1.53436961 15.731088 38.353226 11.458744
#> 414 207 2003-07-26 1.37915684 15.680933 38.166382 11.219985
#> 415 208 2003-07-27 4.14111701 15.629816 37.975044 14.524502
#> 418 209 2003-07-28 5.73725665 15.577768 37.779274 16.401829
#> 420 210 2003-07-29 0.40655363 15.524820 37.579134  9.886832
#> 422 211 2003-07-30 2.23245127 15.471001 37.374685 12.040238
#> 423 212 2003-07-31 3.90763620 15.416342 37.165992 14.001797
#> 426 213 2003-08-01 1.33381630 15.360873 36.953119 10.842638
#> 428 214 2003-08-02 1.20118370 15.304621 36.736132 10.625651
#> 430 215 2003-08-03 2.42597221 15.247616 36.515097 12.033642
#> 431 216 2003-08-04 4.13098980 15.189886 36.290082 14.007184
#> 433 217 2003-08-05 5.03660802 15.131457 36.061156 15.016889
#> 436 218 2003-08-06 2.65735266 15.072356 35.828388 12.115484
#> 438 219 2003-08-07 6.73210440 15.012610 35.591848 16.878188
#> 440 220 2003-08-08 6.12410854 14.952243 35.351609 16.077522
#> 441 221 2003-08-09 6.86008482 14.891281 35.107744 16.863618
#> 444 222 2003-08-10 6.34629986 14.829748 34.860325 16.174213
#> 446 223 2003-08-11 2.45175216 14.767666 34.609429 11.525314
#> 448 224 2003-08-12 5.56140558 14.705060 34.355129 15.085281
#> 449 225 2003-08-13 2.32551712 14.641952 34.097505 11.232155
#> 452 226 2003-08-14 1.51077376 14.578362 33.836632 10.212424
#> 454 227 2003-08-15 6.47005287 14.514313 33.572591 15.875984
#> 456 228 2003-08-16 0.85450179 14.449824 33.305462  9.311138
#> 457 229 2003-08-17 4.18819656 14.384916 33.035326 13.067981
#> 459 230 2003-08-18 4.22708101 14.319607 32.762264 13.026200
#> 462 231 2003-08-19 1.90020732 14.253916 32.486361 10.286989
#> 464 232 2003-08-20 6.51827523 14.187862 32.207701 15.450456
#> 466 233 2003-08-21 6.37336062 14.121461 31.926370 15.186169
#> 467 234 2003-08-22 3.66217831 14.054730 31.642455 12.033080
#> 470 235 2003-08-23 6.99608548 13.987686 31.356043 15.680535
#> 472 236 2003-08-24 1.71596657 13.920345 31.067223  9.681640
#> 474 237 2003-08-25 4.46168030 13.852721 30.776085 12.650197
#> 475 238 2003-08-26 0.37548046 13.784830 30.482721  8.035835
#> 478 239 2003-08-27 6.11772270 13.716686 30.187223 14.278646
#> 480 240 2003-08-28 0.53635900 13.648302 29.889684  8.059732
#> 482 241 2003-08-29 0.34017142 13.579692 29.590199  7.768167
#> 483 242 2003-08-30 5.76594598 13.510869 29.288864 13.571925
#> 486 243 2003-08-31 4.00397607 13.441846 28.985774 11.563497
#> 488 244 2003-09-01 3.48143471 13.372633 28.681027 10.903670
#> 489 245 2003-09-02 6.33935304 13.303244 28.374722 13.854339
#> 492 246 2003-09-03 5.90701229 13.233689 28.066958 13.280747
#> 494 247 2003-09-04 0.74546177 13.163980 27.757836  7.725407
#> 496 248 2003-09-05 5.37367929 13.094126 27.447457 12.493925
#> 497 249 2003-09-06 4.65158140 13.024138 27.135924 11.629789
#> 500 250 2003-09-07 5.49230513 12.954027 26.823338 12.392173
#> 502 251 2003-09-08 5.93292036 12.883802 26.509805 12.731261
#> 504 252 2003-09-09 1.55410129 12.813472 26.195428  8.137433
#> 505 253 2003-09-10 3.67695078 12.743047 25.880314 10.203904
#> 508 254 2003-09-11 2.16604839 12.672536 25.564567  8.575949
#> 510 255 2003-09-12 3.95942812 12.601947 25.248295 10.278477
#> 512 256 2003-09-13 6.70460679 12.531289 24.931604 12.902470
#> 513 257 2003-09-14 0.02961344 12.460572 24.614602  6.182900
#> 515 258 2003-09-15 6.41907870 12.389802 24.297397 12.368514
#> 518 259 2003-09-16 0.13455934 12.318988 23.980097  6.125991
#> 520 260 2003-09-17 2.55531643 12.248139 23.662812  8.384077
#> 522 261 2003-09-18 1.11656655 12.177262 23.345649  6.906726
#> 523 262 2003-09-19 4.01297725 12.106365 23.028718  9.573921
#> 526 263 2003-09-20 5.55728936 12.035456 22.712128 10.921617
#> 528 264 2003-09-21 0.41843246 11.964544 22.395989  5.990621
#> 530 265 2003-09-22 5.73655348 11.893635 22.080409 10.845028
#> 531 266 2003-09-23 3.14142659 11.822738 21.765497  8.333036
#> 534 267 2003-09-24 2.42681313 11.751861 21.451363  7.577743
#> 536 268 2003-09-25 6.81579673 11.681012 21.138114 11.451506
#> 538 269 2003-09-26 6.34638306 11.610198 20.825859 10.898395
#> 539 270 2003-09-27 6.09525700 11.539428 20.514705 10.546726
#> 541 271 2003-09-28 5.91963724 11.468711 20.204761 10.265589
#> 544 272 2003-09-29 0.96488188 11.398053 19.896131  5.816169
#> 546 273 2003-09-30 3.56467578 11.327464 19.588923  7.979481
#> 548 274 2003-10-01 3.28126617 11.256953 19.283241  7.631226
#> 549 275 2003-10-02 2.24911219 11.186528 18.979189  6.652732
#> 552 276 2003-10-03 2.19631028 11.116198 18.676871  6.514282
#> 554 277 2003-10-04 2.96391179 11.045973 18.376389  7.059521
#> 556 278 2003-10-05 4.84625891 10.975862 18.077844  8.510488
#> 557 279 2003-10-06 2.02655584 10.905874 17.781337  6.097420
#> 560 280 2003-10-07 0.11408734 10.836020 17.486965  4.463797
#> 562 281 2003-10-08 6.52858270 10.766311 17.194826  9.512091
#> 564 282 2003-10-09 3.65203997 10.696756 16.905016  7.112073
#> 565 283 2003-10-10 6.86264154 10.627367 16.617629  9.519840
#> 568 284 2003-10-11 1.26686721 10.558154 16.332759  5.063069
#> 569 285 2003-10-12 3.02438697 10.489131 16.050495  6.326586
#> 572 286 2003-10-13 2.59860492 10.420308 15.770928  5.909200
#> 574 287 2003-10-14 3.81286887 10.351698 15.494145  6.727036
#> 576 288 2003-10-15 2.58141072 10.283314 15.220231  5.715418
#> 577 289 2003-10-16 3.60399391 10.215170 14.949271  6.374429
#> 580 290 2003-10-17 1.07790845 10.147279 14.681345  4.450109
#> 582 291 2003-10-18 3.62219272 10.079655 14.416534  6.194473
#> 584 292 2003-10-19 6.33565059 10.012314 14.154915  8.017244
#> 585 293 2003-10-20 6.57856582  9.945270 13.896562  8.070268
#> 588 294 2003-10-21 2.61653916  9.878539 13.641550  5.217013
#> 590 295 2003-10-22 0.22207111  9.812138 13.389948  3.499010
#> 592 296 2003-10-23 6.71359426  9.746084 13.141825  7.811832
#> 593 297 2003-10-24 3.18165044  9.680393 12.897247  5.343778
#> 595 298 2003-10-25 3.80491884  9.615084 12.656277  5.668265
#> 598 299 2003-10-26 2.60333071  9.550176 12.418977  4.797420
#> 600 300 2003-10-27 2.54636642  9.485687 12.185404  4.681894
#> 602 301 2003-10-28 3.36116021  9.421638 11.955615  5.121481
#> 603 302 2003-10-29 2.34835655  9.358048 11.729663  4.404167
#> 606 303 2003-10-30 3.01838006  9.294940 11.507600  4.745353
#> 608 304 2003-10-31 0.14053240  9.232334 11.289473  2.908291
#> 610 305 2003-11-01 0.74228898  9.170252 11.075328  3.217080
#> 611 306 2003-11-02 3.04942520  9.108719 10.865209  4.535035
#> 614 307 2003-11-03 4.57963947  9.047757 10.659156  5.362424
#> 616 308 2003-11-04 0.28417218  8.987390 10.457208  2.779625
#> 618 309 2003-11-05 2.66438421  8.927644 10.259398  4.095767
#> 619 310 2003-11-06 0.45372824  8.868543 10.065762  2.773930
#> 621 311 2003-11-07 6.56347120  8.810114  9.876328  6.147979
#> 624 312 2003-11-08 2.23218383  8.752384  9.691126  3.658580
#> 626 313 2003-11-09 6.70909810  8.695379  9.510180  6.046432
#> 628 314 2003-11-10 6.45136059  8.639127  9.333513  5.818329
#> 629 315 2003-11-11 4.91567277  8.583658  9.161147  4.913481
#> 632 316 2003-11-12 0.80679438  8.528999  8.993100  2.673623
#> 634 317 2003-11-13 6.42173045  8.475180  8.829388  5.552406
#> 636 318 2003-11-14 3.36844820  8.422232  8.670025  3.901283
#> 637 319 2003-11-15 0.57584718  8.370184  8.515023  2.421662
#> 640 320 2003-11-16 0.57582465  8.319067  8.364392  2.380579
#> 642 321 2003-11-17 5.12308561  8.268912  8.218139  4.600349
#> 644 322 2003-11-18 3.08236046  8.219751  8.076271  3.533346
#> 645 323 2003-11-19 2.62861177  8.171616  7.938791  3.261556
#> 648 324 2003-11-20 2.50691055  8.124538  7.805702  3.155691
#> 649 325 2003-11-21 4.61442066  8.078549  7.677004  4.111781
#> 652 326 2003-11-22 2.39720048  8.033683  7.552698  3.015013
#> 654 327 2003-11-23 4.97416863  7.989971  7.432780  4.171839
#> 656 328 2003-11-24 4.54118304  7.947446  7.317248  3.919855
#> 657 329 2003-11-25 1.27529618  7.906140  7.206095  2.382712
#> 660 330 2003-11-26 0.91303388  7.866085  7.099317  2.186846
#> 662 331 2003-11-27 2.24861792  7.827314  6.996907  2.754257
#> 664 332 2003-11-28 0.96690447  7.789856  6.898857  2.152869
#> 665 333 2003-11-29 1.12084267  7.753745  6.805157  2.193149
#> 668 334 2003-11-30 5.56246621  7.719011  6.715800  4.098716
#> 670 335 2003-12-01 5.57758316  7.685684  6.630775  4.063706
#> 672 336 2003-12-02 3.77202692  7.653793  6.550072  3.251557
#> 673 337 2003-12-03 1.65135700  7.623368  6.473680  2.319577
#> 675 338 2003-12-04 4.24168108  7.594437  6.401590  3.388120
#> 678 339 2003-12-05 6.75803277  7.567027  6.333789  4.411768
#> 680 340 2003-12-06 3.05943303  7.541163  6.270268  2.839484
#> 682 341 2003-12-07 3.63436761  7.516872  6.211016  3.054251
#> 683 342 2003-12-08 4.62064936  7.494177  6.156023  3.436801
#> 686 343 2003-12-09 3.42678189  7.473100  6.105278  2.926104
#> 688 344 2003-12-10 1.61689626  7.453664  6.058772  2.171847
#> 690 345 2003-12-11 3.98650858  7.435887  6.016497  3.116898
#> 691 346 2003-12-12 2.29919736  7.419789  5.978443  2.420892
#> 694 347 2003-12-13 3.74399330  7.405386  5.944603  2.988878
#> 696 348 2003-12-14 5.81509372  7.392694  5.914970  3.805101
#> 698 349 2003-12-15 0.90544423  7.381725  5.889539  1.833591
#> 699 350 2003-12-16 6.27752501  7.372493  5.868303  3.965445
#> 701 351 2003-12-17 4.58131275  7.365006  5.851259  3.282667
#> 704 352 2003-12-18 6.84293157  7.359273  5.838404  4.173986
#> 706 353 2003-12-19 3.71000017  7.355300  5.829736  2.927688
#> 708 354 2003-12-20 2.88076558  7.353091  5.825253  2.597411
#> 709 355 2003-12-21 4.64315021  7.352649  5.824956  3.295450
#> 712 356 2003-12-22 4.13353949  7.353975  5.828845  3.095357
#> 714 357 2003-12-23 4.59271725  7.357066  5.836924  3.281108
#> 716 358 2003-12-24 6.92420787  7.361920  5.849194  4.213010
#> 717 359 2003-12-25 3.15566638  7.368531  5.865662  2.722437
#> 720 360 2003-12-26 1.34574356  7.376892  5.886331  2.008495
#> 722 361 2003-12-27 0.30257437  7.386993  5.911209  1.598865
#> 724 362 2003-12-28 3.66404329  7.398825  5.940303  2.955953
#> 725 363 2003-12-29 1.74545936  7.412375  5.973622  2.196737
#> 728 364 2003-12-30 3.23058336  7.427627  6.011174  2.810048
#> 730 365 2003-12-31 4.40325273  7.444567  6.052971  3.303324