Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.346946  2.346946  1.837396  1.837396  2.973457  2.973457  3.275645
#>   [8]  3.275645  2.234568  2.234568  2.844481  2.844481  4.532074  4.532074
#>  [15]  4.037743  4.037743  2.791389  2.791389  2.692472  2.692472  2.337299
#>  [22]  2.337299  3.500585  3.500585  2.279516  2.279516  2.838504  2.838504
#>  [29]  2.961206  2.961206  2.008156  2.008156  2.597079  2.597079  2.514202
#>  [36]  2.514202  2.424443  2.424443  5.111021  5.111021  2.260013  2.260013
#>  [43]  3.901514  3.901514  5.064235  5.064235  5.433639  5.433639  2.459730
#>  [50]  2.459730  5.618244  5.618244  5.626891  5.626891  5.781860  5.781860
#>  [57]  5.632048  5.632048  4.133331  4.133331  4.909147  4.909147  5.331357
#>  [64]  5.331357  3.342894  3.342894  4.151473  4.151473  6.231259  6.231259
#>  [71]  4.452773  4.452773  3.105428  3.105428  5.818025  5.818025  4.766939
#>  [78]  4.766939  4.404053  4.404053  5.219612  5.219612  3.812890  3.812890
#>  [85]  5.949779  5.949779  5.158078  5.158078  3.161494  3.161494  4.707465
#>  [92]  4.707465  7.240449  7.240449  3.650491  3.650491  7.597154  7.597154
#>  [99]  6.457309  6.457309  7.221243  7.221243  7.475605  7.475605  4.311501
#> [106]  4.311501  5.810778  5.810778  4.968299  4.968299  4.617998  4.617998
#> [113]  4.435010  4.435010  7.294009  7.294009  5.435076  5.435076  6.048899
#> [120]  6.048899  5.789623  5.789623  9.255392  9.255392  8.153216  8.153216
#> [127]  9.679551  9.679551  5.311612  5.311612  4.714718  4.714718  6.579404
#> [134]  6.579404  5.589936  5.589936  6.754962  6.754962  5.848506  5.848506
#> [141] 10.657903 10.657903  8.695601  8.695601  6.555423  6.555423  9.686174
#> [148]  9.686174 10.513248 10.513248  6.838679  6.838679  8.189341  8.189341
#> [155]  8.246272  8.246272 11.940552 11.940552 10.239276 10.239276  8.429567
#> [162]  8.429567  8.139600  8.139600 11.438845 11.438845 12.608803 12.608803
#> [169] 10.083569 10.083569  8.990788  8.990788  6.801817  6.801817  8.788780
#> [176]  8.788780  7.701116  7.701116  8.346536  8.346536 13.297489 13.297489
#> [183] 13.087202 13.087202 11.490375 11.490375 10.464797 10.464797 12.878928
#> [190] 12.878928 12.668756 12.668756  7.971107  7.971107 13.581357 13.581357
#> [197] 12.185390 12.185390  8.391971  8.391971 11.251069 11.251069  7.816511
#> [204]  7.816511  9.085574  9.085574 14.152422 14.152422 12.226745 12.226745
#> [211]  9.222135  9.222135 14.317373 14.317373 12.946920 12.946920 11.251607
#> [218] 11.251607 13.161267 13.161267 15.177673 15.177673 12.440201 12.440201
#> [225] 13.902959 13.902959 13.775126 13.775126 13.076714 13.076714 13.677312
#> [232] 13.677312  9.718531  9.718531  9.699829  9.699829 13.474092 13.474092
#> [239] 15.677575 15.677575 10.047223 10.047223 14.716696 14.716696 12.776065
#> [246] 12.776065 12.104267 12.104267 12.242246 12.242246 10.602661 10.602661
#> [253] 15.974199 15.974199 14.479291 14.479291 10.701124 10.701124 11.684467
#> [260] 11.684467 11.741968 11.741968 12.609916 12.609916 15.042152 15.042152
#> [267]  9.718668  9.718668 17.228563 17.228563 14.562447 14.562447 12.630958
#> [274] 12.630958 12.534280 12.534280 11.560176 11.560176 13.428970 13.428970
#> [281] 14.302755 14.302755 15.587481 15.587481 12.415037 12.415037 14.594777
#> [288] 14.594777 10.605336 10.605336 10.030694 10.030694 15.068397 15.068397
#> [295] 14.430594 14.430594 12.546256 12.546256 11.538891 11.538891 17.213231
#> [302] 17.213231 11.112572 11.112572 15.258440 15.258440 18.303502 18.303502
#> [309] 12.574016 12.574016 17.681678 17.681678 13.330149 13.330149 18.685908
#> [316] 18.685908 17.441420 17.441420 15.079900 15.079900 13.738364 13.738364
#> [323] 16.909459 16.909459 15.611933 15.611933 15.279237 15.279237 14.886080
#> [330] 14.886080 12.135573 12.135573 16.148636 16.148636 13.271632 13.271632
#> [337] 16.082579 16.082579 13.415167 13.415167 14.568844 14.568844 17.171904
#> [344] 17.171904 13.218837 13.218837 17.887055 17.887055 14.318309 14.318309
#> [351] 11.859878 11.859878 12.722514 12.722514 14.088409 14.088409 11.928915
#> [358] 11.928915 13.923778 13.923778 10.755177 10.755177 16.563812 16.563812
#> [365] 10.426913 10.426913 18.271917 18.271917 15.463421 15.463421 11.165411
#> [372] 11.165411 17.200351 17.200351 11.778796 11.778796 12.449159 12.449159
#> [379] 11.622534 11.622534 15.788056 15.788056 13.910934 13.910934 10.475531
#> [386] 10.475531 14.614875 14.614875 11.620439 11.620439 15.280656 15.280656
#> [393] 11.182794 11.182794 17.856639 17.856639 17.200671 17.200671 17.781588
#> [400] 17.781588 11.022707 11.022707 14.625043 14.625043 11.683498 11.683498
#> [407] 11.185325 11.185325 11.024089 11.024089 16.016953 16.016953 11.322459
#> [414] 11.322459 10.355316 10.355316 16.421112 16.421112 11.017810 11.017810
#> [421] 14.041723 14.041723 12.027272 12.027272 15.558873 15.558873 10.582380
#> [428] 10.582380 11.782975 11.782975  9.561601  9.561601 14.650734 14.650734
#> [435]  9.672454  9.672454 12.212428 12.212428  9.531153  9.531153 10.268537
#> [442] 10.268537 14.604614 14.604614 12.035648 12.035648  8.966741  8.966741
#> [449] 11.023125 11.023125 14.142967 14.142967 11.030372 11.030372 15.402019
#> [456] 15.402019 13.437249 13.437249 10.760035 10.760035 11.512016 11.512016
#> [463] 13.440074 13.440074  9.952168  9.952168 12.959059 12.959059  9.923775
#> [470]  9.923775 12.265648 12.265648 12.874337 12.874337  8.465921  8.465921
#> [477]  9.022791  9.022791 10.747286 10.747286 12.459384 12.459384  7.716918
#> [484]  7.716918  8.358285  8.358285 10.415664 10.415664 12.806668 12.806668
#> [491]  7.556276  7.556276 13.053753 13.053753 10.660701 10.660701 11.584451
#> [498] 11.584451  8.128325  8.128325 10.925627 10.925627 10.787675 10.787675
#> [505] 13.537335 13.537335 11.385602 11.385602 11.342789 11.342789  9.034555
#> [512]  9.034555 10.738199 10.738199 11.377915 11.377915  9.312533  9.312533
#> [519]  8.449822  8.449822  7.457584  7.457584  5.911883  5.911883  6.191586
#> [526]  6.191586  9.368063  9.368063  9.300147  9.300147  6.129425  6.129425
#> [533] 10.902065 10.902065 11.135532 11.135532  8.500235  8.500235  8.412993
#> [540]  8.412993  6.638725  6.638725  7.757814  7.757814 10.716149 10.716149
#> [547]  6.515776  6.515776  5.803769  5.803769  5.093225  5.093225  5.749979
#> [554]  5.749979  9.887295  9.887295  6.092406  6.092406  4.914806  4.914806
#> [561]  8.471345  8.471345  8.304830  8.304830  5.177898  5.177898  9.112121
#> [568]  9.112121  7.934998  7.934998  5.206499  5.206499  4.059111  4.059111
#> [575]  7.266354  7.266354  6.177174  6.177174  6.183869  6.183869  6.736876
#> [582]  6.736876  6.963498  6.963498  5.129888  5.129888  7.165346  7.165346
#> [589]  5.228137  5.228137  6.404486  6.404486  4.650108  4.650108  5.359444
#> [596]  5.359444  3.637507  3.637507  5.232352  5.232352  5.038700  5.038700
#> [603]  6.863031  6.863031  4.698190  4.698190  5.458443  5.458443  4.500669
#> [610]  4.500669  6.299997  6.299997  5.778856  5.778856  3.355877  3.355877
#> [617]  3.437320  3.437320  4.641705  4.641705  3.497873  3.497873  5.504598
#> [624]  5.504598  4.258909  4.258909  4.026601  4.026601  5.268317  5.268317
#> [631]  3.536459  3.536459  3.532111  3.532111  4.828451  4.828451  3.641636
#> [638]  3.641636  2.639315  2.639315  4.975173  4.975173  5.300898  5.300898
#> [645]  3.454879  3.454879  4.508090  4.508090  2.534479  2.534479  2.770346
#> [652]  2.770346  1.909658  1.909658  2.610389  2.610389  3.164688  3.164688
#> [659]  4.384986  4.384986  3.497150  3.497150  3.404662  3.404662  2.790993
#> [666]  2.790993  2.693251  2.693251  2.543201  2.543201  2.467933  2.467933
#> [673]  3.597230  3.597230  3.137921  3.137921  4.468282  4.468282  2.208167
#> [680]  2.208167  2.596249  2.596249  3.770636  3.770636  1.662953  1.662953
#> [687]  3.094349  3.094349  1.756364  1.756364  3.878019  3.878019  3.900429
#> [694]  3.900429  1.864743  1.864743  3.941486  3.941486  3.088002  3.088002
#> [701]  3.594252  3.594252  2.557269  2.557269  2.336936  2.336936  1.929698
#> [708]  1.929698  3.959845  3.959845  2.423512  2.423512  3.127953  3.127953
#> [715]  1.870091  1.870091  1.814216  1.814216  4.256370  4.256370  2.651239
#> [722]  2.651239  4.185820  4.185820  2.933896  2.933896  3.780344  3.780344
#> [729]  3.990397  3.990397