Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.723961  4.344234  3.673684  2.994827  4.474746  1.776897  2.423950
#>   [8]  1.785985  2.116210  2.748591  2.252917  3.829839  4.560958  4.004966
#>  [15]  3.105198  2.284763  3.559454  4.919463  3.825419  4.610123  2.718422
#>  [22]  4.128664  4.082217  2.566785  5.464735  5.382149  4.908832  2.889067
#>  [29]  4.049353  5.330518  3.746462  3.233243  2.727654  3.426391  4.258919
#>  [36]  5.909595  4.017716  5.720791  3.952353  6.435109  6.941043  5.679224
#>  [43]  3.892993  6.988197  7.516035  7.526976  3.299380  7.051747  6.063918
#>  [50]  6.689986  4.415604  6.011061  8.338151  5.117122  5.485327  8.722906
#>  [57]  6.734626  4.022107  7.430476  8.430022  6.975501  6.629643  9.649704
#>  [64]  9.755066 10.014722  4.937426  8.150648  8.082677  9.016321  7.455100
#>  [71]  9.338442  5.678021  9.053815  8.330326 11.495546  8.482281 10.688340
#>  [78]  7.061287  9.544915  9.584767  6.254848  8.096770  8.165403  9.448070
#>  [85] 12.140424 12.822204 12.018859 12.613030 12.849419 11.145606 12.569182
#>  [92] 11.143073  7.089317 12.076676 12.288486  9.469859  9.825388 13.465333
#>  [99] 12.819679 12.787259 11.191474  7.680384 14.046501 13.623822 10.621513
#> [106] 10.353200 10.047603 15.526153  8.295151 15.855124 10.017203  9.882153
#> [113] 13.305639 15.591076  8.473935 14.749894  9.262995 15.520104 16.789827
#> [120] 11.930870 16.618550 14.370990 11.311946 11.141435 13.281843 15.122681
#> [127] 13.465903 12.351093 12.031648 10.269853 15.831026 15.946889 11.158083
#> [134] 11.644175 15.069166 13.000614 11.902250 18.178357 18.162181 16.135650
#> [141] 17.959012 16.171962 17.389568 13.289805 10.181499 18.254193 10.666292
#> [148] 13.987102 18.643417 10.540604 15.985566 14.087701 16.005691 14.638899
#> [155] 18.845259 10.450910 14.298030 14.118066 17.693911 10.565853 13.554052
#> [162] 17.512049 16.576421 18.933526 15.331336 12.451021 18.220906 12.051407
#> [169] 16.136975 14.931191 18.633898 16.270650 14.665783 15.911005 12.758363
#> [176] 11.446664 11.361290 11.401264 10.727769 15.826799 15.287616 14.913155
#> [183] 12.460277 11.281248 12.063488 14.000137 17.716796 17.873659 17.581607
#> [190] 15.797449 12.144179 12.290586 15.062666 10.633711 11.995187 14.123808
#> [197] 10.915592 18.378118 15.321387 12.237994 18.017859 14.980410 14.459998
#> [204] 14.882981 11.602305 12.857629 17.447629 14.747075 10.350643 12.274393
#> [211] 10.347813 11.589196 16.008635 16.028939 11.740748  9.885508 10.253065
#> [218] 16.847688 13.081536 16.946751 14.756275 16.184042 16.552550 10.742998
#> [225] 15.647950 14.503357 10.517050 16.188747 14.303649 13.339712 10.137152
#> [232]  8.452093 10.251702 15.244325 11.639251 14.072009 14.975659 15.318868
#> [239] 12.911030 10.260078 10.875950  9.450337 10.768110 12.472102 10.155052
#> [246]  7.226264 11.913264  8.043231 11.549151  8.338942 10.426394 11.667656
#> [253] 12.928119  8.862160 11.636097 12.099835  9.154194  9.510715 11.086837
#> [260] 10.063381  9.098615  8.110580 11.253017  5.852955  7.065511  8.592994
#> [267] 11.332543  7.786991  6.641609  6.300018 10.970847 10.382947  7.045228
#> [274]  6.498306  5.832811  5.739524  5.322790  5.582501  4.509915  7.218918
#> [281]  8.494596  7.705093  4.679113  8.138150  4.203309  4.932573  8.724597
#> [288]  6.796521  6.097102  4.064213  6.697533  6.331917  7.536617  6.265837
#> [295]  6.989108  3.994765  7.667325  7.371024  5.878909  6.384860  6.215102
#> [302]  5.225077  4.742868  6.654227  5.294988  6.508080  5.226290  2.718018
#> [309]  3.500598  4.804456  6.184806  4.511496  4.461820  4.726568  2.576828
#> [316]  2.939990  5.637238  3.420574  4.520470  4.525532  2.417420  4.917375
#> [323]  3.077654  2.421011  5.106643  4.219254  3.187642  2.241010  2.613081
#> [330]  3.698251  2.875663  4.565797  3.012262  1.693794  4.019512  2.388579
#> [337]  1.745455  3.850990  4.475415  3.718849  1.697142  3.933645  3.622329
#> [344]  2.194436  3.805441  4.133786  3.272010  3.064227  4.239830  3.160684
#> [351]  3.426654  3.836957  1.657744  4.150405  3.255243  3.879202  2.005369
#> [358]  3.617328  3.525718  3.523638  3.315633  4.178058  1.605615  1.886403
#> [365]  2.158124  2.723961  4.344234  3.673684  2.994827  4.474746  1.776897
#> [372]  2.423950  1.785985  2.116210  2.748591  2.252917  3.829839  4.560958
#> [379]  4.004966  3.105198  2.284763  3.559454  4.919463  3.825419  4.610123
#> [386]  2.718422  4.128664  4.082217  2.566785  5.464735  5.382149  4.908832
#> [393]  2.889067  4.049353  5.330518  3.746462  3.233243  2.727654  3.426391
#> [400]  4.258919  5.909595  4.017716  5.720791  3.952353  6.435109  6.941043
#> [407]  5.679224  3.892993  6.988197  7.516035  7.526976  3.299380  7.051747
#> [414]  6.063918  6.689986  4.415604  6.011061  8.338151  5.117122  5.485327
#> [421]  8.722906  6.734626  4.022107  7.430476  8.430022  6.975501  6.629643
#> [428]  9.649704  9.755066 10.014722  4.937426  8.150648  8.082677  9.016321
#> [435]  7.455100  9.338442  5.678021  9.053815  8.330326 11.495546  8.482281
#> [442] 10.688340  7.061287  9.544915  9.584767  6.254848  8.096770  8.165403
#> [449]  9.448070 12.140424 12.822204 12.018859 12.613030 12.849419 11.145606
#> [456] 12.569182 11.143073  7.089317 12.076676 12.288486  9.469859  9.825388
#> [463] 13.465333 12.819679 12.787259 11.191474  7.680384 14.046501 13.623822
#> [470] 10.621513 10.353200 10.047603 15.526153  8.295151 15.855124 10.017203
#> [477]  9.882153 13.305639 15.591076  8.473935 14.749894  9.262995 15.520104
#> [484] 16.789827 11.930870 16.618550 14.370990 11.311946 11.141435 13.281843
#> [491] 15.122681 13.465903 12.351093 12.031648 10.269853 15.831026 15.946889
#> [498] 11.158083 11.644175 15.069166 13.000614 11.902250 18.178357 18.162181
#> [505] 16.135650 17.959012 16.171962 17.389568 13.289805 10.181499 18.254193
#> [512] 10.666292 13.987102 18.643417 10.540604 15.985566 14.087701 16.005691
#> [519] 14.638899 18.845259 10.450910 14.298030 14.118066 17.693911 10.565853
#> [526] 13.554052 17.512049 16.576421 18.933526 15.331336 12.451021 18.220906
#> [533] 12.051407 16.136975 14.931191 18.633898 16.270650 14.665783 15.911005
#> [540] 12.758363 11.446664 11.361290 11.401264 10.727769 15.826799 15.287616
#> [547] 14.913155 12.460277 11.281248 12.063488 14.000137 17.716796 17.873659
#> [554] 17.581607 15.797449 12.144179 12.290586 15.062666 10.633711 11.995187
#> [561] 14.123808 10.915592 18.378118 15.321387 12.237994 18.017859 14.980410
#> [568] 14.459998 14.882981 11.602305 12.857629 17.447629 14.747075 10.350643
#> [575] 12.274393 10.347813 11.589196 16.008635 16.028939 11.740748  9.885508
#> [582] 10.253065 16.847688 13.081536 16.946751 14.756275 16.184042 16.552550
#> [589] 10.742998 15.647950 14.503357 10.517050 16.188747 14.303649 13.339712
#> [596] 10.137152  8.452093 10.251702 15.244325 11.639251 14.072009 14.975659
#> [603] 15.318868 12.911030 10.260078 10.875950  9.450337 10.768110 12.472102
#> [610] 10.155052  7.226264 11.913264  8.043231 11.549151  8.338942 10.426394
#> [617] 11.667656 12.928119  8.862160 11.636097 12.099835  9.154194  9.510715
#> [624] 11.086837 10.063381  9.098615  8.110580 11.253017  5.852955  7.065511
#> [631]  8.592994 11.332543  7.786991  6.641609  6.300018 10.970847 10.382947
#> [638]  7.045228  6.498306  5.832811  5.739524  5.322790  5.582501  4.509915
#> [645]  7.218918  8.494596  7.705093  4.679113  8.138150  4.203309  4.932573
#> [652]  8.724597  6.796521  6.097102  4.064213  6.697533  6.331917  7.536617
#> [659]  6.265837  6.989108  3.994765  7.667325  7.371024  5.878909  6.384860
#> [666]  6.215102  5.225077  4.742868  6.654227  5.294988  6.508080  5.226290
#> [673]  2.718018  3.500598  4.804456  6.184806  4.511496  4.461820  4.726568
#> [680]  2.576828  2.939990  5.637238  3.420574  4.520470  4.525532  2.417420
#> [687]  4.917375  3.077654  2.421011  5.106643  4.219254  3.187642  2.241010
#> [694]  2.613081  3.698251  2.875663  4.565797  3.012262  1.693794  4.019512
#> [701]  2.388579  1.745455  3.850990  4.475415  3.718849  1.697142  3.933645
#> [708]  3.622329  2.194436  3.805441  4.133786  3.272010  3.064227  4.239830
#> [715]  3.160684  3.426654  3.836957  1.657744  4.150405  3.255243  3.879202
#> [722]  2.005369  3.617328  3.525718  3.523638  3.315633  4.178058  1.605615
#> [729]  1.886403  2.158124
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 6.81651616  7.463176  6.099033  4.310044
#> 3     2 2002-01-02 4.11096523  7.483435  6.149345  3.226383
#> 6     3 2002-01-03 6.02744283  7.505323  6.203934  4.042139
#> 7     4 2002-01-04 0.55602096  7.528820  6.262814  1.796965
#> 9     5 2002-01-05 6.15347674  7.553900  6.325996  4.158107
#> 11    6 2002-01-06 6.32319845  7.580540  6.393496  4.264895
#> 14    7 2002-01-07 2.69039133  7.608714  6.465324  2.759379
#> 15    8 2002-01-08 6.60180260  7.638396  6.541496  4.462254
#> 17    9 2002-01-09 6.66864889  7.669557  6.622023  4.534417
#> 19   10 2002-01-10 2.17075999  7.702170  6.706920  2.621861
#> 22   11 2002-01-11 3.59728135  7.736204  6.796198  3.279142
#> 23   12 2002-01-12 0.63673530  7.771631  6.889871  2.004714
#> 25   13 2002-01-13 2.41500189  7.808419  6.987950  2.827610
#> 27   14 2002-01-14 5.54607177  7.846537  7.090446  4.278438
#> 30   15 2002-01-15 2.01591591  7.885954  7.197371  2.719288
#> 32   16 2002-01-16 0.68819404  7.926639  7.308732  2.144457
#> 33   17 2002-01-17 3.12463912  7.968559  7.424539  3.311794
#> 35   18 2002-01-18 3.44593314  8.011681  7.544799  3.508760
#> 37   19 2002-01-19 1.72892183  8.055974  7.669519  2.740371
#> 40   20 2002-01-20 6.46163421  8.101405  7.798703  5.059776
#> 41   21 2002-01-21 5.19876554  8.147942  7.932354  4.513693
#> 43   22 2002-01-22 3.23705346  8.195553  8.070474  3.611443
#> 45   23 2002-01-23 5.64154386  8.244205  8.213063  4.863382
#> 48   24 2002-01-24 6.00946650  8.293867  8.360119  5.118765
#> 49   25 2002-01-25 2.70345387  8.344507  8.511639  3.506711
#> 51   26 2002-01-26 5.87384799  8.396093  8.667618  5.198807
#> 53   27 2002-01-27 2.11908267  8.448595  8.828046  3.314140
#> 56   28 2002-01-28 5.00933608  8.501983  8.992915  4.897525
#> 58   29 2002-01-29 1.51888064  8.556225  9.162213  3.103780
#> 59   30 2002-01-30 3.89648290  8.611293  9.335925  4.446166
#> 61   31 2002-01-31 1.85597243  8.667157  9.514035  3.397170
#> 63   32 2002-02-01 2.02786180  8.723789  9.696523  3.551119
#> 66   33 2002-02-02 4.98895643  8.781160  9.883368  5.278426
#> 67   34 2002-02-03 0.59119595  8.839243 10.074546  2.855545
#> 69   35 2002-02-04 6.36360197  8.898011 10.270030  6.239923
#> 71   36 2002-02-05 5.88813301  8.957438 10.469790  6.058583
#> 74   37 2002-02-06 1.40590496  9.017498 10.673795  3.500517
#> 75   38 2002-02-07 6.21931201  9.078165 10.882010  6.448052
#> 77   39 2002-02-08 1.21012896  9.139416 11.094398  3.508091
#> 79   40 2002-02-09 5.35795585  9.201226 11.310918  6.120954
#> 82   41 2002-02-10 5.49136003  9.263572 11.531528  6.300774
#> 83   42 2002-02-11 1.84426978  9.326432 11.756182  4.101418
#> 86   43 2002-02-12 0.39923590  9.389784 11.984831  3.250994
#> 87   44 2002-02-13 1.94467798  9.453606 12.217425  4.310965
#> 89   45 2002-02-14 0.85874924  9.517878 12.453910  3.675304
#> 91   46 2002-02-15 3.48857656  9.582579 12.694229  5.484250
#> 94   47 2002-02-16 4.40203654  9.647690 12.938322  6.186322
#> 95   48 2002-02-17 3.98421103  9.713192 13.186128  6.000912
#> 97   49 2002-02-18 3.22111683  9.779067 13.437583  5.572491
#> 99   50 2002-02-19 6.78945449  9.845297 13.692618  8.144465
#> 102  51 2002-02-20 4.16755545  9.911864 13.951164  6.420753
#> 103  52 2002-02-21 1.45267408  9.978754 14.213148  4.587839
#> 105  53 2002-02-22 1.92522427 10.045948 14.478496  5.006967
#> 107  54 2002-02-23 5.43805145 10.113433 14.747130  7.651591
#> 110  55 2002-02-24 3.87515190 10.181192 15.018971  6.612993
#> 112  56 2002-02-25 1.43550549 10.249211 15.293936  4.894519
#> 113  57 2002-02-26 0.20303964 10.317477 15.571940  4.046207
#> 115  58 2002-02-27 4.89588423 10.385975 15.852898  7.699703
#> 117  59 2002-02-28 4.52318019 10.454693 16.136720  7.524923
#> 120  60 2002-03-01 0.26490426 10.523618 16.423316  4.312536
#> 121  61 2002-03-02 4.36823780 10.592738 16.712592  7.624121
#> 123  62 2002-03-03 2.37436683 10.662040 17.004454  6.144504
#> 125  63 2002-03-04 0.33262868 10.731513 17.298805  4.592794
#> 128  64 2002-03-05 0.48084085 10.801147 17.595546  4.790542
#> 129  65 2002-03-06 0.29157237 10.870930 17.894578  4.713622
#> 131  66 2002-03-07 1.47316153 10.940852 18.195799  5.773962
#> 133  67 2002-03-08 5.33910886 11.010902 18.499104  9.109819
#> 136  68 2002-03-09 3.93723299 11.081072 18.804391  8.041807
#> 138  69 2002-03-10 6.10469300 11.151350 19.111552 10.009100
#> 139  70 2002-03-11 1.70554382 11.221729 19.420480  6.330939
#> 141  71 2002-03-12 4.61225684 11.292198 19.731067  8.962308
#> 143  72 2002-03-13 4.73300703 11.362749 20.043203  9.185170
#> 146  73 2002-03-14 2.32741559 11.433374 20.356778  7.161141
#> 147  74 2002-03-15 5.78815145 11.504062 20.671679 10.368292
#> 149  75 2002-03-16 2.40832940 11.574807 20.987796  7.430377
#> 151  76 2002-03-17 3.87663301 11.645600 21.305014  8.872302
#> 154  77 2002-03-18 0.20692377 11.716433 21.623220  5.596749
#> 155  78 2002-03-19 3.20000391 11.787297 21.942300  8.464012
#> 157  79 2002-03-20 4.27913525 11.858185 22.262140  9.582284
#> 159  80 2002-03-21 0.78654443 11.929088 22.582624  6.390149
#> 162  81 2002-03-22 2.89512787 12.000000 22.903638  8.488783
#> 163  82 2002-03-23 5.87282343 12.070912 23.225065 11.456076
#> 166  83 2002-03-24 3.28692029 12.141815 23.546791  9.073883
#> 167  84 2002-03-25 1.69486413 12.212703 23.868700  7.623410
#> 169  85 2002-03-26 6.10527370 12.283567 24.190677 12.059388
#> 171  86 2002-03-27 6.73210827 12.354400 24.512606 12.806805
#> 174  87 2002-03-28 1.94288653 12.425193 24.834373  8.150228
#> 175  88 2002-03-29 1.29477272 12.495938 25.155863  7.592234
#> 177  89 2002-03-30 5.06129257 12.566626 25.476961 11.499748
#> 179  90 2002-03-31 0.99705148 12.637251 25.797553  7.467074
#> 182  91 2002-04-01 6.70217164 12.707802 26.117528 13.416653
#> 183  92 2002-04-02 0.04539211 12.778271 26.436771  6.656148
#> 185  93 2002-04-03 6.94761753 12.848650 26.755170 13.922420
#> 187  94 2002-04-04 5.14043081 12.918928 27.072615 12.154239
#> 190  95 2002-04-05 6.30353822 12.989098 27.388994 13.493113
#> 192  96 2002-04-06 3.29985629 13.059148 27.704199 10.426273
#> 193  97 2002-04-07 0.65467075 13.129070 28.018120  7.703081
#> 195  98 2002-04-08 0.33422863 13.198853 28.330649  7.441364
#> 197  99 2002-04-09 6.77062334 13.268487 28.641680 14.468036
#> 200 100 2002-04-10 3.66864573 13.337960 28.951106 11.219320
#> 201 101 2002-04-11 2.89539416 13.407262 29.258822 10.474032
#> 203 102 2002-04-12 2.06680679 13.476382 29.564726  9.658280
#> 205 103 2002-04-13 5.92747984 13.545307 29.868714 14.002512
#> 208 104 2002-04-14 2.85533862 13.614025 30.170685 10.706597
#> 209 105 2002-04-15 5.46462140 13.682523 30.470539 13.702403
#> 211 106 2002-04-16 3.23637838 13.750789 30.768176 11.312835
#> 213 107 2002-04-17 6.90873697 13.818808 31.063500 15.531000
#> 216 108 2002-04-18 5.45435209 13.886567 31.356413 13.997173
#> 218 109 2002-04-19 6.40977794 13.954052 31.646820 15.180171
#> 219 110 2002-04-20 6.11400434 14.021246 31.934628 14.946250
#> 221 111 2002-04-21 3.86551066 14.088136 32.219743 12.475171
#> 223 112 2002-04-22 3.58619944 14.154703 32.502075 12.242840
#> 226 113 2002-04-23 2.95953238 14.220933 32.781534 11.606482
#> 227 114 2002-04-24 2.28881220 14.286808 33.058030 10.912532
#> 229 115 2002-04-25 2.83922691 14.352310 33.331477 11.629747
#> 231 116 2002-04-26 3.97809642 14.417421 33.601789 13.036198
#> 234 117 2002-04-27 0.65902690 14.482122 33.868881  9.237843
#> 235 118 2002-04-28 5.26336332 14.546394 34.132670 14.708328
#> 237 119 2002-04-29 6.67483067 14.610216 34.393074 16.454687
#> 239 120 2002-04-30 0.03219655 14.673568 34.650012  8.700517
#> 242 121 2002-05-01 5.80371422 14.736428 34.903405 15.598934
#> 243 122 2002-05-02 4.15440174 14.798774 35.153175 13.722500
#> 245 123 2002-05-03 5.33555781 14.860584 35.399245 15.204700
#> 248 124 2002-05-04 0.89817294 14.921835 35.641540  9.983050
#> 249 125 2002-05-05 5.07206407 14.982502 35.879987 15.043267
#> 251 126 2002-05-06 5.05965420 15.042562 36.114511 15.102292
#> 253 127 2002-05-07 6.50669820 15.101989 36.345041 16.915898
#> 256 128 2002-05-08 2.68876838 15.160757 36.571508 12.385865
#> 257 129 2002-05-09 3.26975911 15.218840 36.793843 13.151029
#> 259 130 2002-05-10 2.33627446 15.276211 37.011977 12.083216
#> 261 131 2002-05-11 0.53688007 15.332843 37.225844  9.958193
#> 264 132 2002-05-12 3.67127328 15.388707 37.435378 13.824311
#> 265 133 2002-05-13 0.13183720 15.443775 37.640517  9.570790
#> 267 134 2002-05-14 2.01110229 15.498017 37.841197 11.915533
#> 269 135 2002-05-15 6.07225078 15.551405 38.037357 16.935433
#> 272 136 2002-05-16 3.82423750 15.603907 38.228936 14.241847
#> 274 137 2002-05-17 4.21314685 15.655493 38.415875 14.773136
#> 275 138 2002-05-18 1.37206043 15.706133 38.598118 11.335462
#> 277 139 2002-05-19 2.71929669 15.755795 38.775606 13.040048
#> 279 140 2002-05-20 0.42833791 15.804447 38.948286 10.264867
#> 282 141 2002-05-21 1.68419472 15.852058 39.116103 11.856962
#> 283 142 2002-05-22 4.03482256 15.898595 39.279005 14.803960
#> 285 143 2002-05-23 0.92905815 15.944026 39.436940 11.008230
#> 287 144 2002-05-24 4.29301620 15.988319 39.589859 15.212592
#> 290 145 2002-05-25 4.72645110 16.031441 39.737713 15.792241
#> 291 146 2002-05-26 5.42138403 16.073361 39.880455 16.695753
#> 293 147 2002-05-27 1.02770307 16.114046 40.018039 11.280622
#> 295 148 2002-05-28 5.13065227 16.153463 40.150421 16.413880
#> 298 149 2002-05-29 6.07801570 16.191581 40.277558 17.629109
#> 300 150 2002-05-30 4.88656377 16.228369 40.399408 16.182234
#> 301 151 2002-05-31 6.91762896 16.263796 40.515933 18.745489
#> 303 152 2002-06-01 0.78131857 16.297830 40.627093 11.130605
#> 305 153 2002-06-02 1.34041448 16.330443 40.732853 11.854904
#> 308 154 2002-06-03 6.93765119 16.361604 40.833176 18.865341
#> 309 155 2002-06-04 2.96558720 16.391286 40.928031 13.934452
#> 311 156 2002-06-05 1.78270920 16.419460 41.017385 12.481036
#> 313 157 2002-06-06 0.64246569 16.446100 41.101209 11.078110
#> 316 158 2002-06-07 1.54799727 16.471180 41.179475 12.229937
#> 317 159 2002-06-08 5.96570962 16.494677 41.252156 17.772973
#> 319 160 2002-06-09 4.32579091 16.516565 41.319229 15.740689
#> 321 161 2002-06-10 6.61478094 16.536824 41.380671 18.621366
#> 324 162 2002-06-11 0.76124858 16.555433 41.436461 11.311777
#> 325 163 2002-06-12 0.67424056 16.572373 41.486582 11.215578
#> 328 164 2002-06-13 6.96385771 16.587625 41.531016 19.100581
#> 329 165 2002-06-14 3.21931746 16.601175 41.569750 14.423062
#> 331 166 2002-06-15 2.52291823 16.613007 41.602770 13.559675
#> 333 167 2002-06-16 2.54867600 16.623108 41.630068 13.598904
#> 336 168 2002-06-17 2.50719665 16.631469 41.651634 13.552404
#> 337 169 2002-06-18 3.52247860 16.638080 41.667464 14.827614
#> 339 170 2002-06-19 4.37189991 16.642934 41.677552 15.893485
#> 341 171 2002-06-20 6.85328672 16.646025 41.681897 19.000840
#> 344 172 2002-06-21 2.04756405 16.647351 41.680500 12.983400
#> 345 173 2002-06-22 2.09296112 16.646909 41.673363 13.038068
#> 347 174 2002-06-23 5.67501016 16.644700 41.660491 17.517194
#> 349 175 2002-06-24 6.51824550 16.640727 41.641891 18.566128
#> 352 176 2002-06-25 6.36676532 16.634994 41.617572 18.368607
#> 354 177 2002-06-26 5.59184235 16.627507 41.587545 17.389847
#> 355 178 2002-06-27 2.87008806 16.618275 41.551823 13.976096
#> 357 179 2002-06-28 0.50678018 16.607306 41.510423 11.010961
#> 359 180 2002-06-29 5.68746681 16.594614 41.463360 17.471202
#> 362 181 2002-06-30 0.62761557 16.580211 41.410655 11.136428
#> 363 182 2002-07-01 1.53789628 16.564113 41.352328 12.257762
#> 365 183 2002-07-02 2.12520906 16.546336 41.288405 12.973640
#> 367 184 2002-07-03 2.49644591 16.526900 41.218909 13.417857
#> 370 185 2002-07-04 1.19388600 16.505823 41.143868 11.773960
#> 371 186 2002-07-05 0.69951172 16.483128 41.063311 11.137151
#> 373 187 2002-07-06 2.28776717 16.458837 40.977269 13.092224
#> 375 188 2002-07-07 5.26572827 16.432973 40.885775 16.772096
#> 378 189 2002-07-08 6.53851552 16.405563 40.788864 18.325514
#> 380 190 2002-07-09 1.39705038 16.376632 40.686570 11.907078
#> 381 191 2002-07-10 1.84424429 16.346207 40.578932 12.433872
#> 383 192 2002-07-11 4.94156660 16.314316 40.465990 16.245022
#> 385 193 2002-07-12 0.19890206 16.280989 40.347784 10.333407
#> 388 194 2002-07-13 2.05795845 16.246255 40.224356 12.603755
#> 389 195 2002-07-14 3.73312691 16.210144 40.095751 14.640878
#> 391 196 2002-07-15 2.85135033 16.172686 39.962013 13.513285
#> 393 197 2002-07-16 6.59777443 16.133915 39.823189 18.098409
#> 396 198 2002-07-17 4.70372819 16.093860 39.679327 15.718340
#> 397 199 2002-07-18 6.97611070 16.052554 39.530476 18.472186
#> 399 200 2002-07-19 2.56679318 16.010029 39.376687 13.000688
#> 401 201 2002-07-20 5.40962197 15.966317 39.218012 16.448321
#> 404 202 2002-07-21 2.93909647 15.921451 39.054503 13.368352
#> 405 203 2002-07-22 3.22650038 15.875462 38.886214 13.673135
#> 408 204 2002-07-23 1.42578058 15.828384 38.713200 11.421893
#> 409 205 2002-07-24 6.55543364 15.780249 38.535519 17.638095
#> 411 206 2002-07-25 3.50527372 15.731088 38.353226 13.861328
#> 413 207 2002-07-26 4.93253770 15.680933 38.166382 15.544335
#> 416 208 2002-07-27 4.85668313 15.629816 37.975044 15.393791
#> 417 209 2002-07-28 4.41609321 15.577768 37.779274 14.799783
#> 419 210 2002-07-29 0.68893896 15.524820 37.579134 10.228601
#> 421 211 2002-07-30 1.12687208 15.471001 37.374685 10.704814
#> 424 212 2002-07-31 0.93110720 15.416342 37.165992 10.413863
#> 425 213 2002-08-01 2.25281311 15.360873 36.953119 11.948037
#> 427 214 2002-08-02 6.56854166 15.304621 36.736132 17.067365
#> 429 215 2002-08-03 3.69485156 15.247616 36.515097 13.553002
#> 432 216 2002-08-04 1.87109866 15.189886 36.290082 11.307637
#> 434 217 2002-08-05 3.11094845 15.131457 36.061156 12.722281
#> 435 218 2002-08-06 3.50953046 15.072356 35.828388 13.128336
#> 437 219 2002-08-07 2.43001206 15.012610 35.591848 11.778495
#> 439 220 2002-08-08 2.71502714 14.952243 35.351609 12.047474
#> 442 221 2002-08-09 0.77817075 14.891281 35.107744  9.694245
#> 443 222 2002-08-10 6.45261387 14.829748 34.860325 16.299169
#> 445 223 2002-08-11 0.74547101 14.767666 34.609429  9.525898
#> 447 224 2002-08-12 1.05421952 14.705060 34.355129  9.820258
#> 450 225 2002-08-13 2.19809743 14.641952 34.097505 11.083790
#> 451 226 2002-08-14 2.41664170 14.578362 33.836632 11.263692
#> 453 227 2002-08-15 5.08302336 14.514313 33.572591 14.271837
#> 455 228 2002-08-16 4.13123194 14.449824 33.305462 13.087413
#> 458 229 2002-08-17 1.84628859 14.384916 33.035326 10.378856
#> 460 230 2002-08-18 6.93109153 14.319607 32.762264 16.119493
#> 461 231 2002-08-19 5.83289684 14.253916 32.486361 14.768521
#> 463 232 2002-08-20 3.82599057 14.187862 32.207701 12.394594
#> 465 233 2002-08-21 5.28072873 14.121461 31.926370 13.951035
#> 468 234 2002-08-22 3.98910814 14.054730 31.642455 12.401101
#> 469 235 2002-08-23 1.53539459 13.987686 31.356043  9.559949
#> 471 236 2002-08-24 0.73824277 13.920345 31.067223  8.590605
#> 473 237 2002-08-25 5.41833280 13.852721 30.776085 13.712877
#> 476 238 2002-08-26 0.76829816 13.784830 30.482721  8.470158
#> 477 239 2002-08-27 5.55815313 13.716686 30.187223 13.662905
#> 479 240 2002-08-28 4.01092762 13.648302 29.889684 11.864373
#> 481 241 2002-08-29 5.45376630 13.579692 29.590199 13.339438
#> 484 242 2002-08-30 4.83188715 13.510869 29.288864 12.559499
#> 485 243 2002-08-31 3.93609592 13.441846 28.985774 11.490309
#> 487 244 2002-09-01 4.67327780 13.372633 28.681027 12.181775
#> 490 245 2002-09-02 5.00060067 13.303244 28.374722 12.426615
#> 491 246 2002-09-03 4.81411165 13.233689 28.066958 12.121796
#> 493 247 2002-09-04 3.58384362 13.163980 27.757836 10.717942
#> 495 248 2002-09-05 6.93351707 13.094126 27.447457 14.128765
#> 498 249 2002-09-06 6.24346730 13.024138 27.135924 13.288144
#> 499 250 2002-09-07 6.30522871 12.954027 26.823338 13.233815
#> 501 251 2002-09-08 2.30568382 12.883802 26.509805  8.999547
#> 503 252 2002-09-09 2.37312208 12.813472 26.195428  8.974622
#> 506 253 2002-09-10 6.41750228 12.743047 25.880314 12.986847
#> 507 254 2002-09-11 4.30558717 12.672536 25.564567 10.734016
#> 509 255 2002-09-12 0.41705192 12.601947 25.248295  6.729860
#> 511 256 2002-09-13 1.88001107 12.531289 24.931604  8.103087
#> 514 257 2002-09-14 6.75448884 12.460572 24.614602 12.825056
#> 516 258 2002-09-15 5.96518036 12.389802 24.297397 11.923448
#> 517 259 2002-09-16 4.23597003 12.318988 23.980097 10.117886
#> 519 260 2002-09-17 0.84445755 12.248139 23.662812  6.731428
#> 521 261 2002-09-18 2.69712201 12.177262 23.345649  8.421807
#> 524 262 2002-09-19 2.65048756 12.106365 23.028718  8.278057
#> 525 263 2002-09-20 2.12256496 12.035456 22.712128  7.680780
#> 527 264 2002-09-21 0.96791049 11.964544 22.395989  6.504895
#> 529 265 2002-09-22 4.30259115 11.893635 22.080409  9.513960
#> 532 266 2002-09-23 3.77226017 11.822738 21.765497  8.913714
#> 533 267 2002-09-24 6.22374461 11.751861 21.451363 11.043124
#> 535 268 2002-09-25 5.57640106 11.681012 21.138114 10.330093
#> 537 269 2002-09-26 1.99081192 11.610198 20.825859  6.991980
#> 540 270 2002-09-27 3.68356425 11.539428 20.514705  8.402982
#> 542 271 2002-09-28 5.92767766 11.468711 20.204761 10.272671
#> 543 272 2002-09-29 5.36367648 11.398053 19.896131  9.655377
#> 545 273 2002-09-30 1.20286877 11.327464 19.588923  5.937309
#> 547 274 2002-10-01 2.46615531 11.256953 19.283241  6.933081
#> 550 275 2002-10-02 5.38321367 11.186528 18.979189  9.311408
#> 551 276 2002-10-03 2.97672628 11.116198 18.676871  7.169890
#> 553 277 2002-10-04 3.12711917 11.045973 18.376389  7.195279
#> 555 278 2002-10-05 6.14123372 10.975862 18.077844  9.576935
#> 558 279 2002-10-06 1.26902670 10.905874 17.781337  5.479868
#> 559 280 2002-10-07 6.16404388 10.836020 17.486965  9.345450
#> 561 281 2002-10-08 6.23210676 10.766311 17.194826  9.275341
#> 563 282 2002-10-09 5.09742295 10.696756 16.905016  8.254205
#> 566 283 2002-10-10 1.93867145 10.627367 16.617629  5.670123
#> 567 284 2002-10-11 3.70066513 10.558154 16.332759  6.945530
#> 570 285 2002-10-12 3.79230540 10.489131 16.050495  6.914122
#> 571 286 2002-10-13 5.27154887 10.420308 15.770928  7.931924
#> 573 287 2002-10-14 2.69135938 10.351698 15.494145  5.887714
#> 575 288 2002-10-15 5.86535115 10.283314 15.220231  8.145682
#> 578 289 2002-10-16 3.11901290 10.215170 14.949271  6.019559
#> 579 290 2002-10-17 3.87092373 10.147279 14.681345  6.470613
#> 581 291 2002-10-18 5.13759945 10.079655 14.416534  7.278187
#> 583 292 2002-10-19 3.27283361 10.012314 14.154915  5.852214
#> 586 293 2002-10-20 1.56437010  9.945270 13.896562  4.567091
#> 587 294 2002-10-21 3.83007683  9.878539 13.641550  6.054917
#> 589 295 2002-10-22 4.62640734  9.812138 13.389948  6.504156
#> 591 296 2002-10-23 4.47015012  9.746084 13.141825  6.299279
#> 594 297 2002-10-24 4.39361061  9.680393 12.897247  6.151129
#> 596 298 2002-10-25 5.59970454  9.615084 12.656277  6.849498
#> 597 299 2002-10-26 6.75185071  9.550176 12.418977  7.494772
#> 599 300 2002-10-27 2.23886120  9.485687 12.185404  4.484382
#> 601 301 2002-10-28 1.66044273  9.421638 11.955615  4.042416
#> 604 302 2002-10-29 4.66764785  9.358048 11.729663  5.857702
#> 605 303 2002-10-30 4.29079977  9.294940 11.507600  5.533012
#> 607 304 2002-10-31 0.19196456  9.232334 11.289473  2.939737
#> 609 305 2002-11-01 6.34958995  9.170252 11.075328  6.603176
#> 612 306 2002-11-02 3.58920569  9.108719 10.865209  4.856970
#> 613 307 2002-11-03 6.87503403  9.047757 10.659156  6.714526
#> 615 308 2002-11-04 6.20001596  8.987390 10.457208  6.221292
#> 617 309 2002-11-05 0.83706620  8.927644 10.259398  3.045816
#> 620 310 2002-11-06 5.39404287  8.868543 10.065762  5.577549
#> 622 311 2002-11-07 1.38495433  8.810114  9.876328  3.245364
#> 623 312 2002-11-08 3.58573787  8.752384  9.691126  4.407946
#> 625 313 2002-11-09 5.21425166  8.695379  9.510180  5.228972
#> 627 314 2002-11-10 6.00126653  8.639127  9.333513  5.575193
#> 630 315 2002-11-11 0.87864148  8.583658  9.161147  2.759164
#> 631 316 2002-11-12 1.82691784  8.528999  8.993100  3.211439
#> 633 317 2002-11-13 5.98673457  8.475180  8.829388  5.325818
#> 635 318 2002-11-14 6.72850404  8.422232  8.670025  5.630739
#> 638 319 2002-11-15 3.07742819  8.370184  8.515023  3.694096
#> 639 320 2002-11-16 2.06820487  8.319067  8.364392  3.130835
#> 641 321 2002-11-17 4.02135139  8.268912  8.218139  4.052864
#> 643 322 2002-11-18 0.74209443  8.219751  8.076271  2.383638
#> 646 323 2002-11-19 5.23357406  8.171616  7.938791  4.526927
#> 647 324 2002-11-20 6.07489372  8.124538  7.805702  4.869672
#> 650 325 2002-11-21 2.03353238  8.078549  7.677004  2.885479
#> 651 326 2002-11-22 0.12693248  8.033683  7.552698  1.947841
#> 653 327 2002-11-23 5.00167573  7.989971  7.432780  4.184634
#> 655 328 2002-11-24 1.61122478  7.947446  7.317248  2.571043
#> 658 329 2002-11-25 1.77097829  7.906140  7.206095  2.608608
#> 659 330 2002-11-26 3.64815790  7.866085  7.099317  3.421101
#> 661 331 2002-11-27 5.92679783  7.827314  6.996907  4.398236
#> 663 332 2002-11-28 5.87129415  7.789856  6.898857  4.324583
#> 666 333 2002-11-29 2.57193307  7.753745  6.805157  2.829932
#> 667 334 2002-11-30 2.40262695  7.719011  6.715800  2.724133
#> 669 335 2002-12-01 3.15807119  7.685684  6.630775  3.019997
#> 671 336 2002-12-02 2.28539523  7.653793  6.550072  2.615432
#> 674 337 2002-12-03 0.61902366  7.623368  6.473680  1.881254
#> 676 338 2002-12-04 4.87497517  7.594437  6.401590  3.655032
#> 677 339 2002-12-05 6.51847878  7.567027  6.333789  4.311512
#> 679 340 2002-12-06 6.56965046  7.541163  6.270268  4.298808
#> 681 341 2002-12-07 2.26203499  7.516872  6.211016  2.487288
#> 684 342 2002-12-08 0.06695332  7.494177  6.156023  1.566505
#> 685 343 2002-12-09 4.02899126  7.473100  6.105278  3.172096
#> 687 344 2002-12-10 3.10034322  7.453664  6.058772  2.774763
#> 689 345 2002-12-11 3.14021991  7.435887  6.016497  2.774526
#> 692 346 2002-12-12 2.41808780  7.419789  5.978443  2.468789
#> 693 347 2002-12-13 0.25340833  7.405386  5.944603  1.587861
#> 695 348 2002-12-14 0.02391030  7.392694  5.914970  1.488308
#> 697 349 2002-12-15 3.87203651  7.381725  5.889539  3.017044
#> 700 350 2002-12-16 5.66566520  7.372493  5.868303  3.721933
#> 702 351 2002-12-17 1.57932001  7.365006  5.851259  2.090174
#> 703 352 2002-12-18 4.96476662  7.359273  5.838404  3.428975
#> 705 353 2002-12-19 2.60354924  7.355300  5.829736  2.489207
#> 707 354 2002-12-20 0.61859034  7.353091  5.825253  1.701343
#> 710 355 2002-12-21 6.60801581  7.352649  5.824956  4.073758
#> 711 356 2002-12-22 5.02265773  7.353975  5.828845  3.447720
#> 713 357 2002-12-23 6.33307338  7.357066  5.836924  3.971487
#> 715 358 2002-12-24 2.73019727  7.361920  5.849194  2.546897
#> 718 359 2002-12-25 2.85629146  7.368531  5.865662  2.603280
#> 719 360 2002-12-26 3.28382834  7.376892  5.886331  2.781735
#> 721 361 2002-12-27 3.32164519  7.386993  5.911209  2.806823
#> 723 362 2002-12-28 2.84064925  7.398825  5.940303  2.625413
#> 726 363 2002-12-29 6.74354469  7.412375  5.973622  4.210712
#> 727 364 2002-12-30 2.82357060  7.427627  6.011174  2.645351
#> 729 365 2002-12-31 5.93436651  7.444567  6.052971  3.925777
#> 2     1 2003-01-01 6.81651616  7.463176  6.099033  4.310044
#> 4     2 2003-01-02 4.11096523  7.483435  6.149345  3.226383
#> 5     3 2003-01-03 6.02744283  7.505323  6.203934  4.042139
#> 8     4 2003-01-04 0.55602096  7.528820  6.262814  1.796965
#> 10    5 2003-01-05 6.15347674  7.553900  6.325996  4.158107
#> 12    6 2003-01-06 6.32319845  7.580540  6.393496  4.264895
#> 13    7 2003-01-07 2.69039133  7.608714  6.465324  2.759379
#> 16    8 2003-01-08 6.60180260  7.638396  6.541496  4.462254
#> 18    9 2003-01-09 6.66864889  7.669557  6.622023  4.534417
#> 20   10 2003-01-10 2.17075999  7.702170  6.706920  2.621861
#> 21   11 2003-01-11 3.59728135  7.736204  6.796198  3.279142
#> 24   12 2003-01-12 0.63673530  7.771631  6.889871  2.004714
#> 26   13 2003-01-13 2.41500189  7.808419  6.987950  2.827610
#> 28   14 2003-01-14 5.54607177  7.846537  7.090446  4.278438
#> 29   15 2003-01-15 2.01591591  7.885954  7.197371  2.719288
#> 31   16 2003-01-16 0.68819404  7.926639  7.308732  2.144457
#> 34   17 2003-01-17 3.12463912  7.968559  7.424539  3.311794
#> 36   18 2003-01-18 3.44593314  8.011681  7.544799  3.508760
#> 38   19 2003-01-19 1.72892183  8.055974  7.669519  2.740371
#> 39   20 2003-01-20 6.46163421  8.101405  7.798703  5.059776
#> 42   21 2003-01-21 5.19876554  8.147942  7.932354  4.513693
#> 44   22 2003-01-22 3.23705346  8.195553  8.070474  3.611443
#> 46   23 2003-01-23 5.64154386  8.244205  8.213063  4.863382
#> 47   24 2003-01-24 6.00946650  8.293867  8.360119  5.118765
#> 50   25 2003-01-25 2.70345387  8.344507  8.511639  3.506711
#> 52   26 2003-01-26 5.87384799  8.396093  8.667618  5.198807
#> 54   27 2003-01-27 2.11908267  8.448595  8.828046  3.314140
#> 55   28 2003-01-28 5.00933608  8.501983  8.992915  4.897525
#> 57   29 2003-01-29 1.51888064  8.556225  9.162213  3.103780
#> 60   30 2003-01-30 3.89648290  8.611293  9.335925  4.446166
#> 62   31 2003-01-31 1.85597243  8.667157  9.514035  3.397170
#> 64   32 2003-02-01 2.02786180  8.723789  9.696523  3.551119
#> 65   33 2003-02-02 4.98895643  8.781160  9.883368  5.278426
#> 68   34 2003-02-03 0.59119595  8.839243 10.074546  2.855545
#> 70   35 2003-02-04 6.36360197  8.898011 10.270030  6.239923
#> 72   36 2003-02-05 5.88813301  8.957438 10.469790  6.058583
#> 73   37 2003-02-06 1.40590496  9.017498 10.673795  3.500517
#> 76   38 2003-02-07 6.21931201  9.078165 10.882010  6.448052
#> 78   39 2003-02-08 1.21012896  9.139416 11.094398  3.508091
#> 80   40 2003-02-09 5.35795585  9.201226 11.310918  6.120954
#> 81   41 2003-02-10 5.49136003  9.263572 11.531528  6.300774
#> 84   42 2003-02-11 1.84426978  9.326432 11.756182  4.101418
#> 85   43 2003-02-12 0.39923590  9.389784 11.984831  3.250994
#> 88   44 2003-02-13 1.94467798  9.453606 12.217425  4.310965
#> 90   45 2003-02-14 0.85874924  9.517878 12.453910  3.675304
#> 92   46 2003-02-15 3.48857656  9.582579 12.694229  5.484250
#> 93   47 2003-02-16 4.40203654  9.647690 12.938322  6.186322
#> 96   48 2003-02-17 3.98421103  9.713192 13.186128  6.000912
#> 98   49 2003-02-18 3.22111683  9.779067 13.437583  5.572491
#> 100  50 2003-02-19 6.78945449  9.845297 13.692618  8.144465
#> 101  51 2003-02-20 4.16755545  9.911864 13.951164  6.420753
#> 104  52 2003-02-21 1.45267408  9.978754 14.213148  4.587839
#> 106  53 2003-02-22 1.92522427 10.045948 14.478496  5.006967
#> 108  54 2003-02-23 5.43805145 10.113433 14.747130  7.651591
#> 109  55 2003-02-24 3.87515190 10.181192 15.018971  6.612993
#> 111  56 2003-02-25 1.43550549 10.249211 15.293936  4.894519
#> 114  57 2003-02-26 0.20303964 10.317477 15.571940  4.046207
#> 116  58 2003-02-27 4.89588423 10.385975 15.852898  7.699703
#> 118  59 2003-02-28 4.52318019 10.454693 16.136720  7.524923
#> 119  60 2003-03-01 0.26490426 10.523618 16.423316  4.312536
#> 122  61 2003-03-02 4.36823780 10.592738 16.712592  7.624121
#> 124  62 2003-03-03 2.37436683 10.662040 17.004454  6.144504
#> 126  63 2003-03-04 0.33262868 10.731513 17.298805  4.592794
#> 127  64 2003-03-05 0.48084085 10.801147 17.595546  4.790542
#> 130  65 2003-03-06 0.29157237 10.870930 17.894578  4.713622
#> 132  66 2003-03-07 1.47316153 10.940852 18.195799  5.773962
#> 134  67 2003-03-08 5.33910886 11.010902 18.499104  9.109819
#> 135  68 2003-03-09 3.93723299 11.081072 18.804391  8.041807
#> 137  69 2003-03-10 6.10469300 11.151350 19.111552 10.009100
#> 140  70 2003-03-11 1.70554382 11.221729 19.420480  6.330939
#> 142  71 2003-03-12 4.61225684 11.292198 19.731067  8.962308
#> 144  72 2003-03-13 4.73300703 11.362749 20.043203  9.185170
#> 145  73 2003-03-14 2.32741559 11.433374 20.356778  7.161141
#> 148  74 2003-03-15 5.78815145 11.504062 20.671679 10.368292
#> 150  75 2003-03-16 2.40832940 11.574807 20.987796  7.430377
#> 152  76 2003-03-17 3.87663301 11.645600 21.305014  8.872302
#> 153  77 2003-03-18 0.20692377 11.716433 21.623220  5.596749
#> 156  78 2003-03-19 3.20000391 11.787297 21.942300  8.464012
#> 158  79 2003-03-20 4.27913525 11.858185 22.262140  9.582284
#> 160  80 2003-03-21 0.78654443 11.929088 22.582624  6.390149
#> 161  81 2003-03-22 2.89512787 12.000000 22.903638  8.488783
#> 164  82 2003-03-23 5.87282343 12.070912 23.225065 11.456076
#> 165  83 2003-03-24 3.28692029 12.141815 23.546791  9.073883
#> 168  84 2003-03-25 1.69486413 12.212703 23.868700  7.623410
#> 170  85 2003-03-26 6.10527370 12.283567 24.190677 12.059388
#> 172  86 2003-03-27 6.73210827 12.354400 24.512606 12.806805
#> 173  87 2003-03-28 1.94288653 12.425193 24.834373  8.150228
#> 176  88 2003-03-29 1.29477272 12.495938 25.155863  7.592234
#> 178  89 2003-03-30 5.06129257 12.566626 25.476961 11.499748
#> 180  90 2003-03-31 0.99705148 12.637251 25.797553  7.467074
#> 181  91 2003-04-01 6.70217164 12.707802 26.117528 13.416653
#> 184  92 2003-04-02 0.04539211 12.778271 26.436771  6.656148
#> 186  93 2003-04-03 6.94761753 12.848650 26.755170 13.922420
#> 188  94 2003-04-04 5.14043081 12.918928 27.072615 12.154239
#> 189  95 2003-04-05 6.30353822 12.989098 27.388994 13.493113
#> 191  96 2003-04-06 3.29985629 13.059148 27.704199 10.426273
#> 194  97 2003-04-07 0.65467075 13.129070 28.018120  7.703081
#> 196  98 2003-04-08 0.33422863 13.198853 28.330649  7.441364
#> 198  99 2003-04-09 6.77062334 13.268487 28.641680 14.468036
#> 199 100 2003-04-10 3.66864573 13.337960 28.951106 11.219320
#> 202 101 2003-04-11 2.89539416 13.407262 29.258822 10.474032
#> 204 102 2003-04-12 2.06680679 13.476382 29.564726  9.658280
#> 206 103 2003-04-13 5.92747984 13.545307 29.868714 14.002512
#> 207 104 2003-04-14 2.85533862 13.614025 30.170685 10.706597
#> 210 105 2003-04-15 5.46462140 13.682523 30.470539 13.702403
#> 212 106 2003-04-16 3.23637838 13.750789 30.768176 11.312835
#> 214 107 2003-04-17 6.90873697 13.818808 31.063500 15.531000
#> 215 108 2003-04-18 5.45435209 13.886567 31.356413 13.997173
#> 217 109 2003-04-19 6.40977794 13.954052 31.646820 15.180171
#> 220 110 2003-04-20 6.11400434 14.021246 31.934628 14.946250
#> 222 111 2003-04-21 3.86551066 14.088136 32.219743 12.475171
#> 224 112 2003-04-22 3.58619944 14.154703 32.502075 12.242840
#> 225 113 2003-04-23 2.95953238 14.220933 32.781534 11.606482
#> 228 114 2003-04-24 2.28881220 14.286808 33.058030 10.912532
#> 230 115 2003-04-25 2.83922691 14.352310 33.331477 11.629747
#> 232 116 2003-04-26 3.97809642 14.417421 33.601789 13.036198
#> 233 117 2003-04-27 0.65902690 14.482122 33.868881  9.237843
#> 236 118 2003-04-28 5.26336332 14.546394 34.132670 14.708328
#> 238 119 2003-04-29 6.67483067 14.610216 34.393074 16.454687
#> 240 120 2003-04-30 0.03219655 14.673568 34.650012  8.700517
#> 241 121 2003-05-01 5.80371422 14.736428 34.903405 15.598934
#> 244 122 2003-05-02 4.15440174 14.798774 35.153175 13.722500
#> 246 123 2003-05-03 5.33555781 14.860584 35.399245 15.204700
#> 247 124 2003-05-04 0.89817294 14.921835 35.641540  9.983050
#> 250 125 2003-05-05 5.07206407 14.982502 35.879987 15.043267
#> 252 126 2003-05-06 5.05965420 15.042562 36.114511 15.102292
#> 254 127 2003-05-07 6.50669820 15.101989 36.345041 16.915898
#> 255 128 2003-05-08 2.68876838 15.160757 36.571508 12.385865
#> 258 129 2003-05-09 3.26975911 15.218840 36.793843 13.151029
#> 260 130 2003-05-10 2.33627446 15.276211 37.011977 12.083216
#> 262 131 2003-05-11 0.53688007 15.332843 37.225844  9.958193
#> 263 132 2003-05-12 3.67127328 15.388707 37.435378 13.824311
#> 266 133 2003-05-13 0.13183720 15.443775 37.640517  9.570790
#> 268 134 2003-05-14 2.01110229 15.498017 37.841197 11.915533
#> 270 135 2003-05-15 6.07225078 15.551405 38.037357 16.935433
#> 271 136 2003-05-16 3.82423750 15.603907 38.228936 14.241847
#> 273 137 2003-05-17 4.21314685 15.655493 38.415875 14.773136
#> 276 138 2003-05-18 1.37206043 15.706133 38.598118 11.335462
#> 278 139 2003-05-19 2.71929669 15.755795 38.775606 13.040048
#> 280 140 2003-05-20 0.42833791 15.804447 38.948286 10.264867
#> 281 141 2003-05-21 1.68419472 15.852058 39.116103 11.856962
#> 284 142 2003-05-22 4.03482256 15.898595 39.279005 14.803960
#> 286 143 2003-05-23 0.92905815 15.944026 39.436940 11.008230
#> 288 144 2003-05-24 4.29301620 15.988319 39.589859 15.212592
#> 289 145 2003-05-25 4.72645110 16.031441 39.737713 15.792241
#> 292 146 2003-05-26 5.42138403 16.073361 39.880455 16.695753
#> 294 147 2003-05-27 1.02770307 16.114046 40.018039 11.280622
#> 296 148 2003-05-28 5.13065227 16.153463 40.150421 16.413880
#> 297 149 2003-05-29 6.07801570 16.191581 40.277558 17.629109
#> 299 150 2003-05-30 4.88656377 16.228369 40.399408 16.182234
#> 302 151 2003-05-31 6.91762896 16.263796 40.515933 18.745489
#> 304 152 2003-06-01 0.78131857 16.297830 40.627093 11.130605
#> 306 153 2003-06-02 1.34041448 16.330443 40.732853 11.854904
#> 307 154 2003-06-03 6.93765119 16.361604 40.833176 18.865341
#> 310 155 2003-06-04 2.96558720 16.391286 40.928031 13.934452
#> 312 156 2003-06-05 1.78270920 16.419460 41.017385 12.481036
#> 314 157 2003-06-06 0.64246569 16.446100 41.101209 11.078110
#> 315 158 2003-06-07 1.54799727 16.471180 41.179475 12.229937
#> 318 159 2003-06-08 5.96570962 16.494677 41.252156 17.772973
#> 320 160 2003-06-09 4.32579091 16.516565 41.319229 15.740689
#> 322 161 2003-06-10 6.61478094 16.536824 41.380671 18.621366
#> 323 162 2003-06-11 0.76124858 16.555433 41.436461 11.311777
#> 326 163 2003-06-12 0.67424056 16.572373 41.486582 11.215578
#> 327 164 2003-06-13 6.96385771 16.587625 41.531016 19.100581
#> 330 165 2003-06-14 3.21931746 16.601175 41.569750 14.423062
#> 332 166 2003-06-15 2.52291823 16.613007 41.602770 13.559675
#> 334 167 2003-06-16 2.54867600 16.623108 41.630068 13.598904
#> 335 168 2003-06-17 2.50719665 16.631469 41.651634 13.552404
#> 338 169 2003-06-18 3.52247860 16.638080 41.667464 14.827614
#> 340 170 2003-06-19 4.37189991 16.642934 41.677552 15.893485
#> 342 171 2003-06-20 6.85328672 16.646025 41.681897 19.000840
#> 343 172 2003-06-21 2.04756405 16.647351 41.680500 12.983400
#> 346 173 2003-06-22 2.09296112 16.646909 41.673363 13.038068
#> 348 174 2003-06-23 5.67501016 16.644700 41.660491 17.517194
#> 350 175 2003-06-24 6.51824550 16.640727 41.641891 18.566128
#> 351 176 2003-06-25 6.36676532 16.634994 41.617572 18.368607
#> 353 177 2003-06-26 5.59184235 16.627507 41.587545 17.389847
#> 356 178 2003-06-27 2.87008806 16.618275 41.551823 13.976096
#> 358 179 2003-06-28 0.50678018 16.607306 41.510423 11.010961
#> 360 180 2003-06-29 5.68746681 16.594614 41.463360 17.471202
#> 361 181 2003-06-30 0.62761557 16.580211 41.410655 11.136428
#> 364 182 2003-07-01 1.53789628 16.564113 41.352328 12.257762
#> 366 183 2003-07-02 2.12520906 16.546336 41.288405 12.973640
#> 368 184 2003-07-03 2.49644591 16.526900 41.218909 13.417857
#> 369 185 2003-07-04 1.19388600 16.505823 41.143868 11.773960
#> 372 186 2003-07-05 0.69951172 16.483128 41.063311 11.137151
#> 374 187 2003-07-06 2.28776717 16.458837 40.977269 13.092224
#> 376 188 2003-07-07 5.26572827 16.432973 40.885775 16.772096
#> 377 189 2003-07-08 6.53851552 16.405563 40.788864 18.325514
#> 379 190 2003-07-09 1.39705038 16.376632 40.686570 11.907078
#> 382 191 2003-07-10 1.84424429 16.346207 40.578932 12.433872
#> 384 192 2003-07-11 4.94156660 16.314316 40.465990 16.245022
#> 386 193 2003-07-12 0.19890206 16.280989 40.347784 10.333407
#> 387 194 2003-07-13 2.05795845 16.246255 40.224356 12.603755
#> 390 195 2003-07-14 3.73312691 16.210144 40.095751 14.640878
#> 392 196 2003-07-15 2.85135033 16.172686 39.962013 13.513285
#> 394 197 2003-07-16 6.59777443 16.133915 39.823189 18.098409
#> 395 198 2003-07-17 4.70372819 16.093860 39.679327 15.718340
#> 398 199 2003-07-18 6.97611070 16.052554 39.530476 18.472186
#> 400 200 2003-07-19 2.56679318 16.010029 39.376687 13.000688
#> 402 201 2003-07-20 5.40962197 15.966317 39.218012 16.448321
#> 403 202 2003-07-21 2.93909647 15.921451 39.054503 13.368352
#> 406 203 2003-07-22 3.22650038 15.875462 38.886214 13.673135
#> 407 204 2003-07-23 1.42578058 15.828384 38.713200 11.421893
#> 410 205 2003-07-24 6.55543364 15.780249 38.535519 17.638095
#> 412 206 2003-07-25 3.50527372 15.731088 38.353226 13.861328
#> 414 207 2003-07-26 4.93253770 15.680933 38.166382 15.544335
#> 415 208 2003-07-27 4.85668313 15.629816 37.975044 15.393791
#> 418 209 2003-07-28 4.41609321 15.577768 37.779274 14.799783
#> 420 210 2003-07-29 0.68893896 15.524820 37.579134 10.228601
#> 422 211 2003-07-30 1.12687208 15.471001 37.374685 10.704814
#> 423 212 2003-07-31 0.93110720 15.416342 37.165992 10.413863
#> 426 213 2003-08-01 2.25281311 15.360873 36.953119 11.948037
#> 428 214 2003-08-02 6.56854166 15.304621 36.736132 17.067365
#> 430 215 2003-08-03 3.69485156 15.247616 36.515097 13.553002
#> 431 216 2003-08-04 1.87109866 15.189886 36.290082 11.307637
#> 433 217 2003-08-05 3.11094845 15.131457 36.061156 12.722281
#> 436 218 2003-08-06 3.50953046 15.072356 35.828388 13.128336
#> 438 219 2003-08-07 2.43001206 15.012610 35.591848 11.778495
#> 440 220 2003-08-08 2.71502714 14.952243 35.351609 12.047474
#> 441 221 2003-08-09 0.77817075 14.891281 35.107744  9.694245
#> 444 222 2003-08-10 6.45261387 14.829748 34.860325 16.299169
#> 446 223 2003-08-11 0.74547101 14.767666 34.609429  9.525898
#> 448 224 2003-08-12 1.05421952 14.705060 34.355129  9.820258
#> 449 225 2003-08-13 2.19809743 14.641952 34.097505 11.083790
#> 452 226 2003-08-14 2.41664170 14.578362 33.836632 11.263692
#> 454 227 2003-08-15 5.08302336 14.514313 33.572591 14.271837
#> 456 228 2003-08-16 4.13123194 14.449824 33.305462 13.087413
#> 457 229 2003-08-17 1.84628859 14.384916 33.035326 10.378856
#> 459 230 2003-08-18 6.93109153 14.319607 32.762264 16.119493
#> 462 231 2003-08-19 5.83289684 14.253916 32.486361 14.768521
#> 464 232 2003-08-20 3.82599057 14.187862 32.207701 12.394594
#> 466 233 2003-08-21 5.28072873 14.121461 31.926370 13.951035
#> 467 234 2003-08-22 3.98910814 14.054730 31.642455 12.401101
#> 470 235 2003-08-23 1.53539459 13.987686 31.356043  9.559949
#> 472 236 2003-08-24 0.73824277 13.920345 31.067223  8.590605
#> 474 237 2003-08-25 5.41833280 13.852721 30.776085 13.712877
#> 475 238 2003-08-26 0.76829816 13.784830 30.482721  8.470158
#> 478 239 2003-08-27 5.55815313 13.716686 30.187223 13.662905
#> 480 240 2003-08-28 4.01092762 13.648302 29.889684 11.864373
#> 482 241 2003-08-29 5.45376630 13.579692 29.590199 13.339438
#> 483 242 2003-08-30 4.83188715 13.510869 29.288864 12.559499
#> 486 243 2003-08-31 3.93609592 13.441846 28.985774 11.490309
#> 488 244 2003-09-01 4.67327780 13.372633 28.681027 12.181775
#> 489 245 2003-09-02 5.00060067 13.303244 28.374722 12.426615
#> 492 246 2003-09-03 4.81411165 13.233689 28.066958 12.121796
#> 494 247 2003-09-04 3.58384362 13.163980 27.757836 10.717942
#> 496 248 2003-09-05 6.93351707 13.094126 27.447457 14.128765
#> 497 249 2003-09-06 6.24346730 13.024138 27.135924 13.288144
#> 500 250 2003-09-07 6.30522871 12.954027 26.823338 13.233815
#> 502 251 2003-09-08 2.30568382 12.883802 26.509805  8.999547
#> 504 252 2003-09-09 2.37312208 12.813472 26.195428  8.974622
#> 505 253 2003-09-10 6.41750228 12.743047 25.880314 12.986847
#> 508 254 2003-09-11 4.30558717 12.672536 25.564567 10.734016
#> 510 255 2003-09-12 0.41705192 12.601947 25.248295  6.729860
#> 512 256 2003-09-13 1.88001107 12.531289 24.931604  8.103087
#> 513 257 2003-09-14 6.75448884 12.460572 24.614602 12.825056
#> 515 258 2003-09-15 5.96518036 12.389802 24.297397 11.923448
#> 518 259 2003-09-16 4.23597003 12.318988 23.980097 10.117886
#> 520 260 2003-09-17 0.84445755 12.248139 23.662812  6.731428
#> 522 261 2003-09-18 2.69712201 12.177262 23.345649  8.421807
#> 523 262 2003-09-19 2.65048756 12.106365 23.028718  8.278057
#> 526 263 2003-09-20 2.12256496 12.035456 22.712128  7.680780
#> 528 264 2003-09-21 0.96791049 11.964544 22.395989  6.504895
#> 530 265 2003-09-22 4.30259115 11.893635 22.080409  9.513960
#> 531 266 2003-09-23 3.77226017 11.822738 21.765497  8.913714
#> 534 267 2003-09-24 6.22374461 11.751861 21.451363 11.043124
#> 536 268 2003-09-25 5.57640106 11.681012 21.138114 10.330093
#> 538 269 2003-09-26 1.99081192 11.610198 20.825859  6.991980
#> 539 270 2003-09-27 3.68356425 11.539428 20.514705  8.402982
#> 541 271 2003-09-28 5.92767766 11.468711 20.204761 10.272671
#> 544 272 2003-09-29 5.36367648 11.398053 19.896131  9.655377
#> 546 273 2003-09-30 1.20286877 11.327464 19.588923  5.937309
#> 548 274 2003-10-01 2.46615531 11.256953 19.283241  6.933081
#> 549 275 2003-10-02 5.38321367 11.186528 18.979189  9.311408
#> 552 276 2003-10-03 2.97672628 11.116198 18.676871  7.169890
#> 554 277 2003-10-04 3.12711917 11.045973 18.376389  7.195279
#> 556 278 2003-10-05 6.14123372 10.975862 18.077844  9.576935
#> 557 279 2003-10-06 1.26902670 10.905874 17.781337  5.479868
#> 560 280 2003-10-07 6.16404388 10.836020 17.486965  9.345450
#> 562 281 2003-10-08 6.23210676 10.766311 17.194826  9.275341
#> 564 282 2003-10-09 5.09742295 10.696756 16.905016  8.254205
#> 565 283 2003-10-10 1.93867145 10.627367 16.617629  5.670123
#> 568 284 2003-10-11 3.70066513 10.558154 16.332759  6.945530
#> 569 285 2003-10-12 3.79230540 10.489131 16.050495  6.914122
#> 572 286 2003-10-13 5.27154887 10.420308 15.770928  7.931924
#> 574 287 2003-10-14 2.69135938 10.351698 15.494145  5.887714
#> 576 288 2003-10-15 5.86535115 10.283314 15.220231  8.145682
#> 577 289 2003-10-16 3.11901290 10.215170 14.949271  6.019559
#> 580 290 2003-10-17 3.87092373 10.147279 14.681345  6.470613
#> 582 291 2003-10-18 5.13759945 10.079655 14.416534  7.278187
#> 584 292 2003-10-19 3.27283361 10.012314 14.154915  5.852214
#> 585 293 2003-10-20 1.56437010  9.945270 13.896562  4.567091
#> 588 294 2003-10-21 3.83007683  9.878539 13.641550  6.054917
#> 590 295 2003-10-22 4.62640734  9.812138 13.389948  6.504156
#> 592 296 2003-10-23 4.47015012  9.746084 13.141825  6.299279
#> 593 297 2003-10-24 4.39361061  9.680393 12.897247  6.151129
#> 595 298 2003-10-25 5.59970454  9.615084 12.656277  6.849498
#> 598 299 2003-10-26 6.75185071  9.550176 12.418977  7.494772
#> 600 300 2003-10-27 2.23886120  9.485687 12.185404  4.484382
#> 602 301 2003-10-28 1.66044273  9.421638 11.955615  4.042416
#> 603 302 2003-10-29 4.66764785  9.358048 11.729663  5.857702
#> 606 303 2003-10-30 4.29079977  9.294940 11.507600  5.533012
#> 608 304 2003-10-31 0.19196456  9.232334 11.289473  2.939737
#> 610 305 2003-11-01 6.34958995  9.170252 11.075328  6.603176
#> 611 306 2003-11-02 3.58920569  9.108719 10.865209  4.856970
#> 614 307 2003-11-03 6.87503403  9.047757 10.659156  6.714526
#> 616 308 2003-11-04 6.20001596  8.987390 10.457208  6.221292
#> 618 309 2003-11-05 0.83706620  8.927644 10.259398  3.045816
#> 619 310 2003-11-06 5.39404287  8.868543 10.065762  5.577549
#> 621 311 2003-11-07 1.38495433  8.810114  9.876328  3.245364
#> 624 312 2003-11-08 3.58573787  8.752384  9.691126  4.407946
#> 626 313 2003-11-09 5.21425166  8.695379  9.510180  5.228972
#> 628 314 2003-11-10 6.00126653  8.639127  9.333513  5.575193
#> 629 315 2003-11-11 0.87864148  8.583658  9.161147  2.759164
#> 632 316 2003-11-12 1.82691784  8.528999  8.993100  3.211439
#> 634 317 2003-11-13 5.98673457  8.475180  8.829388  5.325818
#> 636 318 2003-11-14 6.72850404  8.422232  8.670025  5.630739
#> 637 319 2003-11-15 3.07742819  8.370184  8.515023  3.694096
#> 640 320 2003-11-16 2.06820487  8.319067  8.364392  3.130835
#> 642 321 2003-11-17 4.02135139  8.268912  8.218139  4.052864
#> 644 322 2003-11-18 0.74209443  8.219751  8.076271  2.383638
#> 645 323 2003-11-19 5.23357406  8.171616  7.938791  4.526927
#> 648 324 2003-11-20 6.07489372  8.124538  7.805702  4.869672
#> 649 325 2003-11-21 2.03353238  8.078549  7.677004  2.885479
#> 652 326 2003-11-22 0.12693248  8.033683  7.552698  1.947841
#> 654 327 2003-11-23 5.00167573  7.989971  7.432780  4.184634
#> 656 328 2003-11-24 1.61122478  7.947446  7.317248  2.571043
#> 657 329 2003-11-25 1.77097829  7.906140  7.206095  2.608608
#> 660 330 2003-11-26 3.64815790  7.866085  7.099317  3.421101
#> 662 331 2003-11-27 5.92679783  7.827314  6.996907  4.398236
#> 664 332 2003-11-28 5.87129415  7.789856  6.898857  4.324583
#> 665 333 2003-11-29 2.57193307  7.753745  6.805157  2.829932
#> 668 334 2003-11-30 2.40262695  7.719011  6.715800  2.724133
#> 670 335 2003-12-01 3.15807119  7.685684  6.630775  3.019997
#> 672 336 2003-12-02 2.28539523  7.653793  6.550072  2.615432
#> 673 337 2003-12-03 0.61902366  7.623368  6.473680  1.881254
#> 675 338 2003-12-04 4.87497517  7.594437  6.401590  3.655032
#> 678 339 2003-12-05 6.51847878  7.567027  6.333789  4.311512
#> 680 340 2003-12-06 6.56965046  7.541163  6.270268  4.298808
#> 682 341 2003-12-07 2.26203499  7.516872  6.211016  2.487288
#> 683 342 2003-12-08 0.06695332  7.494177  6.156023  1.566505
#> 686 343 2003-12-09 4.02899126  7.473100  6.105278  3.172096
#> 688 344 2003-12-10 3.10034322  7.453664  6.058772  2.774763
#> 690 345 2003-12-11 3.14021991  7.435887  6.016497  2.774526
#> 691 346 2003-12-12 2.41808780  7.419789  5.978443  2.468789
#> 694 347 2003-12-13 0.25340833  7.405386  5.944603  1.587861
#> 696 348 2003-12-14 0.02391030  7.392694  5.914970  1.488308
#> 698 349 2003-12-15 3.87203651  7.381725  5.889539  3.017044
#> 699 350 2003-12-16 5.66566520  7.372493  5.868303  3.721933
#> 701 351 2003-12-17 1.57932001  7.365006  5.851259  2.090174
#> 704 352 2003-12-18 4.96476662  7.359273  5.838404  3.428975
#> 706 353 2003-12-19 2.60354924  7.355300  5.829736  2.489207
#> 708 354 2003-12-20 0.61859034  7.353091  5.825253  1.701343
#> 709 355 2003-12-21 6.60801581  7.352649  5.824956  4.073758
#> 712 356 2003-12-22 5.02265773  7.353975  5.828845  3.447720
#> 714 357 2003-12-23 6.33307338  7.357066  5.836924  3.971487
#> 716 358 2003-12-24 2.73019727  7.361920  5.849194  2.546897
#> 717 359 2003-12-25 2.85629146  7.368531  5.865662  2.603280
#> 720 360 2003-12-26 3.28382834  7.376892  5.886331  2.781735
#> 722 361 2003-12-27 3.32164519  7.386993  5.911209  2.806823
#> 724 362 2003-12-28 2.84064925  7.398825  5.940303  2.625413
#> 725 363 2003-12-29 6.74354469  7.412375  5.973622  4.210712
#> 728 364 2003-12-30 2.82357060  7.427627  6.011174  2.645351
#> 730 365 2003-12-31 5.93436651  7.444567  6.052971  3.925777