Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.090933  2.621345  1.917570  4.221333  4.334474  3.311883  3.200370
#>   [8]  2.714313  3.008997  4.514264  3.205185  2.216427  2.095431  2.844366
#>  [15]  1.966723  2.035575  3.365371  5.169421  2.843768  2.836268  2.627307
#>  [22]  2.578911  4.310719  2.412539  3.666705  2.816551  4.008806  2.846938
#>  [29]  3.407432  5.651612  4.007026  3.050680  4.983045  3.868562  3.051597
#>  [36]  6.473327  5.739545  5.851617  4.402528  2.871422  3.691513  5.782472
#>  [43]  3.411512  5.949342  3.770910  7.104099  4.242180  4.363202  4.656177
#>  [50]  7.635960  7.020598  8.209435  5.144904  3.974738  8.206316  6.165984
#>  [57]  6.191032  8.821952  5.987524  4.477339  7.059613  9.106812  8.056612
#>  [64]  8.279206 10.180905  5.306371  7.942574  5.690571  7.338291  5.436595
#>  [71] 10.967338  6.523901 10.360434 10.249573 11.523865  8.343089  6.258107
#>  [78]  7.410859  9.078461 10.770625 12.220990  7.882357 11.884579  9.337651
#>  [85]  6.908521  9.113673  7.514147 11.613233 13.111722 12.191785 12.495588
#>  [92] 11.441156 13.410370 12.200545 10.294512 14.340476  8.304698 11.113809
#>  [99] 14.198682  8.734032 11.049767 12.276184  7.759731  8.773776  9.341004
#> [106] 10.774628  9.601118  8.371645 10.371856  9.141526 13.010295 12.893714
#> [113] 14.658649 12.358415 15.820047 10.501433 12.426560 14.667865 10.195863
#> [120] 10.640434 15.939508 11.398569 11.954079 10.493009 16.342426 15.667988
#> [127] 11.841886 15.256801 11.908278 16.982792 14.963350 13.235020 13.712184
#> [134] 13.440816 16.845573 15.206236 18.119545 10.113283 10.467355 10.921182
#> [141] 18.044993 16.267990 13.482198 13.111118 17.788112 13.507906 16.771834
#> [148] 11.693444 13.280247 16.882129 15.152296 16.189458 17.620272 18.585996
#> [155] 11.561669 14.662163 11.303081 15.148550 11.931577 18.413896 16.211007
#> [162] 15.619885 17.203349 12.767709 18.674747 12.332618 12.614319 16.532073
#> [169] 10.522954 14.436142 13.471858 12.219125 12.515162 18.125450 15.425807
#> [176] 14.505891 16.338029 15.734540 11.193224 18.895013 18.558959 10.794649
#> [183] 11.668685 11.096143 16.155979 11.517245 12.148187 18.018160 12.197013
#> [190] 13.741328 14.862936 14.222430 10.576576 18.652037 10.961059 10.682448
#> [197] 14.522409 14.722474 16.684417 10.401218 18.153278 11.599246 17.112373
#> [204] 11.485160 17.251185 11.371378 13.735948 13.820621 14.734447 10.360819
#> [211] 10.991508 15.537586 16.194486 10.763662 16.813338 10.238082 13.796920
#> [218]  9.701263 12.483253 12.075194 11.682364  9.657508 16.713356 10.905415
#> [225] 14.882993 10.720435  9.642822  9.149758 16.005335 13.978529  9.129568
#> [232] 12.920577 12.585044 11.862517 11.453710 10.314777 11.981045 11.329812
#> [239] 11.535025 13.359035  8.767588 10.264801  8.407089  7.369006 13.770458
#> [246] 12.352143 13.480586 10.374390 12.292995  6.904522 13.043911  7.821641
#> [253] 11.213471 10.798581  8.548032  7.874748 11.529543  7.962646  9.672257
#> [260] 10.084307  6.208985  8.587597  7.892748 10.185606  9.433677  8.098568
#> [267]  9.375987 11.143872 11.067518 11.048043 10.770014 10.570065  9.977180
#> [274]  7.753656  8.396307  5.965499  9.252656  7.427250  7.678317  8.815669
#> [281]  6.342653  4.329270  7.338992  8.698230  9.118671  9.072982  4.883165
#> [288]  4.957508  6.911943  8.403109  7.314838  8.440428  7.956064  4.959520
#> [295]  4.211147  5.725999  7.312753  3.181870  5.483240  5.107623  6.772257
#> [302]  7.113793  4.500631  4.626201  4.074140  3.445936  3.267701  6.247198
#> [309]  3.645607  4.174307  3.610676  5.266701  5.141076  3.943969  5.420876
#> [316]  4.015908  2.260260  2.435380  3.284691  3.808380  2.487564  2.055548
#> [323]  3.971426  2.969072  3.507587  3.881824  3.911007  3.327619  2.376063
#> [330]  2.134393  1.879726  4.823121  2.431756  2.155620  2.929776  3.844653
#> [337]  2.699469  3.023564  3.489613  1.878755  4.426798  3.548351  4.239191
#> [344]  3.879394  3.338439  2.742394  4.251742  3.022783  2.191584  2.546418
#> [351]  2.659178  3.924758  3.343196  1.590720  1.816784  2.947983  3.984202
#> [358]  2.527272  1.707010  3.580546  2.146110  2.384359  1.530091  3.799923
#> [365]  4.101614  2.090933  2.621345  1.917570  4.221333  4.334474  3.311883
#> [372]  3.200370  2.714313  3.008997  4.514264  3.205185  2.216427  2.095431
#> [379]  2.844366  1.966723  2.035575  3.365371  5.169421  2.843768  2.836268
#> [386]  2.627307  2.578911  4.310719  2.412539  3.666705  2.816551  4.008806
#> [393]  2.846938  3.407432  5.651612  4.007026  3.050680  4.983045  3.868562
#> [400]  3.051597  6.473327  5.739545  5.851617  4.402528  2.871422  3.691513
#> [407]  5.782472  3.411512  5.949342  3.770910  7.104099  4.242180  4.363202
#> [414]  4.656177  7.635960  7.020598  8.209435  5.144904  3.974738  8.206316
#> [421]  6.165984  6.191032  8.821952  5.987524  4.477339  7.059613  9.106812
#> [428]  8.056612  8.279206 10.180905  5.306371  7.942574  5.690571  7.338291
#> [435]  5.436595 10.967338  6.523901 10.360434 10.249573 11.523865  8.343089
#> [442]  6.258107  7.410859  9.078461 10.770625 12.220990  7.882357 11.884579
#> [449]  9.337651  6.908521  9.113673  7.514147 11.613233 13.111722 12.191785
#> [456] 12.495588 11.441156 13.410370 12.200545 10.294512 14.340476  8.304698
#> [463] 11.113809 14.198682  8.734032 11.049767 12.276184  7.759731  8.773776
#> [470]  9.341004 10.774628  9.601118  8.371645 10.371856  9.141526 13.010295
#> [477] 12.893714 14.658649 12.358415 15.820047 10.501433 12.426560 14.667865
#> [484] 10.195863 10.640434 15.939508 11.398569 11.954079 10.493009 16.342426
#> [491] 15.667988 11.841886 15.256801 11.908278 16.982792 14.963350 13.235020
#> [498] 13.712184 13.440816 16.845573 15.206236 18.119545 10.113283 10.467355
#> [505] 10.921182 18.044993 16.267990 13.482198 13.111118 17.788112 13.507906
#> [512] 16.771834 11.693444 13.280247 16.882129 15.152296 16.189458 17.620272
#> [519] 18.585996 11.561669 14.662163 11.303081 15.148550 11.931577 18.413896
#> [526] 16.211007 15.619885 17.203349 12.767709 18.674747 12.332618 12.614319
#> [533] 16.532073 10.522954 14.436142 13.471858 12.219125 12.515162 18.125450
#> [540] 15.425807 14.505891 16.338029 15.734540 11.193224 18.895013 18.558959
#> [547] 10.794649 11.668685 11.096143 16.155979 11.517245 12.148187 18.018160
#> [554] 12.197013 13.741328 14.862936 14.222430 10.576576 18.652037 10.961059
#> [561] 10.682448 14.522409 14.722474 16.684417 10.401218 18.153278 11.599246
#> [568] 17.112373 11.485160 17.251185 11.371378 13.735948 13.820621 14.734447
#> [575] 10.360819 10.991508 15.537586 16.194486 10.763662 16.813338 10.238082
#> [582] 13.796920  9.701263 12.483253 12.075194 11.682364  9.657508 16.713356
#> [589] 10.905415 14.882993 10.720435  9.642822  9.149758 16.005335 13.978529
#> [596]  9.129568 12.920577 12.585044 11.862517 11.453710 10.314777 11.981045
#> [603] 11.329812 11.535025 13.359035  8.767588 10.264801  8.407089  7.369006
#> [610] 13.770458 12.352143 13.480586 10.374390 12.292995  6.904522 13.043911
#> [617]  7.821641 11.213471 10.798581  8.548032  7.874748 11.529543  7.962646
#> [624]  9.672257 10.084307  6.208985  8.587597  7.892748 10.185606  9.433677
#> [631]  8.098568  9.375987 11.143872 11.067518 11.048043 10.770014 10.570065
#> [638]  9.977180  7.753656  8.396307  5.965499  9.252656  7.427250  7.678317
#> [645]  8.815669  6.342653  4.329270  7.338992  8.698230  9.118671  9.072982
#> [652]  4.883165  4.957508  6.911943  8.403109  7.314838  8.440428  7.956064
#> [659]  4.959520  4.211147  5.725999  7.312753  3.181870  5.483240  5.107623
#> [666]  6.772257  7.113793  4.500631  4.626201  4.074140  3.445936  3.267701
#> [673]  6.247198  3.645607  4.174307  3.610676  5.266701  5.141076  3.943969
#> [680]  5.420876  4.015908  2.260260  2.435380  3.284691  3.808380  2.487564
#> [687]  2.055548  3.971426  2.969072  3.507587  3.881824  3.911007  3.327619
#> [694]  2.376063  2.134393  1.879726  4.823121  2.431756  2.155620  2.929776
#> [701]  3.844653  2.699469  3.023564  3.489613  1.878755  4.426798  3.548351
#> [708]  4.239191  3.879394  3.338439  2.742394  4.251742  3.022783  2.191584
#> [715]  2.546418  2.659178  3.924758  3.343196  1.590720  1.816784  2.947983
#> [722]  3.984202  2.527272  1.707010  3.580546  2.146110  2.384359  1.530091
#> [729]  3.799923  4.101614
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates    sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 4.611262846  7.463176  6.099033  3.408959
#> 3     2 2002-01-02 2.386307218  7.483435  6.149345  2.517783
#> 6     3 2002-01-03 1.854255885  7.505323  6.203934  2.317352
#> 7     4 2002-01-04 3.522667380  7.528820  6.262814  3.030861
#> 9     5 2002-01-05 5.714403047  7.553900  6.325996  3.974256
#> 11    6 2002-01-06 5.406210926  7.580540  6.393496  3.878197
#> 14    7 2002-01-07 1.273838189  7.608714  6.465324  2.157538
#> 15    8 2002-01-08 4.064626473  7.638396  6.541496  3.375840
#> 17    9 2002-01-09 5.691023512  7.669557  6.622023  4.112368
#> 19   10 2002-01-10 2.480866319  7.702170  6.706920  2.756878
#> 22   11 2002-01-11 4.321027229  7.736204  6.796198  3.597045
#> 23   12 2002-01-12 0.151938202  7.771631  6.889871  1.789818
#> 25   13 2002-01-13 0.835534656  7.808419  6.987950  2.120858
#> 27   14 2002-01-14 2.808012696  7.846537  7.090446  3.041328
#> 30   15 2002-01-15 2.133019948  7.885954  7.197371  2.772727
#> 32   16 2002-01-16 2.279468778  7.926639  7.308732  2.878071
#> 33   17 2002-01-17 3.899005742  7.968559  7.424539  3.672544
#> 35   18 2002-01-18 6.768910912  8.011681  7.544799  5.073426
#> 37   19 2002-01-19 1.500975901  8.055974  7.669519  2.631866
#> 40   20 2002-01-20 1.492200088  8.101405  7.798703  2.667898
#> 41   21 2002-01-21 4.126250145  8.147942  7.932354  3.991625
#> 43   22 2002-01-22 1.453380998  8.195553  8.070474  2.733218
#> 45   23 2002-01-23 3.900548216  8.244205  8.213063  3.996173
#> 48   24 2002-01-24 6.743044733  8.293867  8.360119  5.488484
#> 49   25 2002-01-25 5.623498152  8.344507  8.511639  4.995976
#> 51   26 2002-01-26 4.464458765  8.396093  8.667618  4.471323
#> 53   27 2002-01-27 2.711565692  8.448595  8.828046  3.623687
#> 56   28 2002-01-28 5.521823498  8.501983  8.992915  5.168565
#> 58   29 2002-01-29 2.221688779  8.556225  9.162213  3.480072
#> 59   30 2002-01-30 5.822266832  8.611293  9.335925  5.490084
#> 61   31 2002-01-31 4.953305518  8.667157  9.514035  5.097158
#> 63   32 2002-02-01 3.406662119  8.723789  9.696523  4.317390
#> 66   33 2002-02-02 4.043295227  8.781160  9.883368  4.746246
#> 67   34 2002-02-03 5.198640356  8.839243 10.074546  5.481217
#> 69   35 2002-02-04 0.863918277  8.898011 10.270030  3.066072
#> 71   36 2002-02-05 6.360066342  8.957438 10.469790  6.334390
#> 74   37 2002-02-06 6.885087076  9.017498 10.673795  6.743305
#> 75   38 2002-02-07 1.210778576  9.078165 10.882010  3.446184
#> 77   39 2002-02-08 4.236163078  9.139416 11.094398  5.344753
#> 79   40 2002-02-09 6.028568071  9.201226 11.310918  6.533140
#> 82   41 2002-02-10 0.347101070  9.263572 11.531528  3.098922
#> 83   42 2002-02-11 2.309723030  9.326432 11.756182  4.394775
#> 86   43 2002-02-12 0.609958315  9.389784 11.984831  3.385474
#> 87   44 2002-02-13 1.447692496  9.453606 12.217425  3.989823
#> 89   45 2002-02-14 0.095307771  9.517878 12.453910  3.175831
#> 91   46 2002-02-15 3.260743607  9.582579 12.694229  5.333342
#> 94   47 2002-02-16 0.204662514  9.647690 12.938322  3.371815
#> 95   48 2002-02-17 1.668058059  9.713192 13.186128  4.428767
#> 97   49 2002-02-18 4.269067556  9.779067 13.437583  6.292495
#> 99   50 2002-02-19 5.574544707  9.845297 13.692618  7.299630
#> 102  51 2002-02-20 2.159777722  9.911864 13.951164  5.007758
#> 103  52 2002-02-21 3.910690197  9.978754 14.213148  6.338365
#> 105  53 2002-02-22 6.609909837 10.045948 14.478496  8.382816
#> 107  54 2002-02-23 1.303444650 10.113433 14.747130  4.637106
#> 110  55 2002-02-24 6.761473027 10.181192 15.018971  8.741898
#> 112  56 2002-02-25 0.497533437 10.249211 15.293936  4.194695
#> 113  57 2002-02-26 6.409647314 10.317477 15.571940  8.729955
#> 115  58 2002-02-27 2.635493300 10.385975 15.852898  5.974601
#> 117  59 2002-02-28 4.259933306 10.454693 16.136720  7.321763
#> 120  60 2002-03-01 1.777947329 10.523618 16.423316  5.493174
#> 121  61 2002-03-02 4.409067284 10.592738 16.712592  7.656330
#> 123  62 2002-03-03 2.102007400 10.662040 17.004454  5.927316
#> 125  63 2002-03-04 5.877780087 10.731513 17.298805  9.062084
#> 128  64 2002-03-05 6.439903690 10.801147 17.595546  9.644331
#> 129  65 2002-03-06 3.594012135 10.870930 17.894578  7.431686
#> 131  66 2002-03-07 0.041286347 10.940852 18.195799  4.583281
#> 133  67 2002-03-08 1.017528177 11.010902 18.499104  5.479536
#> 136  68 2002-03-09 2.072047547 11.081072 18.804391  6.459213
#> 138  69 2002-03-10 6.258400836 11.151350 19.111552 10.140815
#> 139  70 2002-03-11 6.713639717 11.221729 19.420480 10.664479
#> 141  71 2002-03-12 2.571600291 11.292198 19.731067  7.179469
#> 143  72 2002-03-13 3.217142902 11.362749 20.043203  7.848224
#> 146  73 2002-03-14 4.023985261 11.433374 20.356778  8.671487
#> 147  74 2002-03-15 4.092863218 11.504062 20.671679  8.845158
#> 149  75 2002-03-16 2.154470579 11.574807 20.987796  7.200225
#> 151  76 2002-03-17 0.983922642 11.645600 21.305014  6.226271
#> 154  77 2002-03-18 1.154630782 11.716433 21.623220  6.471267
#> 155  78 2002-03-19 0.688907004 11.787297 21.942300  6.126782
#> 157  79 2002-03-20 1.115722286 11.858185 22.262140  6.612844
#> 159  80 2002-03-21 1.942975181 11.929088 22.582624  7.484752
#> 162  81 2002-03-22 5.838376877 12.000000 22.903638 11.297579
#> 163  82 2002-03-23 1.224851055 12.070912 23.225065  6.984605
#> 166  83 2002-03-24 2.806900965 12.141815 23.546791  8.608429
#> 167  84 2002-03-25 1.436214926 12.212703 23.868700  7.370656
#> 169  85 2002-03-26 5.390462664 12.283567 24.190677 11.355531
#> 171  86 2002-03-27 0.591139893 12.354400 24.512606  6.714598
#> 174  87 2002-03-28 3.360516214 12.425193 24.834373  9.566944
#> 175  88 2002-03-29 6.326688315 12.495938 25.155863 12.657167
#> 177  89 2002-03-30 6.540529131 12.566626 25.476961 12.999214
#> 179  90 2002-03-31 5.979488024 12.637251 25.797553 12.552621
#> 182  91 2002-04-01 5.083456849 12.707802 26.117528 11.753233
#> 183  92 2002-04-02 6.997491852 12.778271 26.436771 13.847695
#> 185  93 2002-04-03 1.936760237 12.848650 26.755170  8.705283
#> 187  94 2002-04-04 6.929281596 12.918928 27.072615 14.028577
#> 190  95 2002-04-05 6.591813325 12.989098 27.388994 13.797044
#> 192  96 2002-04-06 1.950361753 13.059148 27.704199  8.994837
#> 193  97 2002-04-07 3.684649445 13.129070 28.018120 10.936147
#> 195  98 2002-04-08 2.405912317 13.198853 28.330649  9.664745
#> 197  99 2002-04-09 3.531690320 13.268487 28.641680 10.972216
#> 200 100 2002-04-10 3.248816334 13.337960 28.951106 10.763684
#> 201 101 2002-04-11 6.556543378 13.407262 29.258822 14.468915
#> 203 102 2002-04-12 6.162946355 13.476382 29.564726 14.151372
#> 205 103 2002-04-13 0.920085492 13.545307 29.868714  8.481617
#> 208 104 2002-04-14 4.973068039 13.614025 30.170685 13.053197
#> 209 105 2002-04-15 1.656187188 13.682523 30.470539  9.461773
#> 211 106 2002-04-16 6.680489243 13.750789 30.768176 15.166032
#> 213 107 2002-04-17 2.668783003 13.818808 31.063500 10.765473
#> 216 108 2002-04-18 6.092803516 13.886567 31.356413 14.717998
#> 218 109 2002-04-19 2.276160974 13.954052 31.646820 10.492792
#> 219 110 2002-04-20 4.622140738 14.021246 31.934628 13.247324
#> 221 111 2002-04-21 1.891243824 14.088136 32.219743 10.217585
#> 223 112 2002-04-22 1.328978937 14.154703 32.502075  9.651322
#> 226 113 2002-04-23 3.705833357 14.220933 32.781534 12.466654
#> 227 114 2002-04-24 6.367192212 14.286808 33.058030 15.630982
#> 229 115 2002-04-25 2.741436963 14.352310 33.331477 11.516195
#> 231 116 2002-04-26 4.260394927 14.417421 33.601789 13.365166
#> 234 117 2002-04-27 4.542204761 14.482122 33.868881 13.778576
#> 235 118 2002-04-28 1.187684727 14.546394 34.132670  9.926601
#> 237 119 2002-04-29 3.721117721 14.610216 34.393074 12.978104
#> 239 120 2002-04-30 6.463157263 14.673568 34.650012 16.293520
#> 242 121 2002-05-01 2.262920687 14.736428 34.903405 11.405729
#> 243 122 2002-05-02 0.123758806 14.798774 35.153175  8.935283
#> 245 123 2002-05-03 0.851378916 14.860584 35.399245  9.863842
#> 248 124 2002-05-04 6.972022764 14.921835 35.641540 17.236896
#> 249 125 2002-05-05 0.898510223 14.982502 35.879987 10.045869
#> 251 126 2002-05-06 5.613266329 15.042562 36.114511 15.766854
#> 253 127 2002-05-07 2.257340275 15.101989 36.345041 11.802562
#> 256 128 2002-05-08 2.606661847 15.160757 36.571508 12.286835
#> 257 129 2002-05-09 3.680756183 15.218840 36.793843 13.647852
#> 259 130 2002-05-10 3.938710346 15.276211 37.011977 14.024447
#> 261 131 2002-05-11 5.257843484 15.332843 37.225844 15.689089
#> 264 132 2002-05-12 3.500852155 15.388707 37.435378 13.617023
#> 265 133 2002-05-13 6.275364066 15.443775 37.640517 17.057481
#> 267 134 2002-05-14 4.392007902 15.498017 37.841197 14.822238
#> 269 135 2002-05-15 4.517446551 15.551405 38.037357 15.033976
#> 272 136 2002-05-16 2.887461698 15.603907 38.228936 13.094316
#> 274 137 2002-05-17 2.913481535 15.655493 38.415875 13.178559
#> 275 138 2002-05-18 3.167894837 15.706133 38.598118 13.542110
#> 277 139 2002-05-19 6.536948200 15.755795 38.775606 17.737740
#> 279 140 2002-05-20 0.045705891 15.804447 38.948286  9.793390
#> 282 141 2002-05-21 4.055711838 15.852058 39.116103 14.782908
#> 283 142 2002-05-22 5.891203906 15.898595 39.279005 17.097144
#> 285 143 2002-05-23 2.203847506 15.944026 39.436940 12.584802
#> 287 144 2002-05-24 5.392472057 15.988319 39.589859 16.573814
#> 290 145 2002-05-25 4.914849321 16.031441 39.737713 16.025736
#> 291 146 2002-05-26 1.009311622 16.073361 39.880455 11.222242
#> 293 147 2002-05-27 2.251410034 16.114046 40.018039 12.800115
#> 295 148 2002-05-28 2.637081859 16.153463 40.150421 13.314919
#> 298 149 2002-05-29 6.185129004 16.191581 40.277558 17.762334
#> 300 150 2002-05-30 3.266295813 16.228369 40.399408 14.165461
#> 301 151 2002-05-31 2.333219598 16.263796 40.515933 13.035210
#> 303 152 2002-06-01 1.161030555 16.297830 40.627093 11.603877
#> 305 153 2002-06-02 6.635071927 16.330443 40.732853 18.458108
#> 308 154 2002-06-03 0.527014628 16.361604 40.833176 10.865922
#> 309 155 2002-06-04 0.364430947 16.391286 40.928031 10.686989
#> 311 156 2002-06-05 1.514192510 16.419460 41.017385 12.145645
#> 313 157 2002-06-06 0.661726309 16.446100 41.101209 11.102178
#> 316 158 2002-06-07 3.602409440 16.471180 41.179475 14.798047
#> 317 159 2002-06-08 4.664494097 16.494677 41.252156 16.145843
#> 319 160 2002-06-09 1.806480620 16.516565 41.319229 12.589429
#> 321 161 2002-06-10 2.590266821 16.536824 41.380671 13.586025
#> 324 162 2002-06-11 4.828831229 16.555433 41.436461 16.402137
#> 325 163 2002-06-12 2.989207186 16.572373 41.486582 14.113173
#> 328 164 2002-06-13 3.990326004 16.587625 41.531016 15.378113
#> 329 165 2002-06-14 5.458190701 16.601175 41.569750 17.226159
#> 331 166 2002-06-15 0.358330484 16.613007 41.602770 10.849363
#> 333 167 2002-06-16 3.036887600 16.623108 41.630068 14.210230
#> 336 168 2002-06-17 6.322465102 16.631469 41.651634 18.329858
#> 337 169 2002-06-18 5.830726353 16.638080 41.667464 17.717936
#> 339 170 2002-06-19 3.840207535 16.642934 41.677552 15.227748
#> 341 171 2002-06-20 5.140014651 16.646025 41.681897 16.855811
#> 344 172 2002-06-21 1.985246242 16.647351 41.680500 12.905387
#> 345 173 2002-06-22 6.518643295 16.646909 41.673363 18.577627
#> 347 174 2002-06-23 6.218011558 16.644700 41.660491 18.196741
#> 349 175 2002-06-24 5.447706526 16.640727 41.641891 17.226665
#> 352 176 2002-06-25 6.627437949 16.634994 41.617572 18.694684
#> 354 177 2002-06-26 2.377887187 16.627507 41.587545 13.370588
#> 355 178 2002-06-27 5.248310023 16.618275 41.551823 16.949312
#> 357 179 2002-06-28 1.747926840 16.607306 41.510423 12.562101
#> 359 180 2002-06-29 2.613705729 16.594614 41.463360 13.631147
#> 362 181 2002-06-30 5.643160499 16.580211 41.410655 17.399829
#> 363 182 2002-07-01 0.588312812 16.564113 41.352328 11.072444
#> 365 183 2002-07-02 5.251806693 16.546336 41.288405 16.874570
#> 367 184 2002-07-03 2.459529572 16.526900 41.218909 13.371822
#> 370 185 2002-07-04 6.630615377 16.505823 41.143868 18.549995
#> 371 186 2002-07-05 6.277682149 16.483128 41.063311 18.085412
#> 373 187 2002-07-06 1.513677215 16.458837 40.977269 12.128605
#> 375 188 2002-07-07 6.206419612 16.432973 40.885775 17.942331
#> 378 189 2002-07-08 3.813768352 16.405563 40.788864 14.938269
#> 380 190 2002-07-09 3.093611901 16.376632 40.686570 14.014571
#> 381 191 2002-07-10 1.472566161 16.346207 40.578932 11.972532
#> 383 192 2002-07-11 5.804832100 16.314316 40.465990 17.315643
#> 385 193 2002-07-12 3.229059896 16.280989 40.347784 14.088098
#> 388 194 2002-07-13 1.890056796 16.246255 40.224356 12.395900
#> 389 195 2002-07-14 0.745064342 16.210144 40.095751 10.945395
#> 391 196 2002-07-15 4.768831418 16.172686 39.962013 15.882292
#> 393 197 2002-07-16 1.772030025 16.133915 39.823189 12.142740
#> 396 198 2002-07-17 3.054132832 16.093860 39.679327 13.684806
#> 397 199 2002-07-18 4.434066389 16.052554 39.530476 15.342210
#> 399 200 2002-07-19 1.201399590 16.010029 39.376687 11.321594
#> 401 201 2002-07-20 5.169396505 15.966317 39.218012 16.153289
#> 404 202 2002-07-21 1.446498648 15.921451 39.054503 11.537719
#> 405 203 2002-07-22 2.654737939 15.875462 38.886214 12.972883
#> 408 204 2002-07-23 4.179130922 15.828384 38.713200 14.788977
#> 409 205 2002-07-24 5.480576848 15.780249 38.535519 16.325690
#> 411 206 2002-07-25 1.250679998 15.731088 38.353226 11.112919
#> 413 207 2002-07-26 4.613111342 15.680933 38.166382 15.155603
#> 416 208 2002-07-27 1.059676717 15.629816 37.975044 10.781085
#> 417 209 2002-07-28 1.872143169 15.577768 37.779274 11.714984
#> 419 210 2002-07-29 6.749606356 15.524820 37.579134 17.563779
#> 421 211 2002-07-30 5.052154410 15.471001 37.374685 15.446143
#> 424 212 2002-07-31 3.915674833 15.416342 37.165992 14.011487
#> 425 213 2002-08-01 5.007207330 15.360873 36.953119 15.261112
#> 427 214 2002-08-02 4.161575598 15.304621 36.736132 14.178609
#> 429 215 2002-08-03 2.867791035 15.247616 36.515097 12.562677
#> 432 216 2002-08-04 1.557822797 15.189886 36.290082 10.933414
#> 434 217 2002-08-05 4.808616636 15.131457 36.061156 14.745215
#> 435 218 2002-08-06 6.962277524 15.072356 35.828388 17.232086
#> 437 219 2002-08-07 2.275078549 15.012610 35.591848 11.594837
#> 439 220 2002-08-08 4.577322428 14.952243 35.351609 14.248987
#> 442 221 2002-08-09 6.947512812 14.891281 35.107744 16.966678
#> 443 222 2002-08-10 4.908489490 14.829748 34.860325 14.484281
#> 445 223 2002-08-11 0.812991344 14.767666 34.609429  9.605018
#> 447 224 2002-08-12 1.466775839 14.705060 34.355129 10.302181
#> 450 225 2002-08-13 4.451967423 14.641952 34.097505 13.708145
#> 451 226 2002-08-14 6.384717726 14.578362 33.836632 15.868678
#> 453 227 2002-08-15 6.319396940 14.514313 33.572591 15.701745
#> 455 228 2002-08-16 3.319251172 14.449824 33.305462 12.151644
#> 458 229 2002-08-17 2.688080645 14.384916 33.035326 11.345454
#> 460 230 2002-08-18 4.469954673 14.319607 32.762264 13.304039
#> 461 231 2002-08-19 5.293986544 14.253916 32.486361 14.154401
#> 463 232 2002-08-20 0.373782803 14.187862 32.207701  8.476185
#> 465 233 2002-08-21 6.855420141 14.121461 31.926370 15.731098
#> 468 234 2002-08-22 1.307089004 14.054730 31.642455  9.381987
#> 469 235 2002-08-23 2.308125706 13.987686 31.356043 10.426061
#> 471 236 2002-08-24 1.019253180 13.920345 31.067223  8.904183
#> 473 237 2002-08-25 0.419050964 13.852721 30.776085  8.159516
#> 476 238 2002-08-26 2.951845796 13.784830 30.482721 10.884424
#> 477 239 2002-08-27 1.754648889 13.716686 30.187223  9.477592
#> 479 240 2002-08-28 6.367400069 13.648302 29.889684 14.444702
#> 481 241 2002-08-29 6.112914078 13.579692 29.590199 14.057581
#> 484 242 2002-08-30 5.349798450 13.510869 29.288864 13.120863
#> 485 243 2002-08-31 5.293575048 13.441846 28.985774 12.953932
#> 487 244 2002-09-01 0.561112848 13.372633 28.681027  7.771982
#> 490 245 2002-09-02 5.337707529 13.303244 28.374722 12.786125
#> 491 246 2002-09-03 5.664905122 13.233689 28.066958 13.024008
#> 493 247 2002-09-04 6.699861557 13.163980 27.757836 14.003191
#> 495 248 2002-09-05 6.604089008 13.094126 27.447457 13.783497
#> 498 249 2002-09-06 1.933557368 13.024138 27.135924  8.798274
#> 499 250 2002-09-07 6.599109725 12.954027 26.823338 13.538079
#> 501 251 2002-09-08 4.637222107 12.883802 26.509805 11.398242
#> 503 252 2002-09-09 4.094891687 12.813472 26.195428 10.734586
#> 506 253 2002-09-10 4.784830776 12.743047 25.880314 11.328921
#> 507 254 2002-09-11 0.292671062 12.672536 25.564567  6.686347
#> 509 255 2002-09-12 1.299265890 12.601947 25.248295  7.613628
#> 511 256 2002-09-13 2.991221502 12.531289 24.931604  9.208490
#> 514 257 2002-09-14 4.601295392 12.460572 24.614602 10.698348
#> 516 258 2002-09-15 6.055681083 12.389802 24.297397 12.012188
#> 517 259 2002-09-16 4.314332724 12.318988 23.980097 10.194157
#> 519 260 2002-09-17 3.305476568 12.248139 23.662812  9.108713
#> 521 261 2002-09-18 0.506177795 12.177262 23.345649  6.321622
#> 524 262 2002-09-19 3.169654622 12.106365 23.028718  8.771837
#> 525 263 2002-09-20 3.064068204 12.035456 22.712128  8.569136
#> 527 264 2002-09-21 0.397929662 11.964544 22.395989  5.971432
#> 529 265 2002-09-22 3.004594050 11.893635 22.080409  8.309101
#> 532 266 2002-09-23 6.123064085 11.822738 21.765497 11.077612
#> 533 267 2002-09-24 6.406179688 11.751861 21.451363 11.209629
#> 535 268 2002-09-25 6.949633145 11.681012 21.138114 11.572602
#> 537 269 2002-09-26 0.893042861 11.610198 20.825859  6.007415
#> 540 270 2002-09-27 0.345366617 11.539428 20.514705  5.435671
#> 542 271 2002-09-28 4.435446441 11.468711 20.204761  8.958218
#> 543 272 2002-09-29 2.354769603 11.398053 19.896131  7.029244
#> 545 273 2002-09-30 3.643687775 11.327464 19.588923  8.047800
#> 547 274 2002-10-01 0.226586546 11.256953 19.283241  5.014882
#> 550 275 2002-10-02 2.869557982 11.186528 18.979189  7.179059
#> 551 276 2002-10-03 1.009328573 11.116198 18.676871  5.517129
#> 553 277 2002-10-04 1.741127761 11.045973 18.376389  6.042392
#> 555 278 2002-10-05 6.857548602 10.975862 18.077844 10.166840
#> 558 279 2002-10-06 0.483707267 10.905874 17.781337  4.839661
#> 559 280 2002-10-07 1.763550752 10.836020 17.486965  5.794734
#> 561 281 2002-10-08 1.269533484 10.766311 17.194826  5.312490
#> 563 282 2002-10-09 0.916894832 10.696756 16.905016  4.950778
#> 566 283 2002-10-10 3.273970849 10.627367 16.617629  6.714102
#> 567 284 2002-10-11 1.411032734 10.558154 16.332759  5.174576
#> 570 285 2002-10-12 2.900171760 10.489131 16.050495  6.231549
#> 571 286 2002-10-13 2.555083296 10.420308 15.770928  5.876266
#> 573 287 2002-10-14 2.003807075 10.351698 15.494145  5.373159
#> 575 288 2002-10-15 0.742040911 10.283314 15.220231  4.354202
#> 578 289 2002-10-16 0.873949433 10.215170 14.949271  4.376803
#> 579 290 2002-10-17 0.382892496 10.147279 14.681345  3.947326
#> 581 291 2002-10-18 4.022908095 10.079655 14.416534  6.481037
#> 583 292 2002-10-19 1.590480776 10.012314 14.154915  4.663000
#> 586 293 2002-10-20 6.876294382  9.945270 13.896562  8.278276
#> 587 294 2002-10-21 0.655488661  9.878539 13.641550  3.862979
#> 589 295 2002-10-22 0.001804234  9.812138 13.389948  3.348718
#> 591 296 2002-10-23 3.224408589  9.746084 13.141825  5.459387
#> 594 297 2002-10-24 5.951067399  9.680393 12.897247  7.188634
#> 596 298 2002-10-25 3.928886968  9.615084 12.656277  5.749854
#> 597 299 2002-10-26 1.640564084  9.550176 12.418977  4.171433
#> 599 300 2002-10-27 0.767576429  9.485687 12.185404  3.539369
#> 601 301 2002-10-28 4.175401482  9.421638 11.955615  5.638098
#> 604 302 2002-10-29 0.402579783  9.358048 11.729663  3.184719
#> 605 303 2002-10-30 0.953374233  9.294940 11.507600  3.467062
#> 607 304 2002-10-31 2.952933467  9.232334 11.289473  4.627820
#> 609 305 2002-11-01 2.676408005  9.170252 11.075328  4.385042
#> 612 306 2002-11-02 3.242400780  9.108719 10.865209  4.650129
#> 613 307 2002-11-03 5.748349817  9.047757 10.659156  6.050853
#> 615 308 2002-11-04 5.676461110  8.987390 10.457208  5.916703
#> 617 309 2002-11-05 0.266227652  8.927644 10.259398  2.717820
#> 620 310 2002-11-06 3.402050495  8.868543 10.065762  4.447097
#> 622 311 2002-11-07 4.676887830  8.810114  9.876328  5.090528
#> 623 312 2002-11-08 5.396738385  8.752384  9.691126  5.410566
#> 625 313 2002-11-09 5.482297761  8.695379  9.510180  5.375553
#> 627 314 2002-11-10 1.575241396  8.639127  9.333513  3.184306
#> 630 315 2002-11-11 6.381676861  8.583658  9.161147  5.695798
#> 631 316 2002-11-12 2.048916358  8.528999  8.993100  3.328479
#> 633 317 2002-11-13 5.278413749  8.475180  8.829388  4.956856
#> 635 318 2002-11-14 6.062478317  8.422232  8.670025  5.287929
#> 638 319 2002-11-15 5.449345143  8.370184  8.515023  4.900577
#> 639 320 2002-11-16 5.548287326  8.319067  8.364392  4.880356
#> 641 321 2002-11-17 3.404244291  8.268912  8.218139  3.746206
#> 643 322 2002-11-18 3.847267092  8.219751  8.076271  3.909123
#> 646 323 2002-11-19 1.803918035  8.171616  7.938791  2.860958
#> 647 324 2002-11-20 3.271851779  8.124538  7.805702  3.523152
#> 650 325 2002-11-21 5.908705917  8.078549  7.677004  4.726758
#> 651 326 2002-11-22 3.766579355  8.033683  7.552698  3.658710
#> 653 327 2002-11-23 6.076577706  7.989971  7.432780  4.684605
#> 655 328 2002-11-24 3.751311821  7.947446  7.317248  3.556236
#> 658 329 2002-11-25 2.248408676  7.906140  7.206095  2.826186
#> 659 330 2002-11-26 6.540115044  7.866085  7.099317  4.726129
#> 661 331 2002-11-27 1.295815796  7.827314  6.996907  2.328398
#> 663 332 2002-11-28 3.689286474  7.789856  6.898857  3.358368
#> 666 333 2002-11-29 5.079096504  7.753745  6.805157  3.930151
#> 667 334 2002-11-30 3.779604355  7.719011  6.715800  3.323142
#> 669 335 2002-12-01 0.508021192  7.685684  6.630775  1.876840
#> 671 336 2002-12-02 0.180046076  7.653793  6.550072  1.714559
#> 674 337 2002-12-03 0.594266347  7.623368  6.473680  1.870742
#> 676 338 2002-12-04 0.033672391  7.594437  6.401590  1.614589
#> 677 339 2002-12-05 3.371931653  7.567027  6.333789  2.994643
#> 679 340 2002-12-06 0.593281978  7.541163  6.270268  1.814216
#> 681 341 2002-12-07 1.136587080  7.516872  6.211016  2.022322
#> 684 342 2002-12-08 3.072690888  7.494177  6.156023  2.801023
#> 685 343 2002-12-09 2.244694140  7.473100  6.105278  2.443240
#> 687 344 2002-12-10 2.915600738  7.453664  6.058772  2.699678
#> 689 345 2002-12-11 5.108126587  7.435887  6.016497  3.570658
#> 692 346 2002-12-12 0.675345022  7.419789  5.978443  1.766688
#> 693 347 2002-12-13 5.621325786  7.405386  5.944603  3.742383
#> 695 348 2002-12-14 4.805234769  7.392694  5.914970  3.401101
#> 697 349 2002-12-15 6.872905885  7.381725  5.889539  4.214172
#> 700 350 2002-12-16 0.062451263  7.372493  5.868303  1.491931
#> 702 351 2002-12-17 0.272974131  7.365006  5.851259  1.571249
#> 703 352 2002-12-18 0.678849538  7.359273  5.838404  1.728880
#> 705 353 2002-12-19 4.580985443  7.355300  5.829736  3.272855
#> 707 354 2002-12-20 4.189577027  7.353091  5.825253  3.115843
#> 710 355 2002-12-21 2.893200642  7.352649  5.824956  2.602273
#> 711 356 2002-12-22 0.826158721  7.353975  5.828845  1.784623
#> 713 357 2002-12-23 1.795680602  7.357066  5.836924  2.171556
#> 715 358 2002-12-24 6.192300929  7.361920  5.849194  3.922253
#> 718 359 2002-12-25 6.842709004  7.368531  5.865662  4.189958
#> 719 360 2002-12-26 5.298458709  7.376892  5.886331  3.585514
#> 721 361 2002-12-27 2.677087447  7.386993  5.911209  2.548930
#> 723 362 2002-12-28 2.230606180  7.398825  5.940303  2.380521
#> 726 363 2002-12-29 1.564384551  7.412375  5.973622  2.123773
#> 727 364 2002-12-30 3.879492762  7.427627  6.011174  3.072629
#> 729 365 2002-12-31 6.817003092  7.444567  6.052971  4.284601
#> 2     1 2003-01-01 4.611262846  7.463176  6.099033  3.408959
#> 4     2 2003-01-02 2.386307218  7.483435  6.149345  2.517783
#> 5     3 2003-01-03 1.854255885  7.505323  6.203934  2.317352
#> 8     4 2003-01-04 3.522667380  7.528820  6.262814  3.030861
#> 10    5 2003-01-05 5.714403047  7.553900  6.325996  3.974256
#> 12    6 2003-01-06 5.406210926  7.580540  6.393496  3.878197
#> 13    7 2003-01-07 1.273838189  7.608714  6.465324  2.157538
#> 16    8 2003-01-08 4.064626473  7.638396  6.541496  3.375840
#> 18    9 2003-01-09 5.691023512  7.669557  6.622023  4.112368
#> 20   10 2003-01-10 2.480866319  7.702170  6.706920  2.756878
#> 21   11 2003-01-11 4.321027229  7.736204  6.796198  3.597045
#> 24   12 2003-01-12 0.151938202  7.771631  6.889871  1.789818
#> 26   13 2003-01-13 0.835534656  7.808419  6.987950  2.120858
#> 28   14 2003-01-14 2.808012696  7.846537  7.090446  3.041328
#> 29   15 2003-01-15 2.133019948  7.885954  7.197371  2.772727
#> 31   16 2003-01-16 2.279468778  7.926639  7.308732  2.878071
#> 34   17 2003-01-17 3.899005742  7.968559  7.424539  3.672544
#> 36   18 2003-01-18 6.768910912  8.011681  7.544799  5.073426
#> 38   19 2003-01-19 1.500975901  8.055974  7.669519  2.631866
#> 39   20 2003-01-20 1.492200088  8.101405  7.798703  2.667898
#> 42   21 2003-01-21 4.126250145  8.147942  7.932354  3.991625
#> 44   22 2003-01-22 1.453380998  8.195553  8.070474  2.733218
#> 46   23 2003-01-23 3.900548216  8.244205  8.213063  3.996173
#> 47   24 2003-01-24 6.743044733  8.293867  8.360119  5.488484
#> 50   25 2003-01-25 5.623498152  8.344507  8.511639  4.995976
#> 52   26 2003-01-26 4.464458765  8.396093  8.667618  4.471323
#> 54   27 2003-01-27 2.711565692  8.448595  8.828046  3.623687
#> 55   28 2003-01-28 5.521823498  8.501983  8.992915  5.168565
#> 57   29 2003-01-29 2.221688779  8.556225  9.162213  3.480072
#> 60   30 2003-01-30 5.822266832  8.611293  9.335925  5.490084
#> 62   31 2003-01-31 4.953305518  8.667157  9.514035  5.097158
#> 64   32 2003-02-01 3.406662119  8.723789  9.696523  4.317390
#> 65   33 2003-02-02 4.043295227  8.781160  9.883368  4.746246
#> 68   34 2003-02-03 5.198640356  8.839243 10.074546  5.481217
#> 70   35 2003-02-04 0.863918277  8.898011 10.270030  3.066072
#> 72   36 2003-02-05 6.360066342  8.957438 10.469790  6.334390
#> 73   37 2003-02-06 6.885087076  9.017498 10.673795  6.743305
#> 76   38 2003-02-07 1.210778576  9.078165 10.882010  3.446184
#> 78   39 2003-02-08 4.236163078  9.139416 11.094398  5.344753
#> 80   40 2003-02-09 6.028568071  9.201226 11.310918  6.533140
#> 81   41 2003-02-10 0.347101070  9.263572 11.531528  3.098922
#> 84   42 2003-02-11 2.309723030  9.326432 11.756182  4.394775
#> 85   43 2003-02-12 0.609958315  9.389784 11.984831  3.385474
#> 88   44 2003-02-13 1.447692496  9.453606 12.217425  3.989823
#> 90   45 2003-02-14 0.095307771  9.517878 12.453910  3.175831
#> 92   46 2003-02-15 3.260743607  9.582579 12.694229  5.333342
#> 93   47 2003-02-16 0.204662514  9.647690 12.938322  3.371815
#> 96   48 2003-02-17 1.668058059  9.713192 13.186128  4.428767
#> 98   49 2003-02-18 4.269067556  9.779067 13.437583  6.292495
#> 100  50 2003-02-19 5.574544707  9.845297 13.692618  7.299630
#> 101  51 2003-02-20 2.159777722  9.911864 13.951164  5.007758
#> 104  52 2003-02-21 3.910690197  9.978754 14.213148  6.338365
#> 106  53 2003-02-22 6.609909837 10.045948 14.478496  8.382816
#> 108  54 2003-02-23 1.303444650 10.113433 14.747130  4.637106
#> 109  55 2003-02-24 6.761473027 10.181192 15.018971  8.741898
#> 111  56 2003-02-25 0.497533437 10.249211 15.293936  4.194695
#> 114  57 2003-02-26 6.409647314 10.317477 15.571940  8.729955
#> 116  58 2003-02-27 2.635493300 10.385975 15.852898  5.974601
#> 118  59 2003-02-28 4.259933306 10.454693 16.136720  7.321763
#> 119  60 2003-03-01 1.777947329 10.523618 16.423316  5.493174
#> 122  61 2003-03-02 4.409067284 10.592738 16.712592  7.656330
#> 124  62 2003-03-03 2.102007400 10.662040 17.004454  5.927316
#> 126  63 2003-03-04 5.877780087 10.731513 17.298805  9.062084
#> 127  64 2003-03-05 6.439903690 10.801147 17.595546  9.644331
#> 130  65 2003-03-06 3.594012135 10.870930 17.894578  7.431686
#> 132  66 2003-03-07 0.041286347 10.940852 18.195799  4.583281
#> 134  67 2003-03-08 1.017528177 11.010902 18.499104  5.479536
#> 135  68 2003-03-09 2.072047547 11.081072 18.804391  6.459213
#> 137  69 2003-03-10 6.258400836 11.151350 19.111552 10.140815
#> 140  70 2003-03-11 6.713639717 11.221729 19.420480 10.664479
#> 142  71 2003-03-12 2.571600291 11.292198 19.731067  7.179469
#> 144  72 2003-03-13 3.217142902 11.362749 20.043203  7.848224
#> 145  73 2003-03-14 4.023985261 11.433374 20.356778  8.671487
#> 148  74 2003-03-15 4.092863218 11.504062 20.671679  8.845158
#> 150  75 2003-03-16 2.154470579 11.574807 20.987796  7.200225
#> 152  76 2003-03-17 0.983922642 11.645600 21.305014  6.226271
#> 153  77 2003-03-18 1.154630782 11.716433 21.623220  6.471267
#> 156  78 2003-03-19 0.688907004 11.787297 21.942300  6.126782
#> 158  79 2003-03-20 1.115722286 11.858185 22.262140  6.612844
#> 160  80 2003-03-21 1.942975181 11.929088 22.582624  7.484752
#> 161  81 2003-03-22 5.838376877 12.000000 22.903638 11.297579
#> 164  82 2003-03-23 1.224851055 12.070912 23.225065  6.984605
#> 165  83 2003-03-24 2.806900965 12.141815 23.546791  8.608429
#> 168  84 2003-03-25 1.436214926 12.212703 23.868700  7.370656
#> 170  85 2003-03-26 5.390462664 12.283567 24.190677 11.355531
#> 172  86 2003-03-27 0.591139893 12.354400 24.512606  6.714598
#> 173  87 2003-03-28 3.360516214 12.425193 24.834373  9.566944
#> 176  88 2003-03-29 6.326688315 12.495938 25.155863 12.657167
#> 178  89 2003-03-30 6.540529131 12.566626 25.476961 12.999214
#> 180  90 2003-03-31 5.979488024 12.637251 25.797553 12.552621
#> 181  91 2003-04-01 5.083456849 12.707802 26.117528 11.753233
#> 184  92 2003-04-02 6.997491852 12.778271 26.436771 13.847695
#> 186  93 2003-04-03 1.936760237 12.848650 26.755170  8.705283
#> 188  94 2003-04-04 6.929281596 12.918928 27.072615 14.028577
#> 189  95 2003-04-05 6.591813325 12.989098 27.388994 13.797044
#> 191  96 2003-04-06 1.950361753 13.059148 27.704199  8.994837
#> 194  97 2003-04-07 3.684649445 13.129070 28.018120 10.936147
#> 196  98 2003-04-08 2.405912317 13.198853 28.330649  9.664745
#> 198  99 2003-04-09 3.531690320 13.268487 28.641680 10.972216
#> 199 100 2003-04-10 3.248816334 13.337960 28.951106 10.763684
#> 202 101 2003-04-11 6.556543378 13.407262 29.258822 14.468915
#> 204 102 2003-04-12 6.162946355 13.476382 29.564726 14.151372
#> 206 103 2003-04-13 0.920085492 13.545307 29.868714  8.481617
#> 207 104 2003-04-14 4.973068039 13.614025 30.170685 13.053197
#> 210 105 2003-04-15 1.656187188 13.682523 30.470539  9.461773
#> 212 106 2003-04-16 6.680489243 13.750789 30.768176 15.166032
#> 214 107 2003-04-17 2.668783003 13.818808 31.063500 10.765473
#> 215 108 2003-04-18 6.092803516 13.886567 31.356413 14.717998
#> 217 109 2003-04-19 2.276160974 13.954052 31.646820 10.492792
#> 220 110 2003-04-20 4.622140738 14.021246 31.934628 13.247324
#> 222 111 2003-04-21 1.891243824 14.088136 32.219743 10.217585
#> 224 112 2003-04-22 1.328978937 14.154703 32.502075  9.651322
#> 225 113 2003-04-23 3.705833357 14.220933 32.781534 12.466654
#> 228 114 2003-04-24 6.367192212 14.286808 33.058030 15.630982
#> 230 115 2003-04-25 2.741436963 14.352310 33.331477 11.516195
#> 232 116 2003-04-26 4.260394927 14.417421 33.601789 13.365166
#> 233 117 2003-04-27 4.542204761 14.482122 33.868881 13.778576
#> 236 118 2003-04-28 1.187684727 14.546394 34.132670  9.926601
#> 238 119 2003-04-29 3.721117721 14.610216 34.393074 12.978104
#> 240 120 2003-04-30 6.463157263 14.673568 34.650012 16.293520
#> 241 121 2003-05-01 2.262920687 14.736428 34.903405 11.405729
#> 244 122 2003-05-02 0.123758806 14.798774 35.153175  8.935283
#> 246 123 2003-05-03 0.851378916 14.860584 35.399245  9.863842
#> 247 124 2003-05-04 6.972022764 14.921835 35.641540 17.236896
#> 250 125 2003-05-05 0.898510223 14.982502 35.879987 10.045869
#> 252 126 2003-05-06 5.613266329 15.042562 36.114511 15.766854
#> 254 127 2003-05-07 2.257340275 15.101989 36.345041 11.802562
#> 255 128 2003-05-08 2.606661847 15.160757 36.571508 12.286835
#> 258 129 2003-05-09 3.680756183 15.218840 36.793843 13.647852
#> 260 130 2003-05-10 3.938710346 15.276211 37.011977 14.024447
#> 262 131 2003-05-11 5.257843484 15.332843 37.225844 15.689089
#> 263 132 2003-05-12 3.500852155 15.388707 37.435378 13.617023
#> 266 133 2003-05-13 6.275364066 15.443775 37.640517 17.057481
#> 268 134 2003-05-14 4.392007902 15.498017 37.841197 14.822238
#> 270 135 2003-05-15 4.517446551 15.551405 38.037357 15.033976
#> 271 136 2003-05-16 2.887461698 15.603907 38.228936 13.094316
#> 273 137 2003-05-17 2.913481535 15.655493 38.415875 13.178559
#> 276 138 2003-05-18 3.167894837 15.706133 38.598118 13.542110
#> 278 139 2003-05-19 6.536948200 15.755795 38.775606 17.737740
#> 280 140 2003-05-20 0.045705891 15.804447 38.948286  9.793390
#> 281 141 2003-05-21 4.055711838 15.852058 39.116103 14.782908
#> 284 142 2003-05-22 5.891203906 15.898595 39.279005 17.097144
#> 286 143 2003-05-23 2.203847506 15.944026 39.436940 12.584802
#> 288 144 2003-05-24 5.392472057 15.988319 39.589859 16.573814
#> 289 145 2003-05-25 4.914849321 16.031441 39.737713 16.025736
#> 292 146 2003-05-26 1.009311622 16.073361 39.880455 11.222242
#> 294 147 2003-05-27 2.251410034 16.114046 40.018039 12.800115
#> 296 148 2003-05-28 2.637081859 16.153463 40.150421 13.314919
#> 297 149 2003-05-29 6.185129004 16.191581 40.277558 17.762334
#> 299 150 2003-05-30 3.266295813 16.228369 40.399408 14.165461
#> 302 151 2003-05-31 2.333219598 16.263796 40.515933 13.035210
#> 304 152 2003-06-01 1.161030555 16.297830 40.627093 11.603877
#> 306 153 2003-06-02 6.635071927 16.330443 40.732853 18.458108
#> 307 154 2003-06-03 0.527014628 16.361604 40.833176 10.865922
#> 310 155 2003-06-04 0.364430947 16.391286 40.928031 10.686989
#> 312 156 2003-06-05 1.514192510 16.419460 41.017385 12.145645
#> 314 157 2003-06-06 0.661726309 16.446100 41.101209 11.102178
#> 315 158 2003-06-07 3.602409440 16.471180 41.179475 14.798047
#> 318 159 2003-06-08 4.664494097 16.494677 41.252156 16.145843
#> 320 160 2003-06-09 1.806480620 16.516565 41.319229 12.589429
#> 322 161 2003-06-10 2.590266821 16.536824 41.380671 13.586025
#> 323 162 2003-06-11 4.828831229 16.555433 41.436461 16.402137
#> 326 163 2003-06-12 2.989207186 16.572373 41.486582 14.113173
#> 327 164 2003-06-13 3.990326004 16.587625 41.531016 15.378113
#> 330 165 2003-06-14 5.458190701 16.601175 41.569750 17.226159
#> 332 166 2003-06-15 0.358330484 16.613007 41.602770 10.849363
#> 334 167 2003-06-16 3.036887600 16.623108 41.630068 14.210230
#> 335 168 2003-06-17 6.322465102 16.631469 41.651634 18.329858
#> 338 169 2003-06-18 5.830726353 16.638080 41.667464 17.717936
#> 340 170 2003-06-19 3.840207535 16.642934 41.677552 15.227748
#> 342 171 2003-06-20 5.140014651 16.646025 41.681897 16.855811
#> 343 172 2003-06-21 1.985246242 16.647351 41.680500 12.905387
#> 346 173 2003-06-22 6.518643295 16.646909 41.673363 18.577627
#> 348 174 2003-06-23 6.218011558 16.644700 41.660491 18.196741
#> 350 175 2003-06-24 5.447706526 16.640727 41.641891 17.226665
#> 351 176 2003-06-25 6.627437949 16.634994 41.617572 18.694684
#> 353 177 2003-06-26 2.377887187 16.627507 41.587545 13.370588
#> 356 178 2003-06-27 5.248310023 16.618275 41.551823 16.949312
#> 358 179 2003-06-28 1.747926840 16.607306 41.510423 12.562101
#> 360 180 2003-06-29 2.613705729 16.594614 41.463360 13.631147
#> 361 181 2003-06-30 5.643160499 16.580211 41.410655 17.399829
#> 364 182 2003-07-01 0.588312812 16.564113 41.352328 11.072444
#> 366 183 2003-07-02 5.251806693 16.546336 41.288405 16.874570
#> 368 184 2003-07-03 2.459529572 16.526900 41.218909 13.371822
#> 369 185 2003-07-04 6.630615377 16.505823 41.143868 18.549995
#> 372 186 2003-07-05 6.277682149 16.483128 41.063311 18.085412
#> 374 187 2003-07-06 1.513677215 16.458837 40.977269 12.128605
#> 376 188 2003-07-07 6.206419612 16.432973 40.885775 17.942331
#> 377 189 2003-07-08 3.813768352 16.405563 40.788864 14.938269
#> 379 190 2003-07-09 3.093611901 16.376632 40.686570 14.014571
#> 382 191 2003-07-10 1.472566161 16.346207 40.578932 11.972532
#> 384 192 2003-07-11 5.804832100 16.314316 40.465990 17.315643
#> 386 193 2003-07-12 3.229059896 16.280989 40.347784 14.088098
#> 387 194 2003-07-13 1.890056796 16.246255 40.224356 12.395900
#> 390 195 2003-07-14 0.745064342 16.210144 40.095751 10.945395
#> 392 196 2003-07-15 4.768831418 16.172686 39.962013 15.882292
#> 394 197 2003-07-16 1.772030025 16.133915 39.823189 12.142740
#> 395 198 2003-07-17 3.054132832 16.093860 39.679327 13.684806
#> 398 199 2003-07-18 4.434066389 16.052554 39.530476 15.342210
#> 400 200 2003-07-19 1.201399590 16.010029 39.376687 11.321594
#> 402 201 2003-07-20 5.169396505 15.966317 39.218012 16.153289
#> 403 202 2003-07-21 1.446498648 15.921451 39.054503 11.537719
#> 406 203 2003-07-22 2.654737939 15.875462 38.886214 12.972883
#> 407 204 2003-07-23 4.179130922 15.828384 38.713200 14.788977
#> 410 205 2003-07-24 5.480576848 15.780249 38.535519 16.325690
#> 412 206 2003-07-25 1.250679998 15.731088 38.353226 11.112919
#> 414 207 2003-07-26 4.613111342 15.680933 38.166382 15.155603
#> 415 208 2003-07-27 1.059676717 15.629816 37.975044 10.781085
#> 418 209 2003-07-28 1.872143169 15.577768 37.779274 11.714984
#> 420 210 2003-07-29 6.749606356 15.524820 37.579134 17.563779
#> 422 211 2003-07-30 5.052154410 15.471001 37.374685 15.446143
#> 423 212 2003-07-31 3.915674833 15.416342 37.165992 14.011487
#> 426 213 2003-08-01 5.007207330 15.360873 36.953119 15.261112
#> 428 214 2003-08-02 4.161575598 15.304621 36.736132 14.178609
#> 430 215 2003-08-03 2.867791035 15.247616 36.515097 12.562677
#> 431 216 2003-08-04 1.557822797 15.189886 36.290082 10.933414
#> 433 217 2003-08-05 4.808616636 15.131457 36.061156 14.745215
#> 436 218 2003-08-06 6.962277524 15.072356 35.828388 17.232086
#> 438 219 2003-08-07 2.275078549 15.012610 35.591848 11.594837
#> 440 220 2003-08-08 4.577322428 14.952243 35.351609 14.248987
#> 441 221 2003-08-09 6.947512812 14.891281 35.107744 16.966678
#> 444 222 2003-08-10 4.908489490 14.829748 34.860325 14.484281
#> 446 223 2003-08-11 0.812991344 14.767666 34.609429  9.605018
#> 448 224 2003-08-12 1.466775839 14.705060 34.355129 10.302181
#> 449 225 2003-08-13 4.451967423 14.641952 34.097505 13.708145
#> 452 226 2003-08-14 6.384717726 14.578362 33.836632 15.868678
#> 454 227 2003-08-15 6.319396940 14.514313 33.572591 15.701745
#> 456 228 2003-08-16 3.319251172 14.449824 33.305462 12.151644
#> 457 229 2003-08-17 2.688080645 14.384916 33.035326 11.345454
#> 459 230 2003-08-18 4.469954673 14.319607 32.762264 13.304039
#> 462 231 2003-08-19 5.293986544 14.253916 32.486361 14.154401
#> 464 232 2003-08-20 0.373782803 14.187862 32.207701  8.476185
#> 466 233 2003-08-21 6.855420141 14.121461 31.926370 15.731098
#> 467 234 2003-08-22 1.307089004 14.054730 31.642455  9.381987
#> 470 235 2003-08-23 2.308125706 13.987686 31.356043 10.426061
#> 472 236 2003-08-24 1.019253180 13.920345 31.067223  8.904183
#> 474 237 2003-08-25 0.419050964 13.852721 30.776085  8.159516
#> 475 238 2003-08-26 2.951845796 13.784830 30.482721 10.884424
#> 478 239 2003-08-27 1.754648889 13.716686 30.187223  9.477592
#> 480 240 2003-08-28 6.367400069 13.648302 29.889684 14.444702
#> 482 241 2003-08-29 6.112914078 13.579692 29.590199 14.057581
#> 483 242 2003-08-30 5.349798450 13.510869 29.288864 13.120863
#> 486 243 2003-08-31 5.293575048 13.441846 28.985774 12.953932
#> 488 244 2003-09-01 0.561112848 13.372633 28.681027  7.771982
#> 489 245 2003-09-02 5.337707529 13.303244 28.374722 12.786125
#> 492 246 2003-09-03 5.664905122 13.233689 28.066958 13.024008
#> 494 247 2003-09-04 6.699861557 13.163980 27.757836 14.003191
#> 496 248 2003-09-05 6.604089008 13.094126 27.447457 13.783497
#> 497 249 2003-09-06 1.933557368 13.024138 27.135924  8.798274
#> 500 250 2003-09-07 6.599109725 12.954027 26.823338 13.538079
#> 502 251 2003-09-08 4.637222107 12.883802 26.509805 11.398242
#> 504 252 2003-09-09 4.094891687 12.813472 26.195428 10.734586
#> 505 253 2003-09-10 4.784830776 12.743047 25.880314 11.328921
#> 508 254 2003-09-11 0.292671062 12.672536 25.564567  6.686347
#> 510 255 2003-09-12 1.299265890 12.601947 25.248295  7.613628
#> 512 256 2003-09-13 2.991221502 12.531289 24.931604  9.208490
#> 513 257 2003-09-14 4.601295392 12.460572 24.614602 10.698348
#> 515 258 2003-09-15 6.055681083 12.389802 24.297397 12.012188
#> 518 259 2003-09-16 4.314332724 12.318988 23.980097 10.194157
#> 520 260 2003-09-17 3.305476568 12.248139 23.662812  9.108713
#> 522 261 2003-09-18 0.506177795 12.177262 23.345649  6.321622
#> 523 262 2003-09-19 3.169654622 12.106365 23.028718  8.771837
#> 526 263 2003-09-20 3.064068204 12.035456 22.712128  8.569136
#> 528 264 2003-09-21 0.397929662 11.964544 22.395989  5.971432
#> 530 265 2003-09-22 3.004594050 11.893635 22.080409  8.309101
#> 531 266 2003-09-23 6.123064085 11.822738 21.765497 11.077612
#> 534 267 2003-09-24 6.406179688 11.751861 21.451363 11.209629
#> 536 268 2003-09-25 6.949633145 11.681012 21.138114 11.572602
#> 538 269 2003-09-26 0.893042861 11.610198 20.825859  6.007415
#> 539 270 2003-09-27 0.345366617 11.539428 20.514705  5.435671
#> 541 271 2003-09-28 4.435446441 11.468711 20.204761  8.958218
#> 544 272 2003-09-29 2.354769603 11.398053 19.896131  7.029244
#> 546 273 2003-09-30 3.643687775 11.327464 19.588923  8.047800
#> 548 274 2003-10-01 0.226586546 11.256953 19.283241  5.014882
#> 549 275 2003-10-02 2.869557982 11.186528 18.979189  7.179059
#> 552 276 2003-10-03 1.009328573 11.116198 18.676871  5.517129
#> 554 277 2003-10-04 1.741127761 11.045973 18.376389  6.042392
#> 556 278 2003-10-05 6.857548602 10.975862 18.077844 10.166840
#> 557 279 2003-10-06 0.483707267 10.905874 17.781337  4.839661
#> 560 280 2003-10-07 1.763550752 10.836020 17.486965  5.794734
#> 562 281 2003-10-08 1.269533484 10.766311 17.194826  5.312490
#> 564 282 2003-10-09 0.916894832 10.696756 16.905016  4.950778
#> 565 283 2003-10-10 3.273970849 10.627367 16.617629  6.714102
#> 568 284 2003-10-11 1.411032734 10.558154 16.332759  5.174576
#> 569 285 2003-10-12 2.900171760 10.489131 16.050495  6.231549
#> 572 286 2003-10-13 2.555083296 10.420308 15.770928  5.876266
#> 574 287 2003-10-14 2.003807075 10.351698 15.494145  5.373159
#> 576 288 2003-10-15 0.742040911 10.283314 15.220231  4.354202
#> 577 289 2003-10-16 0.873949433 10.215170 14.949271  4.376803
#> 580 290 2003-10-17 0.382892496 10.147279 14.681345  3.947326
#> 582 291 2003-10-18 4.022908095 10.079655 14.416534  6.481037
#> 584 292 2003-10-19 1.590480776 10.012314 14.154915  4.663000
#> 585 293 2003-10-20 6.876294382  9.945270 13.896562  8.278276
#> 588 294 2003-10-21 0.655488661  9.878539 13.641550  3.862979
#> 590 295 2003-10-22 0.001804234  9.812138 13.389948  3.348718
#> 592 296 2003-10-23 3.224408589  9.746084 13.141825  5.459387
#> 593 297 2003-10-24 5.951067399  9.680393 12.897247  7.188634
#> 595 298 2003-10-25 3.928886968  9.615084 12.656277  5.749854
#> 598 299 2003-10-26 1.640564084  9.550176 12.418977  4.171433
#> 600 300 2003-10-27 0.767576429  9.485687 12.185404  3.539369
#> 602 301 2003-10-28 4.175401482  9.421638 11.955615  5.638098
#> 603 302 2003-10-29 0.402579783  9.358048 11.729663  3.184719
#> 606 303 2003-10-30 0.953374233  9.294940 11.507600  3.467062
#> 608 304 2003-10-31 2.952933467  9.232334 11.289473  4.627820
#> 610 305 2003-11-01 2.676408005  9.170252 11.075328  4.385042
#> 611 306 2003-11-02 3.242400780  9.108719 10.865209  4.650129
#> 614 307 2003-11-03 5.748349817  9.047757 10.659156  6.050853
#> 616 308 2003-11-04 5.676461110  8.987390 10.457208  5.916703
#> 618 309 2003-11-05 0.266227652  8.927644 10.259398  2.717820
#> 619 310 2003-11-06 3.402050495  8.868543 10.065762  4.447097
#> 621 311 2003-11-07 4.676887830  8.810114  9.876328  5.090528
#> 624 312 2003-11-08 5.396738385  8.752384  9.691126  5.410566
#> 626 313 2003-11-09 5.482297761  8.695379  9.510180  5.375553
#> 628 314 2003-11-10 1.575241396  8.639127  9.333513  3.184306
#> 629 315 2003-11-11 6.381676861  8.583658  9.161147  5.695798
#> 632 316 2003-11-12 2.048916358  8.528999  8.993100  3.328479
#> 634 317 2003-11-13 5.278413749  8.475180  8.829388  4.956856
#> 636 318 2003-11-14 6.062478317  8.422232  8.670025  5.287929
#> 637 319 2003-11-15 5.449345143  8.370184  8.515023  4.900577
#> 640 320 2003-11-16 5.548287326  8.319067  8.364392  4.880356
#> 642 321 2003-11-17 3.404244291  8.268912  8.218139  3.746206
#> 644 322 2003-11-18 3.847267092  8.219751  8.076271  3.909123
#> 645 323 2003-11-19 1.803918035  8.171616  7.938791  2.860958
#> 648 324 2003-11-20 3.271851779  8.124538  7.805702  3.523152
#> 649 325 2003-11-21 5.908705917  8.078549  7.677004  4.726758
#> 652 326 2003-11-22 3.766579355  8.033683  7.552698  3.658710
#> 654 327 2003-11-23 6.076577706  7.989971  7.432780  4.684605
#> 656 328 2003-11-24 3.751311821  7.947446  7.317248  3.556236
#> 657 329 2003-11-25 2.248408676  7.906140  7.206095  2.826186
#> 660 330 2003-11-26 6.540115044  7.866085  7.099317  4.726129
#> 662 331 2003-11-27 1.295815796  7.827314  6.996907  2.328398
#> 664 332 2003-11-28 3.689286474  7.789856  6.898857  3.358368
#> 665 333 2003-11-29 5.079096504  7.753745  6.805157  3.930151
#> 668 334 2003-11-30 3.779604355  7.719011  6.715800  3.323142
#> 670 335 2003-12-01 0.508021192  7.685684  6.630775  1.876840
#> 672 336 2003-12-02 0.180046076  7.653793  6.550072  1.714559
#> 673 337 2003-12-03 0.594266347  7.623368  6.473680  1.870742
#> 675 338 2003-12-04 0.033672391  7.594437  6.401590  1.614589
#> 678 339 2003-12-05 3.371931653  7.567027  6.333789  2.994643
#> 680 340 2003-12-06 0.593281978  7.541163  6.270268  1.814216
#> 682 341 2003-12-07 1.136587080  7.516872  6.211016  2.022322
#> 683 342 2003-12-08 3.072690888  7.494177  6.156023  2.801023
#> 686 343 2003-12-09 2.244694140  7.473100  6.105278  2.443240
#> 688 344 2003-12-10 2.915600738  7.453664  6.058772  2.699678
#> 690 345 2003-12-11 5.108126587  7.435887  6.016497  3.570658
#> 691 346 2003-12-12 0.675345022  7.419789  5.978443  1.766688
#> 694 347 2003-12-13 5.621325786  7.405386  5.944603  3.742383
#> 696 348 2003-12-14 4.805234769  7.392694  5.914970  3.401101
#> 698 349 2003-12-15 6.872905885  7.381725  5.889539  4.214172
#> 699 350 2003-12-16 0.062451263  7.372493  5.868303  1.491931
#> 701 351 2003-12-17 0.272974131  7.365006  5.851259  1.571249
#> 704 352 2003-12-18 0.678849538  7.359273  5.838404  1.728880
#> 706 353 2003-12-19 4.580985443  7.355300  5.829736  3.272855
#> 708 354 2003-12-20 4.189577027  7.353091  5.825253  3.115843
#> 709 355 2003-12-21 2.893200642  7.352649  5.824956  2.602273
#> 712 356 2003-12-22 0.826158721  7.353975  5.828845  1.784623
#> 714 357 2003-12-23 1.795680602  7.357066  5.836924  2.171556
#> 716 358 2003-12-24 6.192300929  7.361920  5.849194  3.922253
#> 717 359 2003-12-25 6.842709004  7.368531  5.865662  4.189958
#> 720 360 2003-12-26 5.298458709  7.376892  5.886331  3.585514
#> 722 361 2003-12-27 2.677087447  7.386993  5.911209  2.548930
#> 724 362 2003-12-28 2.230606180  7.398825  5.940303  2.380521
#> 725 363 2003-12-29 1.564384551  7.412375  5.973622  2.123773
#> 728 364 2003-12-30 3.879492762  7.427627  6.011174  3.072629
#> 730 365 2003-12-31 6.817003092  7.444567  6.052971  4.284601