Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.955298  2.404369  4.246029  2.643164  2.982871  2.552858  3.796426
#>   [8]  4.202809  1.933778  3.018940  4.484283  2.914394  1.807780  3.124534
#>  [15]  2.178192  3.868302  4.260876  3.581426  3.442414  4.295165  4.647672
#>  [22]  2.696597  5.249717  3.989673  5.584942  2.814032  5.092770  5.090393
#>  [29]  3.890943  4.943443  5.643524  3.260203  6.169388  3.030945  5.791008
#>  [36]  4.182985  6.796422  5.529083  5.142677  5.532704  6.604297  6.343747
#>  [43]  3.400387  6.110076  4.502926  6.304040  4.907683  4.253191  3.378629
#>  [50]  3.730129  7.245403  7.336593  7.660325  7.241897  6.896488  8.810851
#>  [57]  7.228644  5.515257  8.745226  8.189800  9.183395  6.044717  6.525361
#>  [64]  5.167206  8.330762  7.982769 10.257845  8.078853 10.207394  8.275740
#>  [71]  7.597108  9.028666  8.541264  5.474795  6.495719 11.074844  7.550402
#>  [78]  6.570258  7.066829  8.245186 11.777028 11.727581  8.665518 12.273395
#>  [85]  7.919721  9.869278  8.787783  7.616904 12.045859  9.852853 10.140304
#>  [92]  9.662589 10.706271  7.162506 13.046218 13.562973  9.250822  8.561158
#>  [99] 11.900917 14.541717 10.036832 14.915390  9.939442 11.625187 11.806618
#> [106] 14.477999  8.073956  9.522254  8.930821 15.350326 15.252361 15.296243
#> [113] 11.782195 12.591023 12.079175 11.790937 10.542652 13.436502  8.777066
#> [120] 16.513308 13.186394 15.944729 16.871045 15.973250 14.076342 10.799241
#> [127]  9.856257 16.987369 11.019265 17.346502 16.309685 11.016829 12.133464
#> [134] 12.206660 13.025424 12.414191 14.037046 15.024409  9.913499 15.642747
#> [141] 17.156139 10.742143 16.593877 11.907483 12.667264 14.192217 15.189913
#> [148] 13.931203 12.345002 16.824389 14.718061 14.395459 17.625781 12.664997
#> [155] 17.658216 18.587841 11.678034 17.665295 12.825406 18.598288 15.709616
#> [162] 14.170584 18.969814 15.089095 12.222912 18.735655 18.358661 10.509505
#> [169] 18.734481 15.008272 11.357386 11.325405 14.304236 11.921923 11.239790
#> [176] 13.292149 17.924458 13.939892 17.441094 12.162806 15.852019 11.634564
#> [183] 14.375372 13.135919 14.803585 15.486930 13.816887 10.265631 13.696203
#> [190] 18.525211 12.699905 17.676514 13.569373 14.271643 10.289723 11.223137
#> [197] 11.159900 12.015687 15.201637 15.058825 10.962675 14.468407  9.996109
#> [204] 10.402355 18.151009 13.414360 15.358971 10.762258 11.424650 15.313488
#> [211] 13.965958 16.329131 13.197776 15.871836 12.083024 16.554395 15.649160
#> [218] 14.074460 16.324204 13.647428 14.916187 12.732848 13.250235  9.835053
#> [225] 11.943675 15.484602  8.926291 10.343735 15.874251 14.850035 13.113002
#> [232] 12.833762 11.468843  8.347291 14.579211 13.976393 15.334984 11.534066
#> [239] 12.873592  7.677795  8.773000  7.958097 10.861336  7.993358 14.052357
#> [246] 11.898553 11.414009  9.502560 11.959111 12.963636 12.307689 12.986578
#> [253]  8.183870  9.021240  7.479592 11.934906  9.654746 11.020916 11.484602
#> [260]  5.953990  7.182336 10.711022  6.256407 11.652008  9.137058  8.541959
#> [267] 11.613179  5.974296  8.318533 11.077405  9.361103  7.544773  5.010352
#> [274]  6.300662 10.357299  9.741495 10.128553  6.583498  9.630734  9.794484
#> [281]  6.268283  6.305181  5.944081  9.408364  8.045497  5.774762  8.083013
#> [288]  6.683530  5.332218  6.732570  5.997091  4.465325  3.519924  3.813506
#> [295]  3.567894  6.054428  6.091388  5.722198  7.120924  6.960958  5.792241
#> [302]  5.182274  3.192443  4.802738  5.902406  5.582712  3.177691  6.658356
#> [309]  3.058296  5.449150  4.605358  2.493378  3.719220  4.923940  4.531984
#> [316]  2.726241  3.352597  3.944438  3.455747  2.112570  2.476980  2.438368
#> [323]  2.936382  4.949848  3.575748  4.185791  3.532706  3.743300  4.581736
#> [330]  2.198148  2.745923  3.575006  4.442893  3.982559  3.704371  2.287104
#> [337]  2.242841  4.548777  2.512500  2.296708  2.441921  3.691649  1.608303
#> [344]  3.306170  2.615416  2.330403  2.890919  3.682267  2.640304  2.443445
#> [351]  3.851960  4.194753  2.670785  2.411930  2.858510  3.691438  2.160996
#> [358]  3.883178  2.553178  2.695748  2.375472  3.744689  3.030528  2.094437
#> [365]  3.573935  1.955298  2.404369  4.246029  2.643164  2.982871  2.552858
#> [372]  3.796426  4.202809  1.933778  3.018940  4.484283  2.914394  1.807780
#> [379]  3.124534  2.178192  3.868302  4.260876  3.581426  3.442414  4.295165
#> [386]  4.647672  2.696597  5.249717  3.989673  5.584942  2.814032  5.092770
#> [393]  5.090393  3.890943  4.943443  5.643524  3.260203  6.169388  3.030945
#> [400]  5.791008  4.182985  6.796422  5.529083  5.142677  5.532704  6.604297
#> [407]  6.343747  3.400387  6.110076  4.502926  6.304040  4.907683  4.253191
#> [414]  3.378629  3.730129  7.245403  7.336593  7.660325  7.241897  6.896488
#> [421]  8.810851  7.228644  5.515257  8.745226  8.189800  9.183395  6.044717
#> [428]  6.525361  5.167206  8.330762  7.982769 10.257845  8.078853 10.207394
#> [435]  8.275740  7.597108  9.028666  8.541264  5.474795  6.495719 11.074844
#> [442]  7.550402  6.570258  7.066829  8.245186 11.777028 11.727581  8.665518
#> [449] 12.273395  7.919721  9.869278  8.787783  7.616904 12.045859  9.852853
#> [456] 10.140304  9.662589 10.706271  7.162506 13.046218 13.562973  9.250822
#> [463]  8.561158 11.900917 14.541717 10.036832 14.915390  9.939442 11.625187
#> [470] 11.806618 14.477999  8.073956  9.522254  8.930821 15.350326 15.252361
#> [477] 15.296243 11.782195 12.591023 12.079175 11.790937 10.542652 13.436502
#> [484]  8.777066 16.513308 13.186394 15.944729 16.871045 15.973250 14.076342
#> [491] 10.799241  9.856257 16.987369 11.019265 17.346502 16.309685 11.016829
#> [498] 12.133464 12.206660 13.025424 12.414191 14.037046 15.024409  9.913499
#> [505] 15.642747 17.156139 10.742143 16.593877 11.907483 12.667264 14.192217
#> [512] 15.189913 13.931203 12.345002 16.824389 14.718061 14.395459 17.625781
#> [519] 12.664997 17.658216 18.587841 11.678034 17.665295 12.825406 18.598288
#> [526] 15.709616 14.170584 18.969814 15.089095 12.222912 18.735655 18.358661
#> [533] 10.509505 18.734481 15.008272 11.357386 11.325405 14.304236 11.921923
#> [540] 11.239790 13.292149 17.924458 13.939892 17.441094 12.162806 15.852019
#> [547] 11.634564 14.375372 13.135919 14.803585 15.486930 13.816887 10.265631
#> [554] 13.696203 18.525211 12.699905 17.676514 13.569373 14.271643 10.289723
#> [561] 11.223137 11.159900 12.015687 15.201637 15.058825 10.962675 14.468407
#> [568]  9.996109 10.402355 18.151009 13.414360 15.358971 10.762258 11.424650
#> [575] 15.313488 13.965958 16.329131 13.197776 15.871836 12.083024 16.554395
#> [582] 15.649160 14.074460 16.324204 13.647428 14.916187 12.732848 13.250235
#> [589]  9.835053 11.943675 15.484602  8.926291 10.343735 15.874251 14.850035
#> [596] 13.113002 12.833762 11.468843  8.347291 14.579211 13.976393 15.334984
#> [603] 11.534066 12.873592  7.677795  8.773000  7.958097 10.861336  7.993358
#> [610] 14.052357 11.898553 11.414009  9.502560 11.959111 12.963636 12.307689
#> [617] 12.986578  8.183870  9.021240  7.479592 11.934906  9.654746 11.020916
#> [624] 11.484602  5.953990  7.182336 10.711022  6.256407 11.652008  9.137058
#> [631]  8.541959 11.613179  5.974296  8.318533 11.077405  9.361103  7.544773
#> [638]  5.010352  6.300662 10.357299  9.741495 10.128553  6.583498  9.630734
#> [645]  9.794484  6.268283  6.305181  5.944081  9.408364  8.045497  5.774762
#> [652]  8.083013  6.683530  5.332218  6.732570  5.997091  4.465325  3.519924
#> [659]  3.813506  3.567894  6.054428  6.091388  5.722198  7.120924  6.960958
#> [666]  5.792241  5.182274  3.192443  4.802738  5.902406  5.582712  3.177691
#> [673]  6.658356  3.058296  5.449150  4.605358  2.493378  3.719220  4.923940
#> [680]  4.531984  2.726241  3.352597  3.944438  3.455747  2.112570  2.476980
#> [687]  2.438368  2.936382  4.949848  3.575748  4.185791  3.532706  3.743300
#> [694]  4.581736  2.198148  2.745923  3.575006  4.442893  3.982559  3.704371
#> [701]  2.287104  2.242841  4.548777  2.512500  2.296708  2.441921  3.691649
#> [708]  1.608303  3.306170  2.615416  2.330403  2.890919  3.682267  2.640304
#> [715]  2.443445  3.851960  4.194753  2.670785  2.411930  2.858510  3.691438
#> [722]  2.160996  3.883178  2.553178  2.695748  2.375472  3.744689  3.030528
#> [729]  2.094437  3.573935
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates    sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 5.346466128  7.463176  6.099033  3.709369
#> 3     2 2002-01-02 4.220213105  7.483435  6.149345  3.271269
#> 6     3 2002-01-03 2.636928513  7.505323  6.203934  2.640832
#> 7     4 2002-01-04 0.195266241  7.528820  6.262814  1.646919
#> 9     5 2002-01-05 6.118228219  7.553900  6.325996  4.143347
#> 11    6 2002-01-06 5.426044006  7.580540  6.393496  3.886561
#> 14    7 2002-01-07 0.210693771  7.608714  6.465324  1.705847
#> 15    8 2002-01-08 2.133529094  7.638396  6.541496  2.548947
#> 17    9 2002-01-09 5.585108983  7.669557  6.622023  4.066644
#> 19   10 2002-01-10 3.757857909  7.702170  6.706920  3.312870
#> 22   11 2002-01-11 3.755143018  7.736204  6.796198  3.348482
#> 23   12 2002-01-12 2.220060847  7.771631  6.889871  2.706555
#> 25   13 2002-01-13 4.109860013  7.808419  6.987950  3.585996
#> 27   14 2002-01-14 2.856462530  7.846537  7.090446  3.063219
#> 30   15 2002-01-15 0.828308818  7.885954  7.197371  2.177334
#> 32   16 2002-01-16 6.566517362  7.926639  7.308732  4.854501
#> 33   17 2002-01-17 0.059141654  7.968559  7.424539  1.883687
#> 35   18 2002-01-18 4.273028879  8.011681  7.544799  3.898209
#> 37   19 2002-01-19 5.796921838  8.055974  7.669519  4.676798
#> 40   20 2002-01-20 0.256513920  8.101405  7.798703  2.073140
#> 41   21 2002-01-21 3.654414976  8.147942  7.932354  3.761949
#> 43   22 2002-01-22 1.578354027  8.195553  8.070474  2.794751
#> 45   23 2002-01-23 0.021041591  8.244205  8.213063  2.063747
#> 48   24 2002-01-24 1.290562234  8.293867  8.360119  2.740465
#> 49   25 2002-01-25 5.177232418  8.344507  8.511639  4.768374
#> 51   26 2002-01-26 2.889118763  8.396093  8.667618  3.658180
#> 53   27 2002-01-27 0.872215553  8.448595  8.828046  2.662706
#> 56   28 2002-01-28 6.387470992  8.501983  8.992915  5.626381
#> 58   29 2002-01-29 4.309051740  8.556225  9.162213  4.597672
#> 59   30 2002-01-30 2.287111194  8.611293  9.335925  3.573766
#> 61   31 2002-01-31 1.480668145  8.667157  9.514035  3.191182
#> 63   32 2002-02-01 4.934802754  8.723789  9.696523  5.166657
#> 66   33 2002-02-02 1.217424934  8.781160  9.883368  3.155960
#> 67   34 2002-02-03 1.675824982  8.839243 10.074546  3.473649
#> 69   35 2002-02-04 0.384682714  8.898011 10.270030  2.789507
#> 71   36 2002-02-05 2.572630246  8.957438 10.469790  4.120941
#> 74   37 2002-02-06 6.684625771  9.017498 10.673795  6.624664
#> 75   38 2002-02-07 5.903855837  9.078165 10.882010  6.258983
#> 77   39 2002-02-08 5.771272356  9.139416 11.094398  6.276493
#> 79   40 2002-02-09 3.574521063  9.201226 11.310918  5.024780
#> 82   41 2002-02-10 4.181449309  9.263572 11.531528  5.485469
#> 83   42 2002-02-11 1.889822650  9.326432 11.756182  4.130128
#> 86   43 2002-02-12 2.241967742  9.389784 11.984831  4.426997
#> 87   44 2002-02-13 6.901964964  9.453606 12.217425  7.514255
#> 89   45 2002-02-14 2.369955791  9.517878 12.453910  4.663992
#> 91   46 2002-02-15 2.954360289  9.582579 12.694229  5.130406
#> 94   47 2002-02-16 4.878565070  9.647690 12.938322  6.505853
#> 95   48 2002-02-17 4.673536489  9.713192 13.186128  6.468808
#> 97   49 2002-02-18 4.703844421  9.779067 13.437583  6.591212
#> 99   50 2002-02-19 3.956472656  9.845297 13.692618  6.174441
#> 102  51 2002-02-20 0.753652334  9.911864 13.951164  4.018182
#> 103  52 2002-02-21 5.359994642  9.978754 14.213148  7.370517
#> 105  53 2002-02-22 6.880672349 10.045948 14.478496  8.577931
#> 107  54 2002-02-23 5.546067693 10.113433 14.747130  7.730344
#> 110  55 2002-02-24 2.503614815 10.181192 15.018971  5.601369
#> 112  56 2002-02-25 1.260371667 10.249211 15.293936  4.763851
#> 113  57 2002-02-26 1.559494340 10.317477 15.571940  5.069840
#> 115  58 2002-02-27 2.695639274 10.385975 15.852898  6.020503
#> 117  59 2002-02-28 5.879764090 10.454693 16.136720  8.571860
#> 120  60 2002-03-01 5.707255055 10.523618 16.423316  8.559243
#> 121  61 2002-03-02 2.541059608 10.592738 16.712592  6.182714
#> 123  62 2002-03-03 6.339195809 10.662040 17.004454  9.306176
#> 125  63 2002-03-04 4.929673170 10.731513 17.298805  8.297927
#> 128  64 2002-03-05 2.081068342 10.801147 17.595546  6.093963
#> 129  65 2002-03-06 4.576830708 10.870930 17.894578  8.240592
#> 131  66 2002-03-07 0.459251327 10.940852 18.195799  4.930842
#> 133  67 2002-03-08 5.729895330 11.010902 18.499104  9.438093
#> 136  68 2002-03-09 3.165345448 11.081072 18.804391  7.386867
#> 138  69 2002-03-10 1.408556739 11.151350 19.111552  5.984904
#> 139  70 2002-03-11 4.458657040 11.221729 19.420480  8.713227
#> 141  71 2002-03-12 3.929892811 11.292198 19.731067  8.366154
#> 143  72 2002-03-13 2.265477794 11.362749 20.043203  7.008884
#> 146  73 2002-03-14 5.946182227 11.433374 20.356778 10.382694
#> 147  74 2002-03-15 3.031708040 11.504062 20.671679  7.891762
#> 149  75 2002-03-16 5.444097594 11.574807 20.987796 10.182651
#> 151  76 2002-03-17 6.802634196 11.645600 21.305014 11.548784
#> 154  77 2002-03-18 3.562667101 11.716433 21.623220  8.693339
#> 155  78 2002-03-19 1.483878894 11.787297 21.942300  6.866711
#> 157  79 2002-03-20 4.632159284 11.858185 22.262140  9.913662
#> 159  80 2002-03-21 0.179379495 11.929088 22.582624  5.815445
#> 162  81 2002-03-22 0.629280239 12.000000 22.903638  6.326443
#> 163  82 2002-03-23 6.568814200 12.070912 23.225065 12.125637
#> 166  83 2002-03-24 6.211200055 12.141815 23.546791 11.909431
#> 167  84 2002-03-25 2.814769043 12.212703 23.868700  8.717790
#> 169  85 2002-03-26 1.959904648 12.283567 24.190677  7.977541
#> 171  86 2002-03-27 1.004798163 12.354400 24.512606  7.124971
#> 174  87 2002-03-28 5.223724755 12.425193 24.834373 11.428952
#> 175  88 2002-03-29 0.726657745 12.495938 25.155863  7.020391
#> 177  89 2002-03-30 5.703427713 12.566626 25.476961 12.150665
#> 179  90 2002-03-31 3.098877233 12.637251 25.797553  9.612396
#> 182  91 2002-04-01 2.145057886 12.707802 26.117528  8.733682
#> 183  92 2002-04-02 2.455308025 12.778271 26.436771  9.149067
#> 185  93 2002-04-03 4.591504807 12.848650 26.755170 11.469314
#> 187  94 2002-04-04 4.238102138 12.918928 27.072615 11.208789
#> 190  95 2002-04-05 2.911666510 12.989098 27.388994  9.917039
#> 192  96 2002-04-06 4.319171169 13.059148 27.704199 11.507481
#> 193  97 2002-04-07 5.177229744 13.129070 28.018120 12.528769
#> 195  98 2002-04-08 0.585099010 13.198853 28.330649  7.710605
#> 197  99 2002-04-09 4.273927257 13.268487 28.641680 11.773322
#> 200 100 2002-04-10 1.856810334 13.337960 28.951106  9.252954
#> 201 101 2002-04-11 0.151547552 13.407262 29.258822  7.480068
#> 203 102 2002-04-12 2.078697973 13.476382 29.564726  9.671324
#> 205 103 2002-04-13 6.897621024 13.545307 29.868714 15.072140
#> 208 104 2002-04-14 6.768781199 13.614025 30.170685 15.042980
#> 209 105 2002-04-15 4.718211779 13.682523 30.470539 12.871288
#> 211 106 2002-04-16 3.157932859 13.750789 30.768176 11.225072
#> 213 107 2002-04-17 5.433631282 13.818808 31.063500 13.873044
#> 216 108 2002-04-18 1.060704988 13.886567 31.356413  9.036660
#> 218 109 2002-04-19 4.309713482 13.954052 31.646820 12.798770
#> 219 110 2002-04-20 5.253752640 14.021246 31.934628 13.966600
#> 221 111 2002-04-21 2.500893773 14.088136 32.219743 10.914724
#> 223 112 2002-04-22 5.405318146 14.154703 32.502075 14.331373
#> 226 113 2002-04-23 0.604094247 14.220933 32.781534  8.891650
#> 227 114 2002-04-24 6.358280688 14.286808 33.058030 15.620672
#> 229 115 2002-04-25 4.726527665 14.352310 33.331477 13.821259
#> 231 116 2002-04-26 3.140455667 14.417421 33.601789 12.060080
#> 234 117 2002-04-27 6.191683189 14.482122 33.868881 15.707367
#> 235 118 2002-04-28 1.836442742 14.546394 34.132670 10.687746
#> 237 119 2002-04-29 2.367340992 14.610216 34.393074 11.384679
#> 239 120 2002-04-30 1.866369228 14.673568 34.650012 10.866116
#> 242 121 2002-05-01 2.580743319 14.736428 34.903405 11.782112
#> 243 122 2002-05-02 5.094530334 14.798774 35.153175 14.839096
#> 245 123 2002-05-03 6.865795782 14.860584 35.399245 17.027282
#> 248 124 2002-05-04 2.285861791 14.921835 35.641540 11.640332
#> 249 125 2002-05-05 0.559039887 14.982502 35.879987  9.639389
#> 251 126 2002-05-06 6.589105384 15.042562 36.114511 16.938262
#> 253 127 2002-05-07 5.195047207 15.101989 36.345041 15.337563
#> 256 128 2002-05-08 2.324032335 15.160757 36.571508 11.945948
#> 257 129 2002-05-09 6.293838559 15.218840 36.793843 16.806613
#> 259 130 2002-05-10 3.870362384 15.276211 37.011977 13.941649
#> 261 131 2002-05-11 6.371933505 15.332843 37.225844 17.041510
#> 264 132 2002-05-12 1.551171798 15.388707 37.435378 11.245576
#> 265 133 2002-05-13 0.984006018 15.443775 37.640517 10.609269
#> 267 134 2002-05-14 4.385435541 15.498017 37.841197 14.814214
#> 269 135 2002-05-15 5.667578293 15.551405 38.037357 16.440536
#> 272 136 2002-05-16 6.603817558 15.603907 38.228936 17.646776
#> 274 137 2002-05-17 2.347675991 15.655493 38.415875 12.484364
#> 275 138 2002-05-18 3.466337159 15.706133 38.598118 13.908824
#> 277 139 2002-05-19 3.054301888 15.755795 38.775606 13.452278
#> 279 140 2002-05-20 1.115012350 15.804447 38.948286 11.110983
#> 282 141 2002-05-21 5.701182109 15.852058 39.116103 16.813066
#> 283 142 2002-05-22 2.371364098 15.898595 39.279005 12.749093
#> 285 143 2002-05-23 5.891228415 15.944026 39.436940 17.145100
#> 287 144 2002-05-24 0.605561723 15.988319 39.589859 10.647203
#> 290 145 2002-05-25 3.754380077 16.031441 39.737713 14.587487
#> 291 146 2002-05-26 3.147743349 16.073361 39.880455 13.875129
#> 293 147 2002-05-27 0.101121502 16.114046 40.018039 10.130074
#> 295 148 2002-05-28 3.921174960 16.153463 40.150421 14.910765
#> 298 149 2002-05-29 5.950183730 16.191581 40.277558 17.470114
#> 300 150 2002-05-30 1.708266983 16.228369 40.399408 12.226159
#> 301 151 2002-05-31 0.956850386 16.263796 40.515933 11.320823
#> 303 152 2002-06-01 0.194074892 16.297830 40.627093 10.398667
#> 305 153 2002-06-02 0.995608123 16.330443 40.732853 11.424881
#> 308 154 2002-06-03 2.563606332 16.361604 40.833176 13.407253
#> 309 155 2002-06-04 3.272007084 16.391286 40.928031 14.317008
#> 311 156 2002-06-05 4.565927499 16.419460 41.017385 15.957409
#> 313 157 2002-06-06 1.814118492 16.446100 41.101209 12.542176
#> 316 158 2002-06-07 4.871101883 16.471180 41.179475 16.383972
#> 317 159 2002-06-08 5.906279777 16.494677 41.252156 17.698658
#> 319 160 2002-06-09 2.380154778 16.516565 41.319229 13.307005
#> 321 161 2002-06-10 0.337609736 16.536824 41.380671 10.767574
#> 324 162 2002-06-11 5.868529785 16.555433 41.436461 17.703263
#> 325 163 2002-06-12 1.167741178 16.572373 41.486582 11.833283
#> 328 164 2002-06-13 4.157983132 16.587625 41.531016 15.587998
#> 329 165 2002-06-14 0.771372520 16.601175 41.569750 11.358205
#> 331 166 2002-06-15 6.348114167 16.613007 41.602770 18.349258
#> 333 167 2002-06-16 3.648740239 16.623108 41.630068 14.976377
#> 336 168 2002-06-17 3.101837576 16.631469 41.651634 14.297009
#> 337 169 2002-06-18 2.383380798 16.638080 41.667464 13.401268
#> 339 170 2002-06-19 2.289234685 16.642934 41.677552 13.285760
#> 341 171 2002-06-20 4.675435271 16.646025 41.681897 16.274154
#> 344 172 2002-06-21 3.858076106 16.647351 41.680500 15.249919
#> 345 173 2002-06-22 6.198646426 16.646909 41.673363 18.177092
#> 347 174 2002-06-23 5.046040092 16.644700 41.660491 16.730061
#> 349 175 2002-06-24 3.837817344 16.640727 41.641891 15.212365
#> 352 176 2002-06-25 2.556718227 16.634994 41.617572 13.602603
#> 354 177 2002-06-26 1.941061665 16.627507 41.587545 12.824309
#> 355 178 2002-06-27 3.351008352 16.618275 41.551823 14.577335
#> 357 179 2002-06-28 1.194959934 16.607306 41.510423 11.871023
#> 359 180 2002-06-29 3.321325910 16.594614 41.463360 14.515178
#> 362 181 2002-06-30 4.586019765 16.580211 41.410655 16.079674
#> 363 182 2002-07-01 6.821231979 16.564113 41.352328 18.852689
#> 365 183 2002-07-02 3.697402181 16.546336 41.288405 14.935202
#> 367 184 2002-07-03 2.023175810 16.526900 41.218909 12.827678
#> 370 185 2002-07-04 3.058722907 16.505823 41.143868 14.098188
#> 371 186 2002-07-05 5.107771533 16.483128 41.063311 16.628152
#> 373 187 2002-07-06 6.971099220 16.458837 40.977269 18.922227
#> 375 188 2002-07-07 2.812090087 16.432973 40.885775 13.719730
#> 378 189 2002-07-08 3.561725251 16.405563 40.788864 14.624943
#> 380 190 2002-07-09 5.643152172 16.376632 40.686570 17.181647
#> 381 191 2002-07-10 5.234334262 16.346207 40.578932 16.641766
#> 383 192 2002-07-11 0.256368269 16.314316 40.465990 10.434445
#> 385 193 2002-07-12 4.144149978 16.280989 40.347784 15.221992
#> 388 194 2002-07-13 5.253237555 16.246255 40.224356 16.559376
#> 389 195 2002-07-14 2.419825571 16.210144 40.095751 13.016654
#> 391 196 2002-07-15 4.224382862 16.172686 39.962013 15.209637
#> 393 197 2002-07-16 4.473516801 16.133915 39.823189 15.476767
#> 396 198 2002-07-17 6.185037801 16.093860 39.679327 17.544421
#> 397 199 2002-07-18 5.935578215 16.052554 39.530476 17.190996
#> 399 200 2002-07-19 5.031648878 16.010029 39.376687 16.031846
#> 401 201 2002-07-20 5.637044312 15.966317 39.218012 16.727630
#> 404 202 2002-07-21 1.525244024 15.921451 39.054503 11.634298
#> 405 203 2002-07-22 0.337674261 15.875462 38.886214 10.135112
#> 408 204 2002-07-23 2.425238556 15.828384 38.713200 12.644135
#> 409 205 2002-07-24 1.433426410 15.780249 38.535519 11.384100
#> 411 206 2002-07-25 2.067890891 15.731088 38.353226 12.109120
#> 413 207 2002-07-26 6.222557041 15.680933 38.166382 17.114247
#> 416 208 2002-07-27 0.145022986 15.629816 37.975044  9.669939
#> 417 209 2002-07-28 1.826441819 15.577768 37.779274 11.659566
#> 419 210 2002-07-29 3.730188907 15.524820 37.579134 13.909402
#> 421 211 2002-07-30 5.536721197 15.471001 37.374685 16.031448
#> 424 212 2002-07-31 5.530686481 15.416342 37.165992 15.958237
#> 425 213 2002-08-01 4.531277812 15.360873 36.953119 14.688649
#> 427 214 2002-08-02 3.086756767 15.304621 36.736132 12.888649
#> 429 215 2002-08-03 4.736843769 15.247616 36.515097 14.800687
#> 432 216 2002-08-04 3.707521536 15.189886 36.290082 13.501331
#> 434 217 2002-08-05 0.944819122 15.131457 36.061156 10.141131
#> 435 218 2002-08-06 4.057890830 15.072356 35.828388 13.780088
#> 437 219 2002-08-07 1.567892114 15.012610 35.591848 10.756539
#> 439 220 2002-08-08 1.928363443 14.952243 35.351609 11.117519
#> 442 221 2002-08-09 4.772439688 14.891281 35.107744 14.402697
#> 443 222 2002-08-10 0.684604814 14.829748 34.860325  9.519732
#> 445 223 2002-08-11 1.759696863 14.767666 34.609429 10.714366
#> 447 224 2002-08-12 0.516269256 14.705060 34.355129  9.191857
#> 450 225 2002-08-13 4.336863183 14.641952 34.097505 13.574120
#> 451 226 2002-08-14 6.309296044 14.578362 33.836632 15.781151
#> 453 227 2002-08-15 4.903855481 14.514313 33.572591 14.064623
#> 455 228 2002-08-16 0.624909819 14.449824 33.305462  9.046544
#> 458 229 2002-08-17 1.322634048 14.384916 33.035326  9.777563
#> 460 230 2002-08-18 3.311052407 14.319607 32.762264 11.978295
#> 461 231 2002-08-19 1.590223627 14.253916 32.486361  9.933744
#> 463 232 2002-08-20 6.767555999 14.187862 32.207701 15.733401
#> 465 233 2002-08-21 4.721694021 14.121461 31.926370 13.319091
#> 468 234 2002-08-22 4.925692849 14.054730 31.642455 13.455402
#> 469 235 2002-08-23 5.461186505 13.987686 31.356043 13.960152
#> 471 236 2002-08-24 6.108819669 13.920345 31.067223 14.583593
#> 473 237 2002-08-25 0.554105908 13.852721 30.776085  8.309540
#> 476 238 2002-08-26 2.808919345 13.784830 30.482721 10.726395
#> 477 239 2002-08-27 3.047115201 13.716686 30.187223 10.899800
#> 479 240 2002-08-28 3.592651900 13.648302 29.889684 11.406362
#> 481 241 2002-08-29 1.525462459 13.579692 29.590199  9.059544
#> 484 242 2002-08-30 1.812724699 13.510869 29.288864  9.287028
#> 485 243 2002-08-31 0.928946492 13.441846 28.985774  8.248026
#> 487 244 2002-09-01 1.483984499 13.372633 28.681027  8.761649
#> 490 245 2002-09-02 3.792539239 13.303244 28.374722 11.138267
#> 491 246 2002-09-03 1.150685783 13.233689 28.066958  8.236968
#> 493 247 2002-09-04 4.303489579 13.163980 27.757836 11.476672
#> 495 248 2002-09-05 2.860763655 13.094126 27.447457  9.860182
#> 498 249 2002-09-06 4.121097219 13.024138 27.135924 11.077155
#> 499 250 2002-09-07 2.025448830 12.954027 26.823338  8.802839
#> 501 251 2002-09-08 4.504122691 12.883802 26.509805 11.261309
#> 503 252 2002-09-09 0.605502350 12.813472 26.195428  7.167791
#> 506 253 2002-09-10 5.940309472 12.743047 25.880314 12.502273
#> 507 254 2002-09-11 1.427932479 12.672536 25.564567  7.831440
#> 509 255 2002-09-12 0.805348260 12.601947 25.248295  7.118841
#> 511 256 2002-09-13 6.712759128 12.531289 24.931604 12.910580
#> 514 257 2002-09-14 5.418190317 12.460572 24.614602 11.505195
#> 516 258 2002-09-15 3.263578591 12.389802 24.297397  9.274419
#> 517 259 2002-09-16 2.663592477 12.318988 23.980097  8.587494
#> 519 260 2002-09-17 6.956307768 12.248139 23.662812 12.635328
#> 521 261 2002-09-18 4.260030818 12.177262 23.345649  9.919973
#> 524 262 2002-09-19 6.303863887 12.106365 23.028718 11.752782
#> 525 263 2002-09-20 5.907775519 12.035456 22.712128 11.252318
#> 527 264 2002-09-21 1.473066509 11.964544 22.395989  6.977687
#> 529 265 2002-09-22 0.428248713 11.893635 22.080409  5.917622
#> 532 266 2002-09-23 0.298457270 11.822738 21.765497  5.716102
#> 533 267 2002-09-24 3.000396808 11.751861 21.451363  8.101241
#> 535 268 2002-09-25 5.851183712 11.681012 21.138114 10.578718
#> 537 269 2002-09-26 2.063183710 11.610198 20.825859  7.056888
#> 540 270 2002-09-27 6.059461448 11.539428 20.514705 10.514908
#> 542 271 2002-09-28 0.214963053 11.468711 20.204761  5.240543
#> 543 272 2002-09-29 1.101319561 11.398053 19.896131  5.935249
#> 545 273 2002-09-30 6.834124056 11.327464 19.588923 10.806458
#> 547 274 2002-10-01 4.088562306 11.256953 19.283241  8.322678
#> 550 275 2002-10-02 3.910389113 11.186528 18.979189  8.062002
#> 551 276 2002-10-03 5.501613360 11.116198 18.676871  9.290983
#> 553 277 2002-10-04 1.126868028 11.045973 18.376389  5.531442
#> 555 278 2002-10-05 2.375180243 10.975862 18.077844  6.475487
#> 558 279 2002-10-06 6.161025902 10.905874 17.781337  9.467915
#> 559 280 2002-10-07 5.668477252 10.836020 17.486965  8.945582
#> 561 281 2002-10-08 3.734187266 10.766311 17.194826  7.280633
#> 563 282 2002-10-09 2.153549145 10.696756 16.905016  5.927975
#> 566 283 2002-10-10 1.530248365 10.627367 16.617629  5.350804
#> 567 284 2002-10-11 5.543729952 10.558154 16.332759  8.371079
#> 570 285 2002-10-12 6.836860875 10.489131 16.050495  9.243515
#> 571 286 2002-10-13 0.682354403 10.420308 15.770928  4.459097
#> 573 287 2002-10-14 0.532752133 10.351698 15.494145  4.272241
#> 575 288 2002-10-15 0.214409488 10.283314 15.220231  3.963730
#> 578 289 2002-10-16 1.763435757 10.215170 14.949271  5.027657
#> 579 290 2002-10-17 3.554873263 10.147279 14.681345  6.241978
#> 581 291 2002-10-18 2.893585083 10.079655 14.416534  5.673424
#> 583 292 2002-10-19 0.473787279 10.012314 14.154915  3.873637
#> 586 293 2002-10-20 0.457034282  9.945270 13.896562  3.793448
#> 587 294 2002-10-21 2.627102702  9.878539 13.641550  5.224307
#> 589 295 2002-10-22 1.486037260  9.812138 13.389948  4.361433
#> 591 296 2002-10-23 6.629201027  9.746084 13.141825  7.754934
#> 594 297 2002-10-24 0.555792708  9.680393 12.897247  3.594555
#> 596 298 2002-10-25 5.759639296  9.615084 12.656277  6.954758
#> 597 299 2002-10-26 0.874462851  9.550176 12.418977  3.673317
#> 599 300 2002-10-27 5.666923532  9.485687 12.185404  6.686243
#> 601 301 2002-10-28 3.734648532  9.421638 11.955615  5.358450
#> 604 302 2002-10-29 4.747809808  9.358048 11.729663  5.907941
#> 605 303 2002-10-30 5.876400917  9.294940 11.507600  6.514539
#> 607 304 2002-10-31 2.844237336  9.232334 11.289473  4.561362
#> 609 305 2002-11-01 2.911861394  9.170252 11.075328  4.527225
#> 612 306 2002-11-02 2.230961108  9.108719 10.865209  4.046888
#> 613 307 2002-11-03 0.431140979  9.047757 10.659156  2.918753
#> 615 308 2002-11-04 3.256310559  8.987390 10.457208  4.508729
#> 617 309 2002-11-05 0.145013645  8.927644 10.259398  2.648172
#> 620 310 2002-11-06 4.465065615  8.868543 10.065762  5.050356
#> 622 311 2002-11-07 2.664774132  8.810114  9.876328  3.962717
#> 623 312 2002-11-08 5.061233402  8.752384  9.691126  5.224821
#> 625 313 2002-11-09 1.595530970  8.695379  9.510180  3.250065
#> 627 314 2002-11-10 1.489968644  8.639127  9.333513  3.138242
#> 630 315 2002-11-11 5.680727547  8.583658  9.161147  5.321744
#> 631 316 2002-11-12 3.718959933  8.528999  8.993100  4.208938
#> 633 317 2002-11-13 3.733010901  8.475180  8.829388  4.151860
#> 635 318 2002-11-14 4.376331537  8.422232  8.670025  4.420051
#> 638 319 2002-11-15 1.887049478  8.370184  8.515023  3.088608
#> 639 320 2002-11-16 1.911829494  8.319067  8.364392  3.052221
#> 641 321 2002-11-17 0.009589607  8.268912  8.218139  2.059300
#> 643 322 2002-11-18 6.652308213  8.219751  8.076271  5.287162
#> 646 323 2002-11-19 6.888120741  8.171616  7.938791  5.330630
#> 647 324 2002-11-20 3.983066921  8.124538  7.805702  3.864804
#> 650 325 2002-11-21 2.850881563  8.078549  7.677004  3.273840
#> 651 326 2002-11-22 6.452425353  8.033683  7.552698  4.921230
#> 653 327 2002-11-23 1.980128279  7.989971  7.432780  2.779216
#> 655 328 2002-11-24 0.455982693  7.947446  7.317248  2.039225
#> 658 329 2002-11-25 4.252517668  7.906140  7.206095  3.739514
#> 659 330 2002-11-26 6.600609051  7.866085  7.099317  4.753428
#> 661 331 2002-11-27 5.717745816  7.827314  6.996907  4.304799
#> 663 332 2002-11-28 4.229055071  7.789856  6.898857  3.597383
#> 666 333 2002-11-29 1.753567855  7.753745  6.805157  2.470808
#> 667 334 2002-11-30 5.931181726  7.719011  6.715800  4.259114
#> 669 335 2002-12-01 1.403344423  7.685684  6.630775  2.263057
#> 671 336 2002-12-02 3.373529711  7.653793  6.550072  3.081041
#> 674 337 2002-12-03 2.668603001  7.623368  6.473680  2.751494
#> 676 338 2002-12-04 3.057438505  7.594437  6.401590  2.889003
#> 677 339 2002-12-05 1.265092603  7.567027  6.333789  2.112904
#> 679 340 2002-12-06 3.322490525  7.541163  6.270268  2.948846
#> 681 341 2002-12-07 0.206629705  7.516872  6.211016  1.638121
#> 684 342 2002-12-08 5.665324077  7.494177  6.156023  3.865870
#> 685 343 2002-12-09 5.051031652  7.473100  6.105278  3.589583
#> 687 344 2002-12-10 6.225139382  7.453664  6.058772  4.044771
#> 689 345 2002-12-11 2.607151900  7.435887  6.016497  2.558869
#> 692 346 2002-12-12 2.343835559  7.419789  5.978443  2.438875
#> 693 347 2002-12-13 0.391743615  7.405386  5.944603  1.643385
#> 695 348 2002-12-14 2.118707435  7.392694  5.914970  2.326342
#> 697 349 2002-12-15 5.485445574  7.381725  5.889539  3.660677
#> 700 350 2002-12-16 0.429328293  7.372493  5.868303  1.637943
#> 702 351 2002-12-17 3.969139552  7.365006  5.851259  3.039491
#> 703 352 2002-12-18 5.390693348  7.359273  5.838404  3.597927
#> 705 353 2002-12-19 0.279529882  7.355300  5.829736  1.568210
#> 707 354 2002-12-20 0.512248491  7.353091  5.825253  1.659220
#> 710 355 2002-12-21 1.470242915  7.352649  5.824956  2.038621
#> 711 356 2002-12-22 5.769641161  7.353975  5.828845  3.743753
#> 713 357 2002-12-23 0.800619052  7.357066  5.836924  1.776827
#> 715 358 2002-12-24 4.327972631  7.361920  5.849194  3.181630
#> 718 359 2002-12-25 3.723941368  7.368531  5.865662  2.948623
#> 719 360 2002-12-26 4.361810666  7.376892  5.886331  3.211819
#> 721 361 2002-12-27 6.419668769  7.386993  5.911209  4.046371
#> 723 362 2002-12-28 6.851674899  7.398825  5.940303  4.235582
#> 726 363 2002-12-29 3.747936110  7.412375  5.973622  3.003634
#> 727 364 2002-12-30 0.155168234  7.427627  6.011174  1.565582
#> 729 365 2002-12-31 4.886359626  7.444567  6.052971  3.499724
#> 2     1 2003-01-01 5.346466128  7.463176  6.099033  3.709369
#> 4     2 2003-01-02 4.220213105  7.483435  6.149345  3.271269
#> 5     3 2003-01-03 2.636928513  7.505323  6.203934  2.640832
#> 8     4 2003-01-04 0.195266241  7.528820  6.262814  1.646919
#> 10    5 2003-01-05 6.118228219  7.553900  6.325996  4.143347
#> 12    6 2003-01-06 5.426044006  7.580540  6.393496  3.886561
#> 13    7 2003-01-07 0.210693771  7.608714  6.465324  1.705847
#> 16    8 2003-01-08 2.133529094  7.638396  6.541496  2.548947
#> 18    9 2003-01-09 5.585108983  7.669557  6.622023  4.066644
#> 20   10 2003-01-10 3.757857909  7.702170  6.706920  3.312870
#> 21   11 2003-01-11 3.755143018  7.736204  6.796198  3.348482
#> 24   12 2003-01-12 2.220060847  7.771631  6.889871  2.706555
#> 26   13 2003-01-13 4.109860013  7.808419  6.987950  3.585996
#> 28   14 2003-01-14 2.856462530  7.846537  7.090446  3.063219
#> 29   15 2003-01-15 0.828308818  7.885954  7.197371  2.177334
#> 31   16 2003-01-16 6.566517362  7.926639  7.308732  4.854501
#> 34   17 2003-01-17 0.059141654  7.968559  7.424539  1.883687
#> 36   18 2003-01-18 4.273028879  8.011681  7.544799  3.898209
#> 38   19 2003-01-19 5.796921838  8.055974  7.669519  4.676798
#> 39   20 2003-01-20 0.256513920  8.101405  7.798703  2.073140
#> 42   21 2003-01-21 3.654414976  8.147942  7.932354  3.761949
#> 44   22 2003-01-22 1.578354027  8.195553  8.070474  2.794751
#> 46   23 2003-01-23 0.021041591  8.244205  8.213063  2.063747
#> 47   24 2003-01-24 1.290562234  8.293867  8.360119  2.740465
#> 50   25 2003-01-25 5.177232418  8.344507  8.511639  4.768374
#> 52   26 2003-01-26 2.889118763  8.396093  8.667618  3.658180
#> 54   27 2003-01-27 0.872215553  8.448595  8.828046  2.662706
#> 55   28 2003-01-28 6.387470992  8.501983  8.992915  5.626381
#> 57   29 2003-01-29 4.309051740  8.556225  9.162213  4.597672
#> 60   30 2003-01-30 2.287111194  8.611293  9.335925  3.573766
#> 62   31 2003-01-31 1.480668145  8.667157  9.514035  3.191182
#> 64   32 2003-02-01 4.934802754  8.723789  9.696523  5.166657
#> 65   33 2003-02-02 1.217424934  8.781160  9.883368  3.155960
#> 68   34 2003-02-03 1.675824982  8.839243 10.074546  3.473649
#> 70   35 2003-02-04 0.384682714  8.898011 10.270030  2.789507
#> 72   36 2003-02-05 2.572630246  8.957438 10.469790  4.120941
#> 73   37 2003-02-06 6.684625771  9.017498 10.673795  6.624664
#> 76   38 2003-02-07 5.903855837  9.078165 10.882010  6.258983
#> 78   39 2003-02-08 5.771272356  9.139416 11.094398  6.276493
#> 80   40 2003-02-09 3.574521063  9.201226 11.310918  5.024780
#> 81   41 2003-02-10 4.181449309  9.263572 11.531528  5.485469
#> 84   42 2003-02-11 1.889822650  9.326432 11.756182  4.130128
#> 85   43 2003-02-12 2.241967742  9.389784 11.984831  4.426997
#> 88   44 2003-02-13 6.901964964  9.453606 12.217425  7.514255
#> 90   45 2003-02-14 2.369955791  9.517878 12.453910  4.663992
#> 92   46 2003-02-15 2.954360289  9.582579 12.694229  5.130406
#> 93   47 2003-02-16 4.878565070  9.647690 12.938322  6.505853
#> 96   48 2003-02-17 4.673536489  9.713192 13.186128  6.468808
#> 98   49 2003-02-18 4.703844421  9.779067 13.437583  6.591212
#> 100  50 2003-02-19 3.956472656  9.845297 13.692618  6.174441
#> 101  51 2003-02-20 0.753652334  9.911864 13.951164  4.018182
#> 104  52 2003-02-21 5.359994642  9.978754 14.213148  7.370517
#> 106  53 2003-02-22 6.880672349 10.045948 14.478496  8.577931
#> 108  54 2003-02-23 5.546067693 10.113433 14.747130  7.730344
#> 109  55 2003-02-24 2.503614815 10.181192 15.018971  5.601369
#> 111  56 2003-02-25 1.260371667 10.249211 15.293936  4.763851
#> 114  57 2003-02-26 1.559494340 10.317477 15.571940  5.069840
#> 116  58 2003-02-27 2.695639274 10.385975 15.852898  6.020503
#> 118  59 2003-02-28 5.879764090 10.454693 16.136720  8.571860
#> 119  60 2003-03-01 5.707255055 10.523618 16.423316  8.559243
#> 122  61 2003-03-02 2.541059608 10.592738 16.712592  6.182714
#> 124  62 2003-03-03 6.339195809 10.662040 17.004454  9.306176
#> 126  63 2003-03-04 4.929673170 10.731513 17.298805  8.297927
#> 127  64 2003-03-05 2.081068342 10.801147 17.595546  6.093963
#> 130  65 2003-03-06 4.576830708 10.870930 17.894578  8.240592
#> 132  66 2003-03-07 0.459251327 10.940852 18.195799  4.930842
#> 134  67 2003-03-08 5.729895330 11.010902 18.499104  9.438093
#> 135  68 2003-03-09 3.165345448 11.081072 18.804391  7.386867
#> 137  69 2003-03-10 1.408556739 11.151350 19.111552  5.984904
#> 140  70 2003-03-11 4.458657040 11.221729 19.420480  8.713227
#> 142  71 2003-03-12 3.929892811 11.292198 19.731067  8.366154
#> 144  72 2003-03-13 2.265477794 11.362749 20.043203  7.008884
#> 145  73 2003-03-14 5.946182227 11.433374 20.356778 10.382694
#> 148  74 2003-03-15 3.031708040 11.504062 20.671679  7.891762
#> 150  75 2003-03-16 5.444097594 11.574807 20.987796 10.182651
#> 152  76 2003-03-17 6.802634196 11.645600 21.305014 11.548784
#> 153  77 2003-03-18 3.562667101 11.716433 21.623220  8.693339
#> 156  78 2003-03-19 1.483878894 11.787297 21.942300  6.866711
#> 158  79 2003-03-20 4.632159284 11.858185 22.262140  9.913662
#> 160  80 2003-03-21 0.179379495 11.929088 22.582624  5.815445
#> 161  81 2003-03-22 0.629280239 12.000000 22.903638  6.326443
#> 164  82 2003-03-23 6.568814200 12.070912 23.225065 12.125637
#> 165  83 2003-03-24 6.211200055 12.141815 23.546791 11.909431
#> 168  84 2003-03-25 2.814769043 12.212703 23.868700  8.717790
#> 170  85 2003-03-26 1.959904648 12.283567 24.190677  7.977541
#> 172  86 2003-03-27 1.004798163 12.354400 24.512606  7.124971
#> 173  87 2003-03-28 5.223724755 12.425193 24.834373 11.428952
#> 176  88 2003-03-29 0.726657745 12.495938 25.155863  7.020391
#> 178  89 2003-03-30 5.703427713 12.566626 25.476961 12.150665
#> 180  90 2003-03-31 3.098877233 12.637251 25.797553  9.612396
#> 181  91 2003-04-01 2.145057886 12.707802 26.117528  8.733682
#> 184  92 2003-04-02 2.455308025 12.778271 26.436771  9.149067
#> 186  93 2003-04-03 4.591504807 12.848650 26.755170 11.469314
#> 188  94 2003-04-04 4.238102138 12.918928 27.072615 11.208789
#> 189  95 2003-04-05 2.911666510 12.989098 27.388994  9.917039
#> 191  96 2003-04-06 4.319171169 13.059148 27.704199 11.507481
#> 194  97 2003-04-07 5.177229744 13.129070 28.018120 12.528769
#> 196  98 2003-04-08 0.585099010 13.198853 28.330649  7.710605
#> 198  99 2003-04-09 4.273927257 13.268487 28.641680 11.773322
#> 199 100 2003-04-10 1.856810334 13.337960 28.951106  9.252954
#> 202 101 2003-04-11 0.151547552 13.407262 29.258822  7.480068
#> 204 102 2003-04-12 2.078697973 13.476382 29.564726  9.671324
#> 206 103 2003-04-13 6.897621024 13.545307 29.868714 15.072140
#> 207 104 2003-04-14 6.768781199 13.614025 30.170685 15.042980
#> 210 105 2003-04-15 4.718211779 13.682523 30.470539 12.871288
#> 212 106 2003-04-16 3.157932859 13.750789 30.768176 11.225072
#> 214 107 2003-04-17 5.433631282 13.818808 31.063500 13.873044
#> 215 108 2003-04-18 1.060704988 13.886567 31.356413  9.036660
#> 217 109 2003-04-19 4.309713482 13.954052 31.646820 12.798770
#> 220 110 2003-04-20 5.253752640 14.021246 31.934628 13.966600
#> 222 111 2003-04-21 2.500893773 14.088136 32.219743 10.914724
#> 224 112 2003-04-22 5.405318146 14.154703 32.502075 14.331373
#> 225 113 2003-04-23 0.604094247 14.220933 32.781534  8.891650
#> 228 114 2003-04-24 6.358280688 14.286808 33.058030 15.620672
#> 230 115 2003-04-25 4.726527665 14.352310 33.331477 13.821259
#> 232 116 2003-04-26 3.140455667 14.417421 33.601789 12.060080
#> 233 117 2003-04-27 6.191683189 14.482122 33.868881 15.707367
#> 236 118 2003-04-28 1.836442742 14.546394 34.132670 10.687746
#> 238 119 2003-04-29 2.367340992 14.610216 34.393074 11.384679
#> 240 120 2003-04-30 1.866369228 14.673568 34.650012 10.866116
#> 241 121 2003-05-01 2.580743319 14.736428 34.903405 11.782112
#> 244 122 2003-05-02 5.094530334 14.798774 35.153175 14.839096
#> 246 123 2003-05-03 6.865795782 14.860584 35.399245 17.027282
#> 247 124 2003-05-04 2.285861791 14.921835 35.641540 11.640332
#> 250 125 2003-05-05 0.559039887 14.982502 35.879987  9.639389
#> 252 126 2003-05-06 6.589105384 15.042562 36.114511 16.938262
#> 254 127 2003-05-07 5.195047207 15.101989 36.345041 15.337563
#> 255 128 2003-05-08 2.324032335 15.160757 36.571508 11.945948
#> 258 129 2003-05-09 6.293838559 15.218840 36.793843 16.806613
#> 260 130 2003-05-10 3.870362384 15.276211 37.011977 13.941649
#> 262 131 2003-05-11 6.371933505 15.332843 37.225844 17.041510
#> 263 132 2003-05-12 1.551171798 15.388707 37.435378 11.245576
#> 266 133 2003-05-13 0.984006018 15.443775 37.640517 10.609269
#> 268 134 2003-05-14 4.385435541 15.498017 37.841197 14.814214
#> 270 135 2003-05-15 5.667578293 15.551405 38.037357 16.440536
#> 271 136 2003-05-16 6.603817558 15.603907 38.228936 17.646776
#> 273 137 2003-05-17 2.347675991 15.655493 38.415875 12.484364
#> 276 138 2003-05-18 3.466337159 15.706133 38.598118 13.908824
#> 278 139 2003-05-19 3.054301888 15.755795 38.775606 13.452278
#> 280 140 2003-05-20 1.115012350 15.804447 38.948286 11.110983
#> 281 141 2003-05-21 5.701182109 15.852058 39.116103 16.813066
#> 284 142 2003-05-22 2.371364098 15.898595 39.279005 12.749093
#> 286 143 2003-05-23 5.891228415 15.944026 39.436940 17.145100
#> 288 144 2003-05-24 0.605561723 15.988319 39.589859 10.647203
#> 289 145 2003-05-25 3.754380077 16.031441 39.737713 14.587487
#> 292 146 2003-05-26 3.147743349 16.073361 39.880455 13.875129
#> 294 147 2003-05-27 0.101121502 16.114046 40.018039 10.130074
#> 296 148 2003-05-28 3.921174960 16.153463 40.150421 14.910765
#> 297 149 2003-05-29 5.950183730 16.191581 40.277558 17.470114
#> 299 150 2003-05-30 1.708266983 16.228369 40.399408 12.226159
#> 302 151 2003-05-31 0.956850386 16.263796 40.515933 11.320823
#> 304 152 2003-06-01 0.194074892 16.297830 40.627093 10.398667
#> 306 153 2003-06-02 0.995608123 16.330443 40.732853 11.424881
#> 307 154 2003-06-03 2.563606332 16.361604 40.833176 13.407253
#> 310 155 2003-06-04 3.272007084 16.391286 40.928031 14.317008
#> 312 156 2003-06-05 4.565927499 16.419460 41.017385 15.957409
#> 314 157 2003-06-06 1.814118492 16.446100 41.101209 12.542176
#> 315 158 2003-06-07 4.871101883 16.471180 41.179475 16.383972
#> 318 159 2003-06-08 5.906279777 16.494677 41.252156 17.698658
#> 320 160 2003-06-09 2.380154778 16.516565 41.319229 13.307005
#> 322 161 2003-06-10 0.337609736 16.536824 41.380671 10.767574
#> 323 162 2003-06-11 5.868529785 16.555433 41.436461 17.703263
#> 326 163 2003-06-12 1.167741178 16.572373 41.486582 11.833283
#> 327 164 2003-06-13 4.157983132 16.587625 41.531016 15.587998
#> 330 165 2003-06-14 0.771372520 16.601175 41.569750 11.358205
#> 332 166 2003-06-15 6.348114167 16.613007 41.602770 18.349258
#> 334 167 2003-06-16 3.648740239 16.623108 41.630068 14.976377
#> 335 168 2003-06-17 3.101837576 16.631469 41.651634 14.297009
#> 338 169 2003-06-18 2.383380798 16.638080 41.667464 13.401268
#> 340 170 2003-06-19 2.289234685 16.642934 41.677552 13.285760
#> 342 171 2003-06-20 4.675435271 16.646025 41.681897 16.274154
#> 343 172 2003-06-21 3.858076106 16.647351 41.680500 15.249919
#> 346 173 2003-06-22 6.198646426 16.646909 41.673363 18.177092
#> 348 174 2003-06-23 5.046040092 16.644700 41.660491 16.730061
#> 350 175 2003-06-24 3.837817344 16.640727 41.641891 15.212365
#> 351 176 2003-06-25 2.556718227 16.634994 41.617572 13.602603
#> 353 177 2003-06-26 1.941061665 16.627507 41.587545 12.824309
#> 356 178 2003-06-27 3.351008352 16.618275 41.551823 14.577335
#> 358 179 2003-06-28 1.194959934 16.607306 41.510423 11.871023
#> 360 180 2003-06-29 3.321325910 16.594614 41.463360 14.515178
#> 361 181 2003-06-30 4.586019765 16.580211 41.410655 16.079674
#> 364 182 2003-07-01 6.821231979 16.564113 41.352328 18.852689
#> 366 183 2003-07-02 3.697402181 16.546336 41.288405 14.935202
#> 368 184 2003-07-03 2.023175810 16.526900 41.218909 12.827678
#> 369 185 2003-07-04 3.058722907 16.505823 41.143868 14.098188
#> 372 186 2003-07-05 5.107771533 16.483128 41.063311 16.628152
#> 374 187 2003-07-06 6.971099220 16.458837 40.977269 18.922227
#> 376 188 2003-07-07 2.812090087 16.432973 40.885775 13.719730
#> 377 189 2003-07-08 3.561725251 16.405563 40.788864 14.624943
#> 379 190 2003-07-09 5.643152172 16.376632 40.686570 17.181647
#> 382 191 2003-07-10 5.234334262 16.346207 40.578932 16.641766
#> 384 192 2003-07-11 0.256368269 16.314316 40.465990 10.434445
#> 386 193 2003-07-12 4.144149978 16.280989 40.347784 15.221992
#> 387 194 2003-07-13 5.253237555 16.246255 40.224356 16.559376
#> 390 195 2003-07-14 2.419825571 16.210144 40.095751 13.016654
#> 392 196 2003-07-15 4.224382862 16.172686 39.962013 15.209637
#> 394 197 2003-07-16 4.473516801 16.133915 39.823189 15.476767
#> 395 198 2003-07-17 6.185037801 16.093860 39.679327 17.544421
#> 398 199 2003-07-18 5.935578215 16.052554 39.530476 17.190996
#> 400 200 2003-07-19 5.031648878 16.010029 39.376687 16.031846
#> 402 201 2003-07-20 5.637044312 15.966317 39.218012 16.727630
#> 403 202 2003-07-21 1.525244024 15.921451 39.054503 11.634298
#> 406 203 2003-07-22 0.337674261 15.875462 38.886214 10.135112
#> 407 204 2003-07-23 2.425238556 15.828384 38.713200 12.644135
#> 410 205 2003-07-24 1.433426410 15.780249 38.535519 11.384100
#> 412 206 2003-07-25 2.067890891 15.731088 38.353226 12.109120
#> 414 207 2003-07-26 6.222557041 15.680933 38.166382 17.114247
#> 415 208 2003-07-27 0.145022986 15.629816 37.975044  9.669939
#> 418 209 2003-07-28 1.826441819 15.577768 37.779274 11.659566
#> 420 210 2003-07-29 3.730188907 15.524820 37.579134 13.909402
#> 422 211 2003-07-30 5.536721197 15.471001 37.374685 16.031448
#> 423 212 2003-07-31 5.530686481 15.416342 37.165992 15.958237
#> 426 213 2003-08-01 4.531277812 15.360873 36.953119 14.688649
#> 428 214 2003-08-02 3.086756767 15.304621 36.736132 12.888649
#> 430 215 2003-08-03 4.736843769 15.247616 36.515097 14.800687
#> 431 216 2003-08-04 3.707521536 15.189886 36.290082 13.501331
#> 433 217 2003-08-05 0.944819122 15.131457 36.061156 10.141131
#> 436 218 2003-08-06 4.057890830 15.072356 35.828388 13.780088
#> 438 219 2003-08-07 1.567892114 15.012610 35.591848 10.756539
#> 440 220 2003-08-08 1.928363443 14.952243 35.351609 11.117519
#> 441 221 2003-08-09 4.772439688 14.891281 35.107744 14.402697
#> 444 222 2003-08-10 0.684604814 14.829748 34.860325  9.519732
#> 446 223 2003-08-11 1.759696863 14.767666 34.609429 10.714366
#> 448 224 2003-08-12 0.516269256 14.705060 34.355129  9.191857
#> 449 225 2003-08-13 4.336863183 14.641952 34.097505 13.574120
#> 452 226 2003-08-14 6.309296044 14.578362 33.836632 15.781151
#> 454 227 2003-08-15 4.903855481 14.514313 33.572591 14.064623
#> 456 228 2003-08-16 0.624909819 14.449824 33.305462  9.046544
#> 457 229 2003-08-17 1.322634048 14.384916 33.035326  9.777563
#> 459 230 2003-08-18 3.311052407 14.319607 32.762264 11.978295
#> 462 231 2003-08-19 1.590223627 14.253916 32.486361  9.933744
#> 464 232 2003-08-20 6.767555999 14.187862 32.207701 15.733401
#> 466 233 2003-08-21 4.721694021 14.121461 31.926370 13.319091
#> 467 234 2003-08-22 4.925692849 14.054730 31.642455 13.455402
#> 470 235 2003-08-23 5.461186505 13.987686 31.356043 13.960152
#> 472 236 2003-08-24 6.108819669 13.920345 31.067223 14.583593
#> 474 237 2003-08-25 0.554105908 13.852721 30.776085  8.309540
#> 475 238 2003-08-26 2.808919345 13.784830 30.482721 10.726395
#> 478 239 2003-08-27 3.047115201 13.716686 30.187223 10.899800
#> 480 240 2003-08-28 3.592651900 13.648302 29.889684 11.406362
#> 482 241 2003-08-29 1.525462459 13.579692 29.590199  9.059544
#> 483 242 2003-08-30 1.812724699 13.510869 29.288864  9.287028
#> 486 243 2003-08-31 0.928946492 13.441846 28.985774  8.248026
#> 488 244 2003-09-01 1.483984499 13.372633 28.681027  8.761649
#> 489 245 2003-09-02 3.792539239 13.303244 28.374722 11.138267
#> 492 246 2003-09-03 1.150685783 13.233689 28.066958  8.236968
#> 494 247 2003-09-04 4.303489579 13.163980 27.757836 11.476672
#> 496 248 2003-09-05 2.860763655 13.094126 27.447457  9.860182
#> 497 249 2003-09-06 4.121097219 13.024138 27.135924 11.077155
#> 500 250 2003-09-07 2.025448830 12.954027 26.823338  8.802839
#> 502 251 2003-09-08 4.504122691 12.883802 26.509805 11.261309
#> 504 252 2003-09-09 0.605502350 12.813472 26.195428  7.167791
#> 505 253 2003-09-10 5.940309472 12.743047 25.880314 12.502273
#> 508 254 2003-09-11 1.427932479 12.672536 25.564567  7.831440
#> 510 255 2003-09-12 0.805348260 12.601947 25.248295  7.118841
#> 512 256 2003-09-13 6.712759128 12.531289 24.931604 12.910580
#> 513 257 2003-09-14 5.418190317 12.460572 24.614602 11.505195
#> 515 258 2003-09-15 3.263578591 12.389802 24.297397  9.274419
#> 518 259 2003-09-16 2.663592477 12.318988 23.980097  8.587494
#> 520 260 2003-09-17 6.956307768 12.248139 23.662812 12.635328
#> 522 261 2003-09-18 4.260030818 12.177262 23.345649  9.919973
#> 523 262 2003-09-19 6.303863887 12.106365 23.028718 11.752782
#> 526 263 2003-09-20 5.907775519 12.035456 22.712128 11.252318
#> 528 264 2003-09-21 1.473066509 11.964544 22.395989  6.977687
#> 530 265 2003-09-22 0.428248713 11.893635 22.080409  5.917622
#> 531 266 2003-09-23 0.298457270 11.822738 21.765497  5.716102
#> 534 267 2003-09-24 3.000396808 11.751861 21.451363  8.101241
#> 536 268 2003-09-25 5.851183712 11.681012 21.138114 10.578718
#> 538 269 2003-09-26 2.063183710 11.610198 20.825859  7.056888
#> 539 270 2003-09-27 6.059461448 11.539428 20.514705 10.514908
#> 541 271 2003-09-28 0.214963053 11.468711 20.204761  5.240543
#> 544 272 2003-09-29 1.101319561 11.398053 19.896131  5.935249
#> 546 273 2003-09-30 6.834124056 11.327464 19.588923 10.806458
#> 548 274 2003-10-01 4.088562306 11.256953 19.283241  8.322678
#> 549 275 2003-10-02 3.910389113 11.186528 18.979189  8.062002
#> 552 276 2003-10-03 5.501613360 11.116198 18.676871  9.290983
#> 554 277 2003-10-04 1.126868028 11.045973 18.376389  5.531442
#> 556 278 2003-10-05 2.375180243 10.975862 18.077844  6.475487
#> 557 279 2003-10-06 6.161025902 10.905874 17.781337  9.467915
#> 560 280 2003-10-07 5.668477252 10.836020 17.486965  8.945582
#> 562 281 2003-10-08 3.734187266 10.766311 17.194826  7.280633
#> 564 282 2003-10-09 2.153549145 10.696756 16.905016  5.927975
#> 565 283 2003-10-10 1.530248365 10.627367 16.617629  5.350804
#> 568 284 2003-10-11 5.543729952 10.558154 16.332759  8.371079
#> 569 285 2003-10-12 6.836860875 10.489131 16.050495  9.243515
#> 572 286 2003-10-13 0.682354403 10.420308 15.770928  4.459097
#> 574 287 2003-10-14 0.532752133 10.351698 15.494145  4.272241
#> 576 288 2003-10-15 0.214409488 10.283314 15.220231  3.963730
#> 577 289 2003-10-16 1.763435757 10.215170 14.949271  5.027657
#> 580 290 2003-10-17 3.554873263 10.147279 14.681345  6.241978
#> 582 291 2003-10-18 2.893585083 10.079655 14.416534  5.673424
#> 584 292 2003-10-19 0.473787279 10.012314 14.154915  3.873637
#> 585 293 2003-10-20 0.457034282  9.945270 13.896562  3.793448
#> 588 294 2003-10-21 2.627102702  9.878539 13.641550  5.224307
#> 590 295 2003-10-22 1.486037260  9.812138 13.389948  4.361433
#> 592 296 2003-10-23 6.629201027  9.746084 13.141825  7.754934
#> 593 297 2003-10-24 0.555792708  9.680393 12.897247  3.594555
#> 595 298 2003-10-25 5.759639296  9.615084 12.656277  6.954758
#> 598 299 2003-10-26 0.874462851  9.550176 12.418977  3.673317
#> 600 300 2003-10-27 5.666923532  9.485687 12.185404  6.686243
#> 602 301 2003-10-28 3.734648532  9.421638 11.955615  5.358450
#> 603 302 2003-10-29 4.747809808  9.358048 11.729663  5.907941
#> 606 303 2003-10-30 5.876400917  9.294940 11.507600  6.514539
#> 608 304 2003-10-31 2.844237336  9.232334 11.289473  4.561362
#> 610 305 2003-11-01 2.911861394  9.170252 11.075328  4.527225
#> 611 306 2003-11-02 2.230961108  9.108719 10.865209  4.046888
#> 614 307 2003-11-03 0.431140979  9.047757 10.659156  2.918753
#> 616 308 2003-11-04 3.256310559  8.987390 10.457208  4.508729
#> 618 309 2003-11-05 0.145013645  8.927644 10.259398  2.648172
#> 619 310 2003-11-06 4.465065615  8.868543 10.065762  5.050356
#> 621 311 2003-11-07 2.664774132  8.810114  9.876328  3.962717
#> 624 312 2003-11-08 5.061233402  8.752384  9.691126  5.224821
#> 626 313 2003-11-09 1.595530970  8.695379  9.510180  3.250065
#> 628 314 2003-11-10 1.489968644  8.639127  9.333513  3.138242
#> 629 315 2003-11-11 5.680727547  8.583658  9.161147  5.321744
#> 632 316 2003-11-12 3.718959933  8.528999  8.993100  4.208938
#> 634 317 2003-11-13 3.733010901  8.475180  8.829388  4.151860
#> 636 318 2003-11-14 4.376331537  8.422232  8.670025  4.420051
#> 637 319 2003-11-15 1.887049478  8.370184  8.515023  3.088608
#> 640 320 2003-11-16 1.911829494  8.319067  8.364392  3.052221
#> 642 321 2003-11-17 0.009589607  8.268912  8.218139  2.059300
#> 644 322 2003-11-18 6.652308213  8.219751  8.076271  5.287162
#> 645 323 2003-11-19 6.888120741  8.171616  7.938791  5.330630
#> 648 324 2003-11-20 3.983066921  8.124538  7.805702  3.864804
#> 649 325 2003-11-21 2.850881563  8.078549  7.677004  3.273840
#> 652 326 2003-11-22 6.452425353  8.033683  7.552698  4.921230
#> 654 327 2003-11-23 1.980128279  7.989971  7.432780  2.779216
#> 656 328 2003-11-24 0.455982693  7.947446  7.317248  2.039225
#> 657 329 2003-11-25 4.252517668  7.906140  7.206095  3.739514
#> 660 330 2003-11-26 6.600609051  7.866085  7.099317  4.753428
#> 662 331 2003-11-27 5.717745816  7.827314  6.996907  4.304799
#> 664 332 2003-11-28 4.229055071  7.789856  6.898857  3.597383
#> 665 333 2003-11-29 1.753567855  7.753745  6.805157  2.470808
#> 668 334 2003-11-30 5.931181726  7.719011  6.715800  4.259114
#> 670 335 2003-12-01 1.403344423  7.685684  6.630775  2.263057
#> 672 336 2003-12-02 3.373529711  7.653793  6.550072  3.081041
#> 673 337 2003-12-03 2.668603001  7.623368  6.473680  2.751494
#> 675 338 2003-12-04 3.057438505  7.594437  6.401590  2.889003
#> 678 339 2003-12-05 1.265092603  7.567027  6.333789  2.112904
#> 680 340 2003-12-06 3.322490525  7.541163  6.270268  2.948846
#> 682 341 2003-12-07 0.206629705  7.516872  6.211016  1.638121
#> 683 342 2003-12-08 5.665324077  7.494177  6.156023  3.865870
#> 686 343 2003-12-09 5.051031652  7.473100  6.105278  3.589583
#> 688 344 2003-12-10 6.225139382  7.453664  6.058772  4.044771
#> 690 345 2003-12-11 2.607151900  7.435887  6.016497  2.558869
#> 691 346 2003-12-12 2.343835559  7.419789  5.978443  2.438875
#> 694 347 2003-12-13 0.391743615  7.405386  5.944603  1.643385
#> 696 348 2003-12-14 2.118707435  7.392694  5.914970  2.326342
#> 698 349 2003-12-15 5.485445574  7.381725  5.889539  3.660677
#> 699 350 2003-12-16 0.429328293  7.372493  5.868303  1.637943
#> 701 351 2003-12-17 3.969139552  7.365006  5.851259  3.039491
#> 704 352 2003-12-18 5.390693348  7.359273  5.838404  3.597927
#> 706 353 2003-12-19 0.279529882  7.355300  5.829736  1.568210
#> 708 354 2003-12-20 0.512248491  7.353091  5.825253  1.659220
#> 709 355 2003-12-21 1.470242915  7.352649  5.824956  2.038621
#> 712 356 2003-12-22 5.769641161  7.353975  5.828845  3.743753
#> 714 357 2003-12-23 0.800619052  7.357066  5.836924  1.776827
#> 716 358 2003-12-24 4.327972631  7.361920  5.849194  3.181630
#> 717 359 2003-12-25 3.723941368  7.368531  5.865662  2.948623
#> 720 360 2003-12-26 4.361810666  7.376892  5.886331  3.211819
#> 722 361 2003-12-27 6.419668769  7.386993  5.911209  4.046371
#> 724 362 2003-12-28 6.851674899  7.398825  5.940303  4.235582
#> 725 363 2003-12-29 3.747936110  7.412375  5.973622  3.003634
#> 728 364 2003-12-30 0.155168234  7.427627  6.011174  1.565582
#> 730 365 2003-12-31 4.886359626  7.444567  6.052971  3.499724