Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.975295  3.094523  3.406165  2.775878  4.468359  2.113073  4.490627
#>   [8]  1.747444  3.801190  4.460268  4.558809  4.606020  1.968204  3.083885
#>  [15]  4.004109  3.003517  3.038518  2.186463  2.807534  2.690816  3.840246
#>  [22]  4.058936  4.911061  3.231163  4.620963  5.149249  2.874099  5.854003
#>  [29]  3.337229  2.744502  5.931883  2.732103  4.696661  2.582465  6.597440
#>  [36]  6.526917  5.223002  6.032925  4.259529  4.002247  3.713938  6.827964
#>  [43]  7.408414  5.641570  3.469598  3.852593  6.920400  3.472714  4.907637
#>  [50]  7.517962  3.958698  4.077555  5.525800  6.098750  4.027189  5.553366
#>  [57]  4.511673  6.058569  7.731170  8.666181  9.640969  7.022303  6.823275
#>  [64]  5.971812  9.030620  6.871803  7.477278  5.882990  7.939201  6.881307
#>  [71]  7.488885  6.493934  9.329862  6.627064  8.964465  7.599426  5.520854
#>  [78]  9.127563  6.300006  6.975263  8.499074  6.998145 11.951089 10.960765
#>  [85] 11.003243 12.672630 11.612675 11.557890 12.888812  8.917462  6.899510
#>  [92] 11.539150  9.415588 10.280467  8.497107  9.178254 14.060785 13.352951
#>  [99] 14.599959 12.530325 14.890507 10.022364 13.153680 11.077546 11.241507
#> [106] 14.008852 12.315991 14.369845 13.909971 15.852228 15.602823 12.791807
#> [113] 11.653650 11.281293 14.047968 16.260251  9.480517  8.910239 15.607943
#> [120] 15.770825  8.793153 15.587982  9.149386 16.226736 12.803296 11.751487
#> [127] 13.059131 12.758507  9.465260  9.809815 14.895122 17.331592 17.780108
#> [134]  9.514342 14.697045 12.643504 11.036494 15.296817  9.972755 10.990404
#> [141] 16.951407 15.258390 17.584386 13.409536 17.335528 15.284240 12.602836
#> [148] 18.591250 10.425489 11.656984 13.707054 10.537321 15.054344 11.798564
#> [155] 10.803005 16.685127 15.026151 11.430731 16.713464 10.869009 13.709222
#> [162] 11.161911 13.477943 17.347026 15.668013 15.615038 17.877987 11.543975
#> [169] 11.966692 12.355682 13.692232 15.750690 10.517173 10.514477 16.766870
#> [176] 15.325256 14.461534 15.163560 14.069155 16.101685 15.847358 13.162970
#> [183] 17.627913 12.540499 15.881606 15.381473 11.839149 12.797465 14.030756
#> [190] 12.006326 18.101367 15.855098 12.484122 17.792132 16.848674 17.856331
#> [197] 13.098782 15.253766 17.271572 14.657872 13.914324 14.602811  9.858836
#> [204] 13.673023 15.458241 13.113331 17.901543 13.124210 10.210974 16.166371
#> [211] 11.892785 14.426254 16.719303 17.398316 10.088533 10.446527 14.622343
#> [218]  9.003995 13.815211 10.601106 16.074602 11.920347 16.800946  8.589381
#> [225] 11.974666 10.644941 11.763343 11.232390 13.502011 12.225124 11.671804
#> [232]  8.472870 11.323199 11.524891 15.414356 13.270037  9.573697  9.301561
#> [239] 10.199944 10.324887 14.157686 12.939665 12.699023 10.954854 12.367078
#> [246]  7.839547  9.033918 12.186688 11.480945  8.607870 12.141590 11.013204
#> [253] 12.309852  8.294338 11.244089  7.612670 10.989290 11.439392  6.523793
#> [260]  8.756115  7.677836  6.378391  8.734232 11.709192  5.533878 10.347415
#> [267]  8.296169  6.877127  5.607099 11.086344 10.350180  6.566850  9.047878
#> [274]  5.512943  7.906852  7.288232  6.278826  9.577393  6.949118  7.315090
#> [281]  9.217622  5.394910  8.287945  9.454542  6.127264  4.322403  4.554658
#> [288]  5.899222  4.417608  4.143067  7.678938  7.911011  3.525593  4.346443
#> [295]  7.655891  7.486892  6.210650  5.815431  3.328515  5.290010  5.284865
#> [302]  3.063362  5.815148  4.125475  5.288068  4.028476  3.709145  4.524922
#> [309]  5.210496  5.531749  5.451599  5.812108  3.980642  4.340056  3.910838
#> [316]  5.202587  4.840079  3.088812  3.536283  3.588467  2.611152  3.240233
#> [323]  2.537126  2.397693  2.572854  2.585776  3.285802  4.608179  3.453458
#> [330]  2.282051  3.493605  4.762193  1.753195  2.149138  2.676188  1.788626
#> [337]  2.748463  3.214938  2.159512  4.064613  3.971627  3.157415  2.059711
#> [344]  1.600235  2.749844  2.617505  3.301724  1.816227  1.597269  2.143375
#> [351]  2.502871  2.731934  3.037358  1.811543  2.860227  3.572598  3.846592
#> [358]  4.179632  2.272967  3.699044  2.264405  1.976238  3.104405  4.291676
#> [365]  3.998313  1.975295  3.094523  3.406165  2.775878  4.468359  2.113073
#> [372]  4.490627  1.747444  3.801190  4.460268  4.558809  4.606020  1.968204
#> [379]  3.083885  4.004109  3.003517  3.038518  2.186463  2.807534  2.690816
#> [386]  3.840246  4.058936  4.911061  3.231163  4.620963  5.149249  2.874099
#> [393]  5.854003  3.337229  2.744502  5.931883  2.732103  4.696661  2.582465
#> [400]  6.597440  6.526917  5.223002  6.032925  4.259529  4.002247  3.713938
#> [407]  6.827964  7.408414  5.641570  3.469598  3.852593  6.920400  3.472714
#> [414]  4.907637  7.517962  3.958698  4.077555  5.525800  6.098750  4.027189
#> [421]  5.553366  4.511673  6.058569  7.731170  8.666181  9.640969  7.022303
#> [428]  6.823275  5.971812  9.030620  6.871803  7.477278  5.882990  7.939201
#> [435]  6.881307  7.488885  6.493934  9.329862  6.627064  8.964465  7.599426
#> [442]  5.520854  9.127563  6.300006  6.975263  8.499074  6.998145 11.951089
#> [449] 10.960765 11.003243 12.672630 11.612675 11.557890 12.888812  8.917462
#> [456]  6.899510 11.539150  9.415588 10.280467  8.497107  9.178254 14.060785
#> [463] 13.352951 14.599959 12.530325 14.890507 10.022364 13.153680 11.077546
#> [470] 11.241507 14.008852 12.315991 14.369845 13.909971 15.852228 15.602823
#> [477] 12.791807 11.653650 11.281293 14.047968 16.260251  9.480517  8.910239
#> [484] 15.607943 15.770825  8.793153 15.587982  9.149386 16.226736 12.803296
#> [491] 11.751487 13.059131 12.758507  9.465260  9.809815 14.895122 17.331592
#> [498] 17.780108  9.514342 14.697045 12.643504 11.036494 15.296817  9.972755
#> [505] 10.990404 16.951407 15.258390 17.584386 13.409536 17.335528 15.284240
#> [512] 12.602836 18.591250 10.425489 11.656984 13.707054 10.537321 15.054344
#> [519] 11.798564 10.803005 16.685127 15.026151 11.430731 16.713464 10.869009
#> [526] 13.709222 11.161911 13.477943 17.347026 15.668013 15.615038 17.877987
#> [533] 11.543975 11.966692 12.355682 13.692232 15.750690 10.517173 10.514477
#> [540] 16.766870 15.325256 14.461534 15.163560 14.069155 16.101685 15.847358
#> [547] 13.162970 17.627913 12.540499 15.881606 15.381473 11.839149 12.797465
#> [554] 14.030756 12.006326 18.101367 15.855098 12.484122 17.792132 16.848674
#> [561] 17.856331 13.098782 15.253766 17.271572 14.657872 13.914324 14.602811
#> [568]  9.858836 13.673023 15.458241 13.113331 17.901543 13.124210 10.210974
#> [575] 16.166371 11.892785 14.426254 16.719303 17.398316 10.088533 10.446527
#> [582] 14.622343  9.003995 13.815211 10.601106 16.074602 11.920347 16.800946
#> [589]  8.589381 11.974666 10.644941 11.763343 11.232390 13.502011 12.225124
#> [596] 11.671804  8.472870 11.323199 11.524891 15.414356 13.270037  9.573697
#> [603]  9.301561 10.199944 10.324887 14.157686 12.939665 12.699023 10.954854
#> [610] 12.367078  7.839547  9.033918 12.186688 11.480945  8.607870 12.141590
#> [617] 11.013204 12.309852  8.294338 11.244089  7.612670 10.989290 11.439392
#> [624]  6.523793  8.756115  7.677836  6.378391  8.734232 11.709192  5.533878
#> [631] 10.347415  8.296169  6.877127  5.607099 11.086344 10.350180  6.566850
#> [638]  9.047878  5.512943  7.906852  7.288232  6.278826  9.577393  6.949118
#> [645]  7.315090  9.217622  5.394910  8.287945  9.454542  6.127264  4.322403
#> [652]  4.554658  5.899222  4.417608  4.143067  7.678938  7.911011  3.525593
#> [659]  4.346443  7.655891  7.486892  6.210650  5.815431  3.328515  5.290010
#> [666]  5.284865  3.063362  5.815148  4.125475  5.288068  4.028476  3.709145
#> [673]  4.524922  5.210496  5.531749  5.451599  5.812108  3.980642  4.340056
#> [680]  3.910838  5.202587  4.840079  3.088812  3.536283  3.588467  2.611152
#> [687]  3.240233  2.537126  2.397693  2.572854  2.585776  3.285802  4.608179
#> [694]  3.453458  2.282051  3.493605  4.762193  1.753195  2.149138  2.676188
#> [701]  1.788626  2.748463  3.214938  2.159512  4.064613  3.971627  3.157415
#> [708]  2.059711  1.600235  2.749844  2.617505  3.301724  1.816227  1.597269
#> [715]  2.143375  2.502871  2.731934  3.037358  1.811543  2.860227  3.572598
#> [722]  3.846592  4.179632  2.272967  3.699044  2.264405  1.976238  3.104405
#> [729]  4.291676  3.998313
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates    sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 4.218025697  7.463176  6.099033  3.248279
#> 3     2 2002-01-02 2.350310716  7.483435  6.149345  2.502994
#> 6     3 2002-01-03 5.635220543  7.505323  6.203934  3.880033
#> 7     4 2002-01-04 5.351562292  7.528820  6.262814  3.791540
#> 9     5 2002-01-05 6.028261216  7.553900  6.325996  4.105676
#> 11    6 2002-01-06 6.379830886  7.580540  6.393496  4.288777
#> 14    7 2002-01-07 5.755360525  7.608714  6.465324  4.061571
#> 15    8 2002-01-08 2.654378710  7.638396  6.541496  2.771974
#> 17    9 2002-01-09 1.250950453  7.669557  6.622023  2.195551
#> 19   10 2002-01-10 1.714485450  7.702170  6.706920  2.423202
#> 22   11 2002-01-11 3.302940199  7.736204  6.796198  3.149854
#> 23   12 2002-01-12 5.463881827  7.771631  6.889871  4.144446
#> 25   13 2002-01-13 6.872662003  7.808419  6.987950  4.822246
#> 27   14 2002-01-14 6.721420841  7.846537  7.090446  4.809485
#> 30   15 2002-01-15 4.924143260  7.885954  7.197371  4.046432
#> 32   16 2002-01-16 0.001466272  7.926639  7.308732  1.827859
#> 33   17 2002-01-17 1.188836980  7.968559  7.424539  2.409972
#> 35   18 2002-01-18 5.745376100  8.011681  7.544799  4.591482
#> 37   19 2002-01-19 1.438952263  8.055974  7.669519  2.602342
#> 40   20 2002-01-20 6.087951251  8.101405  7.798703  4.879915
#> 41   21 2002-01-21 5.691682552  8.147942  7.932354  4.753631
#> 43   22 2002-01-22 5.700842574  8.195553  8.070474  4.824537
#> 45   23 2002-01-23 4.729649344  8.244205  8.213063  4.409157
#> 48   24 2002-01-24 0.722758523  8.293867  8.360119  2.454296
#> 49   25 2002-01-25 6.421414232  8.344507  8.511639  5.402924
#> 51   26 2002-01-26 1.417095419  8.396093  8.667618  2.898366
#> 53   27 2002-01-27 0.091289767  8.448595  8.828046  2.254706
#> 56   28 2002-01-28 6.707902594  8.501983  8.992915  5.795848
#> 58   29 2002-01-29 6.038678029  8.556225  9.162213  5.523735
#> 59   30 2002-01-30 3.438845109  8.611293  9.335925  4.198092
#> 61   31 2002-01-31 3.611597589  8.667157  9.514035  4.360754
#> 63   32 2002-02-01 4.477981275  8.723789  9.696523  4.912777
#> 66   33 2002-02-02 6.219222408  8.781160  9.883368  5.970771
#> 67   34 2002-02-03 6.846978475  8.839243 10.074546  6.420566
#> 69   35 2002-02-04 3.515063129  8.898011 10.270030  4.596040
#> 71   36 2002-02-05 6.230447005  8.957438 10.469790  6.258638
#> 74   37 2002-02-06 1.395006160  9.017498 10.673795  3.494066
#> 75   38 2002-02-07 0.550756876  9.078165 10.882010  3.050599
#> 77   39 2002-02-08 4.334600321  9.139416 11.094398  5.404500
#> 79   40 2002-02-09 2.641469302  9.201226 11.310918  4.451287
#> 82   41 2002-02-10 4.954859186  9.263572 11.531528  5.966849
#> 83   42 2002-02-11 4.512614884  9.326432 11.756182  5.783173
#> 86   43 2002-02-12 0.430211422  9.389784 11.984831  3.270762
#> 87   44 2002-02-13 6.607118283  9.453606 12.217425  7.323731
#> 89   45 2002-02-14 2.913605612  9.517878 12.453910  5.019668
#> 91   46 2002-02-15 6.896918186  9.582579 12.694229  7.741798
#> 94   47 2002-02-16 1.897690138  9.647690 12.938322  4.507058
#> 95   48 2002-02-17 3.043781735  9.713192 13.186128  5.362573
#> 97   49 2002-02-18 4.168892888  9.779067 13.437583  6.223669
#> 99   50 2002-02-19 2.690810873  9.845297 13.692618  5.294314
#> 102  51 2002-02-20 4.078149359  9.911864 13.951164  6.357833
#> 103  52 2002-02-21 6.816455732  9.978754 14.213148  8.407766
#> 105  53 2002-02-22 0.596380766 10.045948 14.478496  4.049384
#> 107  54 2002-02-23 0.099445763 10.113433 14.747130  3.759287
#> 110  55 2002-02-24 4.120855951 10.181192 15.018971  6.794221
#> 112  56 2002-02-25 2.842552606 10.249211 15.293936  5.944321
#> 113  57 2002-02-26 0.257937328 10.317477 15.571940  4.087635
#> 115  58 2002-02-27 0.825786097 10.385975 15.852898  4.593454
#> 117  59 2002-02-28 5.323738605 10.454693 16.136720  8.142750
#> 120  60 2002-03-01 5.089677375 10.523618 16.423316  8.077342
#> 121  61 2002-03-02 6.320918286 10.592738 16.712592  9.164533
#> 123  62 2002-03-03 2.422320955 10.662040 17.004454  6.182744
#> 125  63 2002-03-04 0.737566581 10.731513 17.298805  4.919166
#> 128  64 2002-03-05 6.135701386 10.801147 17.595546  9.396551
#> 129  65 2002-03-06 3.957363001 10.870930 17.894578  7.730741
#> 131  66 2002-03-07 5.002253442 10.940852 18.195799  8.708589
#> 133  67 2002-03-08 5.374239810 11.010902 18.499104  9.139330
#> 136  68 2002-03-09 3.366093179 11.081072 18.804391  7.557199
#> 138  69 2002-03-10 4.341247552 11.151350 19.111552  8.497975
#> 139  70 2002-03-11 2.455163144 11.221729 19.420480  6.979590
#> 141  71 2002-03-12 4.820943577 11.292198 19.731067  9.144629
#> 143  72 2002-03-13 6.177523599 11.362749 20.043203 10.459190
#> 146  73 2002-03-14 3.093284306 11.433374 20.356778  7.842944
#> 147  74 2002-03-15 2.884934733 11.504062 20.671679  7.759893
#> 149  75 2002-03-16 0.920445371 11.574807 20.987796  6.081439
#> 151  76 2002-03-17 2.915815989 11.645600 21.305014  7.993419
#> 154  77 2002-03-18 1.013852384 11.716433 21.623220  6.341361
#> 155  78 2002-03-19 5.172739574 11.787297 21.942300 10.300157
#> 157  79 2002-03-20 3.442604237 11.858185 22.262140  8.797047
#> 159  80 2002-03-21 4.202485326 11.929088 22.582624  9.623460
#> 162  81 2002-03-22 2.103288508 12.000000 22.903638  7.733116
#> 163  82 2002-03-23 5.496203793 12.070912 23.225065 11.093758
#> 166  83 2002-03-24 4.587696616 12.141815 23.546791 10.335190
#> 167  84 2002-03-25 1.641943365 12.212703 23.868700  7.571695
#> 169  85 2002-03-26 2.432948720 12.283567 24.190677  8.443336
#> 171  86 2002-03-27 1.058169709 12.354400 24.512606  7.177919
#> 174  87 2002-03-28 2.277619783 12.425193 24.834373  8.484745
#> 175  88 2002-03-29 6.587070655 12.495938 25.155863 12.919258
#> 177  89 2002-03-30 6.532327242 12.566626 25.476961 12.990900
#> 179  90 2002-03-31 2.979059650 12.637251 25.797553  9.490099
#> 182  91 2002-04-01 6.604773045 12.707802 26.117528 13.316565
#> 183  92 2002-04-02 3.335872958 12.778271 26.436771 10.059961
#> 185  93 2002-04-03 4.654210890 12.848650 26.755170 11.534602
#> 187  94 2002-04-04 2.843534737 12.918928 27.072615  9.747577
#> 190  95 2002-04-05 2.529678463 12.989098 27.388994  9.514306
#> 192  96 2002-04-06 0.892109613 13.059148 27.704199  7.872328
#> 193  97 2002-04-07 5.127616950 13.129070 28.018120 12.475831
#> 195  98 2002-04-08 6.996872225 13.198853 28.330649 14.591873
#> 197  99 2002-04-09 4.930323236 13.268487 28.641680 12.481778
#> 200 100 2002-04-10 3.008076727 13.337960 28.951106 10.502412
#> 201 101 2002-04-11 2.314761337 13.407262 29.258822  9.840471
#> 203 102 2002-04-12 0.159819188 13.476382 29.564726  7.566489
#> 205 103 2002-04-13 6.458478360 13.545307 29.868714 14.587964
#> 208 104 2002-04-14 6.810957315 13.614025 30.170685 15.089714
#> 209 105 2002-04-15 3.914013838 13.682523 30.470539 11.975826
#> 211 106 2002-04-16 1.937724346 13.750789 30.768176  9.859929
#> 213 107 2002-04-17 3.300353488 13.818808 31.063500 11.475331
#> 216 108 2002-04-18 4.021040461 13.886567 31.356413 12.378937
#> 218 109 2002-04-19 5.435460265 13.954052 31.646820 14.075328
#> 219 110 2002-04-20 2.445954583 14.021246 31.934628 10.769096
#> 221 111 2002-04-21 6.822804823 14.088136 32.219743 15.856856
#> 223 112 2002-04-22 0.576854802 14.154703 32.502075  8.787807
#> 226 113 2002-04-23 6.292793965 14.220933 32.781534 15.448333
#> 227 114 2002-04-24 1.773125401 14.286808 33.058030 10.315912
#> 229 115 2002-04-25 1.406931345 14.352310 33.331477  9.966582
#> 231 116 2002-04-26 2.099477617 14.417421 33.601789 10.847008
#> 234 117 2002-04-27 4.337533938 14.482122 33.868881 13.539247
#> 235 118 2002-04-28 4.293976727 14.546394 34.132670 13.571010
#> 237 119 2002-04-29 4.536967778 14.610216 34.393074 13.938377
#> 239 120 2002-04-30 3.490357023 14.673568 34.650012 12.783550
#> 242 121 2002-05-01 1.298726949 14.736428 34.903405 10.263876
#> 243 122 2002-05-02 2.385214576 14.798774 35.153175 11.621227
#> 245 123 2002-05-03 5.443336120 14.860584 35.399245 15.333069
#> 248 124 2002-05-04 4.711774437 14.921835 35.641540 14.537538
#> 249 125 2002-05-05 3.040357125 14.982502 35.879987 12.610509
#> 251 126 2002-05-06 1.923491065 15.042562 36.114511 11.337607
#> 253 127 2002-05-07 2.076513312 15.101989 36.345041 11.584970
#> 256 128 2002-05-08 0.837676536 15.160757 36.571508 10.153219
#> 257 129 2002-05-09 0.980442530 15.218840 36.793843 10.383645
#> 259 130 2002-05-10 5.253401524 15.276211 37.011977 15.617097
#> 261 131 2002-05-11 6.342674886 15.332843 37.225844 17.005993
#> 264 132 2002-05-12 2.460640862 15.388707 37.435378 12.351787
#> 265 133 2002-05-13 3.124755928 15.443775 37.640517 13.218053
#> 267 134 2002-05-14 1.318808831 15.498017 37.841197 11.070354
#> 269 135 2002-05-15 4.716360412 15.551405 38.037357 15.277239
#> 272 136 2002-05-16 5.049121182 15.603907 38.228936 15.742304
#> 274 137 2002-05-17 4.631598995 15.655493 38.415875 15.286541
#> 275 138 2002-05-18 6.472602041 15.706133 38.598118 17.602800
#> 277 139 2002-05-19 6.070583629 15.755795 38.775606 17.163869
#> 279 140 2002-05-20 1.769570091 15.804447 38.948286 11.917525
#> 282 141 2002-05-21 0.349454425 15.852058 39.116103 10.210178
#> 283 142 2002-05-22 6.432474652 15.898595 39.279005 17.765775
#> 285 143 2002-05-23 2.248065760 15.944026 39.436940 12.639488
#> 287 144 2002-05-24 1.471806938 15.988319 39.589859 11.719690
#> 290 145 2002-05-25 2.360133399 16.031441 39.737713 12.859502
#> 291 146 2002-05-26 3.619628088 16.073361 39.880455 14.460538
#> 293 147 2002-05-27 4.774555809 16.114046 40.018039 15.933138
#> 295 148 2002-05-28 6.369006412 16.153463 40.150421 17.952883
#> 298 149 2002-05-29 3.247477293 16.191581 40.277558 14.108539
#> 300 150 2002-05-30 0.901773856 16.228369 40.399408 11.222304
#> 301 151 2002-05-31 6.433595332 16.263796 40.515933 18.142583
#> 303 152 2002-06-01 0.537016976 16.297830 40.627093 10.826109
#> 305 153 2002-06-02 3.386744159 16.330443 40.732853 14.406973
#> 308 154 2002-06-03 1.923592852 16.361604 40.833176 12.608621
#> 309 155 2002-06-04 0.047389833 16.391286 40.928031 10.291173
#> 311 156 2002-06-05 2.518466091 16.419460 41.017385 13.400031
#> 313 157 2002-06-06 4.213676462 16.446100 41.101209 15.540599
#> 316 158 2002-06-07 6.689473284 16.471180 41.179475 18.657020
#> 317 159 2002-06-08 6.000313662 16.494677 41.252156 17.816244
#> 319 160 2002-06-09 0.953232299 16.516565 41.319229 11.522150
#> 321 161 2002-06-10 5.757379242 16.536824 41.380671 17.548613
#> 324 162 2002-06-11 1.712939545 16.555433 41.436461 12.502767
#> 325 163 2002-06-12 1.711797111 16.572373 41.486582 12.514266
#> 328 164 2002-06-13 4.438163790 16.587625 41.531016 15.938747
#> 329 165 2002-06-14 3.424212669 16.601175 41.569750 14.679594
#> 331 166 2002-06-15 1.354784395 16.613007 41.602770 12.097038
#> 333 167 2002-06-16 4.513224004 16.623108 41.630068 16.058861
#> 336 168 2002-06-17 5.490847560 16.631469 41.651634 17.288512
#> 337 169 2002-06-18 1.315403251 16.638080 41.667464 12.063976
#> 339 170 2002-06-19 0.261794883 16.642934 41.677552 10.747184
#> 341 171 2002-06-20 0.061636631 16.646025 41.681897 10.497644
#> 344 172 2002-06-21 4.853320623 16.647351 41.680500 16.495831
#> 345 173 2002-06-22 1.309010799 16.646909 41.673363 12.056810
#> 347 174 2002-06-23 6.642586834 16.644700 41.660491 18.728082
#> 349 175 2002-06-24 6.505938904 16.640727 41.641891 18.550730
#> 352 176 2002-06-25 6.801692838 16.634994 41.617572 18.912660
#> 354 177 2002-06-26 3.095387362 16.627507 41.587545 14.267868
#> 355 178 2002-06-27 0.712566079 16.618275 41.551823 11.278795
#> 357 179 2002-06-28 1.121584516 16.607306 41.510423 11.779321
#> 359 180 2002-06-29 5.956279378 16.594614 41.463360 17.807030
#> 362 181 2002-06-30 1.083993516 16.580211 41.410655 11.706352
#> 363 182 2002-07-01 1.165206472 16.564113 41.352328 11.792552
#> 365 183 2002-07-02 1.161082212 16.546336 41.288405 11.770737
#> 367 184 2002-07-03 4.316793745 16.526900 41.218909 15.687876
#> 370 185 2002-07-04 1.553896320 16.505823 41.143868 12.222656
#> 371 186 2002-07-05 4.583022735 16.483128 41.063311 15.974516
#> 373 187 2002-07-06 1.184716310 16.458837 40.977269 11.719101
#> 375 188 2002-07-07 2.886485089 16.432973 40.885775 13.812279
#> 378 189 2002-07-08 5.515140577 16.405563 40.788864 17.053315
#> 380 190 2002-07-09 4.130905689 16.376632 40.686570 15.303112
#> 381 191 2002-07-10 4.301387777 16.346207 40.578932 15.483762
#> 383 192 2002-07-11 1.403719073 16.314316 40.465990 11.857388
#> 385 193 2002-07-12 3.217764623 16.280989 40.347784 14.074102
#> 388 194 2002-07-13 2.440028817 16.246255 40.224356 13.076742
#> 389 195 2002-07-14 6.561118152 16.210144 40.095751 18.138393
#> 391 196 2002-07-15 6.491596428 16.172686 39.962013 18.010731
#> 393 197 2002-07-16 3.037822752 16.133915 39.823189 13.704912
#> 396 198 2002-07-17 1.422439959 16.093860 39.679327 11.673341
#> 397 199 2002-07-18 1.081612366 16.052554 39.530476 11.214390
#> 399 200 2002-07-19 6.013812438 16.010029 39.376687 17.239662
#> 401 201 2002-07-20 2.746278714 15.966317 39.218012 13.177341
#> 404 202 2002-07-21 5.624729963 15.921451 39.054503 16.662213
#> 405 203 2002-07-22 4.195613581 15.875462 38.886214 14.860035
#> 408 204 2002-07-23 2.348303811 15.828384 38.713200 12.550051
#> 409 205 2002-07-24 5.958259068 15.780249 38.535519 16.908942
#> 411 206 2002-07-25 2.353394791 15.731088 38.353226 12.457157
#> 413 207 2002-07-26 3.769250399 15.680933 38.166382 14.128652
#> 416 208 2002-07-27 4.471825660 15.629816 37.975044 14.926255
#> 417 209 2002-07-28 5.162544357 15.577768 37.779274 15.704932
#> 419 210 2002-07-29 2.995091790 15.524820 37.579134 13.019719
#> 421 211 2002-07-30 5.759564141 15.471001 37.374685 16.300619
#> 424 212 2002-07-31 5.346667747 15.416342 37.165992 15.736419
#> 425 213 2002-08-01 2.184446559 15.360873 36.953119 11.865803
#> 427 214 2002-08-02 2.914980125 15.304621 36.736132 12.682489
#> 429 215 2002-08-03 1.724335391 15.247616 36.515097 11.193499
#> 432 216 2002-08-04 1.710277890 15.189886 36.290082 11.115529
#> 434 217 2002-08-05 2.175719579 15.131457 36.061156 11.607867
#> 435 218 2002-08-06 2.872744079 15.072356 35.828388 12.371486
#> 437 219 2002-08-07 2.107042101 15.012610 35.591848 11.395646
#> 439 220 2002-08-08 1.987707453 14.952243 35.351609 11.187672
#> 442 221 2002-08-09 3.320487209 14.891281 35.107744 12.691133
#> 443 222 2002-08-10 0.188974112 14.829748 34.860325  8.937192
#> 445 223 2002-08-11 0.831197130 14.767666 34.609429  9.626352
#> 447 224 2002-08-12 4.534992892 14.705060 34.355129 13.886288
#> 450 225 2002-08-13 2.502407747 14.641952 34.097505 11.438122
#> 451 226 2002-08-14 5.653033376 14.578362 33.836632 15.019552
#> 453 227 2002-08-15 1.404332883 14.514313 33.572591 10.017306
#> 455 228 2002-08-16 6.746073060 14.449824 33.305462 16.100892
#> 458 229 2002-08-17 6.082119748 14.384916 33.035326 15.242703
#> 460 230 2002-08-18 5.525917383 14.319607 32.762264 14.512023
#> 461 231 2002-08-19 2.155253167 14.253916 32.486361 10.577629
#> 463 232 2002-08-20 3.812782858 14.187862 32.207701 12.379603
#> 465 233 2002-08-21 2.058005774 14.121461 31.926370 10.308004
#> 468 234 2002-08-22 5.320126927 14.054730 31.642455 13.899412
#> 469 235 2002-08-23 0.497383161 13.987686 31.356043  8.396500
#> 471 236 2002-08-24 5.471596280 13.920345 31.067223 13.872520
#> 473 237 2002-08-25 1.682638278 13.852721 30.776085  9.563149
#> 476 238 2002-08-26 0.159216830 13.784830 30.482721  7.796720
#> 477 239 2002-08-27 1.284872721 13.716686 30.187223  8.960658
#> 479 240 2002-08-28 3.944394176 13.648302 29.889684 11.791519
#> 481 241 2002-08-29 1.090554160 13.579692 29.590199  8.585711
#> 484 242 2002-08-30 3.063329049 13.510869 29.288864 10.642558
#> 485 243 2002-08-31 0.132692818 13.441846 28.985774  7.389512
#> 487 244 2002-09-01 0.025618678 13.372633 28.681027  7.197730
#> 490 245 2002-09-02 2.366089029 13.303244 28.374722  9.617017
#> 491 246 2002-09-03 5.865941786 13.233689 28.066958 13.237195
#> 493 247 2002-09-04 0.008197031 13.163980 27.757836  6.948101
#> 495 248 2002-09-05 2.583635072 13.094126 27.447457  9.569728
#> 498 249 2002-09-06 0.048793440 13.024138 27.135924  6.834812
#> 499 250 2002-09-07 6.191431744 12.954027 26.823338 13.115998
#> 501 251 2002-09-08 4.175683721 12.883802 26.509805 10.923410
#> 503 252 2002-09-09 3.492897340 12.813472 26.195428 10.119238
#> 506 253 2002-09-10 0.698747876 12.743047 25.880314  7.179635
#> 507 254 2002-09-11 0.847024361 12.672536 25.564567  7.245502
#> 509 255 2002-09-12 0.125141687 12.601947 25.248295  6.437436
#> 511 256 2002-09-13 2.230350421 12.531289 24.931604  8.451596
#> 514 257 2002-09-14 3.557105622 12.460572 24.614602  9.667002
#> 516 258 2002-09-15 5.420657080 12.389802 24.297397 11.389521
#> 517 259 2002-09-16 5.932486910 12.318988 23.980097 11.769103
#> 519 260 2002-09-17 0.207304377 12.248139 23.662812  6.115954
#> 521 261 2002-09-18 1.887469641 12.177262 23.345649  7.645694
#> 524 262 2002-09-19 4.158145277 12.106365 23.028718  9.711990
#> 525 263 2002-09-20 0.761703555 12.035456 22.712128  6.396738
#> 527 264 2002-09-21 0.233411112 11.964544 22.395989  5.817454
#> 529 265 2002-09-22 3.574864688 11.893635 22.080409  8.838452
#> 532 266 2002-09-23 5.715422823 11.822738 21.765497 10.702381
#> 533 267 2002-09-24 6.058531311 11.751861 21.451363 10.892337
#> 535 268 2002-09-25 6.356969235 11.681012 21.138114 11.036356
#> 537 269 2002-09-26 1.906317284 11.610198 20.825859  6.916198
#> 540 270 2002-09-27 6.628592010 11.539428 20.514705 11.020806
#> 542 271 2002-09-28 4.871494930 11.468711 20.204761  9.342317
#> 543 272 2002-09-29 0.829985858 11.398053 19.896131  5.698433
#> 545 273 2002-09-30 4.542662139 11.327464 19.588923  8.825111
#> 547 274 2002-10-01 2.956903441 11.256953 19.283241  7.353408
#> 550 275 2002-10-02 5.948623402 11.186528 18.979189  9.791048
#> 551 276 2002-10-03 4.140383233 11.116198 18.676871  8.147448
#> 553 277 2002-10-04 3.246177797 11.045973 18.376389  7.294313
#> 555 278 2002-10-05 6.333286001 10.975862 18.077844  9.735095
#> 558 279 2002-10-06 4.172329761 10.905874 17.781337  7.846694
#> 559 280 2002-10-07 0.453603757 10.836020 17.486965  4.737750
#> 561 281 2002-10-08 0.848784184 10.766311 17.194826  4.976501
#> 563 282 2002-10-09 6.736581398 10.696756 16.905016  9.549457
#> 566 283 2002-10-10 5.808569201 10.627367 16.617629  8.695732
#> 567 284 2002-10-11 5.678667163 10.558154 16.332759  8.475449
#> 570 285 2002-10-12 2.977306077 10.489131 16.050495  6.290565
#> 571 286 2002-10-13 0.466854048 10.420308 15.770928  4.296019
#> 573 287 2002-10-14 2.232430891 10.351698 15.494145  5.544258
#> 575 288 2002-10-15 1.698201336 10.283314 15.220231  5.061803
#> 578 289 2002-10-16 0.683212012 10.215170 14.949271  4.237237
#> 579 290 2002-10-17 1.423320913 10.147279 14.681345  4.699985
#> 581 291 2002-10-18 2.375077084 10.079655 14.416534  5.302623
#> 583 292 2002-10-19 4.612421778 10.012314 14.154915  6.799136
#> 586 293 2002-10-20 1.155167689  9.945270 13.896562  4.281201
#> 587 294 2002-10-21 2.079116258  9.878539 13.641550  4.845942
#> 589 295 2002-10-22 6.747920525  9.812138 13.389948  7.951698
#> 591 296 2002-10-23 3.395849496  9.746084 13.141825  5.574974
#> 594 297 2002-10-24 2.907252101  9.680393 12.897247  5.160987
#> 596 298 2002-10-25 5.162400663  9.615084 12.656277  6.561688
#> 597 299 2002-10-26 5.779791278  9.550176 12.418977  6.862743
#> 599 300 2002-10-27 6.283666709  9.485687 12.185404  7.082380
#> 601 301 2002-10-28 4.356036209  9.421638 11.955615  5.752706
#> 604 302 2002-10-29 5.310800898  9.358048 11.729663  6.260776
#> 605 303 2002-10-30 2.596615217  9.294940 11.507600  4.484270
#> 607 304 2002-10-31 3.621186691  9.232334 11.289473  5.036396
#> 609 305 2002-11-01 3.759525961  9.170252 11.075328  5.039107
#> 612 306 2002-11-02 2.768241242  9.108719 10.865209  4.367331
#> 613 307 2002-11-03 6.516555501  9.047757 10.659156  6.503364
#> 615 308 2002-11-04 3.445559008  8.987390 10.457208  4.618829
#> 617 309 2002-11-05 4.923284085  8.927644 10.259398  5.393700
#> 620 310 2002-11-06 4.743495603  8.868543 10.065762  5.208365
#> 622 311 2002-11-07 1.389956403  8.810114  9.876328  3.248168
#> 623 312 2002-11-08 5.083678877  8.752384  9.691126  5.237247
#> 625 313 2002-11-09 3.903173036  8.695379  9.510180  4.512005
#> 627 314 2002-11-10 4.432926817  8.639127  9.333513  4.727994
#> 630 315 2002-11-11 5.151877331  8.583658  9.161147  5.039529
#> 631 316 2002-11-12 6.933435163  8.528999  8.993100  5.903632
#> 633 317 2002-11-13 1.458557163  8.475180  8.829388  2.967105
#> 635 318 2002-11-14 5.248632400  8.422232  8.670025  4.869034
#> 638 319 2002-11-15 6.742362284  8.370184  8.515023  5.558273
#> 639 320 2002-11-16 0.003901486  8.319067  8.364392  2.093059
#> 641 321 2002-11-17 0.017124578  8.268912  8.218139  2.063044
#> 643 322 2002-11-18 4.341518395  8.219751  8.076271  4.151935
#> 646 323 2002-11-19 6.382501164  8.171616  7.938791  5.085023
#> 647 324 2002-11-20 6.640175541  8.124538  7.805702  5.141221
#> 650 325 2002-11-21 2.633275308  8.078549  7.677004  3.170445
#> 651 326 2002-11-22 5.602188285  8.033683  7.552698  4.521564
#> 653 327 2002-11-23 4.027339105  7.989971  7.432780  3.731439
#> 655 328 2002-11-24 1.361747528  7.947446  7.317248  2.456195
#> 658 329 2002-11-25 2.857169457  7.906140  7.206095  3.103615
#> 659 330 2002-11-26 4.828869787  7.866085  7.099317  3.953911
#> 661 331 2002-11-27 4.588119972  7.827314  6.996907  3.799908
#> 663 332 2002-11-28 4.774691253  7.789856  6.898857  3.838996
#> 666 333 2002-11-29 4.552335268  7.753745  6.805157  3.698992
#> 667 334 2002-11-30 6.906555703  7.719011  6.715800  4.683418
#> 669 335 2002-12-01 1.725578418  7.685684  6.630775  2.402060
#> 671 336 2002-12-02 1.572032213  7.653793  6.550072  2.310186
#> 674 337 2002-12-03 0.032943771  7.623368  6.473680  1.632408
#> 676 338 2002-12-04 4.167511935  7.594437  6.401590  3.356861
#> 677 339 2002-12-05 6.178511561  7.567027  6.333789  4.169231
#> 679 340 2002-12-06 5.286428550  7.541163  6.270268  3.765326
#> 681 341 2002-12-07 3.596070545  7.516872  6.211016  3.038429
#> 684 342 2002-12-08 3.225743171  7.494177  6.156023  2.863884
#> 685 343 2002-12-09 0.005631276  7.473100  6.105278  1.528620
#> 687 344 2002-12-10 5.994859724  7.453664  6.058772  3.951179
#> 689 345 2002-12-11 0.886944425  7.435887  6.016497  1.862945
#> 692 346 2002-12-12 6.922263789  7.419789  5.978443  4.283393
#> 693 347 2002-12-13 0.820995779  7.405386  5.944603  1.815674
#> 695 348 2002-12-14 3.307251029  7.392694  5.914970  2.801825
#> 697 349 2002-12-15 5.899310592  7.381725  5.889539  3.825779
#> 700 350 2002-12-16 0.530205437  7.372493  5.868303  1.678090
#> 702 351 2002-12-17 4.912255621  7.365006  5.851259  3.414129
#> 703 352 2002-12-18 2.283240192  7.359273  5.838404  2.365294
#> 705 353 2002-12-19 2.794665916  7.355300  5.829736  2.564946
#> 707 354 2002-12-20 0.796329900  7.353091  5.825253  1.771747
#> 710 355 2002-12-21 2.378266219  7.352649  5.824956  2.398300
#> 711 356 2002-12-22 3.652361379  7.353975  5.828845  2.904663
#> 713 357 2002-12-23 5.068798506  7.357066  5.836924  3.469964
#> 715 358 2002-12-24 2.493088013  7.361920  5.849194  2.452703
#> 718 359 2002-12-25 6.831483717  7.368531  5.865662  4.185490
#> 719 360 2002-12-26 4.526711828  7.376892  5.886331  3.277609
#> 721 361 2002-12-27 0.814352876  7.386993  5.911209  1.803632
#> 723 362 2002-12-28 4.360284353  7.398825  5.940303  3.235449
#> 726 363 2002-12-29 6.528902256  7.412375  5.973622  4.124222
#> 727 364 2002-12-30 6.129454987  7.427627  6.011174  3.983076
#> 729 365 2002-12-31 1.907395097  7.444567  6.052971  2.288668
#> 2     1 2003-01-01 4.218025697  7.463176  6.099033  3.248279
#> 4     2 2003-01-02 2.350310716  7.483435  6.149345  2.502994
#> 5     3 2003-01-03 5.635220543  7.505323  6.203934  3.880033
#> 8     4 2003-01-04 5.351562292  7.528820  6.262814  3.791540
#> 10    5 2003-01-05 6.028261216  7.553900  6.325996  4.105676
#> 12    6 2003-01-06 6.379830886  7.580540  6.393496  4.288777
#> 13    7 2003-01-07 5.755360525  7.608714  6.465324  4.061571
#> 16    8 2003-01-08 2.654378710  7.638396  6.541496  2.771974
#> 18    9 2003-01-09 1.250950453  7.669557  6.622023  2.195551
#> 20   10 2003-01-10 1.714485450  7.702170  6.706920  2.423202
#> 21   11 2003-01-11 3.302940199  7.736204  6.796198  3.149854
#> 24   12 2003-01-12 5.463881827  7.771631  6.889871  4.144446
#> 26   13 2003-01-13 6.872662003  7.808419  6.987950  4.822246
#> 28   14 2003-01-14 6.721420841  7.846537  7.090446  4.809485
#> 29   15 2003-01-15 4.924143260  7.885954  7.197371  4.046432
#> 31   16 2003-01-16 0.001466272  7.926639  7.308732  1.827859
#> 34   17 2003-01-17 1.188836980  7.968559  7.424539  2.409972
#> 36   18 2003-01-18 5.745376100  8.011681  7.544799  4.591482
#> 38   19 2003-01-19 1.438952263  8.055974  7.669519  2.602342
#> 39   20 2003-01-20 6.087951251  8.101405  7.798703  4.879915
#> 42   21 2003-01-21 5.691682552  8.147942  7.932354  4.753631
#> 44   22 2003-01-22 5.700842574  8.195553  8.070474  4.824537
#> 46   23 2003-01-23 4.729649344  8.244205  8.213063  4.409157
#> 47   24 2003-01-24 0.722758523  8.293867  8.360119  2.454296
#> 50   25 2003-01-25 6.421414232  8.344507  8.511639  5.402924
#> 52   26 2003-01-26 1.417095419  8.396093  8.667618  2.898366
#> 54   27 2003-01-27 0.091289767  8.448595  8.828046  2.254706
#> 55   28 2003-01-28 6.707902594  8.501983  8.992915  5.795848
#> 57   29 2003-01-29 6.038678029  8.556225  9.162213  5.523735
#> 60   30 2003-01-30 3.438845109  8.611293  9.335925  4.198092
#> 62   31 2003-01-31 3.611597589  8.667157  9.514035  4.360754
#> 64   32 2003-02-01 4.477981275  8.723789  9.696523  4.912777
#> 65   33 2003-02-02 6.219222408  8.781160  9.883368  5.970771
#> 68   34 2003-02-03 6.846978475  8.839243 10.074546  6.420566
#> 70   35 2003-02-04 3.515063129  8.898011 10.270030  4.596040
#> 72   36 2003-02-05 6.230447005  8.957438 10.469790  6.258638
#> 73   37 2003-02-06 1.395006160  9.017498 10.673795  3.494066
#> 76   38 2003-02-07 0.550756876  9.078165 10.882010  3.050599
#> 78   39 2003-02-08 4.334600321  9.139416 11.094398  5.404500
#> 80   40 2003-02-09 2.641469302  9.201226 11.310918  4.451287
#> 81   41 2003-02-10 4.954859186  9.263572 11.531528  5.966849
#> 84   42 2003-02-11 4.512614884  9.326432 11.756182  5.783173
#> 85   43 2003-02-12 0.430211422  9.389784 11.984831  3.270762
#> 88   44 2003-02-13 6.607118283  9.453606 12.217425  7.323731
#> 90   45 2003-02-14 2.913605612  9.517878 12.453910  5.019668
#> 92   46 2003-02-15 6.896918186  9.582579 12.694229  7.741798
#> 93   47 2003-02-16 1.897690138  9.647690 12.938322  4.507058
#> 96   48 2003-02-17 3.043781735  9.713192 13.186128  5.362573
#> 98   49 2003-02-18 4.168892888  9.779067 13.437583  6.223669
#> 100  50 2003-02-19 2.690810873  9.845297 13.692618  5.294314
#> 101  51 2003-02-20 4.078149359  9.911864 13.951164  6.357833
#> 104  52 2003-02-21 6.816455732  9.978754 14.213148  8.407766
#> 106  53 2003-02-22 0.596380766 10.045948 14.478496  4.049384
#> 108  54 2003-02-23 0.099445763 10.113433 14.747130  3.759287
#> 109  55 2003-02-24 4.120855951 10.181192 15.018971  6.794221
#> 111  56 2003-02-25 2.842552606 10.249211 15.293936  5.944321
#> 114  57 2003-02-26 0.257937328 10.317477 15.571940  4.087635
#> 116  58 2003-02-27 0.825786097 10.385975 15.852898  4.593454
#> 118  59 2003-02-28 5.323738605 10.454693 16.136720  8.142750
#> 119  60 2003-03-01 5.089677375 10.523618 16.423316  8.077342
#> 122  61 2003-03-02 6.320918286 10.592738 16.712592  9.164533
#> 124  62 2003-03-03 2.422320955 10.662040 17.004454  6.182744
#> 126  63 2003-03-04 0.737566581 10.731513 17.298805  4.919166
#> 127  64 2003-03-05 6.135701386 10.801147 17.595546  9.396551
#> 130  65 2003-03-06 3.957363001 10.870930 17.894578  7.730741
#> 132  66 2003-03-07 5.002253442 10.940852 18.195799  8.708589
#> 134  67 2003-03-08 5.374239810 11.010902 18.499104  9.139330
#> 135  68 2003-03-09 3.366093179 11.081072 18.804391  7.557199
#> 137  69 2003-03-10 4.341247552 11.151350 19.111552  8.497975
#> 140  70 2003-03-11 2.455163144 11.221729 19.420480  6.979590
#> 142  71 2003-03-12 4.820943577 11.292198 19.731067  9.144629
#> 144  72 2003-03-13 6.177523599 11.362749 20.043203 10.459190
#> 145  73 2003-03-14 3.093284306 11.433374 20.356778  7.842944
#> 148  74 2003-03-15 2.884934733 11.504062 20.671679  7.759893
#> 150  75 2003-03-16 0.920445371 11.574807 20.987796  6.081439
#> 152  76 2003-03-17 2.915815989 11.645600 21.305014  7.993419
#> 153  77 2003-03-18 1.013852384 11.716433 21.623220  6.341361
#> 156  78 2003-03-19 5.172739574 11.787297 21.942300 10.300157
#> 158  79 2003-03-20 3.442604237 11.858185 22.262140  8.797047
#> 160  80 2003-03-21 4.202485326 11.929088 22.582624  9.623460
#> 161  81 2003-03-22 2.103288508 12.000000 22.903638  7.733116
#> 164  82 2003-03-23 5.496203793 12.070912 23.225065 11.093758
#> 165  83 2003-03-24 4.587696616 12.141815 23.546791 10.335190
#> 168  84 2003-03-25 1.641943365 12.212703 23.868700  7.571695
#> 170  85 2003-03-26 2.432948720 12.283567 24.190677  8.443336
#> 172  86 2003-03-27 1.058169709 12.354400 24.512606  7.177919
#> 173  87 2003-03-28 2.277619783 12.425193 24.834373  8.484745
#> 176  88 2003-03-29 6.587070655 12.495938 25.155863 12.919258
#> 178  89 2003-03-30 6.532327242 12.566626 25.476961 12.990900
#> 180  90 2003-03-31 2.979059650 12.637251 25.797553  9.490099
#> 181  91 2003-04-01 6.604773045 12.707802 26.117528 13.316565
#> 184  92 2003-04-02 3.335872958 12.778271 26.436771 10.059961
#> 186  93 2003-04-03 4.654210890 12.848650 26.755170 11.534602
#> 188  94 2003-04-04 2.843534737 12.918928 27.072615  9.747577
#> 189  95 2003-04-05 2.529678463 12.989098 27.388994  9.514306
#> 191  96 2003-04-06 0.892109613 13.059148 27.704199  7.872328
#> 194  97 2003-04-07 5.127616950 13.129070 28.018120 12.475831
#> 196  98 2003-04-08 6.996872225 13.198853 28.330649 14.591873
#> 198  99 2003-04-09 4.930323236 13.268487 28.641680 12.481778
#> 199 100 2003-04-10 3.008076727 13.337960 28.951106 10.502412
#> 202 101 2003-04-11 2.314761337 13.407262 29.258822  9.840471
#> 204 102 2003-04-12 0.159819188 13.476382 29.564726  7.566489
#> 206 103 2003-04-13 6.458478360 13.545307 29.868714 14.587964
#> 207 104 2003-04-14 6.810957315 13.614025 30.170685 15.089714
#> 210 105 2003-04-15 3.914013838 13.682523 30.470539 11.975826
#> 212 106 2003-04-16 1.937724346 13.750789 30.768176  9.859929
#> 214 107 2003-04-17 3.300353488 13.818808 31.063500 11.475331
#> 215 108 2003-04-18 4.021040461 13.886567 31.356413 12.378937
#> 217 109 2003-04-19 5.435460265 13.954052 31.646820 14.075328
#> 220 110 2003-04-20 2.445954583 14.021246 31.934628 10.769096
#> 222 111 2003-04-21 6.822804823 14.088136 32.219743 15.856856
#> 224 112 2003-04-22 0.576854802 14.154703 32.502075  8.787807
#> 225 113 2003-04-23 6.292793965 14.220933 32.781534 15.448333
#> 228 114 2003-04-24 1.773125401 14.286808 33.058030 10.315912
#> 230 115 2003-04-25 1.406931345 14.352310 33.331477  9.966582
#> 232 116 2003-04-26 2.099477617 14.417421 33.601789 10.847008
#> 233 117 2003-04-27 4.337533938 14.482122 33.868881 13.539247
#> 236 118 2003-04-28 4.293976727 14.546394 34.132670 13.571010
#> 238 119 2003-04-29 4.536967778 14.610216 34.393074 13.938377
#> 240 120 2003-04-30 3.490357023 14.673568 34.650012 12.783550
#> 241 121 2003-05-01 1.298726949 14.736428 34.903405 10.263876
#> 244 122 2003-05-02 2.385214576 14.798774 35.153175 11.621227
#> 246 123 2003-05-03 5.443336120 14.860584 35.399245 15.333069
#> 247 124 2003-05-04 4.711774437 14.921835 35.641540 14.537538
#> 250 125 2003-05-05 3.040357125 14.982502 35.879987 12.610509
#> 252 126 2003-05-06 1.923491065 15.042562 36.114511 11.337607
#> 254 127 2003-05-07 2.076513312 15.101989 36.345041 11.584970
#> 255 128 2003-05-08 0.837676536 15.160757 36.571508 10.153219
#> 258 129 2003-05-09 0.980442530 15.218840 36.793843 10.383645
#> 260 130 2003-05-10 5.253401524 15.276211 37.011977 15.617097
#> 262 131 2003-05-11 6.342674886 15.332843 37.225844 17.005993
#> 263 132 2003-05-12 2.460640862 15.388707 37.435378 12.351787
#> 266 133 2003-05-13 3.124755928 15.443775 37.640517 13.218053
#> 268 134 2003-05-14 1.318808831 15.498017 37.841197 11.070354
#> 270 135 2003-05-15 4.716360412 15.551405 38.037357 15.277239
#> 271 136 2003-05-16 5.049121182 15.603907 38.228936 15.742304
#> 273 137 2003-05-17 4.631598995 15.655493 38.415875 15.286541
#> 276 138 2003-05-18 6.472602041 15.706133 38.598118 17.602800
#> 278 139 2003-05-19 6.070583629 15.755795 38.775606 17.163869
#> 280 140 2003-05-20 1.769570091 15.804447 38.948286 11.917525
#> 281 141 2003-05-21 0.349454425 15.852058 39.116103 10.210178
#> 284 142 2003-05-22 6.432474652 15.898595 39.279005 17.765775
#> 286 143 2003-05-23 2.248065760 15.944026 39.436940 12.639488
#> 288 144 2003-05-24 1.471806938 15.988319 39.589859 11.719690
#> 289 145 2003-05-25 2.360133399 16.031441 39.737713 12.859502
#> 292 146 2003-05-26 3.619628088 16.073361 39.880455 14.460538
#> 294 147 2003-05-27 4.774555809 16.114046 40.018039 15.933138
#> 296 148 2003-05-28 6.369006412 16.153463 40.150421 17.952883
#> 297 149 2003-05-29 3.247477293 16.191581 40.277558 14.108539
#> 299 150 2003-05-30 0.901773856 16.228369 40.399408 11.222304
#> 302 151 2003-05-31 6.433595332 16.263796 40.515933 18.142583
#> 304 152 2003-06-01 0.537016976 16.297830 40.627093 10.826109
#> 306 153 2003-06-02 3.386744159 16.330443 40.732853 14.406973
#> 307 154 2003-06-03 1.923592852 16.361604 40.833176 12.608621
#> 310 155 2003-06-04 0.047389833 16.391286 40.928031 10.291173
#> 312 156 2003-06-05 2.518466091 16.419460 41.017385 13.400031
#> 314 157 2003-06-06 4.213676462 16.446100 41.101209 15.540599
#> 315 158 2003-06-07 6.689473284 16.471180 41.179475 18.657020
#> 318 159 2003-06-08 6.000313662 16.494677 41.252156 17.816244
#> 320 160 2003-06-09 0.953232299 16.516565 41.319229 11.522150
#> 322 161 2003-06-10 5.757379242 16.536824 41.380671 17.548613
#> 323 162 2003-06-11 1.712939545 16.555433 41.436461 12.502767
#> 326 163 2003-06-12 1.711797111 16.572373 41.486582 12.514266
#> 327 164 2003-06-13 4.438163790 16.587625 41.531016 15.938747
#> 330 165 2003-06-14 3.424212669 16.601175 41.569750 14.679594
#> 332 166 2003-06-15 1.354784395 16.613007 41.602770 12.097038
#> 334 167 2003-06-16 4.513224004 16.623108 41.630068 16.058861
#> 335 168 2003-06-17 5.490847560 16.631469 41.651634 17.288512
#> 338 169 2003-06-18 1.315403251 16.638080 41.667464 12.063976
#> 340 170 2003-06-19 0.261794883 16.642934 41.677552 10.747184
#> 342 171 2003-06-20 0.061636631 16.646025 41.681897 10.497644
#> 343 172 2003-06-21 4.853320623 16.647351 41.680500 16.495831
#> 346 173 2003-06-22 1.309010799 16.646909 41.673363 12.056810
#> 348 174 2003-06-23 6.642586834 16.644700 41.660491 18.728082
#> 350 175 2003-06-24 6.505938904 16.640727 41.641891 18.550730
#> 351 176 2003-06-25 6.801692838 16.634994 41.617572 18.912660
#> 353 177 2003-06-26 3.095387362 16.627507 41.587545 14.267868
#> 356 178 2003-06-27 0.712566079 16.618275 41.551823 11.278795
#> 358 179 2003-06-28 1.121584516 16.607306 41.510423 11.779321
#> 360 180 2003-06-29 5.956279378 16.594614 41.463360 17.807030
#> 361 181 2003-06-30 1.083993516 16.580211 41.410655 11.706352
#> 364 182 2003-07-01 1.165206472 16.564113 41.352328 11.792552
#> 366 183 2003-07-02 1.161082212 16.546336 41.288405 11.770737
#> 368 184 2003-07-03 4.316793745 16.526900 41.218909 15.687876
#> 369 185 2003-07-04 1.553896320 16.505823 41.143868 12.222656
#> 372 186 2003-07-05 4.583022735 16.483128 41.063311 15.974516
#> 374 187 2003-07-06 1.184716310 16.458837 40.977269 11.719101
#> 376 188 2003-07-07 2.886485089 16.432973 40.885775 13.812279
#> 377 189 2003-07-08 5.515140577 16.405563 40.788864 17.053315
#> 379 190 2003-07-09 4.130905689 16.376632 40.686570 15.303112
#> 382 191 2003-07-10 4.301387777 16.346207 40.578932 15.483762
#> 384 192 2003-07-11 1.403719073 16.314316 40.465990 11.857388
#> 386 193 2003-07-12 3.217764623 16.280989 40.347784 14.074102
#> 387 194 2003-07-13 2.440028817 16.246255 40.224356 13.076742
#> 390 195 2003-07-14 6.561118152 16.210144 40.095751 18.138393
#> 392 196 2003-07-15 6.491596428 16.172686 39.962013 18.010731
#> 394 197 2003-07-16 3.037822752 16.133915 39.823189 13.704912
#> 395 198 2003-07-17 1.422439959 16.093860 39.679327 11.673341
#> 398 199 2003-07-18 1.081612366 16.052554 39.530476 11.214390
#> 400 200 2003-07-19 6.013812438 16.010029 39.376687 17.239662
#> 402 201 2003-07-20 2.746278714 15.966317 39.218012 13.177341
#> 403 202 2003-07-21 5.624729963 15.921451 39.054503 16.662213
#> 406 203 2003-07-22 4.195613581 15.875462 38.886214 14.860035
#> 407 204 2003-07-23 2.348303811 15.828384 38.713200 12.550051
#> 410 205 2003-07-24 5.958259068 15.780249 38.535519 16.908942
#> 412 206 2003-07-25 2.353394791 15.731088 38.353226 12.457157
#> 414 207 2003-07-26 3.769250399 15.680933 38.166382 14.128652
#> 415 208 2003-07-27 4.471825660 15.629816 37.975044 14.926255
#> 418 209 2003-07-28 5.162544357 15.577768 37.779274 15.704932
#> 420 210 2003-07-29 2.995091790 15.524820 37.579134 13.019719
#> 422 211 2003-07-30 5.759564141 15.471001 37.374685 16.300619
#> 423 212 2003-07-31 5.346667747 15.416342 37.165992 15.736419
#> 426 213 2003-08-01 2.184446559 15.360873 36.953119 11.865803
#> 428 214 2003-08-02 2.914980125 15.304621 36.736132 12.682489
#> 430 215 2003-08-03 1.724335391 15.247616 36.515097 11.193499
#> 431 216 2003-08-04 1.710277890 15.189886 36.290082 11.115529
#> 433 217 2003-08-05 2.175719579 15.131457 36.061156 11.607867
#> 436 218 2003-08-06 2.872744079 15.072356 35.828388 12.371486
#> 438 219 2003-08-07 2.107042101 15.012610 35.591848 11.395646
#> 440 220 2003-08-08 1.987707453 14.952243 35.351609 11.187672
#> 441 221 2003-08-09 3.320487209 14.891281 35.107744 12.691133
#> 444 222 2003-08-10 0.188974112 14.829748 34.860325  8.937192
#> 446 223 2003-08-11 0.831197130 14.767666 34.609429  9.626352
#> 448 224 2003-08-12 4.534992892 14.705060 34.355129 13.886288
#> 449 225 2003-08-13 2.502407747 14.641952 34.097505 11.438122
#> 452 226 2003-08-14 5.653033376 14.578362 33.836632 15.019552
#> 454 227 2003-08-15 1.404332883 14.514313 33.572591 10.017306
#> 456 228 2003-08-16 6.746073060 14.449824 33.305462 16.100892
#> 457 229 2003-08-17 6.082119748 14.384916 33.035326 15.242703
#> 459 230 2003-08-18 5.525917383 14.319607 32.762264 14.512023
#> 462 231 2003-08-19 2.155253167 14.253916 32.486361 10.577629
#> 464 232 2003-08-20 3.812782858 14.187862 32.207701 12.379603
#> 466 233 2003-08-21 2.058005774 14.121461 31.926370 10.308004
#> 467 234 2003-08-22 5.320126927 14.054730 31.642455 13.899412
#> 470 235 2003-08-23 0.497383161 13.987686 31.356043  8.396500
#> 472 236 2003-08-24 5.471596280 13.920345 31.067223 13.872520
#> 474 237 2003-08-25 1.682638278 13.852721 30.776085  9.563149
#> 475 238 2003-08-26 0.159216830 13.784830 30.482721  7.796720
#> 478 239 2003-08-27 1.284872721 13.716686 30.187223  8.960658
#> 480 240 2003-08-28 3.944394176 13.648302 29.889684 11.791519
#> 482 241 2003-08-29 1.090554160 13.579692 29.590199  8.585711
#> 483 242 2003-08-30 3.063329049 13.510869 29.288864 10.642558
#> 486 243 2003-08-31 0.132692818 13.441846 28.985774  7.389512
#> 488 244 2003-09-01 0.025618678 13.372633 28.681027  7.197730
#> 489 245 2003-09-02 2.366089029 13.303244 28.374722  9.617017
#> 492 246 2003-09-03 5.865941786 13.233689 28.066958 13.237195
#> 494 247 2003-09-04 0.008197031 13.163980 27.757836  6.948101
#> 496 248 2003-09-05 2.583635072 13.094126 27.447457  9.569728
#> 497 249 2003-09-06 0.048793440 13.024138 27.135924  6.834812
#> 500 250 2003-09-07 6.191431744 12.954027 26.823338 13.115998
#> 502 251 2003-09-08 4.175683721 12.883802 26.509805 10.923410
#> 504 252 2003-09-09 3.492897340 12.813472 26.195428 10.119238
#> 505 253 2003-09-10 0.698747876 12.743047 25.880314  7.179635
#> 508 254 2003-09-11 0.847024361 12.672536 25.564567  7.245502
#> 510 255 2003-09-12 0.125141687 12.601947 25.248295  6.437436
#> 512 256 2003-09-13 2.230350421 12.531289 24.931604  8.451596
#> 513 257 2003-09-14 3.557105622 12.460572 24.614602  9.667002
#> 515 258 2003-09-15 5.420657080 12.389802 24.297397 11.389521
#> 518 259 2003-09-16 5.932486910 12.318988 23.980097 11.769103
#> 520 260 2003-09-17 0.207304377 12.248139 23.662812  6.115954
#> 522 261 2003-09-18 1.887469641 12.177262 23.345649  7.645694
#> 523 262 2003-09-19 4.158145277 12.106365 23.028718  9.711990
#> 526 263 2003-09-20 0.761703555 12.035456 22.712128  6.396738
#> 528 264 2003-09-21 0.233411112 11.964544 22.395989  5.817454
#> 530 265 2003-09-22 3.574864688 11.893635 22.080409  8.838452
#> 531 266 2003-09-23 5.715422823 11.822738 21.765497 10.702381
#> 534 267 2003-09-24 6.058531311 11.751861 21.451363 10.892337
#> 536 268 2003-09-25 6.356969235 11.681012 21.138114 11.036356
#> 538 269 2003-09-26 1.906317284 11.610198 20.825859  6.916198
#> 539 270 2003-09-27 6.628592010 11.539428 20.514705 11.020806
#> 541 271 2003-09-28 4.871494930 11.468711 20.204761  9.342317
#> 544 272 2003-09-29 0.829985858 11.398053 19.896131  5.698433
#> 546 273 2003-09-30 4.542662139 11.327464 19.588923  8.825111
#> 548 274 2003-10-01 2.956903441 11.256953 19.283241  7.353408
#> 549 275 2003-10-02 5.948623402 11.186528 18.979189  9.791048
#> 552 276 2003-10-03 4.140383233 11.116198 18.676871  8.147448
#> 554 277 2003-10-04 3.246177797 11.045973 18.376389  7.294313
#> 556 278 2003-10-05 6.333286001 10.975862 18.077844  9.735095
#> 557 279 2003-10-06 4.172329761 10.905874 17.781337  7.846694
#> 560 280 2003-10-07 0.453603757 10.836020 17.486965  4.737750
#> 562 281 2003-10-08 0.848784184 10.766311 17.194826  4.976501
#> 564 282 2003-10-09 6.736581398 10.696756 16.905016  9.549457
#> 565 283 2003-10-10 5.808569201 10.627367 16.617629  8.695732
#> 568 284 2003-10-11 5.678667163 10.558154 16.332759  8.475449
#> 569 285 2003-10-12 2.977306077 10.489131 16.050495  6.290565
#> 572 286 2003-10-13 0.466854048 10.420308 15.770928  4.296019
#> 574 287 2003-10-14 2.232430891 10.351698 15.494145  5.544258
#> 576 288 2003-10-15 1.698201336 10.283314 15.220231  5.061803
#> 577 289 2003-10-16 0.683212012 10.215170 14.949271  4.237237
#> 580 290 2003-10-17 1.423320913 10.147279 14.681345  4.699985
#> 582 291 2003-10-18 2.375077084 10.079655 14.416534  5.302623
#> 584 292 2003-10-19 4.612421778 10.012314 14.154915  6.799136
#> 585 293 2003-10-20 1.155167689  9.945270 13.896562  4.281201
#> 588 294 2003-10-21 2.079116258  9.878539 13.641550  4.845942
#> 590 295 2003-10-22 6.747920525  9.812138 13.389948  7.951698
#> 592 296 2003-10-23 3.395849496  9.746084 13.141825  5.574974
#> 593 297 2003-10-24 2.907252101  9.680393 12.897247  5.160987
#> 595 298 2003-10-25 5.162400663  9.615084 12.656277  6.561688
#> 598 299 2003-10-26 5.779791278  9.550176 12.418977  6.862743
#> 600 300 2003-10-27 6.283666709  9.485687 12.185404  7.082380
#> 602 301 2003-10-28 4.356036209  9.421638 11.955615  5.752706
#> 603 302 2003-10-29 5.310800898  9.358048 11.729663  6.260776
#> 606 303 2003-10-30 2.596615217  9.294940 11.507600  4.484270
#> 608 304 2003-10-31 3.621186691  9.232334 11.289473  5.036396
#> 610 305 2003-11-01 3.759525961  9.170252 11.075328  5.039107
#> 611 306 2003-11-02 2.768241242  9.108719 10.865209  4.367331
#> 614 307 2003-11-03 6.516555501  9.047757 10.659156  6.503364
#> 616 308 2003-11-04 3.445559008  8.987390 10.457208  4.618829
#> 618 309 2003-11-05 4.923284085  8.927644 10.259398  5.393700
#> 619 310 2003-11-06 4.743495603  8.868543 10.065762  5.208365
#> 621 311 2003-11-07 1.389956403  8.810114  9.876328  3.248168
#> 624 312 2003-11-08 5.083678877  8.752384  9.691126  5.237247
#> 626 313 2003-11-09 3.903173036  8.695379  9.510180  4.512005
#> 628 314 2003-11-10 4.432926817  8.639127  9.333513  4.727994
#> 629 315 2003-11-11 5.151877331  8.583658  9.161147  5.039529
#> 632 316 2003-11-12 6.933435163  8.528999  8.993100  5.903632
#> 634 317 2003-11-13 1.458557163  8.475180  8.829388  2.967105
#> 636 318 2003-11-14 5.248632400  8.422232  8.670025  4.869034
#> 637 319 2003-11-15 6.742362284  8.370184  8.515023  5.558273
#> 640 320 2003-11-16 0.003901486  8.319067  8.364392  2.093059
#> 642 321 2003-11-17 0.017124578  8.268912  8.218139  2.063044
#> 644 322 2003-11-18 4.341518395  8.219751  8.076271  4.151935
#> 645 323 2003-11-19 6.382501164  8.171616  7.938791  5.085023
#> 648 324 2003-11-20 6.640175541  8.124538  7.805702  5.141221
#> 649 325 2003-11-21 2.633275308  8.078549  7.677004  3.170445
#> 652 326 2003-11-22 5.602188285  8.033683  7.552698  4.521564
#> 654 327 2003-11-23 4.027339105  7.989971  7.432780  3.731439
#> 656 328 2003-11-24 1.361747528  7.947446  7.317248  2.456195
#> 657 329 2003-11-25 2.857169457  7.906140  7.206095  3.103615
#> 660 330 2003-11-26 4.828869787  7.866085  7.099317  3.953911
#> 662 331 2003-11-27 4.588119972  7.827314  6.996907  3.799908
#> 664 332 2003-11-28 4.774691253  7.789856  6.898857  3.838996
#> 665 333 2003-11-29 4.552335268  7.753745  6.805157  3.698992
#> 668 334 2003-11-30 6.906555703  7.719011  6.715800  4.683418
#> 670 335 2003-12-01 1.725578418  7.685684  6.630775  2.402060
#> 672 336 2003-12-02 1.572032213  7.653793  6.550072  2.310186
#> 673 337 2003-12-03 0.032943771  7.623368  6.473680  1.632408
#> 675 338 2003-12-04 4.167511935  7.594437  6.401590  3.356861
#> 678 339 2003-12-05 6.178511561  7.567027  6.333789  4.169231
#> 680 340 2003-12-06 5.286428550  7.541163  6.270268  3.765326
#> 682 341 2003-12-07 3.596070545  7.516872  6.211016  3.038429
#> 683 342 2003-12-08 3.225743171  7.494177  6.156023  2.863884
#> 686 343 2003-12-09 0.005631276  7.473100  6.105278  1.528620
#> 688 344 2003-12-10 5.994859724  7.453664  6.058772  3.951179
#> 690 345 2003-12-11 0.886944425  7.435887  6.016497  1.862945
#> 691 346 2003-12-12 6.922263789  7.419789  5.978443  4.283393
#> 694 347 2003-12-13 0.820995779  7.405386  5.944603  1.815674
#> 696 348 2003-12-14 3.307251029  7.392694  5.914970  2.801825
#> 698 349 2003-12-15 5.899310592  7.381725  5.889539  3.825779
#> 699 350 2003-12-16 0.530205437  7.372493  5.868303  1.678090
#> 701 351 2003-12-17 4.912255621  7.365006  5.851259  3.414129
#> 704 352 2003-12-18 2.283240192  7.359273  5.838404  2.365294
#> 706 353 2003-12-19 2.794665916  7.355300  5.829736  2.564946
#> 708 354 2003-12-20 0.796329900  7.353091  5.825253  1.771747
#> 709 355 2003-12-21 2.378266219  7.352649  5.824956  2.398300
#> 712 356 2003-12-22 3.652361379  7.353975  5.828845  2.904663
#> 714 357 2003-12-23 5.068798506  7.357066  5.836924  3.469964
#> 716 358 2003-12-24 2.493088013  7.361920  5.849194  2.452703
#> 717 359 2003-12-25 6.831483717  7.368531  5.865662  4.185490
#> 720 360 2003-12-26 4.526711828  7.376892  5.886331  3.277609
#> 722 361 2003-12-27 0.814352876  7.386993  5.911209  1.803632
#> 724 362 2003-12-28 4.360284353  7.398825  5.940303  3.235449
#> 725 363 2003-12-29 6.528902256  7.412375  5.973622  4.124222
#> 728 364 2003-12-30 6.129454987  7.427627  6.011174  3.983076
#> 730 365 2003-12-31 1.907395097  7.444567  6.052971  2.288668