Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.014223  3.713330  2.877278  2.426673  3.671609  2.652263  1.755697
#>   [8]  1.865638  3.324315  3.597616  4.706121  3.766697  3.559645  1.908959
#>  [15]  3.403796  4.415907  2.415105  2.128187  4.479617  4.249345  2.287725
#>  [22]  2.458500  2.170210  3.535254  4.791826  3.197019  5.789234  3.826966
#>  [29]  3.470334  6.098690  5.143813  3.025193  5.206582  5.711603  6.378012
#>  [36]  4.550146  3.313993  5.347051  6.035145  4.504189  5.694725  7.211995
#>  [43]  5.341417  5.646709  7.237143  6.356891  4.016147  6.734269  3.961510
#>  [50]  7.413769  4.391228  6.792779  5.626000  3.790109  5.374327  8.701960
#>  [57]  6.909099  8.853084  9.242610  8.832998  7.361838  5.747029  9.530516
#>  [64]  6.981106  6.517625  7.756246  6.348721  6.460027  8.088587  7.995428
#>  [71]  6.557654 10.089691  8.921722  8.348356  9.384139  5.902869  8.117596
#>  [78]  5.746461  6.367914  8.640520 12.175027  8.985835 10.552337 10.901819
#>  [85]  7.571958  8.476184  9.976047  9.254880 12.848838 13.193801  7.781547
#>  [92]  7.106116 10.236779 12.021555  8.348660  7.130438 12.319371  9.914866
#>  [99]  8.239269 11.964608  8.408470 12.904758  9.452238  9.518177 10.621532
#> [106] 11.949017 11.307759 10.063188 13.559241  9.756580 14.071704 11.133554
#> [113]  9.583642 11.255253 10.217706 14.401348 12.156723 12.084777 13.051617
#> [120] 15.494208 15.268371 13.812404 14.048592 13.952374 14.914888 15.850296
#> [127] 15.750638 14.279233 10.083797 10.976501 13.645250 14.709499 13.915988
#> [134] 11.880568 11.113526 14.217071 12.393588 12.993010 13.702312 12.758395
#> [141] 16.120149 12.769001 14.000269 18.108137 14.356868 11.857703 10.871599
#> [148] 14.043075 15.191067 13.500664 18.271733 13.681653 15.062999 17.067523
#> [155] 11.370230 12.461910 15.266166 13.106963 18.763203 11.465957 12.113186
#> [162] 16.511436 14.862445 16.078815 10.582559 12.641159 14.546094 14.848473
#> [169] 11.507627 15.107981 15.032775 16.911276 16.104977 17.779720 13.559285
#> [176] 11.327336 15.623849 10.537984 18.519120 16.955252 15.553638 16.296843
#> [183] 17.901319 14.974654 11.530450 11.440987 17.329136 16.861238 16.855634
#> [190] 15.529243 15.310045 17.796294 12.603641 11.983948 10.538261 13.038539
#> [197] 10.334759 16.683956 13.765271 15.178598 15.817676 14.054255 13.711038
#> [204] 14.727279 17.456982 11.256245 15.338304 10.275900 15.457509 13.920029
#> [211] 12.711439 11.870487  9.949352 16.469873 10.653569 11.829350  9.103912
#> [218] 17.107669 15.377172 12.717651 12.376269 16.610376  8.943685 12.140580
#> [225]  8.924595  8.867323 13.376375  9.491122  9.838539 11.370057 12.860580
#> [232] 10.315715  9.293689 10.383811  9.789758 15.346499 13.684777  9.973249
#> [239]  8.071134 11.770541 10.975483  8.792274  9.542273 12.933503 13.057647
#> [246]  9.890503  8.475417 13.131648 14.001583 12.110497  6.813874 10.777550
#> [253]  6.991125 10.920448 12.015716 11.825672 12.246548  7.616476 10.224479
#> [260] 11.209276 10.695301  8.565206  9.260170  6.015466 11.834518  8.496844
#> [267] 11.480356  7.965241 10.757171  5.883379  6.267698  5.996317  5.160074
#> [274]  8.697669  7.814156  7.697615  8.990935  5.889140  6.892469  4.499498
#> [281]  6.656488  7.566553  6.816421  6.530487  6.534326  6.860568  6.764990
#> [288]  4.931563  7.597825  6.081690  4.942984  7.576815  5.184408  6.129711
#> [295]  5.601402  3.948526  3.727221  4.029083  4.965090  6.345962  3.127751
#> [302]  3.441769  3.068743  3.015681  3.290501  4.237417  4.448666  4.678248
#> [309]  6.447643  6.156730  3.041278  3.587384  4.738698  5.865173  3.351373
#> [316]  5.812346  5.556069  4.125938  3.521257  2.324053  4.382350  3.963356
#> [323]  4.811076  4.681311  3.814577  2.554536  4.914138  3.710214  4.322388
#> [330]  4.606979  4.221511  4.416583  3.424934  4.657710  4.627790  2.616459
#> [337]  4.537483  3.280443  3.750783  1.570304  2.489832  3.199602  3.617570
#> [344]  3.618160  2.889185  3.900436  2.894368  3.516026  2.083875  3.724304
#> [351]  2.660323  1.574202  3.005513  3.928361  1.662965  1.684452  3.861595
#> [358]  3.766753  4.090992  2.621410  2.441506  2.262320  3.855558  3.320609
#> [365]  2.779781  2.014223  3.713330  2.877278  2.426673  3.671609  2.652263
#> [372]  1.755697  1.865638  3.324315  3.597616  4.706121  3.766697  3.559645
#> [379]  1.908959  3.403796  4.415907  2.415105  2.128187  4.479617  4.249345
#> [386]  2.287725  2.458500  2.170210  3.535254  4.791826  3.197019  5.789234
#> [393]  3.826966  3.470334  6.098690  5.143813  3.025193  5.206582  5.711603
#> [400]  6.378012  4.550146  3.313993  5.347051  6.035145  4.504189  5.694725
#> [407]  7.211995  5.341417  5.646709  7.237143  6.356891  4.016147  6.734269
#> [414]  3.961510  7.413769  4.391228  6.792779  5.626000  3.790109  5.374327
#> [421]  8.701960  6.909099  8.853084  9.242610  8.832998  7.361838  5.747029
#> [428]  9.530516  6.981106  6.517625  7.756246  6.348721  6.460027  8.088587
#> [435]  7.995428  6.557654 10.089691  8.921722  8.348356  9.384139  5.902869
#> [442]  8.117596  5.746461  6.367914  8.640520 12.175027  8.985835 10.552337
#> [449] 10.901819  7.571958  8.476184  9.976047  9.254880 12.848838 13.193801
#> [456]  7.781547  7.106116 10.236779 12.021555  8.348660  7.130438 12.319371
#> [463]  9.914866  8.239269 11.964608  8.408470 12.904758  9.452238  9.518177
#> [470] 10.621532 11.949017 11.307759 10.063188 13.559241  9.756580 14.071704
#> [477] 11.133554  9.583642 11.255253 10.217706 14.401348 12.156723 12.084777
#> [484] 13.051617 15.494208 15.268371 13.812404 14.048592 13.952374 14.914888
#> [491] 15.850296 15.750638 14.279233 10.083797 10.976501 13.645250 14.709499
#> [498] 13.915988 11.880568 11.113526 14.217071 12.393588 12.993010 13.702312
#> [505] 12.758395 16.120149 12.769001 14.000269 18.108137 14.356868 11.857703
#> [512] 10.871599 14.043075 15.191067 13.500664 18.271733 13.681653 15.062999
#> [519] 17.067523 11.370230 12.461910 15.266166 13.106963 18.763203 11.465957
#> [526] 12.113186 16.511436 14.862445 16.078815 10.582559 12.641159 14.546094
#> [533] 14.848473 11.507627 15.107981 15.032775 16.911276 16.104977 17.779720
#> [540] 13.559285 11.327336 15.623849 10.537984 18.519120 16.955252 15.553638
#> [547] 16.296843 17.901319 14.974654 11.530450 11.440987 17.329136 16.861238
#> [554] 16.855634 15.529243 15.310045 17.796294 12.603641 11.983948 10.538261
#> [561] 13.038539 10.334759 16.683956 13.765271 15.178598 15.817676 14.054255
#> [568] 13.711038 14.727279 17.456982 11.256245 15.338304 10.275900 15.457509
#> [575] 13.920029 12.711439 11.870487  9.949352 16.469873 10.653569 11.829350
#> [582]  9.103912 17.107669 15.377172 12.717651 12.376269 16.610376  8.943685
#> [589] 12.140580  8.924595  8.867323 13.376375  9.491122  9.838539 11.370057
#> [596] 12.860580 10.315715  9.293689 10.383811  9.789758 15.346499 13.684777
#> [603]  9.973249  8.071134 11.770541 10.975483  8.792274  9.542273 12.933503
#> [610] 13.057647  9.890503  8.475417 13.131648 14.001583 12.110497  6.813874
#> [617] 10.777550  6.991125 10.920448 12.015716 11.825672 12.246548  7.616476
#> [624] 10.224479 11.209276 10.695301  8.565206  9.260170  6.015466 11.834518
#> [631]  8.496844 11.480356  7.965241 10.757171  5.883379  6.267698  5.996317
#> [638]  5.160074  8.697669  7.814156  7.697615  8.990935  5.889140  6.892469
#> [645]  4.499498  6.656488  7.566553  6.816421  6.530487  6.534326  6.860568
#> [652]  6.764990  4.931563  7.597825  6.081690  4.942984  7.576815  5.184408
#> [659]  6.129711  5.601402  3.948526  3.727221  4.029083  4.965090  6.345962
#> [666]  3.127751  3.441769  3.068743  3.015681  3.290501  4.237417  4.448666
#> [673]  4.678248  6.447643  6.156730  3.041278  3.587384  4.738698  5.865173
#> [680]  3.351373  5.812346  5.556069  4.125938  3.521257  2.324053  4.382350
#> [687]  3.963356  4.811076  4.681311  3.814577  2.554536  4.914138  3.710214
#> [694]  4.322388  4.606979  4.221511  4.416583  3.424934  4.657710  4.627790
#> [701]  2.616459  4.537483  3.280443  3.750783  1.570304  2.489832  3.199602
#> [708]  3.617570  3.618160  2.889185  3.900436  2.894368  3.516026  2.083875
#> [715]  3.724304  2.660323  1.574202  3.005513  3.928361  1.662965  1.684452
#> [722]  3.861595  3.766753  4.090992  2.621410  2.441506  2.262320  3.855558
#> [729]  3.320609  2.779781
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 0.39315608  7.463176  6.099033  1.685405
#> 3     2 2002-01-02 6.65767997  7.483435  6.149345  4.272736
#> 6     3 2002-01-03 5.13201547  7.505323  6.203934  3.672057
#> 7     4 2002-01-04 4.81992611  7.528820  6.262814  3.570420
#> 9     5 2002-01-05 5.59136642  7.553900  6.325996  3.922738
#> 11    6 2002-01-06 3.55878803  7.580540  6.393496  3.099131
#> 14    7 2002-01-07 3.77158676  7.608714  6.465324  3.218739
#> 15    8 2002-01-08 4.30155822  7.638396  6.541496  3.477294
#> 17    9 2002-01-09 6.69894899  7.669557  6.622023  4.547498
#> 19   10 2002-01-10 2.57000847  7.702170  6.706920  2.795690
#> 22   11 2002-01-11 4.26265516  7.736204  6.796198  3.571405
#> 23   12 2002-01-12 1.69024424  7.771631  6.889871  2.471703
#> 25   13 2002-01-13 1.13675742  7.808419  6.987950  2.255644
#> 27   14 2002-01-14 5.12163890  7.846537  7.090446  4.086671
#> 30   15 2002-01-15 0.72307882  7.885954  7.197371  2.129313
#> 32   16 2002-01-16 2.58604859  7.926639  7.308732  3.019412
#> 33   17 2002-01-17 0.66155894  7.968559  7.424539  2.164332
#> 35   18 2002-01-18 3.05632233  8.011681  7.544799  3.325307
#> 37   19 2002-01-19 6.08630314  8.055974  7.669519  4.814548
#> 40   20 2002-01-20 6.97782342  8.101405  7.798703  5.308227
#> 41   21 2002-01-21 5.92757777  8.147942  7.932354  4.868458
#> 43   22 2002-01-22 2.10609761  8.195553  8.070474  3.054596
#> 45   23 2002-01-23 0.29736172  8.244205  8.213063  2.201385
#> 48   24 2002-01-24 1.17920243  8.293867  8.360119  2.684341
#> 49   25 2002-01-25 6.71167996  8.344507  8.511639  5.550964
#> 51   26 2002-01-26 2.28902036  8.396093  8.667618  3.348427
#> 53   27 2002-01-27 6.21440782  8.448595  8.828046  5.453769
#> 56   28 2002-01-28 1.52557069  8.501983  8.992915  3.055060
#> 58   29 2002-01-29 2.73618197  8.556225  9.162213  3.755538
#> 59   30 2002-01-30 5.59669586  8.611293  9.335925  5.367808
#> 61   31 2002-01-31 2.72260762  8.667157  9.514035  3.872827
#> 63   32 2002-02-01 4.31220658  8.723789  9.696523  4.820648
#> 66   33 2002-02-02 2.38912630  8.781160  9.883368  3.815346
#> 67   34 2002-02-03 1.15048279  8.839243 10.074546  3.174269
#> 69   35 2002-02-04 0.41929181  8.898011 10.270030  2.809479
#> 71   36 2002-02-05 1.46346990  8.957438 10.469790  3.472727
#> 74   37 2002-02-06 6.54794323  9.017498 10.673795  6.543770
#> 75   38 2002-02-07 2.48032514  9.078165 10.882010  4.207087
#> 77   39 2002-02-08 4.56244619  9.139416 11.094398  5.542792
#> 79   40 2002-02-09 4.96403647  9.201226 11.310918  5.878834
#> 82   41 2002-02-10 2.19433792  9.263572 11.531528  4.248666
#> 83   42 2002-02-11 6.31336605  9.326432 11.756182  6.918117
#> 86   43 2002-02-12 3.43900229  9.389784 11.984831  5.190926
#> 87   44 2002-02-13 2.68645718  9.453606 12.217425  4.790286
#> 89   45 2002-02-14 3.42210893  9.517878 12.453910  5.352350
#> 91   46 2002-02-15 6.34332636  9.582579 12.694229  7.375121
#> 94   47 2002-02-16 2.57755373  9.647690 12.938322  4.962933
#> 95   48 2002-02-17 0.55915326  9.713192 13.186128  3.676071
#> 97   49 2002-02-18 5.78936252  9.779067 13.437583  7.337027
#> 99   50 2002-02-19 1.53305454  9.845297 13.692618  4.489223
#> 102  51 2002-02-20 6.99665694  9.911864 13.951164  8.411764
#> 103  52 2002-02-21 2.53112205  9.978754 14.213148  5.355877
#> 105  53 2002-02-22 0.51804845 10.045948 14.478496  3.992937
#> 107  54 2002-02-23 3.85522296 10.113433 14.747130  6.497573
#> 110  55 2002-02-24 6.59954862 10.181192 15.018971  8.622465
#> 112  56 2002-02-25 6.56223657 10.249211 15.293936  8.719589
#> 113  57 2002-02-26 6.07824633 10.317477 15.571940  8.479867
#> 115  58 2002-02-27 3.90087209 10.385975 15.852898  6.940322
#> 117  59 2002-02-28 4.21124726 10.454693 16.136720  7.284190
#> 120  60 2002-03-01 2.02123090 10.523618 16.423316  5.683010
#> 121  61 2002-03-02 1.90090078 10.592738 16.712592  5.677712
#> 123  62 2002-03-03 0.40881757 10.662040 17.004454  4.577117
#> 125  63 2002-03-04 2.94329109 10.731513 17.298805  6.696940
#> 128  64 2002-03-05 3.10241020 10.801147 17.595546  6.925868
#> 129  65 2002-03-06 2.98566395 10.870930 17.894578  6.930987
#> 131  66 2002-03-07 4.26955535 10.940852 18.195799  8.099312
#> 133  67 2002-03-08 5.02129728 11.010902 18.499104  8.842846
#> 136  68 2002-03-09 0.69387187 11.081072 18.804391  5.289842
#> 138  69 2002-03-10 0.07429720 11.151350 19.111552  4.841555
#> 139  70 2002-03-11 6.27569026 11.221729 19.420480 10.285519
#> 141  71 2002-03-12 3.25527009 11.292198 19.731067  7.776764
#> 143  72 2002-03-13 4.57931470 11.362749 20.043203  9.049618
#> 146  73 2002-03-14 2.94803549 11.433374 20.356778  7.713639
#> 147  74 2002-03-15 0.56132592 11.504062 20.671679  5.672244
#> 149  75 2002-03-16 2.89860182 11.574807 20.987796  7.874866
#> 151  76 2002-03-17 4.42825025 11.645600 21.305014  9.376879
#> 154  77 2002-03-18 0.46920738 11.716433 21.623220  5.838777
#> 155  78 2002-03-19 3.27081998 11.787297 21.942300  8.529925
#> 157  79 2002-03-20 0.50484093 11.858185 22.262140  6.039420
#> 159  80 2002-03-21 3.41424965 11.929088 22.582624  8.877366
#> 162  81 2002-03-22 0.04253733 12.000000 22.903638  5.766504
#> 163  82 2002-03-23 4.81676192 12.070912 23.225065 10.440117
#> 166  83 2002-03-24 1.50025424 12.141815 23.546791  7.341430
#> 167  84 2002-03-25 4.30613721 12.212703 23.868700 10.175166
#> 169  85 2002-03-26 4.55659126 12.283567 24.190677 10.534437
#> 171  86 2002-03-27 1.81894676 12.354400 24.512606  7.932656
#> 174  87 2002-03-28 5.49042420 12.425193 24.834373 11.695480
#> 175  88 2002-03-29 5.06363752 12.495938 25.155863 11.385829
#> 177  89 2002-03-30 1.00916220 12.566626 25.476961  7.392203
#> 179  90 2002-03-31 3.86442496 12.637251 25.797553 10.393787
#> 182  91 2002-04-01 5.98601007 12.707802 26.117528 12.680713
#> 183  92 2002-04-02 2.92174002 12.778271 26.436771  9.631564
#> 185  93 2002-04-03 4.58283422 12.848650 26.755170 11.460287
#> 187  94 2002-04-04 1.97568073 12.918928 27.072615  8.838250
#> 190  95 2002-04-05 2.98881537 12.989098 27.388994  9.998378
#> 192  96 2002-04-06 1.08049905 13.059148 27.704199  8.072157
#> 193  97 2002-04-07 5.73547802 13.129070 28.018120 13.124434
#> 195  98 2002-04-08 3.11708333 13.198853 28.330649 10.427991
#> 197  99 2002-04-09 6.49840857 13.268487 28.641680 14.174232
#> 200 100 2002-04-10 1.10051867 13.337960 28.951106  8.432158
#> 201 101 2002-04-11 5.84834132 13.407262 29.258822 13.696157
#> 203 102 2002-04-12 4.65753479 13.476382 29.564726 12.500073
#> 205 103 2002-04-13 2.80618996 13.545307 29.868714 10.561139
#> 208 104 2002-04-14 4.49402222 13.614025 30.170685 12.522379
#> 209 105 2002-04-15 0.34063993 13.682523 30.470539  7.996932
#> 211 106 2002-04-16 6.11554686 13.750789 30.768176 14.533987
#> 213 107 2002-04-17 5.84264428 13.818808 31.063500 14.332757
#> 216 108 2002-04-18 4.58491883 13.886567 31.356413 13.015566
#> 218 109 2002-04-19 6.41128897 13.954052 31.646820 15.181884
#> 219 110 2002-04-20 4.43835554 14.021246 31.934628 13.038030
#> 221 111 2002-04-21 2.54134760 14.088136 32.219743 10.960983
#> 223 112 2002-04-22 2.19398403 14.154703 32.502075 10.644435
#> 226 113 2002-04-23 5.20741777 14.220933 32.781534 14.197350
#> 227 114 2002-04-24 6.42264234 14.286808 33.058030 15.695135
#> 229 115 2002-04-25 4.31412050 14.352310 33.331477 13.342377
#> 231 116 2002-04-26 4.51156508 14.417421 33.601789 13.657860
#> 234 117 2002-04-27 0.95937214 14.482122 33.868881  9.589047
#> 235 118 2002-04-28 0.14004751 14.546394 34.132670  8.697476
#> 237 119 2002-04-29 4.18791911 14.610216 34.393074 13.527539
#> 239 120 2002-04-30 3.34075040 14.673568 34.650012 12.606910
#> 242 121 2002-05-01 2.07435345 14.736428 34.903405 11.182417
#> 243 122 2002-05-02 4.44830285 14.798774 35.153175 14.071568
#> 245 123 2002-05-03 1.87641940 14.860584 35.399245 11.084711
#> 248 124 2002-05-04 0.61929799 14.921835 35.641540  9.649997
#> 249 125 2002-05-05 5.73482853 14.982502 35.879987 15.836859
#> 251 126 2002-05-06 0.69418756 15.042562 36.114511  9.861938
#> 253 127 2002-05-07 2.99436233 15.101989 36.345041 12.689436
#> 256 128 2002-05-08 0.79592428 15.160757 36.571508 10.102861
#> 257 129 2002-05-09 5.30337431 15.218840 36.793843 15.609315
#> 259 130 2002-05-10 2.78114386 15.276211 37.011977 12.622142
#> 261 131 2002-05-11 5.67647111 15.332843 37.225844 16.197271
#> 264 132 2002-05-12 2.70391072 15.388707 37.435378 12.647682
#> 265 133 2002-05-13 5.12864917 15.443775 37.640517 15.660059
#> 267 134 2002-05-14 6.48864251 15.498017 37.841197 17.381893
#> 269 135 2002-05-15 1.88593473 15.551405 38.037357 11.815754
#> 272 136 2002-05-16 5.93270101 15.603907 38.228936 16.824672
#> 274 137 2002-05-17 2.57525893 15.655493 38.415875 12.763589
#> 275 138 2002-05-18 3.85118051 15.706133 38.598118 14.381704
#> 277 139 2002-05-19 6.89226155 15.755795 38.775606 18.174960
#> 279 140 2002-05-20 6.57648045 15.804447 38.948286 17.840571
#> 282 141 2002-05-21 0.14681792 15.852058 39.116103  9.960168
#> 283 142 2002-05-22 3.25567725 15.898595 39.279005 13.841483
#> 285 143 2002-05-23 6.04588113 15.944026 39.436940 17.336363
#> 287 144 2002-05-24 1.75453204 15.988319 39.589859 12.069728
#> 290 145 2002-05-25 3.66486298 16.031441 39.737713 14.476542
#> 291 146 2002-05-26 1.27905643 16.073361 39.880455 11.556881
#> 293 147 2002-05-27 3.72160548 16.114046 40.018039 14.625676
#> 295 148 2002-05-28 6.93792359 16.153463 40.150421 18.659922
#> 298 149 2002-05-29 6.82845044 16.191581 40.277558 18.562485
#> 300 150 2002-05-30 0.54363758 16.228369 40.399408 10.776526
#> 301 151 2002-05-31 4.55976130 16.263796 40.515933 15.808561
#> 303 152 2002-06-01 3.31892027 16.297830 40.627093 14.293461
#> 305 153 2002-06-02 4.84313694 16.330443 40.732853 16.223306
#> 308 154 2002-06-03 3.05486800 16.361604 40.833176 14.020266
#> 309 155 2002-06-04 5.76451244 16.391286 40.928031 17.428824
#> 311 156 2002-06-05 4.09571490 16.419460 41.017385 15.370091
#> 313 157 2002-06-06 1.38204393 16.446100 41.101209 12.002267
#> 316 158 2002-06-07 0.53273060 16.471180 41.179475 10.960807
#> 317 159 2002-06-08 6.51957291 16.494677 41.252156 18.465562
#> 319 160 2002-06-09 0.50068789 16.516565 41.319229 10.956089
#> 321 161 2002-06-10 0.70037781 16.536824 41.380671 11.221458
#> 324 162 2002-06-11 2.13194594 16.555433 41.436461 13.027130
#> 325 163 2002-06-12 3.93096772 16.572373 41.486582 15.291955
#> 328 164 2002-06-13 3.66423522 16.587625 41.531016 14.969891
#> 329 165 2002-06-14 1.32003546 16.601175 41.569750 12.045138
#> 331 166 2002-06-15 4.95851592 16.613007 41.602770 16.609322
#> 333 167 2002-06-16 3.46809344 16.623108 41.630068 14.750175
#> 336 168 2002-06-17 5.65308047 16.631469 41.651634 17.491659
#> 337 169 2002-06-18 0.73648559 16.638080 41.667464 11.339072
#> 339 170 2002-06-19 2.63041146 16.642934 41.677552 13.712951
#> 341 171 2002-06-20 6.74323877 16.646025 41.681897 18.863060
#> 344 172 2002-06-21 5.29965306 16.647351 41.680500 17.054580
#> 345 173 2002-06-22 2.89389100 16.646909 41.673363 14.040580
#> 347 174 2002-06-23 0.95010626 16.644700 41.660491 11.604147
#> 349 175 2002-06-24 3.84942006 16.640727 41.641891 15.226883
#> 352 176 2002-06-25 2.44958101 16.634994 41.617572 13.468584
#> 354 177 2002-06-26 1.51034004 16.627507 41.587545 12.285664
#> 355 178 2002-06-27 3.61359667 16.618275 41.551823 14.905619
#> 357 179 2002-06-28 2.37510974 16.607306 41.510423 13.345932
#> 359 180 2002-06-29 3.46193259 16.594614 41.463360 14.690838
#> 362 181 2002-06-30 6.13956313 16.580211 41.410655 18.019735
#> 363 182 2002-07-01 4.68517011 16.564113 41.352328 16.186349
#> 365 183 2002-07-02 6.98602528 16.546336 41.288405 19.038286
#> 367 184 2002-07-03 2.26327106 16.526900 41.218909 13.127082
#> 370 185 2002-07-04 3.12668569 16.505823 41.143868 14.182893
#> 371 186 2002-07-05 2.13040549 16.483128 41.063311 12.919496
#> 373 187 2002-07-06 5.32391827 16.458837 40.977269 16.871749
#> 375 188 2002-07-07 3.08077859 16.432973 40.885775 14.053983
#> 378 189 2002-07-08 2.98996951 16.405563 40.788864 13.914170
#> 380 190 2002-07-09 0.43908723 16.376632 40.686570 10.717083
#> 381 191 2002-07-10 0.97942110 16.346207 40.578932 11.360424
#> 383 192 2002-07-11 6.21233535 16.314316 40.465990 17.821028
#> 385 193 2002-07-12 1.91754863 16.280989 40.347784 12.462994
#> 388 194 2002-07-13 0.45357451 16.246255 40.224356 10.617595
#> 389 195 2002-07-14 3.45006372 16.210144 40.095751 14.290800
#> 391 196 2002-07-15 3.93071725 16.172686 39.962013 14.846820
#> 393 197 2002-07-16 5.58018094 16.133915 39.823189 16.842551
#> 396 198 2002-07-17 4.78303066 16.093860 39.679327 15.816100
#> 397 199 2002-07-18 6.90648742 16.052554 39.530476 18.386460
#> 399 200 2002-07-19 6.04464246 16.010029 39.376687 17.277575
#> 401 201 2002-07-20 1.60775896 15.966317 39.218012 11.779070
#> 404 202 2002-07-21 3.03079690 15.921451 39.054503 13.480820
#> 405 203 2002-07-22 3.50949703 15.875462 38.886214 14.019729
#> 408 204 2002-07-23 6.18767513 15.828384 38.713200 17.245235
#> 409 205 2002-07-24 3.81268039 15.780249 38.535519 14.289181
#> 411 206 2002-07-25 0.38054690 15.731088 38.353226 10.052203
#> 413 207 2002-07-26 2.33134659 15.680933 38.166382 12.378769
#> 416 208 2002-07-27 2.29427126 15.629816 37.975044 12.280904
#> 417 209 2002-07-28 2.62476022 15.577768 37.779274 12.627609
#> 419 210 2002-07-29 2.16175479 15.524820 37.579134 12.011138
#> 421 211 2002-07-30 3.98854946 15.471001 37.374685 14.161420
#> 424 212 2002-07-31 2.62982864 15.416342 37.165992 12.461517
#> 425 213 2002-08-01 0.32380961 15.360873 36.953119  9.627769
#> 427 214 2002-08-02 4.33419255 15.304621 36.736132 14.385778
#> 429 215 2002-08-03 5.28217602 15.247616 36.515097 15.453670
#> 432 216 2002-08-04 6.26627086 15.189886 36.290082 16.557879
#> 434 217 2002-08-05 2.12257546 15.131457 36.061156 11.544541
#> 435 218 2002-08-06 1.57700064 15.072356 35.828388 10.831435
#> 437 219 2002-08-07 6.00177342 15.012610 35.591848 16.012455
#> 439 220 2002-08-08 2.02020417 14.952243 35.351609 11.226088
#> 442 221 2002-08-09 0.99813408 14.891281 35.107744  9.953538
#> 443 222 2002-08-10 3.24592311 14.829748 34.860325 12.530181
#> 445 223 2002-08-11 0.10682561 14.767666 34.609429  8.777535
#> 447 224 2002-08-12 2.21368779 14.705060 34.355129 11.174679
#> 450 225 2002-08-13 5.61272817 14.641952 34.097505 15.059708
#> 451 226 2002-08-14 6.78930756 14.578362 33.836632 16.338208
#> 453 227 2002-08-15 0.04480125 14.514313 33.572591  8.444962
#> 455 228 2002-08-16 4.85777199 14.449824 33.305462 13.924715
#> 458 229 2002-08-17 2.66470290 14.384916 33.035326 11.318610
#> 460 230 2002-08-18 3.18386679 14.319607 32.762264 11.832799
#> 461 231 2002-08-19 6.62180226 14.253916 32.486361 15.667526
#> 463 232 2002-08-20 3.15502594 14.187862 32.207701 11.633019
#> 465 233 2002-08-21 3.77608221 14.121461 31.926370 12.250152
#> 468 234 2002-08-22 5.90380556 14.054730 31.642455 14.556451
#> 469 235 2002-08-23 5.31304510 13.987686 31.356043 13.794108
#> 471 236 2002-08-24 1.01437579 13.920345 31.067223  8.898740
#> 473 237 2002-08-25 2.75127665 13.852721 30.776085 10.750227
#> 476 238 2002-08-26 0.88408504 13.784830 30.482721  8.598179
#> 477 239 2002-08-27 6.82069106 13.716686 30.187223 15.052180
#> 479 240 2002-08-28 4.62116200 13.648302 29.889684 12.532577
#> 481 241 2002-08-29 0.76133905 13.579692 29.590199  8.227030
#> 484 242 2002-08-30 3.21078040 13.510869 29.288864 10.802381
#> 485 243 2002-08-31 5.48370162 13.441846 28.985774 13.158925
#> 487 244 2002-09-01 3.98470628 13.372633 28.681027 11.443367
#> 490 245 2002-09-02 4.09066682 13.303244 28.374722 11.456208
#> 491 246 2002-09-03 5.32596748 13.233689 28.066958 12.664587
#> 493 247 2002-09-04 4.58557022 13.163980 27.757836 11.774073
#> 495 248 2002-09-05 6.22809121 13.094126 27.447457 13.389420
#> 498 249 2002-09-06 6.88115889 13.024138 27.135924 13.952463
#> 499 250 2002-09-07 2.13892318 12.954027 26.823338  8.920322
#> 501 251 2002-09-08 0.76188846 12.883802 26.509805  7.411285
#> 503 252 2002-09-09 0.67541700 12.813472 26.195428  7.239257
#> 506 253 2002-09-10 2.27813108 12.743047 25.880314  8.783448
#> 507 254 2002-09-11 5.70527986 12.672536 25.564567 12.145831
#> 509 255 2002-09-12 6.34449657 12.601947 25.248295 12.667747
#> 511 256 2002-09-13 0.78000966 12.531289 24.931604  7.008834
#> 514 257 2002-09-14 0.14060439 12.460572 24.614602  6.292525
#> 516 258 2002-09-15 1.70646887 12.389802 24.297397  7.747610
#> 517 259 2002-09-16 5.41303296 12.318988 23.980097 11.263519
#> 519 260 2002-09-17 4.27688513 12.248139 23.662812 10.047071
#> 521 261 2002-09-18 0.40774727 12.177262 23.345649  6.227269
#> 524 262 2002-09-19 6.74436496 12.106365 23.028718 12.171743
#> 525 263 2002-09-20 1.76499712 12.035456 22.712128  7.343397
#> 527 264 2002-09-21 3.39830693 11.964544 22.395989  8.779580
#> 529 265 2002-09-22 0.33374385 11.893635 22.080409  5.829898
#> 532 266 2002-09-23 3.60039672 11.822738 21.765497  8.755514
#> 533 267 2002-09-24 1.01086204 11.751861 21.451363  6.285434
#> 535 268 2002-09-25 1.23446262 11.681012 21.138114  6.401478
#> 537 269 2002-09-26 3.11134856 11.610198 20.825859  7.996964
#> 540 270 2002-09-27 5.76725504 11.539428 20.514705 10.255167
#> 542 271 2002-09-28 5.71444178 11.468711 20.204761 10.084839
#> 543 272 2002-09-29 5.00761992 11.398053 19.896131  9.344615
#> 545 273 2002-09-30 2.08054037 11.327464 19.588923  6.696201
#> 547 274 2002-10-01 3.92316911 11.256953 19.283241  8.181018
#> 550 275 2002-10-02 2.83916299 11.186528 18.979189  7.153275
#> 551 276 2002-10-03 3.84999165 11.116198 18.676871  7.903498
#> 553 277 2002-10-04 2.68631457 11.045973 18.376389  6.828611
#> 555 278 2002-10-05 5.93862348 10.975862 18.077844  9.410080
#> 558 279 2002-10-06 6.18384814 10.905874 17.781337  9.486520
#> 559 280 2002-10-07 0.32472375 10.836020 17.486965  4.633758
#> 561 281 2002-10-08 4.07199437 10.766311 17.194826  7.550388
#> 563 282 2002-10-09 6.74525178 10.696756 16.905016  9.556309
#> 566 283 2002-10-10 5.71754589 10.627367 16.617629  8.624567
#> 567 284 2002-10-11 5.16081003 10.558154 16.332759  8.074904
#> 570 285 2002-10-12 6.23586266 10.489131 16.050495  8.783690
#> 571 286 2002-10-13 5.23177896 10.420308 15.770928  7.901829
#> 573 287 2002-10-14 5.24398899 10.351698 15.494145  7.798068
#> 575 288 2002-10-15 5.81532545 10.283314 15.220231  8.108661
#> 578 289 2002-10-16 0.64267933 10.215170 14.949271  4.207578
#> 579 290 2002-10-17 0.14146357 10.147279 14.681345  3.772673
#> 581 291 2002-10-18 1.41738011 10.079655 14.416534  4.617745
#> 583 292 2002-10-19 1.22004930 10.012314 14.154915  4.401151
#> 586 293 2002-10-20 3.43659547  9.945270 13.896562  5.875124
#> 587 294 2002-10-21 1.25292492  9.878539 13.641550  4.275487
#> 589 295 2002-10-22 4.63322866  9.812138 13.389948  6.508811
#> 591 296 2002-10-23 4.40396795  9.746084 13.141825  6.254658
#> 594 297 2002-10-24 4.10570135  9.680393 12.897247  5.959337
#> 596 298 2002-10-25 0.80387699  9.615084 12.656277  3.693138
#> 597 299 2002-10-26 3.89459078  9.550176 12.418977  5.636992
#> 599 300 2002-10-27 1.90794640  9.485687 12.185404  4.271834
#> 601 301 2002-10-28 2.40662999  9.421638 11.955615  4.515854
#> 604 302 2002-10-29 4.42589292  9.358048 11.729663  5.706190
#> 605 303 2002-10-30 6.95455280  9.294940 11.507600  7.181942
#> 607 304 2002-10-31 1.68460852  9.232334 11.289473  3.852354
#> 609 305 2002-11-01 2.87824205  9.170252 11.075328  4.506924
#> 612 306 2002-11-02 6.77880148  9.108719 10.865209  6.759302
#> 613 307 2002-11-03 4.74568922  9.047757 10.659156  5.460236
#> 615 308 2002-11-04 2.87547726  8.987390 10.457208  4.287171
#> 617 309 2002-11-05 2.32672356  8.927644 10.259398  3.901752
#> 620 310 2002-11-06 3.91088824  8.868543 10.065762  4.735862
#> 622 311 2002-11-07 4.55059409  8.810114  9.876328  5.019739
#> 623 312 2002-11-08 6.68317258  8.752384  9.691126  6.122772
#> 625 313 2002-11-09 4.75900555  8.695379  9.510180  4.980019
#> 627 314 2002-11-10 3.59117730  8.639127  9.333513  4.273291
#> 630 315 2002-11-11 6.64382889  8.583658  9.161147  5.835692
#> 631 316 2002-11-12 4.21492166  8.528999  8.993100  4.470412
#> 633 317 2002-11-13 3.11429060  8.475180  8.829388  3.829571
#> 635 318 2002-11-14 2.63668067  8.422232  8.670025  3.524634
#> 638 319 2002-11-15 4.69829786  8.370184  8.515023  4.518555
#> 639 320 2002-11-16 5.92987479  8.319067  8.364392  5.072189
#> 641 321 2002-11-17 4.00526015  8.268912  8.218139  4.044868
#> 643 322 2002-11-18 3.79664856  8.219751  8.076271  3.884256
#> 646 323 2002-11-19 0.93596302  8.171616  7.938791  2.439346
#> 647 324 2002-11-20 1.66913569  8.124538  7.805702  2.753242
#> 650 325 2002-11-21 2.40371746  8.078549  7.677004  3.061371
#> 651 326 2002-11-22 2.96623382  8.033683  7.552698  3.282496
#> 653 327 2002-11-23 5.83143640  7.989971  7.432780  4.570582
#> 655 328 2002-11-24 5.31037464  7.947446  7.317248  4.273954
#> 658 329 2002-11-25 5.76833561  7.906140  7.206095  4.430315
#> 659 330 2002-11-26 4.10005119  7.866085  7.099317  3.625023
#> 661 331 2002-11-27 4.61157049  7.827314  6.996907  3.810389
#> 663 332 2002-11-28 1.10700818  7.789856  6.898857  2.214909
#> 666 333 2002-11-29 2.86117171  7.753745  6.805157  2.956858
#> 667 334 2002-11-30 6.81843381  7.719011  6.715800  4.645084
#> 669 335 2002-12-01 3.37092786  7.685684  6.630775  3.111817
#> 671 336 2002-12-02 1.58484554  7.653793  6.550072  2.315669
#> 674 337 2002-12-03 5.35932541  7.623368  6.473680  3.893960
#> 676 338 2002-12-04 0.68241628  7.594437  6.401590  1.888012
#> 677 339 2002-12-05 5.71363287  7.567027  6.333789  3.974674
#> 679 340 2002-12-06 5.99910479  7.541163  6.270268  4.061612
#> 681 341 2002-12-07 6.92585780  7.516872  6.211016  4.414092
#> 684 342 2002-12-08 5.27165478  7.494177  6.156023  3.704182
#> 685 343 2002-12-09 0.76269407  7.473100  6.105278  1.837868
#> 687 344 2002-12-10 4.76264750  7.453664  6.058772  3.450372
#> 689 345 2002-12-11 3.76744222  7.435887  6.016497  3.028273
#> 692 346 2002-12-12 6.23619758  7.419789  5.978443  4.006997
#> 693 347 2002-12-13 3.04955304  7.405386  5.944603  2.710150
#> 695 348 2002-12-14 1.13011269  7.392694  5.914970  1.930850
#> 697 349 2002-12-15 2.33239979  7.381725  5.889539  2.402842
#> 700 350 2002-12-16 2.38042016  7.372493  5.868303  2.414451
#> 702 351 2002-12-17 0.47323944  7.365006  5.851259  1.650802
#> 703 352 2002-12-18 4.67359302  7.359273  5.838404  3.313475
#> 705 353 2002-12-19 5.73714750  7.355300  5.829736  3.731037
#> 707 354 2002-12-20 0.57863179  7.353091  5.825253  1.685515
#> 710 355 2002-12-21 6.70803014  7.352649  5.824956  4.113375
#> 711 356 2002-12-22 2.67151453  7.353975  5.828845  2.515948
#> 713 357 2002-12-23 3.49299459  7.357066  5.836924  2.844861
#> 715 358 2002-12-24 5.61246982  7.361920  5.849194  3.691909
#> 718 359 2002-12-25 1.73915973  7.368531  5.865662  2.158638
#> 719 360 2002-12-26 2.06280198  7.376892  5.886331  2.294581
#> 721 361 2002-12-27 6.09863606  7.386993  5.911209  3.917923
#> 723 362 2002-12-28 0.39716993  7.398825  5.940303  1.644514
#> 726 363 2002-12-29 4.93854392  7.412375  5.973622  3.483388
#> 727 364 2002-12-30 0.06932094  7.427627  6.011174  1.530844
#> 729 365 2002-12-31 6.89584304  7.444567  6.052971  4.316652
#> 2     1 2003-01-01 0.39315608  7.463176  6.099033  1.685405
#> 4     2 2003-01-02 6.65767997  7.483435  6.149345  4.272736
#> 5     3 2003-01-03 5.13201547  7.505323  6.203934  3.672057
#> 8     4 2003-01-04 4.81992611  7.528820  6.262814  3.570420
#> 10    5 2003-01-05 5.59136642  7.553900  6.325996  3.922738
#> 12    6 2003-01-06 3.55878803  7.580540  6.393496  3.099131
#> 13    7 2003-01-07 3.77158676  7.608714  6.465324  3.218739
#> 16    8 2003-01-08 4.30155822  7.638396  6.541496  3.477294
#> 18    9 2003-01-09 6.69894899  7.669557  6.622023  4.547498
#> 20   10 2003-01-10 2.57000847  7.702170  6.706920  2.795690
#> 21   11 2003-01-11 4.26265516  7.736204  6.796198  3.571405
#> 24   12 2003-01-12 1.69024424  7.771631  6.889871  2.471703
#> 26   13 2003-01-13 1.13675742  7.808419  6.987950  2.255644
#> 28   14 2003-01-14 5.12163890  7.846537  7.090446  4.086671
#> 29   15 2003-01-15 0.72307882  7.885954  7.197371  2.129313
#> 31   16 2003-01-16 2.58604859  7.926639  7.308732  3.019412
#> 34   17 2003-01-17 0.66155894  7.968559  7.424539  2.164332
#> 36   18 2003-01-18 3.05632233  8.011681  7.544799  3.325307
#> 38   19 2003-01-19 6.08630314  8.055974  7.669519  4.814548
#> 39   20 2003-01-20 6.97782342  8.101405  7.798703  5.308227
#> 42   21 2003-01-21 5.92757777  8.147942  7.932354  4.868458
#> 44   22 2003-01-22 2.10609761  8.195553  8.070474  3.054596
#> 46   23 2003-01-23 0.29736172  8.244205  8.213063  2.201385
#> 47   24 2003-01-24 1.17920243  8.293867  8.360119  2.684341
#> 50   25 2003-01-25 6.71167996  8.344507  8.511639  5.550964
#> 52   26 2003-01-26 2.28902036  8.396093  8.667618  3.348427
#> 54   27 2003-01-27 6.21440782  8.448595  8.828046  5.453769
#> 55   28 2003-01-28 1.52557069  8.501983  8.992915  3.055060
#> 57   29 2003-01-29 2.73618197  8.556225  9.162213  3.755538
#> 60   30 2003-01-30 5.59669586  8.611293  9.335925  5.367808
#> 62   31 2003-01-31 2.72260762  8.667157  9.514035  3.872827
#> 64   32 2003-02-01 4.31220658  8.723789  9.696523  4.820648
#> 65   33 2003-02-02 2.38912630  8.781160  9.883368  3.815346
#> 68   34 2003-02-03 1.15048279  8.839243 10.074546  3.174269
#> 70   35 2003-02-04 0.41929181  8.898011 10.270030  2.809479
#> 72   36 2003-02-05 1.46346990  8.957438 10.469790  3.472727
#> 73   37 2003-02-06 6.54794323  9.017498 10.673795  6.543770
#> 76   38 2003-02-07 2.48032514  9.078165 10.882010  4.207087
#> 78   39 2003-02-08 4.56244619  9.139416 11.094398  5.542792
#> 80   40 2003-02-09 4.96403647  9.201226 11.310918  5.878834
#> 81   41 2003-02-10 2.19433792  9.263572 11.531528  4.248666
#> 84   42 2003-02-11 6.31336605  9.326432 11.756182  6.918117
#> 85   43 2003-02-12 3.43900229  9.389784 11.984831  5.190926
#> 88   44 2003-02-13 2.68645718  9.453606 12.217425  4.790286
#> 90   45 2003-02-14 3.42210893  9.517878 12.453910  5.352350
#> 92   46 2003-02-15 6.34332636  9.582579 12.694229  7.375121
#> 93   47 2003-02-16 2.57755373  9.647690 12.938322  4.962933
#> 96   48 2003-02-17 0.55915326  9.713192 13.186128  3.676071
#> 98   49 2003-02-18 5.78936252  9.779067 13.437583  7.337027
#> 100  50 2003-02-19 1.53305454  9.845297 13.692618  4.489223
#> 101  51 2003-02-20 6.99665694  9.911864 13.951164  8.411764
#> 104  52 2003-02-21 2.53112205  9.978754 14.213148  5.355877
#> 106  53 2003-02-22 0.51804845 10.045948 14.478496  3.992937
#> 108  54 2003-02-23 3.85522296 10.113433 14.747130  6.497573
#> 109  55 2003-02-24 6.59954862 10.181192 15.018971  8.622465
#> 111  56 2003-02-25 6.56223657 10.249211 15.293936  8.719589
#> 114  57 2003-02-26 6.07824633 10.317477 15.571940  8.479867
#> 116  58 2003-02-27 3.90087209 10.385975 15.852898  6.940322
#> 118  59 2003-02-28 4.21124726 10.454693 16.136720  7.284190
#> 119  60 2003-03-01 2.02123090 10.523618 16.423316  5.683010
#> 122  61 2003-03-02 1.90090078 10.592738 16.712592  5.677712
#> 124  62 2003-03-03 0.40881757 10.662040 17.004454  4.577117
#> 126  63 2003-03-04 2.94329109 10.731513 17.298805  6.696940
#> 127  64 2003-03-05 3.10241020 10.801147 17.595546  6.925868
#> 130  65 2003-03-06 2.98566395 10.870930 17.894578  6.930987
#> 132  66 2003-03-07 4.26955535 10.940852 18.195799  8.099312
#> 134  67 2003-03-08 5.02129728 11.010902 18.499104  8.842846
#> 135  68 2003-03-09 0.69387187 11.081072 18.804391  5.289842
#> 137  69 2003-03-10 0.07429720 11.151350 19.111552  4.841555
#> 140  70 2003-03-11 6.27569026 11.221729 19.420480 10.285519
#> 142  71 2003-03-12 3.25527009 11.292198 19.731067  7.776764
#> 144  72 2003-03-13 4.57931470 11.362749 20.043203  9.049618
#> 145  73 2003-03-14 2.94803549 11.433374 20.356778  7.713639
#> 148  74 2003-03-15 0.56132592 11.504062 20.671679  5.672244
#> 150  75 2003-03-16 2.89860182 11.574807 20.987796  7.874866
#> 152  76 2003-03-17 4.42825025 11.645600 21.305014  9.376879
#> 153  77 2003-03-18 0.46920738 11.716433 21.623220  5.838777
#> 156  78 2003-03-19 3.27081998 11.787297 21.942300  8.529925
#> 158  79 2003-03-20 0.50484093 11.858185 22.262140  6.039420
#> 160  80 2003-03-21 3.41424965 11.929088 22.582624  8.877366
#> 161  81 2003-03-22 0.04253733 12.000000 22.903638  5.766504
#> 164  82 2003-03-23 4.81676192 12.070912 23.225065 10.440117
#> 165  83 2003-03-24 1.50025424 12.141815 23.546791  7.341430
#> 168  84 2003-03-25 4.30613721 12.212703 23.868700 10.175166
#> 170  85 2003-03-26 4.55659126 12.283567 24.190677 10.534437
#> 172  86 2003-03-27 1.81894676 12.354400 24.512606  7.932656
#> 173  87 2003-03-28 5.49042420 12.425193 24.834373 11.695480
#> 176  88 2003-03-29 5.06363752 12.495938 25.155863 11.385829
#> 178  89 2003-03-30 1.00916220 12.566626 25.476961  7.392203
#> 180  90 2003-03-31 3.86442496 12.637251 25.797553 10.393787
#> 181  91 2003-04-01 5.98601007 12.707802 26.117528 12.680713
#> 184  92 2003-04-02 2.92174002 12.778271 26.436771  9.631564
#> 186  93 2003-04-03 4.58283422 12.848650 26.755170 11.460287
#> 188  94 2003-04-04 1.97568073 12.918928 27.072615  8.838250
#> 189  95 2003-04-05 2.98881537 12.989098 27.388994  9.998378
#> 191  96 2003-04-06 1.08049905 13.059148 27.704199  8.072157
#> 194  97 2003-04-07 5.73547802 13.129070 28.018120 13.124434
#> 196  98 2003-04-08 3.11708333 13.198853 28.330649 10.427991
#> 198  99 2003-04-09 6.49840857 13.268487 28.641680 14.174232
#> 199 100 2003-04-10 1.10051867 13.337960 28.951106  8.432158
#> 202 101 2003-04-11 5.84834132 13.407262 29.258822 13.696157
#> 204 102 2003-04-12 4.65753479 13.476382 29.564726 12.500073
#> 206 103 2003-04-13 2.80618996 13.545307 29.868714 10.561139
#> 207 104 2003-04-14 4.49402222 13.614025 30.170685 12.522379
#> 210 105 2003-04-15 0.34063993 13.682523 30.470539  7.996932
#> 212 106 2003-04-16 6.11554686 13.750789 30.768176 14.533987
#> 214 107 2003-04-17 5.84264428 13.818808 31.063500 14.332757
#> 215 108 2003-04-18 4.58491883 13.886567 31.356413 13.015566
#> 217 109 2003-04-19 6.41128897 13.954052 31.646820 15.181884
#> 220 110 2003-04-20 4.43835554 14.021246 31.934628 13.038030
#> 222 111 2003-04-21 2.54134760 14.088136 32.219743 10.960983
#> 224 112 2003-04-22 2.19398403 14.154703 32.502075 10.644435
#> 225 113 2003-04-23 5.20741777 14.220933 32.781534 14.197350
#> 228 114 2003-04-24 6.42264234 14.286808 33.058030 15.695135
#> 230 115 2003-04-25 4.31412050 14.352310 33.331477 13.342377
#> 232 116 2003-04-26 4.51156508 14.417421 33.601789 13.657860
#> 233 117 2003-04-27 0.95937214 14.482122 33.868881  9.589047
#> 236 118 2003-04-28 0.14004751 14.546394 34.132670  8.697476
#> 238 119 2003-04-29 4.18791911 14.610216 34.393074 13.527539
#> 240 120 2003-04-30 3.34075040 14.673568 34.650012 12.606910
#> 241 121 2003-05-01 2.07435345 14.736428 34.903405 11.182417
#> 244 122 2003-05-02 4.44830285 14.798774 35.153175 14.071568
#> 246 123 2003-05-03 1.87641940 14.860584 35.399245 11.084711
#> 247 124 2003-05-04 0.61929799 14.921835 35.641540  9.649997
#> 250 125 2003-05-05 5.73482853 14.982502 35.879987 15.836859
#> 252 126 2003-05-06 0.69418756 15.042562 36.114511  9.861938
#> 254 127 2003-05-07 2.99436233 15.101989 36.345041 12.689436
#> 255 128 2003-05-08 0.79592428 15.160757 36.571508 10.102861
#> 258 129 2003-05-09 5.30337431 15.218840 36.793843 15.609315
#> 260 130 2003-05-10 2.78114386 15.276211 37.011977 12.622142
#> 262 131 2003-05-11 5.67647111 15.332843 37.225844 16.197271
#> 263 132 2003-05-12 2.70391072 15.388707 37.435378 12.647682
#> 266 133 2003-05-13 5.12864917 15.443775 37.640517 15.660059
#> 268 134 2003-05-14 6.48864251 15.498017 37.841197 17.381893
#> 270 135 2003-05-15 1.88593473 15.551405 38.037357 11.815754
#> 271 136 2003-05-16 5.93270101 15.603907 38.228936 16.824672
#> 273 137 2003-05-17 2.57525893 15.655493 38.415875 12.763589
#> 276 138 2003-05-18 3.85118051 15.706133 38.598118 14.381704
#> 278 139 2003-05-19 6.89226155 15.755795 38.775606 18.174960
#> 280 140 2003-05-20 6.57648045 15.804447 38.948286 17.840571
#> 281 141 2003-05-21 0.14681792 15.852058 39.116103  9.960168
#> 284 142 2003-05-22 3.25567725 15.898595 39.279005 13.841483
#> 286 143 2003-05-23 6.04588113 15.944026 39.436940 17.336363
#> 288 144 2003-05-24 1.75453204 15.988319 39.589859 12.069728
#> 289 145 2003-05-25 3.66486298 16.031441 39.737713 14.476542
#> 292 146 2003-05-26 1.27905643 16.073361 39.880455 11.556881
#> 294 147 2003-05-27 3.72160548 16.114046 40.018039 14.625676
#> 296 148 2003-05-28 6.93792359 16.153463 40.150421 18.659922
#> 297 149 2003-05-29 6.82845044 16.191581 40.277558 18.562485
#> 299 150 2003-05-30 0.54363758 16.228369 40.399408 10.776526
#> 302 151 2003-05-31 4.55976130 16.263796 40.515933 15.808561
#> 304 152 2003-06-01 3.31892027 16.297830 40.627093 14.293461
#> 306 153 2003-06-02 4.84313694 16.330443 40.732853 16.223306
#> 307 154 2003-06-03 3.05486800 16.361604 40.833176 14.020266
#> 310 155 2003-06-04 5.76451244 16.391286 40.928031 17.428824
#> 312 156 2003-06-05 4.09571490 16.419460 41.017385 15.370091
#> 314 157 2003-06-06 1.38204393 16.446100 41.101209 12.002267
#> 315 158 2003-06-07 0.53273060 16.471180 41.179475 10.960807
#> 318 159 2003-06-08 6.51957291 16.494677 41.252156 18.465562
#> 320 160 2003-06-09 0.50068789 16.516565 41.319229 10.956089
#> 322 161 2003-06-10 0.70037781 16.536824 41.380671 11.221458
#> 323 162 2003-06-11 2.13194594 16.555433 41.436461 13.027130
#> 326 163 2003-06-12 3.93096772 16.572373 41.486582 15.291955
#> 327 164 2003-06-13 3.66423522 16.587625 41.531016 14.969891
#> 330 165 2003-06-14 1.32003546 16.601175 41.569750 12.045138
#> 332 166 2003-06-15 4.95851592 16.613007 41.602770 16.609322
#> 334 167 2003-06-16 3.46809344 16.623108 41.630068 14.750175
#> 335 168 2003-06-17 5.65308047 16.631469 41.651634 17.491659
#> 338 169 2003-06-18 0.73648559 16.638080 41.667464 11.339072
#> 340 170 2003-06-19 2.63041146 16.642934 41.677552 13.712951
#> 342 171 2003-06-20 6.74323877 16.646025 41.681897 18.863060
#> 343 172 2003-06-21 5.29965306 16.647351 41.680500 17.054580
#> 346 173 2003-06-22 2.89389100 16.646909 41.673363 14.040580
#> 348 174 2003-06-23 0.95010626 16.644700 41.660491 11.604147
#> 350 175 2003-06-24 3.84942006 16.640727 41.641891 15.226883
#> 351 176 2003-06-25 2.44958101 16.634994 41.617572 13.468584
#> 353 177 2003-06-26 1.51034004 16.627507 41.587545 12.285664
#> 356 178 2003-06-27 3.61359667 16.618275 41.551823 14.905619
#> 358 179 2003-06-28 2.37510974 16.607306 41.510423 13.345932
#> 360 180 2003-06-29 3.46193259 16.594614 41.463360 14.690838
#> 361 181 2003-06-30 6.13956313 16.580211 41.410655 18.019735
#> 364 182 2003-07-01 4.68517011 16.564113 41.352328 16.186349
#> 366 183 2003-07-02 6.98602528 16.546336 41.288405 19.038286
#> 368 184 2003-07-03 2.26327106 16.526900 41.218909 13.127082
#> 369 185 2003-07-04 3.12668569 16.505823 41.143868 14.182893
#> 372 186 2003-07-05 2.13040549 16.483128 41.063311 12.919496
#> 374 187 2003-07-06 5.32391827 16.458837 40.977269 16.871749
#> 376 188 2003-07-07 3.08077859 16.432973 40.885775 14.053983
#> 377 189 2003-07-08 2.98996951 16.405563 40.788864 13.914170
#> 379 190 2003-07-09 0.43908723 16.376632 40.686570 10.717083
#> 382 191 2003-07-10 0.97942110 16.346207 40.578932 11.360424
#> 384 192 2003-07-11 6.21233535 16.314316 40.465990 17.821028
#> 386 193 2003-07-12 1.91754863 16.280989 40.347784 12.462994
#> 387 194 2003-07-13 0.45357451 16.246255 40.224356 10.617595
#> 390 195 2003-07-14 3.45006372 16.210144 40.095751 14.290800
#> 392 196 2003-07-15 3.93071725 16.172686 39.962013 14.846820
#> 394 197 2003-07-16 5.58018094 16.133915 39.823189 16.842551
#> 395 198 2003-07-17 4.78303066 16.093860 39.679327 15.816100
#> 398 199 2003-07-18 6.90648742 16.052554 39.530476 18.386460
#> 400 200 2003-07-19 6.04464246 16.010029 39.376687 17.277575
#> 402 201 2003-07-20 1.60775896 15.966317 39.218012 11.779070
#> 403 202 2003-07-21 3.03079690 15.921451 39.054503 13.480820
#> 406 203 2003-07-22 3.50949703 15.875462 38.886214 14.019729
#> 407 204 2003-07-23 6.18767513 15.828384 38.713200 17.245235
#> 410 205 2003-07-24 3.81268039 15.780249 38.535519 14.289181
#> 412 206 2003-07-25 0.38054690 15.731088 38.353226 10.052203
#> 414 207 2003-07-26 2.33134659 15.680933 38.166382 12.378769
#> 415 208 2003-07-27 2.29427126 15.629816 37.975044 12.280904
#> 418 209 2003-07-28 2.62476022 15.577768 37.779274 12.627609
#> 420 210 2003-07-29 2.16175479 15.524820 37.579134 12.011138
#> 422 211 2003-07-30 3.98854946 15.471001 37.374685 14.161420
#> 423 212 2003-07-31 2.62982864 15.416342 37.165992 12.461517
#> 426 213 2003-08-01 0.32380961 15.360873 36.953119  9.627769
#> 428 214 2003-08-02 4.33419255 15.304621 36.736132 14.385778
#> 430 215 2003-08-03 5.28217602 15.247616 36.515097 15.453670
#> 431 216 2003-08-04 6.26627086 15.189886 36.290082 16.557879
#> 433 217 2003-08-05 2.12257546 15.131457 36.061156 11.544541
#> 436 218 2003-08-06 1.57700064 15.072356 35.828388 10.831435
#> 438 219 2003-08-07 6.00177342 15.012610 35.591848 16.012455
#> 440 220 2003-08-08 2.02020417 14.952243 35.351609 11.226088
#> 441 221 2003-08-09 0.99813408 14.891281 35.107744  9.953538
#> 444 222 2003-08-10 3.24592311 14.829748 34.860325 12.530181
#> 446 223 2003-08-11 0.10682561 14.767666 34.609429  8.777535
#> 448 224 2003-08-12 2.21368779 14.705060 34.355129 11.174679
#> 449 225 2003-08-13 5.61272817 14.641952 34.097505 15.059708
#> 452 226 2003-08-14 6.78930756 14.578362 33.836632 16.338208
#> 454 227 2003-08-15 0.04480125 14.514313 33.572591  8.444962
#> 456 228 2003-08-16 4.85777199 14.449824 33.305462 13.924715
#> 457 229 2003-08-17 2.66470290 14.384916 33.035326 11.318610
#> 459 230 2003-08-18 3.18386679 14.319607 32.762264 11.832799
#> 462 231 2003-08-19 6.62180226 14.253916 32.486361 15.667526
#> 464 232 2003-08-20 3.15502594 14.187862 32.207701 11.633019
#> 466 233 2003-08-21 3.77608221 14.121461 31.926370 12.250152
#> 467 234 2003-08-22 5.90380556 14.054730 31.642455 14.556451
#> 470 235 2003-08-23 5.31304510 13.987686 31.356043 13.794108
#> 472 236 2003-08-24 1.01437579 13.920345 31.067223  8.898740
#> 474 237 2003-08-25 2.75127665 13.852721 30.776085 10.750227
#> 475 238 2003-08-26 0.88408504 13.784830 30.482721  8.598179
#> 478 239 2003-08-27 6.82069106 13.716686 30.187223 15.052180
#> 480 240 2003-08-28 4.62116200 13.648302 29.889684 12.532577
#> 482 241 2003-08-29 0.76133905 13.579692 29.590199  8.227030
#> 483 242 2003-08-30 3.21078040 13.510869 29.288864 10.802381
#> 486 243 2003-08-31 5.48370162 13.441846 28.985774 13.158925
#> 488 244 2003-09-01 3.98470628 13.372633 28.681027 11.443367
#> 489 245 2003-09-02 4.09066682 13.303244 28.374722 11.456208
#> 492 246 2003-09-03 5.32596748 13.233689 28.066958 12.664587
#> 494 247 2003-09-04 4.58557022 13.163980 27.757836 11.774073
#> 496 248 2003-09-05 6.22809121 13.094126 27.447457 13.389420
#> 497 249 2003-09-06 6.88115889 13.024138 27.135924 13.952463
#> 500 250 2003-09-07 2.13892318 12.954027 26.823338  8.920322
#> 502 251 2003-09-08 0.76188846 12.883802 26.509805  7.411285
#> 504 252 2003-09-09 0.67541700 12.813472 26.195428  7.239257
#> 505 253 2003-09-10 2.27813108 12.743047 25.880314  8.783448
#> 508 254 2003-09-11 5.70527986 12.672536 25.564567 12.145831
#> 510 255 2003-09-12 6.34449657 12.601947 25.248295 12.667747
#> 512 256 2003-09-13 0.78000966 12.531289 24.931604  7.008834
#> 513 257 2003-09-14 0.14060439 12.460572 24.614602  6.292525
#> 515 258 2003-09-15 1.70646887 12.389802 24.297397  7.747610
#> 518 259 2003-09-16 5.41303296 12.318988 23.980097 11.263519
#> 520 260 2003-09-17 4.27688513 12.248139 23.662812 10.047071
#> 522 261 2003-09-18 0.40774727 12.177262 23.345649  6.227269
#> 523 262 2003-09-19 6.74436496 12.106365 23.028718 12.171743
#> 526 263 2003-09-20 1.76499712 12.035456 22.712128  7.343397
#> 528 264 2003-09-21 3.39830693 11.964544 22.395989  8.779580
#> 530 265 2003-09-22 0.33374385 11.893635 22.080409  5.829898
#> 531 266 2003-09-23 3.60039672 11.822738 21.765497  8.755514
#> 534 267 2003-09-24 1.01086204 11.751861 21.451363  6.285434
#> 536 268 2003-09-25 1.23446262 11.681012 21.138114  6.401478
#> 538 269 2003-09-26 3.11134856 11.610198 20.825859  7.996964
#> 539 270 2003-09-27 5.76725504 11.539428 20.514705 10.255167
#> 541 271 2003-09-28 5.71444178 11.468711 20.204761 10.084839
#> 544 272 2003-09-29 5.00761992 11.398053 19.896131  9.344615
#> 546 273 2003-09-30 2.08054037 11.327464 19.588923  6.696201
#> 548 274 2003-10-01 3.92316911 11.256953 19.283241  8.181018
#> 549 275 2003-10-02 2.83916299 11.186528 18.979189  7.153275
#> 552 276 2003-10-03 3.84999165 11.116198 18.676871  7.903498
#> 554 277 2003-10-04 2.68631457 11.045973 18.376389  6.828611
#> 556 278 2003-10-05 5.93862348 10.975862 18.077844  9.410080
#> 557 279 2003-10-06 6.18384814 10.905874 17.781337  9.486520
#> 560 280 2003-10-07 0.32472375 10.836020 17.486965  4.633758
#> 562 281 2003-10-08 4.07199437 10.766311 17.194826  7.550388
#> 564 282 2003-10-09 6.74525178 10.696756 16.905016  9.556309
#> 565 283 2003-10-10 5.71754589 10.627367 16.617629  8.624567
#> 568 284 2003-10-11 5.16081003 10.558154 16.332759  8.074904
#> 569 285 2003-10-12 6.23586266 10.489131 16.050495  8.783690
#> 572 286 2003-10-13 5.23177896 10.420308 15.770928  7.901829
#> 574 287 2003-10-14 5.24398899 10.351698 15.494145  7.798068
#> 576 288 2003-10-15 5.81532545 10.283314 15.220231  8.108661
#> 577 289 2003-10-16 0.64267933 10.215170 14.949271  4.207578
#> 580 290 2003-10-17 0.14146357 10.147279 14.681345  3.772673
#> 582 291 2003-10-18 1.41738011 10.079655 14.416534  4.617745
#> 584 292 2003-10-19 1.22004930 10.012314 14.154915  4.401151
#> 585 293 2003-10-20 3.43659547  9.945270 13.896562  5.875124
#> 588 294 2003-10-21 1.25292492  9.878539 13.641550  4.275487
#> 590 295 2003-10-22 4.63322866  9.812138 13.389948  6.508811
#> 592 296 2003-10-23 4.40396795  9.746084 13.141825  6.254658
#> 593 297 2003-10-24 4.10570135  9.680393 12.897247  5.959337
#> 595 298 2003-10-25 0.80387699  9.615084 12.656277  3.693138
#> 598 299 2003-10-26 3.89459078  9.550176 12.418977  5.636992
#> 600 300 2003-10-27 1.90794640  9.485687 12.185404  4.271834
#> 602 301 2003-10-28 2.40662999  9.421638 11.955615  4.515854
#> 603 302 2003-10-29 4.42589292  9.358048 11.729663  5.706190
#> 606 303 2003-10-30 6.95455280  9.294940 11.507600  7.181942
#> 608 304 2003-10-31 1.68460852  9.232334 11.289473  3.852354
#> 610 305 2003-11-01 2.87824205  9.170252 11.075328  4.506924
#> 611 306 2003-11-02 6.77880148  9.108719 10.865209  6.759302
#> 614 307 2003-11-03 4.74568922  9.047757 10.659156  5.460236
#> 616 308 2003-11-04 2.87547726  8.987390 10.457208  4.287171
#> 618 309 2003-11-05 2.32672356  8.927644 10.259398  3.901752
#> 619 310 2003-11-06 3.91088824  8.868543 10.065762  4.735862
#> 621 311 2003-11-07 4.55059409  8.810114  9.876328  5.019739
#> 624 312 2003-11-08 6.68317258  8.752384  9.691126  6.122772
#> 626 313 2003-11-09 4.75900555  8.695379  9.510180  4.980019
#> 628 314 2003-11-10 3.59117730  8.639127  9.333513  4.273291
#> 629 315 2003-11-11 6.64382889  8.583658  9.161147  5.835692
#> 632 316 2003-11-12 4.21492166  8.528999  8.993100  4.470412
#> 634 317 2003-11-13 3.11429060  8.475180  8.829388  3.829571
#> 636 318 2003-11-14 2.63668067  8.422232  8.670025  3.524634
#> 637 319 2003-11-15 4.69829786  8.370184  8.515023  4.518555
#> 640 320 2003-11-16 5.92987479  8.319067  8.364392  5.072189
#> 642 321 2003-11-17 4.00526015  8.268912  8.218139  4.044868
#> 644 322 2003-11-18 3.79664856  8.219751  8.076271  3.884256
#> 645 323 2003-11-19 0.93596302  8.171616  7.938791  2.439346
#> 648 324 2003-11-20 1.66913569  8.124538  7.805702  2.753242
#> 649 325 2003-11-21 2.40371746  8.078549  7.677004  3.061371
#> 652 326 2003-11-22 2.96623382  8.033683  7.552698  3.282496
#> 654 327 2003-11-23 5.83143640  7.989971  7.432780  4.570582
#> 656 328 2003-11-24 5.31037464  7.947446  7.317248  4.273954
#> 657 329 2003-11-25 5.76833561  7.906140  7.206095  4.430315
#> 660 330 2003-11-26 4.10005119  7.866085  7.099317  3.625023
#> 662 331 2003-11-27 4.61157049  7.827314  6.996907  3.810389
#> 664 332 2003-11-28 1.10700818  7.789856  6.898857  2.214909
#> 665 333 2003-11-29 2.86117171  7.753745  6.805157  2.956858
#> 668 334 2003-11-30 6.81843381  7.719011  6.715800  4.645084
#> 670 335 2003-12-01 3.37092786  7.685684  6.630775  3.111817
#> 672 336 2003-12-02 1.58484554  7.653793  6.550072  2.315669
#> 673 337 2003-12-03 5.35932541  7.623368  6.473680  3.893960
#> 675 338 2003-12-04 0.68241628  7.594437  6.401590  1.888012
#> 678 339 2003-12-05 5.71363287  7.567027  6.333789  3.974674
#> 680 340 2003-12-06 5.99910479  7.541163  6.270268  4.061612
#> 682 341 2003-12-07 6.92585780  7.516872  6.211016  4.414092
#> 683 342 2003-12-08 5.27165478  7.494177  6.156023  3.704182
#> 686 343 2003-12-09 0.76269407  7.473100  6.105278  1.837868
#> 688 344 2003-12-10 4.76264750  7.453664  6.058772  3.450372
#> 690 345 2003-12-11 3.76744222  7.435887  6.016497  3.028273
#> 691 346 2003-12-12 6.23619758  7.419789  5.978443  4.006997
#> 694 347 2003-12-13 3.04955304  7.405386  5.944603  2.710150
#> 696 348 2003-12-14 1.13011269  7.392694  5.914970  1.930850
#> 698 349 2003-12-15 2.33239979  7.381725  5.889539  2.402842
#> 699 350 2003-12-16 2.38042016  7.372493  5.868303  2.414451
#> 701 351 2003-12-17 0.47323944  7.365006  5.851259  1.650802
#> 704 352 2003-12-18 4.67359302  7.359273  5.838404  3.313475
#> 706 353 2003-12-19 5.73714750  7.355300  5.829736  3.731037
#> 708 354 2003-12-20 0.57863179  7.353091  5.825253  1.685515
#> 709 355 2003-12-21 6.70803014  7.352649  5.824956  4.113375
#> 712 356 2003-12-22 2.67151453  7.353975  5.828845  2.515948
#> 714 357 2003-12-23 3.49299459  7.357066  5.836924  2.844861
#> 716 358 2003-12-24 5.61246982  7.361920  5.849194  3.691909
#> 717 359 2003-12-25 1.73915973  7.368531  5.865662  2.158638
#> 720 360 2003-12-26 2.06280198  7.376892  5.886331  2.294581
#> 722 361 2003-12-27 6.09863606  7.386993  5.911209  3.917923
#> 724 362 2003-12-28 0.39716993  7.398825  5.940303  1.644514
#> 725 363 2003-12-29 4.93854392  7.412375  5.973622  3.483388
#> 728 364 2003-12-30 0.06932094  7.427627  6.011174  1.530844
#> 730 365 2003-12-31 6.89584304  7.444567  6.052971  4.316652