Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  3.230712  3.110289  3.966258  1.958410  2.774166  2.822609  4.121998
#>   [8]  1.840059  2.951120  3.208805  2.297657  2.046024  4.517620  4.848027
#>  [15]  3.968016  2.185826  4.476663  3.279971  4.647565  3.501787  4.782940
#>  [22]  2.721251  3.230761  2.294403  3.417405  3.438153  2.351458  5.703487
#>  [29]  3.581410  5.283797  5.969450  3.159221  5.861914  5.424516  2.974851
#>  [36]  5.871105  4.426526  4.760969  3.745659  3.766448  3.309613  4.792017
#>  [43]  3.440552  6.267766  5.993912  4.141054  5.080168  7.511983  4.533612
#>  [50]  7.206602  7.908292  6.760793  6.197943  4.381430  8.662746  7.391893
#>  [57]  6.753090  8.464060  6.140890  9.166201  9.559050  6.006050  5.014466
#>  [64]  6.362909  8.566745  9.431841  6.401311  6.316070  5.721783  8.647017
#>  [71]  7.144258  7.195950  6.773036 10.317239  8.901508  6.486238  7.994296
#>  [78]  8.938130  7.461917  7.754794 11.731718 10.830151  7.175970  8.731612
#>  [85] 12.893474  6.872511  9.796804 10.239332 13.427412 12.249403  9.907493
#>  [92] 11.770472 13.123331 11.209306 10.740421 10.006010 13.730064 13.343748
#>  [99] 14.116675  9.745595  8.187391 12.579255 14.171207 11.838973 13.008935
#> [106] 14.029697 10.596894 10.762475 11.384228  9.322708 14.708905 11.923657
#> [113] 15.458795 14.942371 15.995086 13.037616  9.129182 16.589181 14.525398
#> [120] 14.976213 15.708068 13.176380 12.122945 15.751663  9.413289  9.398419
#> [127] 14.892835  9.515358 16.832942 15.082143 11.117489 15.625402 15.917348
#> [134] 12.764109 15.670778 13.015553  9.925738 14.147872 13.694284 18.236260
#> [141] 11.452676 15.151826 15.866266 16.019692 10.235869 18.375631 11.171157
#> [148] 17.490180 14.736091 15.607704 15.662691 16.030606 11.294397 11.504623
#> [155] 16.131736 11.705676 18.319865 11.294477 14.559430 16.300930 12.077850
#> [162] 11.763041 14.922982 17.718836 15.908140 17.980019 17.987988 18.546784
#> [169] 15.029106 13.850958 13.508163 18.088881 15.590679 17.121397 15.940580
#> [176] 18.366937 17.308475 18.199546 17.935220 17.950364 17.177804 13.473610
#> [183] 17.414109 18.118601 17.665305 11.838363 13.139711 13.942087 17.776858
#> [190] 14.860952 15.254394 11.882689 18.392832 17.084194 12.559342 11.389272
#> [197] 14.500957 14.203032 16.126003 14.025180 18.382909 10.483167 10.010072
#> [204] 12.473332 18.155006 15.733777 14.120553 13.523461 12.737321 13.638974
#> [211] 10.825494 11.410562 14.176195 15.228004 14.985213 14.380866 12.462397
#> [218] 11.978274 16.913024 13.404252 15.895878  8.798605 14.261656 11.803530
#> [225] 16.331666  9.063051  9.547315 12.842188  8.274750 10.803270  8.146016
#> [232]  9.138498 11.894970  7.942495 12.001607 10.392129 12.652054 13.234728
#> [239] 12.304728 14.023956  8.336492 14.868048  9.790945  8.455680 13.905836
#> [246] 12.982823 12.859969  8.810573  8.447053  9.219475  7.045360 13.530305
#> [253]  9.746582  8.612813 10.292713  7.774144  6.166407  8.119020 10.688542
#> [260] 11.262188  6.866968 11.549314  7.809431  8.645601  9.211405  7.669108
#> [267]  8.223861  7.473650 10.814260  8.114709  7.446926 10.896466  7.185164
#> [274] 10.619322  9.850314  6.909465  6.503846  9.753144  7.802168  4.480319
#> [281]  7.193782  9.022246  6.822572  6.973913  5.053357  9.013984  6.197110
#> [288]  5.164396  3.887635  4.378655  5.237929  5.471815  3.525583  5.151607
#> [295]  6.347854  4.190182  5.206192  6.033909  6.912734  3.316158  3.461629
#> [302]  5.440249  4.225651  4.715663  3.442137  3.809220  3.720148  6.672896
#> [309]  6.217857  2.806569  4.668424  5.385990  3.810821  5.719303  4.705897
#> [316]  2.424984  5.547729  4.825235  4.005332  3.812841  3.415342  2.049526
#> [323]  5.345321  4.723653  2.655829  3.343876  4.793007  4.416392  3.387659
#> [330]  2.499560  2.437860  2.962475  2.668007  2.325961  3.321787  2.503546
#> [337]  2.289002  4.352134  2.563605  3.399213  3.172794  4.004418  4.337788
#> [344]  2.357048  2.782079  2.876256  1.519006  4.110565  3.759977  3.765079
#> [351]  1.949119  1.946511  3.934913  3.175746  3.407469  2.962698  2.355324
#> [358]  2.142521  2.419272  3.886441  4.105734  1.503987  4.203059  3.049543
#> [365]  1.726367  3.230712  3.110289  3.966258  1.958410  2.774166  2.822609
#> [372]  4.121998  1.840059  2.951120  3.208805  2.297657  2.046024  4.517620
#> [379]  4.848027  3.968016  2.185826  4.476663  3.279971  4.647565  3.501787
#> [386]  4.782940  2.721251  3.230761  2.294403  3.417405  3.438153  2.351458
#> [393]  5.703487  3.581410  5.283797  5.969450  3.159221  5.861914  5.424516
#> [400]  2.974851  5.871105  4.426526  4.760969  3.745659  3.766448  3.309613
#> [407]  4.792017  3.440552  6.267766  5.993912  4.141054  5.080168  7.511983
#> [414]  4.533612  7.206602  7.908292  6.760793  6.197943  4.381430  8.662746
#> [421]  7.391893  6.753090  8.464060  6.140890  9.166201  9.559050  6.006050
#> [428]  5.014466  6.362909  8.566745  9.431841  6.401311  6.316070  5.721783
#> [435]  8.647017  7.144258  7.195950  6.773036 10.317239  8.901508  6.486238
#> [442]  7.994296  8.938130  7.461917  7.754794 11.731718 10.830151  7.175970
#> [449]  8.731612 12.893474  6.872511  9.796804 10.239332 13.427412 12.249403
#> [456]  9.907493 11.770472 13.123331 11.209306 10.740421 10.006010 13.730064
#> [463] 13.343748 14.116675  9.745595  8.187391 12.579255 14.171207 11.838973
#> [470] 13.008935 14.029697 10.596894 10.762475 11.384228  9.322708 14.708905
#> [477] 11.923657 15.458795 14.942371 15.995086 13.037616  9.129182 16.589181
#> [484] 14.525398 14.976213 15.708068 13.176380 12.122945 15.751663  9.413289
#> [491]  9.398419 14.892835  9.515358 16.832942 15.082143 11.117489 15.625402
#> [498] 15.917348 12.764109 15.670778 13.015553  9.925738 14.147872 13.694284
#> [505] 18.236260 11.452676 15.151826 15.866266 16.019692 10.235869 18.375631
#> [512] 11.171157 17.490180 14.736091 15.607704 15.662691 16.030606 11.294397
#> [519] 11.504623 16.131736 11.705676 18.319865 11.294477 14.559430 16.300930
#> [526] 12.077850 11.763041 14.922982 17.718836 15.908140 17.980019 17.987988
#> [533] 18.546784 15.029106 13.850958 13.508163 18.088881 15.590679 17.121397
#> [540] 15.940580 18.366937 17.308475 18.199546 17.935220 17.950364 17.177804
#> [547] 13.473610 17.414109 18.118601 17.665305 11.838363 13.139711 13.942087
#> [554] 17.776858 14.860952 15.254394 11.882689 18.392832 17.084194 12.559342
#> [561] 11.389272 14.500957 14.203032 16.126003 14.025180 18.382909 10.483167
#> [568] 10.010072 12.473332 18.155006 15.733777 14.120553 13.523461 12.737321
#> [575] 13.638974 10.825494 11.410562 14.176195 15.228004 14.985213 14.380866
#> [582] 12.462397 11.978274 16.913024 13.404252 15.895878  8.798605 14.261656
#> [589] 11.803530 16.331666  9.063051  9.547315 12.842188  8.274750 10.803270
#> [596]  8.146016  9.138498 11.894970  7.942495 12.001607 10.392129 12.652054
#> [603] 13.234728 12.304728 14.023956  8.336492 14.868048  9.790945  8.455680
#> [610] 13.905836 12.982823 12.859969  8.810573  8.447053  9.219475  7.045360
#> [617] 13.530305  9.746582  8.612813 10.292713  7.774144  6.166407  8.119020
#> [624] 10.688542 11.262188  6.866968 11.549314  7.809431  8.645601  9.211405
#> [631]  7.669108  8.223861  7.473650 10.814260  8.114709  7.446926 10.896466
#> [638]  7.185164 10.619322  9.850314  6.909465  6.503846  9.753144  7.802168
#> [645]  4.480319  7.193782  9.022246  6.822572  6.973913  5.053357  9.013984
#> [652]  6.197110  5.164396  3.887635  4.378655  5.237929  5.471815  3.525583
#> [659]  5.151607  6.347854  4.190182  5.206192  6.033909  6.912734  3.316158
#> [666]  3.461629  5.440249  4.225651  4.715663  3.442137  3.809220  3.720148
#> [673]  6.672896  6.217857  2.806569  4.668424  5.385990  3.810821  5.719303
#> [680]  4.705897  2.424984  5.547729  4.825235  4.005332  3.812841  3.415342
#> [687]  2.049526  5.345321  4.723653  2.655829  3.343876  4.793007  4.416392
#> [694]  3.387659  2.499560  2.437860  2.962475  2.668007  2.325961  3.321787
#> [701]  2.503546  2.289002  4.352134  2.563605  3.399213  3.172794  4.004418
#> [708]  4.337788  2.357048  2.782079  2.876256  1.519006  4.110565  3.759977
#> [715]  3.765079  1.949119  1.946511  3.934913  3.175746  3.407469  2.962698
#> [722]  2.355324  2.142521  2.419272  3.886441  4.105734  1.503987  4.203059
#> [729]  3.049543  1.726367
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 5.10217027  7.463176  6.099033  3.609548
#> 3     2 2002-01-02 3.45027446  7.483435  6.149345  2.954929
#> 6     3 2002-01-03 4.15491456  7.505323  6.203934  3.268219
#> 7     4 2002-01-04 6.06466199  7.528820  6.262814  4.088134
#> 9     5 2002-01-05 2.42542789  7.553900  6.325996  2.597084
#> 11    6 2002-01-06 0.02612027  7.580540  6.393496  1.609389
#> 14    7 2002-01-07 6.51007142  7.608714  6.465324  4.382220
#> 15    8 2002-01-08 6.48282713  7.638396  6.541496  4.411309
#> 17    9 2002-01-09 0.42030169  7.669557  6.622023  1.836953
#> 19   10 2002-01-10 0.24669275  7.702170  6.706920  1.784138
#> 22   11 2002-01-11 4.39958190  7.736204  6.796198  3.631550
#> 23   12 2002-01-12 1.39940451  7.771631  6.889871  2.342783
#> 25   13 2002-01-13 2.90515629  7.808419  6.987950  3.046936
#> 27   14 2002-01-14 0.93101837  7.846537  7.090446  2.193264
#> 30   15 2002-01-15 4.96966653  7.885954  7.197371  4.067206
#> 32   16 2002-01-16 6.13308031  7.926639  7.308732  4.654676
#> 33   17 2002-01-17 2.89273712  7.968559  7.424539  3.203759
#> 35   18 2002-01-18 2.25037225  8.011681  7.544799  2.945816
#> 37   19 2002-01-19 2.49524570  8.055974  7.669519  3.105153
#> 40   20 2002-01-20 1.71221896  8.101405  7.798703  2.773797
#> 41   21 2002-01-21 6.34309426  8.147942  7.932354  5.070719
#> 43   22 2002-01-22 6.88563195  8.195553  8.070474  5.407891
#> 45   23 2002-01-23 0.43539256  8.244205  8.213063  2.270140
#> 48   24 2002-01-24 6.11032768  8.293867  8.360119  5.169599
#> 49   25 2002-01-25 5.93455046  8.344507  8.511639  5.154617
#> 51   26 2002-01-26 5.87958456  8.396093  8.667618  5.201768
#> 53   27 2002-01-27 1.24220447  8.448595  8.828046  2.856009
#> 56   28 2002-01-28 6.22074742  8.501983  8.992915  5.538206
#> 58   29 2002-01-29 1.76242762  8.556225  9.162213  3.234178
#> 59   30 2002-01-30 2.52242096  8.611293  9.335925  3.701322
#> 61   31 2002-01-31 2.60602755  8.667157  9.514035  3.808842
#> 63   32 2002-02-01 6.78186656  8.723789  9.696523  6.193166
#> 66   33 2002-02-02 3.63104149  8.781160  9.883368  4.514246
#> 67   34 2002-02-03 6.93932164  8.839243 10.074546  6.473190
#> 69   35 2002-02-04 5.09395689  8.898011 10.270030  5.507214
#> 71   36 2002-02-05 5.78496199  8.957438 10.469790  5.998288
#> 74   37 2002-02-06 0.04580369  9.017498 10.673795  2.695557
#> 75   38 2002-02-07 1.20414155  9.078165 10.882010  3.442206
#> 77   39 2002-02-08 3.38290887  9.139416 11.094398  4.826867
#> 79   40 2002-02-09 6.42584974  9.201226 11.310918  6.777326
#> 82   41 2002-02-10 2.80506607  9.263572 11.531528  4.628790
#> 83   42 2002-02-11 6.66239141  9.326432 11.756182  7.138094
#> 86   43 2002-02-12 5.91099148  9.389784 11.984831  6.768512
#> 87   44 2002-02-13 4.91937495  9.453606 12.217425  6.233148
#> 89   45 2002-02-14 4.78346847  9.517878 12.453910  6.243003
#> 91   46 2002-02-15 1.85554466  9.582579 12.694229  4.402595
#> 94   47 2002-02-16 0.41633158  9.647690 12.938322  3.513747
#> 95   48 2002-02-17 1.67487729  9.713192 13.186128  4.433396
#> 97   49 2002-02-18 1.65748448  9.779067 13.437583  4.498185
#> 99   50 2002-02-19 0.77148226  9.845297 13.692618  3.959635
#> 102  51 2002-02-20 0.25151837  9.911864 13.951164  3.664800
#> 103  52 2002-02-21 3.44433108  9.978754 14.213148  6.006238
#> 105  53 2002-02-22 0.96272345 10.045948 14.478496  4.313376
#> 107  54 2002-02-23 0.98427605 10.113433 14.747130  4.404405
#> 110  55 2002-02-24 6.40554160 10.181192 15.018971  8.479368
#> 112  56 2002-02-25 1.96797835 10.249211 15.293936  5.291799
#> 113  57 2002-02-26 6.36657399 10.317477 15.571940  8.697450
#> 115  58 2002-02-27 6.19035014 10.385975 15.852898  8.687624
#> 117  59 2002-02-28 6.11837561 10.454693 16.136720  8.756007
#> 120  60 2002-03-01 5.07336054 10.523618 16.423316  8.064610
#> 121  61 2002-03-02 6.05286607 10.592738 16.712592  8.953074
#> 123  62 2002-03-03 2.40698421 10.662040 17.004454  6.170514
#> 125  63 2002-03-04 3.71789606 10.731513 17.298805  7.321257
#> 128  64 2002-03-05 4.71034577 10.801147 17.595546  8.235567
#> 129  65 2002-03-06 6.47631175 10.870930 17.894578  9.803955
#> 131  66 2002-03-07 2.59602666 10.940852 18.195799  6.707684
#> 133  67 2002-03-08 5.56772367 11.010902 18.499104  9.301863
#> 136  68 2002-03-09 3.35337660 11.081072 18.804391  7.546409
#> 138  69 2002-03-10 0.70129788 11.151350 19.111552  5.378842
#> 139  70 2002-03-11 0.58515753 11.221729 19.420480  5.361461
#> 141  71 2002-03-12 2.05172824 11.292198 19.731067  6.725278
#> 143  72 2002-03-13 4.28073948 11.362749 20.043203  8.786284
#> 146  73 2002-03-14 2.68647119 11.433374 20.356778  7.480785
#> 147  74 2002-03-15 4.59842276 11.504062 20.671679  9.299379
#> 149  75 2002-03-16 2.73169685 11.574807 20.987796  7.723547
#> 151  76 2002-03-17 0.25752421 11.645600 21.305014  5.561817
#> 154  77 2002-03-18 2.33345293 11.716433 21.623220  7.559053
#> 155  78 2002-03-19 1.08108723 11.787297 21.942300  6.491808
#> 157  79 2002-03-20 5.36226116 11.858185 22.262140 10.598995
#> 159  80 2002-03-21 1.48460839 11.929088 22.582624  7.050891
#> 162  81 2002-03-22 2.07322733 12.000000 22.903638  7.704428
#> 163  82 2002-03-23 5.54703638 12.070912 23.225065 11.142660
#> 166  83 2002-03-24 0.72358531 12.141815 23.546791  6.588327
#> 167  84 2002-03-25 4.49662421 12.212703 23.868700 10.361312
#> 169  85 2002-03-26 6.69592080 12.283567 24.190677 12.640984
#> 171  86 2002-03-27 4.03653736 12.354400 24.512606 10.132638
#> 174  87 2002-03-28 3.76571012 12.425193 24.834373  9.971877
#> 175  88 2002-03-29 0.54249069 12.495938 25.155863  6.835016
#> 177  89 2002-03-30 2.68192039 12.566626 25.476961  9.087837
#> 179  90 2002-03-31 5.58961055 12.637251 25.797553 12.154675
#> 182  91 2002-04-01 1.48963468 12.707802 26.117528  8.060157
#> 183  92 2002-04-02 4.11707073 12.778271 26.436771 10.868065
#> 185  93 2002-04-03 2.09318538 12.848650 26.755170  8.868147
#> 187  94 2002-04-04 3.99104740 12.918928 27.072615 10.949928
#> 190  95 2002-04-05 5.89471543 12.989098 27.388994 13.062089
#> 192  96 2002-04-06 5.46211195 13.059148 27.704199 12.719821
#> 193  97 2002-04-07 0.05107026 13.129070 28.018120  7.059023
#> 195  98 2002-04-08 2.22981682 13.198853 28.330649  9.475755
#> 197  99 2002-04-09 4.49188021 13.268487 28.641680 12.008561
#> 200 100 2002-04-10 6.03689657 13.337960 28.951106 13.789559
#> 201 101 2002-04-11 3.32655914 13.407262 29.258822 10.944500
#> 203 102 2002-04-12 4.80131553 13.476382 29.564726 12.657787
#> 205 103 2002-04-13 6.21147305 13.545307 29.868714 14.315629
#> 208 104 2002-04-14 1.38907164 13.614025 30.170685  9.081865
#> 209 105 2002-04-15 1.15525168 13.682523 30.470539  8.903989
#> 211 106 2002-04-16 6.68925823 13.750789 30.768176 15.175843
#> 213 107 2002-04-17 3.59652558 13.818808 31.063500 11.808216
#> 216 108 2002-04-18 1.51618105 13.886567 31.356413  9.550901
#> 218 109 2002-04-19 5.07812675 13.954052 31.646820 13.670124
#> 219 110 2002-04-20 1.08990185 14.021246 31.934628  9.224831
#> 221 111 2002-04-21 3.20458761 14.088136 32.219743 11.719402
#> 223 112 2002-04-22 0.68357804 14.154703 32.502075  8.910336
#> 226 113 2002-04-23 4.11705967 14.220933 32.781534 12.940625
#> 227 114 2002-04-24 4.45395069 14.286808 33.058030 13.417472
#> 229 115 2002-04-25 4.91138004 14.352310 33.331477 14.035908
#> 231 116 2002-04-26 1.80080724 14.417421 33.601789 10.498962
#> 234 117 2002-04-27 2.20252989 14.482122 33.868881 11.042714
#> 235 118 2002-04-28 3.04203136 14.546394 34.132670 12.102184
#> 237 119 2002-04-29 2.39492133 14.610216 34.393074 11.417142
#> 239 120 2002-04-30 3.37252518 14.673568 34.650012 12.644426
#> 242 121 2002-05-01 0.11503112 14.736428 34.903405  8.862078
#> 243 122 2002-05-02 6.58495661 14.798774 35.153175 16.609284
#> 245 123 2002-05-03 6.42196418 14.860584 35.399245 16.498659
#> 248 124 2002-05-04 2.51681932 14.921835 35.641540 11.916159
#> 249 125 2002-05-05 6.43675517 14.982502 35.879987 16.677344
#> 251 126 2002-05-06 0.41004192 15.042562 36.114511  9.520846
#> 253 127 2002-05-07 1.79958922 15.101989 36.345041 11.251742
#> 256 128 2002-05-08 1.73978576 15.160757 36.571508 11.241275
#> 257 129 2002-05-09 0.18856734 15.218840 36.793843  9.426406
#> 259 130 2002-05-10 3.70195164 15.276211 37.011977 13.737632
#> 261 131 2002-05-11 6.77243852 15.332843 37.225844 17.527694
#> 264 132 2002-05-12 4.54640406 15.388707 37.435378 14.888755
#> 265 133 2002-05-13 4.54998340 15.443775 37.640517 14.954879
#> 267 134 2002-05-14 6.93243802 15.498017 37.841197 17.923697
#> 269 135 2002-05-15 3.64680822 15.551405 38.037357 13.969224
#> 272 136 2002-05-16 4.95367117 15.603907 38.228936 15.625380
#> 274 137 2002-05-17 5.96266381 15.655493 38.415875 16.919642
#> 275 138 2002-05-18 1.10910662 15.706133 38.598118 11.012355
#> 277 139 2002-05-19 0.35962883 15.755795 38.775606 10.136432
#> 279 140 2002-05-20 1.72686232 15.804447 38.948286 11.864901
#> 282 141 2002-05-21 5.56287812 15.852058 39.116103 16.642429
#> 283 142 2002-05-22 2.61794558 15.898595 39.279005 13.053694
#> 285 143 2002-05-23 3.36434127 15.944026 39.436940 14.020020
#> 287 144 2002-05-24 2.25065942 15.988319 39.589859 12.683977
#> 290 145 2002-05-25 3.32228224 16.031441 39.737713 14.051959
#> 291 146 2002-05-26 2.54460138 16.073361 39.880455 13.126885
#> 293 147 2002-05-27 1.00246456 16.114046 40.018039 11.249283
#> 295 148 2002-05-28 6.03223491 16.153463 40.150421 17.534350
#> 298 149 2002-05-29 6.40962083 16.191581 40.277558 18.041553
#> 300 150 2002-05-30 4.35361678 16.228369 40.399408 15.518867
#> 301 151 2002-05-31 4.22685782 16.263796 40.515933 15.393901
#> 303 152 2002-06-01 4.72555027 16.297830 40.627093 16.046679
#> 305 153 2002-06-02 2.14028017 16.330443 40.732853 12.852452
#> 308 154 2002-06-03 5.33496272 16.361604 40.833176 16.865449
#> 309 155 2002-06-04 5.62081865 16.391286 40.928031 17.249427
#> 311 156 2002-06-05 5.10435515 16.419460 41.017385 16.629931
#> 313 157 2002-06-06 5.96587817 16.446100 41.101209 17.730103
#> 316 158 2002-06-07 4.69637802 16.471180 41.179475 16.165559
#> 317 159 2002-06-08 2.02322923 16.494677 41.252156 12.843024
#> 319 160 2002-06-09 0.11025299 16.516565 41.319229 10.467716
#> 321 161 2002-06-10 6.41900065 16.536824 41.380671 18.376412
#> 324 162 2002-06-11 5.01626571 16.555433 41.436461 16.636701
#> 325 163 2002-06-12 6.68953328 16.572373 41.486582 18.744794
#> 328 164 2002-06-13 4.39231515 16.587625 41.531016 15.881350
#> 329 165 2002-06-14 1.78381510 16.601175 41.569750 12.625796
#> 331 166 2002-06-15 6.13048633 16.613007 41.602770 18.076763
#> 333 167 2002-06-16 1.01716533 16.623108 41.630068 11.681186
#> 336 168 2002-06-17 6.04868781 16.631469 41.651634 17.987036
#> 337 169 2002-06-18 6.58762088 16.638080 41.667464 18.665698
#> 339 170 2002-06-19 0.37761942 16.642934 41.677552 10.892209
#> 341 171 2002-06-20 3.38806792 16.646025 41.681897 14.662361
#> 344 172 2002-06-21 0.43336073 16.647351 41.680500 10.962634
#> 345 173 2002-06-22 3.05883568 16.646909 41.673363 14.247039
#> 347 174 2002-06-23 3.52857549 16.644700 41.660491 14.831008
#> 349 175 2002-06-24 0.51427259 16.640727 41.641891 11.053933
#> 352 176 2002-06-25 4.68680536 16.634994 41.617572 16.267138
#> 354 177 2002-06-26 6.04328995 16.627507 41.587545 17.954411
#> 355 178 2002-06-27 4.54505397 16.618275 41.551823 16.070113
#> 357 179 2002-06-28 6.38959719 16.607306 41.510423 18.363093
#> 359 180 2002-06-29 2.04245638 16.594614 41.463360 12.917484
#> 362 181 2002-06-30 1.33046908 16.580211 41.410655 12.014150
#> 363 182 2002-07-01 3.87611102 16.564113 41.352328 15.176440
#> 365 183 2002-07-02 6.22326353 16.546336 41.288405 18.086618
#> 367 184 2002-07-03 0.75322704 16.526900 41.218909 11.244020
#> 370 185 2002-07-04 0.56969058 16.505823 41.143868 10.995997
#> 371 186 2002-07-05 3.52713596 16.483128 41.063311 14.659286
#> 373 187 2002-07-06 1.47341591 16.458837 40.977269 12.078486
#> 375 188 2002-07-07 6.14271704 16.432973 40.885775 17.863084
#> 378 189 2002-07-08 4.17063574 16.405563 40.788864 15.381905
#> 380 190 2002-07-09 0.15936163 16.376632 40.686570 10.369604
#> 381 191 2002-07-10 2.85534081 16.346207 40.578932 13.688879
#> 383 192 2002-07-11 1.07203324 16.314316 40.465990 11.446032
#> 385 193 2002-07-12 1.15340436 16.280989 40.347784 11.516138
#> 388 194 2002-07-13 6.96863284 16.246255 40.224356 18.682963
#> 389 195 2002-07-14 5.16951551 16.210144 40.095751 16.417330
#> 391 196 2002-07-15 4.09223210 16.172686 39.962013 15.046368
#> 393 197 2002-07-16 6.30614373 16.133915 39.823189 17.738495
#> 396 198 2002-07-17 1.42050950 16.093860 39.679327 11.670961
#> 397 199 2002-07-18 0.83255839 16.052554 39.530476 10.907734
#> 399 200 2002-07-19 1.70251221 16.010029 39.376687 11.937837
#> 401 201 2002-07-20 5.83575362 15.966317 39.218012 16.971674
#> 404 202 2002-07-21 6.27028202 15.921451 39.054503 17.453966
#> 405 203 2002-07-22 2.99178301 15.875462 38.886214 13.385671
#> 408 204 2002-07-23 0.93036306 15.828384 38.713200 10.816045
#> 409 205 2002-07-24 4.55958529 15.780249 38.535519 15.201155
#> 411 206 2002-07-25 4.49048584 15.731088 38.353226 15.062328
#> 413 207 2002-07-26 0.97268542 15.680933 38.166382 10.725322
#> 416 208 2002-07-27 4.22811039 15.629816 37.975044 14.630183
#> 417 209 2002-07-28 3.82715638 15.577768 37.779274 14.085637
#> 419 210 2002-07-29 6.04532332 15.524820 37.579134 16.711391
#> 421 211 2002-07-30 6.69304481 15.471001 37.374685 17.428166
#> 424 212 2002-07-31 5.05508500 15.416342 37.165992 15.384942
#> 425 213 2002-08-01 0.28337002 15.360873 36.953119  9.579127
#> 427 214 2002-08-02 4.81654769 15.304621 36.736132 14.964683
#> 429 215 2002-08-03 1.73288233 15.247616 36.515097 11.203734
#> 432 216 2002-08-04 3.24960618 15.189886 36.290082 12.954330
#> 434 217 2002-08-05 4.15592298 15.131457 36.061156 13.967469
#> 435 218 2002-08-06 1.92646694 15.072356 35.828388 11.246792
#> 437 219 2002-08-07 1.44278112 15.012610 35.591848 10.608233
#> 439 220 2002-08-08 1.82616509 14.952243 35.351609 10.996705
#> 442 221 2002-08-09 2.13175297 14.891281 35.107744 11.289851
#> 443 222 2002-08-10 4.00121542 14.829748 34.860325 13.417915
#> 445 223 2002-08-11 1.33762017 14.767666 34.609429 10.219777
#> 447 224 2002-08-12 3.49065697 14.705060 34.355129 12.666357
#> 450 225 2002-08-13 4.13386325 14.641952 34.097505 13.337752
#> 451 226 2002-08-14 6.44574196 14.578362 33.836632 15.939498
#> 453 227 2002-08-15 4.36397636 14.514313 33.572591 13.440235
#> 455 228 2002-08-16 3.70096714 14.449824 33.305462 12.591553
#> 458 229 2002-08-17 6.18436783 14.384916 33.035326 15.360110
#> 460 230 2002-08-18 6.47548066 14.319607 32.762264 15.598291
#> 461 231 2002-08-19 6.98228860 14.253916 32.486361 16.078321
#> 463 232 2002-08-20 1.60184320 14.187862 32.207701  9.870088
#> 465 233 2002-08-21 0.69529107 14.121461 31.926370  8.767564
#> 468 234 2002-08-22 4.19795456 14.054730 31.642455 12.636197
#> 469 235 2002-08-23 0.73361197 13.987686 31.356043  8.661276
#> 471 236 2002-08-24 5.60924993 13.920345 31.067223 14.026127
#> 473 237 2002-08-25 5.82823833 13.852721 30.776085 14.168213
#> 476 238 2002-08-26 1.00361260 13.784830 30.482721  8.730336
#> 477 239 2002-08-27 3.38400483 13.716686 30.187223 11.270508
#> 479 240 2002-08-28 0.41593608 13.648302 29.889684  7.927870
#> 481 241 2002-08-29 3.96119486 13.579692 29.590199 11.713279
#> 484 242 2002-08-30 2.58947696 13.510869 29.288864 10.128950
#> 485 243 2002-08-31 4.39341146 13.441846 28.985774 11.983383
#> 487 244 2002-09-01 5.56118110 13.372633 28.681027 13.133943
#> 490 245 2002-09-02 4.33519049 13.303244 28.374722 11.716982
#> 491 246 2002-09-03 3.39281605 13.233689 28.066958 10.614603
#> 493 247 2002-09-04 1.36361180 13.163980 27.757836  8.377129
#> 495 248 2002-09-05 1.05783363 13.094126 27.447457  7.970562
#> 498 249 2002-09-06 6.52111510 13.024138 27.135924 13.577385
#> 499 250 2002-09-07 1.74585892 12.954027 26.823338  8.513371
#> 501 251 2002-09-08 0.55859460 12.883802 26.509805  7.202135
#> 503 252 2002-09-09 0.57490573 12.813472 26.195428  7.136516
#> 506 253 2002-09-10 3.18835958 12.743047 25.880314  9.707756
#> 507 254 2002-09-11 6.64376555 12.672536 25.564567 13.092444
#> 509 255 2002-09-12 6.96517497 12.601947 25.248295 13.289519
#> 511 256 2002-09-13 1.85144599 12.531289 24.931604  8.074671
#> 514 257 2002-09-14 2.35409171 12.460572 24.614602  8.478786
#> 516 258 2002-09-15 6.54390308 12.389802 24.297397 12.490909
#> 517 259 2002-09-16 6.53800931 12.318988 23.980097 12.358457
#> 519 260 2002-09-17 3.17042118 12.248139 23.662812  8.978253
#> 521 261 2002-09-18 5.08041145 12.177262 23.345649 10.706370
#> 524 262 2002-09-19 4.80427066 12.106365 23.028718 10.326520
#> 525 263 2002-09-20 2.23556512 12.035456 22.712128  7.787401
#> 527 264 2002-09-21 0.32402803 11.964544 22.395989  5.902265
#> 529 265 2002-09-22 0.38996585 11.893635 22.080409  5.882086
#> 532 266 2002-09-23 3.27713020 11.822738 21.765497  8.457950
#> 533 267 2002-09-24 4.86748608 11.751861 21.451363  9.805295
#> 535 268 2002-09-25 0.05598748 11.681012 21.138114  5.335186
#> 537 269 2002-09-26 5.29487441 11.610198 20.825859  9.955320
#> 540 270 2002-09-27 0.45101929 11.539428 20.514705  5.529586
#> 542 271 2002-09-28 3.93141632 11.468711 20.204761  8.514235
#> 543 272 2002-09-29 2.77610080 11.398053 19.896131  7.396976
#> 545 273 2002-09-30 0.38281455 11.327464 19.588923  5.228237
#> 547 274 2002-10-01 6.36876946 11.256953 19.283241 10.275684
#> 550 275 2002-10-02 3.93448544 11.186528 18.979189  8.082443
#> 551 276 2002-10-03 0.46890529 11.116198 18.676871  5.063133
#> 553 277 2002-10-04 4.42075486 11.045973 18.376389  8.271343
#> 555 278 2002-10-05 5.12188737 10.975862 18.077844  8.737475
#> 558 279 2002-10-06 3.24968029 10.905874 17.781337  7.094533
#> 559 280 2002-10-07 2.19371176 10.836020 17.486965  6.141826
#> 561 281 2002-10-08 1.17820812 10.766311 17.194826  5.239562
#> 563 282 2002-10-09 3.22854922 10.696756 16.905016  6.777433
#> 566 283 2002-10-10 5.87086374 10.627367 16.617629  8.744436
#> 567 284 2002-10-11 0.61009091 10.558154 16.332759  4.555075
#> 570 285 2002-10-12 6.86978860 10.489131 16.050495  9.268708
#> 571 286 2002-10-13 4.10356093 10.420308 15.770928  7.048061
#> 573 287 2002-10-14 1.97128878 10.351698 15.494145  5.348822
#> 575 288 2002-10-15 3.47086799 10.283314 15.220231  6.373656
#> 578 289 2002-10-16 2.45724855 10.215170 14.949271  5.535334
#> 579 290 2002-10-17 5.55470648 10.147279 14.681345  7.688683
#> 581 291 2002-10-18 3.37073869 10.079655 14.416534  6.014651
#> 583 292 2002-10-19 5.60217709 10.012314 14.154915  7.498769
#> 586 293 2002-10-20 6.80943783  9.945270 13.896562  8.231567
#> 587 294 2002-10-21 5.62120973  9.878539 13.641550  7.291630
#> 589 295 2002-10-22 6.91617253  9.812138 13.389948  8.066499
#> 591 296 2002-10-23 2.00025269  9.746084 13.141825  4.634048
#> 594 297 2002-10-24 5.60391534  9.680393 12.897247  6.957377
#> 596 298 2002-10-25 1.92213631  9.615084 12.656277  4.429117
#> 597 299 2002-10-26 2.90527923  9.550176 12.418977  4.993746
#> 599 300 2002-10-27 6.54111775  9.485687 12.185404  7.247742
#> 601 301 2002-10-28 4.65099424  9.421638 11.955615  5.939850
#> 604 302 2002-10-29 3.43738717  9.358048 11.729663  5.086679
#> 605 303 2002-10-30 6.66451982  9.294940 11.507600  7.002404
#> 607 304 2002-10-31 1.39316484  9.232334 11.289473  3.674162
#> 609 305 2002-11-01 6.94406980  9.170252 11.075328  6.962166
#> 612 306 2002-11-02 2.71477602  9.108719 10.865209  4.335444
#> 613 307 2002-11-03 6.46813658  9.047757 10.659156  6.474843
#> 615 308 2002-11-04 1.30542218  8.987390 10.457208  3.373759
#> 617 309 2002-11-05 0.56310686  8.927644 10.259398  2.888403
#> 620 310 2002-11-06 2.78521999  8.868543 10.065762  4.097047
#> 622 311 2002-11-07 1.02815793  8.810114  9.876328  3.045376
#> 623 312 2002-11-08 2.30231613  8.752384  9.691126  3.697408
#> 625 313 2002-11-09 4.44368000  8.695379  9.510180  4.807583
#> 627 314 2002-11-10 3.05400254  8.639127  9.333513  3.983115
#> 630 315 2002-11-11 5.76599476  8.583658  9.161147  5.367246
#> 631 316 2002-11-12 6.84060026  8.528999  8.993100  5.854689
#> 633 317 2002-11-13 0.42610578  8.475180  8.829388  2.429304
#> 635 318 2002-11-14 1.64656439  8.422232  8.670025  3.015011
#> 638 319 2002-11-15 0.70762396  8.370184  8.515023  2.488690
#> 639 320 2002-11-16 1.03848084  8.319067  8.364392  2.613167
#> 641 321 2002-11-17 6.97352652  8.268912  8.218139  5.519889
#> 643 322 2002-11-18 6.59192371  8.219751  8.076271  5.257497
#> 646 323 2002-11-19 4.49622950  8.171616  7.938791  4.168759
#> 647 324 2002-11-20 2.62408554  8.124538  7.805702  3.211979
#> 650 325 2002-11-21 5.97819669  8.078549  7.677004  4.759776
#> 651 326 2002-11-22 5.68928762  8.033683  7.552698  4.562506
#> 653 327 2002-11-23 5.20680562  7.989971  7.432780  4.280046
#> 655 328 2002-11-24 3.14226558  7.947446  7.317248  3.275861
#> 658 329 2002-11-25 1.04987889  7.906140  7.206095  2.279983
#> 659 330 2002-11-26 0.59498777  7.866085  7.099317  2.043324
#> 661 331 2002-11-27 1.77642160  7.827314  6.996907  2.543207
#> 663 332 2002-11-28 1.14641729  7.789856  6.898857  2.232359
#> 666 333 2002-11-29 3.24153391  7.753745  6.805157  3.123773
#> 667 334 2002-11-30 6.91578687  7.719011  6.715800  4.687434
#> 669 335 2002-12-01 2.02849782  7.685684  6.630775  2.532730
#> 671 336 2002-12-02 3.13731872  7.653793  6.550072  2.979967
#> 674 337 2002-12-03 6.33066697  7.623368  6.473680  4.306386
#> 676 338 2002-12-04 6.11362792  7.594437  6.401590  4.177082
#> 677 339 2002-12-05 1.85786245  7.567027  6.333789  2.360986
#> 679 340 2002-12-06 3.73930495  7.541163  6.270268  3.122131
#> 681 341 2002-12-07 2.15260830  7.516872  6.211016  2.442079
#> 684 342 2002-12-08 2.57621274  7.494177  6.156023  2.597109
#> 685 343 2002-12-09 1.92278039  7.473100  6.105278  2.311744
#> 687 344 2002-12-10 5.23933262  7.453664  6.058772  3.644111
#> 689 345 2002-12-11 0.60412384  7.435887  6.016497  1.748527
#> 692 346 2002-12-12 5.43044043  7.419789  5.978443  3.682380
#> 693 347 2002-12-13 1.14125806  7.405386  5.944603  1.944218
#> 695 348 2002-12-14 6.21047829  7.392694  5.914970  3.963276
#> 697 349 2002-12-15 1.74296967  7.381725  5.889539  2.167702
#> 700 350 2002-12-16 3.79491100  7.372493  5.868303  2.977399
#> 702 351 2002-12-17 0.84273628  7.365006  5.851259  1.797578
#> 703 352 2002-12-18 4.06568922  7.359273  5.838404  3.072338
#> 705 353 2002-12-19 6.41460395  7.355300  5.829736  3.999509
#> 707 354 2002-12-20 2.42253244  7.353091  5.825253  2.415901
#> 710 355 2002-12-21 3.92450491  7.352649  5.824956  3.010785
#> 711 356 2002-12-22 0.08692587  7.353975  5.828845  1.491661
#> 713 357 2002-12-23 3.48792940  7.357066  5.836924  2.842852
#> 715 358 2002-12-24 3.25348978  7.361920  5.849194  2.754780
#> 718 359 2002-12-25 5.94819254  7.368531  5.865662  3.833922
#> 719 360 2002-12-26 2.48395249  7.376892  5.886331  2.462608
#> 721 361 2002-12-27 1.97288956  7.386993  5.911209  2.267174
#> 723 362 2002-12-28 0.07843627  7.398825  5.940303  1.516563
#> 726 363 2002-12-29 1.03003390  7.412375  5.973622  1.908457
#> 727 364 2002-12-30 4.91110641  7.427627  6.011174  3.490071
#> 729 365 2002-12-31 1.96097878  7.444567  6.052971  2.310451
#> 2     1 2003-01-01 5.10217027  7.463176  6.099033  3.609548
#> 4     2 2003-01-02 3.45027446  7.483435  6.149345  2.954929
#> 5     3 2003-01-03 4.15491456  7.505323  6.203934  3.268219
#> 8     4 2003-01-04 6.06466199  7.528820  6.262814  4.088134
#> 10    5 2003-01-05 2.42542789  7.553900  6.325996  2.597084
#> 12    6 2003-01-06 0.02612027  7.580540  6.393496  1.609389
#> 13    7 2003-01-07 6.51007142  7.608714  6.465324  4.382220
#> 16    8 2003-01-08 6.48282713  7.638396  6.541496  4.411309
#> 18    9 2003-01-09 0.42030169  7.669557  6.622023  1.836953
#> 20   10 2003-01-10 0.24669275  7.702170  6.706920  1.784138
#> 21   11 2003-01-11 4.39958190  7.736204  6.796198  3.631550
#> 24   12 2003-01-12 1.39940451  7.771631  6.889871  2.342783
#> 26   13 2003-01-13 2.90515629  7.808419  6.987950  3.046936
#> 28   14 2003-01-14 0.93101837  7.846537  7.090446  2.193264
#> 29   15 2003-01-15 4.96966653  7.885954  7.197371  4.067206
#> 31   16 2003-01-16 6.13308031  7.926639  7.308732  4.654676
#> 34   17 2003-01-17 2.89273712  7.968559  7.424539  3.203759
#> 36   18 2003-01-18 2.25037225  8.011681  7.544799  2.945816
#> 38   19 2003-01-19 2.49524570  8.055974  7.669519  3.105153
#> 39   20 2003-01-20 1.71221896  8.101405  7.798703  2.773797
#> 42   21 2003-01-21 6.34309426  8.147942  7.932354  5.070719
#> 44   22 2003-01-22 6.88563195  8.195553  8.070474  5.407891
#> 46   23 2003-01-23 0.43539256  8.244205  8.213063  2.270140
#> 47   24 2003-01-24 6.11032768  8.293867  8.360119  5.169599
#> 50   25 2003-01-25 5.93455046  8.344507  8.511639  5.154617
#> 52   26 2003-01-26 5.87958456  8.396093  8.667618  5.201768
#> 54   27 2003-01-27 1.24220447  8.448595  8.828046  2.856009
#> 55   28 2003-01-28 6.22074742  8.501983  8.992915  5.538206
#> 57   29 2003-01-29 1.76242762  8.556225  9.162213  3.234178
#> 60   30 2003-01-30 2.52242096  8.611293  9.335925  3.701322
#> 62   31 2003-01-31 2.60602755  8.667157  9.514035  3.808842
#> 64   32 2003-02-01 6.78186656  8.723789  9.696523  6.193166
#> 65   33 2003-02-02 3.63104149  8.781160  9.883368  4.514246
#> 68   34 2003-02-03 6.93932164  8.839243 10.074546  6.473190
#> 70   35 2003-02-04 5.09395689  8.898011 10.270030  5.507214
#> 72   36 2003-02-05 5.78496199  8.957438 10.469790  5.998288
#> 73   37 2003-02-06 0.04580369  9.017498 10.673795  2.695557
#> 76   38 2003-02-07 1.20414155  9.078165 10.882010  3.442206
#> 78   39 2003-02-08 3.38290887  9.139416 11.094398  4.826867
#> 80   40 2003-02-09 6.42584974  9.201226 11.310918  6.777326
#> 81   41 2003-02-10 2.80506607  9.263572 11.531528  4.628790
#> 84   42 2003-02-11 6.66239141  9.326432 11.756182  7.138094
#> 85   43 2003-02-12 5.91099148  9.389784 11.984831  6.768512
#> 88   44 2003-02-13 4.91937495  9.453606 12.217425  6.233148
#> 90   45 2003-02-14 4.78346847  9.517878 12.453910  6.243003
#> 92   46 2003-02-15 1.85554466  9.582579 12.694229  4.402595
#> 93   47 2003-02-16 0.41633158  9.647690 12.938322  3.513747
#> 96   48 2003-02-17 1.67487729  9.713192 13.186128  4.433396
#> 98   49 2003-02-18 1.65748448  9.779067 13.437583  4.498185
#> 100  50 2003-02-19 0.77148226  9.845297 13.692618  3.959635
#> 101  51 2003-02-20 0.25151837  9.911864 13.951164  3.664800
#> 104  52 2003-02-21 3.44433108  9.978754 14.213148  6.006238
#> 106  53 2003-02-22 0.96272345 10.045948 14.478496  4.313376
#> 108  54 2003-02-23 0.98427605 10.113433 14.747130  4.404405
#> 109  55 2003-02-24 6.40554160 10.181192 15.018971  8.479368
#> 111  56 2003-02-25 1.96797835 10.249211 15.293936  5.291799
#> 114  57 2003-02-26 6.36657399 10.317477 15.571940  8.697450
#> 116  58 2003-02-27 6.19035014 10.385975 15.852898  8.687624
#> 118  59 2003-02-28 6.11837561 10.454693 16.136720  8.756007
#> 119  60 2003-03-01 5.07336054 10.523618 16.423316  8.064610
#> 122  61 2003-03-02 6.05286607 10.592738 16.712592  8.953074
#> 124  62 2003-03-03 2.40698421 10.662040 17.004454  6.170514
#> 126  63 2003-03-04 3.71789606 10.731513 17.298805  7.321257
#> 127  64 2003-03-05 4.71034577 10.801147 17.595546  8.235567
#> 130  65 2003-03-06 6.47631175 10.870930 17.894578  9.803955
#> 132  66 2003-03-07 2.59602666 10.940852 18.195799  6.707684
#> 134  67 2003-03-08 5.56772367 11.010902 18.499104  9.301863
#> 135  68 2003-03-09 3.35337660 11.081072 18.804391  7.546409
#> 137  69 2003-03-10 0.70129788 11.151350 19.111552  5.378842
#> 140  70 2003-03-11 0.58515753 11.221729 19.420480  5.361461
#> 142  71 2003-03-12 2.05172824 11.292198 19.731067  6.725278
#> 144  72 2003-03-13 4.28073948 11.362749 20.043203  8.786284
#> 145  73 2003-03-14 2.68647119 11.433374 20.356778  7.480785
#> 148  74 2003-03-15 4.59842276 11.504062 20.671679  9.299379
#> 150  75 2003-03-16 2.73169685 11.574807 20.987796  7.723547
#> 152  76 2003-03-17 0.25752421 11.645600 21.305014  5.561817
#> 153  77 2003-03-18 2.33345293 11.716433 21.623220  7.559053
#> 156  78 2003-03-19 1.08108723 11.787297 21.942300  6.491808
#> 158  79 2003-03-20 5.36226116 11.858185 22.262140 10.598995
#> 160  80 2003-03-21 1.48460839 11.929088 22.582624  7.050891
#> 161  81 2003-03-22 2.07322733 12.000000 22.903638  7.704428
#> 164  82 2003-03-23 5.54703638 12.070912 23.225065 11.142660
#> 165  83 2003-03-24 0.72358531 12.141815 23.546791  6.588327
#> 168  84 2003-03-25 4.49662421 12.212703 23.868700 10.361312
#> 170  85 2003-03-26 6.69592080 12.283567 24.190677 12.640984
#> 172  86 2003-03-27 4.03653736 12.354400 24.512606 10.132638
#> 173  87 2003-03-28 3.76571012 12.425193 24.834373  9.971877
#> 176  88 2003-03-29 0.54249069 12.495938 25.155863  6.835016
#> 178  89 2003-03-30 2.68192039 12.566626 25.476961  9.087837
#> 180  90 2003-03-31 5.58961055 12.637251 25.797553 12.154675
#> 181  91 2003-04-01 1.48963468 12.707802 26.117528  8.060157
#> 184  92 2003-04-02 4.11707073 12.778271 26.436771 10.868065
#> 186  93 2003-04-03 2.09318538 12.848650 26.755170  8.868147
#> 188  94 2003-04-04 3.99104740 12.918928 27.072615 10.949928
#> 189  95 2003-04-05 5.89471543 12.989098 27.388994 13.062089
#> 191  96 2003-04-06 5.46211195 13.059148 27.704199 12.719821
#> 194  97 2003-04-07 0.05107026 13.129070 28.018120  7.059023
#> 196  98 2003-04-08 2.22981682 13.198853 28.330649  9.475755
#> 198  99 2003-04-09 4.49188021 13.268487 28.641680 12.008561
#> 199 100 2003-04-10 6.03689657 13.337960 28.951106 13.789559
#> 202 101 2003-04-11 3.32655914 13.407262 29.258822 10.944500
#> 204 102 2003-04-12 4.80131553 13.476382 29.564726 12.657787
#> 206 103 2003-04-13 6.21147305 13.545307 29.868714 14.315629
#> 207 104 2003-04-14 1.38907164 13.614025 30.170685  9.081865
#> 210 105 2003-04-15 1.15525168 13.682523 30.470539  8.903989
#> 212 106 2003-04-16 6.68925823 13.750789 30.768176 15.175843
#> 214 107 2003-04-17 3.59652558 13.818808 31.063500 11.808216
#> 215 108 2003-04-18 1.51618105 13.886567 31.356413  9.550901
#> 217 109 2003-04-19 5.07812675 13.954052 31.646820 13.670124
#> 220 110 2003-04-20 1.08990185 14.021246 31.934628  9.224831
#> 222 111 2003-04-21 3.20458761 14.088136 32.219743 11.719402
#> 224 112 2003-04-22 0.68357804 14.154703 32.502075  8.910336
#> 225 113 2003-04-23 4.11705967 14.220933 32.781534 12.940625
#> 228 114 2003-04-24 4.45395069 14.286808 33.058030 13.417472
#> 230 115 2003-04-25 4.91138004 14.352310 33.331477 14.035908
#> 232 116 2003-04-26 1.80080724 14.417421 33.601789 10.498962
#> 233 117 2003-04-27 2.20252989 14.482122 33.868881 11.042714
#> 236 118 2003-04-28 3.04203136 14.546394 34.132670 12.102184
#> 238 119 2003-04-29 2.39492133 14.610216 34.393074 11.417142
#> 240 120 2003-04-30 3.37252518 14.673568 34.650012 12.644426
#> 241 121 2003-05-01 0.11503112 14.736428 34.903405  8.862078
#> 244 122 2003-05-02 6.58495661 14.798774 35.153175 16.609284
#> 246 123 2003-05-03 6.42196418 14.860584 35.399245 16.498659
#> 247 124 2003-05-04 2.51681932 14.921835 35.641540 11.916159
#> 250 125 2003-05-05 6.43675517 14.982502 35.879987 16.677344
#> 252 126 2003-05-06 0.41004192 15.042562 36.114511  9.520846
#> 254 127 2003-05-07 1.79958922 15.101989 36.345041 11.251742
#> 255 128 2003-05-08 1.73978576 15.160757 36.571508 11.241275
#> 258 129 2003-05-09 0.18856734 15.218840 36.793843  9.426406
#> 260 130 2003-05-10 3.70195164 15.276211 37.011977 13.737632
#> 262 131 2003-05-11 6.77243852 15.332843 37.225844 17.527694
#> 263 132 2003-05-12 4.54640406 15.388707 37.435378 14.888755
#> 266 133 2003-05-13 4.54998340 15.443775 37.640517 14.954879
#> 268 134 2003-05-14 6.93243802 15.498017 37.841197 17.923697
#> 270 135 2003-05-15 3.64680822 15.551405 38.037357 13.969224
#> 271 136 2003-05-16 4.95367117 15.603907 38.228936 15.625380
#> 273 137 2003-05-17 5.96266381 15.655493 38.415875 16.919642
#> 276 138 2003-05-18 1.10910662 15.706133 38.598118 11.012355
#> 278 139 2003-05-19 0.35962883 15.755795 38.775606 10.136432
#> 280 140 2003-05-20 1.72686232 15.804447 38.948286 11.864901
#> 281 141 2003-05-21 5.56287812 15.852058 39.116103 16.642429
#> 284 142 2003-05-22 2.61794558 15.898595 39.279005 13.053694
#> 286 143 2003-05-23 3.36434127 15.944026 39.436940 14.020020
#> 288 144 2003-05-24 2.25065942 15.988319 39.589859 12.683977
#> 289 145 2003-05-25 3.32228224 16.031441 39.737713 14.051959
#> 292 146 2003-05-26 2.54460138 16.073361 39.880455 13.126885
#> 294 147 2003-05-27 1.00246456 16.114046 40.018039 11.249283
#> 296 148 2003-05-28 6.03223491 16.153463 40.150421 17.534350
#> 297 149 2003-05-29 6.40962083 16.191581 40.277558 18.041553
#> 299 150 2003-05-30 4.35361678 16.228369 40.399408 15.518867
#> 302 151 2003-05-31 4.22685782 16.263796 40.515933 15.393901
#> 304 152 2003-06-01 4.72555027 16.297830 40.627093 16.046679
#> 306 153 2003-06-02 2.14028017 16.330443 40.732853 12.852452
#> 307 154 2003-06-03 5.33496272 16.361604 40.833176 16.865449
#> 310 155 2003-06-04 5.62081865 16.391286 40.928031 17.249427
#> 312 156 2003-06-05 5.10435515 16.419460 41.017385 16.629931
#> 314 157 2003-06-06 5.96587817 16.446100 41.101209 17.730103
#> 315 158 2003-06-07 4.69637802 16.471180 41.179475 16.165559
#> 318 159 2003-06-08 2.02322923 16.494677 41.252156 12.843024
#> 320 160 2003-06-09 0.11025299 16.516565 41.319229 10.467716
#> 322 161 2003-06-10 6.41900065 16.536824 41.380671 18.376412
#> 323 162 2003-06-11 5.01626571 16.555433 41.436461 16.636701
#> 326 163 2003-06-12 6.68953328 16.572373 41.486582 18.744794
#> 327 164 2003-06-13 4.39231515 16.587625 41.531016 15.881350
#> 330 165 2003-06-14 1.78381510 16.601175 41.569750 12.625796
#> 332 166 2003-06-15 6.13048633 16.613007 41.602770 18.076763
#> 334 167 2003-06-16 1.01716533 16.623108 41.630068 11.681186
#> 335 168 2003-06-17 6.04868781 16.631469 41.651634 17.987036
#> 338 169 2003-06-18 6.58762088 16.638080 41.667464 18.665698
#> 340 170 2003-06-19 0.37761942 16.642934 41.677552 10.892209
#> 342 171 2003-06-20 3.38806792 16.646025 41.681897 14.662361
#> 343 172 2003-06-21 0.43336073 16.647351 41.680500 10.962634
#> 346 173 2003-06-22 3.05883568 16.646909 41.673363 14.247039
#> 348 174 2003-06-23 3.52857549 16.644700 41.660491 14.831008
#> 350 175 2003-06-24 0.51427259 16.640727 41.641891 11.053933
#> 351 176 2003-06-25 4.68680536 16.634994 41.617572 16.267138
#> 353 177 2003-06-26 6.04328995 16.627507 41.587545 17.954411
#> 356 178 2003-06-27 4.54505397 16.618275 41.551823 16.070113
#> 358 179 2003-06-28 6.38959719 16.607306 41.510423 18.363093
#> 360 180 2003-06-29 2.04245638 16.594614 41.463360 12.917484
#> 361 181 2003-06-30 1.33046908 16.580211 41.410655 12.014150
#> 364 182 2003-07-01 3.87611102 16.564113 41.352328 15.176440
#> 366 183 2003-07-02 6.22326353 16.546336 41.288405 18.086618
#> 368 184 2003-07-03 0.75322704 16.526900 41.218909 11.244020
#> 369 185 2003-07-04 0.56969058 16.505823 41.143868 10.995997
#> 372 186 2003-07-05 3.52713596 16.483128 41.063311 14.659286
#> 374 187 2003-07-06 1.47341591 16.458837 40.977269 12.078486
#> 376 188 2003-07-07 6.14271704 16.432973 40.885775 17.863084
#> 377 189 2003-07-08 4.17063574 16.405563 40.788864 15.381905
#> 379 190 2003-07-09 0.15936163 16.376632 40.686570 10.369604
#> 382 191 2003-07-10 2.85534081 16.346207 40.578932 13.688879
#> 384 192 2003-07-11 1.07203324 16.314316 40.465990 11.446032
#> 386 193 2003-07-12 1.15340436 16.280989 40.347784 11.516138
#> 387 194 2003-07-13 6.96863284 16.246255 40.224356 18.682963
#> 390 195 2003-07-14 5.16951551 16.210144 40.095751 16.417330
#> 392 196 2003-07-15 4.09223210 16.172686 39.962013 15.046368
#> 394 197 2003-07-16 6.30614373 16.133915 39.823189 17.738495
#> 395 198 2003-07-17 1.42050950 16.093860 39.679327 11.670961
#> 398 199 2003-07-18 0.83255839 16.052554 39.530476 10.907734
#> 400 200 2003-07-19 1.70251221 16.010029 39.376687 11.937837
#> 402 201 2003-07-20 5.83575362 15.966317 39.218012 16.971674
#> 403 202 2003-07-21 6.27028202 15.921451 39.054503 17.453966
#> 406 203 2003-07-22 2.99178301 15.875462 38.886214 13.385671
#> 407 204 2003-07-23 0.93036306 15.828384 38.713200 10.816045
#> 410 205 2003-07-24 4.55958529 15.780249 38.535519 15.201155
#> 412 206 2003-07-25 4.49048584 15.731088 38.353226 15.062328
#> 414 207 2003-07-26 0.97268542 15.680933 38.166382 10.725322
#> 415 208 2003-07-27 4.22811039 15.629816 37.975044 14.630183
#> 418 209 2003-07-28 3.82715638 15.577768 37.779274 14.085637
#> 420 210 2003-07-29 6.04532332 15.524820 37.579134 16.711391
#> 422 211 2003-07-30 6.69304481 15.471001 37.374685 17.428166
#> 423 212 2003-07-31 5.05508500 15.416342 37.165992 15.384942
#> 426 213 2003-08-01 0.28337002 15.360873 36.953119  9.579127
#> 428 214 2003-08-02 4.81654769 15.304621 36.736132 14.964683
#> 430 215 2003-08-03 1.73288233 15.247616 36.515097 11.203734
#> 431 216 2003-08-04 3.24960618 15.189886 36.290082 12.954330
#> 433 217 2003-08-05 4.15592298 15.131457 36.061156 13.967469
#> 436 218 2003-08-06 1.92646694 15.072356 35.828388 11.246792
#> 438 219 2003-08-07 1.44278112 15.012610 35.591848 10.608233
#> 440 220 2003-08-08 1.82616509 14.952243 35.351609 10.996705
#> 441 221 2003-08-09 2.13175297 14.891281 35.107744 11.289851
#> 444 222 2003-08-10 4.00121542 14.829748 34.860325 13.417915
#> 446 223 2003-08-11 1.33762017 14.767666 34.609429 10.219777
#> 448 224 2003-08-12 3.49065697 14.705060 34.355129 12.666357
#> 449 225 2003-08-13 4.13386325 14.641952 34.097505 13.337752
#> 452 226 2003-08-14 6.44574196 14.578362 33.836632 15.939498
#> 454 227 2003-08-15 4.36397636 14.514313 33.572591 13.440235
#> 456 228 2003-08-16 3.70096714 14.449824 33.305462 12.591553
#> 457 229 2003-08-17 6.18436783 14.384916 33.035326 15.360110
#> 459 230 2003-08-18 6.47548066 14.319607 32.762264 15.598291
#> 462 231 2003-08-19 6.98228860 14.253916 32.486361 16.078321
#> 464 232 2003-08-20 1.60184320 14.187862 32.207701  9.870088
#> 466 233 2003-08-21 0.69529107 14.121461 31.926370  8.767564
#> 467 234 2003-08-22 4.19795456 14.054730 31.642455 12.636197
#> 470 235 2003-08-23 0.73361197 13.987686 31.356043  8.661276
#> 472 236 2003-08-24 5.60924993 13.920345 31.067223 14.026127
#> 474 237 2003-08-25 5.82823833 13.852721 30.776085 14.168213
#> 475 238 2003-08-26 1.00361260 13.784830 30.482721  8.730336
#> 478 239 2003-08-27 3.38400483 13.716686 30.187223 11.270508
#> 480 240 2003-08-28 0.41593608 13.648302 29.889684  7.927870
#> 482 241 2003-08-29 3.96119486 13.579692 29.590199 11.713279
#> 483 242 2003-08-30 2.58947696 13.510869 29.288864 10.128950
#> 486 243 2003-08-31 4.39341146 13.441846 28.985774 11.983383
#> 488 244 2003-09-01 5.56118110 13.372633 28.681027 13.133943
#> 489 245 2003-09-02 4.33519049 13.303244 28.374722 11.716982
#> 492 246 2003-09-03 3.39281605 13.233689 28.066958 10.614603
#> 494 247 2003-09-04 1.36361180 13.163980 27.757836  8.377129
#> 496 248 2003-09-05 1.05783363 13.094126 27.447457  7.970562
#> 497 249 2003-09-06 6.52111510 13.024138 27.135924 13.577385
#> 500 250 2003-09-07 1.74585892 12.954027 26.823338  8.513371
#> 502 251 2003-09-08 0.55859460 12.883802 26.509805  7.202135
#> 504 252 2003-09-09 0.57490573 12.813472 26.195428  7.136516
#> 505 253 2003-09-10 3.18835958 12.743047 25.880314  9.707756
#> 508 254 2003-09-11 6.64376555 12.672536 25.564567 13.092444
#> 510 255 2003-09-12 6.96517497 12.601947 25.248295 13.289519
#> 512 256 2003-09-13 1.85144599 12.531289 24.931604  8.074671
#> 513 257 2003-09-14 2.35409171 12.460572 24.614602  8.478786
#> 515 258 2003-09-15 6.54390308 12.389802 24.297397 12.490909
#> 518 259 2003-09-16 6.53800931 12.318988 23.980097 12.358457
#> 520 260 2003-09-17 3.17042118 12.248139 23.662812  8.978253
#> 522 261 2003-09-18 5.08041145 12.177262 23.345649 10.706370
#> 523 262 2003-09-19 4.80427066 12.106365 23.028718 10.326520
#> 526 263 2003-09-20 2.23556512 12.035456 22.712128  7.787401
#> 528 264 2003-09-21 0.32402803 11.964544 22.395989  5.902265
#> 530 265 2003-09-22 0.38996585 11.893635 22.080409  5.882086
#> 531 266 2003-09-23 3.27713020 11.822738 21.765497  8.457950
#> 534 267 2003-09-24 4.86748608 11.751861 21.451363  9.805295
#> 536 268 2003-09-25 0.05598748 11.681012 21.138114  5.335186
#> 538 269 2003-09-26 5.29487441 11.610198 20.825859  9.955320
#> 539 270 2003-09-27 0.45101929 11.539428 20.514705  5.529586
#> 541 271 2003-09-28 3.93141632 11.468711 20.204761  8.514235
#> 544 272 2003-09-29 2.77610080 11.398053 19.896131  7.396976
#> 546 273 2003-09-30 0.38281455 11.327464 19.588923  5.228237
#> 548 274 2003-10-01 6.36876946 11.256953 19.283241 10.275684
#> 549 275 2003-10-02 3.93448544 11.186528 18.979189  8.082443
#> 552 276 2003-10-03 0.46890529 11.116198 18.676871  5.063133
#> 554 277 2003-10-04 4.42075486 11.045973 18.376389  8.271343
#> 556 278 2003-10-05 5.12188737 10.975862 18.077844  8.737475
#> 557 279 2003-10-06 3.24968029 10.905874 17.781337  7.094533
#> 560 280 2003-10-07 2.19371176 10.836020 17.486965  6.141826
#> 562 281 2003-10-08 1.17820812 10.766311 17.194826  5.239562
#> 564 282 2003-10-09 3.22854922 10.696756 16.905016  6.777433
#> 565 283 2003-10-10 5.87086374 10.627367 16.617629  8.744436
#> 568 284 2003-10-11 0.61009091 10.558154 16.332759  4.555075
#> 569 285 2003-10-12 6.86978860 10.489131 16.050495  9.268708
#> 572 286 2003-10-13 4.10356093 10.420308 15.770928  7.048061
#> 574 287 2003-10-14 1.97128878 10.351698 15.494145  5.348822
#> 576 288 2003-10-15 3.47086799 10.283314 15.220231  6.373656
#> 577 289 2003-10-16 2.45724855 10.215170 14.949271  5.535334
#> 580 290 2003-10-17 5.55470648 10.147279 14.681345  7.688683
#> 582 291 2003-10-18 3.37073869 10.079655 14.416534  6.014651
#> 584 292 2003-10-19 5.60217709 10.012314 14.154915  7.498769
#> 585 293 2003-10-20 6.80943783  9.945270 13.896562  8.231567
#> 588 294 2003-10-21 5.62120973  9.878539 13.641550  7.291630
#> 590 295 2003-10-22 6.91617253  9.812138 13.389948  8.066499
#> 592 296 2003-10-23 2.00025269  9.746084 13.141825  4.634048
#> 593 297 2003-10-24 5.60391534  9.680393 12.897247  6.957377
#> 595 298 2003-10-25 1.92213631  9.615084 12.656277  4.429117
#> 598 299 2003-10-26 2.90527923  9.550176 12.418977  4.993746
#> 600 300 2003-10-27 6.54111775  9.485687 12.185404  7.247742
#> 602 301 2003-10-28 4.65099424  9.421638 11.955615  5.939850
#> 603 302 2003-10-29 3.43738717  9.358048 11.729663  5.086679
#> 606 303 2003-10-30 6.66451982  9.294940 11.507600  7.002404
#> 608 304 2003-10-31 1.39316484  9.232334 11.289473  3.674162
#> 610 305 2003-11-01 6.94406980  9.170252 11.075328  6.962166
#> 611 306 2003-11-02 2.71477602  9.108719 10.865209  4.335444
#> 614 307 2003-11-03 6.46813658  9.047757 10.659156  6.474843
#> 616 308 2003-11-04 1.30542218  8.987390 10.457208  3.373759
#> 618 309 2003-11-05 0.56310686  8.927644 10.259398  2.888403
#> 619 310 2003-11-06 2.78521999  8.868543 10.065762  4.097047
#> 621 311 2003-11-07 1.02815793  8.810114  9.876328  3.045376
#> 624 312 2003-11-08 2.30231613  8.752384  9.691126  3.697408
#> 626 313 2003-11-09 4.44368000  8.695379  9.510180  4.807583
#> 628 314 2003-11-10 3.05400254  8.639127  9.333513  3.983115
#> 629 315 2003-11-11 5.76599476  8.583658  9.161147  5.367246
#> 632 316 2003-11-12 6.84060026  8.528999  8.993100  5.854689
#> 634 317 2003-11-13 0.42610578  8.475180  8.829388  2.429304
#> 636 318 2003-11-14 1.64656439  8.422232  8.670025  3.015011
#> 637 319 2003-11-15 0.70762396  8.370184  8.515023  2.488690
#> 640 320 2003-11-16 1.03848084  8.319067  8.364392  2.613167
#> 642 321 2003-11-17 6.97352652  8.268912  8.218139  5.519889
#> 644 322 2003-11-18 6.59192371  8.219751  8.076271  5.257497
#> 645 323 2003-11-19 4.49622950  8.171616  7.938791  4.168759
#> 648 324 2003-11-20 2.62408554  8.124538  7.805702  3.211979
#> 649 325 2003-11-21 5.97819669  8.078549  7.677004  4.759776
#> 652 326 2003-11-22 5.68928762  8.033683  7.552698  4.562506
#> 654 327 2003-11-23 5.20680562  7.989971  7.432780  4.280046
#> 656 328 2003-11-24 3.14226558  7.947446  7.317248  3.275861
#> 657 329 2003-11-25 1.04987889  7.906140  7.206095  2.279983
#> 660 330 2003-11-26 0.59498777  7.866085  7.099317  2.043324
#> 662 331 2003-11-27 1.77642160  7.827314  6.996907  2.543207
#> 664 332 2003-11-28 1.14641729  7.789856  6.898857  2.232359
#> 665 333 2003-11-29 3.24153391  7.753745  6.805157  3.123773
#> 668 334 2003-11-30 6.91578687  7.719011  6.715800  4.687434
#> 670 335 2003-12-01 2.02849782  7.685684  6.630775  2.532730
#> 672 336 2003-12-02 3.13731872  7.653793  6.550072  2.979967
#> 673 337 2003-12-03 6.33066697  7.623368  6.473680  4.306386
#> 675 338 2003-12-04 6.11362792  7.594437  6.401590  4.177082
#> 678 339 2003-12-05 1.85786245  7.567027  6.333789  2.360986
#> 680 340 2003-12-06 3.73930495  7.541163  6.270268  3.122131
#> 682 341 2003-12-07 2.15260830  7.516872  6.211016  2.442079
#> 683 342 2003-12-08 2.57621274  7.494177  6.156023  2.597109
#> 686 343 2003-12-09 1.92278039  7.473100  6.105278  2.311744
#> 688 344 2003-12-10 5.23933262  7.453664  6.058772  3.644111
#> 690 345 2003-12-11 0.60412384  7.435887  6.016497  1.748527
#> 691 346 2003-12-12 5.43044043  7.419789  5.978443  3.682380
#> 694 347 2003-12-13 1.14125806  7.405386  5.944603  1.944218
#> 696 348 2003-12-14 6.21047829  7.392694  5.914970  3.963276
#> 698 349 2003-12-15 1.74296967  7.381725  5.889539  2.167702
#> 699 350 2003-12-16 3.79491100  7.372493  5.868303  2.977399
#> 701 351 2003-12-17 0.84273628  7.365006  5.851259  1.797578
#> 704 352 2003-12-18 4.06568922  7.359273  5.838404  3.072338
#> 706 353 2003-12-19 6.41460395  7.355300  5.829736  3.999509
#> 708 354 2003-12-20 2.42253244  7.353091  5.825253  2.415901
#> 709 355 2003-12-21 3.92450491  7.352649  5.824956  3.010785
#> 712 356 2003-12-22 0.08692587  7.353975  5.828845  1.491661
#> 714 357 2003-12-23 3.48792940  7.357066  5.836924  2.842852
#> 716 358 2003-12-24 3.25348978  7.361920  5.849194  2.754780
#> 717 359 2003-12-25 5.94819254  7.368531  5.865662  3.833922
#> 720 360 2003-12-26 2.48395249  7.376892  5.886331  2.462608
#> 722 361 2003-12-27 1.97288956  7.386993  5.911209  2.267174
#> 724 362 2003-12-28 0.07843627  7.398825  5.940303  1.516563
#> 725 363 2003-12-29 1.03003390  7.412375  5.973622  1.908457
#> 728 364 2003-12-30 4.91110641  7.427627  6.011174  3.490071
#> 730 365 2003-12-31 1.96097878  7.444567  6.052971  2.310451