Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.580495  2.580495  3.624219  3.624219  2.513676  2.513676  3.342494
#>   [8]  3.342494  2.286347  2.286347  4.031513  4.031513  1.747460  1.747460
#>  [15]  2.662705  2.662705  4.336854  4.336854  3.752560  3.752560  3.662307
#>  [22]  3.662307  3.210884  3.210884  4.236565  4.236565  2.320670  2.320670
#>  [29]  2.174049  2.174049  4.932705  4.932705  4.186108  4.186108  3.768003
#>  [36]  3.768003  2.453467  2.453467  3.082446  3.082446  4.052462  4.052462
#>  [43]  5.452764  5.452764  3.469508  3.469508  4.569327  4.569327  3.097827
#>  [50]  3.097827  3.316959  3.316959  2.355042  2.355042  3.605069  3.605069
#>  [57]  5.955364  5.955364  4.981610  4.981610  2.567513  2.567513  3.379669
#>  [64]  3.379669  3.322172  3.322172  4.135920  4.135920  2.952059  2.952059
#>  [71]  5.236763  5.236763  3.179901  3.179901  5.655947  5.655947  5.130954
#>  [78]  5.130954  6.828032  6.828032  5.004285  5.004285  5.810671  5.810671
#>  [85]  4.479332  4.479332  7.294330  7.294330  3.386306  3.386306  6.142574
#>  [92]  6.142574  4.357951  4.357951  7.751765  7.751765  3.431321  3.431321
#>  [99]  3.950737  3.950737  4.101975  4.101975  5.769481  5.769481  8.220442
#> [106]  8.220442  5.133553  5.133553  8.795323  8.795323  5.866811  5.866811
#> [113]  8.949952  8.949952  5.612256  5.612256  4.048408  4.048408  7.711686
#> [120]  7.711686  7.905413  7.905413  5.761485  5.761485  6.106330  6.106330
#> [127]  6.238287  6.238287  6.968631  6.968631  6.334058  6.334058  8.678225
#> [134]  8.678225  7.949377  7.949377  7.803176  7.803176 10.022315 10.022315
#> [141]  6.075987  6.075987  8.684216  8.684216  8.678102  8.678102  6.540741
#> [148]  6.540741  8.013778  8.013778 11.609686 11.609686  9.554231  9.554231
#> [155]  9.265984  9.265984  9.062092  9.062092  5.953068  5.953068  7.032200
#> [162]  7.032200  5.923481  5.923481 11.941197 11.941197 10.963696 10.963696
#> [169]  6.146639  6.146639  6.272834  6.272834  8.831896  8.831896  6.701500
#> [176]  6.701500 11.038922 11.038922  8.461356  8.461356 12.155261 12.155261
#> [183] 12.385960 12.385960  7.455615  7.455615  9.078868  9.078868 12.436167
#> [190] 12.436167  8.144395  8.144395  7.548640  7.548640 13.562832 13.562832
#> [197] 11.725397 11.725397 11.003496 11.003496 12.190196 12.190196 14.260616
#> [204] 14.260616 10.456580 10.456580  9.903712  9.903712  9.108735  9.108735
#> [211] 12.803017 12.803017 10.560814 10.560814 10.653867 10.653867 12.681380
#> [218] 12.681380 13.625600 13.625600 13.066135 13.066135 15.636720 15.636720
#> [225]  8.609412  8.609412 13.978692 13.978692 15.656240 15.656240 13.178508
#> [232] 13.178508 15.684569 15.684569 11.967935 11.967935 10.275770 10.275770
#> [239] 10.598418 10.598418 15.337456 15.337456 15.483288 15.483288 12.719343
#> [246] 12.719343 14.246877 14.246877 16.691493 16.691493 12.488441 12.488441
#> [253] 10.054327 10.054327 16.603087 16.603087 16.331972 16.331972 12.794597
#> [260] 12.794597 11.688198 11.688198 15.050472 15.050472 17.499000 17.499000
#> [267] 14.625940 14.625940 11.876600 11.876600 13.864728 13.864728 12.684840
#> [274] 12.684840 17.109496 17.109496 13.860447 13.860447 16.210292 16.210292
#> [281] 17.181029 17.181029 17.599867 17.599867 17.663198 17.663198 15.803493
#> [288] 15.803493 16.651928 16.651928 13.518800 13.518800 15.149233 15.149233
#> [295] 12.481434 12.481434 13.305937 13.305937 16.191016 16.191016 18.845878
#> [302] 18.845878 14.009719 14.009719 12.460024 12.460024 14.953189 14.953189
#> [309] 11.696691 11.696691 18.498311 18.498311 16.565663 16.565663 10.862475
#> [316] 10.862475 18.760539 18.760539 12.674163 12.674163 11.247713 11.247713
#> [323] 14.479043 14.479043 16.733209 16.733209 15.599943 15.599943 15.310344
#> [330] 15.310344 14.526685 14.526685 15.539960 15.539960 15.380149 15.380149
#> [337] 15.958602 15.958602 15.230457 15.230457 17.750539 17.750539 16.122393
#> [344] 16.122393 10.897123 10.897123 12.539021 12.539021 11.700545 11.700545
#> [351] 13.067132 13.067132 11.215915 11.215915 16.422568 16.422568 16.518818
#> [358] 16.518818 15.171560 15.171560 15.227965 15.227965 11.594778 11.594778
#> [365] 14.512680 14.512680 12.117945 12.117945 17.385334 17.385334 13.287687
#> [372] 13.287687 13.486009 13.486009 11.617864 11.617864 10.763531 10.763531
#> [379] 17.121145 17.121145 15.310665 15.310665 18.331768 18.331768 15.818519
#> [386] 15.818519 10.648605 10.648605 14.539360 14.539360 15.138419 15.138419
#> [393] 17.902176 17.902176 18.438958 18.438958 12.174540 12.174540 17.195985
#> [400] 17.195985 14.834329 14.834329 11.556222 11.556222 13.729470 13.729470
#> [407] 13.856534 13.856534 14.104451 14.104451 12.326199 12.326199 10.146440
#> [414] 10.146440 15.744712 15.744712 17.600940 17.600940 15.848028 15.848028
#> [421] 15.667541 15.667541 14.193065 14.193065  9.321800  9.321800 11.137900
#> [428] 11.137900 12.739735 12.739735 13.402234 13.402234 11.877069 11.877069
#> [435] 16.640016 16.640016 13.586656 13.586656 13.889035 13.889035 14.447394
#> [442] 14.447394 15.238698 15.238698 10.486996 10.486996 12.926120 12.926120
#> [449] 10.406343 10.406343 16.534258 16.534258 10.456522 10.456522 12.637387
#> [456] 12.637387  9.362386  9.362386  8.692583  8.692583 11.644262 11.644262
#> [463] 12.571417 12.571417 10.506048 10.506048  8.908864  8.908864 14.369127
#> [470] 14.369127 12.861220 12.861220 10.604793 10.604793 10.922852 10.922852
#> [477] 13.199256 13.199256  8.751915  8.751915  8.123743  8.123743  7.617699
#> [484]  7.617699 13.468516 13.468516  8.440499  8.440499 13.145432 13.145432
#> [491]  7.975795  7.975795 11.764864 11.764864  8.504568  8.504568  6.890305
#> [498]  6.890305  8.080905  8.080905 11.188138 11.188138 11.120540 11.120540
#> [505] 12.138950 12.138950  6.706139  6.706139  7.067125  7.067125  6.315237
#> [512]  6.315237  7.130711  7.130711 10.712667 10.712667 11.522102 11.522102
#> [519]  9.545839  9.545839 10.210020 10.210020 12.196032 12.196032  5.951098
#> [526]  5.951098 11.649873 11.649873  6.123980  6.123980 11.829734 11.829734
#> [533]  6.216028  6.216028  5.962734  5.962734  5.973134  5.973134  9.293220
#> [540]  9.293220 10.372002 10.372002  7.405059  7.405059  9.504547  9.504547
#> [547]  8.944947  8.944947 10.566862 10.566862  5.652284  5.652284  9.305762
#> [554]  9.305762  9.313084  9.313084  6.644221  6.644221  4.727074  4.727074
#> [561]  9.024904  9.024904  5.298672  5.298672  7.018978  7.018978  7.945417
#> [568]  7.945417  7.998161  7.998161  5.619949  5.619949  5.270529  5.270529
#> [575]  6.506627  6.506627  4.419249  4.419249  7.017675  7.017675  5.342237
#> [582]  5.342237  6.183387  6.183387  5.691875  5.691875  7.585833  7.585833
#> [589]  6.980758  6.980758  5.823203  5.823203  3.341670  3.341670  6.315397
#> [596]  6.315397  3.321693  3.321693  3.958907  3.958907  5.534616  5.534616
#> [603]  6.917107  6.917107  4.728151  4.728151  6.800959  6.800959  5.795198
#> [610]  5.795198  3.006315  3.006315  2.719399  2.719399  4.571179  4.571179
#> [617]  3.560318  3.560318  2.680699  2.680699  6.175540  6.175540  5.229406
#> [624]  5.229406  5.936559  5.936559  5.721771  5.721771  5.716393  5.716393
#> [631]  3.596481  3.596481  4.542165  4.542165  5.727153  5.727153  2.316982
#> [638]  2.316982  4.532704  4.532704  4.628302  4.628302  2.997841  2.997841
#> [645]  3.843066  3.843066  3.163075  3.163075  4.270525  4.270525  1.910515
#> [652]  1.910515  4.870571  4.870571  3.764383  3.764383  2.475294  2.475294
#> [659]  2.195877  2.195877  4.603509  4.603509  4.163866  4.163866  3.458625
#> [666]  3.458625  3.012924  3.012924  2.464341  2.464341  3.336970  3.336970
#> [673]  3.114098  3.114098  2.598264  2.598264  3.951320  3.951320  2.414183
#> [680]  2.414183  2.230483  2.230483  4.050499  4.050499  4.217600  4.217600
#> [687]  3.258140  3.258140  3.833046  3.833046  3.117476  3.117476  2.179964
#> [694]  2.179964  2.070158  2.070158  2.718387  2.718387  2.365316  2.365316
#> [701]  4.201670  4.201670  2.230869  2.230869  3.358544  3.358544  3.798938
#> [708]  3.798938  3.232955  3.232955  2.233375  2.233375  2.501491  2.501491
#> [715]  3.636204  3.636204  3.724774  3.724774  3.821802  3.821802  3.274679
#> [722]  3.274679  3.517744  3.517744  1.741451  1.741451  3.751539  3.751539
#> [729]  4.064184  4.064184