Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.884850  1.884850  4.055555  4.055555  2.093853  2.093853  2.586440
#>   [8]  2.586440  1.635228  1.635228  3.409809  3.409809  3.797151  3.797151
#>  [15]  2.348207  2.348207  3.679267  3.679267  4.236857  4.236857  2.222169
#>  [22]  2.222169  4.210970  4.210970  3.080245  3.080245  4.762088  4.762088
#>  [29]  3.455903  3.455903  1.905775  1.905775  3.843633  3.843633  3.173103
#>  [36]  3.173103  4.027534  4.027534  2.747578  2.747578  2.926327  2.926327
#>  [43]  3.454889  3.454889  3.851280  3.851280  4.813574  4.813574  4.775456
#>  [50]  4.775456  2.320513  2.320513  5.690315  5.690315  5.121304  5.121304
#>  [57]  3.484793  3.484793  3.674300  3.674300  5.745835  5.745835  4.180810
#>  [64]  4.180810  6.316108  6.316108  4.065432  4.065432  2.672336  2.672336
#>  [71]  4.764127  4.764127  2.852200  2.852200  5.867945  5.867945  6.782861
#>  [78]  6.782861  4.791221  4.791221  6.691743  6.691743  4.345954  4.345954
#>  [85]  3.238871  3.238871  4.004389  4.004389  6.545443  6.545443  7.214124
#>  [92]  7.214124  6.142092  6.142092  4.108701  4.108701  7.137948  7.137948
#>  [99]  8.121454  8.121454  7.578722  7.578722  5.583872  5.583872  3.719082
#> [106]  3.719082  7.109014  7.109014  4.244124  4.244124  8.854880  8.854880
#> [113]  8.899768  8.899768  4.416554  4.416554  8.339540  8.339540  6.004376
#> [120]  6.004376  6.857305  6.857305  7.108105  7.108105  8.426271  8.426271
#> [127]  7.953941  7.953941  9.060290  9.060290  5.129924  5.129924  4.895926
#> [134]  4.895926  7.707902  7.707902  5.773505  5.773505  6.472760  6.472760
#> [141]  5.608786  5.608786  9.031475  9.031475  9.373946  9.373946  8.635547
#> [148]  8.635547  8.259490  8.259490 10.908222 10.908222  8.730791  8.730791
#> [155]  7.408028  7.408028  6.595516  6.595516  6.816001  6.816001  8.913247
#> [162]  8.913247  7.130750  7.130750  8.338120  8.338120  9.243988  9.243988
#> [169] 12.784534 12.784534 12.083561 12.083561 12.636454 12.636454  6.603640
#> [176]  6.603640  8.974340  8.974340  9.575424  9.575424  7.802128  7.802128
#> [183] 13.818326 13.818326 10.523016 10.523016  6.831978  6.831978  9.840294
#> [190]  9.840294 10.735597 10.735597 10.910723 10.910723  8.179555  8.179555
#> [197] 14.074979 14.074979  7.883326  7.883326 12.547844 12.547844 14.424441
#> [204] 14.424441 13.993200 13.993200 12.906985 12.906985 11.833225 11.833225
#> [211] 11.005434 11.005434  9.566107  9.566107 11.760148 11.760148 14.448564
#> [218] 14.448564 12.193125 12.193125 14.544295 14.544295 11.283173 11.283173
#> [225] 16.142721 16.142721  8.290230  8.290230 10.103668 10.103668 12.422338
#> [232] 12.422338 16.151528 16.151528 14.666850 14.666850 12.049984 12.049984
#> [239] 16.820846 16.820846 11.821717 11.821717 15.123619 15.123619 10.554498
#> [246] 10.554498 16.797534 16.797534 11.947784 11.947784 14.770210 14.770210
#> [253] 14.578852 14.578852 15.837632 15.837632 11.771489 11.771489 16.066739
#> [260] 16.066739 11.470150 11.470150  9.554854  9.554854 17.147108 17.147108
#> [267] 10.642681 10.642681 15.312352 15.312352 15.587181 15.587181 10.839168
#> [274] 10.839168 11.245667 11.245667 12.341242 12.341242 16.461415 16.461415
#> [281] 11.602041 11.602041 10.178962 10.178962 12.983462 12.983462 16.305253
#> [288] 16.305253 16.321864 16.321864 17.429875 17.429875 15.078291 15.078291
#> [295] 11.828198 11.828198 16.267667 16.267667 14.961687 14.961687 14.320768
#> [302] 14.320768 16.506851 16.506851 14.291168 14.291168 14.676146 14.676146
#> [309] 10.912606 10.912606 11.876710 11.876710 11.734007 11.734007 11.138022
#> [316] 11.138022 15.526228 15.526228 18.489941 18.489941 18.061996 18.061996
#> [323] 14.648084 14.648084 10.479818 10.479818 17.732276 17.732276 16.120865
#> [330] 16.120865 11.238155 11.238155 17.812173 17.812173 12.858605 12.858605
#> [337] 11.852926 11.852926 16.507407 16.507407 10.821763 10.821763 13.262218
#> [344] 13.262218 15.002062 15.002062 17.216247 17.216247 13.751493 13.751493
#> [351] 15.046309 15.046309 13.795383 13.795383 18.447444 18.447444 15.410746
#> [358] 15.410746 18.122283 18.122283 13.397248 13.397248 13.562796 13.562796
#> [365] 17.287782 17.287782 14.488173 14.488173 18.243310 18.243310 11.917381
#> [372] 11.917381 13.020817 13.020817 16.924120 16.924120 13.783667 13.783667
#> [379] 16.609878 16.609878 11.658129 11.658129 13.264327 13.264327 16.413526
#> [386] 16.413526 10.784221 10.784221 16.479975 16.479975 18.205117 18.205117
#> [393] 15.599497 15.599497 16.846567 16.846567 10.554937 10.554937 18.124381
#> [400] 18.124381 12.905686 12.905686 10.018415 10.018415 14.453796 14.453796
#> [407] 16.594617 16.594617 13.634738 13.634738 11.887203 11.887203 12.165864
#> [414] 12.165864 17.534687 17.534687 13.897794 13.897794 15.520204 15.520204
#> [421] 13.244743 13.244743 17.133313 17.133313 17.560838 17.560838 16.929053
#> [428] 16.929053 10.840900 10.840900 12.616817 12.616817 10.274623 10.274623
#> [435] 15.853567 15.853567 12.312574 12.312574 14.697711 14.697711 16.134301
#> [442] 16.134301  8.940798  8.940798 10.889648 10.889648 15.805952 15.805952
#> [449] 13.501070 13.501070 15.613937 15.613937 11.358925 11.358925 11.734674
#> [456] 11.734674  8.902445  8.902445  8.840616  8.840616 11.039248 11.039248
#> [463]  8.801333  8.801333 13.138754 13.138754  9.580439  9.580439 12.955483
#> [470] 12.955483  8.843536  8.843536  8.625377  8.625377 14.828613 14.828613
#> [477] 10.244060 10.244060  8.973951  8.973951 11.060024 11.060024 12.264157
#> [484] 12.264157 12.508530 12.508530  9.714904  9.714904 13.999610 13.999610
#> [491] 10.998209 10.998209 11.311629 11.311629  9.068660  9.068660 12.696519
#> [498] 12.696519  8.177378  8.177378 12.789219 12.789219 12.105927 12.105927
#> [505] 12.407665 12.407665  8.487066  8.487066 11.104378 11.104378 10.872304
#> [512] 10.872304 11.614733 11.614733  6.696129  6.696129  9.734486  9.734486
#> [519]  9.094561  9.094561 11.020807 11.020807 11.367179 11.367179  6.564905
#> [526]  6.564905  8.176399  8.176399  9.074384  9.074384  8.273780  8.273780
#> [533]  5.799329  5.799329  5.530734  5.530734 10.310182 10.310182  5.876439
#> [540]  5.876439  6.864886  6.864886 10.042803 10.042803  5.922881  5.922881
#> [547]  8.119035  8.119035  7.763084  7.763084  7.547268  7.547268  9.891761
#> [554]  9.891761 10.276592 10.276592  5.401729  5.401729  5.961375  5.961375
#> [561]  6.407930  6.407930  9.417870  9.417870  9.102205  9.102205  8.077252
#> [568]  8.077252  5.337604  5.337604  9.161892  9.161892  4.319056  4.319056
#> [575]  4.384249  4.384249  6.998394  6.998394  5.921914  5.921914  5.141581
#> [582]  5.141581  7.259344  7.259344  5.495288  5.495288  8.195921  8.195921
#> [589]  7.372486  7.372486  5.078688  5.078688  3.917316  3.917316  6.391379
#> [596]  6.391379  3.204921  3.204921  6.920699  6.920699  4.065841  4.065841
#> [603]  3.497381  3.497381  4.816425  4.816425  3.758545  3.758545  4.117309
#> [610]  4.117309  5.173816  5.173816  5.424782  5.424782  3.522287  3.522287
#> [617]  6.497685  6.497685  3.459859  3.459859  5.948665  5.948665  6.251389
#> [624]  6.251389  5.579764  5.579764  4.101271  4.101271  3.184710  3.184710
#> [631]  5.849734  5.849734  3.519971  3.519971  4.072267  4.072267  2.547215
#> [638]  2.547215  3.125883  3.125883  2.908974  2.908974  3.640909  3.640909
#> [645]  4.142204  4.142204  3.313215  3.313215  3.195489  3.195489  3.280304
#> [652]  3.280304  4.918193  4.918193  4.355532  4.355532  4.443323  4.443323
#> [659]  3.977840  3.977840  1.913496  1.913496  2.295929  2.295929  2.037354
#> [666]  2.037354  3.074205  3.074205  2.177736  2.177736  3.957596  3.957596
#> [673]  1.822285  1.822285  4.543945  4.543945  4.057896  4.057896  4.226361
#> [680]  4.226361  3.082091  3.082091  4.277290  4.277290  4.046521  4.046521
#> [687]  2.184132  2.184132  2.736049  2.736049  4.011129  4.011129  1.533980
#> [694]  1.533980  3.409808  3.409808  2.281634  2.281634  2.520214  2.520214
#> [701]  3.523530  3.523530  2.218033  2.218033  3.935245  3.935245  2.310557
#> [708]  2.310557  1.737103  1.737103  2.012274  2.012274  3.896509  3.896509
#> [715]  3.163100  3.163100  2.709483  2.709483  3.725901  3.725901  3.056836
#> [722]  3.056836  4.280543  4.280543  4.106140  4.106140  4.109613  4.109613
#> [729]  1.557046  1.557046