Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  3.325386  2.059171  4.332406  3.173470  1.660888  2.245657  3.866924
#>   [8]  3.500089  3.095959  4.692719  3.445403  2.219754  4.023063  2.137872
#>  [15]  4.673494  4.751804  3.100343  2.346859  2.037245  3.626265  4.638598
#>  [22]  4.565689  5.187018  2.775289  4.457914  4.642178  2.659171  3.526040
#>  [29]  4.446556  5.450193  4.109487  4.237926  2.713329  6.061377  6.135121
#>  [36]  4.938575  5.490980  2.962232  6.361092  4.576033  5.691511  6.194516
#>  [43]  5.669421  3.674955  5.722310  4.352295  4.163938  4.339380  3.632899
#>  [50]  7.645871  4.469921  4.255465  8.518371  4.475351  3.995403  6.688111
#>  [57]  3.993704  7.841128  4.842250  7.903829  5.726306  7.614457  8.383455
#>  [64]  8.181550  4.969101  7.557405 10.196736  5.968896  6.365711 10.602756
#>  [71]  5.764317  5.341250  7.769308  9.160076  9.211545  7.757770  7.640440
#>  [78] 10.966022 10.123547  8.837977 11.441129  8.246275 10.470140  6.642297
#>  [85] 11.458562  6.144124 12.084566  9.711821 12.056794 13.434922  8.157209
#>  [92] 10.727572  7.525497 13.953726 10.188994  6.990004 12.367535  7.874522
#>  [99]  7.520404 10.373781 10.604339  9.594078  9.511317 10.362376  9.994939
#> [106] 12.295702  8.262164 12.616396 15.274576 15.033417 12.510752 15.577402
#> [113] 11.699907 15.015729 12.792212  8.536106 14.505300 14.450256  9.841450
#> [120] 16.753923 10.243302 17.012555 14.444575 11.757529 14.120663  9.862067
#> [127] 13.426638 10.984483 11.423792 12.719512 14.340241 10.216917 15.793966
#> [134] 12.227220 11.957340 17.945378 13.734829 16.156253 12.046685 14.319887
#> [141] 17.158395 15.134587 12.995331 14.076105 17.215859 10.171087 17.268558
#> [148] 14.469733 16.640908 13.479915 10.378615 17.058649 14.025126 16.233575
#> [155] 15.475781 16.199011 12.072784 14.767926 11.370914 15.114814 12.418676
#> [162] 17.006160 16.350680 17.139412 12.155521 10.622340 14.441198 15.817464
#> [169] 11.455285 15.040576 19.106100 18.633045 15.215634 16.424435 12.114403
#> [176] 12.153310 16.210096 13.392805 18.728650 13.642935 18.762806 17.889672
#> [183] 12.308652 14.612879 14.634191 17.740003 16.293480 14.351499 16.746578
#> [190] 17.454139 15.126673 18.641043 17.861291 16.788041 13.818998 15.214179
#> [197] 15.273293 18.361988 10.389112 12.091902 16.643491 12.866398  9.943170
#> [204] 13.382730 12.677572 10.875985 16.483564 11.197112 16.619590 13.495448
#> [211]  9.531561 14.627093 15.027163 15.795984 15.270093 13.001427 10.922107
#> [218] 14.466653 11.215096 15.565061 13.581182 14.551852 13.339477 15.131619
#> [225] 11.747294 14.153316 11.491158 14.338338 14.122961 11.796732 14.332972
#> [232]  8.230637 15.033704 10.952880 10.604001  9.017129 12.715751 14.229233
#> [239]  8.522965 11.165189  8.253987  8.133563 11.839373  9.203929  9.223147
#> [246] 13.846168 14.046668  8.751922 10.922893 12.427220  8.168390 11.717807
#> [253] 13.274767 10.315471 10.507461 12.687322  8.510219 11.161378 11.778363
#> [260] 11.661647  9.197397  8.974666  8.479873  9.228413  5.733951  7.182655
#> [267]  7.253058 11.371103  5.461328  6.714683  7.251428  8.950629  9.562148
#> [274]  7.205452  6.788842  6.795661 10.270414  6.407253  6.223593  6.432073
#> [281]  9.153879  5.082208  7.526027  8.878972  5.511022  6.429291  5.926766
#> [288]  6.132161  4.894856  7.214209  4.667076  3.662270  4.837733  7.291510
#> [295]  6.332246  4.999472  7.183378  3.866441  4.934172  3.906091  5.674161
#> [302]  4.636374  3.613195  3.090189  3.523395  3.243339  2.928970  4.313396
#> [309]  6.162440  3.341726  6.041868  2.672661  4.159786  4.870365  2.710964
#> [316]  4.253248  2.493361  2.363026  4.471924  5.304842  3.928226  4.528105
#> [323]  5.323517  3.883294  3.185333  5.115944  3.592647  3.142990  4.392355
#> [330]  3.132811  2.760398  4.532445  4.443667  3.813904  2.885724  1.791978
#> [337]  3.863299  4.346353  4.285682  4.088586  3.380414  3.122847  3.572476
#> [344]  2.199163  2.636463  2.319613  3.646192  1.946592  3.662431  2.951451
#> [351]  2.472590  3.334081  3.425045  2.109775  2.711063  3.038146  3.914660
#> [358]  1.476207  1.871132  3.253974  4.066127  4.223615  2.986061  1.794579
#> [365]  1.972209  3.325386  2.059171  4.332406  3.173470  1.660888  2.245657
#> [372]  3.866924  3.500089  3.095959  4.692719  3.445403  2.219754  4.023063
#> [379]  2.137872  4.673494  4.751804  3.100343  2.346859  2.037245  3.626265
#> [386]  4.638598  4.565689  5.187018  2.775289  4.457914  4.642178  2.659171
#> [393]  3.526040  4.446556  5.450193  4.109487  4.237926  2.713329  6.061377
#> [400]  6.135121  4.938575  5.490980  2.962232  6.361092  4.576033  5.691511
#> [407]  6.194516  5.669421  3.674955  5.722310  4.352295  4.163938  4.339380
#> [414]  3.632899  7.645871  4.469921  4.255465  8.518371  4.475351  3.995403
#> [421]  6.688111  3.993704  7.841128  4.842250  7.903829  5.726306  7.614457
#> [428]  8.383455  8.181550  4.969101  7.557405 10.196736  5.968896  6.365711
#> [435] 10.602756  5.764317  5.341250  7.769308  9.160076  9.211545  7.757770
#> [442]  7.640440 10.966022 10.123547  8.837977 11.441129  8.246275 10.470140
#> [449]  6.642297 11.458562  6.144124 12.084566  9.711821 12.056794 13.434922
#> [456]  8.157209 10.727572  7.525497 13.953726 10.188994  6.990004 12.367535
#> [463]  7.874522  7.520404 10.373781 10.604339  9.594078  9.511317 10.362376
#> [470]  9.994939 12.295702  8.262164 12.616396 15.274576 15.033417 12.510752
#> [477] 15.577402 11.699907 15.015729 12.792212  8.536106 14.505300 14.450256
#> [484]  9.841450 16.753923 10.243302 17.012555 14.444575 11.757529 14.120663
#> [491]  9.862067 13.426638 10.984483 11.423792 12.719512 14.340241 10.216917
#> [498] 15.793966 12.227220 11.957340 17.945378 13.734829 16.156253 12.046685
#> [505] 14.319887 17.158395 15.134587 12.995331 14.076105 17.215859 10.171087
#> [512] 17.268558 14.469733 16.640908 13.479915 10.378615 17.058649 14.025126
#> [519] 16.233575 15.475781 16.199011 12.072784 14.767926 11.370914 15.114814
#> [526] 12.418676 17.006160 16.350680 17.139412 12.155521 10.622340 14.441198
#> [533] 15.817464 11.455285 15.040576 19.106100 18.633045 15.215634 16.424435
#> [540] 12.114403 12.153310 16.210096 13.392805 18.728650 13.642935 18.762806
#> [547] 17.889672 12.308652 14.612879 14.634191 17.740003 16.293480 14.351499
#> [554] 16.746578 17.454139 15.126673 18.641043 17.861291 16.788041 13.818998
#> [561] 15.214179 15.273293 18.361988 10.389112 12.091902 16.643491 12.866398
#> [568]  9.943170 13.382730 12.677572 10.875985 16.483564 11.197112 16.619590
#> [575] 13.495448  9.531561 14.627093 15.027163 15.795984 15.270093 13.001427
#> [582] 10.922107 14.466653 11.215096 15.565061 13.581182 14.551852 13.339477
#> [589] 15.131619 11.747294 14.153316 11.491158 14.338338 14.122961 11.796732
#> [596] 14.332972  8.230637 15.033704 10.952880 10.604001  9.017129 12.715751
#> [603] 14.229233  8.522965 11.165189  8.253987  8.133563 11.839373  9.203929
#> [610]  9.223147 13.846168 14.046668  8.751922 10.922893 12.427220  8.168390
#> [617] 11.717807 13.274767 10.315471 10.507461 12.687322  8.510219 11.161378
#> [624] 11.778363 11.661647  9.197397  8.974666  8.479873  9.228413  5.733951
#> [631]  7.182655  7.253058 11.371103  5.461328  6.714683  7.251428  8.950629
#> [638]  9.562148  7.205452  6.788842  6.795661 10.270414  6.407253  6.223593
#> [645]  6.432073  9.153879  5.082208  7.526027  8.878972  5.511022  6.429291
#> [652]  5.926766  6.132161  4.894856  7.214209  4.667076  3.662270  4.837733
#> [659]  7.291510  6.332246  4.999472  7.183378  3.866441  4.934172  3.906091
#> [666]  5.674161  4.636374  3.613195  3.090189  3.523395  3.243339  2.928970
#> [673]  4.313396  6.162440  3.341726  6.041868  2.672661  4.159786  4.870365
#> [680]  2.710964  4.253248  2.493361  2.363026  4.471924  5.304842  3.928226
#> [687]  4.528105  5.323517  3.883294  3.185333  5.115944  3.592647  3.142990
#> [694]  4.392355  3.132811  2.760398  4.532445  4.443667  3.813904  2.885724
#> [701]  1.791978  3.863299  4.346353  4.285682  4.088586  3.380414  3.122847
#> [708]  3.572476  2.199163  2.636463  2.319613  3.646192  1.946592  3.662431
#> [715]  2.951451  2.472590  3.334081  3.425045  2.109775  2.711063  3.038146
#> [722]  3.914660  1.476207  1.871132  3.253974  4.066127  4.223615  2.986061
#> [729]  1.794579  1.972209
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 5.88728035  7.463176  6.099033  3.930351
#> 3     2 2002-01-02 0.99732387  7.483435  6.149345  1.947100
#> 6     3 2002-01-03 2.84786472  7.505323  6.203934  2.728012
#> 7     4 2002-01-04 3.26925127  7.528820  6.262814  2.925459
#> 9     5 2002-01-05 5.25707035  7.553900  6.325996  3.782760
#> 11    6 2002-01-06 2.29643399  7.580540  6.393496  2.566790
#> 14    7 2002-01-07 3.23908570  7.608714  6.465324  2.992499
#> 15    8 2002-01-08 4.21218119  7.638396  6.541496  3.439022
#> 17    9 2002-01-09 1.08838615  7.669557  6.622023  2.125371
#> 19   10 2002-01-10 3.10146092  7.702170  6.706920  3.027080
#> 22   11 2002-01-11 3.67700706  7.736204  6.796198  3.314161
#> 23   12 2002-01-12 0.51325373  7.771631  6.889871  1.949978
#> 25   13 2002-01-13 5.15675579  7.808419  6.987950  4.054443
#> 27   14 2002-01-14 4.74845599  7.846537  7.090446  3.918059
#> 30   15 2002-01-15 5.55273449  7.885954  7.197371  4.333284
#> 32   16 2002-01-16 5.78980453  7.926639  7.308732  4.496418
#> 33   17 2002-01-17 3.20784723  7.968559  7.424539  3.350557
#> 35   18 2002-01-18 5.84591469  8.011681  7.544799  4.638822
#> 37   19 2002-01-19 2.88741765  8.055974  7.669519  3.291832
#> 40   20 2002-01-20 3.71239546  8.101405  7.798703  3.736518
#> 41   21 2002-01-21 4.75973718  8.147942  7.932354  4.299987
#> 43   22 2002-01-22 6.05019733  8.195553  8.070474  4.996548
#> 45   23 2002-01-23 2.43948897  8.244205  8.213063  3.268403
#> 48   24 2002-01-24 1.47660848  8.293867  8.360119  2.834232
#> 49   25 2002-01-25 5.48118218  8.344507  8.511639  4.923392
#> 51   26 2002-01-26 1.45409590  8.396093  8.667618  2.917465
#> 53   27 2002-01-27 3.58509151  8.448595  8.828046  4.080066
#> 56   28 2002-01-28 5.72015713  8.501983  8.992915  5.273458
#> 58   29 2002-01-29 1.04847024  8.556225  9.162213  2.851917
#> 59   30 2002-01-30 5.87394299  8.611293  9.335925  5.518096
#> 61   31 2002-01-31 1.26045481  8.667157  9.514035  3.070316
#> 63   32 2002-02-01 1.22747450  8.723789  9.696523  3.106302
#> 66   33 2002-02-02 0.28777974  8.781160  9.883368  2.632793
#> 67   34 2002-02-03 1.89536112  8.839243 10.074546  3.598757
#> 69   35 2002-02-04 3.89385942  8.898011 10.270030  4.814642
#> 71   36 2002-02-05 1.57338627  8.957438 10.469790  3.536964
#> 74   37 2002-02-06 2.31660395  9.017498 10.673795  4.039503
#> 75   38 2002-02-07 6.60370847  9.078165 10.882010  6.678440
#> 77   39 2002-02-08 6.68120436  9.139416 11.094398  6.828779
#> 79   40 2002-02-09 6.21240482  9.201226 11.310918  6.646134
#> 82   41 2002-02-10 4.25149131  9.263572 11.531528  5.529064
#> 83   42 2002-02-11 6.31000183  9.326432 11.756182  6.915996
#> 86   43 2002-02-12 6.02982746  9.389784 11.984831  6.844351
#> 87   44 2002-02-13 2.50558272  9.453606 12.217425  4.673409
#> 89   45 2002-02-14 2.05624792  9.517878 12.453910  4.458753
#> 91   46 2002-02-15 5.09160490  9.582579 12.694229  6.546031
#> 94   47 2002-02-16 5.26053531  9.647690 12.938322  6.761979
#> 95   48 2002-02-17 0.32570719  9.713192 13.186128  3.517614
#> 97   49 2002-02-18 4.85953925  9.779067 13.437583  6.698184
#> 99   50 2002-02-19 1.26433085  9.845297 13.692618  4.302356
#> 102  51 2002-02-20 2.88056090  9.911864 13.951164  5.515017
#> 103  52 2002-02-21 0.66390348  9.978754 14.213148  4.026099
#> 105  53 2002-02-22 4.81117134 10.045948 14.478496  7.086620
#> 107  54 2002-02-23 5.52526683 10.113433 14.747130  7.715179
#> 110  55 2002-02-24 5.14944320 10.181192 15.018971  7.552890
#> 112  56 2002-02-25 1.23525433 10.249211 15.293936  4.745111
#> 113  57 2002-02-26 3.14220071 10.317477 15.571940  6.264212
#> 115  58 2002-02-27 0.52519995 10.385975 15.852898  4.364051
#> 117  59 2002-02-28 6.56673574 10.454693 16.136720  9.102027
#> 120  60 2002-03-01 1.98978406 10.523618 16.423316  5.658472
#> 121  61 2002-03-02 3.81935214 10.592738 16.712592  7.191121
#> 123  62 2002-03-03 5.34658917 10.662040 17.004454  8.514642
#> 125  63 2002-03-04 1.13942735 10.731513 17.298805  5.243059
#> 128  64 2002-03-05 1.72251390 10.801147 17.595546  5.801912
#> 129  65 2002-03-06 0.43445051 10.870930 17.894578  4.831218
#> 131  66 2002-03-07 5.44505691 10.940852 18.195799  9.076804
#> 133  67 2002-03-08 1.03712861 11.010902 18.499104  5.496001
#> 136  68 2002-03-09 3.54932891 11.081072 18.804391  7.712673
#> 138  69 2002-03-10 3.54015221 11.151350 19.111552  7.811503
#> 139  70 2002-03-11 4.41186251 11.221729 19.420480  8.672735
#> 141  71 2002-03-12 2.50394176 11.292198 19.731067  7.120359
#> 143  72 2002-03-13 4.74630098 11.362749 20.043203  9.196895
#> 146  73 2002-03-14 1.30476653 11.433374 20.356778  6.250743
#> 147  74 2002-03-15 6.79417484 11.504062 20.671679 11.272155
#> 149  75 2002-03-16 6.13429639 11.574807 20.987796 10.808396
#> 151  76 2002-03-17 1.20368147 11.645600 21.305014  6.427289
#> 154  77 2002-03-18 0.10094065 11.716433 21.623220  5.498950
#> 155  78 2002-03-19 2.15468747 11.787297 21.942300  7.491073
#> 157  79 2002-03-20 1.80660154 11.858185 22.262140  7.261360
#> 159  80 2002-03-21 5.10413904 11.929088 22.582624 10.476907
#> 162  81 2002-03-22 1.45835114 12.000000 22.903638  7.117640
#> 163  82 2002-03-23 2.97238925 12.070912 23.225065  8.665782
#> 166  83 2002-03-24 2.84127477 12.141815 23.546791  8.641760
#> 167  84 2002-03-25 5.72040147 12.212703 23.868700 11.557197
#> 169  85 2002-03-26 0.53359560 12.283567 24.190677  6.573088
#> 171  86 2002-03-27 2.05538092 12.354400 24.512606  8.167212
#> 174  87 2002-03-28 5.86581058 12.425193 24.834373 12.070624
#> 175  88 2002-03-29 1.14005263 12.495938 25.155863  7.436499
#> 177  89 2002-03-30 6.01107667 12.566626 25.476961 12.462521
#> 179  90 2002-03-31 5.12630353 12.637251 25.797553 11.681780
#> 182  91 2002-04-01 2.85697422 12.707802 26.117528  9.465260
#> 183  92 2002-04-02 5.31393526 12.778271 26.436771 12.106153
#> 185  93 2002-04-03 1.10388759 12.848650 26.755170  7.838123
#> 187  94 2002-04-04 3.98454133 12.918928 27.072615 10.943111
#> 190  95 2002-04-05 6.36909915 12.989098 27.388994 13.562235
#> 192  96 2002-04-06 1.01323336 13.059148 27.704199  8.000807
#> 193  97 2002-04-07 4.30891293 13.129070 28.018120 11.602252
#> 195  98 2002-04-08 6.74153856 13.198853 28.330649 14.317843
#> 197  99 2002-04-09 3.37783604 13.268487 28.641680 10.806160
#> 200 100 2002-04-10 2.30317720 13.337960 28.951106  9.737391
#> 201 101 2002-04-11 1.74502295 13.407262 29.258822  9.218797
#> 203 102 2002-04-12 0.11536448 13.476382 29.564726  7.517726
#> 205 103 2002-04-13 3.24824827 13.545307 29.868714 11.048530
#> 208 104 2002-04-14 1.24683109 13.614025 30.170685  8.924252
#> 209 105 2002-04-15 1.41331321 13.682523 30.470539  9.191336
#> 211 106 2002-04-16 1.12342335 13.750789 30.768176  8.948906
#> 213 107 2002-04-17 6.82726190 13.818808 31.063500 15.439426
#> 216 108 2002-04-18 3.30327812 13.886567 31.356413 11.568569
#> 218 109 2002-04-19 1.60965273 13.954052 31.646820  9.736995
#> 219 110 2002-04-20 5.38213215 14.021246 31.934628 14.112798
#> 221 111 2002-04-21 5.90863398 14.088136 32.219743 14.811496
#> 223 112 2002-04-22 0.36247495 14.154703 32.502075  8.541677
#> 226 113 2002-04-23 0.85556289 14.220933 32.781534  9.181488
#> 227 114 2002-04-24 5.91556462 14.286808 33.058030 15.108475
#> 229 115 2002-04-25 2.71564674 14.352310 33.331477 11.486247
#> 231 116 2002-04-26 3.54015250 14.417421 33.601789 12.525854
#> 234 117 2002-04-27 3.95238602 14.482122 33.868881 13.088880
#> 235 118 2002-04-28 1.29461218 14.546394 34.132670 10.052052
#> 237 119 2002-04-29 2.38837070 14.610216 34.393074 11.409432
#> 239 120 2002-04-30 4.83033089 14.673568 34.650012 14.365650
#> 242 121 2002-05-01 6.01926500 14.736428 34.903405 15.854202
#> 243 122 2002-05-02 5.20032755 14.798774 35.153175 14.964752
#> 245 123 2002-05-03 5.32767265 14.860584 35.399245 15.195308
#> 248 124 2002-05-04 4.79261749 14.921835 35.641540 14.634087
#> 249 125 2002-05-05 5.23822931 14.982502 35.879987 15.242233
#> 251 126 2002-05-06 6.16545781 15.042562 36.114511 16.429710
#> 253 127 2002-05-07 3.70335098 15.101989 36.345041 13.542576
#> 256 128 2002-05-08 1.61800555 15.160757 36.571508 11.094393
#> 257 129 2002-05-09 3.31901326 15.218840 36.793843 13.210568
#> 259 130 2002-05-10 6.83192193 15.276211 37.011977 17.529357
#> 261 131 2002-05-11 3.78336872 15.332843 37.225844 13.899187
#> 264 132 2002-05-12 6.07490291 15.388707 37.435378 16.747909
#> 265 133 2002-05-13 3.41780829 15.443775 37.640517 13.575175
#> 267 134 2002-05-14 5.50101514 15.498017 37.841197 16.176158
#> 269 135 2002-05-15 2.81680716 15.551405 38.037357 12.954169
#> 272 136 2002-05-16 6.27856947 15.603907 38.228936 17.248354
#> 274 137 2002-05-17 4.70683661 15.655493 38.415875 15.378851
#> 275 138 2002-05-18 6.93302998 15.706133 38.598118 18.168555
#> 277 139 2002-05-19 5.74707964 15.755795 38.775606 16.765792
#> 279 140 2002-05-20 1.66296568 15.804447 38.948286 11.786168
#> 282 141 2002-05-21 2.60827330 15.852058 39.116103 12.997078
#> 283 142 2002-05-22 1.41935723 15.898595 39.279005 11.573080
#> 285 143 2002-05-23 3.33232411 15.944026 39.436940 13.980423
#> 287 144 2002-05-24 4.65878583 15.988319 39.589859 15.665447
#> 290 145 2002-05-25 0.09039124 16.031441 39.737713 10.046456
#> 291 146 2002-05-26 6.86248397 16.073361 39.880455 18.483547
#> 293 147 2002-05-27 1.63530860 16.114046 40.018039 12.035094
#> 295 148 2002-05-28 1.78143731 16.153463 40.150421 12.251541
#> 298 149 2002-05-29 5.42816096 16.191581 40.277558 16.820832
#> 300 150 2002-05-30 2.20244448 16.228369 40.399408 12.841269
#> 301 151 2002-05-31 3.62758437 16.263796 40.515933 14.647454
#> 303 152 2002-06-01 0.38452855 16.297830 40.627093 10.636048
#> 305 153 2002-06-02 0.24229191 16.330443 40.732853 10.485386
#> 308 154 2002-06-03 3.41114486 16.361604 40.833176 14.464841
#> 309 155 2002-06-04 6.43917722 16.391286 40.928031 18.271122
#> 311 156 2002-06-05 2.44914370 16.419460 41.017385 13.313444
#> 313 157 2002-06-06 0.30308071 16.446100 41.101209 10.654024
#> 316 158 2002-06-07 5.23819867 16.471180 41.179475 16.842859
#> 317 159 2002-06-08 6.62362369 16.494677 41.252156 18.595674
#> 319 160 2002-06-09 1.73860669 16.516565 41.319229 12.504530
#> 321 161 2002-06-10 1.86176268 16.536824 41.380671 12.674544
#> 324 162 2002-06-11 1.41043472 16.555433 41.436461 12.124198
#> 325 163 2002-06-12 1.49834244 16.572373 41.486582 12.247089
#> 328 164 2002-06-13 2.82621875 16.587625 41.531016 13.920805
#> 329 165 2002-06-14 5.06706744 16.601175 41.569750 16.736468
#> 331 166 2002-06-15 2.57999292 16.613007 41.602770 13.631139
#> 333 167 2002-06-16 1.22499382 16.623108 41.630068 11.941423
#> 336 168 2002-06-17 1.38687576 16.631469 41.651634 12.149545
#> 337 169 2002-06-18 5.39234658 16.638080 41.667464 17.169009
#> 339 170 2002-06-19 1.20476382 16.642934 41.677552 11.927884
#> 341 171 2002-06-20 2.71721455 16.646025 41.681897 13.822447
#> 344 172 2002-06-21 4.38452174 16.647351 41.680500 15.908958
#> 345 173 2002-06-22 6.33025521 16.646909 41.673363 18.341824
#> 347 174 2002-06-23 1.89638725 16.644700 41.660491 12.788383
#> 349 175 2002-06-24 3.52499008 16.640727 41.641891 14.820955
#> 352 176 2002-06-25 4.01601651 16.634994 41.617572 15.428045
#> 354 177 2002-06-26 5.52822270 16.627507 41.587545 17.310286
#> 355 178 2002-06-27 3.66946178 16.618275 41.551823 14.975461
#> 357 179 2002-06-28 4.27421470 16.607306 41.510423 15.719365
#> 359 180 2002-06-29 6.56320521 16.594614 41.463360 18.565264
#> 362 181 2002-06-30 5.08622640 16.580211 41.410655 16.704331
#> 363 182 2002-07-01 5.01932683 16.564113 41.352328 16.603460
#> 365 183 2002-07-02 0.74497104 16.546336 41.288405 11.251572
#> 367 184 2002-07-03 6.15051224 16.526900 41.218909 17.974569
#> 370 185 2002-07-04 0.70178876 16.505823 41.143868 11.160637
#> 371 186 2002-07-05 2.65275712 16.483128 41.063311 13.570146
#> 373 187 2002-07-06 6.84061609 16.458837 40.977269 18.759797
#> 375 188 2002-07-07 3.82715192 16.432973 40.885775 14.982483
#> 378 189 2002-07-08 1.67899347 16.405563 40.788864 12.284442
#> 380 190 2002-07-09 2.58602967 16.376632 40.686570 13.384045
#> 381 191 2002-07-10 6.32366585 16.346207 40.578932 17.993882
#> 383 192 2002-07-11 0.69608870 16.314316 40.465990 10.979786
#> 385 193 2002-07-12 1.42652246 16.280989 40.347784 11.854560
#> 388 194 2002-07-13 5.30415639 16.246255 40.224356 16.622411
#> 389 195 2002-07-14 1.97069053 16.210144 40.095751 12.461187
#> 391 196 2002-07-15 2.95948325 16.172686 39.962013 13.646881
#> 393 197 2002-07-16 0.98090762 16.133915 39.823189 11.166380
#> 396 198 2002-07-17 1.38086050 16.093860 39.679327 11.622084
#> 397 199 2002-07-18 4.91280391 16.052554 39.530476 15.931672
#> 399 200 2002-07-19 6.99385679 16.010029 39.376687 18.444872
#> 401 201 2002-07-20 3.36466227 15.966317 39.218012 13.936807
#> 404 202 2002-07-21 1.00323047 15.921451 39.054503 10.994062
#> 405 203 2002-07-22 5.32549010 15.875462 38.886214 16.243825
#> 408 204 2002-07-23 1.91458463 15.828384 38.713200 12.019654
#> 409 205 2002-07-24 6.26178181 15.780249 38.535519 17.279545
#> 411 206 2002-07-25 0.04395864 15.731088 38.353226  9.641893
#> 413 207 2002-07-26 0.21610383 15.680933 38.166382  9.804587
#> 416 208 2002-07-27 0.33781811 15.629816 37.975044  9.904152
#> 417 209 2002-07-28 0.37846093 15.577768 37.779274  9.903741
#> 419 210 2002-07-29 6.24418354 15.524820 37.579134 16.952070
#> 421 211 2002-07-30 0.36011785 15.471001 37.374685  9.778656
#> 424 212 2002-07-31 1.18588409 15.416342 37.165992 10.720973
#> 425 213 2002-08-01 6.08868604 15.360873 36.953119 16.561950
#> 427 214 2002-08-02 5.37204911 15.304621 36.736132 15.631377
#> 429 215 2002-08-03 6.87876059 15.247616 36.515097 17.365426
#> 432 216 2002-08-04 5.72375200 15.189886 36.290082 15.909814
#> 434 217 2002-08-05 3.23556616 15.131457 36.061156 12.870775
#> 435 218 2002-08-06 1.38520249 15.072356 35.828388 10.603474
#> 437 219 2002-08-07 1.13516957 15.012610 35.591848 10.243590
#> 439 220 2002-08-08 4.57884990 14.952243 35.351609 14.250793
#> 442 221 2002-08-09 2.86260037 14.891281 35.107744 12.151375
#> 443 222 2002-08-10 1.48811970 14.829748 34.860325 10.464145
#> 445 223 2002-08-11 2.88665677 14.767666 34.609429 12.034934
#> 447 224 2002-08-12 2.27588864 14.705060 34.355129 11.247338
#> 450 225 2002-08-13 3.02023703 14.641952 34.097505 12.041071
#> 451 226 2002-08-14 1.91527511 14.578362 33.836632 10.681851
#> 453 227 2002-08-15 0.34182251 14.514313 33.572591  8.788477
#> 455 228 2002-08-16 1.68905642 14.449824 33.305462 10.272922
#> 458 229 2002-08-17 4.37532574 14.384916 33.035326 13.282855
#> 460 230 2002-08-18 4.15792726 14.319607 32.762264 12.947090
#> 461 231 2002-08-19 3.51504194 14.253916 32.486361 12.127188
#> 463 232 2002-08-20 2.97701307 14.187862 32.207701 11.430967
#> 465 233 2002-08-21 5.77263300 14.121461 31.926370 14.507094
#> 468 234 2002-08-22 1.03401976 14.054730 31.642455  9.074596
#> 469 235 2002-08-23 5.77071359 13.987686 31.356043 14.307083
#> 471 236 2002-08-24 2.08587978 13.920345 31.067223 10.094424
#> 473 237 2002-08-25 4.12055349 13.852721 30.776085 12.271263
#> 476 238 2002-08-26 2.37878049 13.784830 30.482721 10.250807
#> 477 239 2002-08-27 3.52747393 13.716686 30.187223 11.428379
#> 479 240 2002-08-28 4.79023155 13.648302 29.889684 12.717708
#> 481 241 2002-08-29 0.10543136 13.579692 29.590199  7.512418
#> 484 242 2002-08-30 6.07479514 13.510869 29.288864 13.906687
#> 485 243 2002-08-31 3.85800195 13.441846 28.985774 11.406109
#> 487 244 2002-09-01 5.47615957 13.372633 28.681027 13.042768
#> 490 245 2002-09-02 3.14763470 13.303244 28.374722 10.450503
#> 491 246 2002-09-03 3.20644738 13.233689 28.066958 10.416971
#> 493 247 2002-09-04 2.94275668 13.163980 27.757836 10.042037
#> 495 248 2002-09-05 0.78073314 13.094126 27.447457  7.680137
#> 498 249 2002-09-06 2.75294725 13.024138 27.135924  9.651878
#> 499 250 2002-09-07 4.81818423 12.954027 26.823338 11.694236
#> 501 251 2002-09-08 2.68563869 12.883802 26.509805  9.390446
#> 503 252 2002-09-09 6.10870607 12.813472 26.195428 12.793073
#> 506 253 2002-09-10 0.26465842 12.743047 25.880314  6.738831
#> 507 254 2002-09-11 1.95264934 12.672536 25.564567  8.360702
#> 509 255 2002-09-12 0.28683350 12.601947 25.248295  6.599412
#> 511 256 2002-09-13 1.14666165 12.531289 24.931604  7.373570
#> 514 257 2002-09-14 4.72673764 12.460572 24.614602 10.822247
#> 516 258 2002-09-15 0.72138783 12.389802 24.297397  6.781699
#> 517 259 2002-09-16 1.49054584 12.318988 23.980097  7.445770
#> 519 260 2002-09-17 3.03948061 12.248139 23.662812  8.851768
#> 521 261 2002-09-18 3.67744383 12.177262 23.345649  9.361520
#> 524 262 2002-09-19 5.29623662 12.106365 23.028718 10.794428
#> 525 263 2002-09-20 4.25644613 12.035456 22.712128  9.694205
#> 527 264 2002-09-21 6.62175461 11.964544 22.395989 11.796507
#> 529 265 2002-09-22 0.78878993 11.893635 22.080409  6.252292
#> 532 266 2002-09-23 0.39913146 11.822738 21.765497  5.808772
#> 533 267 2002-09-24 5.15150318 11.751861 21.451363 10.064511
#> 535 268 2002-09-25 2.20902129 11.681012 21.138114  7.283266
#> 537 269 2002-09-26 5.01942311 11.610198 20.825859  9.708274
#> 540 270 2002-09-27 6.05193950 11.539428 20.514705 10.508222
#> 542 271 2002-09-28 6.85721121 11.468711 20.204761 11.091464
#> 543 272 2002-09-29 0.54553327 11.398053 19.896131  5.450167
#> 545 273 2002-09-30 5.56853630 11.327464 19.588923  9.712149
#> 547 274 2002-10-01 6.85283944 11.256953 19.283241 10.690292
#> 550 275 2002-10-02 0.28911170 11.186528 18.979189  4.990052
#> 551 276 2002-10-03 5.59270993 11.116198 18.676871  9.367511
#> 553 277 2002-10-04 0.40498212 11.045973 18.376389  4.930967
#> 555 278 2002-10-05 5.89758519 10.975862 18.077844  9.376284
#> 558 279 2002-10-06 1.22314244 10.905874 17.781337  5.442462
#> 559 280 2002-10-07 1.14459081 10.836020 17.486965  5.295301
#> 561 281 2002-10-08 1.56535655 10.766311 17.194826  5.548718
#> 563 282 2002-10-09 2.55513277 10.696756 16.905016  6.245304
#> 566 283 2002-10-10 6.93780649 10.627367 16.617629  9.578606
#> 567 284 2002-10-11 3.39590945 10.558154 16.332759  6.709812
#> 570 285 2002-10-12 4.21120062 10.489131 16.050495  7.234619
#> 571 286 2002-10-13 1.17079780 10.420308 15.770928  4.828722
#> 573 287 2002-10-14 6.90687267 10.351698 15.494145  9.042547
#> 575 288 2002-10-15 4.86203239 10.283314 15.220231  7.403181
#> 578 289 2002-10-16 2.83884472 10.215170 14.949271  5.814555
#> 579 290 2002-10-17 6.37888233 10.147279 14.681345  8.284902
#> 581 291 2002-10-18 0.71384971 10.079655 14.416534  4.114629
#> 583 292 2002-10-19 3.02674855 10.012314 14.154915  5.678262
#> 586 293 2002-10-20 5.00591853  9.945270 13.896562  6.971535
#> 587 294 2002-10-21 0.59270937  9.878539 13.641550  3.819632
#> 589 295 2002-10-22 5.12180450  9.812138 13.389948  6.842174
#> 591 296 2002-10-23 4.67359087  9.746084 13.141825  6.436441
#> 594 297 2002-10-24 6.34715464  9.680393 12.897247  7.452488
#> 596 298 2002-10-25 4.55498821  9.615084 12.656277  6.161921
#> 597 299 2002-10-26 0.05503942  9.550176 12.418977  3.140531
#> 599 300 2002-10-27 3.72531985  9.485687 12.185404  5.439142
#> 601 301 2002-10-28 0.89342249  9.421638 11.955615  3.555759
#> 604 302 2002-10-29 0.29341253  9.358048 11.729663  3.116302
#> 605 303 2002-10-30 3.45319860  9.294940 11.507600  5.014516
#> 607 304 2002-10-31 2.63543961  9.232334 11.289473  4.433701
#> 609 305 2002-11-01 0.49285411  9.170252 11.075328  3.066453
#> 612 306 2002-11-02 2.90120311  9.108719 10.865209  4.446632
#> 613 307 2002-11-03 5.98206789  9.047757 10.659156  6.188524
#> 615 308 2002-11-04 4.79602577  8.987390 10.457208  5.404491
#> 617 309 2002-11-05 4.59883431  8.927644 10.259398  5.207275
#> 620 310 2002-11-06 2.51137937  8.868543 10.065762  3.941643
#> 622 311 2002-11-07 2.64702610  8.810114  9.876328  3.952769
#> 623 312 2002-11-08 2.56012977  8.752384  9.691126  3.840140
#> 625 313 2002-11-09 5.03703609  8.695379  9.510180  5.132061
#> 627 314 2002-11-10 6.85480024  8.639127  9.333513  6.036262
#> 630 315 2002-11-11 1.04248484  8.583658  9.161147  2.846597
#> 631 316 2002-11-12 0.73438306  8.528999  8.993100  2.635447
#> 633 317 2002-11-13 6.11942979  8.475180  8.829388  5.394938
#> 635 318 2002-11-14 5.75815793  8.422232  8.670025  5.131292
#> 638 319 2002-11-15 1.70617585  8.370184  8.515023  2.996606
#> 639 320 2002-11-16 3.05545171  8.319067  8.364392  3.627147
#> 641 321 2002-11-17 4.49455891  8.268912  8.218139  4.288015
#> 643 322 2002-11-18 3.34700290  8.219751  8.076271  3.663357
#> 646 323 2002-11-19 2.93149448  8.171616  7.938791  3.408683
#> 647 324 2002-11-20 4.65003184  8.124538  7.805702  4.185199
#> 650 325 2002-11-21 1.49552096  8.078549  7.677004  2.629844
#> 651 326 2002-11-22 1.20670358  8.033683  7.552698  2.455403
#> 653 327 2002-11-23 6.82443730  7.989971  7.432780  5.032458
#> 655 328 2002-11-24 4.08519840  7.947446  7.317248  3.709942
#> 658 329 2002-11-25 0.26143867  7.906140  7.206095  1.920669
#> 659 330 2002-11-26 0.50577865  7.866085  7.099317  2.003068
#> 661 331 2002-11-27 0.69507027  7.827314  6.996907  2.059892
#> 663 332 2002-11-28 6.20668641  7.789856  6.898857  4.473098
#> 666 333 2002-11-29 5.34687789  7.753745  6.805157  4.047661
#> 667 334 2002-11-30 0.58470086  7.719011  6.715800  1.933305
#> 669 335 2002-12-01 5.20011287  7.685684  6.630775  3.900876
#> 671 336 2002-12-02 0.78443102  7.653793  6.550072  1.973174
#> 674 337 2002-12-03 5.63648862  7.623368  6.473680  4.011642
#> 676 338 2002-12-04 0.26845705  7.594437  6.401590  1.713543
#> 677 339 2002-12-05 2.99332023  7.567027  6.333789  2.836189
#> 679 340 2002-12-06 5.86007453  7.541163  6.270268  4.003812
#> 681 341 2002-12-07 2.18633500  7.516872  6.211016  2.456013
#> 684 342 2002-12-08 1.64351635  7.494177  6.156023  2.214031
#> 685 343 2002-12-09 1.53896590  7.473100  6.105278  2.154962
#> 687 344 2002-12-10 6.78803236  7.453664  6.058772  4.273547
#> 689 345 2002-12-11 6.09612254  7.435887  6.016497  3.970360
#> 692 346 2002-12-12 2.89731546  7.419789  5.978443  2.661856
#> 693 347 2002-12-13 5.17010975  7.405386  5.944603  3.561279
#> 695 348 2002-12-14 5.38236430  7.392694  5.914970  3.631985
#> 697 349 2002-12-15 1.88925763  7.381725  5.889539  2.226060
#> 700 350 2002-12-16 6.91680206  7.372493  5.868303  4.219869
#> 702 351 2002-12-17 6.09615712  7.365006  5.851259  3.884415
#> 703 352 2002-12-18 6.66239082  7.359273  5.838404  4.102371
#> 705 353 2002-12-19 0.27659963  7.355300  5.829736  1.567049
#> 707 354 2002-12-20 3.52696713  7.353091  5.825253  2.853377
#> 710 355 2002-12-21 0.05606027  7.352649  5.824956  1.478445
#> 711 356 2002-12-22 5.89632310  7.353975  5.828845  3.793958
#> 713 357 2002-12-23 6.26343999  7.357066  5.836924  3.943864
#> 715 358 2002-12-24 2.41077964  7.361920  5.849194  2.420005
#> 718 359 2002-12-25 2.61848106  7.368531  5.865662  2.508626
#> 719 360 2002-12-26 0.09894429  7.376892  5.886331  1.511059
#> 721 361 2002-12-27 4.47753295  7.386993  5.911209  3.269305
#> 723 362 2002-12-28 5.20409578  7.398825  5.940303  3.574185
#> 726 363 2002-12-29 2.44451003  7.412375  5.973622  2.478419
#> 727 364 2002-12-30 6.86648030  7.427627  6.011174  4.281313
#> 729 365 2002-12-31 3.87294596  7.444567  6.052971  3.087735
#> 2     1 2003-01-01 5.88728035  7.463176  6.099033  3.930351
#> 4     2 2003-01-02 0.99732387  7.483435  6.149345  1.947100
#> 5     3 2003-01-03 2.84786472  7.505323  6.203934  2.728012
#> 8     4 2003-01-04 3.26925127  7.528820  6.262814  2.925459
#> 10    5 2003-01-05 5.25707035  7.553900  6.325996  3.782760
#> 12    6 2003-01-06 2.29643399  7.580540  6.393496  2.566790
#> 13    7 2003-01-07 3.23908570  7.608714  6.465324  2.992499
#> 16    8 2003-01-08 4.21218119  7.638396  6.541496  3.439022
#> 18    9 2003-01-09 1.08838615  7.669557  6.622023  2.125371
#> 20   10 2003-01-10 3.10146092  7.702170  6.706920  3.027080
#> 21   11 2003-01-11 3.67700706  7.736204  6.796198  3.314161
#> 24   12 2003-01-12 0.51325373  7.771631  6.889871  1.949978
#> 26   13 2003-01-13 5.15675579  7.808419  6.987950  4.054443
#> 28   14 2003-01-14 4.74845599  7.846537  7.090446  3.918059
#> 29   15 2003-01-15 5.55273449  7.885954  7.197371  4.333284
#> 31   16 2003-01-16 5.78980453  7.926639  7.308732  4.496418
#> 34   17 2003-01-17 3.20784723  7.968559  7.424539  3.350557
#> 36   18 2003-01-18 5.84591469  8.011681  7.544799  4.638822
#> 38   19 2003-01-19 2.88741765  8.055974  7.669519  3.291832
#> 39   20 2003-01-20 3.71239546  8.101405  7.798703  3.736518
#> 42   21 2003-01-21 4.75973718  8.147942  7.932354  4.299987
#> 44   22 2003-01-22 6.05019733  8.195553  8.070474  4.996548
#> 46   23 2003-01-23 2.43948897  8.244205  8.213063  3.268403
#> 47   24 2003-01-24 1.47660848  8.293867  8.360119  2.834232
#> 50   25 2003-01-25 5.48118218  8.344507  8.511639  4.923392
#> 52   26 2003-01-26 1.45409590  8.396093  8.667618  2.917465
#> 54   27 2003-01-27 3.58509151  8.448595  8.828046  4.080066
#> 55   28 2003-01-28 5.72015713  8.501983  8.992915  5.273458
#> 57   29 2003-01-29 1.04847024  8.556225  9.162213  2.851917
#> 60   30 2003-01-30 5.87394299  8.611293  9.335925  5.518096
#> 62   31 2003-01-31 1.26045481  8.667157  9.514035  3.070316
#> 64   32 2003-02-01 1.22747450  8.723789  9.696523  3.106302
#> 65   33 2003-02-02 0.28777974  8.781160  9.883368  2.632793
#> 68   34 2003-02-03 1.89536112  8.839243 10.074546  3.598757
#> 70   35 2003-02-04 3.89385942  8.898011 10.270030  4.814642
#> 72   36 2003-02-05 1.57338627  8.957438 10.469790  3.536964
#> 73   37 2003-02-06 2.31660395  9.017498 10.673795  4.039503
#> 76   38 2003-02-07 6.60370847  9.078165 10.882010  6.678440
#> 78   39 2003-02-08 6.68120436  9.139416 11.094398  6.828779
#> 80   40 2003-02-09 6.21240482  9.201226 11.310918  6.646134
#> 81   41 2003-02-10 4.25149131  9.263572 11.531528  5.529064
#> 84   42 2003-02-11 6.31000183  9.326432 11.756182  6.915996
#> 85   43 2003-02-12 6.02982746  9.389784 11.984831  6.844351
#> 88   44 2003-02-13 2.50558272  9.453606 12.217425  4.673409
#> 90   45 2003-02-14 2.05624792  9.517878 12.453910  4.458753
#> 92   46 2003-02-15 5.09160490  9.582579 12.694229  6.546031
#> 93   47 2003-02-16 5.26053531  9.647690 12.938322  6.761979
#> 96   48 2003-02-17 0.32570719  9.713192 13.186128  3.517614
#> 98   49 2003-02-18 4.85953925  9.779067 13.437583  6.698184
#> 100  50 2003-02-19 1.26433085  9.845297 13.692618  4.302356
#> 101  51 2003-02-20 2.88056090  9.911864 13.951164  5.515017
#> 104  52 2003-02-21 0.66390348  9.978754 14.213148  4.026099
#> 106  53 2003-02-22 4.81117134 10.045948 14.478496  7.086620
#> 108  54 2003-02-23 5.52526683 10.113433 14.747130  7.715179
#> 109  55 2003-02-24 5.14944320 10.181192 15.018971  7.552890
#> 111  56 2003-02-25 1.23525433 10.249211 15.293936  4.745111
#> 114  57 2003-02-26 3.14220071 10.317477 15.571940  6.264212
#> 116  58 2003-02-27 0.52519995 10.385975 15.852898  4.364051
#> 118  59 2003-02-28 6.56673574 10.454693 16.136720  9.102027
#> 119  60 2003-03-01 1.98978406 10.523618 16.423316  5.658472
#> 122  61 2003-03-02 3.81935214 10.592738 16.712592  7.191121
#> 124  62 2003-03-03 5.34658917 10.662040 17.004454  8.514642
#> 126  63 2003-03-04 1.13942735 10.731513 17.298805  5.243059
#> 127  64 2003-03-05 1.72251390 10.801147 17.595546  5.801912
#> 130  65 2003-03-06 0.43445051 10.870930 17.894578  4.831218
#> 132  66 2003-03-07 5.44505691 10.940852 18.195799  9.076804
#> 134  67 2003-03-08 1.03712861 11.010902 18.499104  5.496001
#> 135  68 2003-03-09 3.54932891 11.081072 18.804391  7.712673
#> 137  69 2003-03-10 3.54015221 11.151350 19.111552  7.811503
#> 140  70 2003-03-11 4.41186251 11.221729 19.420480  8.672735
#> 142  71 2003-03-12 2.50394176 11.292198 19.731067  7.120359
#> 144  72 2003-03-13 4.74630098 11.362749 20.043203  9.196895
#> 145  73 2003-03-14 1.30476653 11.433374 20.356778  6.250743
#> 148  74 2003-03-15 6.79417484 11.504062 20.671679 11.272155
#> 150  75 2003-03-16 6.13429639 11.574807 20.987796 10.808396
#> 152  76 2003-03-17 1.20368147 11.645600 21.305014  6.427289
#> 153  77 2003-03-18 0.10094065 11.716433 21.623220  5.498950
#> 156  78 2003-03-19 2.15468747 11.787297 21.942300  7.491073
#> 158  79 2003-03-20 1.80660154 11.858185 22.262140  7.261360
#> 160  80 2003-03-21 5.10413904 11.929088 22.582624 10.476907
#> 161  81 2003-03-22 1.45835114 12.000000 22.903638  7.117640
#> 164  82 2003-03-23 2.97238925 12.070912 23.225065  8.665782
#> 165  83 2003-03-24 2.84127477 12.141815 23.546791  8.641760
#> 168  84 2003-03-25 5.72040147 12.212703 23.868700 11.557197
#> 170  85 2003-03-26 0.53359560 12.283567 24.190677  6.573088
#> 172  86 2003-03-27 2.05538092 12.354400 24.512606  8.167212
#> 173  87 2003-03-28 5.86581058 12.425193 24.834373 12.070624
#> 176  88 2003-03-29 1.14005263 12.495938 25.155863  7.436499
#> 178  89 2003-03-30 6.01107667 12.566626 25.476961 12.462521
#> 180  90 2003-03-31 5.12630353 12.637251 25.797553 11.681780
#> 181  91 2003-04-01 2.85697422 12.707802 26.117528  9.465260
#> 184  92 2003-04-02 5.31393526 12.778271 26.436771 12.106153
#> 186  93 2003-04-03 1.10388759 12.848650 26.755170  7.838123
#> 188  94 2003-04-04 3.98454133 12.918928 27.072615 10.943111
#> 189  95 2003-04-05 6.36909915 12.989098 27.388994 13.562235
#> 191  96 2003-04-06 1.01323336 13.059148 27.704199  8.000807
#> 194  97 2003-04-07 4.30891293 13.129070 28.018120 11.602252
#> 196  98 2003-04-08 6.74153856 13.198853 28.330649 14.317843
#> 198  99 2003-04-09 3.37783604 13.268487 28.641680 10.806160
#> 199 100 2003-04-10 2.30317720 13.337960 28.951106  9.737391
#> 202 101 2003-04-11 1.74502295 13.407262 29.258822  9.218797
#> 204 102 2003-04-12 0.11536448 13.476382 29.564726  7.517726
#> 206 103 2003-04-13 3.24824827 13.545307 29.868714 11.048530
#> 207 104 2003-04-14 1.24683109 13.614025 30.170685  8.924252
#> 210 105 2003-04-15 1.41331321 13.682523 30.470539  9.191336
#> 212 106 2003-04-16 1.12342335 13.750789 30.768176  8.948906
#> 214 107 2003-04-17 6.82726190 13.818808 31.063500 15.439426
#> 215 108 2003-04-18 3.30327812 13.886567 31.356413 11.568569
#> 217 109 2003-04-19 1.60965273 13.954052 31.646820  9.736995
#> 220 110 2003-04-20 5.38213215 14.021246 31.934628 14.112798
#> 222 111 2003-04-21 5.90863398 14.088136 32.219743 14.811496
#> 224 112 2003-04-22 0.36247495 14.154703 32.502075  8.541677
#> 225 113 2003-04-23 0.85556289 14.220933 32.781534  9.181488
#> 228 114 2003-04-24 5.91556462 14.286808 33.058030 15.108475
#> 230 115 2003-04-25 2.71564674 14.352310 33.331477 11.486247
#> 232 116 2003-04-26 3.54015250 14.417421 33.601789 12.525854
#> 233 117 2003-04-27 3.95238602 14.482122 33.868881 13.088880
#> 236 118 2003-04-28 1.29461218 14.546394 34.132670 10.052052
#> 238 119 2003-04-29 2.38837070 14.610216 34.393074 11.409432
#> 240 120 2003-04-30 4.83033089 14.673568 34.650012 14.365650
#> 241 121 2003-05-01 6.01926500 14.736428 34.903405 15.854202
#> 244 122 2003-05-02 5.20032755 14.798774 35.153175 14.964752
#> 246 123 2003-05-03 5.32767265 14.860584 35.399245 15.195308
#> 247 124 2003-05-04 4.79261749 14.921835 35.641540 14.634087
#> 250 125 2003-05-05 5.23822931 14.982502 35.879987 15.242233
#> 252 126 2003-05-06 6.16545781 15.042562 36.114511 16.429710
#> 254 127 2003-05-07 3.70335098 15.101989 36.345041 13.542576
#> 255 128 2003-05-08 1.61800555 15.160757 36.571508 11.094393
#> 258 129 2003-05-09 3.31901326 15.218840 36.793843 13.210568
#> 260 130 2003-05-10 6.83192193 15.276211 37.011977 17.529357
#> 262 131 2003-05-11 3.78336872 15.332843 37.225844 13.899187
#> 263 132 2003-05-12 6.07490291 15.388707 37.435378 16.747909
#> 266 133 2003-05-13 3.41780829 15.443775 37.640517 13.575175
#> 268 134 2003-05-14 5.50101514 15.498017 37.841197 16.176158
#> 270 135 2003-05-15 2.81680716 15.551405 38.037357 12.954169
#> 271 136 2003-05-16 6.27856947 15.603907 38.228936 17.248354
#> 273 137 2003-05-17 4.70683661 15.655493 38.415875 15.378851
#> 276 138 2003-05-18 6.93302998 15.706133 38.598118 18.168555
#> 278 139 2003-05-19 5.74707964 15.755795 38.775606 16.765792
#> 280 140 2003-05-20 1.66296568 15.804447 38.948286 11.786168
#> 281 141 2003-05-21 2.60827330 15.852058 39.116103 12.997078
#> 284 142 2003-05-22 1.41935723 15.898595 39.279005 11.573080
#> 286 143 2003-05-23 3.33232411 15.944026 39.436940 13.980423
#> 288 144 2003-05-24 4.65878583 15.988319 39.589859 15.665447
#> 289 145 2003-05-25 0.09039124 16.031441 39.737713 10.046456
#> 292 146 2003-05-26 6.86248397 16.073361 39.880455 18.483547
#> 294 147 2003-05-27 1.63530860 16.114046 40.018039 12.035094
#> 296 148 2003-05-28 1.78143731 16.153463 40.150421 12.251541
#> 297 149 2003-05-29 5.42816096 16.191581 40.277558 16.820832
#> 299 150 2003-05-30 2.20244448 16.228369 40.399408 12.841269
#> 302 151 2003-05-31 3.62758437 16.263796 40.515933 14.647454
#> 304 152 2003-06-01 0.38452855 16.297830 40.627093 10.636048
#> 306 153 2003-06-02 0.24229191 16.330443 40.732853 10.485386
#> 307 154 2003-06-03 3.41114486 16.361604 40.833176 14.464841
#> 310 155 2003-06-04 6.43917722 16.391286 40.928031 18.271122
#> 312 156 2003-06-05 2.44914370 16.419460 41.017385 13.313444
#> 314 157 2003-06-06 0.30308071 16.446100 41.101209 10.654024
#> 315 158 2003-06-07 5.23819867 16.471180 41.179475 16.842859
#> 318 159 2003-06-08 6.62362369 16.494677 41.252156 18.595674
#> 320 160 2003-06-09 1.73860669 16.516565 41.319229 12.504530
#> 322 161 2003-06-10 1.86176268 16.536824 41.380671 12.674544
#> 323 162 2003-06-11 1.41043472 16.555433 41.436461 12.124198
#> 326 163 2003-06-12 1.49834244 16.572373 41.486582 12.247089
#> 327 164 2003-06-13 2.82621875 16.587625 41.531016 13.920805
#> 330 165 2003-06-14 5.06706744 16.601175 41.569750 16.736468
#> 332 166 2003-06-15 2.57999292 16.613007 41.602770 13.631139
#> 334 167 2003-06-16 1.22499382 16.623108 41.630068 11.941423
#> 335 168 2003-06-17 1.38687576 16.631469 41.651634 12.149545
#> 338 169 2003-06-18 5.39234658 16.638080 41.667464 17.169009
#> 340 170 2003-06-19 1.20476382 16.642934 41.677552 11.927884
#> 342 171 2003-06-20 2.71721455 16.646025 41.681897 13.822447
#> 343 172 2003-06-21 4.38452174 16.647351 41.680500 15.908958
#> 346 173 2003-06-22 6.33025521 16.646909 41.673363 18.341824
#> 348 174 2003-06-23 1.89638725 16.644700 41.660491 12.788383
#> 350 175 2003-06-24 3.52499008 16.640727 41.641891 14.820955
#> 351 176 2003-06-25 4.01601651 16.634994 41.617572 15.428045
#> 353 177 2003-06-26 5.52822270 16.627507 41.587545 17.310286
#> 356 178 2003-06-27 3.66946178 16.618275 41.551823 14.975461
#> 358 179 2003-06-28 4.27421470 16.607306 41.510423 15.719365
#> 360 180 2003-06-29 6.56320521 16.594614 41.463360 18.565264
#> 361 181 2003-06-30 5.08622640 16.580211 41.410655 16.704331
#> 364 182 2003-07-01 5.01932683 16.564113 41.352328 16.603460
#> 366 183 2003-07-02 0.74497104 16.546336 41.288405 11.251572
#> 368 184 2003-07-03 6.15051224 16.526900 41.218909 17.974569
#> 369 185 2003-07-04 0.70178876 16.505823 41.143868 11.160637
#> 372 186 2003-07-05 2.65275712 16.483128 41.063311 13.570146
#> 374 187 2003-07-06 6.84061609 16.458837 40.977269 18.759797
#> 376 188 2003-07-07 3.82715192 16.432973 40.885775 14.982483
#> 377 189 2003-07-08 1.67899347 16.405563 40.788864 12.284442
#> 379 190 2003-07-09 2.58602967 16.376632 40.686570 13.384045
#> 382 191 2003-07-10 6.32366585 16.346207 40.578932 17.993882
#> 384 192 2003-07-11 0.69608870 16.314316 40.465990 10.979786
#> 386 193 2003-07-12 1.42652246 16.280989 40.347784 11.854560
#> 387 194 2003-07-13 5.30415639 16.246255 40.224356 16.622411
#> 390 195 2003-07-14 1.97069053 16.210144 40.095751 12.461187
#> 392 196 2003-07-15 2.95948325 16.172686 39.962013 13.646881
#> 394 197 2003-07-16 0.98090762 16.133915 39.823189 11.166380
#> 395 198 2003-07-17 1.38086050 16.093860 39.679327 11.622084
#> 398 199 2003-07-18 4.91280391 16.052554 39.530476 15.931672
#> 400 200 2003-07-19 6.99385679 16.010029 39.376687 18.444872
#> 402 201 2003-07-20 3.36466227 15.966317 39.218012 13.936807
#> 403 202 2003-07-21 1.00323047 15.921451 39.054503 10.994062
#> 406 203 2003-07-22 5.32549010 15.875462 38.886214 16.243825
#> 407 204 2003-07-23 1.91458463 15.828384 38.713200 12.019654
#> 410 205 2003-07-24 6.26178181 15.780249 38.535519 17.279545
#> 412 206 2003-07-25 0.04395864 15.731088 38.353226  9.641893
#> 414 207 2003-07-26 0.21610383 15.680933 38.166382  9.804587
#> 415 208 2003-07-27 0.33781811 15.629816 37.975044  9.904152
#> 418 209 2003-07-28 0.37846093 15.577768 37.779274  9.903741
#> 420 210 2003-07-29 6.24418354 15.524820 37.579134 16.952070
#> 422 211 2003-07-30 0.36011785 15.471001 37.374685  9.778656
#> 423 212 2003-07-31 1.18588409 15.416342 37.165992 10.720973
#> 426 213 2003-08-01 6.08868604 15.360873 36.953119 16.561950
#> 428 214 2003-08-02 5.37204911 15.304621 36.736132 15.631377
#> 430 215 2003-08-03 6.87876059 15.247616 36.515097 17.365426
#> 431 216 2003-08-04 5.72375200 15.189886 36.290082 15.909814
#> 433 217 2003-08-05 3.23556616 15.131457 36.061156 12.870775
#> 436 218 2003-08-06 1.38520249 15.072356 35.828388 10.603474
#> 438 219 2003-08-07 1.13516957 15.012610 35.591848 10.243590
#> 440 220 2003-08-08 4.57884990 14.952243 35.351609 14.250793
#> 441 221 2003-08-09 2.86260037 14.891281 35.107744 12.151375
#> 444 222 2003-08-10 1.48811970 14.829748 34.860325 10.464145
#> 446 223 2003-08-11 2.88665677 14.767666 34.609429 12.034934
#> 448 224 2003-08-12 2.27588864 14.705060 34.355129 11.247338
#> 449 225 2003-08-13 3.02023703 14.641952 34.097505 12.041071
#> 452 226 2003-08-14 1.91527511 14.578362 33.836632 10.681851
#> 454 227 2003-08-15 0.34182251 14.514313 33.572591  8.788477
#> 456 228 2003-08-16 1.68905642 14.449824 33.305462 10.272922
#> 457 229 2003-08-17 4.37532574 14.384916 33.035326 13.282855
#> 459 230 2003-08-18 4.15792726 14.319607 32.762264 12.947090
#> 462 231 2003-08-19 3.51504194 14.253916 32.486361 12.127188
#> 464 232 2003-08-20 2.97701307 14.187862 32.207701 11.430967
#> 466 233 2003-08-21 5.77263300 14.121461 31.926370 14.507094
#> 467 234 2003-08-22 1.03401976 14.054730 31.642455  9.074596
#> 470 235 2003-08-23 5.77071359 13.987686 31.356043 14.307083
#> 472 236 2003-08-24 2.08587978 13.920345 31.067223 10.094424
#> 474 237 2003-08-25 4.12055349 13.852721 30.776085 12.271263
#> 475 238 2003-08-26 2.37878049 13.784830 30.482721 10.250807
#> 478 239 2003-08-27 3.52747393 13.716686 30.187223 11.428379
#> 480 240 2003-08-28 4.79023155 13.648302 29.889684 12.717708
#> 482 241 2003-08-29 0.10543136 13.579692 29.590199  7.512418
#> 483 242 2003-08-30 6.07479514 13.510869 29.288864 13.906687
#> 486 243 2003-08-31 3.85800195 13.441846 28.985774 11.406109
#> 488 244 2003-09-01 5.47615957 13.372633 28.681027 13.042768
#> 489 245 2003-09-02 3.14763470 13.303244 28.374722 10.450503
#> 492 246 2003-09-03 3.20644738 13.233689 28.066958 10.416971
#> 494 247 2003-09-04 2.94275668 13.163980 27.757836 10.042037
#> 496 248 2003-09-05 0.78073314 13.094126 27.447457  7.680137
#> 497 249 2003-09-06 2.75294725 13.024138 27.135924  9.651878
#> 500 250 2003-09-07 4.81818423 12.954027 26.823338 11.694236
#> 502 251 2003-09-08 2.68563869 12.883802 26.509805  9.390446
#> 504 252 2003-09-09 6.10870607 12.813472 26.195428 12.793073
#> 505 253 2003-09-10 0.26465842 12.743047 25.880314  6.738831
#> 508 254 2003-09-11 1.95264934 12.672536 25.564567  8.360702
#> 510 255 2003-09-12 0.28683350 12.601947 25.248295  6.599412
#> 512 256 2003-09-13 1.14666165 12.531289 24.931604  7.373570
#> 513 257 2003-09-14 4.72673764 12.460572 24.614602 10.822247
#> 515 258 2003-09-15 0.72138783 12.389802 24.297397  6.781699
#> 518 259 2003-09-16 1.49054584 12.318988 23.980097  7.445770
#> 520 260 2003-09-17 3.03948061 12.248139 23.662812  8.851768
#> 522 261 2003-09-18 3.67744383 12.177262 23.345649  9.361520
#> 523 262 2003-09-19 5.29623662 12.106365 23.028718 10.794428
#> 526 263 2003-09-20 4.25644613 12.035456 22.712128  9.694205
#> 528 264 2003-09-21 6.62175461 11.964544 22.395989 11.796507
#> 530 265 2003-09-22 0.78878993 11.893635 22.080409  6.252292
#> 531 266 2003-09-23 0.39913146 11.822738 21.765497  5.808772
#> 534 267 2003-09-24 5.15150318 11.751861 21.451363 10.064511
#> 536 268 2003-09-25 2.20902129 11.681012 21.138114  7.283266
#> 538 269 2003-09-26 5.01942311 11.610198 20.825859  9.708274
#> 539 270 2003-09-27 6.05193950 11.539428 20.514705 10.508222
#> 541 271 2003-09-28 6.85721121 11.468711 20.204761 11.091464
#> 544 272 2003-09-29 0.54553327 11.398053 19.896131  5.450167
#> 546 273 2003-09-30 5.56853630 11.327464 19.588923  9.712149
#> 548 274 2003-10-01 6.85283944 11.256953 19.283241 10.690292
#> 549 275 2003-10-02 0.28911170 11.186528 18.979189  4.990052
#> 552 276 2003-10-03 5.59270993 11.116198 18.676871  9.367511
#> 554 277 2003-10-04 0.40498212 11.045973 18.376389  4.930967
#> 556 278 2003-10-05 5.89758519 10.975862 18.077844  9.376284
#> 557 279 2003-10-06 1.22314244 10.905874 17.781337  5.442462
#> 560 280 2003-10-07 1.14459081 10.836020 17.486965  5.295301
#> 562 281 2003-10-08 1.56535655 10.766311 17.194826  5.548718
#> 564 282 2003-10-09 2.55513277 10.696756 16.905016  6.245304
#> 565 283 2003-10-10 6.93780649 10.627367 16.617629  9.578606
#> 568 284 2003-10-11 3.39590945 10.558154 16.332759  6.709812
#> 569 285 2003-10-12 4.21120062 10.489131 16.050495  7.234619
#> 572 286 2003-10-13 1.17079780 10.420308 15.770928  4.828722
#> 574 287 2003-10-14 6.90687267 10.351698 15.494145  9.042547
#> 576 288 2003-10-15 4.86203239 10.283314 15.220231  7.403181
#> 577 289 2003-10-16 2.83884472 10.215170 14.949271  5.814555
#> 580 290 2003-10-17 6.37888233 10.147279 14.681345  8.284902
#> 582 291 2003-10-18 0.71384971 10.079655 14.416534  4.114629
#> 584 292 2003-10-19 3.02674855 10.012314 14.154915  5.678262
#> 585 293 2003-10-20 5.00591853  9.945270 13.896562  6.971535
#> 588 294 2003-10-21 0.59270937  9.878539 13.641550  3.819632
#> 590 295 2003-10-22 5.12180450  9.812138 13.389948  6.842174
#> 592 296 2003-10-23 4.67359087  9.746084 13.141825  6.436441
#> 593 297 2003-10-24 6.34715464  9.680393 12.897247  7.452488
#> 595 298 2003-10-25 4.55498821  9.615084 12.656277  6.161921
#> 598 299 2003-10-26 0.05503942  9.550176 12.418977  3.140531
#> 600 300 2003-10-27 3.72531985  9.485687 12.185404  5.439142
#> 602 301 2003-10-28 0.89342249  9.421638 11.955615  3.555759
#> 603 302 2003-10-29 0.29341253  9.358048 11.729663  3.116302
#> 606 303 2003-10-30 3.45319860  9.294940 11.507600  5.014516
#> 608 304 2003-10-31 2.63543961  9.232334 11.289473  4.433701
#> 610 305 2003-11-01 0.49285411  9.170252 11.075328  3.066453
#> 611 306 2003-11-02 2.90120311  9.108719 10.865209  4.446632
#> 614 307 2003-11-03 5.98206789  9.047757 10.659156  6.188524
#> 616 308 2003-11-04 4.79602577  8.987390 10.457208  5.404491
#> 618 309 2003-11-05 4.59883431  8.927644 10.259398  5.207275
#> 619 310 2003-11-06 2.51137937  8.868543 10.065762  3.941643
#> 621 311 2003-11-07 2.64702610  8.810114  9.876328  3.952769
#> 624 312 2003-11-08 2.56012977  8.752384  9.691126  3.840140
#> 626 313 2003-11-09 5.03703609  8.695379  9.510180  5.132061
#> 628 314 2003-11-10 6.85480024  8.639127  9.333513  6.036262
#> 629 315 2003-11-11 1.04248484  8.583658  9.161147  2.846597
#> 632 316 2003-11-12 0.73438306  8.528999  8.993100  2.635447
#> 634 317 2003-11-13 6.11942979  8.475180  8.829388  5.394938
#> 636 318 2003-11-14 5.75815793  8.422232  8.670025  5.131292
#> 637 319 2003-11-15 1.70617585  8.370184  8.515023  2.996606
#> 640 320 2003-11-16 3.05545171  8.319067  8.364392  3.627147
#> 642 321 2003-11-17 4.49455891  8.268912  8.218139  4.288015
#> 644 322 2003-11-18 3.34700290  8.219751  8.076271  3.663357
#> 645 323 2003-11-19 2.93149448  8.171616  7.938791  3.408683
#> 648 324 2003-11-20 4.65003184  8.124538  7.805702  4.185199
#> 649 325 2003-11-21 1.49552096  8.078549  7.677004  2.629844
#> 652 326 2003-11-22 1.20670358  8.033683  7.552698  2.455403
#> 654 327 2003-11-23 6.82443730  7.989971  7.432780  5.032458
#> 656 328 2003-11-24 4.08519840  7.947446  7.317248  3.709942
#> 657 329 2003-11-25 0.26143867  7.906140  7.206095  1.920669
#> 660 330 2003-11-26 0.50577865  7.866085  7.099317  2.003068
#> 662 331 2003-11-27 0.69507027  7.827314  6.996907  2.059892
#> 664 332 2003-11-28 6.20668641  7.789856  6.898857  4.473098
#> 665 333 2003-11-29 5.34687789  7.753745  6.805157  4.047661
#> 668 334 2003-11-30 0.58470086  7.719011  6.715800  1.933305
#> 670 335 2003-12-01 5.20011287  7.685684  6.630775  3.900876
#> 672 336 2003-12-02 0.78443102  7.653793  6.550072  1.973174
#> 673 337 2003-12-03 5.63648862  7.623368  6.473680  4.011642
#> 675 338 2003-12-04 0.26845705  7.594437  6.401590  1.713543
#> 678 339 2003-12-05 2.99332023  7.567027  6.333789  2.836189
#> 680 340 2003-12-06 5.86007453  7.541163  6.270268  4.003812
#> 682 341 2003-12-07 2.18633500  7.516872  6.211016  2.456013
#> 683 342 2003-12-08 1.64351635  7.494177  6.156023  2.214031
#> 686 343 2003-12-09 1.53896590  7.473100  6.105278  2.154962
#> 688 344 2003-12-10 6.78803236  7.453664  6.058772  4.273547
#> 690 345 2003-12-11 6.09612254  7.435887  6.016497  3.970360
#> 691 346 2003-12-12 2.89731546  7.419789  5.978443  2.661856
#> 694 347 2003-12-13 5.17010975  7.405386  5.944603  3.561279
#> 696 348 2003-12-14 5.38236430  7.392694  5.914970  3.631985
#> 698 349 2003-12-15 1.88925763  7.381725  5.889539  2.226060
#> 699 350 2003-12-16 6.91680206  7.372493  5.868303  4.219869
#> 701 351 2003-12-17 6.09615712  7.365006  5.851259  3.884415
#> 704 352 2003-12-18 6.66239082  7.359273  5.838404  4.102371
#> 706 353 2003-12-19 0.27659963  7.355300  5.829736  1.567049
#> 708 354 2003-12-20 3.52696713  7.353091  5.825253  2.853377
#> 709 355 2003-12-21 0.05606027  7.352649  5.824956  1.478445
#> 712 356 2003-12-22 5.89632310  7.353975  5.828845  3.793958
#> 714 357 2003-12-23 6.26343999  7.357066  5.836924  3.943864
#> 716 358 2003-12-24 2.41077964  7.361920  5.849194  2.420005
#> 717 359 2003-12-25 2.61848106  7.368531  5.865662  2.508626
#> 720 360 2003-12-26 0.09894429  7.376892  5.886331  1.511059
#> 722 361 2003-12-27 4.47753295  7.386993  5.911209  3.269305
#> 724 362 2003-12-28 5.20409578  7.398825  5.940303  3.574185
#> 725 363 2003-12-29 2.44451003  7.412375  5.973622  2.478419
#> 728 364 2003-12-30 6.86648030  7.427627  6.011174  4.281313
#> 730 365 2003-12-31 3.87294596  7.444567  6.052971  3.087735