Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.735678  2.735678  3.676012  3.676012  3.106618  3.106618  4.154038
#>   [8]  4.154038  3.666457  3.666457  2.877112  2.877112  2.611482  2.611482
#>  [15]  2.007587  2.007587  3.709889  3.709889  2.865151  2.865151  2.604388
#>  [22]  2.604388  3.624282  3.624282  4.820612  4.820612  4.379906  4.379906
#>  [29]  3.001804  3.001804  3.165302  3.165302  1.997430  1.997430  4.767174
#>  [36]  4.767174  3.378960  3.378960  2.737442  2.737442  5.015083  5.015083
#>  [43]  3.596000  3.596000  3.839520  3.839520  3.192823  3.192823  4.458520
#>  [50]  4.458520  3.258437  3.258437  4.854255  4.854255  3.538395  3.538395
#>  [57]  3.337293  3.337293  4.006907  4.006907  6.196801  6.196801  6.057339
#>  [64]  6.057339  5.187439  5.187439  3.769477  3.769477  4.221469  4.221469
#>  [71]  6.487058  6.487058  2.987466  2.987466  4.270171  4.270171  6.400080
#>  [78]  6.400080  5.604940  5.604940  6.636443  6.636443  4.129318  4.129318
#>  [85]  4.091666  4.091666  6.354158  6.354158  4.741290  4.741290  6.288063
#>  [92]  6.288063  5.291911  5.291911  4.864335  4.864335  6.418776  6.418776
#>  [99]  4.822634  4.822634  8.408726  8.408726  4.314439  4.314439  6.236394
#> [106]  6.236394  3.820156  3.820156  5.279642  5.279642  7.234768  7.234768
#> [113]  6.472826  6.472826  6.681893  6.681893  5.168733  5.168733  9.461668
#> [120]  9.461668  7.648504  7.648504  4.415058  4.415058  9.346234  9.346234
#> [127]  8.492798  8.492798  9.143190  9.143190  7.931062  7.931062  9.182201
#> [134]  9.182201  7.202191  7.202191  7.719221  7.719221  5.300490  5.300490
#> [141]  7.974076  7.974076  9.181830  9.181830  7.249182  7.249182  5.384028
#> [148]  5.384028  6.386149  6.386149  5.832776  5.832776  7.393012  7.393012
#> [155]  7.393943  7.393943  8.371699  8.371699  7.698508  7.698508 11.995744
#> [162] 11.995744 11.916873 11.916873  6.845070  6.845070 12.576889 12.576889
#> [169]  7.435293  7.435293 12.839670 12.839670 12.067591 12.067591  8.010368
#> [176]  8.010368 11.575078 11.575078 11.477049 11.477049 12.033309 12.033309
#> [183] 12.453494 12.453494 10.906280 10.906280  8.820344  8.820344  7.205002
#> [190]  7.205002  8.220220  8.220220 11.535451 11.535451  9.906982  9.906982
#> [197] 10.556076 10.556076  8.896340  8.896340  7.743363  7.743363 14.514880
#> [204] 14.514880 14.225545 14.225545  8.474890  8.474890 11.317501 11.317501
#> [211]  8.366592  8.366592 15.277323 15.277323  9.368330  9.368330  9.706231
#> [218]  9.706231  8.891679  8.891679 10.108329 10.108329 11.037949 11.037949
#> [225] 15.509273 15.509273 10.499345 10.499345 10.572005 10.572005 12.744740
#> [232] 12.744740  8.531374  8.531374 13.246600 13.246600 11.045213 11.045213
#> [239] 10.348484 10.348484  9.453250  9.453250 13.901026 13.901026 11.005563
#> [246] 11.005563 12.962948 12.962948 11.668831 11.668831 13.248188 13.248188
#> [253] 13.913537 13.913537  9.604891  9.604891 10.383918 10.383918  9.276187
#> [260]  9.276187 17.452418 17.452418 14.718656 14.718656 12.952153 12.952153
#> [267] 17.387446 17.387446 17.379861 17.379861 17.682104 17.682104 11.694973
#> [274] 11.694973 11.835437 11.835437 17.419476 17.419476 11.213088 11.213088
#> [281] 14.660034 14.660034 17.011164 17.011164 13.882057 13.882057 18.328212
#> [288] 18.328212 18.152856 18.152856 15.178357 15.178357 13.628276 13.628276
#> [295] 12.141690 12.141690 16.058951 16.058951 10.591897 10.591897 13.230554
#> [302] 13.230554 18.214922 18.214922 14.932996 14.932996 14.523777 14.523777
#> [309] 14.885373 14.885373 16.735603 16.735603 12.938222 12.938222 17.171372
#> [316] 17.171372 10.643442 10.643442 11.312232 11.312232 11.550678 11.550678
#> [323] 10.724217 10.724217 11.655693 11.655693 12.285467 12.285467 15.189627
#> [330] 15.189627 13.541718 13.541718 15.070025 15.070025 11.895651 11.895651
#> [337] 10.909191 10.909191 16.730062 16.730062 15.227599 15.227599 15.303045
#> [344] 15.303045 11.866724 11.866724 10.799479 10.799479 16.450637 16.450637
#> [351] 18.548763 18.548763 12.117024 12.117024 12.129077 12.129077 13.667140
#> [358] 13.667140 11.990896 11.990896 11.782785 11.782785 11.565775 11.565775
#> [365] 17.727952 17.727952 17.087247 17.087247 14.083346 14.083346 11.440416
#> [372] 11.440416 17.892875 17.892875 10.877842 10.877842 17.911057 17.911057
#> [379] 15.476553 15.476553 15.373408 15.373408 11.853679 11.853679 18.129564
#> [386] 18.129564 17.112823 17.112823 13.340376 13.340376 12.768296 12.768296
#> [393] 10.362815 10.362815 15.112348 15.112348 11.088014 11.088014 12.260393
#> [400] 12.260393 17.340500 17.340500 11.628410 11.628410 17.850179 17.850179
#> [407] 17.906096 17.906096  9.771092  9.771092 10.748428 10.748428 13.891213
#> [414] 13.891213 13.454756 13.454756 12.734044 12.734044 10.996859 10.996859
#> [421] 16.733587 16.733587 11.931697 11.931697 14.149504 14.149504 17.493616
#> [428] 17.493616 14.495313 14.495313 10.669328 10.669328  9.983056  9.983056
#> [435] 14.025814 14.025814 14.655989 14.655989 12.358753 12.358753 13.447817
#> [442] 13.447817 11.266086 11.266086  9.216949  9.216949 12.897319 12.897319
#> [449]  9.783254  9.783254 13.012957 13.012957  8.726946  8.726946 15.741486
#> [456] 15.741486 12.094798 12.094798 12.900051 12.900051  8.648771  8.648771
#> [463]  8.620432  8.620432 14.651278 14.651278 14.744451 14.744451  8.163370
#> [470]  8.163370 11.585188 11.585188 13.810990 13.810990 11.548346 11.548346
#> [477] 11.080373 11.080373 12.353039 12.353039  7.407932  7.407932 11.473361
#> [484] 11.473361  8.678290  8.678290  7.203683  7.203683  7.913729  7.913729
#> [491]  7.775104  7.775104 10.085386 10.085386 13.035650 13.035650 10.302447
#> [498] 10.302447 10.880244 10.880244 11.984413 11.984413 12.488537 12.488537
#> [505] 13.318908 13.318908 10.239464 10.239464 12.631421 12.631421  8.453247
#> [512]  8.453247  6.967615  6.967615 11.964009 11.964009  6.698539  6.698539
#> [519] 12.495029 12.495029  6.600998  6.600998  7.387473  7.387473 10.588048
#> [526] 10.588048 11.882820 11.882820 10.019194 10.019194  5.714731  5.714731
#> [533] 10.009174 10.009174  9.525155  9.525155 10.935951 10.935951  5.460944
#> [540]  5.460944  6.816442  6.816442  4.980523  4.980523 10.742565 10.742565
#> [547]  7.240295  7.240295  4.922350  4.922350  8.837734  8.837734  7.069620
#> [554]  7.069620  9.566458  9.566458  9.891889  9.891889  8.735787  8.735787
#> [561]  7.654932  7.654932  4.365043  4.365043  8.949802  8.949802  6.350134
#> [568]  6.350134  4.304508  4.304508  7.890870  7.890870  5.946313  5.946313
#> [575]  4.436300  4.436300  8.652491  8.652491  5.072764  5.072764  5.624095
#> [582]  5.624095  7.855965  7.855965  3.480780  3.480780  5.561872  5.561872
#> [589]  5.612592  5.612592  4.721255  4.721255  7.340060  7.340060  3.264549
#> [596]  3.264549  5.147828  5.147828  4.294784  4.294784  5.693551  5.693551
#> [603]  4.111773  4.111773  5.806002  5.806002  3.941567  3.941567  6.073434
#> [610]  6.073434  4.989320  4.989320  2.896776  2.896776  4.329432  4.329432
#> [617]  5.127070  5.127070  2.668324  2.668324  4.958364  4.958364  4.127770
#> [624]  4.127770  4.744564  4.744564  3.046748  3.046748  4.969641  4.969641
#> [631]  2.864549  2.864549  4.309481  4.309481  2.572115  2.572115  5.015953
#> [638]  5.015953  4.602114  4.602114  4.420910  4.420910  3.555219  3.555219
#> [645]  5.060305  5.060305  4.393895  4.393895  2.833167  2.833167  2.385012
#> [652]  2.385012  3.674419  3.674419  2.572589  2.572589  2.872442  2.872442
#> [659]  3.831618  3.831618  4.142239  4.142239  3.799749  3.799749  3.046721
#> [666]  3.046721  2.465315  2.465315  2.696037  2.696037  3.352552  3.352552
#> [673]  2.987115  2.987115  2.911233  2.911233  2.475722  2.475722  1.883110
#> [680]  1.883110  3.839949  3.839949  2.998218  2.998218  3.657013  3.657013
#> [687]  3.709117  3.709117  3.388649  3.388649  2.293120  2.293120  4.210548
#> [694]  4.210548  3.360581  3.360581  4.028967  4.028967  2.123874  2.123874
#> [701]  4.237796  4.237796  2.791257  2.791257  3.365490  3.365490  3.219816
#> [708]  3.219816  2.133013  2.133013  1.712925  1.712925  3.694585  3.694585
#> [715]  2.723765  2.723765  3.658641  3.658641  2.238277  2.238277  3.587238
#> [722]  3.587238  2.184467  2.184467  2.567997  2.567997  1.838176  1.838176
#> [729]  3.912588  3.912588