Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.287963  2.287963  2.824376  2.824376  4.254937  4.254937  2.218337
#>   [8]  2.218337  3.739679  3.739679  3.586744  3.586744  3.768777  3.768777
#>  [15]  2.070965  2.070965  2.779137  2.779137  4.720502  4.720502  3.637487
#>  [22]  3.637487  2.359784  2.359784  4.223581  4.223581  2.626233  2.626233
#>  [29]  1.800651  1.800651  1.929087  1.929087  2.763984  2.763984  3.115799
#>  [36]  3.115799  4.300453  4.300453  4.909910  4.909910  3.537834  3.537834
#>  [43]  4.179474  4.179474  5.321812  5.321812  4.737429  4.737429  2.885697
#>  [50]  2.885697  5.258157  5.258157  4.256704  4.256704  3.579407  3.579407
#>  [57]  5.760896  5.760896  3.963938  3.963938  5.511237  5.511237  5.498118
#>  [64]  5.498118  2.567021  2.567021  5.987840  5.987840  5.916693  5.916693
#>  [71]  4.657164  4.657164  3.567309  3.567309  2.733163  2.733163  5.167012
#>  [78]  5.167012  7.054883  7.054883  5.834123  5.834123  5.169167  5.169167
#>  [85]  7.432358  7.432358  7.387388  7.387388  7.142641  7.142641  3.709290
#>  [92]  3.709290  5.576773  5.576773  5.617596  5.617596  6.075948  6.075948
#>  [99]  7.231117  7.231117  5.146657  5.146657  5.114931  5.114931  6.955408
#> [106]  6.955408  7.647003  7.647003  8.198921  8.198921  7.898574  7.898574
#> [113]  8.503751  8.503751  7.279030  7.279030  8.108354  8.108354  7.844457
#> [120]  7.844457  7.622269  7.622269  6.773832  6.773832  4.620480  4.620480
#> [127]  6.405002  6.405002  4.929910  4.929910  7.887659  7.887659  7.883456
#> [134]  7.883456  6.947675  6.947675 10.562326 10.562326  5.920777  5.920777
#> [141]  9.250739  9.250739  9.711196  9.711196  8.014253  8.014253  5.295548
#> [148]  5.295548  5.648545  5.648545  7.406689  7.406689  9.577202  9.577202
#> [155] 11.989501 11.989501  7.914883  7.914883  9.956097  9.956097  6.334357
#> [162]  6.334357  5.923806  5.923806 12.217250 12.217250 12.765200 12.765200
#> [169] 11.350851 11.350851 10.785272 10.785272  7.765483  7.765483  9.658127
#> [176]  9.658127 10.853581 10.853581 12.415976 12.415976  6.899951  6.899951
#> [183]  9.715354  9.715354  9.107686  9.107686  9.761574  9.761574 10.259762
#> [190] 10.259762  9.188532  9.188532 12.266154 12.266154  7.971501  7.971501
#> [197] 11.828127 11.828127  8.784262  8.784262  8.211366  8.211366  8.667910
#> [204]  8.667910  8.988383  8.988383 10.464670 10.464670 10.749814 10.749814
#> [211]  9.526527  9.526527 12.934004 12.934004 10.423067 10.423067  8.558891
#> [218]  8.558891 12.423659 12.423659 11.657818 11.657818 14.827980 14.827980
#> [225]  8.668897  8.668897 16.247099 16.247099 12.316660 12.316660 10.456418
#> [232] 10.456418 16.114529 16.114529 12.158718 12.158718 11.916113 11.916113
#> [239] 16.105940 16.105940  9.006812  9.006812 13.043048 13.043048 12.520940
#> [246] 12.520940  9.376839  9.376839 12.275992 12.275992 11.194382 11.194382
#> [253] 14.143665 14.143665 13.707883 13.707883 13.316736 13.316736 15.114844
#> [260] 15.114844 16.346065 16.346065  9.613937  9.613937 10.764486 10.764486
#> [267] 10.499379 10.499379  9.666636  9.666636 17.501254 17.501254 17.523228
#> [274] 17.523228 17.281564 17.281564  9.956905  9.956905 10.841052 10.841052
#> [281] 10.607322 10.607322 16.534960 16.534960 16.375484 16.375484 15.390766
#> [288] 15.390766 17.565735 17.565735 10.592411 10.592411 17.973457 17.973457
#> [295] 15.011287 15.011287 14.935398 14.935398 12.310167 12.310167 18.256578
#> [302] 18.256578 15.641677 15.641677 16.929943 16.929943 15.956107 15.956107
#> [309] 16.162597 16.162597 12.257608 12.257608 15.820945 15.820945 18.003715
#> [316] 18.003715 15.824627 15.824627 18.179129 18.179129 12.909577 12.909577
#> [323] 16.568252 16.568252 12.710062 12.710062 14.359273 14.359273 15.472497
#> [330] 15.472497 13.342024 13.342024 13.864999 13.864999 12.667526 12.667526
#> [337] 12.628053 12.628053 14.681053 14.681053 18.062797 18.062797 14.809544
#> [344] 14.809544 14.403973 14.403973 17.463022 17.463022 13.829141 13.829141
#> [351] 12.143553 12.143553 10.893120 10.893120 15.976663 15.976663 18.924631
#> [358] 18.924631 10.410056 10.410056 14.877758 14.877758 18.946780 18.946780
#> [365] 11.629286 11.629286 17.308085 17.308085 14.532809 14.532809 10.727279
#> [372] 10.727279 17.181045 17.181045 17.321743 17.321743 18.790307 18.790307
#> [379] 16.235771 16.235771 13.918955 13.918955 14.858002 14.858002 16.816481
#> [386] 16.816481 15.975388 15.975388 10.057232 10.057232 14.004272 14.004272
#> [393] 12.310971 12.310971 17.401088 17.401088 14.069998 14.069998 18.025497
#> [400] 18.025497 17.622266 17.622266 13.835070 13.835070 15.717762 15.717762
#> [407] 17.515638 17.515638 14.889305 14.889305 11.010836 11.010836  9.634404
#> [414]  9.634404 16.932637 16.932637 13.004442 13.004442 10.196788 10.196788
#> [421] 11.250391 11.250391 12.347052 12.347052 15.651190 15.651190 14.462236
#> [428] 14.462236 11.593232 11.593232  9.734494  9.734494 17.030020 17.030020
#> [435]  9.450369  9.450369 14.398995 14.398995  9.845560  9.845560 15.308876
#> [442] 15.308876  9.809296  9.809296 16.464317 16.464317 16.011055 16.011055
#> [449] 11.040233 11.040233 14.956168 14.956168 12.253545 12.253545 11.286034
#> [456] 11.286034 16.036868 16.036868 10.357421 10.357421 10.524904 10.524904
#> [463] 13.004416 13.004416 15.442344 15.442344  8.438274  8.438274 14.039566
#> [470] 14.039566  9.533721  9.533721 13.465038 13.465038 14.190042 14.190042
#> [477]  7.649696  7.649696 14.719564 14.719564  7.704405  7.704405  9.160885
#> [484]  9.160885 13.248976 13.248976 11.450044 11.450044 13.001314 13.001314
#> [491]  8.613554  8.613554 12.538739 12.538739  7.547392  7.547392  8.747937
#> [498]  8.747937  7.307563  7.307563 11.312368 11.312368 12.797931 12.797931
#> [505] 13.244639 13.244639 12.287079 12.287079  8.586072  8.586072  7.744034
#> [512]  7.744034 10.763832 10.763832 12.488760 12.488760 11.635384 11.635384
#> [519]  9.871404  9.871404  6.634920  6.634920  9.772878  9.772878  7.741697
#> [526]  7.741697 11.389241 11.389241  7.073463  7.073463 10.240416 10.240416
#> [533]  9.747586  9.747586  6.621642  6.621642  7.989123  7.989123  9.149032
#> [540]  9.149032 10.850779 10.850779  8.016570  8.016570 10.048414 10.048414
#> [547]  7.505273  7.505273 10.351734 10.351734  7.001331  7.001331 10.136796
#> [554] 10.136796  4.768316  4.768316  6.952953  6.952953  4.686712  4.686712
#> [561]  6.872464  6.872464  7.247519  7.247519  6.374008  6.374008  5.361130
#> [568]  5.361130  6.802771  6.802771  5.247939  5.247939  5.627343  5.627343
#> [575]  5.463031  5.463031  4.920236  4.920236  6.574156  6.574156  3.979336
#> [582]  3.979336  7.331112  7.331112  7.877486  7.877486  3.859016  3.859016
#> [589]  3.495391  3.495391  5.263587  5.263587  3.897721  3.897721  4.755819
#> [596]  4.755819  5.346658  5.346658  6.788058  6.788058  3.375757  3.375757
#> [603]  3.728425  3.728425  5.402078  5.402078  6.388014  6.388014  5.480785
#> [610]  5.480785  3.540880  3.540880  4.148759  4.148759  3.263723  3.263723
#> [617]  3.040754  3.040754  5.414158  5.414158  3.853836  3.853836  4.042849
#> [624]  4.042849  6.123231  6.123231  5.394630  5.394630  3.450952  3.450952
#> [631]  5.191521  5.191521  3.855989  3.855989  4.340194  4.340194  4.100765
#> [638]  4.100765  5.489035  5.489035  5.037617  5.037617  2.086136  2.086136
#> [645]  4.388707  4.388707  2.293009  2.293009  2.655185  2.655185  3.489559
#> [652]  3.489559  4.193603  4.193603  3.853804  3.853804  3.353656  3.353656
#> [659]  4.510284  4.510284  3.356705  3.356705  3.176984  3.176984  4.144375
#> [666]  4.144375  2.838078  2.838078  3.629093  3.629093  1.764410  1.764410
#> [673]  3.813118  3.813118  4.120915  4.120915  3.412673  3.412673  2.560744
#> [680]  2.560744  3.408353  3.408353  3.849592  3.849592  2.208444  2.208444
#> [687]  1.918773  1.918773  1.685429  1.685429  2.586591  2.586591  3.508929
#> [694]  3.508929  1.587509  1.587509  3.058595  3.058595  3.049389  3.049389
#> [701]  2.331757  2.331757  2.004958  2.004958  3.912312  3.912312  4.042193
#> [708]  4.042193  3.951196  3.951196  3.770284  3.770284  2.584587  2.584587
#> [715]  1.540727  1.540727  3.894628  3.894628  4.104089  4.104089  2.098412
#> [722]  2.098412  1.751925  1.751925  3.083384  3.083384  2.614970  2.614970
#> [729]  2.144704  2.144704