Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.641749  3.642171  4.401396  2.078757  3.987946  2.201706  3.653223
#>   [8]  3.954826  2.900536  1.858373  2.790763  2.667350  4.568607  4.481355
#>  [15]  4.825436  4.227077  2.758322  3.424081  4.231798  2.273533  4.697429
#>  [22]  2.991800  2.992137  2.096246  2.431080  3.290299  5.098569  3.384217
#>  [29]  2.951720  3.334733  2.559723  5.968988  5.996202  6.056197  3.302512
#>  [36]  5.987866  3.280038  4.182624  6.684077  6.515611  3.226001  6.625963
#>  [43]  5.421219  6.504521  5.125212  4.237543  4.767346  6.409812  4.628236
#>  [50]  7.839415  7.576698  3.942787  3.747189  5.422319  5.574634  4.202562
#>  [57]  5.274324  5.340090  4.322203  6.612290  5.013219  7.086603  8.420248
#>  [64]  9.290502  7.877576  6.599613  6.703152  6.534819  8.307595  6.441160
#>  [71]  9.803423  9.832394  5.689361  8.406420 10.912811  8.211626  8.123507
#>  [78] 10.728931 11.797974  9.664122 11.835076  7.374797  9.366149  7.897041
#>  [85] 12.815783 12.930270  6.470922 11.126121  6.923038  9.540647  6.670017
#>  [92] 12.833090 11.217133 10.149828 12.189199  7.755276  7.578079  7.417942
#>  [99] 10.969729 11.455896  8.050885 12.456048 14.253481  8.205710  8.917162
#> [106] 12.328464 10.778143  7.885379  8.875025  8.390236 16.049429 13.414474
#> [113] 15.114952 10.707823 12.717248 10.354317 13.213369 11.892709 16.026115
#> [120] 15.806163 13.162240 12.793892 15.606113 13.182654 15.519530 14.787761
#> [127] 14.224169 16.892953 17.038152 12.509986 12.231917 15.060028 17.775847
#> [134] 10.418459  9.889063 16.647243 11.196215 16.526296 12.166447 11.887846
#> [141] 17.949102 12.415232 12.786876 18.544377 17.546161 16.206956 12.478067
#> [148] 16.599207 17.491079 12.863884 11.041489 16.906989 14.806433 13.578379
#> [155] 15.196017 12.935113 13.614199 13.239888 11.996017 18.339156 15.039379
#> [162] 13.263145 14.753433 14.599646 16.042258 17.102231 17.910808 14.873481
#> [169] 17.510371 12.548210 14.069720 15.941186 14.347056 14.401713 12.387365
#> [176] 15.762378 13.070720 11.967159 11.004106 10.677538 13.275499 17.910748
#> [183] 18.871914 13.080867 15.637171 10.851112 18.140296 14.019565 16.425369
#> [190] 16.631009 12.059226 14.802180 16.860697 16.094654 13.299854 13.382824
#> [197] 15.372163 14.356180 17.641142 11.716108 14.708506 10.355856 16.425888
#> [204] 14.563254  9.776595 14.380041  9.924177 12.743208 16.617768 14.019232
#> [211] 11.326076 11.555011 14.391667 11.157836 16.926236 16.626466 15.050680
#> [218] 12.258319 16.420800 14.190563 16.878783 10.481836  8.968711 16.263768
#> [225] 10.822221 15.344868 13.098000 13.547849 11.048180 14.191112 10.859703
#> [232] 11.533399  9.695832 13.009501 10.843228 10.668699  9.632946  8.554413
#> [239]  7.824717 12.687364  9.273408  9.670464 13.939420 13.295487 10.652831
#> [246] 13.782085  8.757742 10.124064  7.795828 12.752006 13.526760 12.580522
#> [253] 10.965563  8.691927 12.534311  9.176351  8.809113  6.672715 11.315187
#> [260] 11.948837 11.186048  6.124991  8.629782 10.890510 11.098085  8.062560
#> [267]  9.287142 10.600602  8.205633  9.723192  8.652428  5.011181  8.129885
#> [274]  9.378070  6.442680  8.130182  6.747466  5.662553  7.618408  5.689005
#> [281]  5.531735  4.878693  8.795309  8.157273  5.703923  6.609016  4.937943
#> [288]  8.843469  5.960482  7.500486  4.203757  6.319697  4.571766  4.919903
#> [295]  7.023321  7.626407  5.025073  5.901600  6.949463  6.614490  5.436053
#> [302]  3.313175  5.354745  2.925645  6.038094  6.429256  6.395169  2.658107
#> [309]  6.217426  4.073423  5.897919  6.219359  3.997693  4.982870  4.520877
#> [316]  5.491125  2.663806  3.491082  2.430554  4.831246  4.326232  5.109378
#> [323]  2.440918  4.409168  4.062134  4.428769  3.938202  4.521463  3.314970
#> [330]  2.686781  3.738054  3.531827  2.561374  4.205357  4.050178  2.086138
#> [337]  2.564509  1.748040  3.379914  2.022611  3.443444  2.282742  2.416499
#> [344]  4.062526  2.410355  4.256480  3.393659  1.794211  3.242771  4.105606
#> [351]  3.418031  3.841242  3.713500  2.612222  2.760731  3.939602  3.960965
#> [358]  2.865940  2.624235  3.735225  3.188088  2.805502  3.549711  2.785490
#> [365]  4.123010  1.641749  3.642171  4.401396  2.078757  3.987946  2.201706
#> [372]  3.653223  3.954826  2.900536  1.858373  2.790763  2.667350  4.568607
#> [379]  4.481355  4.825436  4.227077  2.758322  3.424081  4.231798  2.273533
#> [386]  4.697429  2.991800  2.992137  2.096246  2.431080  3.290299  5.098569
#> [393]  3.384217  2.951720  3.334733  2.559723  5.968988  5.996202  6.056197
#> [400]  3.302512  5.987866  3.280038  4.182624  6.684077  6.515611  3.226001
#> [407]  6.625963  5.421219  6.504521  5.125212  4.237543  4.767346  6.409812
#> [414]  4.628236  7.839415  7.576698  3.942787  3.747189  5.422319  5.574634
#> [421]  4.202562  5.274324  5.340090  4.322203  6.612290  5.013219  7.086603
#> [428]  8.420248  9.290502  7.877576  6.599613  6.703152  6.534819  8.307595
#> [435]  6.441160  9.803423  9.832394  5.689361  8.406420 10.912811  8.211626
#> [442]  8.123507 10.728931 11.797974  9.664122 11.835076  7.374797  9.366149
#> [449]  7.897041 12.815783 12.930270  6.470922 11.126121  6.923038  9.540647
#> [456]  6.670017 12.833090 11.217133 10.149828 12.189199  7.755276  7.578079
#> [463]  7.417942 10.969729 11.455896  8.050885 12.456048 14.253481  8.205710
#> [470]  8.917162 12.328464 10.778143  7.885379  8.875025  8.390236 16.049429
#> [477] 13.414474 15.114952 10.707823 12.717248 10.354317 13.213369 11.892709
#> [484] 16.026115 15.806163 13.162240 12.793892 15.606113 13.182654 15.519530
#> [491] 14.787761 14.224169 16.892953 17.038152 12.509986 12.231917 15.060028
#> [498] 17.775847 10.418459  9.889063 16.647243 11.196215 16.526296 12.166447
#> [505] 11.887846 17.949102 12.415232 12.786876 18.544377 17.546161 16.206956
#> [512] 12.478067 16.599207 17.491079 12.863884 11.041489 16.906989 14.806433
#> [519] 13.578379 15.196017 12.935113 13.614199 13.239888 11.996017 18.339156
#> [526] 15.039379 13.263145 14.753433 14.599646 16.042258 17.102231 17.910808
#> [533] 14.873481 17.510371 12.548210 14.069720 15.941186 14.347056 14.401713
#> [540] 12.387365 15.762378 13.070720 11.967159 11.004106 10.677538 13.275499
#> [547] 17.910748 18.871914 13.080867 15.637171 10.851112 18.140296 14.019565
#> [554] 16.425369 16.631009 12.059226 14.802180 16.860697 16.094654 13.299854
#> [561] 13.382824 15.372163 14.356180 17.641142 11.716108 14.708506 10.355856
#> [568] 16.425888 14.563254  9.776595 14.380041  9.924177 12.743208 16.617768
#> [575] 14.019232 11.326076 11.555011 14.391667 11.157836 16.926236 16.626466
#> [582] 15.050680 12.258319 16.420800 14.190563 16.878783 10.481836  8.968711
#> [589] 16.263768 10.822221 15.344868 13.098000 13.547849 11.048180 14.191112
#> [596] 10.859703 11.533399  9.695832 13.009501 10.843228 10.668699  9.632946
#> [603]  8.554413  7.824717 12.687364  9.273408  9.670464 13.939420 13.295487
#> [610] 10.652831 13.782085  8.757742 10.124064  7.795828 12.752006 13.526760
#> [617] 12.580522 10.965563  8.691927 12.534311  9.176351  8.809113  6.672715
#> [624] 11.315187 11.948837 11.186048  6.124991  8.629782 10.890510 11.098085
#> [631]  8.062560  9.287142 10.600602  8.205633  9.723192  8.652428  5.011181
#> [638]  8.129885  9.378070  6.442680  8.130182  6.747466  5.662553  7.618408
#> [645]  5.689005  5.531735  4.878693  8.795309  8.157273  5.703923  6.609016
#> [652]  4.937943  8.843469  5.960482  7.500486  4.203757  6.319697  4.571766
#> [659]  4.919903  7.023321  7.626407  5.025073  5.901600  6.949463  6.614490
#> [666]  5.436053  3.313175  5.354745  2.925645  6.038094  6.429256  6.395169
#> [673]  2.658107  6.217426  4.073423  5.897919  6.219359  3.997693  4.982870
#> [680]  4.520877  5.491125  2.663806  3.491082  2.430554  4.831246  4.326232
#> [687]  5.109378  2.440918  4.409168  4.062134  4.428769  3.938202  4.521463
#> [694]  3.314970  2.686781  3.738054  3.531827  2.561374  4.205357  4.050178
#> [701]  2.086138  2.564509  1.748040  3.379914  2.022611  3.443444  2.282742
#> [708]  2.416499  4.062526  2.410355  4.256480  3.393659  1.794211  3.242771
#> [715]  4.105606  3.418031  3.841242  3.713500  2.612222  2.760731  3.939602
#> [722]  3.960965  2.865940  2.624235  3.735225  3.188088  2.805502  3.549711
#> [729]  2.785490  4.123010
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 2.31982961  7.463176  6.099033  2.472660
#> 3     2 2002-01-02 0.37740624  7.483435  6.149345  1.692399
#> 6     3 2002-01-03 5.39469713  7.505323  6.203934  3.780624
#> 7     4 2002-01-04 3.29654447  7.528820  6.262814  2.936811
#> 9     5 2002-01-05 6.13985549  7.553900  6.325996  4.152403
#> 11    6 2002-01-06 3.98233910  7.580540  6.393496  3.277744
#> 14    7 2002-01-07 3.25216836  7.608714  6.465324  2.998057
#> 15    8 2002-01-08 1.38528470  7.638396  6.541496  2.228550
#> 17    9 2002-01-09 1.50048071  7.669557  6.622023  2.303276
#> 19   10 2002-01-10 0.43546683  7.702170  6.706920  1.866328
#> 22   11 2002-01-11 1.44677032  7.736204  6.796198  2.334538
#> 23   12 2002-01-12 6.32413442  7.771631  6.889871  4.525771
#> 25   13 2002-01-13 4.90925353  7.808419  6.987950  3.943695
#> 27   14 2002-01-14 0.01716437  7.846537  7.090446  1.780367
#> 30   15 2002-01-15 2.41108846  7.885954  7.197371  2.899621
#> 32   16 2002-01-16 1.34632829  7.926639  7.308732  2.447872
#> 33   17 2002-01-17 3.77833869  7.968559  7.424539  3.616329
#> 35   18 2002-01-18 3.95555264  8.011681  7.544799  3.748721
#> 37   19 2002-01-19 0.24293017  8.055974  7.669519  2.033018
#> 40   20 2002-01-20 6.31002067  8.101405  7.798703  4.986801
#> 41   21 2002-01-21 6.24165115  8.147942  7.932354  5.021339
#> 43   22 2002-01-22 1.99516932  8.195553  8.070474  2.999978
#> 45   23 2002-01-23 6.41847176  8.244205  8.213063  5.250379
#> 48   24 2002-01-24 1.22896735  8.293867  8.360119  2.709422
#> 49   25 2002-01-25 2.56352263  8.344507  8.511639  3.435344
#> 51   26 2002-01-26 2.23909882  8.396093  8.667618  3.322659
#> 53   27 2002-01-27 5.20698487  8.448595  8.828046  4.927434
#> 56   28 2002-01-28 3.68550512  8.501983  8.992915  4.197388
#> 58   29 2002-01-29 0.84714189  8.556225  9.162213  2.744123
#> 59   30 2002-01-30 4.70549184  8.611293  9.335925  4.884709
#> 61   31 2002-01-31 3.59263422  8.667157  9.514035  4.350346
#> 63   32 2002-02-01 0.69916822  8.723789  9.696523  2.812695
#> 66   33 2002-02-02 5.89606197  8.781160  9.883368  5.788909
#> 67   34 2002-02-03 3.23854344  8.839243 10.074546  4.364205
#> 69   35 2002-02-04 4.18834701  8.898011 10.270030  4.984590
#> 71   36 2002-02-05 1.67306024  8.957438 10.469790  3.595215
#> 74   37 2002-02-06 0.19174598  9.017498 10.673795  2.781931
#> 75   38 2002-02-07 6.29735015  9.078165 10.882010  6.494824
#> 77   39 2002-02-08 6.04013345  9.139416 11.094398  6.439679
#> 79   40 2002-02-09 2.95545410  9.201226 11.310918  4.644276
#> 82   41 2002-02-10 2.01245718  9.263572 11.531528  4.135461
#> 83   42 2002-02-11 3.10622049  9.326432 11.756182  4.896777
#> 86   43 2002-02-12 0.79812337  9.389784 11.984831  3.505558
#> 87   44 2002-02-13 3.51148390  9.453606 12.217425  5.323400
#> 89   45 2002-02-14 0.58087339  9.517878 12.453910  3.493507
#> 91   46 2002-02-15 3.20562213  9.582579 12.694229  5.296832
#> 94   47 2002-02-16 0.86947721  9.647690 12.938322  3.817600
#> 95   48 2002-02-17 0.81480127  9.713192 13.186128  3.849598
#> 97   49 2002-02-18 4.61605587  9.779067 13.437583  6.530896
#> 99   50 2002-02-19 1.15747411  9.845297 13.692618  4.228049
#> 102  51 2002-02-20 6.80609747  9.911864 13.951164  8.277656
#> 103  52 2002-02-21 5.83591628  9.978754 14.213148  7.709454
#> 105  53 2002-02-22 0.65508159 10.045948 14.478496  4.091685
#> 107  54 2002-02-23 3.99584588 10.113433 14.747130  6.600099
#> 110  55 2002-02-24 3.73801083 10.181192 15.018971  6.511840
#> 112  56 2002-02-25 0.79868082 10.249211 15.293936  4.419382
#> 113  57 2002-02-26 2.82367195 10.317477 15.571940  6.023838
#> 115  58 2002-02-27 5.47075880 10.385975 15.852898  8.138440
#> 117  59 2002-02-28 0.88137666 10.454693 16.136720  4.714378
#> 120  60 2002-03-01 5.94351226 10.523618 16.423316  8.743596
#> 121  61 2002-03-02 0.64098230 10.592738 16.712592  4.683800
#> 123  62 2002-03-03 0.36244195 10.662040 17.004454  4.540135
#> 125  63 2002-03-04 4.61734403 10.731513 17.298805  8.046195
#> 128  64 2002-03-05 3.46967307 10.801147 17.595546  7.225012
#> 129  65 2002-03-06 3.82658430 10.870930 17.894578  7.623104
#> 131  66 2002-03-07 2.19606105 10.940852 18.195799  6.375091
#> 133  67 2002-03-08 2.28912471 11.010902 18.499104  6.547723
#> 136  68 2002-03-09 3.55853999 11.081072 18.804391  7.720489
#> 138  69 2002-03-10 2.87123565 11.151350 19.111552  7.238297
#> 139  70 2002-03-11 4.24175578 11.221729 19.420480  8.525541
#> 141  71 2002-03-12 3.36703035 11.292198 19.731067  7.874404
#> 143  72 2002-03-13 1.77457315 11.362749 20.043203  6.575921
#> 146  73 2002-03-14 5.09506915 11.433374 20.356778  9.625003
#> 147  74 2002-03-15 2.42814351 11.504062 20.671679  7.349488
#> 149  75 2002-03-16 2.40277040 11.574807 20.987796  7.425337
#> 151  76 2002-03-17 1.65151918 11.645600 21.305014  6.836937
#> 154  77 2002-03-18 2.05384672 11.716433 21.623220  7.301040
#> 155  78 2002-03-19 6.98850504 11.787297 21.942300 11.990199
#> 157  79 2002-03-20 0.13760043 11.858185 22.262140  5.694698
#> 159  80 2002-03-21 3.48833681 11.929088 22.582624  8.947492
#> 162  81 2002-03-22 2.48763280 12.000000 22.903638  8.099903
#> 163  82 2002-03-23 0.63211065 12.070912 23.225065  6.414373
#> 166  83 2002-03-24 2.51639287 12.141815 23.546791  8.326736
#> 167  84 2002-03-25 1.83655491 12.212703 23.868700  7.761871
#> 169  85 2002-03-26 4.45069999 12.283567 24.190677 10.430168
#> 171  86 2002-03-27 0.15910870 12.354400 24.512606  6.285997
#> 174  87 2002-03-28 4.66662603 12.425193 24.834373 10.872212
#> 175  88 2002-03-29 3.27189458 12.495938 25.155863  9.582329
#> 177  89 2002-03-30 4.49403120 12.566626 25.476961 10.924729
#> 179  90 2002-03-31 4.09229789 12.637251 25.797553 10.626376
#> 182  91 2002-04-01 5.25661765 12.707802 26.117528 11.931176
#> 183  92 2002-04-02 6.33856456 12.778271 26.436771 13.166073
#> 185  93 2002-04-03 6.62861669 12.848650 26.755170 13.590287
#> 187  94 2002-04-04 1.51535569 12.918928 27.072615  8.355926
#> 190  95 2002-04-05 2.08631224 12.989098 27.388994  9.046862
#> 192  96 2002-04-06 6.76553481 13.059148 27.704199 14.102388
#> 193  97 2002-04-07 3.50652106 13.129070 28.018120 10.746079
#> 195  98 2002-04-08 4.05350981 13.198853 28.330649 11.432986
#> 197  99 2002-04-09 1.11356547 13.268487 28.641680  8.362305
#> 200 100 2002-04-10 0.31882543 13.337960 28.951106  7.583794
#> 201 101 2002-04-11 3.77790298 13.407262 29.258822 11.436986
#> 203 102 2002-04-12 0.77011642 13.476382 29.564726  8.235929
#> 205 103 2002-04-13 0.21242301 13.545307 29.868714  7.701385
#> 208 104 2002-04-14 3.31570779 13.614025 30.170685 11.216720
#> 209 105 2002-04-15 6.45395597 13.682523 30.470539 14.804011
#> 211 106 2002-04-16 1.20596903 13.750789 30.768176  9.041256
#> 213 107 2002-04-17 0.92748313 13.818808 31.063500  8.808326
#> 216 108 2002-04-18 3.07687128 13.886567 31.356413 11.312951
#> 218 109 2002-04-19 2.43653206 13.954052 31.646820 10.674648
#> 219 110 2002-04-20 6.01752400 14.021246 31.934628 14.836378
#> 221 111 2002-04-21 6.47545620 14.088136 32.219743 15.459661
#> 223 112 2002-04-22 5.17704991 14.154703 32.502075 14.069298
#> 226 113 2002-04-23 4.79464611 14.220933 32.781534 13.721598
#> 227 114 2002-04-24 0.59298202 14.286808 33.058030  8.950554
#> 229 115 2002-04-25 3.55596464 14.352310 33.331477 12.462015
#> 231 116 2002-04-26 4.96504957 14.417421 33.601789 14.186314
#> 234 117 2002-04-27 0.30950482 14.482122 33.868881  8.829135
#> 235 118 2002-04-28 3.54396253 14.546394 34.132670 12.691068
#> 237 119 2002-04-29 4.19541795 14.610216 34.393074 13.536365
#> 239 120 2002-04-30 0.25084732 14.673568 34.650012  8.958677
#> 242 121 2002-05-01 6.13772666 14.736428 34.903405 15.994491
#> 243 122 2002-05-02 5.15388478 14.798774 35.153175 14.909592
#> 245 123 2002-05-03 2.81561952 14.860584 35.399245 12.203340
#> 248 124 2002-05-04 4.04375968 14.921835 35.641540 13.739745
#> 249 125 2002-05-05 1.62235810 14.982502 35.879987 10.912602
#> 251 126 2002-05-06 0.54063557 15.042562 36.114511  9.677613
#> 253 127 2002-05-07 4.68451999 15.101989 36.345041 14.723236
#> 256 128 2002-05-08 3.64271159 15.160757 36.571508 13.536439
#> 257 129 2002-05-09 2.06085679 15.218840 36.793843 11.689677
#> 259 130 2002-05-10 1.04547000 15.276211 37.011977 10.519503
#> 261 131 2002-05-11 3.57203087 15.332843 37.225844 13.642639
#> 264 132 2002-05-12 4.79490317 15.388707 37.435378 15.191011
#> 265 133 2002-05-13 1.63284221 15.443775 37.640517 11.399961
#> 267 134 2002-05-14 0.47149105 15.498017 37.841197 10.035914
#> 269 135 2002-05-15 5.14441830 15.551405 38.037357 15.800735
#> 272 136 2002-05-16 3.02056057 15.603907 38.228936 13.257359
#> 274 137 2002-05-17 3.80208363 15.655493 38.415875 14.268797
#> 275 138 2002-05-18 0.64289230 15.706133 38.598118 10.439489
#> 277 139 2002-05-19 5.35720453 15.755795 38.775606 16.286043
#> 279 140 2002-05-20 6.30054754 15.804447 38.948286 17.500568
#> 282 141 2002-05-21 5.07167578 15.852058 39.116103 16.036390
#> 283 142 2002-05-22 1.97658298 15.898595 39.279005 12.261420
#> 285 143 2002-05-23 6.63844567 15.944026 39.436940 18.069206
#> 287 144 2002-05-24 1.46384443 15.988319 39.589859 11.709832
#> 290 145 2002-05-25 5.33798597 16.031441 39.737713 16.550158
#> 291 146 2002-05-26 6.07219047 16.073361 39.880455 17.503128
#> 293 147 2002-05-27 2.52151679 16.114046 40.018039 13.135510
#> 295 148 2002-05-28 6.27738399 16.153463 40.150421 17.839016
#> 298 149 2002-05-29 3.30143202 16.191581 40.277558 14.175647
#> 300 150 2002-05-30 2.22380580 16.228369 40.399408 12.867858
#> 301 151 2002-05-31 1.76078214 16.263796 40.515933 12.322190
#> 303 152 2002-06-01 0.49225904 16.297830 40.627093 10.770323
#> 305 153 2002-06-02 4.69508835 16.330443 40.732853 16.038668
#> 308 154 2002-06-03 4.47887882 16.361604 40.833176 15.797198
#> 309 155 2002-06-04 0.73720028 16.391286 40.928031 11.152380
#> 311 156 2002-06-05 5.75494726 16.419460 41.017385 17.442551
#> 313 157 2002-06-06 0.44626107 16.446100 41.101209 10.832938
#> 316 158 2002-06-07 0.91227935 16.471180 41.179475 11.435260
#> 317 159 2002-06-08 1.94698379 16.494677 41.252156 12.747682
#> 319 160 2002-06-09 4.26336284 16.516565 41.319229 15.662601
#> 321 161 2002-06-10 0.75061440 16.536824 41.380671 11.284312
#> 324 162 2002-06-11 1.13166055 16.555433 41.436461 11.775327
#> 325 163 2002-06-12 0.42574253 16.572373 41.486582 10.904538
#> 328 164 2002-06-13 0.11710879 16.587625 41.531016 10.529359
#> 329 165 2002-06-14 2.83196881 16.601175 41.569750 13.938097
#> 331 166 2002-06-15 5.15712312 16.613007 41.602770 16.858001
#> 333 167 2002-06-16 1.55369273 16.623108 41.630068 12.353012
#> 336 168 2002-06-17 3.68254385 16.631469 41.651634 15.024166
#> 337 169 2002-06-18 2.79115511 16.638080 41.667464 13.911871
#> 339 170 2002-06-19 3.61308343 16.642934 41.677552 14.943364
#> 341 171 2002-06-20 2.77070820 16.646025 41.681897 13.889422
#> 344 172 2002-06-21 3.58123211 16.647351 41.680500 14.903347
#> 345 173 2002-06-22 0.05052060 16.646909 41.673363 10.481577
#> 347 174 2002-06-23 4.19014296 16.644700 41.660491 15.658936
#> 349 175 2002-06-24 2.60867397 16.640727 41.641891 13.674456
#> 352 176 2002-06-25 0.34962118 16.634994 41.617572 10.841736
#> 354 177 2002-06-26 3.65154469 16.627507 41.587545 14.963379
#> 355 178 2002-06-27 0.63808583 16.618275 41.551823 11.185681
#> 357 179 2002-06-28 4.45485718 16.607306 41.510423 15.945125
#> 359 180 2002-06-29 1.71007846 16.594614 41.463360 12.502244
#> 362 181 2002-06-30 4.88732231 16.580211 41.410655 16.455940
#> 363 182 2002-07-01 6.48323245 16.564113 41.352328 18.430781
#> 365 183 2002-07-02 0.69230939 16.546336 41.288405 11.185868
#> 367 184 2002-07-03 3.47972997 16.526900 41.218909 14.644037
#> 370 185 2002-07-04 1.22235259 16.505823 41.143868 11.809439
#> 371 186 2002-07-05 6.07821540 16.483128 41.063311 17.836953
#> 373 187 2002-07-06 2.02821412 16.458837 40.977269 12.769121
#> 375 188 2002-07-07 6.15829320 16.432973 40.885775 17.882461
#> 378 189 2002-07-08 0.82191101 16.405563 40.788864 11.218967
#> 380 190 2002-07-09 3.33782149 16.376632 40.686570 14.317932
#> 381 191 2002-07-10 6.62362675 16.346207 40.578932 18.366203
#> 383 192 2002-07-11 0.44028888 16.314316 40.465990 10.662543
#> 385 193 2002-07-12 4.39990853 16.280989 40.347784 15.538905
#> 388 194 2002-07-13 3.17699873 16.246255 40.224356 13.989079
#> 389 195 2002-07-14 1.45329146 16.210144 40.095751 11.821294
#> 391 196 2002-07-15 5.52208313 16.172686 39.962013 16.812918
#> 393 197 2002-07-16 2.51113466 16.133915 39.823189 13.054902
#> 396 198 2002-07-17 1.64050491 16.093860 39.679327 11.942160
#> 397 199 2002-07-18 5.36115489 16.052554 39.530476 16.483719
#> 399 200 2002-07-19 1.28768039 16.010029 39.376687 11.427698
#> 401 201 2002-07-20 1.36746734 15.966317 39.218012 11.483956
#> 404 202 2002-07-21 2.69287622 15.921451 39.054503 13.066369
#> 405 203 2002-07-22 5.87806153 15.875462 38.886214 16.920574
#> 408 204 2002-07-23 0.01294838 15.828384 38.713200  9.694135
#> 409 205 2002-07-24 1.06823007 15.780249 38.535519 10.938194
#> 411 206 2002-07-25 2.99441956 15.731088 38.353226 13.238583
#> 413 207 2002-07-26 5.03935166 15.680933 38.166382 15.674324
#> 416 208 2002-07-27 6.81250680 15.629816 37.975044 17.769778
#> 417 209 2002-07-28 4.08781449 15.577768 37.779274 14.401712
#> 419 210 2002-07-29 0.98475310 15.524820 37.579134 10.586622
#> 421 211 2002-07-30 3.34945004 15.471001 37.374685 13.389455
#> 424 212 2002-07-31 5.39234244 15.416342 37.165992 15.791476
#> 425 213 2002-08-01 0.15151088 15.360873 36.953119  9.420522
#> 427 214 2002-08-02 3.27735071 15.304621 36.736132 13.117394
#> 429 215 2002-08-03 4.25231779 15.247616 36.515097 14.220514
#> 432 216 2002-08-04 3.02125815 15.189886 36.290082 12.681557
#> 434 217 2002-08-05 4.41001241 15.131457 36.061156 14.270240
#> 435 218 2002-08-06 1.63312592 15.072356 35.828388 10.898143
#> 437 219 2002-08-07 6.81834417 15.012610 35.591848 16.980416
#> 439 220 2002-08-08 3.48466790 14.952243 35.351609 12.957305
#> 442 221 2002-08-09 5.60358538 14.891281 35.107744 15.382454
#> 443 222 2002-08-10 5.47713699 14.829748 34.860325 15.152641
#> 445 223 2002-08-11 2.59530234 14.767666 34.609429 11.693526
#> 447 224 2002-08-12 4.69832167 14.705060 34.355129 14.077078
#> 450 225 2002-08-13 0.07651361 14.641952 34.097505  8.613467
#> 451 226 2002-08-14 5.25620987 14.578362 33.836632 14.559035
#> 453 227 2002-08-15 4.46266085 14.514313 33.572591 13.554367
#> 455 228 2002-08-16 3.78961516 14.449824 33.305462 12.693716
#> 458 229 2002-08-17 3.82299639 14.384916 33.035326 12.648636
#> 460 230 2002-08-18 1.58834965 14.319607 32.762264 10.007583
#> 461 231 2002-08-19 0.64724391 14.253916 32.486361  8.859163
#> 463 232 2002-08-20 6.61293661 14.187862 32.207701 15.557901
#> 465 233 2002-08-21 3.08045483 14.121461 31.926370 11.463801
#> 468 234 2002-08-22 0.02577488 14.054730 31.642455  7.939628
#> 469 235 2002-08-23 4.64119835 13.987686 31.356043 13.041072
#> 471 236 2002-08-24 6.29471952 13.920345 31.067223 14.791037
#> 473 237 2002-08-25 5.91592738 13.852721 30.776085 14.265621
#> 476 238 2002-08-26 0.18406861 13.784830 30.482721  7.824198
#> 477 239 2002-08-27 2.38429029 13.716686 30.187223 10.170439
#> 479 240 2002-08-28 0.40483590 13.648302 29.889684  7.915715
#> 481 241 2002-08-29 4.62567329 13.579692 29.590199 12.437229
#> 484 242 2002-08-30 4.63439455 13.510869 29.288864 12.345436
#> 485 243 2002-08-31 4.59918903 13.441846 28.985774 12.205251
#> 487 244 2002-09-01 0.98946529 13.372633 28.681027  8.231337
#> 490 245 2002-09-02 6.45762181 13.303244 28.374722 13.980468
#> 491 246 2002-09-03 3.35266269 13.233689 28.066958 10.572023
#> 493 247 2002-09-04 6.93014104 13.163980 27.757836 14.245977
#> 495 248 2002-09-05 6.76425235 13.094126 27.447457 13.951361
#> 498 249 2002-09-06 5.41706421 13.024138 27.135924 12.427235
#> 499 250 2002-09-07 5.25481889 12.954027 26.823338 12.146296
#> 501 251 2002-09-08 6.23105346 12.883802 26.509805 13.037982
#> 503 252 2002-09-09 1.08186031 12.813472 26.195428  7.654716
#> 506 253 2002-09-10 2.04290454 12.743047 25.880314  8.544583
#> 507 254 2002-09-11 6.75003738 12.672536 25.564567 13.199636
#> 509 255 2002-09-12 1.32384361 12.601947 25.248295  7.638249
#> 511 256 2002-09-13 5.51786427 12.531289 24.931604 11.721929
#> 514 257 2002-09-14 3.05742053 12.460572 24.614602  9.173463
#> 516 258 2002-09-15 6.43247133 12.389802 24.297397 12.381646
#> 517 259 2002-09-16 0.34303491 12.318988 23.980097  6.328900
#> 519 260 2002-09-17 1.62301490 12.248139 23.662812  7.483496
#> 521 261 2002-09-18 5.35627440 12.177262 23.345649 10.970806
#> 524 262 2002-09-19 4.74483565 12.106365 23.028718 10.269991
#> 525 263 2002-09-20 4.78614349 12.035456 22.712128 10.194002
#> 527 264 2002-09-21 3.88812912 11.964544 22.395989  9.238020
#> 529 265 2002-09-22 5.12559950 11.893635 22.080409 10.277913
#> 532 266 2002-09-23 1.70971159 11.822738 21.765497  7.015152
#> 533 267 2002-09-24 3.46299276 11.751861 21.451363  8.523443
#> 535 268 2002-09-25 3.05471116 11.681012 21.138114  8.048451
#> 537 269 2002-09-26 2.37905312 11.610198 20.825859  7.340185
#> 540 270 2002-09-27 6.80059401 11.539428 20.514705 11.173698
#> 542 271 2002-09-28 1.40070429 11.468711 20.204761  6.285021
#> 543 272 2002-09-29 3.94987286 11.398053 19.896131  8.421428
#> 545 273 2002-09-30 1.43069631 11.327464 19.588923  6.134304
#> 547 274 2002-10-01 2.20602904 11.256953 19.283241  6.710282
#> 550 275 2002-10-02 4.02084829 11.186528 18.979189  8.155705
#> 551 276 2002-10-03 2.58339402 11.116198 18.676871  6.839461
#> 553 277 2002-10-04 5.63466647 11.045973 18.376389  9.281092
#> 555 278 2002-10-05 5.33847261 10.975862 18.077844  8.915839
#> 558 279 2002-10-06 3.63799932 10.905874 17.781337  7.411098
#> 559 280 2002-10-07 3.29688756 10.836020 17.486965  7.031969
#> 561 281 2002-10-08 3.20441074 10.766311 17.194826  6.857581
#> 563 282 2002-10-09 5.83051944 10.696756 16.905016  8.833493
#> 566 283 2002-10-10 0.40885058 10.627367 16.617629  4.474060
#> 567 284 2002-10-11 1.51159607 10.558154 16.332759  5.252359
#> 570 285 2002-10-12 2.08420439 10.489131 16.050495  5.607251
#> 571 286 2002-10-13 0.36265274 10.420308 15.770928  4.217166
#> 573 287 2002-10-14 3.19667106 10.351698 15.494145  6.265882
#> 575 288 2002-10-15 2.03635202 10.283314 15.220231  5.312050
#> 578 289 2002-10-16 0.52170300 10.215170 14.949271  4.119058
#> 579 290 2002-10-17 4.54323990 10.147279 14.681345  6.956975
#> 581 291 2002-10-18 3.56137276 10.079655 14.416534  6.150979
#> 583 292 2002-10-19 3.82180907 10.012314 14.154915  6.240271
#> 586 293 2002-10-20 6.19760572  9.945270 13.896562  7.804109
#> 587 294 2002-10-21 4.17322475  9.878539 13.641550  6.291849
#> 589 295 2002-10-22 0.08858591  9.812138 13.389948  3.407931
#> 591 296 2002-10-23 6.06704602  9.746084 13.141825  7.375923
#> 594 297 2002-10-24 0.18914340  9.680393 12.897247  3.350310
#> 596 298 2002-10-25 0.17378459  9.615084 12.656277  3.278445
#> 597 299 2002-10-26 1.17246754  9.550176 12.418977  3.867078
#> 599 300 2002-10-27 6.00422060  9.485687 12.185404  6.902890
#> 601 301 2002-10-28 3.58030246  9.421638 11.955615  5.260521
#> 604 302 2002-10-29 4.02446894  9.358048 11.729663  5.454612
#> 605 303 2002-10-30 6.95840858  9.294940 11.507600  7.184329
#> 607 304 2002-10-31 3.71936284  9.232334 11.289473  5.096422
#> 609 305 2002-11-01 1.02622910  9.170252 11.075328  3.388544
#> 612 306 2002-11-02 3.28013573  9.108719 10.865209  4.672635
#> 613 307 2002-11-03 1.57521119  9.047757 10.659156  3.592667
#> 615 308 2002-11-04 2.27316034  8.987390 10.457208  3.936761
#> 617 309 2002-11-05 4.59891911  8.927644 10.259398  5.207324
#> 620 310 2002-11-06 6.94886532  8.868543 10.065762  6.459908
#> 622 311 2002-11-07 6.38045766  8.810114  9.876328  6.045397
#> 623 312 2002-11-08 4.32311000  8.752384  9.691126  4.816175
#> 625 313 2002-11-09 4.70794386  8.695379  9.510180  4.952096
#> 627 314 2002-11-10 4.64087400  8.639127  9.333513  4.840325
#> 630 315 2002-11-11 2.91784478  8.583658  9.161147  3.847362
#> 631 316 2002-11-12 1.68704218  8.528999  8.993100  3.137696
#> 633 317 2002-11-13 5.61655294  8.475180  8.829388  5.132991
#> 635 318 2002-11-14 5.71261607  8.422232  8.670025  5.107851
#> 638 319 2002-11-15 3.31266724  8.370184  8.515023  3.813751
#> 639 320 2002-11-16 4.24243666  8.319067  8.364392  4.223873
#> 641 321 2002-11-17 4.71223966  8.268912  8.218139  4.396188
#> 643 322 2002-11-18 0.97188974  8.219751  8.076271  2.496530
#> 646 323 2002-11-19 5.93045569  8.171616  7.938791  4.865441
#> 647 324 2002-11-20 2.24720792  8.124538  7.805702  3.030935
#> 650 325 2002-11-21 4.30504823  8.078549  7.677004  3.964784
#> 651 326 2002-11-22 3.83145318  8.033683  7.552698  3.689205
#> 653 327 2002-11-23 3.16062950  7.989971  7.432780  3.328305
#> 655 328 2002-11-24 0.58300059  7.947446  7.317248  2.097697
#> 658 329 2002-11-25 6.68681602  7.906140  7.206095  4.848892
#> 659 330 2002-11-26 6.69771738  7.866085  7.099317  4.797249
#> 661 331 2002-11-27 2.81026358  7.827314  6.996907  3.005287
#> 663 332 2002-11-28 5.00262820  7.789856  6.898857  3.939929
#> 666 333 2002-11-29 6.53343134  7.753745  6.805157  4.568357
#> 667 334 2002-11-30 0.61344839  7.719011  6.715800  1.945810
#> 669 335 2002-12-01 1.86366704  7.685684  6.630775  2.461627
#> 671 336 2002-12-02 5.18185073  7.653793  6.550072  3.854817
#> 674 337 2002-12-03 3.74016836  7.623368  6.473680  3.206475
#> 676 338 2002-12-04 3.32863548  7.594437  6.401590  3.003303
#> 677 339 2002-12-05 1.42888268  7.567027  6.333789  2.181452
#> 679 340 2002-12-06 2.44315115  7.541163  6.270268  2.583273
#> 681 341 2002-12-07 0.82921767  7.516872  6.211016  1.895336
#> 684 342 2002-12-08 5.20518093  7.494177  6.156023  3.676880
#> 685 343 2002-12-09 2.88560994  7.473100  6.105278  2.705044
#> 687 344 2002-12-10 3.96462149  7.453664  6.058772  3.126031
#> 689 345 2002-12-11 0.12494967  7.435887  6.016497  1.554674
#> 692 346 2002-12-12 3.28691284  7.419789  5.978443  2.818814
#> 693 347 2002-12-13 2.24987743  7.405386  5.944603  2.389184
#> 695 348 2002-12-14 1.94252978  7.392694  5.914970  2.255862
#> 697 349 2002-12-15 3.23223833  7.381725  5.889539  2.761812
#> 700 350 2002-12-16 1.00146772  7.372493  5.868303  1.865646
#> 702 351 2002-12-17 3.29568326  7.365006  5.851259  2.771972
#> 703 352 2002-12-18 4.85437867  7.359273  5.838404  3.385187
#> 705 353 2002-12-19 3.57671756  7.355300  5.829736  2.874869
#> 707 354 2002-12-20 5.76014524  7.353091  5.825253  3.737959
#> 710 355 2002-12-21 0.48302121  7.352649  5.824956  1.647570
#> 711 356 2002-12-22 6.67302760  7.353975  5.828845  4.101771
#> 713 357 2002-12-23 5.62080526  7.357066  5.836924  3.688938
#> 715 358 2002-12-24 4.53916610  7.361920  5.849194  3.265528
#> 718 359 2002-12-25 6.81737215  7.368531  5.865662  4.179873
#> 719 360 2002-12-26 4.13856635  7.376892  5.886331  3.122751
#> 721 361 2002-12-27 3.83624688  7.386993  5.911209  3.012720
#> 723 362 2002-12-28 4.96262396  7.398825  5.940303  3.477250
#> 726 363 2002-12-29 5.05177740  7.412375  5.973622  3.529016
#> 727 364 2002-12-30 2.09577172  7.427627  6.011174  2.350847
#> 729 365 2002-12-31 5.70615258  7.444567  6.052971  3.833000
#> 2     1 2003-01-01 2.31982961  7.463176  6.099033  2.472660
#> 4     2 2003-01-02 0.37740624  7.483435  6.149345  1.692399
#> 5     3 2003-01-03 5.39469713  7.505323  6.203934  3.780624
#> 8     4 2003-01-04 3.29654447  7.528820  6.262814  2.936811
#> 10    5 2003-01-05 6.13985549  7.553900  6.325996  4.152403
#> 12    6 2003-01-06 3.98233910  7.580540  6.393496  3.277744
#> 13    7 2003-01-07 3.25216836  7.608714  6.465324  2.998057
#> 16    8 2003-01-08 1.38528470  7.638396  6.541496  2.228550
#> 18    9 2003-01-09 1.50048071  7.669557  6.622023  2.303276
#> 20   10 2003-01-10 0.43546683  7.702170  6.706920  1.866328
#> 21   11 2003-01-11 1.44677032  7.736204  6.796198  2.334538
#> 24   12 2003-01-12 6.32413442  7.771631  6.889871  4.525771
#> 26   13 2003-01-13 4.90925353  7.808419  6.987950  3.943695
#> 28   14 2003-01-14 0.01716437  7.846537  7.090446  1.780367
#> 29   15 2003-01-15 2.41108846  7.885954  7.197371  2.899621
#> 31   16 2003-01-16 1.34632829  7.926639  7.308732  2.447872
#> 34   17 2003-01-17 3.77833869  7.968559  7.424539  3.616329
#> 36   18 2003-01-18 3.95555264  8.011681  7.544799  3.748721
#> 38   19 2003-01-19 0.24293017  8.055974  7.669519  2.033018
#> 39   20 2003-01-20 6.31002067  8.101405  7.798703  4.986801
#> 42   21 2003-01-21 6.24165115  8.147942  7.932354  5.021339
#> 44   22 2003-01-22 1.99516932  8.195553  8.070474  2.999978
#> 46   23 2003-01-23 6.41847176  8.244205  8.213063  5.250379
#> 47   24 2003-01-24 1.22896735  8.293867  8.360119  2.709422
#> 50   25 2003-01-25 2.56352263  8.344507  8.511639  3.435344
#> 52   26 2003-01-26 2.23909882  8.396093  8.667618  3.322659
#> 54   27 2003-01-27 5.20698487  8.448595  8.828046  4.927434
#> 55   28 2003-01-28 3.68550512  8.501983  8.992915  4.197388
#> 57   29 2003-01-29 0.84714189  8.556225  9.162213  2.744123
#> 60   30 2003-01-30 4.70549184  8.611293  9.335925  4.884709
#> 62   31 2003-01-31 3.59263422  8.667157  9.514035  4.350346
#> 64   32 2003-02-01 0.69916822  8.723789  9.696523  2.812695
#> 65   33 2003-02-02 5.89606197  8.781160  9.883368  5.788909
#> 68   34 2003-02-03 3.23854344  8.839243 10.074546  4.364205
#> 70   35 2003-02-04 4.18834701  8.898011 10.270030  4.984590
#> 72   36 2003-02-05 1.67306024  8.957438 10.469790  3.595215
#> 73   37 2003-02-06 0.19174598  9.017498 10.673795  2.781931
#> 76   38 2003-02-07 6.29735015  9.078165 10.882010  6.494824
#> 78   39 2003-02-08 6.04013345  9.139416 11.094398  6.439679
#> 80   40 2003-02-09 2.95545410  9.201226 11.310918  4.644276
#> 81   41 2003-02-10 2.01245718  9.263572 11.531528  4.135461
#> 84   42 2003-02-11 3.10622049  9.326432 11.756182  4.896777
#> 85   43 2003-02-12 0.79812337  9.389784 11.984831  3.505558
#> 88   44 2003-02-13 3.51148390  9.453606 12.217425  5.323400
#> 90   45 2003-02-14 0.58087339  9.517878 12.453910  3.493507
#> 92   46 2003-02-15 3.20562213  9.582579 12.694229  5.296832
#> 93   47 2003-02-16 0.86947721  9.647690 12.938322  3.817600
#> 96   48 2003-02-17 0.81480127  9.713192 13.186128  3.849598
#> 98   49 2003-02-18 4.61605587  9.779067 13.437583  6.530896
#> 100  50 2003-02-19 1.15747411  9.845297 13.692618  4.228049
#> 101  51 2003-02-20 6.80609747  9.911864 13.951164  8.277656
#> 104  52 2003-02-21 5.83591628  9.978754 14.213148  7.709454
#> 106  53 2003-02-22 0.65508159 10.045948 14.478496  4.091685
#> 108  54 2003-02-23 3.99584588 10.113433 14.747130  6.600099
#> 109  55 2003-02-24 3.73801083 10.181192 15.018971  6.511840
#> 111  56 2003-02-25 0.79868082 10.249211 15.293936  4.419382
#> 114  57 2003-02-26 2.82367195 10.317477 15.571940  6.023838
#> 116  58 2003-02-27 5.47075880 10.385975 15.852898  8.138440
#> 118  59 2003-02-28 0.88137666 10.454693 16.136720  4.714378
#> 119  60 2003-03-01 5.94351226 10.523618 16.423316  8.743596
#> 122  61 2003-03-02 0.64098230 10.592738 16.712592  4.683800
#> 124  62 2003-03-03 0.36244195 10.662040 17.004454  4.540135
#> 126  63 2003-03-04 4.61734403 10.731513 17.298805  8.046195
#> 127  64 2003-03-05 3.46967307 10.801147 17.595546  7.225012
#> 130  65 2003-03-06 3.82658430 10.870930 17.894578  7.623104
#> 132  66 2003-03-07 2.19606105 10.940852 18.195799  6.375091
#> 134  67 2003-03-08 2.28912471 11.010902 18.499104  6.547723
#> 135  68 2003-03-09 3.55853999 11.081072 18.804391  7.720489
#> 137  69 2003-03-10 2.87123565 11.151350 19.111552  7.238297
#> 140  70 2003-03-11 4.24175578 11.221729 19.420480  8.525541
#> 142  71 2003-03-12 3.36703035 11.292198 19.731067  7.874404
#> 144  72 2003-03-13 1.77457315 11.362749 20.043203  6.575921
#> 145  73 2003-03-14 5.09506915 11.433374 20.356778  9.625003
#> 148  74 2003-03-15 2.42814351 11.504062 20.671679  7.349488
#> 150  75 2003-03-16 2.40277040 11.574807 20.987796  7.425337
#> 152  76 2003-03-17 1.65151918 11.645600 21.305014  6.836937
#> 153  77 2003-03-18 2.05384672 11.716433 21.623220  7.301040
#> 156  78 2003-03-19 6.98850504 11.787297 21.942300 11.990199
#> 158  79 2003-03-20 0.13760043 11.858185 22.262140  5.694698
#> 160  80 2003-03-21 3.48833681 11.929088 22.582624  8.947492
#> 161  81 2003-03-22 2.48763280 12.000000 22.903638  8.099903
#> 164  82 2003-03-23 0.63211065 12.070912 23.225065  6.414373
#> 165  83 2003-03-24 2.51639287 12.141815 23.546791  8.326736
#> 168  84 2003-03-25 1.83655491 12.212703 23.868700  7.761871
#> 170  85 2003-03-26 4.45069999 12.283567 24.190677 10.430168
#> 172  86 2003-03-27 0.15910870 12.354400 24.512606  6.285997
#> 173  87 2003-03-28 4.66662603 12.425193 24.834373 10.872212
#> 176  88 2003-03-29 3.27189458 12.495938 25.155863  9.582329
#> 178  89 2003-03-30 4.49403120 12.566626 25.476961 10.924729
#> 180  90 2003-03-31 4.09229789 12.637251 25.797553 10.626376
#> 181  91 2003-04-01 5.25661765 12.707802 26.117528 11.931176
#> 184  92 2003-04-02 6.33856456 12.778271 26.436771 13.166073
#> 186  93 2003-04-03 6.62861669 12.848650 26.755170 13.590287
#> 188  94 2003-04-04 1.51535569 12.918928 27.072615  8.355926
#> 189  95 2003-04-05 2.08631224 12.989098 27.388994  9.046862
#> 191  96 2003-04-06 6.76553481 13.059148 27.704199 14.102388
#> 194  97 2003-04-07 3.50652106 13.129070 28.018120 10.746079
#> 196  98 2003-04-08 4.05350981 13.198853 28.330649 11.432986
#> 198  99 2003-04-09 1.11356547 13.268487 28.641680  8.362305
#> 199 100 2003-04-10 0.31882543 13.337960 28.951106  7.583794
#> 202 101 2003-04-11 3.77790298 13.407262 29.258822 11.436986
#> 204 102 2003-04-12 0.77011642 13.476382 29.564726  8.235929
#> 206 103 2003-04-13 0.21242301 13.545307 29.868714  7.701385
#> 207 104 2003-04-14 3.31570779 13.614025 30.170685 11.216720
#> 210 105 2003-04-15 6.45395597 13.682523 30.470539 14.804011
#> 212 106 2003-04-16 1.20596903 13.750789 30.768176  9.041256
#> 214 107 2003-04-17 0.92748313 13.818808 31.063500  8.808326
#> 215 108 2003-04-18 3.07687128 13.886567 31.356413 11.312951
#> 217 109 2003-04-19 2.43653206 13.954052 31.646820 10.674648
#> 220 110 2003-04-20 6.01752400 14.021246 31.934628 14.836378
#> 222 111 2003-04-21 6.47545620 14.088136 32.219743 15.459661
#> 224 112 2003-04-22 5.17704991 14.154703 32.502075 14.069298
#> 225 113 2003-04-23 4.79464611 14.220933 32.781534 13.721598
#> 228 114 2003-04-24 0.59298202 14.286808 33.058030  8.950554
#> 230 115 2003-04-25 3.55596464 14.352310 33.331477 12.462015
#> 232 116 2003-04-26 4.96504957 14.417421 33.601789 14.186314
#> 233 117 2003-04-27 0.30950482 14.482122 33.868881  8.829135
#> 236 118 2003-04-28 3.54396253 14.546394 34.132670 12.691068
#> 238 119 2003-04-29 4.19541795 14.610216 34.393074 13.536365
#> 240 120 2003-04-30 0.25084732 14.673568 34.650012  8.958677
#> 241 121 2003-05-01 6.13772666 14.736428 34.903405 15.994491
#> 244 122 2003-05-02 5.15388478 14.798774 35.153175 14.909592
#> 246 123 2003-05-03 2.81561952 14.860584 35.399245 12.203340
#> 247 124 2003-05-04 4.04375968 14.921835 35.641540 13.739745
#> 250 125 2003-05-05 1.62235810 14.982502 35.879987 10.912602
#> 252 126 2003-05-06 0.54063557 15.042562 36.114511  9.677613
#> 254 127 2003-05-07 4.68451999 15.101989 36.345041 14.723236
#> 255 128 2003-05-08 3.64271159 15.160757 36.571508 13.536439
#> 258 129 2003-05-09 2.06085679 15.218840 36.793843 11.689677
#> 260 130 2003-05-10 1.04547000 15.276211 37.011977 10.519503
#> 262 131 2003-05-11 3.57203087 15.332843 37.225844 13.642639
#> 263 132 2003-05-12 4.79490317 15.388707 37.435378 15.191011
#> 266 133 2003-05-13 1.63284221 15.443775 37.640517 11.399961
#> 268 134 2003-05-14 0.47149105 15.498017 37.841197 10.035914
#> 270 135 2003-05-15 5.14441830 15.551405 38.037357 15.800735
#> 271 136 2003-05-16 3.02056057 15.603907 38.228936 13.257359
#> 273 137 2003-05-17 3.80208363 15.655493 38.415875 14.268797
#> 276 138 2003-05-18 0.64289230 15.706133 38.598118 10.439489
#> 278 139 2003-05-19 5.35720453 15.755795 38.775606 16.286043
#> 280 140 2003-05-20 6.30054754 15.804447 38.948286 17.500568
#> 281 141 2003-05-21 5.07167578 15.852058 39.116103 16.036390
#> 284 142 2003-05-22 1.97658298 15.898595 39.279005 12.261420
#> 286 143 2003-05-23 6.63844567 15.944026 39.436940 18.069206
#> 288 144 2003-05-24 1.46384443 15.988319 39.589859 11.709832
#> 289 145 2003-05-25 5.33798597 16.031441 39.737713 16.550158
#> 292 146 2003-05-26 6.07219047 16.073361 39.880455 17.503128
#> 294 147 2003-05-27 2.52151679 16.114046 40.018039 13.135510
#> 296 148 2003-05-28 6.27738399 16.153463 40.150421 17.839016
#> 297 149 2003-05-29 3.30143202 16.191581 40.277558 14.175647
#> 299 150 2003-05-30 2.22380580 16.228369 40.399408 12.867858
#> 302 151 2003-05-31 1.76078214 16.263796 40.515933 12.322190
#> 304 152 2003-06-01 0.49225904 16.297830 40.627093 10.770323
#> 306 153 2003-06-02 4.69508835 16.330443 40.732853 16.038668
#> 307 154 2003-06-03 4.47887882 16.361604 40.833176 15.797198
#> 310 155 2003-06-04 0.73720028 16.391286 40.928031 11.152380
#> 312 156 2003-06-05 5.75494726 16.419460 41.017385 17.442551
#> 314 157 2003-06-06 0.44626107 16.446100 41.101209 10.832938
#> 315 158 2003-06-07 0.91227935 16.471180 41.179475 11.435260
#> 318 159 2003-06-08 1.94698379 16.494677 41.252156 12.747682
#> 320 160 2003-06-09 4.26336284 16.516565 41.319229 15.662601
#> 322 161 2003-06-10 0.75061440 16.536824 41.380671 11.284312
#> 323 162 2003-06-11 1.13166055 16.555433 41.436461 11.775327
#> 326 163 2003-06-12 0.42574253 16.572373 41.486582 10.904538
#> 327 164 2003-06-13 0.11710879 16.587625 41.531016 10.529359
#> 330 165 2003-06-14 2.83196881 16.601175 41.569750 13.938097
#> 332 166 2003-06-15 5.15712312 16.613007 41.602770 16.858001
#> 334 167 2003-06-16 1.55369273 16.623108 41.630068 12.353012
#> 335 168 2003-06-17 3.68254385 16.631469 41.651634 15.024166
#> 338 169 2003-06-18 2.79115511 16.638080 41.667464 13.911871
#> 340 170 2003-06-19 3.61308343 16.642934 41.677552 14.943364
#> 342 171 2003-06-20 2.77070820 16.646025 41.681897 13.889422
#> 343 172 2003-06-21 3.58123211 16.647351 41.680500 14.903347
#> 346 173 2003-06-22 0.05052060 16.646909 41.673363 10.481577
#> 348 174 2003-06-23 4.19014296 16.644700 41.660491 15.658936
#> 350 175 2003-06-24 2.60867397 16.640727 41.641891 13.674456
#> 351 176 2003-06-25 0.34962118 16.634994 41.617572 10.841736
#> 353 177 2003-06-26 3.65154469 16.627507 41.587545 14.963379
#> 356 178 2003-06-27 0.63808583 16.618275 41.551823 11.185681
#> 358 179 2003-06-28 4.45485718 16.607306 41.510423 15.945125
#> 360 180 2003-06-29 1.71007846 16.594614 41.463360 12.502244
#> 361 181 2003-06-30 4.88732231 16.580211 41.410655 16.455940
#> 364 182 2003-07-01 6.48323245 16.564113 41.352328 18.430781
#> 366 183 2003-07-02 0.69230939 16.546336 41.288405 11.185868
#> 368 184 2003-07-03 3.47972997 16.526900 41.218909 14.644037
#> 369 185 2003-07-04 1.22235259 16.505823 41.143868 11.809439
#> 372 186 2003-07-05 6.07821540 16.483128 41.063311 17.836953
#> 374 187 2003-07-06 2.02821412 16.458837 40.977269 12.769121
#> 376 188 2003-07-07 6.15829320 16.432973 40.885775 17.882461
#> 377 189 2003-07-08 0.82191101 16.405563 40.788864 11.218967
#> 379 190 2003-07-09 3.33782149 16.376632 40.686570 14.317932
#> 382 191 2003-07-10 6.62362675 16.346207 40.578932 18.366203
#> 384 192 2003-07-11 0.44028888 16.314316 40.465990 10.662543
#> 386 193 2003-07-12 4.39990853 16.280989 40.347784 15.538905
#> 387 194 2003-07-13 3.17699873 16.246255 40.224356 13.989079
#> 390 195 2003-07-14 1.45329146 16.210144 40.095751 11.821294
#> 392 196 2003-07-15 5.52208313 16.172686 39.962013 16.812918
#> 394 197 2003-07-16 2.51113466 16.133915 39.823189 13.054902
#> 395 198 2003-07-17 1.64050491 16.093860 39.679327 11.942160
#> 398 199 2003-07-18 5.36115489 16.052554 39.530476 16.483719
#> 400 200 2003-07-19 1.28768039 16.010029 39.376687 11.427698
#> 402 201 2003-07-20 1.36746734 15.966317 39.218012 11.483956
#> 403 202 2003-07-21 2.69287622 15.921451 39.054503 13.066369
#> 406 203 2003-07-22 5.87806153 15.875462 38.886214 16.920574
#> 407 204 2003-07-23 0.01294838 15.828384 38.713200  9.694135
#> 410 205 2003-07-24 1.06823007 15.780249 38.535519 10.938194
#> 412 206 2003-07-25 2.99441956 15.731088 38.353226 13.238583
#> 414 207 2003-07-26 5.03935166 15.680933 38.166382 15.674324
#> 415 208 2003-07-27 6.81250680 15.629816 37.975044 17.769778
#> 418 209 2003-07-28 4.08781449 15.577768 37.779274 14.401712
#> 420 210 2003-07-29 0.98475310 15.524820 37.579134 10.586622
#> 422 211 2003-07-30 3.34945004 15.471001 37.374685 13.389455
#> 423 212 2003-07-31 5.39234244 15.416342 37.165992 15.791476
#> 426 213 2003-08-01 0.15151088 15.360873 36.953119  9.420522
#> 428 214 2003-08-02 3.27735071 15.304621 36.736132 13.117394
#> 430 215 2003-08-03 4.25231779 15.247616 36.515097 14.220514
#> 431 216 2003-08-04 3.02125815 15.189886 36.290082 12.681557
#> 433 217 2003-08-05 4.41001241 15.131457 36.061156 14.270240
#> 436 218 2003-08-06 1.63312592 15.072356 35.828388 10.898143
#> 438 219 2003-08-07 6.81834417 15.012610 35.591848 16.980416
#> 440 220 2003-08-08 3.48466790 14.952243 35.351609 12.957305
#> 441 221 2003-08-09 5.60358538 14.891281 35.107744 15.382454
#> 444 222 2003-08-10 5.47713699 14.829748 34.860325 15.152641
#> 446 223 2003-08-11 2.59530234 14.767666 34.609429 11.693526
#> 448 224 2003-08-12 4.69832167 14.705060 34.355129 14.077078
#> 449 225 2003-08-13 0.07651361 14.641952 34.097505  8.613467
#> 452 226 2003-08-14 5.25620987 14.578362 33.836632 14.559035
#> 454 227 2003-08-15 4.46266085 14.514313 33.572591 13.554367
#> 456 228 2003-08-16 3.78961516 14.449824 33.305462 12.693716
#> 457 229 2003-08-17 3.82299639 14.384916 33.035326 12.648636
#> 459 230 2003-08-18 1.58834965 14.319607 32.762264 10.007583
#> 462 231 2003-08-19 0.64724391 14.253916 32.486361  8.859163
#> 464 232 2003-08-20 6.61293661 14.187862 32.207701 15.557901
#> 466 233 2003-08-21 3.08045483 14.121461 31.926370 11.463801
#> 467 234 2003-08-22 0.02577488 14.054730 31.642455  7.939628
#> 470 235 2003-08-23 4.64119835 13.987686 31.356043 13.041072
#> 472 236 2003-08-24 6.29471952 13.920345 31.067223 14.791037
#> 474 237 2003-08-25 5.91592738 13.852721 30.776085 14.265621
#> 475 238 2003-08-26 0.18406861 13.784830 30.482721  7.824198
#> 478 239 2003-08-27 2.38429029 13.716686 30.187223 10.170439
#> 480 240 2003-08-28 0.40483590 13.648302 29.889684  7.915715
#> 482 241 2003-08-29 4.62567329 13.579692 29.590199 12.437229
#> 483 242 2003-08-30 4.63439455 13.510869 29.288864 12.345436
#> 486 243 2003-08-31 4.59918903 13.441846 28.985774 12.205251
#> 488 244 2003-09-01 0.98946529 13.372633 28.681027  8.231337
#> 489 245 2003-09-02 6.45762181 13.303244 28.374722 13.980468
#> 492 246 2003-09-03 3.35266269 13.233689 28.066958 10.572023
#> 494 247 2003-09-04 6.93014104 13.163980 27.757836 14.245977
#> 496 248 2003-09-05 6.76425235 13.094126 27.447457 13.951361
#> 497 249 2003-09-06 5.41706421 13.024138 27.135924 12.427235
#> 500 250 2003-09-07 5.25481889 12.954027 26.823338 12.146296
#> 502 251 2003-09-08 6.23105346 12.883802 26.509805 13.037982
#> 504 252 2003-09-09 1.08186031 12.813472 26.195428  7.654716
#> 505 253 2003-09-10 2.04290454 12.743047 25.880314  8.544583
#> 508 254 2003-09-11 6.75003738 12.672536 25.564567 13.199636
#> 510 255 2003-09-12 1.32384361 12.601947 25.248295  7.638249
#> 512 256 2003-09-13 5.51786427 12.531289 24.931604 11.721929
#> 513 257 2003-09-14 3.05742053 12.460572 24.614602  9.173463
#> 515 258 2003-09-15 6.43247133 12.389802 24.297397 12.381646
#> 518 259 2003-09-16 0.34303491 12.318988 23.980097  6.328900
#> 520 260 2003-09-17 1.62301490 12.248139 23.662812  7.483496
#> 522 261 2003-09-18 5.35627440 12.177262 23.345649 10.970806
#> 523 262 2003-09-19 4.74483565 12.106365 23.028718 10.269991
#> 526 263 2003-09-20 4.78614349 12.035456 22.712128 10.194002
#> 528 264 2003-09-21 3.88812912 11.964544 22.395989  9.238020
#> 530 265 2003-09-22 5.12559950 11.893635 22.080409 10.277913
#> 531 266 2003-09-23 1.70971159 11.822738 21.765497  7.015152
#> 534 267 2003-09-24 3.46299276 11.751861 21.451363  8.523443
#> 536 268 2003-09-25 3.05471116 11.681012 21.138114  8.048451
#> 538 269 2003-09-26 2.37905312 11.610198 20.825859  7.340185
#> 539 270 2003-09-27 6.80059401 11.539428 20.514705 11.173698
#> 541 271 2003-09-28 1.40070429 11.468711 20.204761  6.285021
#> 544 272 2003-09-29 3.94987286 11.398053 19.896131  8.421428
#> 546 273 2003-09-30 1.43069631 11.327464 19.588923  6.134304
#> 548 274 2003-10-01 2.20602904 11.256953 19.283241  6.710282
#> 549 275 2003-10-02 4.02084829 11.186528 18.979189  8.155705
#> 552 276 2003-10-03 2.58339402 11.116198 18.676871  6.839461
#> 554 277 2003-10-04 5.63466647 11.045973 18.376389  9.281092
#> 556 278 2003-10-05 5.33847261 10.975862 18.077844  8.915839
#> 557 279 2003-10-06 3.63799932 10.905874 17.781337  7.411098
#> 560 280 2003-10-07 3.29688756 10.836020 17.486965  7.031969
#> 562 281 2003-10-08 3.20441074 10.766311 17.194826  6.857581
#> 564 282 2003-10-09 5.83051944 10.696756 16.905016  8.833493
#> 565 283 2003-10-10 0.40885058 10.627367 16.617629  4.474060
#> 568 284 2003-10-11 1.51159607 10.558154 16.332759  5.252359
#> 569 285 2003-10-12 2.08420439 10.489131 16.050495  5.607251
#> 572 286 2003-10-13 0.36265274 10.420308 15.770928  4.217166
#> 574 287 2003-10-14 3.19667106 10.351698 15.494145  6.265882
#> 576 288 2003-10-15 2.03635202 10.283314 15.220231  5.312050
#> 577 289 2003-10-16 0.52170300 10.215170 14.949271  4.119058
#> 580 290 2003-10-17 4.54323990 10.147279 14.681345  6.956975
#> 582 291 2003-10-18 3.56137276 10.079655 14.416534  6.150979
#> 584 292 2003-10-19 3.82180907 10.012314 14.154915  6.240271
#> 585 293 2003-10-20 6.19760572  9.945270 13.896562  7.804109
#> 588 294 2003-10-21 4.17322475  9.878539 13.641550  6.291849
#> 590 295 2003-10-22 0.08858591  9.812138 13.389948  3.407931
#> 592 296 2003-10-23 6.06704602  9.746084 13.141825  7.375923
#> 593 297 2003-10-24 0.18914340  9.680393 12.897247  3.350310
#> 595 298 2003-10-25 0.17378459  9.615084 12.656277  3.278445
#> 598 299 2003-10-26 1.17246754  9.550176 12.418977  3.867078
#> 600 300 2003-10-27 6.00422060  9.485687 12.185404  6.902890
#> 602 301 2003-10-28 3.58030246  9.421638 11.955615  5.260521
#> 603 302 2003-10-29 4.02446894  9.358048 11.729663  5.454612
#> 606 303 2003-10-30 6.95840858  9.294940 11.507600  7.184329
#> 608 304 2003-10-31 3.71936284  9.232334 11.289473  5.096422
#> 610 305 2003-11-01 1.02622910  9.170252 11.075328  3.388544
#> 611 306 2003-11-02 3.28013573  9.108719 10.865209  4.672635
#> 614 307 2003-11-03 1.57521119  9.047757 10.659156  3.592667
#> 616 308 2003-11-04 2.27316034  8.987390 10.457208  3.936761
#> 618 309 2003-11-05 4.59891911  8.927644 10.259398  5.207324
#> 619 310 2003-11-06 6.94886532  8.868543 10.065762  6.459908
#> 621 311 2003-11-07 6.38045766  8.810114  9.876328  6.045397
#> 624 312 2003-11-08 4.32311000  8.752384  9.691126  4.816175
#> 626 313 2003-11-09 4.70794386  8.695379  9.510180  4.952096
#> 628 314 2003-11-10 4.64087400  8.639127  9.333513  4.840325
#> 629 315 2003-11-11 2.91784478  8.583658  9.161147  3.847362
#> 632 316 2003-11-12 1.68704218  8.528999  8.993100  3.137696
#> 634 317 2003-11-13 5.61655294  8.475180  8.829388  5.132991
#> 636 318 2003-11-14 5.71261607  8.422232  8.670025  5.107851
#> 637 319 2003-11-15 3.31266724  8.370184  8.515023  3.813751
#> 640 320 2003-11-16 4.24243666  8.319067  8.364392  4.223873
#> 642 321 2003-11-17 4.71223966  8.268912  8.218139  4.396188
#> 644 322 2003-11-18 0.97188974  8.219751  8.076271  2.496530
#> 645 323 2003-11-19 5.93045569  8.171616  7.938791  4.865441
#> 648 324 2003-11-20 2.24720792  8.124538  7.805702  3.030935
#> 649 325 2003-11-21 4.30504823  8.078549  7.677004  3.964784
#> 652 326 2003-11-22 3.83145318  8.033683  7.552698  3.689205
#> 654 327 2003-11-23 3.16062950  7.989971  7.432780  3.328305
#> 656 328 2003-11-24 0.58300059  7.947446  7.317248  2.097697
#> 657 329 2003-11-25 6.68681602  7.906140  7.206095  4.848892
#> 660 330 2003-11-26 6.69771738  7.866085  7.099317  4.797249
#> 662 331 2003-11-27 2.81026358  7.827314  6.996907  3.005287
#> 664 332 2003-11-28 5.00262820  7.789856  6.898857  3.939929
#> 665 333 2003-11-29 6.53343134  7.753745  6.805157  4.568357
#> 668 334 2003-11-30 0.61344839  7.719011  6.715800  1.945810
#> 670 335 2003-12-01 1.86366704  7.685684  6.630775  2.461627
#> 672 336 2003-12-02 5.18185073  7.653793  6.550072  3.854817
#> 673 337 2003-12-03 3.74016836  7.623368  6.473680  3.206475
#> 675 338 2003-12-04 3.32863548  7.594437  6.401590  3.003303
#> 678 339 2003-12-05 1.42888268  7.567027  6.333789  2.181452
#> 680 340 2003-12-06 2.44315115  7.541163  6.270268  2.583273
#> 682 341 2003-12-07 0.82921767  7.516872  6.211016  1.895336
#> 683 342 2003-12-08 5.20518093  7.494177  6.156023  3.676880
#> 686 343 2003-12-09 2.88560994  7.473100  6.105278  2.705044
#> 688 344 2003-12-10 3.96462149  7.453664  6.058772  3.126031
#> 690 345 2003-12-11 0.12494967  7.435887  6.016497  1.554674
#> 691 346 2003-12-12 3.28691284  7.419789  5.978443  2.818814
#> 694 347 2003-12-13 2.24987743  7.405386  5.944603  2.389184
#> 696 348 2003-12-14 1.94252978  7.392694  5.914970  2.255862
#> 698 349 2003-12-15 3.23223833  7.381725  5.889539  2.761812
#> 699 350 2003-12-16 1.00146772  7.372493  5.868303  1.865646
#> 701 351 2003-12-17 3.29568326  7.365006  5.851259  2.771972
#> 704 352 2003-12-18 4.85437867  7.359273  5.838404  3.385187
#> 706 353 2003-12-19 3.57671756  7.355300  5.829736  2.874869
#> 708 354 2003-12-20 5.76014524  7.353091  5.825253  3.737959
#> 709 355 2003-12-21 0.48302121  7.352649  5.824956  1.647570
#> 712 356 2003-12-22 6.67302760  7.353975  5.828845  4.101771
#> 714 357 2003-12-23 5.62080526  7.357066  5.836924  3.688938
#> 716 358 2003-12-24 4.53916610  7.361920  5.849194  3.265528
#> 717 359 2003-12-25 6.81737215  7.368531  5.865662  4.179873
#> 720 360 2003-12-26 4.13856635  7.376892  5.886331  3.122751
#> 722 361 2003-12-27 3.83624688  7.386993  5.911209  3.012720
#> 724 362 2003-12-28 4.96262396  7.398825  5.940303  3.477250
#> 725 363 2003-12-29 5.05177740  7.412375  5.973622  3.529016
#> 728 364 2003-12-30 2.09577172  7.427627  6.011174  2.350847
#> 730 365 2003-12-31 5.70615258  7.444567  6.052971  3.833000