Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.911189  3.368655  2.593628  3.314579  3.546738  2.917817  2.741659
#>   [8]  1.823106  3.107628  4.529526  3.461743  3.718174  4.160444  2.206255
#>  [15]  4.648996  2.477495  3.823971  5.110693  3.559419  2.050214  5.270312
#>  [22]  3.517825  3.609351  3.678931  3.862794  4.535494  3.697307  3.083692
#>  [29]  3.873522  2.546434  3.814317  5.404939  4.446457  2.952818  3.141137
#>  [36]  5.008561  4.694266  5.642326  3.587543  7.057061  3.727736  3.760745
#>  [43]  6.940573  6.559936  6.430907  5.504212  7.013805  3.368581  6.502488
#>  [50]  7.732943  6.241947  7.401157  7.896869  4.796361  8.163011  5.303730
#>  [57]  6.933370  8.534607  5.022482  4.409647  8.596759  9.365774  9.934329
#>  [64]  5.724050  7.161945  8.751780  8.163218  9.392371  7.708626  8.034661
#>  [71]  5.441767  6.400111 11.137091  9.651431 10.540098  8.878882  9.849611
#>  [78]  9.158972  9.854829  6.984899  6.189486  7.997456  8.822792  8.266978
#>  [85]  6.636544  8.363241  7.801785 11.616553 12.451203  7.608400  7.607534
#>  [92] 13.582566 10.040758 11.890451 11.579654 11.131774  9.806930 11.669364
#>  [99] 13.300728 14.158130  8.733873  8.222940  9.262188 14.607970 11.354517
#> [106] 10.976471 13.231334 12.194581 12.048897 14.718567 12.386083 15.806292
#> [113] 14.697634 15.062882 13.608431 13.935568 15.875530 14.996176 13.964124
#> [120] 15.990156 11.954627 12.775872 16.607118 13.898275 16.291305 15.348875
#> [127] 10.849364 11.898074  9.673026 12.787799 12.659469 14.542671 15.642180
#> [134] 14.793656 10.609375 17.516235  9.617591 17.727938 17.397161 18.335511
#> [141] 17.643587 16.096892 17.421963 14.380040 14.240082 15.396048 16.594008
#> [148] 18.080151 14.194679 10.292733 18.269380 11.301686 16.166239 16.290787
#> [155] 17.525820 12.162329 15.903419 13.973712 15.840424 15.823645 13.561326
#> [162] 12.751195 18.376724 12.520023 13.518113 13.048396 15.982845 12.691911
#> [169] 11.343438 11.062775 11.661377 16.921677 14.587813 17.579686 10.559865
#> [176] 18.200826 17.980136 13.049577 15.891512 17.810518 18.532242 16.316082
#> [183] 16.483835 13.615422 11.120015 10.627806 11.393102 14.608171 17.438452
#> [190] 15.740973 18.764118 11.499795 16.622241 13.890843 12.450569 13.250360
#> [197] 12.333994 13.367657 14.840351 16.723728 12.767062 11.206180 13.323471
#> [204] 11.605042 16.946435 13.472593 16.852146 13.726545 10.378331 12.363686
#> [211] 12.976371 14.994260 14.202264 13.264888  9.233000 16.676842 15.697228
#> [218] 15.098804  9.780219 14.745607 12.429924 16.423338 15.328548 11.290023
#> [225] 10.531651 11.587019 16.057624 10.809129  9.043064 10.223463 10.238353
#> [232] 14.777290 15.616524 14.129176 11.699542 11.982070 14.780306  8.890759
#> [239] 14.472766 10.291501 10.658964 12.766342 12.924186 11.421117  9.007412
#> [246]  9.261448  9.936598  9.018606  7.941574  7.494310 10.669067  8.909056
#> [253] 11.289558  9.978966  8.047219 11.691379  6.946521 12.477365 11.672029
#> [260] 12.648347 11.037974  9.119843  6.873027  6.571407  5.944908  6.303442
#> [267] 11.736687  6.315880 10.270640  6.254396  5.773243  6.758651  6.177428
#> [274]  5.213768  7.329147  5.931059  7.511329 10.124011  9.495241  6.143556
#> [281]  4.890146  6.118737  8.435372  8.246575  9.057807  6.692664  7.689976
#> [288]  8.300935  7.159873  5.184576  8.011299  7.207634  6.925242  4.202616
#> [295]  4.220063  3.756362  7.133351  5.073612  4.867144  4.916879  6.654072
#> [302]  5.799409  3.298401  2.910682  2.850060  4.690060  3.181898  6.360667
#> [309]  5.320003  5.306723  4.792052  2.812901  2.998057  3.003935  4.587042
#> [316]  5.353142  5.015167  4.101713  3.668094  4.019028  5.498144  4.427863
#> [323]  4.036999  3.945650  5.169590  2.909425  2.635097  2.899548  3.872926
#> [330]  3.927312  4.666932  2.116098  3.468282  3.115128  3.161513  3.682185
#> [337]  1.786553  4.171534  2.360612  2.950300  1.850512  3.706876  4.353926
#> [344]  3.167681  2.359769  2.536744  2.954598  2.667001  2.593257  2.463906
#> [351]  2.193751  2.195185  3.301235  3.106513  3.350764  2.521079  3.368768
#> [358]  4.049844  2.526365  2.324646  2.693182  1.573785  4.093953  4.227968
#> [365]  2.514577  1.911189  3.368655  2.593628  3.314579  3.546738  2.917817
#> [372]  2.741659  1.823106  3.107628  4.529526  3.461743  3.718174  4.160444
#> [379]  2.206255  4.648996  2.477495  3.823971  5.110693  3.559419  2.050214
#> [386]  5.270312  3.517825  3.609351  3.678931  3.862794  4.535494  3.697307
#> [393]  3.083692  3.873522  2.546434  3.814317  5.404939  4.446457  2.952818
#> [400]  3.141137  5.008561  4.694266  5.642326  3.587543  7.057061  3.727736
#> [407]  3.760745  6.940573  6.559936  6.430907  5.504212  7.013805  3.368581
#> [414]  6.502488  7.732943  6.241947  7.401157  7.896869  4.796361  8.163011
#> [421]  5.303730  6.933370  8.534607  5.022482  4.409647  8.596759  9.365774
#> [428]  9.934329  5.724050  7.161945  8.751780  8.163218  9.392371  7.708626
#> [435]  8.034661  5.441767  6.400111 11.137091  9.651431 10.540098  8.878882
#> [442]  9.849611  9.158972  9.854829  6.984899  6.189486  7.997456  8.822792
#> [449]  8.266978  6.636544  8.363241  7.801785 11.616553 12.451203  7.608400
#> [456]  7.607534 13.582566 10.040758 11.890451 11.579654 11.131774  9.806930
#> [463] 11.669364 13.300728 14.158130  8.733873  8.222940  9.262188 14.607970
#> [470] 11.354517 10.976471 13.231334 12.194581 12.048897 14.718567 12.386083
#> [477] 15.806292 14.697634 15.062882 13.608431 13.935568 15.875530 14.996176
#> [484] 13.964124 15.990156 11.954627 12.775872 16.607118 13.898275 16.291305
#> [491] 15.348875 10.849364 11.898074  9.673026 12.787799 12.659469 14.542671
#> [498] 15.642180 14.793656 10.609375 17.516235  9.617591 17.727938 17.397161
#> [505] 18.335511 17.643587 16.096892 17.421963 14.380040 14.240082 15.396048
#> [512] 16.594008 18.080151 14.194679 10.292733 18.269380 11.301686 16.166239
#> [519] 16.290787 17.525820 12.162329 15.903419 13.973712 15.840424 15.823645
#> [526] 13.561326 12.751195 18.376724 12.520023 13.518113 13.048396 15.982845
#> [533] 12.691911 11.343438 11.062775 11.661377 16.921677 14.587813 17.579686
#> [540] 10.559865 18.200826 17.980136 13.049577 15.891512 17.810518 18.532242
#> [547] 16.316082 16.483835 13.615422 11.120015 10.627806 11.393102 14.608171
#> [554] 17.438452 15.740973 18.764118 11.499795 16.622241 13.890843 12.450569
#> [561] 13.250360 12.333994 13.367657 14.840351 16.723728 12.767062 11.206180
#> [568] 13.323471 11.605042 16.946435 13.472593 16.852146 13.726545 10.378331
#> [575] 12.363686 12.976371 14.994260 14.202264 13.264888  9.233000 16.676842
#> [582] 15.697228 15.098804  9.780219 14.745607 12.429924 16.423338 15.328548
#> [589] 11.290023 10.531651 11.587019 16.057624 10.809129  9.043064 10.223463
#> [596] 10.238353 14.777290 15.616524 14.129176 11.699542 11.982070 14.780306
#> [603]  8.890759 14.472766 10.291501 10.658964 12.766342 12.924186 11.421117
#> [610]  9.007412  9.261448  9.936598  9.018606  7.941574  7.494310 10.669067
#> [617]  8.909056 11.289558  9.978966  8.047219 11.691379  6.946521 12.477365
#> [624] 11.672029 12.648347 11.037974  9.119843  6.873027  6.571407  5.944908
#> [631]  6.303442 11.736687  6.315880 10.270640  6.254396  5.773243  6.758651
#> [638]  6.177428  5.213768  7.329147  5.931059  7.511329 10.124011  9.495241
#> [645]  6.143556  4.890146  6.118737  8.435372  8.246575  9.057807  6.692664
#> [652]  7.689976  8.300935  7.159873  5.184576  8.011299  7.207634  6.925242
#> [659]  4.202616  4.220063  3.756362  7.133351  5.073612  4.867144  4.916879
#> [666]  6.654072  5.799409  3.298401  2.910682  2.850060  4.690060  3.181898
#> [673]  6.360667  5.320003  5.306723  4.792052  2.812901  2.998057  3.003935
#> [680]  4.587042  5.353142  5.015167  4.101713  3.668094  4.019028  5.498144
#> [687]  4.427863  4.036999  3.945650  5.169590  2.909425  2.635097  2.899548
#> [694]  3.872926  3.927312  4.666932  2.116098  3.468282  3.115128  3.161513
#> [701]  3.682185  1.786553  4.171534  2.360612  2.950300  1.850512  3.706876
#> [708]  4.353926  3.167681  2.359769  2.536744  2.954598  2.667001  2.593257
#> [715]  2.463906  2.193751  2.195185  3.301235  3.106513  3.350764  2.521079
#> [722]  3.368768  4.049844  2.526365  2.324646  2.693182  1.573785  4.093953
#> [729]  4.227968  2.514577
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 4.36626400  7.463176  6.099033  3.308851
#> 3     2 2002-01-02 2.09345375  7.483435  6.149345  2.397460
#> 6     3 2002-01-03 3.45797883  7.505323  6.203934  2.980174
#> 7     4 2002-01-04 3.35945646  7.528820  6.262814  2.962978
#> 9     5 2002-01-05 4.69472922  7.553900  6.325996  3.547294
#> 11    6 2002-01-06 2.54066738  7.580540  6.393496  2.669785
#> 14    7 2002-01-07 2.55523060  7.608714  6.465324  2.701954
#> 15    8 2002-01-08 4.08010622  7.638396  6.541496  3.382468
#> 17    9 2002-01-09 3.77962228  7.669557  6.622023  3.287200
#> 19   10 2002-01-10 6.53908849  7.702170  6.706920  4.523794
#> 22   11 2002-01-11 4.56943240  7.736204  6.796198  3.706156
#> 23   12 2002-01-12 3.78038265  7.771631  6.889871  3.398200
#> 25   13 2002-01-13 6.20123572  7.808419  6.987950  4.521808
#> 27   14 2002-01-14 2.62188763  7.846537  7.090446  2.957233
#> 30   15 2002-01-15 1.80467962  7.885954  7.197371  2.622892
#> 32   16 2002-01-16 0.22559585  7.926639  7.308732  1.931188
#> 33   17 2002-01-17 4.79394651  7.968559  7.424539  4.089465
#> 35   18 2002-01-18 4.28846728  8.011681  7.544799  3.905478
#> 37   19 2002-01-19 3.19238786  8.055974  7.669519  3.437002
#> 40   20 2002-01-20 5.29245948  8.101405  7.798703  4.497031
#> 41   21 2002-01-21 6.03323255  8.147942  7.932354  4.919887
#> 43   22 2002-01-22 4.96260846  8.195553  8.070474  4.461053
#> 45   23 2002-01-23 1.76148943  8.244205  8.213063  2.930683
#> 48   24 2002-01-24 0.94737659  8.293867  8.360119  2.567502
#> 49   25 2002-01-25 5.07180608  8.344507  8.511639  4.714605
#> 51   26 2002-01-26 6.73268425  8.396093  8.667618  5.642112
#> 53   27 2002-01-27 3.35836539  8.448595  8.828046  3.961611
#> 56   28 2002-01-28 5.80326946  8.501983  8.992915  5.317414
#> 58   29 2002-01-29 4.92946742  8.556225  9.162213  4.929850
#> 59   30 2002-01-30 6.63069089  8.611293  9.335925  5.928310
#> 61   31 2002-01-31 2.12052614  8.667157  9.514035  3.542371
#> 63   32 2002-02-01 6.64053598  8.723789  9.696523  6.114621
#> 66   33 2002-02-02 2.61309101  8.781160  9.883368  3.941385
#> 67   34 2002-02-03 0.16650115  8.839243 10.074546  2.613522
#> 69   35 2002-02-04 2.10628883  8.898011 10.270030  3.783040
#> 71   36 2002-02-05 5.25065551  8.957438 10.469790  5.686029
#> 74   37 2002-02-06 0.26477067  9.017498 10.673795  2.825150
#> 75   38 2002-02-07 4.56422103  9.078165 10.882010  5.456072
#> 77   39 2002-02-08 1.01349596  9.139416 11.094398  3.388744
#> 79   40 2002-02-09 4.57964102  9.201226 11.310918  5.642569
#> 82   41 2002-02-10 1.09521968  9.263572 11.531528  3.564561
#> 83   42 2002-02-11 0.14671444  9.326432 11.756182  3.031514
#> 86   43 2002-02-12 0.51273036  9.389784 11.984831  3.323424
#> 87   44 2002-02-13 6.33677574  9.453606 12.217425  7.149042
#> 89   45 2002-02-14 0.81345621  9.517878 12.453910  3.645671
#> 91   46 2002-02-15 1.36338543  9.582579 12.694229  4.076609
#> 94   47 2002-02-16 1.02449984  9.647690 12.938322  3.921549
#> 95   48 2002-02-17 0.40055940  9.713192 13.186128  3.568421
#> 97   49 2002-02-18 6.94652528  9.779067 13.437583  8.132065
#> 99   50 2002-02-19 6.95574367  9.845297 13.692618  8.260101
#> 102  51 2002-02-20 1.29044912  9.911864 13.951164  4.395958
#> 103  52 2002-02-21 1.96061111  9.978754 14.213148  4.949576
#> 105  53 2002-02-22 5.28464060 10.045948 14.478496  7.427808
#> 107  54 2002-02-23 1.34983946 10.113433 14.747130  4.670932
#> 110  55 2002-02-24 4.30295599 10.181192 15.018971  6.928535
#> 112  56 2002-02-25 5.28904828 10.249211 15.293936  7.769659
#> 113  57 2002-02-26 2.12483565 10.317477 15.571940  5.496469
#> 115  58 2002-02-27 4.34539211 10.385975 15.852898  7.279574
#> 117  59 2002-02-28 5.42596463 10.454693 16.136720  8.221643
#> 120  60 2002-03-01 1.89843644 10.523618 16.423316  5.587193
#> 121  61 2002-03-02 5.87674747 10.592738 16.712592  8.814139
#> 123  62 2002-03-03 2.77179545 10.662040 17.004454  6.461425
#> 125  63 2002-03-04 5.10545370 10.731513 17.298805  8.439603
#> 128  64 2002-03-05 3.00522124 10.801147 17.595546  6.846706
#> 129  65 2002-03-06 4.78712388 10.870930 17.894578  8.413674
#> 131  66 2002-03-07 6.36441162 10.940852 18.195799  9.841296
#> 133  67 2002-03-08 6.67807721 11.010902 18.499104 10.234600
#> 136  68 2002-03-09 2.03796164 11.081072 18.804391  6.430291
#> 138  69 2002-03-10 3.28930250 11.151350 19.111552  7.596546
#> 139  70 2002-03-11 1.46814907 11.221729 19.420480  6.125520
#> 141  71 2002-03-12 6.74694803 11.292198 19.731067 10.827300
#> 143  72 2002-03-13 0.73286707 11.362749 20.043203  5.657167
#> 146  73 2002-03-14 5.47836462 11.433374 20.356778  9.966226
#> 147  74 2002-03-15 3.93978591 11.504062 20.671679  8.707626
#> 149  75 2002-03-16 3.32472818 11.574807 20.987796  8.261198
#> 151  76 2002-03-17 2.28889192 11.645600 21.305014  7.419957
#> 154  77 2002-03-18 4.89286453 11.716433 21.623220  9.920809
#> 155  78 2002-03-19 4.57684112 11.787297 21.942300  9.745518
#> 157  79 2002-03-20 2.36800359 11.858185 22.262140  7.788338
#> 159  80 2002-03-21 1.26843107 11.929088 22.582624  6.846272
#> 162  81 2002-03-22 2.99330070 12.000000 22.903638  8.582471
#> 163  82 2002-03-23 4.13169562 12.070912 23.225065  9.781066
#> 166  83 2002-03-24 6.15713202 12.141815 23.546791 11.857004
#> 167  84 2002-03-25 1.15230151 12.212703 23.868700  7.093213
#> 169  85 2002-03-26 6.97405143 12.283567 24.190677 12.914853
#> 171  86 2002-03-27 3.19374751 12.354400 24.512606  9.296540
#> 174  87 2002-03-28 5.84916739 12.425193 24.834373 12.053992
#> 175  88 2002-03-29 1.64530302 12.495938 25.155863  7.945065
#> 177  89 2002-03-30 5.47355409 12.566626 25.476961 11.917647
#> 179  90 2002-03-31 4.14159515 12.637251 25.797553 10.676693
#> 182  91 2002-04-01 2.39057153 12.707802 26.117528  8.985976
#> 183  92 2002-04-02 1.11130048 12.778271 26.436771  7.758769
#> 185  93 2002-04-03 2.58348276 12.848650 26.755170  9.378629
#> 187  94 2002-04-04 6.56473279 12.918928 27.072615 13.646607
#> 190  95 2002-04-05 2.64156914 12.989098 27.388994  9.632274
#> 192  96 2002-04-06 0.41718371 13.059148 27.704199  7.368565
#> 193  97 2002-04-07 6.28415077 13.129070 28.018120 13.709883
#> 195  98 2002-04-08 0.50342726 13.198853 28.330649  7.622952
#> 197  99 2002-04-09 5.18404378 13.268487 28.641680 12.755622
#> 200 100 2002-04-10 4.08834679 13.337960 28.951106 11.674818
#> 201 101 2002-04-11 2.65321283 13.407262 29.258822 10.209774
#> 203 102 2002-04-12 4.97133531 13.476382 29.564726 12.844283
#> 205 103 2002-04-13 0.36968458 13.545307 29.868714  7.874774
#> 208 104 2002-04-14 6.18753173 13.614025 30.170685 14.398912
#> 209 105 2002-04-15 0.77844743 13.682523 30.470539  8.484423
#> 211 106 2002-04-16 2.99414673 13.750789 30.768176 11.041831
#> 213 107 2002-04-17 0.67874501 13.818808 31.063500  8.528755
#> 216 108 2002-04-18 3.96015076 13.886567 31.356413 12.310191
#> 218 109 2002-04-19 5.68284712 13.954052 31.646820 14.355856
#> 219 110 2002-04-20 4.50922066 14.021246 31.934628 13.118731
#> 221 111 2002-04-21 6.59433284 14.088136 32.219743 15.595597
#> 223 112 2002-04-22 4.66460017 14.154703 32.502075 13.480954
#> 226 113 2002-04-23 2.01867679 14.220933 32.781534 10.522070
#> 227 114 2002-04-24 4.85377832 14.286808 33.058030 13.880050
#> 229 115 2002-04-25 6.65064622 14.352310 33.331477 16.055524
#> 231 116 2002-04-26 0.87334810 14.417421 33.601789  9.418176
#> 234 117 2002-04-27 5.36348063 14.482122 33.868881 14.738921
#> 235 118 2002-04-28 6.32832904 14.546394 34.132670 15.957783
#> 237 119 2002-04-29 4.34512271 14.610216 34.393074 13.712571
#> 239 120 2002-04-30 5.32850257 14.673568 34.650012 14.953839
#> 242 121 2002-05-01 2.96170166 14.736428 34.903405 12.233264
#> 243 122 2002-05-02 1.49131136 14.798774 35.153175 10.559533
#> 245 123 2002-05-03 1.81317932 14.860584 35.399245 11.009389
#> 248 124 2002-05-04 1.94253376 14.921835 35.641540 11.230304
#> 249 125 2002-05-05 2.18263013 14.982502 35.879987 11.583470
#> 251 126 2002-05-06 0.99305203 15.042562 36.114511 10.220698
#> 253 127 2002-05-07 4.34639720 15.101989 36.345041 14.316366
#> 256 128 2002-05-08 6.34839382 15.160757 36.571508 16.799828
#> 257 129 2002-05-09 6.63850837 15.218840 36.793843 17.223259
#> 259 130 2002-05-10 5.47340235 15.276211 37.011977 15.883612
#> 261 131 2002-05-11 0.33814163 15.332843 37.225844  9.716939
#> 264 132 2002-05-12 4.48936879 15.388707 37.435378 14.819382
#> 265 133 2002-05-13 6.24159806 15.443775 37.640517 17.016332
#> 267 134 2002-05-14 6.72781681 15.498017 37.841197 17.673887
#> 269 135 2002-05-15 1.35015154 15.551405 38.037357 11.160515
#> 272 136 2002-05-16 3.21911913 15.603907 38.228936 13.500589
#> 274 137 2002-05-17 6.05309213 15.655493 38.415875 17.030590
#> 275 138 2002-05-18 4.54218652 15.706133 38.598118 15.230784
#> 277 139 2002-05-19 2.58407349 15.755795 38.775606 12.873653
#> 279 140 2002-05-20 1.31553101 15.804447 38.948286 11.358061
#> 282 141 2002-05-21 1.32482392 15.852058 39.116103 11.413575
#> 283 142 2002-05-22 1.63242736 15.898595 39.279005 11.836286
#> 285 143 2002-05-23 3.11864000 15.944026 39.436940 13.716154
#> 287 144 2002-05-24 1.50747718 15.988319 39.589859 11.763853
#> 290 145 2002-05-25 3.98895657 16.031441 39.737713 14.878214
#> 291 146 2002-05-26 5.38682515 16.073361 39.880455 16.652880
#> 293 147 2002-05-27 3.60951645 16.114046 40.018039 14.486493
#> 295 148 2002-05-28 0.55642230 16.153463 40.150421 10.729116
#> 298 149 2002-05-29 4.12807312 16.191581 40.277558 15.203808
#> 300 150 2002-05-30 5.23678831 16.228369 40.399408 16.618164
#> 301 151 2002-05-31 1.85956505 16.263796 40.515933 12.445233
#> 303 152 2002-06-01 5.34051849 16.297830 40.627093 16.813173
#> 305 153 2002-06-02 3.98422535 16.330443 40.732853 15.152119
#> 308 154 2002-06-03 1.89457834 16.361604 40.833176 12.572416
#> 309 155 2002-06-04 2.33506356 16.391286 40.928031 13.147263
#> 311 156 2002-06-05 6.91669636 16.419460 41.017385 18.893632
#> 313 157 2002-06-06 6.20263958 16.446100 41.101209 18.025954
#> 316 158 2002-06-07 4.10819738 16.471180 41.179475 15.430306
#> 317 159 2002-06-08 0.21793018 16.494677 41.252156 10.585554
#> 319 160 2002-06-09 3.48878880 16.516565 41.319229 14.693731
#> 321 161 2002-06-10 3.74341709 16.536824 41.380671 15.028809
#> 324 162 2002-06-11 1.36686415 16.555433 41.436461 12.069672
#> 325 163 2002-06-12 2.24517994 16.572373 41.486582 13.181890
#> 328 164 2002-06-13 3.60873883 16.587625 41.531016 14.900417
#> 329 165 2002-06-14 5.93034393 16.601175 41.569750 17.817301
#> 331 166 2002-06-15 1.47395073 16.613007 41.602770 12.246248
#> 333 167 2002-06-16 2.39127773 16.623108 41.630068 13.401814
#> 336 168 2002-06-17 3.77381908 16.631469 41.651634 15.138460
#> 337 169 2002-06-18 3.40133470 16.638080 41.667464 14.675921
#> 339 170 2002-06-19 4.95355816 16.642934 41.677552 16.621785
#> 341 171 2002-06-20 3.50561643 16.646025 41.681897 14.809532
#> 344 172 2002-06-21 4.45383999 16.647351 41.680500 15.995735
#> 345 173 2002-06-22 6.57291712 16.646909 41.673363 18.645561
#> 347 174 2002-06-23 5.07157075 16.644700 41.660491 16.762012
#> 349 175 2002-06-24 5.90446207 16.640727 41.641891 17.798160
#> 352 176 2002-06-25 4.95372417 16.634994 41.617572 16.601028
#> 354 177 2002-06-26 2.15971304 16.627507 41.587545 13.097747
#> 355 178 2002-06-27 1.87012605 16.618275 41.551823 12.725959
#> 357 179 2002-06-28 6.65200293 16.607306 41.510423 18.691039
#> 359 180 2002-06-29 2.89167982 16.594614 41.463360 13.978421
#> 362 181 2002-06-30 6.86868665 16.580211 41.410655 18.930263
#> 363 182 2002-07-01 0.61054516 16.564113 41.352328 11.100195
#> 365 183 2002-07-02 6.69778817 16.546336 41.288405 18.678664
#> 367 184 2002-07-03 0.19873115 16.526900 41.218909 10.552550
#> 370 185 2002-07-04 2.31349617 16.505823 41.143868 13.169379
#> 371 186 2002-07-05 6.75402869 16.483128 41.063311 18.678757
#> 373 187 2002-07-06 4.29983100 16.458837 40.977269 15.596923
#> 375 188 2002-07-07 5.23323861 16.432973 40.885775 16.731678
#> 378 189 2002-07-08 0.37067786 16.405563 40.788864 10.658021
#> 380 190 2002-07-09 3.12682734 16.376632 40.686570 14.055832
#> 381 191 2002-07-10 5.34448532 16.346207 40.578932 16.778490
#> 383 192 2002-07-11 5.88964210 16.314316 40.465990 17.420825
#> 385 193 2002-07-12 3.83499181 16.280989 40.347784 14.838912
#> 388 194 2002-07-13 5.78828803 16.246255 40.224356 17.221746
#> 389 195 2002-07-14 0.43148854 16.210144 40.095751 10.557581
#> 391 196 2002-07-15 0.04715465 16.172686 39.962013 10.048762
#> 393 197 2002-07-16 2.73103470 16.133915 39.823189 13.326291
#> 396 198 2002-07-17 4.69658844 16.093860 39.679327 15.709539
#> 397 199 2002-07-18 0.29873699 16.052554 39.530476 10.250449
#> 399 200 2002-07-19 3.61517324 16.010029 39.376687 14.289934
#> 401 201 2002-07-20 3.76218287 15.966317 39.218012 14.425022
#> 404 202 2002-07-21 3.03665221 15.921451 39.054503 13.488002
#> 405 203 2002-07-22 5.35502596 15.875462 38.886214 16.279998
#> 408 204 2002-07-23 5.67490724 15.828384 38.713200 16.618169
#> 409 205 2002-07-24 4.44277747 15.780249 38.535519 15.058532
#> 411 206 2002-07-25 5.78122432 15.731088 38.353226 16.635772
#> 413 207 2002-07-26 0.88871100 15.680933 38.166382 10.623128
#> 416 208 2002-07-27 4.12141165 15.629816 37.975044 14.500563
#> 417 209 2002-07-28 6.34850874 15.577768 37.779274 17.143035
#> 419 210 2002-07-29 4.00870011 15.524820 37.579134 14.246482
#> 421 211 2002-07-30 2.83672761 15.471001 37.374685 12.770140
#> 424 212 2002-07-31 4.23434447 15.416342 37.165992 14.395615
#> 425 213 2002-08-01 3.06310310 15.360873 36.953119 12.922680
#> 427 214 2002-08-02 1.01109499 15.304621 36.736132 10.397513
#> 429 215 2002-08-03 5.11565519 15.247616 36.515097 15.254277
#> 432 216 2002-08-04 1.81667129 15.189886 36.290082 11.242621
#> 434 217 2002-08-05 5.33921030 15.131457 36.061156 15.377469
#> 435 218 2002-08-06 3.48020302 15.072356 35.828388 13.093479
#> 437 219 2002-08-07 0.68225764 15.012610 35.591848  9.706709
#> 439 220 2002-08-08 1.06728051 14.952243 35.351609 10.099589
#> 442 221 2002-08-09 0.33370697 14.891281 35.107744  9.170310
#> 443 222 2002-08-10 2.87307041 14.829748 34.860325 12.091948
#> 445 223 2002-08-11 5.41143185 14.767666 34.609429 14.993459
#> 447 224 2002-08-12 4.51219456 14.705060 34.355129 13.859656
#> 450 225 2002-08-13 4.19728429 14.641952 34.097505 13.411598
#> 451 226 2002-08-14 6.19384931 14.578362 33.836632 15.647174
#> 453 227 2002-08-15 4.76504788 14.514313 33.572591 13.904087
#> 455 228 2002-08-16 2.31412261 14.449824 33.305462 10.993281
#> 458 229 2002-08-17 0.04534070 14.384916 33.035326  8.310894
#> 460 230 2002-08-18 1.34033415 14.319607 32.762264  9.723862
#> 461 231 2002-08-19 5.77687191 14.253916 32.486361 14.704677
#> 463 232 2002-08-20 2.15027117 14.187862 32.207701 10.492578
#> 465 233 2002-08-21 4.59289875 14.121461 31.926370 13.173498
#> 468 234 2002-08-22 0.74375596 14.054730 31.642455  8.747850
#> 469 235 2002-08-23 6.82458695 13.987686 31.356043 15.488311
#> 471 236 2002-08-24 3.67760737 13.920345 31.067223 11.870621
#> 473 237 2002-08-25 3.77805048 13.852721 30.776085 11.890800
#> 476 238 2002-08-26 4.44685400 13.784830 30.482721 12.537397
#> 477 239 2002-08-27 0.41547005 13.716686 30.187223  8.003982
#> 479 240 2002-08-28 1.12680966 13.648302 29.889684  8.706274
#> 481 241 2002-08-29 4.71069886 13.579692 29.590199 12.529865
#> 484 242 2002-08-30 0.31413087 13.510869 29.288864  7.662702
#> 485 243 2002-08-31 2.74398736 13.441846 28.985774 10.204988
#> 487 244 2002-09-01 1.57795695 13.372633 28.681027  8.862422
#> 490 245 2002-09-02 4.31275607 13.303244 28.374722 11.693057
#> 491 246 2002-09-03 4.77823861 13.233689 28.066958 12.083755
#> 493 247 2002-09-04 0.78511327 13.163980 27.757836  7.767212
#> 495 248 2002-09-05 0.08157123 13.094126 27.447457  6.947358
#> 498 249 2002-09-06 1.45943466 13.024138 27.135924  8.304354
#> 499 250 2002-09-07 0.83957396 12.954027 26.823338  7.575069
#> 501 251 2002-09-08 2.27023718 12.883802 26.509805  8.963080
#> 503 252 2002-09-09 4.78701999 12.813472 26.195428 11.442068
#> 506 253 2002-09-10 1.02137036 12.743047 25.880314  7.507247
#> 507 254 2002-09-11 1.37963771 12.672536 25.564567  7.782727
#> 509 255 2002-09-12 1.10468022 12.601947 25.248295  7.418700
#> 511 256 2002-09-13 0.11947378 12.531289 24.931604  6.351750
#> 514 257 2002-09-14 5.82756766 12.460572 24.614602 11.909536
#> 516 258 2002-09-15 5.56800619 12.389802 24.297397 11.534003
#> 517 259 2002-09-16 6.38872366 12.318988 23.980097 12.213157
#> 519 260 2002-09-17 1.91331186 12.248139 23.662812  7.763916
#> 521 261 2002-09-18 3.33153604 12.177262 23.345649  9.029941
#> 524 262 2002-09-19 2.88354709 12.106365 23.028718  8.499720
#> 525 263 2002-09-20 4.93443682 12.035456 22.712128 10.333924
#> 527 264 2002-09-21 4.08230557 11.964544 22.395989  9.419756
#> 529 265 2002-09-22 2.09706334 11.893635 22.080409  7.466690
#> 532 266 2002-09-23 6.73647128 11.822738 21.765497 11.642249
#> 533 267 2002-09-24 2.31702383 11.751861 21.451363  7.477540
#> 535 268 2002-09-25 4.24953963 11.681012 21.138114  9.129540
#> 537 269 2002-09-26 4.23570379 11.610198 20.825859  9.005373
#> 540 270 2002-09-27 2.05825203 11.539428 20.514705  6.958248
#> 542 271 2002-09-28 4.75397696 11.468711 20.204761  9.238800
#> 543 272 2002-09-29 0.23621563 11.398053 19.896131  5.180199
#> 545 273 2002-09-30 4.27447636 11.327464 19.588923  8.593221
#> 547 274 2002-10-01 5.94297409 11.256953 19.283241  9.910989
#> 550 275 2002-10-02 5.29325569 11.186528 18.979189  9.235096
#> 551 276 2002-10-03 5.89624537 11.116198 18.676871  9.622504
#> 553 277 2002-10-04 0.38241835 11.045973 18.376389  4.912198
#> 555 278 2002-10-05 1.40761226 10.975862 18.077844  5.678668
#> 558 279 2002-10-06 4.90662005 10.905874 17.781337  8.445301
#> 559 280 2002-10-07 3.00150033 10.836020 17.486965  6.793623
#> 561 281 2002-10-08 1.08532853 10.766311 17.194826  5.165393
#> 563 282 2002-10-09 4.24790299 10.696756 16.905016  7.582920
#> 566 283 2002-10-10 6.75439153 10.627367 16.617629  9.435206
#> 567 284 2002-10-11 0.67541046 10.558154 16.332759  4.605597
#> 570 285 2002-10-12 5.76710543 10.489131 16.050495  8.425044
#> 571 286 2002-10-13 4.33586197 10.420308 15.770928  7.223852
#> 573 287 2002-10-14 6.86603206 10.351698 15.494145  9.011983
#> 575 288 2002-10-15 2.88056706 10.283314 15.220231  5.936807
#> 578 289 2002-10-16 1.80764097 10.215170 14.949271  5.060003
#> 579 290 2002-10-17 0.53258179 10.147279 14.681345  4.055613
#> 581 291 2002-10-18 6.99717708 10.079655 14.416534  8.608027
#> 583 292 2002-10-19 2.79369535 10.012314 14.154915  5.513523
#> 586 293 2002-10-20 4.46797838  9.945270 13.896562  6.595702
#> 587 294 2002-10-21 0.84603300  9.878539 13.641550  3.994543
#> 589 295 2002-10-22 5.42690545  9.812138 13.389948  7.050349
#> 591 296 2002-10-23 5.68022001  9.746084 13.141825  7.115121
#> 594 297 2002-10-24 3.57442563  9.680393 12.897247  5.605426
#> 596 298 2002-10-25 3.63920037  9.615084 12.656277  5.559198
#> 597 299 2002-10-26 5.71119467  9.550176 12.418977  6.818141
#> 599 300 2002-10-27 2.35046263  9.485687 12.185404  4.556064
#> 601 301 2002-10-28 0.57500944  9.421638 11.955615  3.353734
#> 604 302 2002-10-29 2.43792060  9.358048 11.729663  4.460298
#> 605 303 2002-10-30 0.28854526  9.294940 11.507600  3.055517
#> 607 304 2002-10-31 1.76796768  9.232334 11.289473  3.903321
#> 609 305 2002-11-01 4.88595578  9.170252 11.075328  5.719327
#> 612 306 2002-11-02 0.91360995  9.108719 10.865209  3.261196
#> 613 307 2002-11-03 5.12001628  9.047757 10.659156  5.680733
#> 615 308 2002-11-04 1.92779609  8.987390 10.457208  3.735838
#> 617 309 2002-11-05 0.42357854  8.927644 10.259398  2.808232
#> 620 310 2002-11-06 5.82225235  8.868543 10.065762  5.820557
#> 622 311 2002-11-07 6.20616280  8.810114  9.876328  5.947703
#> 623 312 2002-11-08 5.36534590  8.752384  9.691126  5.393186
#> 625 313 2002-11-09 4.84699297  8.695379  9.510180  5.028135
#> 627 314 2002-11-10 0.30526730  8.639127  9.333513  2.498280
#> 630 315 2002-11-11 1.38114239  8.583658  9.161147  3.027318
#> 631 316 2002-11-12 0.29301815  8.528999  8.993100  2.402756
#> 633 317 2002-11-13 3.91768170  8.475180  8.829388  4.248055
#> 635 318 2002-11-14 6.84436320  8.422232  8.670025  5.690373
#> 638 319 2002-11-15 1.74181489  8.370184  8.515023  3.014734
#> 639 320 2002-11-16 0.29427875  8.319067  8.364392  2.239039
#> 641 321 2002-11-17 1.67673434  8.268912  8.218139  2.887754
#> 643 322 2002-11-18 4.28299426  8.219751  8.076271  4.123184
#> 646 323 2002-11-19 2.48633754  8.171616  7.938791  3.192446
#> 647 324 2002-11-20 6.04282200  8.124538  7.805702  4.854265
#> 650 325 2002-11-21 1.72399054  8.078549  7.677004  2.738401
#> 651 326 2002-11-22 0.09203771  8.033683  7.552698  1.931438
#> 653 327 2002-11-23 2.69284060  7.989971  7.432780  3.110721
#> 655 328 2002-11-24 2.56330073  7.947446  7.317248  3.009333
#> 658 329 2002-11-25 0.15158806  7.906140  7.206095  1.870607
#> 659 330 2002-11-26 2.48421711  7.866085  7.099317  2.895860
#> 661 331 2002-11-27 5.99096279  7.827314  6.996907  4.426915
#> 663 332 2002-11-28 1.98706842  7.789856  6.898857  2.604608
#> 666 333 2002-11-29 1.42319371  7.753745  6.805157  2.325830
#> 667 334 2002-11-30 6.27853611  7.719011  6.715800  4.410219
#> 669 335 2002-12-01 4.93788991  7.685684  6.630775  3.787760
#> 671 336 2002-12-02 4.71598149  7.653793  6.550072  3.655473
#> 674 337 2002-12-03 3.63731362  7.623368  6.473680  3.162803
#> 676 338 2002-12-04 4.43005219  7.594437  6.401590  3.467512
#> 677 339 2002-12-05 6.35018293  7.567027  6.333789  4.241078
#> 679 340 2002-12-06 3.22409351  7.541163  6.270268  2.907939
#> 681 341 2002-12-07 4.42247217  7.516872  6.211016  3.379847
#> 684 342 2002-12-08 1.92459163  7.494177  6.156023  2.329475
#> 685 343 2002-12-09 2.11033188  7.473100  6.105278  2.388356
#> 687 344 2002-12-10 4.39716060  7.453664  6.058772  3.301827
#> 689 345 2002-12-11 0.50074258  7.435887  6.016497  1.706704
#> 692 346 2002-12-12 5.20085744  7.419789  5.978443  3.589888
#> 693 347 2002-12-13 6.11969596  7.405386  5.944603  3.942414
#> 695 348 2002-12-14 6.11527010  7.392694  5.914970  3.925188
#> 697 349 2002-12-15 5.06172381  7.381725  5.889539  3.491643
#> 700 350 2002-12-16 2.43493208  7.372493  5.868303  2.436146
#> 702 351 2002-12-17 2.49371770  7.365006  5.851259  2.453404
#> 703 352 2002-12-18 1.48043281  7.359273  5.838404  2.046844
#> 705 353 2002-12-19 4.47667946  7.355300  5.829736  3.231519
#> 707 354 2002-12-20 3.72875394  7.353091  5.825253  2.933307
#> 710 355 2002-12-21 4.62022699  7.352649  5.824956  3.286370
#> 711 356 2002-12-22 2.65330770  7.353975  5.828845  2.508732
#> 713 357 2002-12-23 5.60275181  7.357066  5.836924  3.681777
#> 715 358 2002-12-24 6.90340111  7.361920  5.849194  4.204745
#> 718 359 2002-12-25 0.90245373  7.368531  5.865662  1.825611
#> 719 360 2002-12-26 6.82980457  7.376892  5.886331  4.196476
#> 721 361 2002-12-27 6.32085602  7.386993  5.911209  4.006835
#> 723 362 2002-12-28 6.78915653  7.398825  5.940303  4.210485
#> 726 363 2002-12-29 0.45051499  7.412375  5.973622  1.674940
#> 727 364 2002-12-30 4.32881511  7.427627  6.011174  3.254447
#> 729 365 2002-12-31 3.29735228  7.444567  6.052971  2.853735
#> 2     1 2003-01-01 4.36626400  7.463176  6.099033  3.308851
#> 4     2 2003-01-02 2.09345375  7.483435  6.149345  2.397460
#> 5     3 2003-01-03 3.45797883  7.505323  6.203934  2.980174
#> 8     4 2003-01-04 3.35945646  7.528820  6.262814  2.962978
#> 10    5 2003-01-05 4.69472922  7.553900  6.325996  3.547294
#> 12    6 2003-01-06 2.54066738  7.580540  6.393496  2.669785
#> 13    7 2003-01-07 2.55523060  7.608714  6.465324  2.701954
#> 16    8 2003-01-08 4.08010622  7.638396  6.541496  3.382468
#> 18    9 2003-01-09 3.77962228  7.669557  6.622023  3.287200
#> 20   10 2003-01-10 6.53908849  7.702170  6.706920  4.523794
#> 21   11 2003-01-11 4.56943240  7.736204  6.796198  3.706156
#> 24   12 2003-01-12 3.78038265  7.771631  6.889871  3.398200
#> 26   13 2003-01-13 6.20123572  7.808419  6.987950  4.521808
#> 28   14 2003-01-14 2.62188763  7.846537  7.090446  2.957233
#> 29   15 2003-01-15 1.80467962  7.885954  7.197371  2.622892
#> 31   16 2003-01-16 0.22559585  7.926639  7.308732  1.931188
#> 34   17 2003-01-17 4.79394651  7.968559  7.424539  4.089465
#> 36   18 2003-01-18 4.28846728  8.011681  7.544799  3.905478
#> 38   19 2003-01-19 3.19238786  8.055974  7.669519  3.437002
#> 39   20 2003-01-20 5.29245948  8.101405  7.798703  4.497031
#> 42   21 2003-01-21 6.03323255  8.147942  7.932354  4.919887
#> 44   22 2003-01-22 4.96260846  8.195553  8.070474  4.461053
#> 46   23 2003-01-23 1.76148943  8.244205  8.213063  2.930683
#> 47   24 2003-01-24 0.94737659  8.293867  8.360119  2.567502
#> 50   25 2003-01-25 5.07180608  8.344507  8.511639  4.714605
#> 52   26 2003-01-26 6.73268425  8.396093  8.667618  5.642112
#> 54   27 2003-01-27 3.35836539  8.448595  8.828046  3.961611
#> 55   28 2003-01-28 5.80326946  8.501983  8.992915  5.317414
#> 57   29 2003-01-29 4.92946742  8.556225  9.162213  4.929850
#> 60   30 2003-01-30 6.63069089  8.611293  9.335925  5.928310
#> 62   31 2003-01-31 2.12052614  8.667157  9.514035  3.542371
#> 64   32 2003-02-01 6.64053598  8.723789  9.696523  6.114621
#> 65   33 2003-02-02 2.61309101  8.781160  9.883368  3.941385
#> 68   34 2003-02-03 0.16650115  8.839243 10.074546  2.613522
#> 70   35 2003-02-04 2.10628883  8.898011 10.270030  3.783040
#> 72   36 2003-02-05 5.25065551  8.957438 10.469790  5.686029
#> 73   37 2003-02-06 0.26477067  9.017498 10.673795  2.825150
#> 76   38 2003-02-07 4.56422103  9.078165 10.882010  5.456072
#> 78   39 2003-02-08 1.01349596  9.139416 11.094398  3.388744
#> 80   40 2003-02-09 4.57964102  9.201226 11.310918  5.642569
#> 81   41 2003-02-10 1.09521968  9.263572 11.531528  3.564561
#> 84   42 2003-02-11 0.14671444  9.326432 11.756182  3.031514
#> 85   43 2003-02-12 0.51273036  9.389784 11.984831  3.323424
#> 88   44 2003-02-13 6.33677574  9.453606 12.217425  7.149042
#> 90   45 2003-02-14 0.81345621  9.517878 12.453910  3.645671
#> 92   46 2003-02-15 1.36338543  9.582579 12.694229  4.076609
#> 93   47 2003-02-16 1.02449984  9.647690 12.938322  3.921549
#> 96   48 2003-02-17 0.40055940  9.713192 13.186128  3.568421
#> 98   49 2003-02-18 6.94652528  9.779067 13.437583  8.132065
#> 100  50 2003-02-19 6.95574367  9.845297 13.692618  8.260101
#> 101  51 2003-02-20 1.29044912  9.911864 13.951164  4.395958
#> 104  52 2003-02-21 1.96061111  9.978754 14.213148  4.949576
#> 106  53 2003-02-22 5.28464060 10.045948 14.478496  7.427808
#> 108  54 2003-02-23 1.34983946 10.113433 14.747130  4.670932
#> 109  55 2003-02-24 4.30295599 10.181192 15.018971  6.928535
#> 111  56 2003-02-25 5.28904828 10.249211 15.293936  7.769659
#> 114  57 2003-02-26 2.12483565 10.317477 15.571940  5.496469
#> 116  58 2003-02-27 4.34539211 10.385975 15.852898  7.279574
#> 118  59 2003-02-28 5.42596463 10.454693 16.136720  8.221643
#> 119  60 2003-03-01 1.89843644 10.523618 16.423316  5.587193
#> 122  61 2003-03-02 5.87674747 10.592738 16.712592  8.814139
#> 124  62 2003-03-03 2.77179545 10.662040 17.004454  6.461425
#> 126  63 2003-03-04 5.10545370 10.731513 17.298805  8.439603
#> 127  64 2003-03-05 3.00522124 10.801147 17.595546  6.846706
#> 130  65 2003-03-06 4.78712388 10.870930 17.894578  8.413674
#> 132  66 2003-03-07 6.36441162 10.940852 18.195799  9.841296
#> 134  67 2003-03-08 6.67807721 11.010902 18.499104 10.234600
#> 135  68 2003-03-09 2.03796164 11.081072 18.804391  6.430291
#> 137  69 2003-03-10 3.28930250 11.151350 19.111552  7.596546
#> 140  70 2003-03-11 1.46814907 11.221729 19.420480  6.125520
#> 142  71 2003-03-12 6.74694803 11.292198 19.731067 10.827300
#> 144  72 2003-03-13 0.73286707 11.362749 20.043203  5.657167
#> 145  73 2003-03-14 5.47836462 11.433374 20.356778  9.966226
#> 148  74 2003-03-15 3.93978591 11.504062 20.671679  8.707626
#> 150  75 2003-03-16 3.32472818 11.574807 20.987796  8.261198
#> 152  76 2003-03-17 2.28889192 11.645600 21.305014  7.419957
#> 153  77 2003-03-18 4.89286453 11.716433 21.623220  9.920809
#> 156  78 2003-03-19 4.57684112 11.787297 21.942300  9.745518
#> 158  79 2003-03-20 2.36800359 11.858185 22.262140  7.788338
#> 160  80 2003-03-21 1.26843107 11.929088 22.582624  6.846272
#> 161  81 2003-03-22 2.99330070 12.000000 22.903638  8.582471
#> 164  82 2003-03-23 4.13169562 12.070912 23.225065  9.781066
#> 165  83 2003-03-24 6.15713202 12.141815 23.546791 11.857004
#> 168  84 2003-03-25 1.15230151 12.212703 23.868700  7.093213
#> 170  85 2003-03-26 6.97405143 12.283567 24.190677 12.914853
#> 172  86 2003-03-27 3.19374751 12.354400 24.512606  9.296540
#> 173  87 2003-03-28 5.84916739 12.425193 24.834373 12.053992
#> 176  88 2003-03-29 1.64530302 12.495938 25.155863  7.945065
#> 178  89 2003-03-30 5.47355409 12.566626 25.476961 11.917647
#> 180  90 2003-03-31 4.14159515 12.637251 25.797553 10.676693
#> 181  91 2003-04-01 2.39057153 12.707802 26.117528  8.985976
#> 184  92 2003-04-02 1.11130048 12.778271 26.436771  7.758769
#> 186  93 2003-04-03 2.58348276 12.848650 26.755170  9.378629
#> 188  94 2003-04-04 6.56473279 12.918928 27.072615 13.646607
#> 189  95 2003-04-05 2.64156914 12.989098 27.388994  9.632274
#> 191  96 2003-04-06 0.41718371 13.059148 27.704199  7.368565
#> 194  97 2003-04-07 6.28415077 13.129070 28.018120 13.709883
#> 196  98 2003-04-08 0.50342726 13.198853 28.330649  7.622952
#> 198  99 2003-04-09 5.18404378 13.268487 28.641680 12.755622
#> 199 100 2003-04-10 4.08834679 13.337960 28.951106 11.674818
#> 202 101 2003-04-11 2.65321283 13.407262 29.258822 10.209774
#> 204 102 2003-04-12 4.97133531 13.476382 29.564726 12.844283
#> 206 103 2003-04-13 0.36968458 13.545307 29.868714  7.874774
#> 207 104 2003-04-14 6.18753173 13.614025 30.170685 14.398912
#> 210 105 2003-04-15 0.77844743 13.682523 30.470539  8.484423
#> 212 106 2003-04-16 2.99414673 13.750789 30.768176 11.041831
#> 214 107 2003-04-17 0.67874501 13.818808 31.063500  8.528755
#> 215 108 2003-04-18 3.96015076 13.886567 31.356413 12.310191
#> 217 109 2003-04-19 5.68284712 13.954052 31.646820 14.355856
#> 220 110 2003-04-20 4.50922066 14.021246 31.934628 13.118731
#> 222 111 2003-04-21 6.59433284 14.088136 32.219743 15.595597
#> 224 112 2003-04-22 4.66460017 14.154703 32.502075 13.480954
#> 225 113 2003-04-23 2.01867679 14.220933 32.781534 10.522070
#> 228 114 2003-04-24 4.85377832 14.286808 33.058030 13.880050
#> 230 115 2003-04-25 6.65064622 14.352310 33.331477 16.055524
#> 232 116 2003-04-26 0.87334810 14.417421 33.601789  9.418176
#> 233 117 2003-04-27 5.36348063 14.482122 33.868881 14.738921
#> 236 118 2003-04-28 6.32832904 14.546394 34.132670 15.957783
#> 238 119 2003-04-29 4.34512271 14.610216 34.393074 13.712571
#> 240 120 2003-04-30 5.32850257 14.673568 34.650012 14.953839
#> 241 121 2003-05-01 2.96170166 14.736428 34.903405 12.233264
#> 244 122 2003-05-02 1.49131136 14.798774 35.153175 10.559533
#> 246 123 2003-05-03 1.81317932 14.860584 35.399245 11.009389
#> 247 124 2003-05-04 1.94253376 14.921835 35.641540 11.230304
#> 250 125 2003-05-05 2.18263013 14.982502 35.879987 11.583470
#> 252 126 2003-05-06 0.99305203 15.042562 36.114511 10.220698
#> 254 127 2003-05-07 4.34639720 15.101989 36.345041 14.316366
#> 255 128 2003-05-08 6.34839382 15.160757 36.571508 16.799828
#> 258 129 2003-05-09 6.63850837 15.218840 36.793843 17.223259
#> 260 130 2003-05-10 5.47340235 15.276211 37.011977 15.883612
#> 262 131 2003-05-11 0.33814163 15.332843 37.225844  9.716939
#> 263 132 2003-05-12 4.48936879 15.388707 37.435378 14.819382
#> 266 133 2003-05-13 6.24159806 15.443775 37.640517 17.016332
#> 268 134 2003-05-14 6.72781681 15.498017 37.841197 17.673887
#> 270 135 2003-05-15 1.35015154 15.551405 38.037357 11.160515
#> 271 136 2003-05-16 3.21911913 15.603907 38.228936 13.500589
#> 273 137 2003-05-17 6.05309213 15.655493 38.415875 17.030590
#> 276 138 2003-05-18 4.54218652 15.706133 38.598118 15.230784
#> 278 139 2003-05-19 2.58407349 15.755795 38.775606 12.873653
#> 280 140 2003-05-20 1.31553101 15.804447 38.948286 11.358061
#> 281 141 2003-05-21 1.32482392 15.852058 39.116103 11.413575
#> 284 142 2003-05-22 1.63242736 15.898595 39.279005 11.836286
#> 286 143 2003-05-23 3.11864000 15.944026 39.436940 13.716154
#> 288 144 2003-05-24 1.50747718 15.988319 39.589859 11.763853
#> 289 145 2003-05-25 3.98895657 16.031441 39.737713 14.878214
#> 292 146 2003-05-26 5.38682515 16.073361 39.880455 16.652880
#> 294 147 2003-05-27 3.60951645 16.114046 40.018039 14.486493
#> 296 148 2003-05-28 0.55642230 16.153463 40.150421 10.729116
#> 297 149 2003-05-29 4.12807312 16.191581 40.277558 15.203808
#> 299 150 2003-05-30 5.23678831 16.228369 40.399408 16.618164
#> 302 151 2003-05-31 1.85956505 16.263796 40.515933 12.445233
#> 304 152 2003-06-01 5.34051849 16.297830 40.627093 16.813173
#> 306 153 2003-06-02 3.98422535 16.330443 40.732853 15.152119
#> 307 154 2003-06-03 1.89457834 16.361604 40.833176 12.572416
#> 310 155 2003-06-04 2.33506356 16.391286 40.928031 13.147263
#> 312 156 2003-06-05 6.91669636 16.419460 41.017385 18.893632
#> 314 157 2003-06-06 6.20263958 16.446100 41.101209 18.025954
#> 315 158 2003-06-07 4.10819738 16.471180 41.179475 15.430306
#> 318 159 2003-06-08 0.21793018 16.494677 41.252156 10.585554
#> 320 160 2003-06-09 3.48878880 16.516565 41.319229 14.693731
#> 322 161 2003-06-10 3.74341709 16.536824 41.380671 15.028809
#> 323 162 2003-06-11 1.36686415 16.555433 41.436461 12.069672
#> 326 163 2003-06-12 2.24517994 16.572373 41.486582 13.181890
#> 327 164 2003-06-13 3.60873883 16.587625 41.531016 14.900417
#> 330 165 2003-06-14 5.93034393 16.601175 41.569750 17.817301
#> 332 166 2003-06-15 1.47395073 16.613007 41.602770 12.246248
#> 334 167 2003-06-16 2.39127773 16.623108 41.630068 13.401814
#> 335 168 2003-06-17 3.77381908 16.631469 41.651634 15.138460
#> 338 169 2003-06-18 3.40133470 16.638080 41.667464 14.675921
#> 340 170 2003-06-19 4.95355816 16.642934 41.677552 16.621785
#> 342 171 2003-06-20 3.50561643 16.646025 41.681897 14.809532
#> 343 172 2003-06-21 4.45383999 16.647351 41.680500 15.995735
#> 346 173 2003-06-22 6.57291712 16.646909 41.673363 18.645561
#> 348 174 2003-06-23 5.07157075 16.644700 41.660491 16.762012
#> 350 175 2003-06-24 5.90446207 16.640727 41.641891 17.798160
#> 351 176 2003-06-25 4.95372417 16.634994 41.617572 16.601028
#> 353 177 2003-06-26 2.15971304 16.627507 41.587545 13.097747
#> 356 178 2003-06-27 1.87012605 16.618275 41.551823 12.725959
#> 358 179 2003-06-28 6.65200293 16.607306 41.510423 18.691039
#> 360 180 2003-06-29 2.89167982 16.594614 41.463360 13.978421
#> 361 181 2003-06-30 6.86868665 16.580211 41.410655 18.930263
#> 364 182 2003-07-01 0.61054516 16.564113 41.352328 11.100195
#> 366 183 2003-07-02 6.69778817 16.546336 41.288405 18.678664
#> 368 184 2003-07-03 0.19873115 16.526900 41.218909 10.552550
#> 369 185 2003-07-04 2.31349617 16.505823 41.143868 13.169379
#> 372 186 2003-07-05 6.75402869 16.483128 41.063311 18.678757
#> 374 187 2003-07-06 4.29983100 16.458837 40.977269 15.596923
#> 376 188 2003-07-07 5.23323861 16.432973 40.885775 16.731678
#> 377 189 2003-07-08 0.37067786 16.405563 40.788864 10.658021
#> 379 190 2003-07-09 3.12682734 16.376632 40.686570 14.055832
#> 382 191 2003-07-10 5.34448532 16.346207 40.578932 16.778490
#> 384 192 2003-07-11 5.88964210 16.314316 40.465990 17.420825
#> 386 193 2003-07-12 3.83499181 16.280989 40.347784 14.838912
#> 387 194 2003-07-13 5.78828803 16.246255 40.224356 17.221746
#> 390 195 2003-07-14 0.43148854 16.210144 40.095751 10.557581
#> 392 196 2003-07-15 0.04715465 16.172686 39.962013 10.048762
#> 394 197 2003-07-16 2.73103470 16.133915 39.823189 13.326291
#> 395 198 2003-07-17 4.69658844 16.093860 39.679327 15.709539
#> 398 199 2003-07-18 0.29873699 16.052554 39.530476 10.250449
#> 400 200 2003-07-19 3.61517324 16.010029 39.376687 14.289934
#> 402 201 2003-07-20 3.76218287 15.966317 39.218012 14.425022
#> 403 202 2003-07-21 3.03665221 15.921451 39.054503 13.488002
#> 406 203 2003-07-22 5.35502596 15.875462 38.886214 16.279998
#> 407 204 2003-07-23 5.67490724 15.828384 38.713200 16.618169
#> 410 205 2003-07-24 4.44277747 15.780249 38.535519 15.058532
#> 412 206 2003-07-25 5.78122432 15.731088 38.353226 16.635772
#> 414 207 2003-07-26 0.88871100 15.680933 38.166382 10.623128
#> 415 208 2003-07-27 4.12141165 15.629816 37.975044 14.500563
#> 418 209 2003-07-28 6.34850874 15.577768 37.779274 17.143035
#> 420 210 2003-07-29 4.00870011 15.524820 37.579134 14.246482
#> 422 211 2003-07-30 2.83672761 15.471001 37.374685 12.770140
#> 423 212 2003-07-31 4.23434447 15.416342 37.165992 14.395615
#> 426 213 2003-08-01 3.06310310 15.360873 36.953119 12.922680
#> 428 214 2003-08-02 1.01109499 15.304621 36.736132 10.397513
#> 430 215 2003-08-03 5.11565519 15.247616 36.515097 15.254277
#> 431 216 2003-08-04 1.81667129 15.189886 36.290082 11.242621
#> 433 217 2003-08-05 5.33921030 15.131457 36.061156 15.377469
#> 436 218 2003-08-06 3.48020302 15.072356 35.828388 13.093479
#> 438 219 2003-08-07 0.68225764 15.012610 35.591848  9.706709
#> 440 220 2003-08-08 1.06728051 14.952243 35.351609 10.099589
#> 441 221 2003-08-09 0.33370697 14.891281 35.107744  9.170310
#> 444 222 2003-08-10 2.87307041 14.829748 34.860325 12.091948
#> 446 223 2003-08-11 5.41143185 14.767666 34.609429 14.993459
#> 448 224 2003-08-12 4.51219456 14.705060 34.355129 13.859656
#> 449 225 2003-08-13 4.19728429 14.641952 34.097505 13.411598
#> 452 226 2003-08-14 6.19384931 14.578362 33.836632 15.647174
#> 454 227 2003-08-15 4.76504788 14.514313 33.572591 13.904087
#> 456 228 2003-08-16 2.31412261 14.449824 33.305462 10.993281
#> 457 229 2003-08-17 0.04534070 14.384916 33.035326  8.310894
#> 459 230 2003-08-18 1.34033415 14.319607 32.762264  9.723862
#> 462 231 2003-08-19 5.77687191 14.253916 32.486361 14.704677
#> 464 232 2003-08-20 2.15027117 14.187862 32.207701 10.492578
#> 466 233 2003-08-21 4.59289875 14.121461 31.926370 13.173498
#> 467 234 2003-08-22 0.74375596 14.054730 31.642455  8.747850
#> 470 235 2003-08-23 6.82458695 13.987686 31.356043 15.488311
#> 472 236 2003-08-24 3.67760737 13.920345 31.067223 11.870621
#> 474 237 2003-08-25 3.77805048 13.852721 30.776085 11.890800
#> 475 238 2003-08-26 4.44685400 13.784830 30.482721 12.537397
#> 478 239 2003-08-27 0.41547005 13.716686 30.187223  8.003982
#> 480 240 2003-08-28 1.12680966 13.648302 29.889684  8.706274
#> 482 241 2003-08-29 4.71069886 13.579692 29.590199 12.529865
#> 483 242 2003-08-30 0.31413087 13.510869 29.288864  7.662702
#> 486 243 2003-08-31 2.74398736 13.441846 28.985774 10.204988
#> 488 244 2003-09-01 1.57795695 13.372633 28.681027  8.862422
#> 489 245 2003-09-02 4.31275607 13.303244 28.374722 11.693057
#> 492 246 2003-09-03 4.77823861 13.233689 28.066958 12.083755
#> 494 247 2003-09-04 0.78511327 13.163980 27.757836  7.767212
#> 496 248 2003-09-05 0.08157123 13.094126 27.447457  6.947358
#> 497 249 2003-09-06 1.45943466 13.024138 27.135924  8.304354
#> 500 250 2003-09-07 0.83957396 12.954027 26.823338  7.575069
#> 502 251 2003-09-08 2.27023718 12.883802 26.509805  8.963080
#> 504 252 2003-09-09 4.78701999 12.813472 26.195428 11.442068
#> 505 253 2003-09-10 1.02137036 12.743047 25.880314  7.507247
#> 508 254 2003-09-11 1.37963771 12.672536 25.564567  7.782727
#> 510 255 2003-09-12 1.10468022 12.601947 25.248295  7.418700
#> 512 256 2003-09-13 0.11947378 12.531289 24.931604  6.351750
#> 513 257 2003-09-14 5.82756766 12.460572 24.614602 11.909536
#> 515 258 2003-09-15 5.56800619 12.389802 24.297397 11.534003
#> 518 259 2003-09-16 6.38872366 12.318988 23.980097 12.213157
#> 520 260 2003-09-17 1.91331186 12.248139 23.662812  7.763916
#> 522 261 2003-09-18 3.33153604 12.177262 23.345649  9.029941
#> 523 262 2003-09-19 2.88354709 12.106365 23.028718  8.499720
#> 526 263 2003-09-20 4.93443682 12.035456 22.712128 10.333924
#> 528 264 2003-09-21 4.08230557 11.964544 22.395989  9.419756
#> 530 265 2003-09-22 2.09706334 11.893635 22.080409  7.466690
#> 531 266 2003-09-23 6.73647128 11.822738 21.765497 11.642249
#> 534 267 2003-09-24 2.31702383 11.751861 21.451363  7.477540
#> 536 268 2003-09-25 4.24953963 11.681012 21.138114  9.129540
#> 538 269 2003-09-26 4.23570379 11.610198 20.825859  9.005373
#> 539 270 2003-09-27 2.05825203 11.539428 20.514705  6.958248
#> 541 271 2003-09-28 4.75397696 11.468711 20.204761  9.238800
#> 544 272 2003-09-29 0.23621563 11.398053 19.896131  5.180199
#> 546 273 2003-09-30 4.27447636 11.327464 19.588923  8.593221
#> 548 274 2003-10-01 5.94297409 11.256953 19.283241  9.910989
#> 549 275 2003-10-02 5.29325569 11.186528 18.979189  9.235096
#> 552 276 2003-10-03 5.89624537 11.116198 18.676871  9.622504
#> 554 277 2003-10-04 0.38241835 11.045973 18.376389  4.912198
#> 556 278 2003-10-05 1.40761226 10.975862 18.077844  5.678668
#> 557 279 2003-10-06 4.90662005 10.905874 17.781337  8.445301
#> 560 280 2003-10-07 3.00150033 10.836020 17.486965  6.793623
#> 562 281 2003-10-08 1.08532853 10.766311 17.194826  5.165393
#> 564 282 2003-10-09 4.24790299 10.696756 16.905016  7.582920
#> 565 283 2003-10-10 6.75439153 10.627367 16.617629  9.435206
#> 568 284 2003-10-11 0.67541046 10.558154 16.332759  4.605597
#> 569 285 2003-10-12 5.76710543 10.489131 16.050495  8.425044
#> 572 286 2003-10-13 4.33586197 10.420308 15.770928  7.223852
#> 574 287 2003-10-14 6.86603206 10.351698 15.494145  9.011983
#> 576 288 2003-10-15 2.88056706 10.283314 15.220231  5.936807
#> 577 289 2003-10-16 1.80764097 10.215170 14.949271  5.060003
#> 580 290 2003-10-17 0.53258179 10.147279 14.681345  4.055613
#> 582 291 2003-10-18 6.99717708 10.079655 14.416534  8.608027
#> 584 292 2003-10-19 2.79369535 10.012314 14.154915  5.513523
#> 585 293 2003-10-20 4.46797838  9.945270 13.896562  6.595702
#> 588 294 2003-10-21 0.84603300  9.878539 13.641550  3.994543
#> 590 295 2003-10-22 5.42690545  9.812138 13.389948  7.050349
#> 592 296 2003-10-23 5.68022001  9.746084 13.141825  7.115121
#> 593 297 2003-10-24 3.57442563  9.680393 12.897247  5.605426
#> 595 298 2003-10-25 3.63920037  9.615084 12.656277  5.559198
#> 598 299 2003-10-26 5.71119467  9.550176 12.418977  6.818141
#> 600 300 2003-10-27 2.35046263  9.485687 12.185404  4.556064
#> 602 301 2003-10-28 0.57500944  9.421638 11.955615  3.353734
#> 603 302 2003-10-29 2.43792060  9.358048 11.729663  4.460298
#> 606 303 2003-10-30 0.28854526  9.294940 11.507600  3.055517
#> 608 304 2003-10-31 1.76796768  9.232334 11.289473  3.903321
#> 610 305 2003-11-01 4.88595578  9.170252 11.075328  5.719327
#> 611 306 2003-11-02 0.91360995  9.108719 10.865209  3.261196
#> 614 307 2003-11-03 5.12001628  9.047757 10.659156  5.680733
#> 616 308 2003-11-04 1.92779609  8.987390 10.457208  3.735838
#> 618 309 2003-11-05 0.42357854  8.927644 10.259398  2.808232
#> 619 310 2003-11-06 5.82225235  8.868543 10.065762  5.820557
#> 621 311 2003-11-07 6.20616280  8.810114  9.876328  5.947703
#> 624 312 2003-11-08 5.36534590  8.752384  9.691126  5.393186
#> 626 313 2003-11-09 4.84699297  8.695379  9.510180  5.028135
#> 628 314 2003-11-10 0.30526730  8.639127  9.333513  2.498280
#> 629 315 2003-11-11 1.38114239  8.583658  9.161147  3.027318
#> 632 316 2003-11-12 0.29301815  8.528999  8.993100  2.402756
#> 634 317 2003-11-13 3.91768170  8.475180  8.829388  4.248055
#> 636 318 2003-11-14 6.84436320  8.422232  8.670025  5.690373
#> 637 319 2003-11-15 1.74181489  8.370184  8.515023  3.014734
#> 640 320 2003-11-16 0.29427875  8.319067  8.364392  2.239039
#> 642 321 2003-11-17 1.67673434  8.268912  8.218139  2.887754
#> 644 322 2003-11-18 4.28299426  8.219751  8.076271  4.123184
#> 645 323 2003-11-19 2.48633754  8.171616  7.938791  3.192446
#> 648 324 2003-11-20 6.04282200  8.124538  7.805702  4.854265
#> 649 325 2003-11-21 1.72399054  8.078549  7.677004  2.738401
#> 652 326 2003-11-22 0.09203771  8.033683  7.552698  1.931438
#> 654 327 2003-11-23 2.69284060  7.989971  7.432780  3.110721
#> 656 328 2003-11-24 2.56330073  7.947446  7.317248  3.009333
#> 657 329 2003-11-25 0.15158806  7.906140  7.206095  1.870607
#> 660 330 2003-11-26 2.48421711  7.866085  7.099317  2.895860
#> 662 331 2003-11-27 5.99096279  7.827314  6.996907  4.426915
#> 664 332 2003-11-28 1.98706842  7.789856  6.898857  2.604608
#> 665 333 2003-11-29 1.42319371  7.753745  6.805157  2.325830
#> 668 334 2003-11-30 6.27853611  7.719011  6.715800  4.410219
#> 670 335 2003-12-01 4.93788991  7.685684  6.630775  3.787760
#> 672 336 2003-12-02 4.71598149  7.653793  6.550072  3.655473
#> 673 337 2003-12-03 3.63731362  7.623368  6.473680  3.162803
#> 675 338 2003-12-04 4.43005219  7.594437  6.401590  3.467512
#> 678 339 2003-12-05 6.35018293  7.567027  6.333789  4.241078
#> 680 340 2003-12-06 3.22409351  7.541163  6.270268  2.907939
#> 682 341 2003-12-07 4.42247217  7.516872  6.211016  3.379847
#> 683 342 2003-12-08 1.92459163  7.494177  6.156023  2.329475
#> 686 343 2003-12-09 2.11033188  7.473100  6.105278  2.388356
#> 688 344 2003-12-10 4.39716060  7.453664  6.058772  3.301827
#> 690 345 2003-12-11 0.50074258  7.435887  6.016497  1.706704
#> 691 346 2003-12-12 5.20085744  7.419789  5.978443  3.589888
#> 694 347 2003-12-13 6.11969596  7.405386  5.944603  3.942414
#> 696 348 2003-12-14 6.11527010  7.392694  5.914970  3.925188
#> 698 349 2003-12-15 5.06172381  7.381725  5.889539  3.491643
#> 699 350 2003-12-16 2.43493208  7.372493  5.868303  2.436146
#> 701 351 2003-12-17 2.49371770  7.365006  5.851259  2.453404
#> 704 352 2003-12-18 1.48043281  7.359273  5.838404  2.046844
#> 706 353 2003-12-19 4.47667946  7.355300  5.829736  3.231519
#> 708 354 2003-12-20 3.72875394  7.353091  5.825253  2.933307
#> 709 355 2003-12-21 4.62022699  7.352649  5.824956  3.286370
#> 712 356 2003-12-22 2.65330770  7.353975  5.828845  2.508732
#> 714 357 2003-12-23 5.60275181  7.357066  5.836924  3.681777
#> 716 358 2003-12-24 6.90340111  7.361920  5.849194  4.204745
#> 717 359 2003-12-25 0.90245373  7.368531  5.865662  1.825611
#> 720 360 2003-12-26 6.82980457  7.376892  5.886331  4.196476
#> 722 361 2003-12-27 6.32085602  7.386993  5.911209  4.006835
#> 724 362 2003-12-28 6.78915653  7.398825  5.940303  4.210485
#> 725 363 2003-12-29 0.45051499  7.412375  5.973622  1.674940
#> 728 364 2003-12-30 4.32881511  7.427627  6.011174  3.254447
#> 730 365 2003-12-31 3.29735228  7.444567  6.052971  2.853735