Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.836761  2.369913  4.251008  1.985222  3.477779  3.826023  2.782798
#>   [8]  3.679978  2.774110  2.069111  2.307761  2.292501  4.454323  3.263203
#>  [15]  3.576886  1.927143  4.948140  4.032237  3.645435  2.426532  4.095931
#>  [22]  4.617368  2.171640  4.620602  2.862116  2.459626  5.413953  2.891541
#>  [29]  5.665489  4.161597  4.163142  6.149218  5.977029  2.978446  4.173688
#>  [36]  4.769133  3.534037  6.661946  5.392539  4.778391  4.186141  4.488915
#>  [43]  4.484369  5.927368  4.580660  7.127441  3.504590  3.647479  5.833584
#>  [50]  4.070851  3.755231  8.025616  8.026922  5.657848  5.765792  8.624556
#>  [57]  6.708885  5.231231  6.946544  8.245548  7.308499  6.972099  7.302731
#>  [64]  8.891030  7.449111  5.288550  6.287065  6.202935  5.621474  5.693400
#>  [71]  9.570735  5.364633  8.724710  9.889925  8.570612 11.057491  5.494355
#>  [78] 11.148213 10.623118  9.972715  8.617209 12.110003  9.813998 11.333814
#>  [85]  7.990991  7.605844  7.876044 11.339738 12.735332  8.595260 11.270448
#>  [92]  9.076759  7.312625 10.824182 10.826261  9.687875 11.955187 13.684518
#>  [99] 11.641936  7.864516  8.836305 11.140637 13.222426 14.422907 11.044600
#> [106]  8.387127 11.355009 14.981417 12.968968 12.682494 10.945673 12.336317
#> [113] 15.590373  8.746388 11.858226  8.571209  9.666223 16.000531  9.570980
#> [120] 12.275002 12.161820  9.911735 11.550635 11.930562 10.852838 16.357397
#> [127] 13.288581 12.548707 16.073528 14.997096  9.550152 12.142817 16.804874
#> [134] 13.046579  9.674176 13.199064 12.759060 12.929423 16.866418 13.854974
#> [141] 12.454829 14.823977 13.873880 10.076004  9.944225 15.802238 10.316306
#> [148] 11.785807 12.783248 12.505641 14.230950 16.762665 17.728129 13.164351
#> [155] 17.569861 10.344737 15.227732 16.445879 11.335124 11.405657 15.461777
#> [162] 15.844846 15.229838 18.381583 17.853967 17.092836 13.742390 15.765204
#> [169] 12.475137 18.744485 14.546939 17.257302 15.370209 15.914448 11.584226
#> [176] 15.245360 14.130068 14.101329 12.901089 18.252786 14.301590 10.597957
#> [183] 16.607646 16.119954 12.711204 13.348209 15.742748 12.353999 14.045030
#> [190] 11.754199 13.517808 18.070033 14.134875 15.605046 18.178509 12.187375
#> [197] 12.574162 11.484549 13.629590 14.791802 14.555146 10.383350 12.643461
#> [204] 10.882189 10.110745 11.348005 16.675063 14.393055 10.445659 11.441531
#> [211] 10.964940 11.132074 15.517200 13.073036 14.418493 16.234256 12.796826
#> [218]  9.910751  9.880288 11.480264 16.026254 10.178790 13.783304 11.927670
#> [225] 15.475051 14.241000 12.519041 11.686183 12.488470 10.921541 13.861338
#> [232] 12.473858  8.068050 10.291510 10.413804  9.610786 10.387723  8.274283
#> [239] 13.475690 11.314579  8.359236  9.338218 11.273633  7.503077  8.232747
#> [246]  8.274210 13.658615  8.975580 11.311098 12.609254  7.118542 11.550979
#> [253] 11.623341  6.468235 10.802218  7.658803  7.463363 10.840846  6.288305
#> [260]  6.470251 10.858986  7.968078 11.161950  8.406869  9.009581 10.864546
#> [267]  8.159426 10.139107  5.374085 11.128440  6.849580  8.835606  8.602052
#> [274]  5.449752 10.046106  9.759258  4.598935  8.935327  8.697868  7.205468
#> [281]  6.142924  7.993626  4.776107  4.139702  6.338759  4.888170  5.446743
#> [288]  8.597507  5.714441  4.528835  7.454984  3.583675  3.829430  5.851397
#> [295]  4.259797  6.461104  3.714434  7.616511  7.307936  7.155599  5.777837
#> [302]  7.005801  6.820042  6.215185  5.727488  3.950939  5.388150  5.206089
#> [309]  6.165159  5.855518  3.966782  5.363790  3.876809  4.169709  3.423076
#> [316]  2.482178  5.065249  5.760666  4.521680  4.421309  3.303516  5.276352
#> [323]  4.826255  2.958586  4.205249  3.064729  3.828850  4.524900  2.387704
#> [330]  3.198144  2.622022  2.711235  1.857238  2.950073  2.247102  3.099334
#> [337]  1.822286  2.270675  3.266464  1.985527  2.823762  2.749020  4.233826
#> [344]  4.060258  4.296654  3.994969  2.520161  2.152405  1.554024  2.706125
#> [351]  4.089667  1.485004  2.862458  2.988144  3.046597  1.761543  3.108609
#> [358]  2.675384  2.481899  3.719629  1.895276  3.802251  3.163047  3.523928
#> [365]  2.602031  1.836761  2.369913  4.251008  1.985222  3.477779  3.826023
#> [372]  2.782798  3.679978  2.774110  2.069111  2.307761  2.292501  4.454323
#> [379]  3.263203  3.576886  1.927143  4.948140  4.032237  3.645435  2.426532
#> [386]  4.095931  4.617368  2.171640  4.620602  2.862116  2.459626  5.413953
#> [393]  2.891541  5.665489  4.161597  4.163142  6.149218  5.977029  2.978446
#> [400]  4.173688  4.769133  3.534037  6.661946  5.392539  4.778391  4.186141
#> [407]  4.488915  4.484369  5.927368  4.580660  7.127441  3.504590  3.647479
#> [414]  5.833584  4.070851  3.755231  8.025616  8.026922  5.657848  5.765792
#> [421]  8.624556  6.708885  5.231231  6.946544  8.245548  7.308499  6.972099
#> [428]  7.302731  8.891030  7.449111  5.288550  6.287065  6.202935  5.621474
#> [435]  5.693400  9.570735  5.364633  8.724710  9.889925  8.570612 11.057491
#> [442]  5.494355 11.148213 10.623118  9.972715  8.617209 12.110003  9.813998
#> [449] 11.333814  7.990991  7.605844  7.876044 11.339738 12.735332  8.595260
#> [456] 11.270448  9.076759  7.312625 10.824182 10.826261  9.687875 11.955187
#> [463] 13.684518 11.641936  7.864516  8.836305 11.140637 13.222426 14.422907
#> [470] 11.044600  8.387127 11.355009 14.981417 12.968968 12.682494 10.945673
#> [477] 12.336317 15.590373  8.746388 11.858226  8.571209  9.666223 16.000531
#> [484]  9.570980 12.275002 12.161820  9.911735 11.550635 11.930562 10.852838
#> [491] 16.357397 13.288581 12.548707 16.073528 14.997096  9.550152 12.142817
#> [498] 16.804874 13.046579  9.674176 13.199064 12.759060 12.929423 16.866418
#> [505] 13.854974 12.454829 14.823977 13.873880 10.076004  9.944225 15.802238
#> [512] 10.316306 11.785807 12.783248 12.505641 14.230950 16.762665 17.728129
#> [519] 13.164351 17.569861 10.344737 15.227732 16.445879 11.335124 11.405657
#> [526] 15.461777 15.844846 15.229838 18.381583 17.853967 17.092836 13.742390
#> [533] 15.765204 12.475137 18.744485 14.546939 17.257302 15.370209 15.914448
#> [540] 11.584226 15.245360 14.130068 14.101329 12.901089 18.252786 14.301590
#> [547] 10.597957 16.607646 16.119954 12.711204 13.348209 15.742748 12.353999
#> [554] 14.045030 11.754199 13.517808 18.070033 14.134875 15.605046 18.178509
#> [561] 12.187375 12.574162 11.484549 13.629590 14.791802 14.555146 10.383350
#> [568] 12.643461 10.882189 10.110745 11.348005 16.675063 14.393055 10.445659
#> [575] 11.441531 10.964940 11.132074 15.517200 13.073036 14.418493 16.234256
#> [582] 12.796826  9.910751  9.880288 11.480264 16.026254 10.178790 13.783304
#> [589] 11.927670 15.475051 14.241000 12.519041 11.686183 12.488470 10.921541
#> [596] 13.861338 12.473858  8.068050 10.291510 10.413804  9.610786 10.387723
#> [603]  8.274283 13.475690 11.314579  8.359236  9.338218 11.273633  7.503077
#> [610]  8.232747  8.274210 13.658615  8.975580 11.311098 12.609254  7.118542
#> [617] 11.550979 11.623341  6.468235 10.802218  7.658803  7.463363 10.840846
#> [624]  6.288305  6.470251 10.858986  7.968078 11.161950  8.406869  9.009581
#> [631] 10.864546  8.159426 10.139107  5.374085 11.128440  6.849580  8.835606
#> [638]  8.602052  5.449752 10.046106  9.759258  4.598935  8.935327  8.697868
#> [645]  7.205468  6.142924  7.993626  4.776107  4.139702  6.338759  4.888170
#> [652]  5.446743  8.597507  5.714441  4.528835  7.454984  3.583675  3.829430
#> [659]  5.851397  4.259797  6.461104  3.714434  7.616511  7.307936  7.155599
#> [666]  5.777837  7.005801  6.820042  6.215185  5.727488  3.950939  5.388150
#> [673]  5.206089  6.165159  5.855518  3.966782  5.363790  3.876809  4.169709
#> [680]  3.423076  2.482178  5.065249  5.760666  4.521680  4.421309  3.303516
#> [687]  5.276352  4.826255  2.958586  4.205249  3.064729  3.828850  4.524900
#> [694]  2.387704  3.198144  2.622022  2.711235  1.857238  2.950073  2.247102
#> [701]  3.099334  1.822286  2.270675  3.266464  1.985527  2.823762  2.749020
#> [708]  4.233826  4.060258  4.296654  3.994969  2.520161  2.152405  1.554024
#> [715]  2.706125  4.089667  1.485004  2.862458  2.988144  3.046597  1.761543
#> [722]  3.108609  2.675384  2.481899  3.719629  1.895276  3.802251  3.163047
#> [729]  3.523928  2.602031
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates   sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 4.86764756  7.463176  6.099033  3.513720
#> 3     2 2002-01-02 1.51227825  7.483435  6.149345  2.158676
#> 6     3 2002-01-03 2.54081300  7.505323  6.203934  2.601107
#> 7     4 2002-01-04 6.25141663  7.528820  6.262814  4.165809
#> 9     5 2002-01-05 2.30999243  7.553900  6.325996  2.548748
#> 11    6 2002-01-06 5.11412033  7.580540  6.393496  3.755021
#> 14    7 2002-01-07 2.00475320  7.608714  6.465324  2.468077
#> 15    8 2002-01-08 0.90595827  7.638396  6.541496  2.023304
#> 17    9 2002-01-09 3.72058544  7.669557  6.622023  3.261713
#> 19   10 2002-01-10 0.87905577  7.702170  6.706920  2.059463
#> 22   11 2002-01-11 2.89750738  7.736204  6.796198  2.971769
#> 23   12 2002-01-12 3.33735115  7.771631  6.889871  3.201817
#> 25   13 2002-01-13 3.59742698  7.808419  6.987950  3.356701
#> 27   14 2002-01-14 4.84240343  7.846537  7.090446  3.960507
#> 30   15 2002-01-15 0.23802075  7.885954  7.197371  1.907961
#> 32   16 2002-01-16 2.19773657  7.926639  7.308732  2.840391
#> 33   17 2002-01-17 4.71981679  7.968559  7.424539  4.054930
#> 35   18 2002-01-18 3.98478200  8.011681  7.544799  3.762484
#> 37   19 2002-01-19 5.84389318  8.055974  7.669519  4.699157
#> 40   20 2002-01-20 5.67503200  8.101405  7.798703  4.681170
#> 41   21 2002-01-21 2.91879380  8.147942  7.932354  3.403871
#> 43   22 2002-01-22 2.86922169  8.195553  8.070474  3.430334
#> 45   23 2002-01-23 1.79607601  8.244205  8.213063  2.947911
#> 48   24 2002-01-24 4.72241127  8.293867  8.360119  4.470097
#> 49   25 2002-01-25 3.59305794  8.344507  8.511639  3.960422
#> 51   26 2002-01-26 4.06542318  8.396093  8.667618  4.265353
#> 53   27 2002-01-27 2.66495124  8.448595  8.828046  3.599332
#> 56   28 2002-01-28 3.54200700  8.501983  8.992915  4.121496
#> 58   29 2002-01-29 1.93085322  8.556225  9.162213  3.324356
#> 59   30 2002-01-30 6.70524893  8.611293  9.335925  5.968726
#> 61   31 2002-01-31 1.45610878  8.667157  9.514035  3.177702
#> 63   32 2002-02-01 5.66463783  8.723789  9.696523  5.572264
#> 66   33 2002-02-02 6.93123902  8.781160  9.883368  6.371465
#> 67   34 2002-02-03 4.99620425  8.839243 10.074546  5.365854
#> 69   35 2002-02-04 6.95004633  8.898011 10.270030  6.578358
#> 71   36 2002-02-05 6.67932393  8.957438 10.469790  6.520970
#> 74   37 2002-02-06 6.19494335  9.017498 10.673795  6.334852
#> 75   38 2002-02-07 5.80099525  9.078165 10.882010  6.197334
#> 77   39 2002-02-08 1.97892264  9.139416 11.094398  3.974713
#> 79   40 2002-02-09 0.62457286  9.201226 11.310918  3.211618
#> 82   41 2002-02-10 4.21642702  9.263572 11.531528  5.507239
#> 83   42 2002-02-11 1.97697984  9.326432 11.756182  4.185060
#> 86   43 2002-02-12 5.40443701  9.389784 11.984831  6.445236
#> 87   44 2002-02-13 4.59503021  9.453606 12.217425  6.023564
#> 89   45 2002-02-14 3.07639373  9.517878 12.453910  5.126170
#> 91   46 2002-02-15 3.04628539  9.582579 12.694229  5.191294
#> 94   47 2002-02-16 2.04758665  9.647690 12.938322  4.607569
#> 95   48 2002-02-17 2.17747689  9.713192 13.186128  4.774547
#> 97   49 2002-02-18 1.42326699  9.779067 13.437583  4.337263
#> 99   50 2002-02-19 6.98644711  9.845297 13.692618  8.281451
#> 102  51 2002-02-20 3.51236562  9.911864 13.951164  5.959656
#> 103  52 2002-02-21 0.31011750  9.978754 14.213148  3.774144
#> 105  53 2002-02-22 6.09354584 10.045948 14.478496  8.010717
#> 107  54 2002-02-23 1.24845687 10.113433 14.747130  4.597015
#> 110  55 2002-02-24 0.05156412 10.181192 15.018971  3.792776
#> 112  56 2002-02-25 6.72432046 10.249211 15.293936  8.840520
#> 113  57 2002-02-26 4.96681018 10.317477 15.571940  7.641134
#> 115  58 2002-02-27 3.78364173 10.385975 15.852898  6.850853
#> 117  59 2002-02-28 6.84885989 10.454693 16.136720  9.319755
#> 120  60 2002-03-01 6.29093799 10.523618 16.423316  9.014695
#> 121  61 2002-03-02 2.93238389 10.592738 16.712592  6.491418
#> 123  62 2002-03-03 0.42748831 10.662040 17.004454  4.592005
#> 125  63 2002-03-04 2.99757750 10.731513 17.298805  6.740693
#> 128  64 2002-03-05 0.10769550 10.801147 17.595546  4.486607
#> 129  65 2002-03-06 3.89402724 10.870930 17.894578  7.678613
#> 131  66 2002-03-07 5.30842257 10.940852 18.195799  8.963185
#> 133  67 2002-03-08 4.92259788 11.010902 18.499104  8.759935
#> 136  68 2002-03-09 0.53582281 11.081072 18.804391  5.155739
#> 138  69 2002-03-10 1.20123556 11.151350 19.111552  5.807247
#> 139  70 2002-03-11 5.09511539 11.221729 19.420480  9.263959
#> 141  71 2002-03-12 6.23344574 11.292198 19.731067 10.378674
#> 143  72 2002-03-13 1.13253247 11.362749 20.043203  6.009660
#> 146  73 2002-03-14 4.47085601 11.433374 20.356778  9.069307
#> 147  74 2002-03-15 2.52380621 11.504062 20.671679  7.435437
#> 149  75 2002-03-16 1.60423776 11.574807 20.987796  6.701375
#> 151  76 2002-03-17 4.34431653 11.645600 21.305014  9.300103
#> 154  77 2002-03-18 3.37701685 11.716433 21.623220  8.522026
#> 155  78 2002-03-19 5.45891153 11.787297 21.942300 10.566514
#> 157  79 2002-03-20 1.89203338 11.858185 22.262140  7.341553
#> 159  80 2002-03-21 6.52439760 11.929088 22.582624 11.821233
#> 162  81 2002-03-22 4.94651940 12.000000 22.903638 10.446463
#> 163  82 2002-03-23 2.07182152 12.070912 23.225065  7.799413
#> 166  83 2002-03-24 5.81928199 12.141815 23.546791 11.529405
#> 167  84 2002-03-25 2.55828955 12.212703 23.868700  8.467156
#> 169  85 2002-03-26 3.22120671 12.283567 24.190677  9.219515
#> 171  86 2002-03-27 0.08292269 12.354400 24.512606  6.210416
#> 174  87 2002-03-28 1.52833057 12.425193 24.834373  7.735939
#> 175  88 2002-03-29 5.59526619 12.495938 25.155863 11.920946
#> 177  89 2002-03-30 4.01161027 12.566626 25.476961 10.435711
#> 179  90 2002-03-31 5.02997664 12.637251 25.797553 11.583460
#> 182  91 2002-04-01 5.79812152 12.707802 26.117528 12.487635
#> 183  92 2002-04-02 4.60019416 12.778271 26.436771 11.367829
#> 185  93 2002-04-03 4.04953692 12.848650 26.755170 10.905035
#> 187  94 2002-04-04 5.90830475 12.918928 27.072615 12.958809
#> 190  95 2002-04-05 1.86403002 12.989098 27.388994  8.812509
#> 192  96 2002-04-06 2.77354081 13.059148 27.704199  9.868000
#> 193  97 2002-04-07 0.67984694 13.129070 28.018120  7.729944
#> 195  98 2002-04-08 2.48580862 13.198853 28.330649  9.750491
#> 197  99 2002-04-09 2.25631797 13.268487 28.641680  9.595691
#> 200 100 2002-04-10 3.86941417 13.337960 28.951106 11.437212
#> 201 101 2002-04-11 6.70897941 13.407262 29.258822 14.635247
#> 203 102 2002-04-12 2.05254836 13.476382 29.564726  9.642640
#> 205 103 2002-04-13 6.27950716 13.545307 29.868714 14.390639
#> 208 104 2002-04-14 2.92462122 13.614025 30.170685 10.783367
#> 209 105 2002-04-15 5.15944844 13.682523 30.470539 13.362598
#> 211 106 2002-04-16 5.98466327 13.750789 30.768176 14.387557
#> 213 107 2002-04-17 6.62349325 13.818808 31.063500 15.210398
#> 216 108 2002-04-18 3.05458735 13.886567 31.356413 11.287792
#> 218 109 2002-04-19 4.56142744 13.954052 31.646820 13.084205
#> 219 110 2002-04-20 4.55568646 14.021246 31.934628 13.171646
#> 221 111 2002-04-21 0.63306389 14.088136 32.219743  8.778848
#> 223 112 2002-04-22 6.43955732 14.154703 32.502075 15.518785
#> 226 113 2002-04-23 6.79408061 14.220933 32.781534 16.026106
#> 227 114 2002-04-24 1.41304176 14.286808 33.058030  9.899316
#> 229 115 2002-04-25 6.31176715 14.352310 33.331477 15.662021
#> 231 116 2002-04-26 4.86217505 14.417421 33.601789 14.066432
#> 234 117 2002-04-27 2.80552678 14.482122 33.868881 11.747819
#> 235 118 2002-04-28 3.90209693 14.546394 34.132670 13.111243
#> 237 119 2002-04-29 5.49861979 14.610216 34.393074 15.070262
#> 239 120 2002-04-30 5.98350161 14.673568 34.650012 15.727193
#> 242 121 2002-05-01 2.12476448 14.736428 34.903405 11.242116
#> 243 122 2002-05-02 0.25295827 14.798774 35.153175  9.088734
#> 245 123 2002-05-03 0.52013018 14.860584 35.399245  9.469310
#> 248 124 2002-05-04 6.25949902 14.921835 35.641540 16.385947
#> 249 125 2002-05-05 5.62238095 14.982502 35.879987 15.702215
#> 251 126 2002-05-06 0.28698326 15.042562 36.114511  9.373126
#> 253 127 2002-05-07 5.03888181 15.101989 36.345041 15.149646
#> 256 128 2002-05-08 6.36907121 15.160757 36.571508 16.824767
#> 257 129 2002-05-09 4.66241772 15.218840 36.793843 14.834510
#> 259 130 2002-05-10 5.04710985 15.276211 37.011977 15.367191
#> 261 131 2002-05-11 1.93774883 15.332843 37.225844 11.658743
#> 264 132 2002-05-12 1.52594571 15.388707 37.435378 11.214893
#> 265 133 2002-05-13 3.97730217 15.443775 37.640517 14.256992
#> 267 134 2002-05-14 1.32361169 15.498017 37.841197 11.076217
#> 269 135 2002-05-15 4.17071663 15.551405 38.037357 14.609941
#> 272 136 2002-05-16 4.22766927 15.603907 38.228936 14.736043
#> 274 137 2002-05-17 4.86062344 15.655493 38.415875 15.567534
#> 275 138 2002-05-18 1.95536176 15.706133 38.598118 12.052199
#> 277 139 2002-05-19 1.19215394 15.755795 38.775606 11.160870
#> 279 140 2002-05-20 2.25297044 15.804447 38.948286 12.513168
#> 282 141 2002-05-21 4.91139475 15.852058 39.116103 15.838637
#> 283 142 2002-05-22 5.31262217 15.898595 39.279005 16.382423
#> 285 143 2002-05-23 4.46889803 15.944026 39.436940 15.386060
#> 287 144 2002-05-24 6.43350463 15.988319 39.589859 17.862703
#> 290 145 2002-05-25 4.71487771 16.031441 39.737713 15.777897
#> 291 146 2002-05-26 4.18172698 16.073361 39.880455 15.157864
#> 293 147 2002-05-27 1.04288303 16.114046 40.018039 11.299471
#> 295 148 2002-05-28 1.17851264 16.153463 40.150421 11.502238
#> 298 149 2002-05-29 6.56407494 16.191581 40.277558 18.233660
#> 300 150 2002-05-30 4.40444119 16.228369 40.399408 15.582129
#> 301 151 2002-05-31 0.63467587 16.263796 40.515933 10.919527
#> 303 152 2002-06-01 1.31690730 16.297830 40.627093 11.798161
#> 305 153 2002-06-02 5.33557703 16.330443 40.732853 16.837450
#> 308 154 2002-06-03 0.29776007 16.361604 40.833176 10.579850
#> 309 155 2002-06-04 1.88601276 16.391286 40.928031 12.586637
#> 311 156 2002-06-05 5.80817179 16.419460 41.017385 17.509031
#> 313 157 2002-06-06 4.94446839 16.446100 41.101209 16.453777
#> 316 158 2002-06-07 6.82913917 16.471180 41.179475 18.831609
#> 317 159 2002-06-08 2.58556601 16.494677 41.252156 13.546209
#> 319 160 2002-06-09 4.70110661 16.516565 41.319229 16.210149
#> 321 161 2002-06-10 3.83445576 16.536824 41.380671 15.142714
#> 324 162 2002-06-11 4.70912505 16.555433 41.436461 16.252331
#> 325 163 2002-06-12 1.41541036 16.572373 41.486582 12.143285
#> 328 164 2002-06-13 3.06658476 16.587625 41.531016 14.221712
#> 329 165 2002-06-14 4.43928665 16.601175 41.569750 15.950479
#> 331 166 2002-06-15 2.63372539 16.613007 41.602770 13.698418
#> 333 167 2002-06-16 5.66422088 16.623108 41.630068 17.500110
#> 336 168 2002-06-17 6.21666850 16.631469 41.651634 18.197380
#> 337 169 2002-06-18 3.95294074 16.638080 41.667464 15.366626
#> 339 170 2002-06-19 1.23490851 16.642934 41.677552 11.965629
#> 341 171 2002-06-20 2.58754926 16.646025 41.681897 13.660105
#> 344 172 2002-06-21 6.98601401 16.647351 41.680500 19.165677
#> 345 173 2002-06-22 6.55588885 16.646909 41.673363 18.624247
#> 347 174 2002-06-23 3.99452428 16.644700 41.660491 15.414127
#> 349 175 2002-06-24 2.77383808 16.640727 41.641891 13.881110
#> 352 176 2002-06-25 5.00319677 16.634994 41.617572 16.662913
#> 354 177 2002-06-26 0.70036028 16.627507 41.587545 11.272732
#> 355 178 2002-06-27 1.80968467 16.618275 41.551823 12.650396
#> 357 179 2002-06-28 2.07292020 16.607306 41.510423 12.968266
#> 359 180 2002-06-29 3.45422126 16.594614 41.463360 14.681205
#> 362 181 2002-06-30 2.62367757 16.580211 41.410655 13.629106
#> 363 182 2002-07-01 5.28211735 16.564113 41.352328 16.931488
#> 365 183 2002-07-02 0.09395082 16.546336 41.288405 10.439320
#> 367 184 2002-07-03 5.53733939 16.526900 41.218909 17.209927
#> 370 185 2002-07-04 5.80397805 16.505823 41.143868 17.519720
#> 371 186 2002-07-05 5.25870987 16.483128 41.063311 16.816163
#> 373 187 2002-07-06 4.93182187 16.458837 40.977269 16.383651
#> 375 188 2002-07-07 4.62470526 16.432973 40.885775 15.974653
#> 378 189 2002-07-08 3.90600358 16.405563 40.788864 15.052930
#> 380 190 2002-07-09 6.96354279 16.376632 40.686570 18.821855
#> 381 191 2002-07-10 0.37011588 16.346207 40.578932 10.604133
#> 383 192 2002-07-11 3.33412428 16.314316 40.465990 14.251474
#> 385 193 2002-07-12 4.36885700 16.280989 40.347784 15.500429
#> 388 194 2002-07-13 5.59149651 16.246255 40.224356 16.978126
#> 389 195 2002-07-14 5.81376381 16.210144 40.095751 17.214103
#> 391 196 2002-07-15 6.19398410 16.172686 39.962013 17.643037
#> 393 197 2002-07-16 1.75968541 16.133915 39.823189 12.127505
#> 396 198 2002-07-17 1.33295035 16.093860 39.679327 11.563023
#> 397 199 2002-07-18 3.28753194 16.052554 39.530476 13.930502
#> 399 200 2002-07-19 0.59755080 16.010029 39.376687 10.579010
#> 401 201 2002-07-20 6.13889247 15.966317 39.218012 17.343974
#> 404 202 2002-07-21 4.47430607 15.921451 39.054503 15.251247
#> 405 203 2002-07-22 4.04783730 15.875462 38.886214 14.679049
#> 408 204 2002-07-23 4.09503446 15.828384 38.713200 14.686136
#> 409 205 2002-07-24 3.87566363 15.780249 38.535519 14.366083
#> 411 206 2002-07-25 4.24234396 15.731088 38.353226 14.759836
#> 413 207 2002-07-26 3.36573545 15.680933 38.166382 13.637587
#> 416 208 2002-07-27 5.84830099 15.629816 37.975044 16.598435
#> 417 209 2002-07-28 4.94219282 15.577768 37.779274 15.437733
#> 419 210 2002-07-29 5.33047667 15.524820 37.579134 15.846217
#> 421 211 2002-07-30 6.88472128 15.471001 37.374685 17.659691
#> 424 212 2002-07-31 0.41905223 15.416342 37.165992  9.796627
#> 425 213 2002-08-01 1.84936553 15.360873 36.953119 11.462757
#> 427 214 2002-08-02 0.34721873 15.304621 36.736132  9.600753
#> 429 215 2002-08-03 0.88743623 15.247616 36.515097 10.191393
#> 432 216 2002-08-04 3.79162767 15.189886 36.290082 13.601800
#> 434 217 2002-08-05 3.27809507 15.131457 36.061156 12.921453
#> 435 218 2002-08-06 4.01434874 15.072356 35.828388 13.728336
#> 437 219 2002-08-07 4.23037204 15.012610 35.591848 13.912638
#> 439 220 2002-08-08 0.70780063 14.952243 35.351609  9.674629
#> 442 221 2002-08-09 3.14148020 14.891281 35.107744 12.480119
#> 443 222 2002-08-10 1.58051339 14.829748 34.860325 10.572740
#> 445 223 2002-08-11 5.58252484 14.767666 34.609429 15.193946
#> 447 224 2002-08-12 2.35790961 14.705060 34.355129 11.343150
#> 450 225 2002-08-13 0.15496775 14.641952 34.097505  8.704817
#> 451 226 2002-08-14 4.84573795 14.578362 33.836632 14.082679
#> 453 227 2002-08-15 2.19941786 14.514313 33.572591 10.936849
#> 455 228 2002-08-16 6.67312688 14.449824 33.305462 16.016825
#> 458 229 2002-08-17 1.55394996 14.384916 33.035326 10.043174
#> 460 230 2002-08-18 6.32556650 14.319607 32.762264 15.426794
#> 461 231 2002-08-19 1.44414896 14.253916 32.486361  9.767283
#> 463 232 2002-08-20 0.63435782 14.187862 32.207701  8.771950
#> 465 233 2002-08-21 4.58206059 14.121461 31.926370 13.161247
#> 468 234 2002-08-22 2.37296205 14.054730 31.642455 10.581826
#> 469 235 2002-08-23 4.91500910 13.987686 31.356043 13.347972
#> 471 236 2002-08-24 3.74859649 13.920345 31.067223 11.949837
#> 473 237 2002-08-25 2.99599983 13.852721 30.776085 11.022073
#> 476 238 2002-08-26 4.59688902 13.784830 30.482721 12.703285
#> 477 239 2002-08-27 2.99694430 13.716686 30.187223 10.844593
#> 479 240 2002-08-28 2.34732321 13.648302 29.889684 10.042732
#> 481 241 2002-08-29 3.88401663 13.579692 29.590199 11.629193
#> 484 242 2002-08-30 5.55184582 13.510869 29.288864 13.339862
#> 485 243 2002-08-31 5.14572880 13.441846 28.985774 12.794526
#> 487 244 2002-09-01 5.34404612 13.372633 28.681027 12.901092
#> 490 245 2002-09-02 2.13060604 13.303244 28.374722  9.365884
#> 491 246 2002-09-03 4.84763128 13.233689 28.066958 12.157342
#> 493 247 2002-09-04 1.46182646 13.163980 27.757836  8.480678
#> 495 248 2002-09-05 2.50916805 13.094126 27.447457  9.491680
#> 498 249 2002-09-06 2.53505334 13.024138 27.135924  9.424885
#> 499 250 2002-09-07 6.71481434 12.954027 26.823338 13.657871
#> 501 251 2002-09-08 0.20079510 12.883802 26.509805  6.834030
#> 503 252 2002-09-09 3.42656477 12.813472 26.195428 10.051434
#> 506 253 2002-09-10 4.89859346 12.743047 25.880314 11.444443
#> 507 254 2002-09-11 2.89874308 12.672536 25.564567  9.314989
#> 509 255 2002-09-12 1.55337380 12.601947 25.248295  7.868184
#> 511 256 2002-09-13 0.88907507 12.531289 24.931604  7.117330
#> 514 257 2002-09-14 1.90753891 12.460572 24.614602  8.037726
#> 516 258 2002-09-15 0.23523168 12.389802 24.297397  6.305003
#> 517 259 2002-09-16 4.93238377 12.318988 23.980097 10.795704
#> 519 260 2002-09-17 0.97029747 12.248139 23.662812  6.852987
#> 521 261 2002-09-18 5.16943270 12.177262 23.345649 10.791704
#> 524 262 2002-09-19 3.70910604 12.106365 23.028718  9.284909
#> 525 263 2002-09-20 2.20414355 12.035456 22.712128  7.757753
#> 527 264 2002-09-21 6.09234828 11.964544 22.395989 11.301018
#> 529 265 2002-09-22 2.51955513 11.893635 22.080409  7.858866
#> 532 266 2002-09-23 4.45422189 11.822738 21.765497  9.541455
#> 533 267 2002-09-24 5.86585045 11.751861 21.451363 10.716481
#> 535 268 2002-09-25 2.70469920 11.681012 21.138114  7.731758
#> 537 269 2002-09-26 5.32019278 11.610198 20.825859  9.978028
#> 540 270 2002-09-27 2.26453030 11.539428 20.514705  7.141608
#> 542 271 2002-09-28 1.42635667 11.468711 20.204761  6.307617
#> 543 272 2002-09-29 0.21768086 11.398053 19.896131  5.164022
#> 545 273 2002-09-30 4.84925075 11.327464 19.588923  9.090208
#> 547 274 2002-10-01 5.68930484 11.256953 19.283241  9.693720
#> 550 275 2002-10-02 1.55023948 11.186528 18.979189  6.059874
#> 551 276 2002-10-03 0.38815190 11.116198 18.676871  4.995294
#> 553 277 2002-10-04 4.35413271 11.045973 18.376389  8.215926
#> 555 278 2002-10-05 3.30349471 10.975862 18.077844  7.239979
#> 558 279 2002-10-06 3.13428513 10.905874 17.781337  7.000461
#> 559 280 2002-10-07 5.10249132 10.836020 17.486965  8.488893
#> 561 281 2002-10-08 0.95555859 10.766311 17.194826  5.061766
#> 563 282 2002-10-09 3.45477686 10.696756 16.905016  6.956197
#> 566 283 2002-10-10 6.23656039 10.627367 16.617629  9.030349
#> 567 284 2002-10-11 6.50926897 10.558154 16.332759  9.117892
#> 570 285 2002-10-12 2.64597941 10.489131 16.050495  6.037066
#> 571 286 2002-10-13 4.00006245 10.420308 15.770928  6.969739
#> 573 287 2002-10-14 2.30587890 10.351698 15.494145  5.599225
#> 575 288 2002-10-15 2.85001356 10.283314 15.220231  5.914196
#> 578 289 2002-10-16 4.63687298 10.215170 14.949271  7.130206
#> 579 290 2002-10-17 0.99471829 10.147279 14.681345  4.389928
#> 581 291 2002-10-18 2.71886938 10.079655 14.416534  5.548479
#> 583 292 2002-10-19 1.79414191 10.012314 14.154915  4.806963
#> 586 293 2002-10-20 0.26171398  9.945270 13.896562  3.656988
#> 587 294 2002-10-21 5.11821485  9.878539 13.641550  6.944330
#> 589 295 2002-10-22 1.35907915  9.812138 13.389948  4.274808
#> 591 296 2002-10-23 2.33435112  9.746084 13.141825  4.859300
#> 594 297 2002-10-24 6.86193859  9.680393 12.897247  7.795413
#> 596 298 2002-10-25 0.07467544  9.615084 12.656277  3.213217
#> 597 299 2002-10-26 3.61518840  9.550176 12.418977  5.455326
#> 599 300 2002-10-27 2.68746281  9.485687 12.185404  4.772521
#> 601 301 2002-10-28 4.78648605  9.421638 11.955615  6.025817
#> 604 302 2002-10-29 4.32193688  9.358048 11.729663  5.641040
#> 605 303 2002-10-30 2.91296123  9.294940 11.507600  4.680096
#> 607 304 2002-10-31 5.43646019  9.232334 11.289473  6.146272
#> 609 305 2002-11-01 0.40793236  9.170252 11.075328  3.015171
#> 612 306 2002-11-02 3.97141773  9.108719 10.865209  5.084928
#> 613 307 2002-11-03 1.26355169  9.047757 10.659156  3.409084
#> 615 308 2002-11-04 0.93204555  8.987390 10.457208  3.156539
#> 617 309 2002-11-05 5.44953800  8.927644 10.259398  5.696078
#> 620 310 2002-11-06 1.07433485  8.868543 10.065762  3.126123
#> 622 311 2002-11-07 4.69733777  8.810114  9.876328  5.101991
#> 623 312 2002-11-08 2.97431932  8.752384  9.691126  4.069447
#> 625 313 2002-11-09 4.06038364  8.695379  9.510180  4.597976
#> 627 314 2002-11-10 4.60683143  8.639127  9.333513  4.821935
#> 630 315 2002-11-11 4.77318029  8.583658  9.161147  4.837442
#> 631 316 2002-11-12 4.83136741  8.528999  8.993100  4.795407
#> 633 317 2002-11-13 0.60788695  8.475180  8.829388  2.523993
#> 635 318 2002-11-14 0.31428548  8.422232  8.670025  2.329272
#> 638 319 2002-11-15 3.29980377  8.370184  8.515023  3.807208
#> 639 320 2002-11-16 5.76332510  8.319067  8.364392  4.988461
#> 641 321 2002-11-17 1.30908503  8.268912  8.218139  2.705058
#> 643 322 2002-11-18 4.83924701  8.219751  8.076271  4.396455
#> 646 323 2002-11-19 6.64528384  8.171616  7.938791  5.212671
#> 647 324 2002-11-20 1.77466369  8.124538  7.805702  2.803935
#> 650 325 2002-11-21 2.35867143  8.078549  7.677004  3.039968
#> 651 326 2002-11-22 2.02351310  8.033683  7.552698  2.839356
#> 653 327 2002-11-23 6.54179126  7.989971  7.432780  4.900991
#> 655 328 2002-11-24 5.28239359  7.947446  7.317248  4.261073
#> 658 329 2002-11-25 0.57618446  7.906140  7.206095  2.064107
#> 659 330 2002-11-26 1.43775804  7.866085  7.099317  2.423634
#> 661 331 2002-11-27 4.14304056  7.827314  6.996907  3.600978
#> 663 332 2002-11-28 4.78612593  7.789856  6.898857  3.844060
#> 666 333 2002-11-29 3.69172809  7.753745  6.805157  3.321332
#> 667 334 2002-11-30 3.72951673  7.719011  6.715800  3.301353
#> 669 335 2002-12-01 0.14868236  7.685684  6.630775  1.721831
#> 671 336 2002-12-02 5.87231344  7.653793  6.550072  4.150264
#> 674 337 2002-12-03 0.20907533  7.623368  6.473680  1.707192
#> 676 338 2002-12-04 1.12987739  7.594437  6.401590  2.076602
#> 677 339 2002-12-05 5.80798370  7.567027  6.333789  4.014161
#> 679 340 2002-12-06 6.20069089  7.541163  6.270268  4.145418
#> 681 341 2002-12-07 6.02061286  7.516872  6.211016  4.040100
#> 684 342 2002-12-08 1.57218543  7.494177  6.156023  2.184734
#> 685 343 2002-12-09 1.50864036  7.473100  6.105278  2.142574
#> 687 344 2002-12-10 0.14753576  7.453664  6.058772  1.574656
#> 689 345 2002-12-11 0.85204842  7.435887  6.016497  1.848827
#> 692 346 2002-12-12 3.61074768  7.419789  5.978443  2.949278
#> 693 347 2002-12-13 4.36167407  7.405386  5.944603  3.236797
#> 695 348 2002-12-14 1.16525747  7.392694  5.914970  1.944910
#> 697 349 2002-12-15 1.22566868  7.381725  5.889539  1.961337
#> 700 350 2002-12-16 4.61959561  7.372493  5.868303  3.305612
#> 702 351 2002-12-17 0.29875454  7.365006  5.851259  1.581490
#> 703 352 2002-12-18 1.46742799  7.359273  5.838404  2.041686
#> 705 353 2002-12-19 6.12941772  7.355300  5.829736  3.886491
#> 707 354 2002-12-20 2.83386218  7.353091  5.825253  2.578832
#> 710 355 2002-12-21 6.20503477  7.352649  5.824956  3.914132
#> 711 356 2002-12-22 0.05135896  7.353975  5.828845  1.477565
#> 713 357 2002-12-23 5.81985937  7.357066  5.836924  3.767901
#> 715 358 2002-12-24 4.22789113  7.361920  5.849194  3.141871
#> 718 359 2002-12-25 1.68581696  7.368531  5.865662  2.137406
#> 719 360 2002-12-26 2.45404486  7.376892  5.886331  2.450675
#> 721 361 2002-12-27 0.17882616  7.386993  5.911209  1.549352
#> 723 362 2002-12-28 1.69919213  7.398825  5.940303  2.167192
#> 726 363 2002-12-29 0.04394250  7.412375  5.973622  1.511112
#> 727 364 2002-12-30 4.57135372  7.427627  6.011174  3.352591
#> 729 365 2002-12-31 6.99903017  7.444567  6.052971  4.358601
#> 2     1 2003-01-01 4.86764756  7.463176  6.099033  3.513720
#> 4     2 2003-01-02 1.51227825  7.483435  6.149345  2.158676
#> 5     3 2003-01-03 2.54081300  7.505323  6.203934  2.601107
#> 8     4 2003-01-04 6.25141663  7.528820  6.262814  4.165809
#> 10    5 2003-01-05 2.30999243  7.553900  6.325996  2.548748
#> 12    6 2003-01-06 5.11412033  7.580540  6.393496  3.755021
#> 13    7 2003-01-07 2.00475320  7.608714  6.465324  2.468077
#> 16    8 2003-01-08 0.90595827  7.638396  6.541496  2.023304
#> 18    9 2003-01-09 3.72058544  7.669557  6.622023  3.261713
#> 20   10 2003-01-10 0.87905577  7.702170  6.706920  2.059463
#> 21   11 2003-01-11 2.89750738  7.736204  6.796198  2.971769
#> 24   12 2003-01-12 3.33735115  7.771631  6.889871  3.201817
#> 26   13 2003-01-13 3.59742698  7.808419  6.987950  3.356701
#> 28   14 2003-01-14 4.84240343  7.846537  7.090446  3.960507
#> 29   15 2003-01-15 0.23802075  7.885954  7.197371  1.907961
#> 31   16 2003-01-16 2.19773657  7.926639  7.308732  2.840391
#> 34   17 2003-01-17 4.71981679  7.968559  7.424539  4.054930
#> 36   18 2003-01-18 3.98478200  8.011681  7.544799  3.762484
#> 38   19 2003-01-19 5.84389318  8.055974  7.669519  4.699157
#> 39   20 2003-01-20 5.67503200  8.101405  7.798703  4.681170
#> 42   21 2003-01-21 2.91879380  8.147942  7.932354  3.403871
#> 44   22 2003-01-22 2.86922169  8.195553  8.070474  3.430334
#> 46   23 2003-01-23 1.79607601  8.244205  8.213063  2.947911
#> 47   24 2003-01-24 4.72241127  8.293867  8.360119  4.470097
#> 50   25 2003-01-25 3.59305794  8.344507  8.511639  3.960422
#> 52   26 2003-01-26 4.06542318  8.396093  8.667618  4.265353
#> 54   27 2003-01-27 2.66495124  8.448595  8.828046  3.599332
#> 55   28 2003-01-28 3.54200700  8.501983  8.992915  4.121496
#> 57   29 2003-01-29 1.93085322  8.556225  9.162213  3.324356
#> 60   30 2003-01-30 6.70524893  8.611293  9.335925  5.968726
#> 62   31 2003-01-31 1.45610878  8.667157  9.514035  3.177702
#> 64   32 2003-02-01 5.66463783  8.723789  9.696523  5.572264
#> 65   33 2003-02-02 6.93123902  8.781160  9.883368  6.371465
#> 68   34 2003-02-03 4.99620425  8.839243 10.074546  5.365854
#> 70   35 2003-02-04 6.95004633  8.898011 10.270030  6.578358
#> 72   36 2003-02-05 6.67932393  8.957438 10.469790  6.520970
#> 73   37 2003-02-06 6.19494335  9.017498 10.673795  6.334852
#> 76   38 2003-02-07 5.80099525  9.078165 10.882010  6.197334
#> 78   39 2003-02-08 1.97892264  9.139416 11.094398  3.974713
#> 80   40 2003-02-09 0.62457286  9.201226 11.310918  3.211618
#> 81   41 2003-02-10 4.21642702  9.263572 11.531528  5.507239
#> 84   42 2003-02-11 1.97697984  9.326432 11.756182  4.185060
#> 85   43 2003-02-12 5.40443701  9.389784 11.984831  6.445236
#> 88   44 2003-02-13 4.59503021  9.453606 12.217425  6.023564
#> 90   45 2003-02-14 3.07639373  9.517878 12.453910  5.126170
#> 92   46 2003-02-15 3.04628539  9.582579 12.694229  5.191294
#> 93   47 2003-02-16 2.04758665  9.647690 12.938322  4.607569
#> 96   48 2003-02-17 2.17747689  9.713192 13.186128  4.774547
#> 98   49 2003-02-18 1.42326699  9.779067 13.437583  4.337263
#> 100  50 2003-02-19 6.98644711  9.845297 13.692618  8.281451
#> 101  51 2003-02-20 3.51236562  9.911864 13.951164  5.959656
#> 104  52 2003-02-21 0.31011750  9.978754 14.213148  3.774144
#> 106  53 2003-02-22 6.09354584 10.045948 14.478496  8.010717
#> 108  54 2003-02-23 1.24845687 10.113433 14.747130  4.597015
#> 109  55 2003-02-24 0.05156412 10.181192 15.018971  3.792776
#> 111  56 2003-02-25 6.72432046 10.249211 15.293936  8.840520
#> 114  57 2003-02-26 4.96681018 10.317477 15.571940  7.641134
#> 116  58 2003-02-27 3.78364173 10.385975 15.852898  6.850853
#> 118  59 2003-02-28 6.84885989 10.454693 16.136720  9.319755
#> 119  60 2003-03-01 6.29093799 10.523618 16.423316  9.014695
#> 122  61 2003-03-02 2.93238389 10.592738 16.712592  6.491418
#> 124  62 2003-03-03 0.42748831 10.662040 17.004454  4.592005
#> 126  63 2003-03-04 2.99757750 10.731513 17.298805  6.740693
#> 127  64 2003-03-05 0.10769550 10.801147 17.595546  4.486607
#> 130  65 2003-03-06 3.89402724 10.870930 17.894578  7.678613
#> 132  66 2003-03-07 5.30842257 10.940852 18.195799  8.963185
#> 134  67 2003-03-08 4.92259788 11.010902 18.499104  8.759935
#> 135  68 2003-03-09 0.53582281 11.081072 18.804391  5.155739
#> 137  69 2003-03-10 1.20123556 11.151350 19.111552  5.807247
#> 140  70 2003-03-11 5.09511539 11.221729 19.420480  9.263959
#> 142  71 2003-03-12 6.23344574 11.292198 19.731067 10.378674
#> 144  72 2003-03-13 1.13253247 11.362749 20.043203  6.009660
#> 145  73 2003-03-14 4.47085601 11.433374 20.356778  9.069307
#> 148  74 2003-03-15 2.52380621 11.504062 20.671679  7.435437
#> 150  75 2003-03-16 1.60423776 11.574807 20.987796  6.701375
#> 152  76 2003-03-17 4.34431653 11.645600 21.305014  9.300103
#> 153  77 2003-03-18 3.37701685 11.716433 21.623220  8.522026
#> 156  78 2003-03-19 5.45891153 11.787297 21.942300 10.566514
#> 158  79 2003-03-20 1.89203338 11.858185 22.262140  7.341553
#> 160  80 2003-03-21 6.52439760 11.929088 22.582624 11.821233
#> 161  81 2003-03-22 4.94651940 12.000000 22.903638 10.446463
#> 164  82 2003-03-23 2.07182152 12.070912 23.225065  7.799413
#> 165  83 2003-03-24 5.81928199 12.141815 23.546791 11.529405
#> 168  84 2003-03-25 2.55828955 12.212703 23.868700  8.467156
#> 170  85 2003-03-26 3.22120671 12.283567 24.190677  9.219515
#> 172  86 2003-03-27 0.08292269 12.354400 24.512606  6.210416
#> 173  87 2003-03-28 1.52833057 12.425193 24.834373  7.735939
#> 176  88 2003-03-29 5.59526619 12.495938 25.155863 11.920946
#> 178  89 2003-03-30 4.01161027 12.566626 25.476961 10.435711
#> 180  90 2003-03-31 5.02997664 12.637251 25.797553 11.583460
#> 181  91 2003-04-01 5.79812152 12.707802 26.117528 12.487635
#> 184  92 2003-04-02 4.60019416 12.778271 26.436771 11.367829
#> 186  93 2003-04-03 4.04953692 12.848650 26.755170 10.905035
#> 188  94 2003-04-04 5.90830475 12.918928 27.072615 12.958809
#> 189  95 2003-04-05 1.86403002 12.989098 27.388994  8.812509
#> 191  96 2003-04-06 2.77354081 13.059148 27.704199  9.868000
#> 194  97 2003-04-07 0.67984694 13.129070 28.018120  7.729944
#> 196  98 2003-04-08 2.48580862 13.198853 28.330649  9.750491
#> 198  99 2003-04-09 2.25631797 13.268487 28.641680  9.595691
#> 199 100 2003-04-10 3.86941417 13.337960 28.951106 11.437212
#> 202 101 2003-04-11 6.70897941 13.407262 29.258822 14.635247
#> 204 102 2003-04-12 2.05254836 13.476382 29.564726  9.642640
#> 206 103 2003-04-13 6.27950716 13.545307 29.868714 14.390639
#> 207 104 2003-04-14 2.92462122 13.614025 30.170685 10.783367
#> 210 105 2003-04-15 5.15944844 13.682523 30.470539 13.362598
#> 212 106 2003-04-16 5.98466327 13.750789 30.768176 14.387557
#> 214 107 2003-04-17 6.62349325 13.818808 31.063500 15.210398
#> 215 108 2003-04-18 3.05458735 13.886567 31.356413 11.287792
#> 217 109 2003-04-19 4.56142744 13.954052 31.646820 13.084205
#> 220 110 2003-04-20 4.55568646 14.021246 31.934628 13.171646
#> 222 111 2003-04-21 0.63306389 14.088136 32.219743  8.778848
#> 224 112 2003-04-22 6.43955732 14.154703 32.502075 15.518785
#> 225 113 2003-04-23 6.79408061 14.220933 32.781534 16.026106
#> 228 114 2003-04-24 1.41304176 14.286808 33.058030  9.899316
#> 230 115 2003-04-25 6.31176715 14.352310 33.331477 15.662021
#> 232 116 2003-04-26 4.86217505 14.417421 33.601789 14.066432
#> 233 117 2003-04-27 2.80552678 14.482122 33.868881 11.747819
#> 236 118 2003-04-28 3.90209693 14.546394 34.132670 13.111243
#> 238 119 2003-04-29 5.49861979 14.610216 34.393074 15.070262
#> 240 120 2003-04-30 5.98350161 14.673568 34.650012 15.727193
#> 241 121 2003-05-01 2.12476448 14.736428 34.903405 11.242116
#> 244 122 2003-05-02 0.25295827 14.798774 35.153175  9.088734
#> 246 123 2003-05-03 0.52013018 14.860584 35.399245  9.469310
#> 247 124 2003-05-04 6.25949902 14.921835 35.641540 16.385947
#> 250 125 2003-05-05 5.62238095 14.982502 35.879987 15.702215
#> 252 126 2003-05-06 0.28698326 15.042562 36.114511  9.373126
#> 254 127 2003-05-07 5.03888181 15.101989 36.345041 15.149646
#> 255 128 2003-05-08 6.36907121 15.160757 36.571508 16.824767
#> 258 129 2003-05-09 4.66241772 15.218840 36.793843 14.834510
#> 260 130 2003-05-10 5.04710985 15.276211 37.011977 15.367191
#> 262 131 2003-05-11 1.93774883 15.332843 37.225844 11.658743
#> 263 132 2003-05-12 1.52594571 15.388707 37.435378 11.214893
#> 266 133 2003-05-13 3.97730217 15.443775 37.640517 14.256992
#> 268 134 2003-05-14 1.32361169 15.498017 37.841197 11.076217
#> 270 135 2003-05-15 4.17071663 15.551405 38.037357 14.609941
#> 271 136 2003-05-16 4.22766927 15.603907 38.228936 14.736043
#> 273 137 2003-05-17 4.86062344 15.655493 38.415875 15.567534
#> 276 138 2003-05-18 1.95536176 15.706133 38.598118 12.052199
#> 278 139 2003-05-19 1.19215394 15.755795 38.775606 11.160870
#> 280 140 2003-05-20 2.25297044 15.804447 38.948286 12.513168
#> 281 141 2003-05-21 4.91139475 15.852058 39.116103 15.838637
#> 284 142 2003-05-22 5.31262217 15.898595 39.279005 16.382423
#> 286 143 2003-05-23 4.46889803 15.944026 39.436940 15.386060
#> 288 144 2003-05-24 6.43350463 15.988319 39.589859 17.862703
#> 289 145 2003-05-25 4.71487771 16.031441 39.737713 15.777897
#> 292 146 2003-05-26 4.18172698 16.073361 39.880455 15.157864
#> 294 147 2003-05-27 1.04288303 16.114046 40.018039 11.299471
#> 296 148 2003-05-28 1.17851264 16.153463 40.150421 11.502238
#> 297 149 2003-05-29 6.56407494 16.191581 40.277558 18.233660
#> 299 150 2003-05-30 4.40444119 16.228369 40.399408 15.582129
#> 302 151 2003-05-31 0.63467587 16.263796 40.515933 10.919527
#> 304 152 2003-06-01 1.31690730 16.297830 40.627093 11.798161
#> 306 153 2003-06-02 5.33557703 16.330443 40.732853 16.837450
#> 307 154 2003-06-03 0.29776007 16.361604 40.833176 10.579850
#> 310 155 2003-06-04 1.88601276 16.391286 40.928031 12.586637
#> 312 156 2003-06-05 5.80817179 16.419460 41.017385 17.509031
#> 314 157 2003-06-06 4.94446839 16.446100 41.101209 16.453777
#> 315 158 2003-06-07 6.82913917 16.471180 41.179475 18.831609
#> 318 159 2003-06-08 2.58556601 16.494677 41.252156 13.546209
#> 320 160 2003-06-09 4.70110661 16.516565 41.319229 16.210149
#> 322 161 2003-06-10 3.83445576 16.536824 41.380671 15.142714
#> 323 162 2003-06-11 4.70912505 16.555433 41.436461 16.252331
#> 326 163 2003-06-12 1.41541036 16.572373 41.486582 12.143285
#> 327 164 2003-06-13 3.06658476 16.587625 41.531016 14.221712
#> 330 165 2003-06-14 4.43928665 16.601175 41.569750 15.950479
#> 332 166 2003-06-15 2.63372539 16.613007 41.602770 13.698418
#> 334 167 2003-06-16 5.66422088 16.623108 41.630068 17.500110
#> 335 168 2003-06-17 6.21666850 16.631469 41.651634 18.197380
#> 338 169 2003-06-18 3.95294074 16.638080 41.667464 15.366626
#> 340 170 2003-06-19 1.23490851 16.642934 41.677552 11.965629
#> 342 171 2003-06-20 2.58754926 16.646025 41.681897 13.660105
#> 343 172 2003-06-21 6.98601401 16.647351 41.680500 19.165677
#> 346 173 2003-06-22 6.55588885 16.646909 41.673363 18.624247
#> 348 174 2003-06-23 3.99452428 16.644700 41.660491 15.414127
#> 350 175 2003-06-24 2.77383808 16.640727 41.641891 13.881110
#> 351 176 2003-06-25 5.00319677 16.634994 41.617572 16.662913
#> 353 177 2003-06-26 0.70036028 16.627507 41.587545 11.272732
#> 356 178 2003-06-27 1.80968467 16.618275 41.551823 12.650396
#> 358 179 2003-06-28 2.07292020 16.607306 41.510423 12.968266
#> 360 180 2003-06-29 3.45422126 16.594614 41.463360 14.681205
#> 361 181 2003-06-30 2.62367757 16.580211 41.410655 13.629106
#> 364 182 2003-07-01 5.28211735 16.564113 41.352328 16.931488
#> 366 183 2003-07-02 0.09395082 16.546336 41.288405 10.439320
#> 368 184 2003-07-03 5.53733939 16.526900 41.218909 17.209927
#> 369 185 2003-07-04 5.80397805 16.505823 41.143868 17.519720
#> 372 186 2003-07-05 5.25870987 16.483128 41.063311 16.816163
#> 374 187 2003-07-06 4.93182187 16.458837 40.977269 16.383651
#> 376 188 2003-07-07 4.62470526 16.432973 40.885775 15.974653
#> 377 189 2003-07-08 3.90600358 16.405563 40.788864 15.052930
#> 379 190 2003-07-09 6.96354279 16.376632 40.686570 18.821855
#> 382 191 2003-07-10 0.37011588 16.346207 40.578932 10.604133
#> 384 192 2003-07-11 3.33412428 16.314316 40.465990 14.251474
#> 386 193 2003-07-12 4.36885700 16.280989 40.347784 15.500429
#> 387 194 2003-07-13 5.59149651 16.246255 40.224356 16.978126
#> 390 195 2003-07-14 5.81376381 16.210144 40.095751 17.214103
#> 392 196 2003-07-15 6.19398410 16.172686 39.962013 17.643037
#> 394 197 2003-07-16 1.75968541 16.133915 39.823189 12.127505
#> 395 198 2003-07-17 1.33295035 16.093860 39.679327 11.563023
#> 398 199 2003-07-18 3.28753194 16.052554 39.530476 13.930502
#> 400 200 2003-07-19 0.59755080 16.010029 39.376687 10.579010
#> 402 201 2003-07-20 6.13889247 15.966317 39.218012 17.343974
#> 403 202 2003-07-21 4.47430607 15.921451 39.054503 15.251247
#> 406 203 2003-07-22 4.04783730 15.875462 38.886214 14.679049
#> 407 204 2003-07-23 4.09503446 15.828384 38.713200 14.686136
#> 410 205 2003-07-24 3.87566363 15.780249 38.535519 14.366083
#> 412 206 2003-07-25 4.24234396 15.731088 38.353226 14.759836
#> 414 207 2003-07-26 3.36573545 15.680933 38.166382 13.637587
#> 415 208 2003-07-27 5.84830099 15.629816 37.975044 16.598435
#> 418 209 2003-07-28 4.94219282 15.577768 37.779274 15.437733
#> 420 210 2003-07-29 5.33047667 15.524820 37.579134 15.846217
#> 422 211 2003-07-30 6.88472128 15.471001 37.374685 17.659691
#> 423 212 2003-07-31 0.41905223 15.416342 37.165992  9.796627
#> 426 213 2003-08-01 1.84936553 15.360873 36.953119 11.462757
#> 428 214 2003-08-02 0.34721873 15.304621 36.736132  9.600753
#> 430 215 2003-08-03 0.88743623 15.247616 36.515097 10.191393
#> 431 216 2003-08-04 3.79162767 15.189886 36.290082 13.601800
#> 433 217 2003-08-05 3.27809507 15.131457 36.061156 12.921453
#> 436 218 2003-08-06 4.01434874 15.072356 35.828388 13.728336
#> 438 219 2003-08-07 4.23037204 15.012610 35.591848 13.912638
#> 440 220 2003-08-08 0.70780063 14.952243 35.351609  9.674629
#> 441 221 2003-08-09 3.14148020 14.891281 35.107744 12.480119
#> 444 222 2003-08-10 1.58051339 14.829748 34.860325 10.572740
#> 446 223 2003-08-11 5.58252484 14.767666 34.609429 15.193946
#> 448 224 2003-08-12 2.35790961 14.705060 34.355129 11.343150
#> 449 225 2003-08-13 0.15496775 14.641952 34.097505  8.704817
#> 452 226 2003-08-14 4.84573795 14.578362 33.836632 14.082679
#> 454 227 2003-08-15 2.19941786 14.514313 33.572591 10.936849
#> 456 228 2003-08-16 6.67312688 14.449824 33.305462 16.016825
#> 457 229 2003-08-17 1.55394996 14.384916 33.035326 10.043174
#> 459 230 2003-08-18 6.32556650 14.319607 32.762264 15.426794
#> 462 231 2003-08-19 1.44414896 14.253916 32.486361  9.767283
#> 464 232 2003-08-20 0.63435782 14.187862 32.207701  8.771950
#> 466 233 2003-08-21 4.58206059 14.121461 31.926370 13.161247
#> 467 234 2003-08-22 2.37296205 14.054730 31.642455 10.581826
#> 470 235 2003-08-23 4.91500910 13.987686 31.356043 13.347972
#> 472 236 2003-08-24 3.74859649 13.920345 31.067223 11.949837
#> 474 237 2003-08-25 2.99599983 13.852721 30.776085 11.022073
#> 475 238 2003-08-26 4.59688902 13.784830 30.482721 12.703285
#> 478 239 2003-08-27 2.99694430 13.716686 30.187223 10.844593
#> 480 240 2003-08-28 2.34732321 13.648302 29.889684 10.042732
#> 482 241 2003-08-29 3.88401663 13.579692 29.590199 11.629193
#> 483 242 2003-08-30 5.55184582 13.510869 29.288864 13.339862
#> 486 243 2003-08-31 5.14572880 13.441846 28.985774 12.794526
#> 488 244 2003-09-01 5.34404612 13.372633 28.681027 12.901092
#> 489 245 2003-09-02 2.13060604 13.303244 28.374722  9.365884
#> 492 246 2003-09-03 4.84763128 13.233689 28.066958 12.157342
#> 494 247 2003-09-04 1.46182646 13.163980 27.757836  8.480678
#> 496 248 2003-09-05 2.50916805 13.094126 27.447457  9.491680
#> 497 249 2003-09-06 2.53505334 13.024138 27.135924  9.424885
#> 500 250 2003-09-07 6.71481434 12.954027 26.823338 13.657871
#> 502 251 2003-09-08 0.20079510 12.883802 26.509805  6.834030
#> 504 252 2003-09-09 3.42656477 12.813472 26.195428 10.051434
#> 505 253 2003-09-10 4.89859346 12.743047 25.880314 11.444443
#> 508 254 2003-09-11 2.89874308 12.672536 25.564567  9.314989
#> 510 255 2003-09-12 1.55337380 12.601947 25.248295  7.868184
#> 512 256 2003-09-13 0.88907507 12.531289 24.931604  7.117330
#> 513 257 2003-09-14 1.90753891 12.460572 24.614602  8.037726
#> 515 258 2003-09-15 0.23523168 12.389802 24.297397  6.305003
#> 518 259 2003-09-16 4.93238377 12.318988 23.980097 10.795704
#> 520 260 2003-09-17 0.97029747 12.248139 23.662812  6.852987
#> 522 261 2003-09-18 5.16943270 12.177262 23.345649 10.791704
#> 523 262 2003-09-19 3.70910604 12.106365 23.028718  9.284909
#> 526 263 2003-09-20 2.20414355 12.035456 22.712128  7.757753
#> 528 264 2003-09-21 6.09234828 11.964544 22.395989 11.301018
#> 530 265 2003-09-22 2.51955513 11.893635 22.080409  7.858866
#> 531 266 2003-09-23 4.45422189 11.822738 21.765497  9.541455
#> 534 267 2003-09-24 5.86585045 11.751861 21.451363 10.716481
#> 536 268 2003-09-25 2.70469920 11.681012 21.138114  7.731758
#> 538 269 2003-09-26 5.32019278 11.610198 20.825859  9.978028
#> 539 270 2003-09-27 2.26453030 11.539428 20.514705  7.141608
#> 541 271 2003-09-28 1.42635667 11.468711 20.204761  6.307617
#> 544 272 2003-09-29 0.21768086 11.398053 19.896131  5.164022
#> 546 273 2003-09-30 4.84925075 11.327464 19.588923  9.090208
#> 548 274 2003-10-01 5.68930484 11.256953 19.283241  9.693720
#> 549 275 2003-10-02 1.55023948 11.186528 18.979189  6.059874
#> 552 276 2003-10-03 0.38815190 11.116198 18.676871  4.995294
#> 554 277 2003-10-04 4.35413271 11.045973 18.376389  8.215926
#> 556 278 2003-10-05 3.30349471 10.975862 18.077844  7.239979
#> 557 279 2003-10-06 3.13428513 10.905874 17.781337  7.000461
#> 560 280 2003-10-07 5.10249132 10.836020 17.486965  8.488893
#> 562 281 2003-10-08 0.95555859 10.766311 17.194826  5.061766
#> 564 282 2003-10-09 3.45477686 10.696756 16.905016  6.956197
#> 565 283 2003-10-10 6.23656039 10.627367 16.617629  9.030349
#> 568 284 2003-10-11 6.50926897 10.558154 16.332759  9.117892
#> 569 285 2003-10-12 2.64597941 10.489131 16.050495  6.037066
#> 572 286 2003-10-13 4.00006245 10.420308 15.770928  6.969739
#> 574 287 2003-10-14 2.30587890 10.351698 15.494145  5.599225
#> 576 288 2003-10-15 2.85001356 10.283314 15.220231  5.914196
#> 577 289 2003-10-16 4.63687298 10.215170 14.949271  7.130206
#> 580 290 2003-10-17 0.99471829 10.147279 14.681345  4.389928
#> 582 291 2003-10-18 2.71886938 10.079655 14.416534  5.548479
#> 584 292 2003-10-19 1.79414191 10.012314 14.154915  4.806963
#> 585 293 2003-10-20 0.26171398  9.945270 13.896562  3.656988
#> 588 294 2003-10-21 5.11821485  9.878539 13.641550  6.944330
#> 590 295 2003-10-22 1.35907915  9.812138 13.389948  4.274808
#> 592 296 2003-10-23 2.33435112  9.746084 13.141825  4.859300
#> 593 297 2003-10-24 6.86193859  9.680393 12.897247  7.795413
#> 595 298 2003-10-25 0.07467544  9.615084 12.656277  3.213217
#> 598 299 2003-10-26 3.61518840  9.550176 12.418977  5.455326
#> 600 300 2003-10-27 2.68746281  9.485687 12.185404  4.772521
#> 602 301 2003-10-28 4.78648605  9.421638 11.955615  6.025817
#> 603 302 2003-10-29 4.32193688  9.358048 11.729663  5.641040
#> 606 303 2003-10-30 2.91296123  9.294940 11.507600  4.680096
#> 608 304 2003-10-31 5.43646019  9.232334 11.289473  6.146272
#> 610 305 2003-11-01 0.40793236  9.170252 11.075328  3.015171
#> 611 306 2003-11-02 3.97141773  9.108719 10.865209  5.084928
#> 614 307 2003-11-03 1.26355169  9.047757 10.659156  3.409084
#> 616 308 2003-11-04 0.93204555  8.987390 10.457208  3.156539
#> 618 309 2003-11-05 5.44953800  8.927644 10.259398  5.696078
#> 619 310 2003-11-06 1.07433485  8.868543 10.065762  3.126123
#> 621 311 2003-11-07 4.69733777  8.810114  9.876328  5.101991
#> 624 312 2003-11-08 2.97431932  8.752384  9.691126  4.069447
#> 626 313 2003-11-09 4.06038364  8.695379  9.510180  4.597976
#> 628 314 2003-11-10 4.60683143  8.639127  9.333513  4.821935
#> 629 315 2003-11-11 4.77318029  8.583658  9.161147  4.837442
#> 632 316 2003-11-12 4.83136741  8.528999  8.993100  4.795407
#> 634 317 2003-11-13 0.60788695  8.475180  8.829388  2.523993
#> 636 318 2003-11-14 0.31428548  8.422232  8.670025  2.329272
#> 637 319 2003-11-15 3.29980377  8.370184  8.515023  3.807208
#> 640 320 2003-11-16 5.76332510  8.319067  8.364392  4.988461
#> 642 321 2003-11-17 1.30908503  8.268912  8.218139  2.705058
#> 644 322 2003-11-18 4.83924701  8.219751  8.076271  4.396455
#> 645 323 2003-11-19 6.64528384  8.171616  7.938791  5.212671
#> 648 324 2003-11-20 1.77466369  8.124538  7.805702  2.803935
#> 649 325 2003-11-21 2.35867143  8.078549  7.677004  3.039968
#> 652 326 2003-11-22 2.02351310  8.033683  7.552698  2.839356
#> 654 327 2003-11-23 6.54179126  7.989971  7.432780  4.900991
#> 656 328 2003-11-24 5.28239359  7.947446  7.317248  4.261073
#> 657 329 2003-11-25 0.57618446  7.906140  7.206095  2.064107
#> 660 330 2003-11-26 1.43775804  7.866085  7.099317  2.423634
#> 662 331 2003-11-27 4.14304056  7.827314  6.996907  3.600978
#> 664 332 2003-11-28 4.78612593  7.789856  6.898857  3.844060
#> 665 333 2003-11-29 3.69172809  7.753745  6.805157  3.321332
#> 668 334 2003-11-30 3.72951673  7.719011  6.715800  3.301353
#> 670 335 2003-12-01 0.14868236  7.685684  6.630775  1.721831
#> 672 336 2003-12-02 5.87231344  7.653793  6.550072  4.150264
#> 673 337 2003-12-03 0.20907533  7.623368  6.473680  1.707192
#> 675 338 2003-12-04 1.12987739  7.594437  6.401590  2.076602
#> 678 339 2003-12-05 5.80798370  7.567027  6.333789  4.014161
#> 680 340 2003-12-06 6.20069089  7.541163  6.270268  4.145418
#> 682 341 2003-12-07 6.02061286  7.516872  6.211016  4.040100
#> 683 342 2003-12-08 1.57218543  7.494177  6.156023  2.184734
#> 686 343 2003-12-09 1.50864036  7.473100  6.105278  2.142574
#> 688 344 2003-12-10 0.14753576  7.453664  6.058772  1.574656
#> 690 345 2003-12-11 0.85204842  7.435887  6.016497  1.848827
#> 691 346 2003-12-12 3.61074768  7.419789  5.978443  2.949278
#> 694 347 2003-12-13 4.36167407  7.405386  5.944603  3.236797
#> 696 348 2003-12-14 1.16525747  7.392694  5.914970  1.944910
#> 698 349 2003-12-15 1.22566868  7.381725  5.889539  1.961337
#> 699 350 2003-12-16 4.61959561  7.372493  5.868303  3.305612
#> 701 351 2003-12-17 0.29875454  7.365006  5.851259  1.581490
#> 704 352 2003-12-18 1.46742799  7.359273  5.838404  2.041686
#> 706 353 2003-12-19 6.12941772  7.355300  5.829736  3.886491
#> 708 354 2003-12-20 2.83386218  7.353091  5.825253  2.578832
#> 709 355 2003-12-21 6.20503477  7.352649  5.824956  3.914132
#> 712 356 2003-12-22 0.05135896  7.353975  5.828845  1.477565
#> 714 357 2003-12-23 5.81985937  7.357066  5.836924  3.767901
#> 716 358 2003-12-24 4.22789113  7.361920  5.849194  3.141871
#> 717 359 2003-12-25 1.68581696  7.368531  5.865662  2.137406
#> 720 360 2003-12-26 2.45404486  7.376892  5.886331  2.450675
#> 722 361 2003-12-27 0.17882616  7.386993  5.911209  1.549352
#> 724 362 2003-12-28 1.69919213  7.398825  5.940303  2.167192
#> 725 363 2003-12-29 0.04394250  7.412375  5.973622  1.511112
#> 728 364 2003-12-30 4.57135372  7.427627  6.011174  3.352591
#> 730 365 2003-12-31 6.99903017  7.444567  6.052971  4.358601