Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

Usage

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.081501  2.081501  2.467927  2.467927  2.553810  2.553810  1.839109
#>   [8]  1.839109  3.158180  3.158180  4.322021  4.322021  2.039250  2.039250
#>  [15]  3.483562  3.483562  1.945384  1.945384  2.664661  2.664661  4.729045
#>  [22]  4.729045  3.166551  3.166551  2.065368  2.065368  4.169492  4.169492
#>  [29]  2.423345  2.423345  3.570106  3.570106  3.544807  3.544807  4.780907
#>  [36]  4.780907  4.945919  4.945919  4.623416  4.623416  2.021544  2.021544
#>  [43]  4.593240  4.593240  4.923421  4.923421  4.079770  4.079770  3.692598
#>  [50]  3.692598  4.984723  4.984723  5.644671  5.644671  5.545185  5.545185
#>  [57]  3.516180  3.516180  4.125155  4.125155  6.138532  6.138532  4.723388
#>  [64]  4.723388  3.316381  3.316381  3.706630  3.706630  2.799672  2.799672
#>  [71]  3.421695  3.421695  2.946222  2.946222  5.397148  5.397148  5.193651
#>  [78]  5.193651  3.533905  3.533905  4.150489  4.150489  7.037466  7.037466
#>  [85]  6.716181  6.716181  4.247325  4.247325  4.340681  4.340681  3.911774
#>  [92]  3.911774  7.191486  7.191486  6.011316  6.011316  5.628113  5.628113
#>  [99]  4.289713  4.289713  8.168001  8.168001  4.637678  4.637678  6.843619
#> [106]  6.843619  8.306531  8.306531  5.530822  5.530822  4.045849  4.045849
#> [113]  6.922055  6.922055  7.153980  7.153980  6.018452  6.018452  5.521603
#> [120]  5.521603  6.110255  6.110255  4.740921  4.740921  6.748801  6.748801
#> [127]  9.559600  9.559600  6.207166  6.207166  6.624156  6.624156  7.234270
#> [134]  7.234270  4.888915  4.888915  4.949106  4.949106  6.730327  6.730327
#> [141] 10.151914 10.151914  6.966179  6.966179  9.645144  9.645144  7.886862
#> [148]  7.886862 10.604793 10.604793  6.665773  6.665773 10.432410 10.432410
#> [155]  7.711941  7.711941  6.605169  6.605169 11.719833 11.719833  9.728647
#> [162]  9.728647  7.042104  7.042104  9.432487  9.432487  9.266531  9.266531
#> [169]  6.213390  6.213390 11.417313 11.417313 11.324207 11.324207  9.666567
#> [176]  9.666567 12.556877 12.556877  7.702025  7.702025 12.571638 12.571638
#> [183]  9.798268  9.798268  6.957224  6.957224  8.608678  8.608678 11.756860
#> [190] 11.756860  9.630658  9.630658 11.564749 11.564749 10.386311 10.386311
#> [197] 12.389949 12.389949 12.173652 12.173652  9.801004  9.801004 10.727082
#> [204] 10.727082  7.688476  7.688476  9.602546  9.602546  9.459869  9.459869
#> [211] 12.647822 12.647822 10.432360 10.432360 10.813150 10.813150 14.902388
#> [218] 14.902388 12.626525 12.626525 15.779473 15.779473  8.427624  8.427624
#> [225] 15.119725 15.119725 10.454698 10.454698  9.981874  9.981874 11.179231
#> [232] 11.179231 12.424756 12.424756 10.302736 10.302736 12.101378 12.101378
#> [239] 16.360828 16.360828 15.422307 15.422307 10.328810 10.328810 16.879772
#> [246] 16.879772 11.707048 11.707048  9.001195  9.001195 15.908412 15.908412
#> [253] 12.375790 12.375790 16.029606 16.029606 13.435994 13.435994 11.606676
#> [260] 11.606676 13.320280 13.320280 12.918279 12.918279 12.692754 12.692754
#> [267] 14.975597 14.975597 10.795618 10.795618 18.074830 18.074830 15.632764
#> [274] 15.632764 15.593188 15.593188 12.103895 12.103895 11.872275 11.872275
#> [281] 14.462639 14.462639 10.162060 10.162060 11.994133 11.994133 17.751865
#> [288] 17.751865 13.460775 13.460775 11.228469 11.228469 16.639750 16.639750
#> [295] 10.782694 10.782694 14.881851 14.881851 11.654539 11.654539 17.376754
#> [302] 17.376754 13.881630 13.881630 16.069821 16.069821 11.860349 11.860349
#> [309] 10.600495 10.600495 15.939198 15.939198 16.525238 16.525238 17.769588
#> [316] 17.769588 17.054020 17.054020 18.923578 18.923578 16.679224 16.679224
#> [323] 15.279220 15.279220 17.089002 17.089002 12.549818 12.549818 17.643187
#> [330] 17.643187 13.927455 13.927455 15.813355 15.813355 11.415552 11.415552
#> [337] 14.823984 14.823984 12.260077 12.260077 11.203968 11.203968 16.937991
#> [344] 16.937991 16.852672 16.852672 11.255298 11.255298 11.087740 11.087740
#> [351] 16.432641 16.432641 10.686582 10.686582 17.274372 17.274372 14.049477
#> [358] 14.049477 16.312283 16.312283 15.334448 15.334448 11.379984 11.379984
#> [365] 18.681266 18.681266 16.271739 16.271739 17.370901 17.370901 15.668261
#> [372] 15.668261 16.745864 16.745864 18.512190 18.512190 17.486090 17.486090
#> [379] 11.913661 11.913661 18.692798 18.692798 16.184892 16.184892 14.736699
#> [386] 14.736699 17.496297 17.496297 13.976085 13.976085 10.545732 10.545732
#> [393] 16.769639 16.769639 12.590398 12.590398 14.548799 14.548799 12.530333
#> [400] 12.530333 14.278519 14.278519 13.508318 13.508318 11.626764 11.626764
#> [407] 13.640407 13.640407 14.789837 14.789837 12.489019 12.489019 11.857634
#> [414] 11.857634 13.359694 13.359694 10.874870 10.874870 10.679857 10.679857
#> [421] 10.997452 10.997452 14.604375 14.604375 11.832524 11.832524 12.620297
#> [428] 12.620297 16.140836 16.140836 11.441499 11.441499 11.972284 11.972284
#> [435] 15.351667 15.351667  9.992982  9.992982 13.935803 13.935803 12.704079
#> [442] 12.704079 10.604465 10.604465 12.605492 12.605492 10.733404 10.733404
#> [449] 10.015942 10.015942 15.328410 15.328410 11.806406 11.806406 14.681209
#> [456] 14.681209 15.988305 15.988305 13.438362 13.438362 11.771836 11.771836
#> [463]  9.985244  9.985244 12.109865 12.109865 12.963048 12.963048 10.248336
#> [470] 10.248336 10.362537 10.362537  9.571825  9.571825 12.100457 12.100457
#> [477]  9.764300  9.764300 10.149779 10.149779 11.260163 11.260163 10.895933
#> [484] 10.895933  9.567807  9.567807 11.947015 11.947015  7.687141  7.687141
#> [491] 13.511005 13.511005 11.759375 11.759375 10.346917 10.346917 12.975769
#> [498] 12.975769  8.432280  8.432280 11.201704 11.201704  7.302976  7.302976
#> [505]  7.383582  7.383582 10.230932 10.230932  9.024542  9.024542  8.515851
#> [512]  8.515851  6.395280  6.395280  7.006269  7.006269 11.709498 11.709498
#> [519]  6.737069  6.737069 11.128433 11.128433 11.694389 11.694389 10.710336
#> [526] 10.710336  8.124521  8.124521  8.759412  8.759412  6.284338  6.284338
#> [533]  5.710191  5.710191 10.090854 10.090854  9.073331  9.073331  8.421543
#> [540]  8.421543  9.298642  9.298642  8.677740  8.677740  6.952893  6.952893
#> [547]  5.490182  5.490182  8.256216  8.256216  6.723828  6.723828  6.498666
#> [554]  6.498666  4.810155  4.810155  5.476297  5.476297  8.166209  8.166209
#> [561]  9.775771  9.775771  7.219538  7.219538  7.536320  7.536320  9.441182
#> [568]  9.441182  6.847211  6.847211  6.542446  6.542446  8.210697  8.210697
#> [575]  6.332791  6.332791  7.650434  7.650434  6.478263  6.478263  4.860697
#> [582]  4.860697  3.796271  3.796271  6.652096  6.652096  3.810731  3.810731
#> [589]  4.443886  4.443886  6.779827  6.779827  4.708619  4.708619  7.454662
#> [596]  7.454662  3.605016  3.605016  4.258425  4.258425  7.085171  7.085171
#> [603]  3.563177  3.563177  4.241682  4.241682  6.468816  6.468816  5.514342
#> [610]  5.514342  5.523283  5.523283  4.522984  4.522984  6.015873  6.015873
#> [617]  4.390124  4.390124  4.943126  4.943126  3.322686  3.322686  5.726366
#> [624]  5.726366  6.105456  6.105456  3.189482  3.189482  4.575811  4.575811
#> [631]  3.537029  3.537029  4.478482  4.478482  4.795603  4.795603  2.193524
#> [638]  2.193524  2.636720  2.636720  2.258004  2.258004  3.032503  3.032503
#> [645]  2.941141  2.941141  3.608843  3.608843  3.697887  3.697887  3.111572
#> [652]  3.111572  4.455765  4.455765  3.187479  3.187479  1.862898  1.862898
#> [659]  3.216365  3.216365  3.011918  3.011918  2.396100  2.396100  2.052819
#> [666]  2.052819  1.882467  1.882467  2.056852  2.056852  3.278704  3.278704
#> [673]  3.985975  3.985975  3.935067  3.935067  4.144257  4.144257  3.250070
#> [680]  3.250070  2.061863  2.061863  2.304758  2.304758  3.214944  3.214944
#> [687]  2.862697  2.862697  4.006510  4.006510  3.913821  3.913821  3.366698
#> [694]  3.366698  1.986790  1.986790  3.451898  3.451898  2.414853  2.414853
#> [701]  3.721975  3.721975  2.713315  2.713315  3.630453  3.630453  4.096658
#> [708]  4.096658  3.691663  3.691663  3.465277  3.465277  3.917543  3.917543
#> [715]  1.862082  1.862082  4.184727  4.184727  3.010816  3.010816  1.876898
#> [722]  1.876898  1.563580  1.563580  4.166007  4.166007  1.781369  1.781369
#> [729]  2.773606  2.773606