Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.545923  2.878705  2.991111  2.409349  3.729625  3.878804  4.217430
#>   [8]  2.159739  1.758982  2.653183  2.936098  2.329611  3.010965  1.973956
#>  [15]  3.041005  4.975436  2.801489  4.122167  4.367539  2.609898  5.324172
#>  [22]  4.573334  2.232647  3.960597  4.612072  4.654787  2.321227  3.083284
#>  [29]  3.418033  4.749064  4.218916  4.105391  5.253710  6.302634  3.296018
#>  [36]  3.504770  5.486273  4.813387  5.499886  5.668599  3.301249  6.316743
#>  [43]  6.434570  7.535592  7.558138  4.978010  5.399288  4.794378  4.199484
#>  [50]  6.010706  5.919608  7.438291  4.649560  7.327680  4.091459  5.673407
#>  [57]  8.252078  5.426048  7.113686  5.939578  7.470766  5.320169  9.671863
#>  [64]  7.491923  7.611285  6.170628  7.251499  6.907666  4.946209  7.677674
#>  [71]  7.318049  5.134679  7.438343  8.689301 10.686265  7.790215  8.815839
#>  [78]  9.398923  7.282948  7.567456  8.932921 12.001752  8.606626  7.425373
#>  [85] 10.677980  6.535192 13.183580  7.339059 10.048777 12.494801 11.696129
#>  [92]  8.356571 10.675736 12.891025  7.053568 10.411247 13.022301 13.198277
#>  [99] 10.212042  8.897203 10.510188 12.527013 11.385142 12.664759 11.606734
#> [106] 14.235631 13.342353 14.748073  8.002827 15.064391 16.030239 12.145413
#> [113] 11.091551 14.540216 13.083671 13.571940 15.495694 13.188869 10.682749
#> [120] 16.256287 15.916004 10.854629 12.208759 15.344434  9.971496 10.664625
#> [127] 10.472466 14.742243 16.446608 17.110087 14.000276 14.272146 15.274344
#> [134] 11.551632  9.891293 17.359107 10.211007 18.224108 14.964173 11.225453
#> [141] 17.633417 10.143595 14.997699 11.951257 17.797080 17.081233 16.087420
#> [148] 11.951490 16.407618 11.991311 14.106903 13.060375 15.144940 12.411273
#> [155] 14.287154 18.277760 18.784782 17.462105 18.216585 15.420187 17.115309
#> [162] 19.076514 16.601000 12.266311 12.949426 16.726773 18.003596 12.503035
#> [169] 10.456277 18.689089 14.260332 16.997708 16.269604 13.989070 13.486860
#> [176] 16.867366 16.212010 11.133782 17.867347 11.039068 17.807877 11.267316
#> [183] 14.556185 12.462746 16.275877 11.692510 18.547125 13.024190 13.343282
#> [190] 17.890841 17.339050 10.990336 17.945812 16.752342 13.342870 18.635847
#> [197] 12.973398 18.094461 11.745181 10.120434 11.053789 17.098785 11.548901
#> [204] 11.478629  9.971664 17.650245 11.628082 16.136262 11.891438 16.810855
#> [211] 11.844316 17.590350 14.204611 15.561839 16.136610 15.451089 12.495808
#> [218] 10.105855  9.568796 14.266195 13.744439 14.120488 11.353608 16.597921
#> [225] 14.353600 15.548023 16.353555 10.089549 13.600217 11.310100  8.489036
#> [232] 12.953503 12.717290 11.116543 14.573377 11.810542 15.308833  7.752375
#> [239] 12.734161 10.318196 14.398770 12.466239 12.266577 12.845588 11.146074
#> [246]  8.792987 10.695181  9.923158 12.085099 11.327331  9.482247 13.414225
#> [253]  8.593022  6.745593 10.352491  7.750270  7.023808 12.513624 11.454182
#> [260] 11.041533  9.409940  9.397640  6.311613  8.143279  6.639995  9.892033
#> [267]  9.676583 11.278025  6.438359 11.155803  7.438046  8.947312  9.827700
#> [274]  5.246281  7.873189  9.158848  7.129222  7.704294  5.609641  4.546983
#> [281]  9.719208  5.214226  8.413866  8.878664  8.492387  7.149478  8.624412
#> [288]  3.991082  4.410448  4.146496  7.091170  5.546326  3.795141  4.021758
#> [295]  7.824372  4.306657  6.330346  4.103446  5.601589  6.766908  3.555430
#> [302]  4.078241  4.126661  2.885578  6.377615  4.444922  5.761480  5.542413
#> [309]  4.637661  6.388964  3.884620  3.083131  5.375382  2.450311  5.239145
#> [316]  2.863118  2.312041  4.999269  5.067318  5.487846  3.371642  2.618766
#> [323]  4.037617  4.659916  2.041248  4.299781  2.559106  1.880881  2.211778
#> [330]  3.941985  3.757657  2.739356  2.891561  3.833372  3.764336  4.410119
#> [337]  2.982695  3.358617  2.066861  3.217361  4.149966  3.248139  3.904442
#> [344]  3.202942  3.710096  2.616367  3.873974  3.556196  2.359810  1.778206
#> [351]  1.743924  3.681224  2.511431  1.602441  4.191620  3.133275  1.872373
#> [358]  2.960494  1.816883  4.161554  1.609256  1.938644  4.134904  4.252295
#> [365]  3.534507  1.545923  2.878705  2.991111  2.409349  3.729625  3.878804
#> [372]  4.217430  2.159739  1.758982  2.653183  2.936098  2.329611  3.010965
#> [379]  1.973956  3.041005  4.975436  2.801489  4.122167  4.367539  2.609898
#> [386]  5.324172  4.573334  2.232647  3.960597  4.612072  4.654787  2.321227
#> [393]  3.083284  3.418033  4.749064  4.218916  4.105391  5.253710  6.302634
#> [400]  3.296018  3.504770  5.486273  4.813387  5.499886  5.668599  3.301249
#> [407]  6.316743  6.434570  7.535592  7.558138  4.978010  5.399288  4.794378
#> [414]  4.199484  6.010706  5.919608  7.438291  4.649560  7.327680  4.091459
#> [421]  5.673407  8.252078  5.426048  7.113686  5.939578  7.470766  5.320169
#> [428]  9.671863  7.491923  7.611285  6.170628  7.251499  6.907666  4.946209
#> [435]  7.677674  7.318049  5.134679  7.438343  8.689301 10.686265  7.790215
#> [442]  8.815839  9.398923  7.282948  7.567456  8.932921 12.001752  8.606626
#> [449]  7.425373 10.677980  6.535192 13.183580  7.339059 10.048777 12.494801
#> [456] 11.696129  8.356571 10.675736 12.891025  7.053568 10.411247 13.022301
#> [463] 13.198277 10.212042  8.897203 10.510188 12.527013 11.385142 12.664759
#> [470] 11.606734 14.235631 13.342353 14.748073  8.002827 15.064391 16.030239
#> [477] 12.145413 11.091551 14.540216 13.083671 13.571940 15.495694 13.188869
#> [484] 10.682749 16.256287 15.916004 10.854629 12.208759 15.344434  9.971496
#> [491] 10.664625 10.472466 14.742243 16.446608 17.110087 14.000276 14.272146
#> [498] 15.274344 11.551632  9.891293 17.359107 10.211007 18.224108 14.964173
#> [505] 11.225453 17.633417 10.143595 14.997699 11.951257 17.797080 17.081233
#> [512] 16.087420 11.951490 16.407618 11.991311 14.106903 13.060375 15.144940
#> [519] 12.411273 14.287154 18.277760 18.784782 17.462105 18.216585 15.420187
#> [526] 17.115309 19.076514 16.601000 12.266311 12.949426 16.726773 18.003596
#> [533] 12.503035 10.456277 18.689089 14.260332 16.997708 16.269604 13.989070
#> [540] 13.486860 16.867366 16.212010 11.133782 17.867347 11.039068 17.807877
#> [547] 11.267316 14.556185 12.462746 16.275877 11.692510 18.547125 13.024190
#> [554] 13.343282 17.890841 17.339050 10.990336 17.945812 16.752342 13.342870
#> [561] 18.635847 12.973398 18.094461 11.745181 10.120434 11.053789 17.098785
#> [568] 11.548901 11.478629  9.971664 17.650245 11.628082 16.136262 11.891438
#> [575] 16.810855 11.844316 17.590350 14.204611 15.561839 16.136610 15.451089
#> [582] 12.495808 10.105855  9.568796 14.266195 13.744439 14.120488 11.353608
#> [589] 16.597921 14.353600 15.548023 16.353555 10.089549 13.600217 11.310100
#> [596]  8.489036 12.953503 12.717290 11.116543 14.573377 11.810542 15.308833
#> [603]  7.752375 12.734161 10.318196 14.398770 12.466239 12.266577 12.845588
#> [610] 11.146074  8.792987 10.695181  9.923158 12.085099 11.327331  9.482247
#> [617] 13.414225  8.593022  6.745593 10.352491  7.750270  7.023808 12.513624
#> [624] 11.454182 11.041533  9.409940  9.397640  6.311613  8.143279  6.639995
#> [631]  9.892033  9.676583 11.278025  6.438359 11.155803  7.438046  8.947312
#> [638]  9.827700  5.246281  7.873189  9.158848  7.129222  7.704294  5.609641
#> [645]  4.546983  9.719208  5.214226  8.413866  8.878664  8.492387  7.149478
#> [652]  8.624412  3.991082  4.410448  4.146496  7.091170  5.546326  3.795141
#> [659]  4.021758  7.824372  4.306657  6.330346  4.103446  5.601589  6.766908
#> [666]  3.555430  4.078241  4.126661  2.885578  6.377615  4.444922  5.761480
#> [673]  5.542413  4.637661  6.388964  3.884620  3.083131  5.375382  2.450311
#> [680]  5.239145  2.863118  2.312041  4.999269  5.067318  5.487846  3.371642
#> [687]  2.618766  4.037617  4.659916  2.041248  4.299781  2.559106  1.880881
#> [694]  2.211778  3.941985  3.757657  2.739356  2.891561  3.833372  3.764336
#> [701]  4.410119  2.982695  3.358617  2.066861  3.217361  4.149966  3.248139
#> [708]  3.904442  3.202942  3.710096  2.616367  3.873974  3.556196  2.359810
#> [715]  1.778206  1.743924  3.681224  2.511431  1.602441  4.191620  3.133275
#> [722]  1.872373  2.960494  1.816883  4.161554  1.609256  1.938644  4.134904
#> [729]  4.252295  3.534507
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates     sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 6.1956365127  7.463176  6.099033  4.056348
#> 3     2 2002-01-02 6.8381272161  7.483435  6.149345  4.346875
#> 6     3 2002-01-03 0.2442975054  7.505323  6.203934  1.651952
#> 7     4 2002-01-04 3.0713008842  7.528820  6.262814  2.843127
#> 9     5 2002-01-05 4.5112883241  7.553900  6.325996  3.470483
#> 11    6 2002-01-06 6.9638268326  7.580540  6.393496  4.535051
#> 14    7 2002-01-07 2.2026383446  7.608714  6.465324  2.552151
#> 15    8 2002-01-08 5.9911737633  7.638396  6.541496  4.200784
#> 17    9 2002-01-09 3.7823637708  7.669557  6.622023  3.288384
#> 19   10 2002-01-10 6.1139214663  7.702170  6.706920  4.338680
#> 22   11 2002-01-11 3.6090993364  7.736204  6.796198  3.284333
#> 23   12 2002-01-12 6.0862150858  7.771631  6.889871  4.420308
#> 25   13 2002-01-13 5.9956581620  7.808419  6.987950  4.429820
#> 27   14 2002-01-14 2.4197822879  7.846537  7.090446  2.865918
#> 30   15 2002-01-15 0.0007258395  7.885954  7.197371  1.799674
#> 32   16 2002-01-16 1.4362582173  7.926639  7.308732  2.489332
#> 33   17 2002-01-17 6.6165434348  7.968559  7.424539  4.938548
#> 35   18 2002-01-18 1.9693345916  8.011681  7.544799  2.813486
#> 37   19 2002-01-19 6.1665395207  8.055974  7.669519  4.852741
#> 40   20 2002-01-20 6.7278139368  8.101405  7.798703  5.187893
#> 41   21 2002-01-21 0.0281291772  8.147942  7.932354  1.996781
#> 43   22 2002-01-22 3.8703014392  8.195553  8.070474  3.923235
#> 45   23 2002-01-23 1.5346518091  8.244205  8.213063  2.817693
#> 48   24 2002-01-24 4.5270900032  8.293867  8.360119  4.371656
#> 49   25 2002-01-25 5.0939143272  8.344507  8.511639  4.725880
#> 51   26 2002-01-26 6.1081120968  8.396093  8.667618  5.319727
#> 53   27 2002-01-27 2.6738708224  8.448595  8.828046  3.603993
#> 56   28 2002-01-28 6.2492199291  8.501983  8.992915  5.553264
#> 58   29 2002-01-29 5.9070919997  8.556225  9.162213  5.453282
#> 59   30 2002-01-30 5.1083284605  8.611293  9.335925  5.103076
#> 61   31 2002-01-31 0.2689048359  8.667157  9.514035  2.526099
#> 63   32 2002-02-01 2.3367632220  8.723789  9.696523  3.722791
#> 66   33 2002-02-02 5.2449085654  8.781160  9.883368  5.422466
#> 67   34 2002-02-03 6.3921060748  8.839243 10.074546  6.161345
#> 69   35 2002-02-04 1.4303034407  8.898011 10.270030  3.392931
#> 71   36 2002-02-05 3.3022892803  8.957438 10.469790  4.547367
#> 74   37 2002-02-06 2.6962817025  9.017498 10.673795  4.264211
#> 75   38 2002-02-07 2.7492635022  9.078165 10.882010  4.368276
#> 77   39 2002-02-08 1.5283690128  9.139416 11.094398  3.701248
#> 79   40 2002-02-09 0.0504625840  9.201226 11.310918  2.858746
#> 82   41 2002-02-10 2.5459850021  9.263572 11.531528  4.467535
#> 83   42 2002-02-11 3.6827471091  9.326432 11.756182  5.260139
#> 86   43 2002-02-12 5.9091437536  9.389784 11.984831  6.767332
#> 87   44 2002-02-13 0.9519381954  9.453606 12.217425  3.669478
#> 89   45 2002-02-14 2.6352713159  9.517878 12.453910  4.837572
#> 91   46 2002-02-15 1.8431605394  9.582579 12.694229  4.394393
#> 94   47 2002-02-16 3.4849707878  9.647690 12.938322  5.571393
#> 95   48 2002-02-17 0.1366767546  9.713192 13.186128  3.389305
#> 97   49 2002-02-18 0.6102866689  9.779067 13.437583  3.778698
#> 99   50 2002-02-19 6.9717386514  9.845297 13.692618  8.271223
#> 102  51 2002-02-20 6.5949254383  9.911864 13.951164  8.129041
#> 103  52 2002-02-21 0.8706141578  9.978754 14.213148  4.173313
#> 105  53 2002-02-22 0.8063251157 10.045948 14.478496  4.200673
#> 107  54 2002-02-23 2.5668167255 10.113433 14.747130  5.558213
#> 110  55 2002-02-24 3.2972802492 10.181192 15.018971  6.186764
#> 112  56 2002-02-25 5.4892071187 10.249211 15.293936  7.918998
#> 113  57 2002-02-26 3.4327496893 10.317477 15.571940  6.483472
#> 115  58 2002-02-27 0.7136566842 10.385975 15.852898  4.507879
#> 117  59 2002-02-28 6.6042180702 10.454693 16.136720  9.130954
#> 120  60 2002-03-01 4.9429765164 10.523618 16.423316  7.962870
#> 121  61 2002-03-02 3.5158496560 10.592738 16.712592  6.951697
#> 123  62 2002-03-03 3.7095434354 10.662040 17.004454  7.209213
#> 125  63 2002-03-04 1.8089844391 10.731513 17.298805  5.782709
#> 128  64 2002-03-05 4.9822739193 10.801147 17.595546  8.457059
#> 129  65 2002-03-06 6.7792378489 10.870930 17.894578 10.053278
#> 131  66 2002-03-07 4.6275445281 10.940852 18.195799  8.396999
#> 133  67 2002-03-08 5.5737541325 11.010902 18.499104  9.306929
#> 136  68 2002-03-09 0.5124639245 11.081072 18.804391  5.135919
#> 138  69 2002-03-10 6.1932001910 11.151350 19.111552 10.084943
#> 139  70 2002-03-11 5.7869026377 11.221729 19.420480  9.862567
#> 141  71 2002-03-12 5.0803477154 11.292198 19.731067  9.371260
#> 143  72 2002-03-13 2.5713031518 11.362749 20.043203  7.278612
#> 146  73 2002-03-14 3.6639989160 11.433374 20.356778  8.351014
#> 147  74 2002-03-15 3.2891523887 11.504062 20.671679  8.123063
#> 149  75 2002-03-16 0.6899646998 11.574807 20.987796  5.872481
#> 151  76 2002-03-17 3.6122443716 11.645600 21.305014  8.630459
#> 154  77 2002-03-18 3.1481367925 11.716433 21.623220  8.310821
#> 155  78 2002-03-19 3.8462308343 11.787297 21.942300  9.065495
#> 157  79 2002-03-20 4.5465433290 11.858185 22.262140  9.833295
#> 159  80 2002-03-21 1.1132437827 11.929088 22.582624  6.699381
#> 162  81 2002-03-22 2.0562555443 12.000000 22.903638  7.688232
#> 163  82 2002-03-23 6.2389940687 12.070912 23.225065 11.808342
#> 166  83 2002-03-24 4.1165503818 12.141815 23.546791  9.878340
#> 167  84 2002-03-25 4.4326209428 12.212703 23.868700 10.298767
#> 169  85 2002-03-26 1.8165663346 12.283567 24.190677  7.836399
#> 171  86 2002-03-27 2.2274890896 12.354400 24.512606  8.337954
#> 174  87 2002-03-28 2.3747771671 12.425193 24.834373  8.581840
#> 175  88 2002-03-29 5.3253279696 12.495938 25.155863 11.649236
#> 177  89 2002-03-30 0.0761072491 12.566626 25.476961  6.446388
#> 179  90 2002-03-31 0.4419613727 12.637251 25.797553  6.900496
#> 182  91 2002-04-01 1.8468925457 12.707802 26.117528  8.427282
#> 183  92 2002-04-02 4.1568561199 12.778271 26.436771 10.909221
#> 185  93 2002-04-03 2.0829202044 12.848650 26.755170  8.857460
#> 187  94 2002-04-04 0.6623594053 12.918928 27.072615  7.462166
#> 190  95 2002-04-05 5.3287875215 12.989098 27.388994 12.465427
#> 192  96 2002-04-06 5.4073379326 13.059148 27.704199 12.661721
#> 193  97 2002-04-07 2.8139327522 13.129070 28.018120 10.007070
#> 195  98 2002-04-08 0.4135806712 13.198853 28.330649  7.526527
#> 197  99 2002-04-09 5.2517351902 13.268487 28.641680 12.828682
#> 200 100 2002-04-10 3.6982564875 13.337960 28.951106 11.251457
#> 201 101 2002-04-11 1.9306358229 13.407262 29.258822  9.421330
#> 203 102 2002-04-12 1.8286628504 13.476382 29.564726  9.397058
#> 205 103 2002-04-13 3.8068181823 13.545307 29.868714 11.664380
#> 208 104 2002-04-14 1.3930054551 13.614025 30.170685  9.086224
#> 209 105 2002-04-15 4.5153834075 13.682523 30.470539 12.645442
#> 211 106 2002-04-16 5.2516789224 13.750789 30.768176 13.567510
#> 213 107 2002-04-17 1.8968448413 13.818808 31.063500  9.897847
#> 216 108 2002-04-18 6.1293533649 13.886567 31.356413 14.759263
#> 218 109 2002-04-19 5.9622022181 13.954052 31.646820 14.672635
#> 219 110 2002-04-20 0.3209451749 14.021246 31.934628  8.349147
#> 221 111 2002-04-21 2.6673980125 14.088136 32.219743 11.105122
#> 223 112 2002-04-22 1.2831843798 14.154703 32.502075  9.598745
#> 226 113 2002-04-23 6.4525974893 14.220933 32.781534 15.632520
#> 227 114 2002-04-24 6.3433079422 14.286808 33.058030 15.603350
#> 229 115 2002-04-25 1.3506032247 14.352310 33.331477  9.901174
#> 231 116 2002-04-26 6.0506269371 14.417421 33.601789 15.451358
#> 234 117 2002-04-27 0.0929082131 14.482122 33.868881  8.575861
#> 235 118 2002-04-28 6.1996223773 14.546394 34.132670 15.806780
#> 237 119 2002-04-29 2.7407947497 14.610216 34.393074 11.824243
#> 239 120 2002-04-30 0.5332341893 14.673568 34.650012  9.292090
#> 242 121 2002-05-01 4.5274691340 14.736428 34.903405 14.087534
#> 243 122 2002-05-02 6.4165772784 14.798774 35.153175 16.409299
#> 245 123 2002-05-03 5.8415603414 14.860584 35.399245 15.807372
#> 248 124 2002-05-04 3.5087294918 14.921835 35.641540 13.100772
#> 249 125 2002-05-05 4.2150133890 14.982502 35.879987 14.017038
#> 251 126 2002-05-06 3.2827132393 15.042562 36.114511 12.969232
#> 253 127 2002-05-07 6.2046196673 15.101989 36.345041 16.552401
#> 256 128 2002-05-08 4.6938394464 15.160757 36.571508 14.804230
#> 257 129 2002-05-09 2.7564440363 15.218840 36.793843 12.530520
#> 259 130 2002-05-10 3.8590347059 15.276211 37.011977 13.927926
#> 261 131 2002-05-11 1.6974487305 15.332843 37.225844 11.367036
#> 264 132 2002-05-12 1.4224618033 15.388707 37.435378 11.089022
#> 265 133 2002-05-13 0.4783919237 15.443775 37.640517  9.993112
#> 267 134 2002-05-14 2.1509161601 15.498017 37.841197 12.086224
#> 269 135 2002-05-15 6.9518266600 15.551405 38.037357 18.011115
#> 272 136 2002-05-16 0.8141642399 15.603907 38.228936 10.554569
#> 274 137 2002-05-17 5.0079615365 15.655493 38.415875 15.748305
#> 275 138 2002-05-18 1.5827611203 15.706133 38.598118 11.594362
#> 277 139 2002-05-19 0.9334398333 15.755795 38.775606 10.842517
#> 279 140 2002-05-20 6.4909723308 15.804447 38.948286 17.735208
#> 282 141 2002-05-21 6.2588585999 15.852058 39.116103 17.501120
#> 283 142 2002-05-22 1.4256507456 15.898595 39.279005 11.580855
#> 285 143 2002-05-23 1.7988804334 15.944026 39.436940 12.083966
#> 287 144 2002-05-24 4.2980392252 15.988319 39.589859 15.218811
#> 290 145 2002-05-25 3.0904943133 16.031441 39.737713 13.764688
#> 291 146 2002-05-26 2.2091627968 16.073361 39.880455 12.710748
#> 293 147 2002-05-27 0.7091963268 16.114046 40.018039 10.885128
#> 295 148 2002-05-28 1.9085233700 16.153463 40.150421 12.409481
#> 298 149 2002-05-29 4.5758957798 16.191581 40.277558 15.760801
#> 300 150 2002-05-30 6.4950907156 16.228369 40.399408 18.184393
#> 301 151 2002-05-31 0.1863726901 16.263796 40.515933 10.361127
#> 303 152 2002-06-01 3.9164383798 16.297830 40.627093 15.038205
#> 305 153 2002-06-02 5.9792155649 16.330443 40.732853 17.640161
#> 308 154 2002-06-03 0.5311813466 16.361604 40.833176 10.871121
#> 309 155 2002-06-04 1.4408019935 16.391286 40.928031 12.030805
#> 311 156 2002-06-05 3.4712275269 16.419460 41.017385 14.590076
#> 313 157 2002-06-06 1.8329986085 16.446100 41.101209 12.565768
#> 316 158 2002-06-07 2.9117301547 16.471180 41.179475 13.934666
#> 317 159 2002-06-08 3.0740499967 16.494677 41.252156 14.157043
#> 319 160 2002-06-09 1.2766454353 16.516565 41.319229 11.926689
#> 321 161 2002-06-10 0.5482978623 16.536824 41.380671 11.031180
#> 324 162 2002-06-11 0.1243896922 16.555433 41.436461 10.514782
#> 325 163 2002-06-12 5.0972598132 16.572373 41.486582 16.751778
#> 328 164 2002-06-13 6.7959151736 16.587625 41.531016 18.890339
#> 329 165 2002-06-14 3.7465231165 16.601175 41.569750 15.083130
#> 331 166 2002-06-15 1.1654353749 16.613007 41.602770 11.859951
#> 333 167 2002-06-16 0.4072295546 16.623108 41.630068 10.917439
#> 336 168 2002-06-17 3.2139731331 16.631469 41.651634 14.437425
#> 337 169 2002-06-18 4.3595917500 16.638080 41.667464 15.875823
#> 339 170 2002-06-19 4.1637344097 16.642934 41.677552 15.632839
#> 341 171 2002-06-20 5.2011276619 16.646025 41.681897 16.932325
#> 344 172 2002-06-21 3.6699609405 16.647351 41.680500 15.014424
#> 345 173 2002-06-22 3.4777767523 16.646909 41.673363 14.771421
#> 347 174 2002-06-23 0.1030316185 16.644700 41.660491 10.544063
#> 349 175 2002-06-24 6.7054338716 16.640727 41.641891 18.800339
#> 352 176 2002-06-25 3.1716246458 16.634994 41.617572 14.371792
#> 354 177 2002-06-26 6.8707229439 16.627507 41.587545 18.989169
#> 355 178 2002-06-27 3.8129361621 16.618275 41.551823 15.154830
#> 357 179 2002-06-28 2.5202968079 16.607306 41.510423 13.527382
#> 359 180 2002-06-29 3.5220227316 16.594614 41.463360 14.765909
#> 362 181 2002-06-30 0.5224354262 16.580211 41.410655 11.005080
#> 363 182 2002-07-01 6.7885432325 16.564113 41.352328 18.811885
#> 365 183 2002-07-02 4.5395091963 16.546336 41.288405 15.985865
#> 367 184 2002-07-03 6.8919582665 16.526900 41.218909 18.899170
#> 370 185 2002-07-04 3.7529327988 16.505823 41.143868 14.963412
#> 371 186 2002-07-05 5.3274158353 16.483128 41.063311 16.901745
#> 373 187 2002-07-06 3.9507000893 16.458837 40.977269 15.162311
#> 375 188 2002-07-07 5.4491801246 16.432973 40.885775 17.000313
#> 378 189 2002-07-08 4.4734269199 16.405563 40.788864 15.758317
#> 380 190 2002-07-09 2.6527251494 16.376632 40.686570 13.466895
#> 381 191 2002-07-10 2.5167146940 16.346207 40.578932 13.268565
#> 383 192 2002-07-11 5.9096199137 16.314316 40.465990 17.445601
#> 385 193 2002-07-12 3.1972442656 16.280989 40.347784 14.048675
#> 388 194 2002-07-13 1.2532400342 16.246255 40.224356 11.607547
#> 389 195 2002-07-14 6.7634643705 16.210144 40.095751 18.388644
#> 391 196 2002-07-15 4.9938836617 16.172686 39.962013 16.160339
#> 393 197 2002-07-16 0.1742022182 16.133915 39.823189 10.170788
#> 396 198 2002-07-17 4.3851954595 16.093860 39.679327 15.325670
#> 397 199 2002-07-18 5.7954492546 16.052554 39.530476 17.018458
#> 399 200 2002-07-19 4.5014832215 16.010029 39.376687 15.379874
#> 401 201 2002-07-20 4.9593898323 15.966317 39.218012 15.895370
#> 404 202 2002-07-21 6.0950900745 15.921451 39.054503 17.239098
#> 405 203 2002-07-22 1.9726880719 15.875462 38.886214 12.137558
#> 408 204 2002-07-23 1.2125738936 15.828384 38.713200 11.161162
#> 409 205 2002-07-24 0.7676901331 15.780249 38.535519 10.571233
#> 411 206 2002-07-25 4.1035553410 15.731088 38.353226 14.590649
#> 413 207 2002-07-26 5.9550492915 15.680933 38.166382 16.788699
#> 416 208 2002-07-27 6.4506832347 15.629816 37.975044 17.330225
#> 417 209 2002-07-28 4.7296333136 15.577768 37.779274 15.179983
#> 419 210 2002-07-29 2.4541171370 15.524820 37.579134 12.364982
#> 421 211 2002-07-30 0.0606065234 15.471001 37.374685  9.416878
#> 424 212 2002-07-31 3.0162915480 15.416342 37.165992 12.927363
#> 425 213 2002-08-01 5.0745456801 15.360873 36.953119 15.342109
#> 427 214 2002-08-02 2.8857947094 15.304621 36.736132 12.647462
#> 429 215 2002-08-03 3.8560367664 15.247616 36.515097 13.746006
#> 432 216 2002-08-04 1.0945239908 15.189886 36.290082 10.379981
#> 434 217 2002-08-05 0.3010090524 15.131457 36.061156  9.373970
#> 435 218 2002-08-06 4.7925087914 15.072356 35.828388 14.653216
#> 437 219 2002-08-07 3.6008174797 15.012610 35.591848 13.166365
#> 439 220 2002-08-08 6.6198149386 14.952243 35.351609 16.663521
#> 442 221 2002-08-09 5.9016785952 14.891281 35.107744 15.733847
#> 443 222 2002-08-10 0.4541816534 14.829748 34.860325  9.248904
#> 445 223 2002-08-11 5.3347764020 14.767666 34.609429 14.903635
#> 447 224 2002-08-12 0.5486897659 14.705060 34.355129  9.229729
#> 450 225 2002-08-13 0.9345536388 14.641952 34.097505  9.612549
#> 451 226 2002-08-14 5.6678734070 14.578362 33.836632 15.036774
#> 453 227 2002-08-15 3.9156826755 14.514313 33.572591 12.921768
#> 455 228 2002-08-16 5.1361078029 14.449824 33.305462 14.245484
#> 458 229 2002-08-17 2.4478683304 14.384916 33.035326 11.069627
#> 460 230 2002-08-18 6.8786806664 14.319607 32.762264 16.059537
#> 461 231 2002-08-19 2.8619771593 14.253916 32.486361 11.382983
#> 463 232 2002-08-20 2.5261158552 14.187862 32.207701 10.919179
#> 465 233 2002-08-21 4.5675702749 14.121461 31.926370 13.144867
#> 468 234 2002-08-22 1.0049145808 14.054730 31.642455  9.041833
#> 469 235 2002-08-23 4.7655903073 13.987686 31.356043 13.180496
#> 471 236 2002-08-24 6.0150012553 13.920345 31.067223 14.478902
#> 473 237 2002-08-25 5.8704099276 13.852721 30.776085 14.215059
#> 476 238 2002-08-26 6.1691904743 13.784830 30.482721 14.441719
#> 477 239 2002-08-27 1.6638841627 13.716686 30.187223  9.377716
#> 479 240 2002-08-28 4.0504159203 13.648302 29.889684 11.907612
#> 481 241 2002-08-29 6.2853175399 13.579692 29.590199 14.245415
#> 484 242 2002-08-30 2.4110032138 13.510869 29.288864  9.935502
#> 485 243 2002-08-31 5.3260076088 13.441846 28.985774 12.988901
#> 487 244 2002-09-01 5.1177255162 13.372633 28.681027 12.658391
#> 490 245 2002-09-02 5.9670266353 13.303244 28.374722 13.457268
#> 491 246 2002-09-03 6.8307262855 13.233689 28.066958 14.260287
#> 493 247 2002-09-04 0.7913325059 13.163980 27.757836  7.773769
#> 495 248 2002-09-05 6.7918359581 13.094126 27.447457 13.980271
#> 498 249 2002-09-06 4.5365794727 13.024138 27.135924 11.509985
#> 499 250 2002-09-07 0.3925113028 12.954027 26.823338  7.112213
#> 501 251 2002-09-08 3.3119596706 12.883802 26.509805 10.034807
#> 503 252 2002-09-09 2.0623750919 12.813472 26.195428  8.656982
#> 506 253 2002-09-10 4.2720973536 12.743047 25.880314 10.808257
#> 507 254 2002-09-11 0.8477621619 12.672536 25.564567  7.246246
#> 509 255 2002-09-12 4.4059155553 12.601947 25.248295 10.725751
#> 511 256 2002-09-13 4.9836871200 12.531289 24.931604 11.190544
#> 514 257 2002-09-14 4.2845795064 12.460572 24.614602 10.385528
#> 516 258 2002-09-15 0.2410046314 12.389802 24.297397  6.310664
#> 517 259 2002-09-16 4.6612355092 12.318988 23.980097 10.531796
#> 519 260 2002-09-17 4.5707709356 12.248139 23.662812 10.330957
#> 521 261 2002-09-18 3.1630090780 12.177262 23.345649  8.868395
#> 524 262 2002-09-19 3.6176287131 12.106365 23.028718  9.197905
#> 525 263 2002-09-20 4.7362755802 12.035456 22.712128 10.146949
#> 527 264 2002-09-21 6.1665316585 11.964544 22.395989 11.370449
#> 529 265 2002-09-22 5.1186040824 11.893635 22.080409 10.271420
#> 532 266 2002-09-23 2.7903387067 11.822738 21.765497  8.009862
#> 533 267 2002-09-24 3.0136445761 11.751861 21.451363  8.113332
#> 535 268 2002-09-25 1.0178514274 11.681012 21.138114  6.205487
#> 537 269 2002-09-26 3.1578423053 11.610198 20.825859  8.038663
#> 540 270 2002-09-27 4.0514704413 11.539428 20.514705  8.730013
#> 542 271 2002-09-28 0.4929198441 11.468711 20.204761  5.485386
#> 543 272 2002-09-29 5.1962832615 11.398053 19.896131  9.509278
#> 545 273 2002-09-30 3.8445449444 11.327464 19.588923  8.221474
#> 547 274 2002-10-01 4.4418099055 11.256953 19.283241  8.625236
#> 550 275 2002-10-02 0.9564000864 11.186528 18.979189  5.556117
#> 551 276 2002-10-03 6.5116344769 11.116198 18.676871 10.139476
#> 553 277 2002-10-04 4.5578304774 11.045973 18.376389  8.385364
#> 555 278 2002-10-05 3.7009649067 10.975862 18.077844  7.567307
#> 558 279 2002-10-06 1.5758077204 10.905874 17.781337  5.729961
#> 559 280 2002-10-07 6.8828680869 10.836020 17.486965  9.925462
#> 561 281 2002-10-08 0.7633414995 10.766311 17.194826  4.908271
#> 563 282 2002-10-09 0.4896186206 10.696756 16.905016  4.613147
#> 566 283 2002-10-10 5.4435185483 10.627367 16.617629  8.410324
#> 567 284 2002-10-11 4.6082556229 10.558154 16.332759  7.647521
#> 570 285 2002-10-12 1.8144192055 10.489131 16.050495  5.400838
#> 571 286 2002-10-13 6.4283649134 10.420308 15.770928  8.807333
#> 573 287 2002-10-14 1.0409724275 10.351698 15.494145  4.652586
#> 575 288 2002-10-15 4.4847633946 10.283314 15.220231  7.123985
#> 578 289 2002-10-16 1.4086689684 10.215170 14.949271  4.768068
#> 579 290 2002-10-17 5.2053492058 10.147279 14.681345  7.435953
#> 581 291 2002-10-18 5.9334619991 10.079655 14.416534  7.847332
#> 583 292 2002-10-19 5.9754511106 10.012314 14.154915  7.762627
#> 586 293 2002-10-20 5.6238880090  9.945270 13.896562  7.403280
#> 587 294 2002-10-21 0.1671373537  9.878539 13.641550  3.525790
#> 589 295 2002-10-22 5.5676440035  9.812138 13.389948  7.146377
#> 591 296 2002-10-23 6.2058535737  9.746084 13.141825  7.469508
#> 594 297 2002-10-24 4.2002613691  9.680393 12.897247  6.022329
#> 596 298 2002-10-25 3.0295171652  9.615084 12.656277  5.157937
#> 597 299 2002-10-26 6.4419306223  9.550176 12.418977  7.293263
#> 599 300 2002-10-27 4.3121009958  9.485687 12.185404  5.816034
#> 601 301 2002-10-28 3.0500271111  9.421638 11.955615  4.924074
#> 604 302 2002-10-29 6.2811308708  9.358048 11.729663  6.868896
#> 605 303 2002-10-30 5.3830946553  9.294940 11.507600  6.209170
#> 607 304 2002-10-31 0.8968103081  9.232334 11.289473  3.370687
#> 609 305 2002-11-01 3.7704318434  9.170252 11.075328  5.045693
#> 612 306 2002-11-02 0.9031135815  9.108719 10.865209  3.254936
#> 613 307 2002-11-03 3.5767541877  9.047757 10.659156  4.771675
#> 615 308 2002-11-04 3.8065488376  8.987390 10.457208  4.828842
#> 617 309 2002-11-05 1.6924326292  8.927644 10.259398  3.537298
#> 620 310 2002-11-06 5.6220088068  8.868543 10.065762  5.706919
#> 622 311 2002-11-07 6.6353960712  8.810114  9.876328  6.188293
#> 623 312 2002-11-08 2.8515294176  8.752384  9.691126  4.001467
#> 625 313 2002-11-09 3.6032119396  8.695379  9.510180  4.347971
#> 627 314 2002-11-10 6.8707929526  8.639127  9.333513  6.044901
#> 630 315 2002-11-11 2.4793070969  8.583658  9.161147  3.613342
#> 631 316 2002-11-12 3.1908810346  8.528999  8.993100  3.930531
#> 633 317 2002-11-13 1.5401729443  8.475180  8.829388  3.009618
#> 635 318 2002-11-14 0.5728311650  8.422232  8.670025  2.462349
#> 638 319 2002-11-15 2.5846524800  8.370184  8.515023  3.443445
#> 639 320 2002-11-16 0.8866660867  8.319067  8.364392  2.536846
#> 641 321 2002-11-17 5.1791075380  8.268912  8.218139  4.628188
#> 643 322 2002-11-18 6.6910640008  8.219751  8.076271  5.306202
#> 646 323 2002-11-19 4.6807215177  8.171616  7.938791  4.258377
#> 647 324 2002-11-20 3.2766254421  8.124538  7.805702  3.525445
#> 650 325 2002-11-21 3.7673471260  8.078549  7.677004  3.709297
#> 651 326 2002-11-22 2.3166759131  8.033683  7.552698  2.977162
#> 653 327 2002-11-23 4.4007747553  7.989971  7.432780  3.905136
#> 655 328 2002-11-24 3.1678673008  7.947446  7.317248  3.287646
#> 658 329 2002-11-25 6.2161062430  7.906140  7.206095  4.634376
#> 659 330 2002-11-26 5.4894127378  7.866085  7.099317  4.251988
#> 661 331 2002-11-27 0.9664168563  7.827314  6.996907  2.181171
#> 663 332 2002-11-28 1.9691914173  7.789856  6.898857  2.596692
#> 666 333 2002-11-29 4.9486254861  7.753745  6.805157  3.872896
#> 667 334 2002-11-30 3.0953084414  7.719011  6.715800  3.025461
#> 669 335 2002-12-01 4.7438896613  7.685684  6.630775  3.704074
#> 671 336 2002-12-02 4.9050048993  7.653793  6.550072  3.736355
#> 674 337 2002-12-03 0.4788449355  7.623368  6.473680  1.821735
#> 676 338 2002-12-04 6.7095246441  7.594437  6.401590  4.428232
#> 677 339 2002-12-05 1.7113846468  7.567027  6.333789  2.299683
#> 679 340 2002-12-06 6.4839833423  7.541163  6.270268  4.263193
#> 681 341 2002-12-07 3.5192295376  7.516872  6.211016  3.006683
#> 684 342 2002-12-08 6.5486199944  7.494177  6.156023  4.228658
#> 685 343 2002-12-09 3.0715974274  7.473100  6.105278  2.781017
#> 687 344 2002-12-10 2.4534057572  7.453664  6.058772  2.511829
#> 689 345 2002-12-11 3.3355286187  7.435887  6.016497  2.853539
#> 692 346 2002-12-12 6.7631166757  7.419789  5.978443  4.219277
#> 693 347 2002-12-13 2.0801334889  7.405386  5.944603  2.321054
#> 695 348 2002-12-14 2.0345381584  7.392694  5.914970  2.292670
#> 697 349 2002-12-15 6.0180557331  7.381725  5.889539  3.873149
#> 700 350 2002-12-16 1.3531886679  7.372493  5.868303  2.005626
#> 702 351 2002-12-17 5.6637585997  7.365006  5.851259  3.712652
#> 703 352 2002-12-18 2.5584055870  7.359273  5.838404  2.474444
#> 705 353 2002-12-19 1.7688975721  7.355300  5.829736  2.158439
#> 707 354 2002-12-20 5.9925670966  7.353091  5.825253  3.830024
#> 710 355 2002-12-21 0.9101705870  7.352649  5.824956  1.816769
#> 711 356 2002-12-22 5.5433536230  7.353975  5.828845  3.654074
#> 713 357 2002-12-23 2.4913105716  7.357066  5.836924  2.447505
#> 715 358 2002-12-24 6.7504068639  7.361920  5.849194  4.143966
#> 718 359 2002-12-25 0.8307824642  7.368531  5.865662  1.797084
#> 719 360 2002-12-26 3.2071245466  7.376892  5.886331  2.751132
#> 721 361 2002-12-27 3.6655530566  7.386993  5.911209  2.944424
#> 723 362 2002-12-28 5.8594320805  7.398825  5.940303  3.837260
#> 726 363 2002-12-29 0.2345282619  7.412375  5.973622  1.587909
#> 727 364 2002-12-30 0.3032990755  7.427627  6.011174  1.625523
#> 729 365 2002-12-31 5.5447206120  7.444567  6.052971  3.767372
#> 2     1 2003-01-01 6.1956365127  7.463176  6.099033  4.056348
#> 4     2 2003-01-02 6.8381272161  7.483435  6.149345  4.346875
#> 5     3 2003-01-03 0.2442975054  7.505323  6.203934  1.651952
#> 8     4 2003-01-04 3.0713008842  7.528820  6.262814  2.843127
#> 10    5 2003-01-05 4.5112883241  7.553900  6.325996  3.470483
#> 12    6 2003-01-06 6.9638268326  7.580540  6.393496  4.535051
#> 13    7 2003-01-07 2.2026383446  7.608714  6.465324  2.552151
#> 16    8 2003-01-08 5.9911737633  7.638396  6.541496  4.200784
#> 18    9 2003-01-09 3.7823637708  7.669557  6.622023  3.288384
#> 20   10 2003-01-10 6.1139214663  7.702170  6.706920  4.338680
#> 21   11 2003-01-11 3.6090993364  7.736204  6.796198  3.284333
#> 24   12 2003-01-12 6.0862150858  7.771631  6.889871  4.420308
#> 26   13 2003-01-13 5.9956581620  7.808419  6.987950  4.429820
#> 28   14 2003-01-14 2.4197822879  7.846537  7.090446  2.865918
#> 29   15 2003-01-15 0.0007258395  7.885954  7.197371  1.799674
#> 31   16 2003-01-16 1.4362582173  7.926639  7.308732  2.489332
#> 34   17 2003-01-17 6.6165434348  7.968559  7.424539  4.938548
#> 36   18 2003-01-18 1.9693345916  8.011681  7.544799  2.813486
#> 38   19 2003-01-19 6.1665395207  8.055974  7.669519  4.852741
#> 39   20 2003-01-20 6.7278139368  8.101405  7.798703  5.187893
#> 42   21 2003-01-21 0.0281291772  8.147942  7.932354  1.996781
#> 44   22 2003-01-22 3.8703014392  8.195553  8.070474  3.923235
#> 46   23 2003-01-23 1.5346518091  8.244205  8.213063  2.817693
#> 47   24 2003-01-24 4.5270900032  8.293867  8.360119  4.371656
#> 50   25 2003-01-25 5.0939143272  8.344507  8.511639  4.725880
#> 52   26 2003-01-26 6.1081120968  8.396093  8.667618  5.319727
#> 54   27 2003-01-27 2.6738708224  8.448595  8.828046  3.603993
#> 55   28 2003-01-28 6.2492199291  8.501983  8.992915  5.553264
#> 57   29 2003-01-29 5.9070919997  8.556225  9.162213  5.453282
#> 60   30 2003-01-30 5.1083284605  8.611293  9.335925  5.103076
#> 62   31 2003-01-31 0.2689048359  8.667157  9.514035  2.526099
#> 64   32 2003-02-01 2.3367632220  8.723789  9.696523  3.722791
#> 65   33 2003-02-02 5.2449085654  8.781160  9.883368  5.422466
#> 68   34 2003-02-03 6.3921060748  8.839243 10.074546  6.161345
#> 70   35 2003-02-04 1.4303034407  8.898011 10.270030  3.392931
#> 72   36 2003-02-05 3.3022892803  8.957438 10.469790  4.547367
#> 73   37 2003-02-06 2.6962817025  9.017498 10.673795  4.264211
#> 76   38 2003-02-07 2.7492635022  9.078165 10.882010  4.368276
#> 78   39 2003-02-08 1.5283690128  9.139416 11.094398  3.701248
#> 80   40 2003-02-09 0.0504625840  9.201226 11.310918  2.858746
#> 81   41 2003-02-10 2.5459850021  9.263572 11.531528  4.467535
#> 84   42 2003-02-11 3.6827471091  9.326432 11.756182  5.260139
#> 85   43 2003-02-12 5.9091437536  9.389784 11.984831  6.767332
#> 88   44 2003-02-13 0.9519381954  9.453606 12.217425  3.669478
#> 90   45 2003-02-14 2.6352713159  9.517878 12.453910  4.837572
#> 92   46 2003-02-15 1.8431605394  9.582579 12.694229  4.394393
#> 93   47 2003-02-16 3.4849707878  9.647690 12.938322  5.571393
#> 96   48 2003-02-17 0.1366767546  9.713192 13.186128  3.389305
#> 98   49 2003-02-18 0.6102866689  9.779067 13.437583  3.778698
#> 100  50 2003-02-19 6.9717386514  9.845297 13.692618  8.271223
#> 101  51 2003-02-20 6.5949254383  9.911864 13.951164  8.129041
#> 104  52 2003-02-21 0.8706141578  9.978754 14.213148  4.173313
#> 106  53 2003-02-22 0.8063251157 10.045948 14.478496  4.200673
#> 108  54 2003-02-23 2.5668167255 10.113433 14.747130  5.558213
#> 109  55 2003-02-24 3.2972802492 10.181192 15.018971  6.186764
#> 111  56 2003-02-25 5.4892071187 10.249211 15.293936  7.918998
#> 114  57 2003-02-26 3.4327496893 10.317477 15.571940  6.483472
#> 116  58 2003-02-27 0.7136566842 10.385975 15.852898  4.507879
#> 118  59 2003-02-28 6.6042180702 10.454693 16.136720  9.130954
#> 119  60 2003-03-01 4.9429765164 10.523618 16.423316  7.962870
#> 122  61 2003-03-02 3.5158496560 10.592738 16.712592  6.951697
#> 124  62 2003-03-03 3.7095434354 10.662040 17.004454  7.209213
#> 126  63 2003-03-04 1.8089844391 10.731513 17.298805  5.782709
#> 127  64 2003-03-05 4.9822739193 10.801147 17.595546  8.457059
#> 130  65 2003-03-06 6.7792378489 10.870930 17.894578 10.053278
#> 132  66 2003-03-07 4.6275445281 10.940852 18.195799  8.396999
#> 134  67 2003-03-08 5.5737541325 11.010902 18.499104  9.306929
#> 135  68 2003-03-09 0.5124639245 11.081072 18.804391  5.135919
#> 137  69 2003-03-10 6.1932001910 11.151350 19.111552 10.084943
#> 140  70 2003-03-11 5.7869026377 11.221729 19.420480  9.862567
#> 142  71 2003-03-12 5.0803477154 11.292198 19.731067  9.371260
#> 144  72 2003-03-13 2.5713031518 11.362749 20.043203  7.278612
#> 145  73 2003-03-14 3.6639989160 11.433374 20.356778  8.351014
#> 148  74 2003-03-15 3.2891523887 11.504062 20.671679  8.123063
#> 150  75 2003-03-16 0.6899646998 11.574807 20.987796  5.872481
#> 152  76 2003-03-17 3.6122443716 11.645600 21.305014  8.630459
#> 153  77 2003-03-18 3.1481367925 11.716433 21.623220  8.310821
#> 156  78 2003-03-19 3.8462308343 11.787297 21.942300  9.065495
#> 158  79 2003-03-20 4.5465433290 11.858185 22.262140  9.833295
#> 160  80 2003-03-21 1.1132437827 11.929088 22.582624  6.699381
#> 161  81 2003-03-22 2.0562555443 12.000000 22.903638  7.688232
#> 164  82 2003-03-23 6.2389940687 12.070912 23.225065 11.808342
#> 165  83 2003-03-24 4.1165503818 12.141815 23.546791  9.878340
#> 168  84 2003-03-25 4.4326209428 12.212703 23.868700 10.298767
#> 170  85 2003-03-26 1.8165663346 12.283567 24.190677  7.836399
#> 172  86 2003-03-27 2.2274890896 12.354400 24.512606  8.337954
#> 173  87 2003-03-28 2.3747771671 12.425193 24.834373  8.581840
#> 176  88 2003-03-29 5.3253279696 12.495938 25.155863 11.649236
#> 178  89 2003-03-30 0.0761072491 12.566626 25.476961  6.446388
#> 180  90 2003-03-31 0.4419613727 12.637251 25.797553  6.900496
#> 181  91 2003-04-01 1.8468925457 12.707802 26.117528  8.427282
#> 184  92 2003-04-02 4.1568561199 12.778271 26.436771 10.909221
#> 186  93 2003-04-03 2.0829202044 12.848650 26.755170  8.857460
#> 188  94 2003-04-04 0.6623594053 12.918928 27.072615  7.462166
#> 189  95 2003-04-05 5.3287875215 12.989098 27.388994 12.465427
#> 191  96 2003-04-06 5.4073379326 13.059148 27.704199 12.661721
#> 194  97 2003-04-07 2.8139327522 13.129070 28.018120 10.007070
#> 196  98 2003-04-08 0.4135806712 13.198853 28.330649  7.526527
#> 198  99 2003-04-09 5.2517351902 13.268487 28.641680 12.828682
#> 199 100 2003-04-10 3.6982564875 13.337960 28.951106 11.251457
#> 202 101 2003-04-11 1.9306358229 13.407262 29.258822  9.421330
#> 204 102 2003-04-12 1.8286628504 13.476382 29.564726  9.397058
#> 206 103 2003-04-13 3.8068181823 13.545307 29.868714 11.664380
#> 207 104 2003-04-14 1.3930054551 13.614025 30.170685  9.086224
#> 210 105 2003-04-15 4.5153834075 13.682523 30.470539 12.645442
#> 212 106 2003-04-16 5.2516789224 13.750789 30.768176 13.567510
#> 214 107 2003-04-17 1.8968448413 13.818808 31.063500  9.897847
#> 215 108 2003-04-18 6.1293533649 13.886567 31.356413 14.759263
#> 217 109 2003-04-19 5.9622022181 13.954052 31.646820 14.672635
#> 220 110 2003-04-20 0.3209451749 14.021246 31.934628  8.349147
#> 222 111 2003-04-21 2.6673980125 14.088136 32.219743 11.105122
#> 224 112 2003-04-22 1.2831843798 14.154703 32.502075  9.598745
#> 225 113 2003-04-23 6.4525974893 14.220933 32.781534 15.632520
#> 228 114 2003-04-24 6.3433079422 14.286808 33.058030 15.603350
#> 230 115 2003-04-25 1.3506032247 14.352310 33.331477  9.901174
#> 232 116 2003-04-26 6.0506269371 14.417421 33.601789 15.451358
#> 233 117 2003-04-27 0.0929082131 14.482122 33.868881  8.575861
#> 236 118 2003-04-28 6.1996223773 14.546394 34.132670 15.806780
#> 238 119 2003-04-29 2.7407947497 14.610216 34.393074 11.824243
#> 240 120 2003-04-30 0.5332341893 14.673568 34.650012  9.292090
#> 241 121 2003-05-01 4.5274691340 14.736428 34.903405 14.087534
#> 244 122 2003-05-02 6.4165772784 14.798774 35.153175 16.409299
#> 246 123 2003-05-03 5.8415603414 14.860584 35.399245 15.807372
#> 247 124 2003-05-04 3.5087294918 14.921835 35.641540 13.100772
#> 250 125 2003-05-05 4.2150133890 14.982502 35.879987 14.017038
#> 252 126 2003-05-06 3.2827132393 15.042562 36.114511 12.969232
#> 254 127 2003-05-07 6.2046196673 15.101989 36.345041 16.552401
#> 255 128 2003-05-08 4.6938394464 15.160757 36.571508 14.804230
#> 258 129 2003-05-09 2.7564440363 15.218840 36.793843 12.530520
#> 260 130 2003-05-10 3.8590347059 15.276211 37.011977 13.927926
#> 262 131 2003-05-11 1.6974487305 15.332843 37.225844 11.367036
#> 263 132 2003-05-12 1.4224618033 15.388707 37.435378 11.089022
#> 266 133 2003-05-13 0.4783919237 15.443775 37.640517  9.993112
#> 268 134 2003-05-14 2.1509161601 15.498017 37.841197 12.086224
#> 270 135 2003-05-15 6.9518266600 15.551405 38.037357 18.011115
#> 271 136 2003-05-16 0.8141642399 15.603907 38.228936 10.554569
#> 273 137 2003-05-17 5.0079615365 15.655493 38.415875 15.748305
#> 276 138 2003-05-18 1.5827611203 15.706133 38.598118 11.594362
#> 278 139 2003-05-19 0.9334398333 15.755795 38.775606 10.842517
#> 280 140 2003-05-20 6.4909723308 15.804447 38.948286 17.735208
#> 281 141 2003-05-21 6.2588585999 15.852058 39.116103 17.501120
#> 284 142 2003-05-22 1.4256507456 15.898595 39.279005 11.580855
#> 286 143 2003-05-23 1.7988804334 15.944026 39.436940 12.083966
#> 288 144 2003-05-24 4.2980392252 15.988319 39.589859 15.218811
#> 289 145 2003-05-25 3.0904943133 16.031441 39.737713 13.764688
#> 292 146 2003-05-26 2.2091627968 16.073361 39.880455 12.710748
#> 294 147 2003-05-27 0.7091963268 16.114046 40.018039 10.885128
#> 296 148 2003-05-28 1.9085233700 16.153463 40.150421 12.409481
#> 297 149 2003-05-29 4.5758957798 16.191581 40.277558 15.760801
#> 299 150 2003-05-30 6.4950907156 16.228369 40.399408 18.184393
#> 302 151 2003-05-31 0.1863726901 16.263796 40.515933 10.361127
#> 304 152 2003-06-01 3.9164383798 16.297830 40.627093 15.038205
#> 306 153 2003-06-02 5.9792155649 16.330443 40.732853 17.640161
#> 307 154 2003-06-03 0.5311813466 16.361604 40.833176 10.871121
#> 310 155 2003-06-04 1.4408019935 16.391286 40.928031 12.030805
#> 312 156 2003-06-05 3.4712275269 16.419460 41.017385 14.590076
#> 314 157 2003-06-06 1.8329986085 16.446100 41.101209 12.565768
#> 315 158 2003-06-07 2.9117301547 16.471180 41.179475 13.934666
#> 318 159 2003-06-08 3.0740499967 16.494677 41.252156 14.157043
#> 320 160 2003-06-09 1.2766454353 16.516565 41.319229 11.926689
#> 322 161 2003-06-10 0.5482978623 16.536824 41.380671 11.031180
#> 323 162 2003-06-11 0.1243896922 16.555433 41.436461 10.514782
#> 326 163 2003-06-12 5.0972598132 16.572373 41.486582 16.751778
#> 327 164 2003-06-13 6.7959151736 16.587625 41.531016 18.890339
#> 330 165 2003-06-14 3.7465231165 16.601175 41.569750 15.083130
#> 332 166 2003-06-15 1.1654353749 16.613007 41.602770 11.859951
#> 334 167 2003-06-16 0.4072295546 16.623108 41.630068 10.917439
#> 335 168 2003-06-17 3.2139731331 16.631469 41.651634 14.437425
#> 338 169 2003-06-18 4.3595917500 16.638080 41.667464 15.875823
#> 340 170 2003-06-19 4.1637344097 16.642934 41.677552 15.632839
#> 342 171 2003-06-20 5.2011276619 16.646025 41.681897 16.932325
#> 343 172 2003-06-21 3.6699609405 16.647351 41.680500 15.014424
#> 346 173 2003-06-22 3.4777767523 16.646909 41.673363 14.771421
#> 348 174 2003-06-23 0.1030316185 16.644700 41.660491 10.544063
#> 350 175 2003-06-24 6.7054338716 16.640727 41.641891 18.800339
#> 351 176 2003-06-25 3.1716246458 16.634994 41.617572 14.371792
#> 353 177 2003-06-26 6.8707229439 16.627507 41.587545 18.989169
#> 356 178 2003-06-27 3.8129361621 16.618275 41.551823 15.154830
#> 358 179 2003-06-28 2.5202968079 16.607306 41.510423 13.527382
#> 360 180 2003-06-29 3.5220227316 16.594614 41.463360 14.765909
#> 361 181 2003-06-30 0.5224354262 16.580211 41.410655 11.005080
#> 364 182 2003-07-01 6.7885432325 16.564113 41.352328 18.811885
#> 366 183 2003-07-02 4.5395091963 16.546336 41.288405 15.985865
#> 368 184 2003-07-03 6.8919582665 16.526900 41.218909 18.899170
#> 369 185 2003-07-04 3.7529327988 16.505823 41.143868 14.963412
#> 372 186 2003-07-05 5.3274158353 16.483128 41.063311 16.901745
#> 374 187 2003-07-06 3.9507000893 16.458837 40.977269 15.162311
#> 376 188 2003-07-07 5.4491801246 16.432973 40.885775 17.000313
#> 377 189 2003-07-08 4.4734269199 16.405563 40.788864 15.758317
#> 379 190 2003-07-09 2.6527251494 16.376632 40.686570 13.466895
#> 382 191 2003-07-10 2.5167146940 16.346207 40.578932 13.268565
#> 384 192 2003-07-11 5.9096199137 16.314316 40.465990 17.445601
#> 386 193 2003-07-12 3.1972442656 16.280989 40.347784 14.048675
#> 387 194 2003-07-13 1.2532400342 16.246255 40.224356 11.607547
#> 390 195 2003-07-14 6.7634643705 16.210144 40.095751 18.388644
#> 392 196 2003-07-15 4.9938836617 16.172686 39.962013 16.160339
#> 394 197 2003-07-16 0.1742022182 16.133915 39.823189 10.170788
#> 395 198 2003-07-17 4.3851954595 16.093860 39.679327 15.325670
#> 398 199 2003-07-18 5.7954492546 16.052554 39.530476 17.018458
#> 400 200 2003-07-19 4.5014832215 16.010029 39.376687 15.379874
#> 402 201 2003-07-20 4.9593898323 15.966317 39.218012 15.895370
#> 403 202 2003-07-21 6.0950900745 15.921451 39.054503 17.239098
#> 406 203 2003-07-22 1.9726880719 15.875462 38.886214 12.137558
#> 407 204 2003-07-23 1.2125738936 15.828384 38.713200 11.161162
#> 410 205 2003-07-24 0.7676901331 15.780249 38.535519 10.571233
#> 412 206 2003-07-25 4.1035553410 15.731088 38.353226 14.590649
#> 414 207 2003-07-26 5.9550492915 15.680933 38.166382 16.788699
#> 415 208 2003-07-27 6.4506832347 15.629816 37.975044 17.330225
#> 418 209 2003-07-28 4.7296333136 15.577768 37.779274 15.179983
#> 420 210 2003-07-29 2.4541171370 15.524820 37.579134 12.364982
#> 422 211 2003-07-30 0.0606065234 15.471001 37.374685  9.416878
#> 423 212 2003-07-31 3.0162915480 15.416342 37.165992 12.927363
#> 426 213 2003-08-01 5.0745456801 15.360873 36.953119 15.342109
#> 428 214 2003-08-02 2.8857947094 15.304621 36.736132 12.647462
#> 430 215 2003-08-03 3.8560367664 15.247616 36.515097 13.746006
#> 431 216 2003-08-04 1.0945239908 15.189886 36.290082 10.379981
#> 433 217 2003-08-05 0.3010090524 15.131457 36.061156  9.373970
#> 436 218 2003-08-06 4.7925087914 15.072356 35.828388 14.653216
#> 438 219 2003-08-07 3.6008174797 15.012610 35.591848 13.166365
#> 440 220 2003-08-08 6.6198149386 14.952243 35.351609 16.663521
#> 441 221 2003-08-09 5.9016785952 14.891281 35.107744 15.733847
#> 444 222 2003-08-10 0.4541816534 14.829748 34.860325  9.248904
#> 446 223 2003-08-11 5.3347764020 14.767666 34.609429 14.903635
#> 448 224 2003-08-12 0.5486897659 14.705060 34.355129  9.229729
#> 449 225 2003-08-13 0.9345536388 14.641952 34.097505  9.612549
#> 452 226 2003-08-14 5.6678734070 14.578362 33.836632 15.036774
#> 454 227 2003-08-15 3.9156826755 14.514313 33.572591 12.921768
#> 456 228 2003-08-16 5.1361078029 14.449824 33.305462 14.245484
#> 457 229 2003-08-17 2.4478683304 14.384916 33.035326 11.069627
#> 459 230 2003-08-18 6.8786806664 14.319607 32.762264 16.059537
#> 462 231 2003-08-19 2.8619771593 14.253916 32.486361 11.382983
#> 464 232 2003-08-20 2.5261158552 14.187862 32.207701 10.919179
#> 466 233 2003-08-21 4.5675702749 14.121461 31.926370 13.144867
#> 467 234 2003-08-22 1.0049145808 14.054730 31.642455  9.041833
#> 470 235 2003-08-23 4.7655903073 13.987686 31.356043 13.180496
#> 472 236 2003-08-24 6.0150012553 13.920345 31.067223 14.478902
#> 474 237 2003-08-25 5.8704099276 13.852721 30.776085 14.215059
#> 475 238 2003-08-26 6.1691904743 13.784830 30.482721 14.441719
#> 478 239 2003-08-27 1.6638841627 13.716686 30.187223  9.377716
#> 480 240 2003-08-28 4.0504159203 13.648302 29.889684 11.907612
#> 482 241 2003-08-29 6.2853175399 13.579692 29.590199 14.245415
#> 483 242 2003-08-30 2.4110032138 13.510869 29.288864  9.935502
#> 486 243 2003-08-31 5.3260076088 13.441846 28.985774 12.988901
#> 488 244 2003-09-01 5.1177255162 13.372633 28.681027 12.658391
#> 489 245 2003-09-02 5.9670266353 13.303244 28.374722 13.457268
#> 492 246 2003-09-03 6.8307262855 13.233689 28.066958 14.260287
#> 494 247 2003-09-04 0.7913325059 13.163980 27.757836  7.773769
#> 496 248 2003-09-05 6.7918359581 13.094126 27.447457 13.980271
#> 497 249 2003-09-06 4.5365794727 13.024138 27.135924 11.509985
#> 500 250 2003-09-07 0.3925113028 12.954027 26.823338  7.112213
#> 502 251 2003-09-08 3.3119596706 12.883802 26.509805 10.034807
#> 504 252 2003-09-09 2.0623750919 12.813472 26.195428  8.656982
#> 505 253 2003-09-10 4.2720973536 12.743047 25.880314 10.808257
#> 508 254 2003-09-11 0.8477621619 12.672536 25.564567  7.246246
#> 510 255 2003-09-12 4.4059155553 12.601947 25.248295 10.725751
#> 512 256 2003-09-13 4.9836871200 12.531289 24.931604 11.190544
#> 513 257 2003-09-14 4.2845795064 12.460572 24.614602 10.385528
#> 515 258 2003-09-15 0.2410046314 12.389802 24.297397  6.310664
#> 518 259 2003-09-16 4.6612355092 12.318988 23.980097 10.531796
#> 520 260 2003-09-17 4.5707709356 12.248139 23.662812 10.330957
#> 522 261 2003-09-18 3.1630090780 12.177262 23.345649  8.868395
#> 523 262 2003-09-19 3.6176287131 12.106365 23.028718  9.197905
#> 526 263 2003-09-20 4.7362755802 12.035456 22.712128 10.146949
#> 528 264 2003-09-21 6.1665316585 11.964544 22.395989 11.370449
#> 530 265 2003-09-22 5.1186040824 11.893635 22.080409 10.271420
#> 531 266 2003-09-23 2.7903387067 11.822738 21.765497  8.009862
#> 534 267 2003-09-24 3.0136445761 11.751861 21.451363  8.113332
#> 536 268 2003-09-25 1.0178514274 11.681012 21.138114  6.205487
#> 538 269 2003-09-26 3.1578423053 11.610198 20.825859  8.038663
#> 539 270 2003-09-27 4.0514704413 11.539428 20.514705  8.730013
#> 541 271 2003-09-28 0.4929198441 11.468711 20.204761  5.485386
#> 544 272 2003-09-29 5.1962832615 11.398053 19.896131  9.509278
#> 546 273 2003-09-30 3.8445449444 11.327464 19.588923  8.221474
#> 548 274 2003-10-01 4.4418099055 11.256953 19.283241  8.625236
#> 549 275 2003-10-02 0.9564000864 11.186528 18.979189  5.556117
#> 552 276 2003-10-03 6.5116344769 11.116198 18.676871 10.139476
#> 554 277 2003-10-04 4.5578304774 11.045973 18.376389  8.385364
#> 556 278 2003-10-05 3.7009649067 10.975862 18.077844  7.567307
#> 557 279 2003-10-06 1.5758077204 10.905874 17.781337  5.729961
#> 560 280 2003-10-07 6.8828680869 10.836020 17.486965  9.925462
#> 562 281 2003-10-08 0.7633414995 10.766311 17.194826  4.908271
#> 564 282 2003-10-09 0.4896186206 10.696756 16.905016  4.613147
#> 565 283 2003-10-10 5.4435185483 10.627367 16.617629  8.410324
#> 568 284 2003-10-11 4.6082556229 10.558154 16.332759  7.647521
#> 569 285 2003-10-12 1.8144192055 10.489131 16.050495  5.400838
#> 572 286 2003-10-13 6.4283649134 10.420308 15.770928  8.807333
#> 574 287 2003-10-14 1.0409724275 10.351698 15.494145  4.652586
#> 576 288 2003-10-15 4.4847633946 10.283314 15.220231  7.123985
#> 577 289 2003-10-16 1.4086689684 10.215170 14.949271  4.768068
#> 580 290 2003-10-17 5.2053492058 10.147279 14.681345  7.435953
#> 582 291 2003-10-18 5.9334619991 10.079655 14.416534  7.847332
#> 584 292 2003-10-19 5.9754511106 10.012314 14.154915  7.762627
#> 585 293 2003-10-20 5.6238880090  9.945270 13.896562  7.403280
#> 588 294 2003-10-21 0.1671373537  9.878539 13.641550  3.525790
#> 590 295 2003-10-22 5.5676440035  9.812138 13.389948  7.146377
#> 592 296 2003-10-23 6.2058535737  9.746084 13.141825  7.469508
#> 593 297 2003-10-24 4.2002613691  9.680393 12.897247  6.022329
#> 595 298 2003-10-25 3.0295171652  9.615084 12.656277  5.157937
#> 598 299 2003-10-26 6.4419306223  9.550176 12.418977  7.293263
#> 600 300 2003-10-27 4.3121009958  9.485687 12.185404  5.816034
#> 602 301 2003-10-28 3.0500271111  9.421638 11.955615  4.924074
#> 603 302 2003-10-29 6.2811308708  9.358048 11.729663  6.868896
#> 606 303 2003-10-30 5.3830946553  9.294940 11.507600  6.209170
#> 608 304 2003-10-31 0.8968103081  9.232334 11.289473  3.370687
#> 610 305 2003-11-01 3.7704318434  9.170252 11.075328  5.045693
#> 611 306 2003-11-02 0.9031135815  9.108719 10.865209  3.254936
#> 614 307 2003-11-03 3.5767541877  9.047757 10.659156  4.771675
#> 616 308 2003-11-04 3.8065488376  8.987390 10.457208  4.828842
#> 618 309 2003-11-05 1.6924326292  8.927644 10.259398  3.537298
#> 619 310 2003-11-06 5.6220088068  8.868543 10.065762  5.706919
#> 621 311 2003-11-07 6.6353960712  8.810114  9.876328  6.188293
#> 624 312 2003-11-08 2.8515294176  8.752384  9.691126  4.001467
#> 626 313 2003-11-09 3.6032119396  8.695379  9.510180  4.347971
#> 628 314 2003-11-10 6.8707929526  8.639127  9.333513  6.044901
#> 629 315 2003-11-11 2.4793070969  8.583658  9.161147  3.613342
#> 632 316 2003-11-12 3.1908810346  8.528999  8.993100  3.930531
#> 634 317 2003-11-13 1.5401729443  8.475180  8.829388  3.009618
#> 636 318 2003-11-14 0.5728311650  8.422232  8.670025  2.462349
#> 637 319 2003-11-15 2.5846524800  8.370184  8.515023  3.443445
#> 640 320 2003-11-16 0.8866660867  8.319067  8.364392  2.536846
#> 642 321 2003-11-17 5.1791075380  8.268912  8.218139  4.628188
#> 644 322 2003-11-18 6.6910640008  8.219751  8.076271  5.306202
#> 645 323 2003-11-19 4.6807215177  8.171616  7.938791  4.258377
#> 648 324 2003-11-20 3.2766254421  8.124538  7.805702  3.525445
#> 649 325 2003-11-21 3.7673471260  8.078549  7.677004  3.709297
#> 652 326 2003-11-22 2.3166759131  8.033683  7.552698  2.977162
#> 654 327 2003-11-23 4.4007747553  7.989971  7.432780  3.905136
#> 656 328 2003-11-24 3.1678673008  7.947446  7.317248  3.287646
#> 657 329 2003-11-25 6.2161062430  7.906140  7.206095  4.634376
#> 660 330 2003-11-26 5.4894127378  7.866085  7.099317  4.251988
#> 662 331 2003-11-27 0.9664168563  7.827314  6.996907  2.181171
#> 664 332 2003-11-28 1.9691914173  7.789856  6.898857  2.596692
#> 665 333 2003-11-29 4.9486254861  7.753745  6.805157  3.872896
#> 668 334 2003-11-30 3.0953084414  7.719011  6.715800  3.025461
#> 670 335 2003-12-01 4.7438896613  7.685684  6.630775  3.704074
#> 672 336 2003-12-02 4.9050048993  7.653793  6.550072  3.736355
#> 673 337 2003-12-03 0.4788449355  7.623368  6.473680  1.821735
#> 675 338 2003-12-04 6.7095246441  7.594437  6.401590  4.428232
#> 678 339 2003-12-05 1.7113846468  7.567027  6.333789  2.299683
#> 680 340 2003-12-06 6.4839833423  7.541163  6.270268  4.263193
#> 682 341 2003-12-07 3.5192295376  7.516872  6.211016  3.006683
#> 683 342 2003-12-08 6.5486199944  7.494177  6.156023  4.228658
#> 686 343 2003-12-09 3.0715974274  7.473100  6.105278  2.781017
#> 688 344 2003-12-10 2.4534057572  7.453664  6.058772  2.511829
#> 690 345 2003-12-11 3.3355286187  7.435887  6.016497  2.853539
#> 691 346 2003-12-12 6.7631166757  7.419789  5.978443  4.219277
#> 694 347 2003-12-13 2.0801334889  7.405386  5.944603  2.321054
#> 696 348 2003-12-14 2.0345381584  7.392694  5.914970  2.292670
#> 698 349 2003-12-15 6.0180557331  7.381725  5.889539  3.873149
#> 699 350 2003-12-16 1.3531886679  7.372493  5.868303  2.005626
#> 701 351 2003-12-17 5.6637585997  7.365006  5.851259  3.712652
#> 704 352 2003-12-18 2.5584055870  7.359273  5.838404  2.474444
#> 706 353 2003-12-19 1.7688975721  7.355300  5.829736  2.158439
#> 708 354 2003-12-20 5.9925670966  7.353091  5.825253  3.830024
#> 709 355 2003-12-21 0.9101705870  7.352649  5.824956  1.816769
#> 712 356 2003-12-22 5.5433536230  7.353975  5.828845  3.654074
#> 714 357 2003-12-23 2.4913105716  7.357066  5.836924  2.447505
#> 716 358 2003-12-24 6.7504068639  7.361920  5.849194  4.143966
#> 717 359 2003-12-25 0.8307824642  7.368531  5.865662  1.797084
#> 720 360 2003-12-26 3.2071245466  7.376892  5.886331  2.751132
#> 722 361 2003-12-27 3.6655530566  7.386993  5.911209  2.944424
#> 724 362 2003-12-28 5.8594320805  7.398825  5.940303  3.837260
#> 725 363 2003-12-29 0.2345282619  7.412375  5.973622  1.587909
#> 728 364 2003-12-30 0.3032990755  7.427627  6.011174  1.625523
#> 730 365 2003-12-31 5.5447206120  7.444567  6.052971  3.767372