Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.601768  1.601768  4.039080  4.039080  3.873766  3.873766  2.635593
#>   [8]  2.635593  2.489244  2.489244  1.798942  1.798942  2.840999  2.840999
#>  [15]  2.030117  2.030117  4.240063  4.240063  1.829316  1.829316  3.520620
#>  [22]  3.520620  1.894883  1.894883  2.371260  2.371260  3.978525  3.978525
#>  [29]  3.017149  3.017149  4.273271  4.273271  1.867317  1.867317  2.761599
#>  [36]  2.761599  4.958729  4.958729  2.868344  2.868344  4.840547  4.840547
#>  [43]  2.647289  2.647289  4.337323  4.337323  5.553779  5.553779  3.623887
#>  [50]  3.623887  5.587445  5.587445  4.824362  4.824362  3.872465  3.872465
#>  [57]  4.081522  4.081522  3.922153  3.922153  4.377138  4.377138  3.966740
#>  [64]  3.966740  3.923130  3.923130  4.181601  4.181601  4.079120  4.079120
#>  [71]  3.943665  3.943665  5.096638  5.096638  4.427046  4.427046  6.115527
#>  [78]  6.115527  3.976369  3.976369  3.752186  3.752186  5.456942  5.456942
#>  [85]  4.647752  4.647752  3.126345  3.126345  7.247915  7.247915  7.585382
#>  [92]  7.585382  3.463594  3.463594  3.855597  3.855597  4.121900  4.121900
#>  [99]  5.656047  5.656047  5.841313  5.841313  6.401941  6.401941  4.214516
#> [106]  4.214516  7.802588  7.802588  7.793982  7.793982  7.996657  7.996657
#> [113]  8.562602  8.562602  5.824826  5.824826  6.684622  6.684622  8.755223
#> [120]  8.755223  9.507008  9.507008  4.823065  4.823065  5.109122  5.109122
#> [127]  9.734216  9.734216  7.533764  7.533764  8.227447  8.227447  7.634625
#> [134]  7.634625  6.044771  6.044771  8.135675  8.135675  7.850604  7.850604
#> [141]  6.308817  6.308817 10.424234 10.424234  6.092068  6.092068  8.202407
#> [148]  8.202407  9.541447  9.541447 11.032377 11.032377  7.949214  7.949214
#> [155]  7.375356  7.375356 11.239429 11.239429 11.088527 11.088527 10.471957
#> [162] 10.471957  6.957552  6.957552 10.845444 10.845444  9.568186  9.568186
#> [169]  6.898519  6.898519 12.340706 12.340706  9.911105  9.911105  9.959297
#> [176]  9.959297 10.044463 10.044463  9.867184  9.867184  9.962361  9.962361
#> [183]  8.790337  8.790337  6.904265  6.904265 13.603791 13.603791  7.546770
#> [190]  7.546770  8.034413  8.034413  7.967530  7.967530 12.282553 12.282553
#> [197]  9.103521  9.103521 14.220458 14.220458 11.756603 11.756603 12.596647
#> [204] 12.596647  9.506589  9.506589 15.004484 15.004484  9.254682  9.254682
#> [211] 10.887657 10.887657 13.048517 13.048517 13.623386 13.623386 14.185685
#> [218] 14.185685 11.738105 11.738105 11.322313 11.322313 13.142585 13.142585
#> [225] 14.197254 14.197254 12.902265 12.902265 14.079183 14.079183  8.472167
#> [232]  8.472167 15.490256 15.490256 11.232276 11.232276  9.737930  9.737930
#> [239] 16.886675 16.886675 14.224744 14.224744 11.820978 11.820978  9.905446
#> [246]  9.905446 16.385300 16.385300 13.946458 13.946458 10.963852 10.963852
#> [253] 16.721986 16.721986 11.327540 11.327540 13.331465 13.331465 10.357740
#> [260] 10.357740 10.985225 10.985225 11.488916 11.488916 17.695798 17.695798
#> [267] 16.556660 16.556660 11.617669 11.617669 16.093704 16.093704 10.740381
#> [274] 10.740381 10.496150 10.496150 11.616267 11.616267 10.861861 10.861861
#> [281] 14.589335 14.589335 11.686629 11.686629 16.815762 16.815762 17.462525
#> [288] 17.462525 14.310984 14.310984 14.643937 14.643937 17.425702 17.425702
#> [295] 18.376166 18.376166 10.524747 10.524747 15.210192 15.210192 12.635078
#> [302] 12.635078 15.658170 15.658170 18.784327 18.784327 15.674472 15.674472
#> [309] 17.335623 17.335623 11.939942 11.939942 18.549496 18.549496 13.183783
#> [316] 13.183783 12.879771 12.879771 10.728282 10.728282 19.030300 19.030300
#> [323] 15.433976 15.433976 14.242614 14.242614 15.818432 15.818432 11.241381
#> [330] 11.241381 17.343821 17.343821 13.499756 13.499756 17.275415 17.275415
#> [337] 18.868424 18.868424 17.474020 17.474020 10.842998 10.842998 10.804873
#> [344] 10.804873 16.991260 16.991260 16.721441 16.721441 14.523022 14.523022
#> [351] 12.363719 12.363719 15.987035 15.987035 16.433459 16.433459 11.760084
#> [358] 11.760084 15.324855 15.324855 14.826859 14.826859 14.617210 14.617210
#> [365] 11.251430 11.251430 17.273840 17.273840 11.665292 11.665292 10.517273
#> [372] 10.517273 10.567685 10.567685 11.932583 11.932583 17.412147 17.412147
#> [379] 10.269931 10.269931 17.417670 17.417670 17.746643 17.746643 13.366985
#> [386] 13.366985 16.803978 16.803978 17.070570 17.070570 14.570941 14.570941
#> [393] 10.766133 10.766133 16.548500 16.548500 11.991092 11.991092 13.049233
#> [400] 13.049233 15.425174 15.425174 14.584122 14.584122 17.042243 17.042243
#> [407] 11.791604 11.791604 17.010329 17.010329 12.861737 12.861737 10.382033
#> [414] 10.382033 16.633062 16.633062 11.384535 11.384535 15.706892 15.706892
#> [421]  9.364748  9.364748 14.761314 14.761314 15.937865 15.937865 10.241027
#> [428] 10.241027 16.971668 16.971668 13.393461 13.393461 13.974837 13.974837
#> [435] 11.107972 11.107972 16.854875 16.854875 13.779882 13.779882 15.867411
#> [442] 15.867411  9.484985  9.484985 16.250184 16.250184 15.003590 15.003590
#> [449] 16.464822 16.464822 11.159352 11.159352 15.858338 15.858338  9.919560
#> [456]  9.919560 15.570288 15.570288 10.235584 10.235584  9.975677  9.975677
#> [463] 11.993128 11.993128  9.490770  9.490770  8.064118  8.064118 14.660436
#> [470] 14.660436 10.586518 10.586518 14.202971 14.202971  9.999743  9.999743
#> [477] 12.048058 12.048058 12.175527 12.175527  7.976428  7.976428 14.550280
#> [484] 14.550280 12.985616 12.985616 10.718640 10.718640 14.122768 14.122768
#> [491]  9.118611  9.118611 13.622436 13.622436  7.780417  7.780417 13.768374
#> [498] 13.768374  6.808553  6.808553 13.444078 13.444078 10.280606 10.280606
#> [505]  7.287320  7.287320 10.058672 10.058672 11.149210 11.149210  7.092244
#> [512]  7.092244 11.804071 11.804071  9.049253  9.049253 11.228642 11.228642
#> [519]  6.521733  6.521733 12.075568 12.075568 11.595119 11.595119  5.921768
#> [526]  5.921768  6.164408  6.164408  9.301837  9.301837  8.191837  8.191837
#> [533] 10.834242 10.834242 11.017983 11.017983 10.968505 10.968505  6.217729
#> [540]  6.217729  9.771353  9.771353  7.433898  7.433898  5.813280  5.813280
#> [547] 10.145062 10.145062  9.897462  9.897462  5.980783  5.980783  5.985919
#> [554]  5.985919 10.150373 10.150373  8.126142  8.126142  7.982532  7.982532
#> [561]  9.754624  9.754624  5.761615  5.761615  4.327242  4.327242  8.456808
#> [568]  8.456808  8.304416  8.304416  8.707732  8.707732  4.655705  4.655705
#> [575]  5.731830  5.731830  5.692033  5.692033  4.359444  4.359444  3.830904
#> [582]  3.830904  7.184359  7.184359  6.116250  6.116250  7.559046  7.559046
#> [589]  6.160467  6.160467  5.257762  5.257762  7.581108  7.581108  5.169672
#> [596]  5.169672  3.638840  3.638840  3.630286  3.630286  6.323953  6.323953
#> [603]  4.103909  4.103909  4.626551  4.626551  6.997821  6.997821  6.882456
#> [610]  6.882456  3.206972  3.206972  3.606785  3.606785  4.009293  4.009293
#> [617]  2.942487  2.942487  6.317035  6.317035  4.729747  4.729747  4.588452
#> [624]  4.588452  6.189936  6.189936  4.106536  4.106536  4.361383  4.361383
#> [631]  3.613553  3.613553  4.421625  4.421625  2.543725  2.543725  2.710636
#> [638]  2.710636  4.895145  4.895145  3.129492  3.129492  3.301122  3.301122
#> [645]  3.726040  3.726040  3.454415  3.454415  3.860141  3.860141  2.386011
#> [652]  2.386011  4.665645  4.665645  3.353569  3.353569  3.922403  3.922403
#> [659]  2.711871  2.711871  2.095318  2.095318  4.292517  4.292517  4.718195
#> [666]  4.718195  1.804357  1.804357  3.383942  3.383942  3.175437  3.175437
#> [673]  2.336147  2.336147  4.168759  4.168759  4.411734  4.411734  1.928203
#> [680]  1.928203  3.203647  3.203647  1.977981  1.977981  3.891587  3.891587
#> [687]  3.505735  3.505735  3.872080  3.872080  4.035147  4.035147  3.067895
#> [694]  3.067895  3.489016  3.489016  2.828326  2.828326  3.871524  3.871524
#> [701]  1.972011  1.972011  1.534405  1.534405  2.237313  2.237313  2.987050
#> [708]  2.987050  1.485343  1.485343  2.681916  2.681916  3.106191  3.106191
#> [715]  3.237301  3.237301  1.584387  1.584387  3.870211  3.870211  1.488180
#> [722]  1.488180  1.817052  1.817052  4.116688  4.116688  4.151533  4.151533
#> [729]  2.113648  2.113648