Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.289837  2.289837  1.547695  1.547695  1.578673  1.578673  2.104927
#>   [8]  2.104927  4.026753  4.026753  3.110062  3.110062  4.047013  4.047013
#>  [15]  4.175801  4.175801  1.813724  1.813724  2.734304  2.734304  4.304299
#>  [22]  4.304299  4.001392  4.001392  3.406887  3.406887  3.428380  3.428380
#>  [29]  4.536910  4.536910  4.089675  4.089675  3.224583  3.224583  4.974039
#>  [36]  4.974039  2.386528  2.386528  2.250291  2.250291  4.314205  4.314205
#>  [43]  3.278000  3.278000  3.719376  3.719376  4.926485  4.926485  4.567634
#>  [50]  4.567634  4.783049  4.783049  2.488058  2.488058  4.951476  4.951476
#>  [57]  2.718120  2.718120  4.952857  4.952857  5.022897  5.022897  6.024652
#>  [64]  6.024652  5.331690  5.331690  3.325315  3.325315  2.724124  2.724124
#>  [71]  4.322063  4.322063  4.344756  4.344756  3.375614  3.375614  6.699923
#>  [78]  6.699923  4.505495  4.505495  3.640214  3.640214  6.396299  6.396299
#>  [85]  5.898218  5.898218  3.194585  3.194585  4.490655  4.490655  4.218196
#>  [92]  4.218196  6.090811  6.090811  4.852966  4.852966  4.349286  4.349286
#>  [99]  5.795177  5.795177  6.856898  6.856898  7.099672  7.099672  6.252633
#> [106]  6.252633  4.426235  4.426235  5.699660  5.699660  7.760663  7.760663
#> [113]  4.806420  4.806420  5.566491  5.566491  6.328859  6.328859  7.629023
#> [120]  7.629023  5.454947  5.454947  9.272011  9.272011  4.786832  4.786832
#> [127]  9.984831  9.984831  5.160923  5.160923  7.975403  7.975403 10.082363
#> [134] 10.082363  7.906495  7.906495  5.837874  5.837874  9.245018  9.245018
#> [141]  8.158899  8.158899  7.599563  7.599563  5.712733  5.712733 10.286252
#> [148] 10.286252  6.009853  6.009853  7.806166  7.806166 10.554274 10.554274
#> [155]  5.958663  5.958663 10.823941 10.823941  8.042517  8.042517 10.924421
#> [162] 10.924421  7.284257  7.284257 11.477113 11.477113  6.427442  6.427442
#> [169] 10.583466 10.583466 11.233031 11.233031 10.182051 10.182051 11.258660
#> [176] 11.258660  7.673890  7.673890  7.718582  7.718582 10.861588 10.861588
#> [183] 12.775173 12.775173 10.926065 10.926065 13.294505 13.294505  7.591404
#> [190]  7.591404  9.039898  9.039898  7.703471  7.703471 11.250066 11.250066
#> [197] 10.470442 10.470442  8.858253  8.858253 10.163948 10.163948 12.776956
#> [204] 12.776956  8.514437  8.514437 13.194269 13.194269 13.761740 13.761740
#> [211] 11.803636 11.803636 12.817796 12.817796 11.264247 11.264247  9.368507
#> [218]  9.368507  8.921556  8.921556 13.844682 13.844682 11.711924 11.711924
#> [225] 13.663891 13.663891 12.506504 12.506504  8.663082  8.663082  9.795320
#> [232]  9.795320 14.664547 14.664547 15.892668 15.892668 15.628078 15.628078
#> [239] 16.758560 16.758560 13.295783 13.295783 13.274848 13.274848 13.318567
#> [246] 13.318567 11.456712 11.456712 11.740428 11.740428 12.174947 12.174947
#> [253] 15.297542 15.297542 11.586015 11.586015 16.671863 16.671863 12.677527
#> [260] 12.677527 14.248549 14.248549 13.342370 13.342370 17.601881 17.601881
#> [267] 16.444162 16.444162  9.614030  9.614030 15.810740 15.810740 14.048856
#> [274] 14.048856 18.096234 18.096234  9.969758  9.969758  9.896877  9.896877
#> [281] 10.524983 10.524983 14.477391 14.477391 13.986582 13.986582 12.891298
#> [288] 12.891298 11.413473 11.413473 11.743738 11.743738 15.298092 15.298092
#> [295] 11.081937 11.081937 15.909939 15.909939 12.623528 12.623528 12.112101
#> [302] 12.112101 15.249003 15.249003 18.811647 18.811647 14.626723 14.626723
#> [309] 17.041276 17.041276 10.506655 10.506655 12.395870 12.395870 16.486608
#> [316] 16.486608 14.597292 14.597292 13.616763 13.616763 13.065044 13.065044
#> [323] 11.823662 11.823662 18.703803 18.703803 16.677635 16.677635 19.111122
#> [330] 19.111122 18.564451 18.564451 14.672899 14.672899 19.062056 19.062056
#> [337] 14.143383 14.143383 17.821624 17.821624 13.047142 13.047142 17.829983
#> [344] 17.829983 12.957888 12.957888 17.079784 17.079784 12.072212 12.072212
#> [351] 17.104471 17.104471 11.622429 11.622429 14.124190 14.124190 14.374928
#> [358] 14.374928 17.923672 17.923672 14.254734 14.254734 16.482617 16.482617
#> [365] 16.851487 16.851487 11.893848 11.893848 13.033771 13.033771 17.809127
#> [372] 17.809127 17.711477 17.711477 16.001967 16.001967 14.238031 14.238031
#> [379] 11.807209 11.807209 13.721071 13.721071 17.704489 17.704489 16.251562
#> [386] 16.251562 16.379189 16.379189 14.698960 14.698960 18.562744 18.562744
#> [393] 14.974500 14.974500  9.963246  9.963246 11.895909 11.895909 15.398471
#> [400] 15.398471 15.436360 15.436360 16.984476 16.984476 13.682524 13.682524
#> [407] 10.582107 10.582107 16.131588 16.131588 12.105679 12.105679  9.997474
#> [414]  9.997474 17.007711 17.007711 10.850560 10.850560 11.778147 11.778147
#> [421] 14.829602 14.829602 10.212120 10.212120 10.928686 10.928686 11.610616
#> [428] 11.610616 11.575774 11.575774 10.990352 10.990352 12.981466 12.981466
#> [435] 16.930569 16.930569  9.342979  9.342979 17.079786 17.079786 15.329604
#> [442] 15.329604  9.878713  9.878713  9.502118  9.502118 10.873123 10.873123
#> [449]  9.281633  9.281633  9.000966  9.000966 10.424739 10.424739 11.190651
#> [456] 11.190651 13.137851 13.137851 12.824605 12.824605  8.640593  8.640593
#> [463] 12.559679 12.559679 13.561726 13.561726  8.507276  8.507276  9.632386
#> [470]  9.632386 15.071768 15.071768 14.244368 14.244368 14.878925 14.878925
#> [477] 10.437202 10.437202 14.714481 14.714481  8.129394  8.129394 10.274247
#> [484] 10.274247 14.561274 14.561274 10.167274 10.167274 12.407209 12.407209
#> [491]  7.522449  7.522449  7.706049  7.706049 13.367245 13.367245 12.351540
#> [498] 12.351540 11.417278 11.417278 13.799540 13.799540  9.393882  9.393882
#> [505] 10.918638 10.918638 10.371469 10.371469 10.977973 10.977973  8.317605
#> [512]  8.317605  7.043847  7.043847 10.235980 10.235980 10.162467 10.162467
#> [519]  7.837154  7.837154 10.658039 10.658039 10.655437 10.655437 12.077866
#> [526] 12.077866  9.498847  9.498847  6.228470  6.228470  7.945625  7.945625
#> [533] 10.538246 10.538246 10.047222 10.047222  6.921816  6.921816  7.322546
#> [540]  7.322546  6.548160  6.548160  8.605183  8.605183  9.655624  9.655624
#> [547]  7.549547  7.549547  6.099821  6.099821  6.981368  6.981368  8.041996
#> [554]  8.041996  8.652085  8.652085  7.488182  7.488182  8.799747  8.799747
#> [561]  4.477753  4.477753  4.470433  4.470433  5.279964  5.279964  9.260597
#> [568]  9.260597  6.557884  6.557884  4.594912  4.594912  7.114685  7.114685
#> [575]  6.701371  6.701371  8.252649  8.252649  4.384474  4.384474  3.625490
#> [582]  3.625490  5.593600  5.593600  4.888582  4.888582  5.940554  5.940554
#> [589]  7.220011  7.220011  5.522372  5.522372  7.541980  7.541980  7.729283
#> [596]  7.729283  4.796914  4.796914  5.340838  5.340838  7.197874  7.197874
#> [603]  4.848711  4.848711  5.914669  5.914669  6.150915  6.150915  4.332170
#> [610]  4.332170  5.339279  5.339279  5.839002  5.839002  3.562110  3.562110
#> [617]  3.445132  3.445132  5.338035  5.338035  3.661573  3.661573  5.000411
#> [624]  5.000411  5.619291  5.619291  3.194942  3.194942  4.521839  4.521839
#> [631]  4.924684  4.924684  5.044227  5.044227  4.363753  4.363753  4.827377
#> [638]  4.827377  2.513560  2.513560  4.841335  4.841335  4.201552  4.201552
#> [645]  5.310864  5.310864  3.794927  3.794927  4.912474  4.912474  2.841779
#> [652]  2.841779  5.001423  5.001423  3.448792  3.448792  3.064373  3.064373
#> [659]  4.627360  4.627360  2.405252  2.405252  2.701279  2.701279  3.094500
#> [666]  3.094500  2.272022  2.272022  3.012386  3.012386  3.697961  3.697961
#> [673]  4.251058  4.251058  3.788314  3.788314  2.964513  2.964513  2.507045
#> [680]  2.507045  2.972782  2.972782  2.481148  2.481148  3.703342  3.703342
#> [687]  4.337410  4.337410  3.655343  3.655343  2.767524  2.767524  3.749273
#> [694]  3.749273  3.374288  3.374288  3.747898  3.747898  2.299773  2.299773
#> [701]  3.823435  3.823435  2.590503  2.590503  1.967858  1.967858  3.822008
#> [708]  3.822008  3.527655  3.527655  2.330500  2.330500  2.016963  2.016963
#> [715]  1.819009  1.819009  1.717833  1.717833  4.103716  4.103716  3.714606
#> [722]  3.714606  3.695466  3.695466  2.948169  2.948169  2.767996  2.767996
#> [729]  4.217580  4.217580