Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

Usage

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  2.841109  2.841109  2.945858  2.945858  2.470824  2.470824  3.970617
#>   [8]  3.970617  3.128069  3.128069  3.233359  3.233359  1.682679  1.682679
#>  [15]  4.322206  4.322206  1.861169  1.861169  4.337073  4.337073  1.915929
#>  [22]  1.915929  2.901347  2.901347  4.036539  4.036539  3.644001  3.644001
#>  [29]  2.772374  2.772374  4.845811  4.845811  2.347067  2.347067  2.241513
#>  [36]  2.241513  4.862814  4.862814  4.361677  4.361677  2.255441  2.255441
#>  [43]  4.747429  4.747429  5.161967  5.161967  3.693489  3.693489  5.443259
#>  [50]  5.443259  5.573046  5.573046  5.192781  5.192781  4.569619  4.569619
#>  [57]  2.472162  2.472162  4.627935  4.627935  3.944191  3.944191  3.050611
#>  [64]  3.050611  5.351242  5.351242  2.841518  2.841518  4.864615  4.864615
#>  [71]  2.784507  2.784507  5.141727  5.141727  5.290278  5.290278  5.547001
#>  [78]  5.547001  5.243992  5.243992  3.009459  3.009459  3.799080  3.799080
#>  [85]  3.727107  3.727107  6.566351  6.566351  3.675466  3.675466  6.564695
#>  [92]  6.564695  6.286270  6.286270  6.633131  6.633131  6.404964  6.404964
#>  [99]  6.579699  6.579699  5.229123  5.229123  7.647606  7.647606  8.386132
#> [106]  8.386132  4.699823  4.699823  3.923121  3.923121  5.941258  5.941258
#> [113]  5.917422  5.917422  7.599959  7.599959  8.057647  8.057647  7.441619
#> [120]  7.441619  8.743401  8.743401  6.103689  6.103689  8.939182  8.939182
#> [127]  5.478390  5.478390  6.523954  6.523954  7.590663  7.590663  9.470793
#> [134]  9.470793 10.415129 10.415129  6.534375  6.534375  8.342407  8.342407
#> [141]  7.458071  7.458071 10.253267 10.253267  5.530310  5.530310 10.493177
#> [148] 10.493177  7.678679  7.678679  6.982424  6.982424  7.648391  7.648391
#> [155] 11.433003 11.433003  8.473149  8.473149 12.001318 12.001318  8.783192
#> [162]  8.783192  9.131558  9.131558  8.703326  8.703326 11.884840 11.884840
#> [169]  6.241149  6.241149 11.041694 11.041694  9.454407  9.454407  6.589205
#> [176]  6.589205  8.434789  8.434789 12.400969 12.400969  7.376753  7.376753
#> [183]  9.565744  9.565744 10.763960 10.763960 11.437735 11.437735 11.846898
#> [190] 11.846898 14.231167 14.231167 10.211782 10.211782 13.587253 13.587253
#> [197] 14.141970 14.141970  8.817786  8.817786 12.368190 12.368190 11.135614
#> [204] 11.135614 13.787631 13.787631  9.992681  9.992681 11.263493 11.263493
#> [211]  8.050951  8.050951 11.707760 11.707760  7.868158  7.868158 12.892923
#> [218] 12.892923 11.420595 11.420595 15.767572 15.767572 13.937899 13.937899
#> [225]  8.846165  8.846165 14.067474 14.067474  9.222769  9.222769 13.320513
#> [232] 13.320513 14.832686 14.832686  8.940740  8.940740  8.740751  8.740751
#> [239] 11.349379 11.349379 13.839914 13.839914 15.175409 15.175409 15.818098
#> [246] 15.818098 16.849028 16.849028 16.632812 16.632812 11.579312 11.579312
#> [253] 11.250563 11.250563 11.034917 11.034917 14.102683 14.102683 10.846521
#> [260] 10.846521 11.041213 11.041213 11.721586 11.721586 12.050037 12.050037
#> [267] 15.829145 15.829145 10.819569 10.819569 16.129550 16.129550 18.025963
#> [274] 18.025963 11.426298 11.426298 16.750578 16.750578 16.182487 16.182487
#> [281] 17.272031 17.272031  9.865104  9.865104 13.263222 13.263222 12.770039
#> [288] 12.770039 16.285834 16.285834 16.464820 16.464820 18.163632 18.163632
#> [295] 11.399258 11.399258 17.589473 17.589473 17.458150 17.458150 13.524756
#> [302] 13.524756 18.054274 18.054274 15.726797 15.726797 12.670727 12.670727
#> [309] 10.844434 10.844434 15.303577 15.303577 18.422977 18.422977 11.121871
#> [316] 11.121871 16.982603 16.982603 16.842103 16.842103 17.462754 17.462754
#> [323] 11.915876 11.915876 15.834930 15.834930 16.050938 16.050938 11.291743
#> [330] 11.291743 17.486448 17.486448 16.297731 16.297731 13.823809 13.823809
#> [337] 10.607804 10.607804 10.692216 10.692216 15.127232 15.127232 15.461293
#> [344] 15.461293 10.659400 10.659400 12.664376 12.664376 17.677513 17.677513
#> [351] 16.251185 16.251185 18.439199 18.439199 14.132335 14.132335 15.202649
#> [358] 15.202649 16.414663 16.414663 14.051604 14.051604 18.777047 18.777047
#> [365] 14.481089 14.481089 11.847059 11.847059 10.542679 10.542679 10.919237
#> [372] 10.919237 12.030255 12.030255 16.876454 16.876454 18.013395 18.013395
#> [379] 18.622018 18.622018 14.338989 14.338989 11.048298 11.048298 12.130043
#> [386] 12.130043 10.227205 10.227205 14.818012 14.818012 17.813680 17.813680
#> [393] 16.944511 16.944511 15.847180 15.847180 11.400440 11.400440 17.717925
#> [400] 17.717925 14.968867 14.968867  9.977592  9.977592 17.053094 17.053094
#> [407] 11.632469 11.632469 11.679112 11.679112  9.736917  9.736917 12.517723
#> [414] 12.517723 12.543916 12.543916 15.685515 15.685515 16.219557 16.219557
#> [421]  9.580761  9.580761 14.079537 14.079537  9.465936  9.465936 12.688176
#> [428] 12.688176 11.434300 11.434300 14.814449 14.814449 13.451087 13.451087
#> [435] 12.744058 12.744058 14.977288 14.977288 12.388962 12.388962 12.588575
#> [442] 12.588575 12.622003 12.622003 14.500111 14.500111 10.605023 10.605023
#> [449]  9.845307  9.845307 11.339437 11.339437 13.927580 13.927580 13.198531
#> [456] 13.198531 11.161120 11.161120 13.974594 13.974594 14.931999 14.931999
#> [463] 10.397218 10.397218  8.225106  8.225106 11.995454 11.995454 10.570850
#> [470] 10.570850 11.132268 11.132268  8.874908  8.874908 11.369330 11.369330
#> [477]  7.819391  7.819391 11.592523 11.592523 14.688953 14.688953  8.645772
#> [484]  8.645772  9.546018  9.546018  9.490677  9.490677 10.758885 10.758885
#> [491] 12.056768 12.056768 10.617045 10.617045 14.057872 14.057872 10.485643
#> [498] 10.485643  8.869981  8.869981 10.796304 10.796304 10.062234 10.062234
#> [505]  8.411625  8.411625  7.698149  7.698149 13.310857 13.310857  7.497458
#> [512]  7.497458 12.704393 12.704393  6.121607  6.121607  7.039910  7.039910
#> [519]  7.963617  7.963617 10.205418 10.205418 10.815888 10.815888  7.666489
#> [526]  7.666489  7.291684  7.291684  8.921212  8.921212 11.661337 11.661337
#> [533]  9.409427  9.409427 10.892962 10.892962  7.782375  7.782375  6.471846
#> [540]  6.471846 10.725147 10.725147 11.036604 11.036604  9.345284  9.345284
#> [547]  9.017716  9.017716  5.492979  5.492979  6.991739  6.991739  5.377561
#> [554]  5.377561  7.855654  7.855654  8.456645  8.456645  4.509557  4.509557
#> [561]  8.011858  8.011858  8.944818  8.944818  6.034850  6.034850  8.909997
#> [568]  8.909997  6.579114  6.579114  6.009753  6.009753  8.399555  8.399555
#> [575]  6.554485  6.554485  7.528822  7.528822  5.997886  5.997886  7.322648
#> [582]  7.322648  7.925861  7.925861  7.050502  7.050502  3.488712  3.488712
#> [589]  6.291872  6.291872  4.438370  4.438370  4.693159  4.693159  7.289259
#> [596]  7.289259  6.519700  6.519700  6.285513  6.285513  7.014240  7.014240
#> [603]  3.285033  3.285033  3.569141  3.569141  3.636780  3.636780  3.475374
#> [610]  3.475374  3.961473  3.961473  2.697965  2.697965  2.904950  2.904950
#> [617]  3.011386  3.011386  3.360594  3.360594  2.664862  2.664862  3.120904
#> [624]  3.120904  5.546045  5.546045  3.294264  3.294264  2.791596  2.791596
#> [631]  4.443802  4.443802  5.821651  5.821651  5.016958  5.016958  4.131735
#> [638]  4.131735  2.360481  2.360481  2.441930  2.441930  4.978945  4.978945
#> [645]  2.718119  2.718119  3.386854  3.386854  4.300786  4.300786  2.638614
#> [652]  2.638614  2.897185  2.897185  3.544317  3.544317  3.360307  3.360307
#> [659]  3.870127  3.870127  4.508749  4.508749  3.742129  3.742129  2.235298
#> [666]  2.235298  2.349267  2.349267  2.701763  2.701763  3.659238  3.659238
#> [673]  2.336633  2.336633  1.819257  1.819257  3.730025  3.730025  4.042174
#> [680]  4.042174  4.086877  4.086877  3.173589  3.173589  2.645528  2.645528
#> [687]  1.637824  1.637824  2.379814  2.379814  1.756851  1.756851  3.678521
#> [694]  3.678521  2.174477  2.174477  4.086925  4.086925  3.439713  3.439713
#> [701]  1.968066  1.968066  3.481943  3.481943  4.040475  4.040475  3.824789
#> [708]  3.824789  2.652376  2.652376  2.037360  2.037360  3.330355  3.330355
#> [715]  1.593958  1.593958  3.897926  3.897926  1.678502  1.678502  1.542558
#> [722]  1.542558  2.560395  2.560395  2.809327  2.809327  2.430849  2.430849
#> [729]  2.225417  2.225417