Calculate global solar radiation from sunshine duration hours

Uses functions taken from the 'sirad' package to determine astronomical daylength and extraterrestrial radiation, from which global radiation is calculated using the Angström-formula.

calc_globrad(dates, sunhours, lat, a0 = 0.25, b0 = 0.5, full_output = FALSE)

Arguments

dates

Date vector

sunhours

Vector of sunshine duration hours, same length as dates.

lat

Latitude in decimal degrees.

a0

Angström parameter a, defaults to 0.25.

b0

Angström parameter b, defaults to 0.5.

full_output

Return extraterrestrial radiation and daylength along with global radiation?

Value

A sequence of global radiation in MJ/(m² d) with the length of dates, or (if full_output = TRUE) a data.frame holding day of year, dates, sunhours, daylength, and extraterrestrial and calculated global solar radiation. A warning is generated if some sunshine duration hours are higher than the expected daylength at the specified latitude.

Examples

dates <- seq.Date(as.Date("2002-01-01"), as.Date("2003-12-31"), by = 'day')
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8)
#>   [1]  1.755725  3.936917  3.289054  2.023409  1.603187  2.975135  3.096742
#>   [8]  2.503919  3.870244  4.031175  4.388205  2.265290  1.854240  2.785909
#>  [15]  3.084535  2.458642  3.172094  2.096030  3.212572  5.236516  2.970865
#>  [22]  4.355713  4.617162  2.781357  5.628536  4.846161  2.395161  4.211129
#>  [29]  4.898421  4.946722  2.498495  3.301628  3.655911  5.057582  4.502613
#>  [36]  4.385431  5.595109  6.700213  3.539800  3.760940  5.846276  5.139868
#>  [43]  5.862777  6.040985  3.553236  6.723276  6.847264  8.003822  8.027028
#>  [50]  5.317588  5.759746  5.124828  4.500741  6.399723  6.303434  7.893596
#>  [57]  4.971552  7.774417  4.386492  6.040560  8.734975  5.779569  7.540821
#>  [64]  6.313041  7.908912  5.663752 10.197866  7.923128  8.044645  6.542624
#>  [71]  7.664627  7.304432  5.264060  8.098587  7.722208  5.454732  7.840816
#>  [78]  9.133588 11.197240  8.195310  9.252515  9.851069  7.660782  7.951250
#>  [85]  9.356700 12.517055  9.011827  7.790966 11.135915  6.868178 13.701624
#>  [92]  7.688373 10.468120 12.974039 12.148186  8.717816 11.090497 13.354167
#>  [99]  7.371633 10.803703 13.468562 13.641762 10.584499  9.236935 10.878506
#> [106] 12.930272 11.759916 13.058013 11.974180 14.644313 13.728406 15.151259
#> [113]  8.288001 15.458083 16.431496 12.480626 11.405487 14.897252 13.413848
#> [120] 13.902232 15.844015 13.501419 10.959127 16.591509 16.239930 11.117068
#> [127] 12.479825 15.640764 10.209519 10.904003 10.704884 15.005574 16.717070
#> [134] 17.378454 14.238072 14.505490 15.508042 11.754428 10.078217 17.584892
#> [141] 10.389464 18.439958 15.156113 11.392965 17.824353 10.295341 15.163674
#> [148] 12.099183 17.960036 17.234700 16.230870 12.076226 16.538584 12.104494
#> [155] 14.219133 13.164126 15.246690 12.501751 14.374680 18.364746 18.865504
#> [162] 17.534377 18.282841 15.477307 17.167150 19.122748 16.639338 12.296957
#> [169] 12.974180 16.745443 18.015540 12.509704 10.457736 18.680562 14.247866
#> [176] 16.976924 16.242914 13.958249 13.450452 16.820533 16.159570 11.082755
#> [183] 17.799234 10.976728 17.726533 11.193100 14.468834 12.374256 16.171733
#> [190] 11.594428 18.423482 12.910643 13.222803 17.748838 17.192258 10.861015
#> [197] 17.783192 16.587901 13.186884 18.449435 12.807440 17.896657 11.574198
#> [204]  9.953024 10.875899 16.878923 11.357366 11.282199  9.780279 17.398958
#> [211] 11.414602 15.882503 11.665187 16.538049 11.607700 17.296596 13.935224
#> [218] 15.273161 15.835539 15.150275 12.220086  9.851086  9.315495 13.952738
#> [225] 13.431092 13.796304 11.059213 16.227946 14.007382 15.178377 15.965160
#> [232]  9.788476 13.239980 10.980304  8.200422 12.585679 12.347656 10.769578
#> [239] 14.158235 11.440435 14.866144  7.451319 12.329184  9.957247 13.946824
#> [246] 12.049406 11.848398 12.408620 10.742504  8.440026 10.292312  9.534501
#> [253] 11.637432 10.893598  9.091832 12.914180  8.218541  6.418919  9.920639
#> [260]  7.389145  6.680911 12.003116 10.970339 10.565402  8.980558  8.964684
#> [267]  5.975692  7.744198  6.288515  9.426561  9.214561 10.755426  6.084880
#> [274] 10.628300  7.043429  8.493579  9.337245  4.928551  7.451706  8.684025
#> [281]  6.732446  7.282175  5.271031  4.251173  9.206093  4.888878  7.950784
#> [288]  8.393319  8.021624  6.735856  8.143892  3.716415  4.116851  3.864859
#> [295]  6.674516  5.200282  3.530380  3.746824  7.370884  4.019225  5.947242
#> [302]  3.826930  5.254043  6.363964  3.308032  3.806735  3.854033  2.674447
#> [309]  5.998614  4.161240  5.415742  5.209244  4.350308  6.019545  3.637961
#> [316]  2.877111  5.062895  2.278924  4.939538  2.677388  2.154412  4.721879
#> [323]  4.790974  5.197189  3.178624  2.461984  3.823960  4.424891  1.918535
#> [330]  4.089968  2.422746  1.774319  2.096332  3.768093  3.596101  2.617988
#> [337]  2.770079  3.687707  3.627289  4.261370  2.880219  3.252400  1.998150
#> [344]  3.127554  4.047395  3.171082  3.822860  3.140515  3.647651  2.575230
#> [351]  3.826030  3.519476  2.339005  1.765976  1.736555  3.676795  2.514274
#> [358]  1.608851  4.214974  3.158918  1.894508  2.999018  1.848212  4.231052
#> [365]  1.646808  1.755725  3.936917  3.289054  2.023409  1.603187  2.975135
#> [372]  3.096742  2.503919  3.870244  4.031175  4.388205  2.265290  1.854240
#> [379]  2.785909  3.084535  2.458642  3.172094  2.096030  3.212572  5.236516
#> [386]  2.970865  4.355713  4.617162  2.781357  5.628536  4.846161  2.395161
#> [393]  4.211129  4.898421  4.946722  2.498495  3.301628  3.655911  5.057582
#> [400]  4.502613  4.385431  5.595109  6.700213  3.539800  3.760940  5.846276
#> [407]  5.139868  5.862777  6.040985  3.553236  6.723276  6.847264  8.003822
#> [414]  8.027028  5.317588  5.759746  5.124828  4.500741  6.399723  6.303434
#> [421]  7.893596  4.971552  7.774417  4.386492  6.040560  8.734975  5.779569
#> [428]  7.540821  6.313041  7.908912  5.663752 10.197866  7.923128  8.044645
#> [435]  6.542624  7.664627  7.304432  5.264060  8.098587  7.722208  5.454732
#> [442]  7.840816  9.133588 11.197240  8.195310  9.252515  9.851069  7.660782
#> [449]  7.951250  9.356700 12.517055  9.011827  7.790966 11.135915  6.868178
#> [456] 13.701624  7.688373 10.468120 12.974039 12.148186  8.717816 11.090497
#> [463] 13.354167  7.371633 10.803703 13.468562 13.641762 10.584499  9.236935
#> [470] 10.878506 12.930272 11.759916 13.058013 11.974180 14.644313 13.728406
#> [477] 15.151259  8.288001 15.458083 16.431496 12.480626 11.405487 14.897252
#> [484] 13.413848 13.902232 15.844015 13.501419 10.959127 16.591509 16.239930
#> [491] 11.117068 12.479825 15.640764 10.209519 10.904003 10.704884 15.005574
#> [498] 16.717070 17.378454 14.238072 14.505490 15.508042 11.754428 10.078217
#> [505] 17.584892 10.389464 18.439958 15.156113 11.392965 17.824353 10.295341
#> [512] 15.163674 12.099183 17.960036 17.234700 16.230870 12.076226 16.538584
#> [519] 12.104494 14.219133 13.164126 15.246690 12.501751 14.374680 18.364746
#> [526] 18.865504 17.534377 18.282841 15.477307 17.167150 19.122748 16.639338
#> [533] 12.296957 12.974180 16.745443 18.015540 12.509704 10.457736 18.680562
#> [540] 14.247866 16.976924 16.242914 13.958249 13.450452 16.820533 16.159570
#> [547] 11.082755 17.799234 10.976728 17.726533 11.193100 14.468834 12.374256
#> [554] 16.171733 11.594428 18.423482 12.910643 13.222803 17.748838 17.192258
#> [561] 10.861015 17.783192 16.587901 13.186884 18.449435 12.807440 17.896657
#> [568] 11.574198  9.953024 10.875899 16.878923 11.357366 11.282199  9.780279
#> [575] 17.398958 11.414602 15.882503 11.665187 16.538049 11.607700 17.296596
#> [582] 13.935224 15.273161 15.835539 15.150275 12.220086  9.851086  9.315495
#> [589] 13.952738 13.431092 13.796304 11.059213 16.227946 14.007382 15.178377
#> [596] 15.965160  9.788476 13.239980 10.980304  8.200422 12.585679 12.347656
#> [603] 10.769578 14.158235 11.440435 14.866144  7.451319 12.329184  9.957247
#> [610] 13.946824 12.049406 11.848398 12.408620 10.742504  8.440026 10.292312
#> [617]  9.534501 11.637432 10.893598  9.091832 12.914180  8.218541  6.418919
#> [624]  9.920639  7.389145  6.680911 12.003116 10.970339 10.565402  8.980558
#> [631]  8.964684  5.975692  7.744198  6.288515  9.426561  9.214561 10.755426
#> [638]  6.084880 10.628300  7.043429  8.493579  9.337245  4.928551  7.451706
#> [645]  8.684025  6.732446  7.282175  5.271031  4.251173  9.206093  4.888878
#> [652]  7.950784  8.393319  8.021624  6.735856  8.143892  3.716415  4.116851
#> [659]  3.864859  6.674516  5.200282  3.530380  3.746824  7.370884  4.019225
#> [666]  5.947242  3.826930  5.254043  6.363964  3.308032  3.806735  3.854033
#> [673]  2.674447  5.998614  4.161240  5.415742  5.209244  4.350308  6.019545
#> [680]  3.637961  2.877111  5.062895  2.278924  4.939538  2.677388  2.154412
#> [687]  4.721879  4.790974  5.197189  3.178624  2.461984  3.823960  4.424891
#> [694]  1.918535  4.089968  2.422746  1.774319  2.096332  3.768093  3.596101
#> [701]  2.617988  2.770079  3.687707  3.627289  4.261370  2.880219  3.252400
#> [708]  1.998150  3.127554  4.047395  3.171082  3.822860  3.140515  3.647651
#> [715]  2.575230  3.826030  3.519476  2.339005  1.765976  1.736555  3.676795
#> [722]  2.514274  1.608851  4.214974  3.158918  1.894508  2.999018  1.848212
#> [729]  4.231052  1.646808
calc_globrad(dates, sunhours = runif(365, 0, 7),lat = 52.8, full_output = TRUE)
#>     doy      dates     sunhours daylength    extrat   globrad
#> 1     1 2002-01-01 1.1298642773  7.463176  6.099033  1.986430
#> 3     2 2002-01-02 6.5554119824  7.483435  6.149345  4.230717
#> 6     3 2002-01-03 6.7947695362  7.505323  6.203934  4.359277
#> 7     4 2002-01-04 4.9719183680  7.528820  6.262814  3.633637
#> 9     5 2002-01-05 6.1956365127  7.553900  6.325996  4.175760
#> 11    6 2002-01-06 6.8381272161  7.580540  6.393496  4.482043
#> 14    7 2002-01-07 0.2442975054  7.608714  6.465324  1.720124
#> 15    8 2002-01-08 3.0713008842  7.638396  6.541496  2.950500
#> 17    9 2002-01-09 4.5112883241  7.669557  6.622023  3.603066
#> 19   10 2002-01-10 6.9638268326  7.702170  6.706920  4.708721
#> 22   11 2002-01-11 2.2026383446  7.736204  6.796198  2.666550
#> 23   12 2002-01-12 5.9911737633  7.771631  6.889871  4.378179
#> 25   13 2002-01-13 3.7823637708  7.808419  6.987950  3.439454
#> 27   14 2002-01-14 6.1139214663  7.846537  7.090446  4.535004
#> 30   15 2002-01-15 3.6090993364  7.885954  7.197371  3.446323
#> 32   16 2002-01-16 6.0862150858  7.926639  7.308732  4.633071
#> 33   17 2002-01-17 5.9956581620  7.968559  7.424539  4.649300
#> 35   18 2002-01-18 2.4197822879  8.011681  7.544799  3.025584
#> 37   19 2002-01-19 0.0007258395  8.055974  7.669519  1.917725
#> 40   20 2002-01-20 1.4362582173  8.101405  7.798703  2.640972
#> 41   21 2002-01-21 6.6165434348  8.147942  7.932354  5.203826
#> 43   22 2002-01-22 1.9693345916  8.195553  8.070474  2.987258
#> 45   23 2002-01-23 6.1665395207  8.244205  8.213063  5.124888
#> 48   24 2002-01-24 6.7278139368  8.293867  8.360119  5.480808
#> 49   25 2002-01-25 0.0281291772  8.344507  8.511639  2.142256
#> 51   26 2002-01-26 3.8703014392  8.396093  8.667618  4.164637
#> 53   27 2002-01-27 1.5346518091  8.448595  8.828046  3.008800
#> 56   28 2002-01-28 4.5270900032  8.501983  8.992915  4.642478
#> 58   29 2002-01-29 5.0939143272  8.556225  9.162213  5.017897
#> 59   30 2002-01-30 6.1081120968  8.611293  9.335925  5.645033
#> 61   31 2002-01-31 2.6738708224  8.667157  9.514035  3.846078
#> 63   32 2002-02-01 6.2492199291  8.723789  9.696523  5.897146
#> 66   33 2002-02-02 5.9070919997  8.781160  9.883368  5.795116
#> 67   34 2002-02-03 5.1083284605  8.839243 10.074546  5.429751
#> 69   35 2002-02-04 0.2689048359  8.898011 10.270030  2.722692
#> 71   36 2002-02-05 2.3367632220  8.957438 10.469790  3.983096
#> 74   37 2002-02-06 5.2449085654  9.017498 10.673795  5.772585
#> 75   38 2002-02-07 6.3921060748  9.078165 10.882010  6.551616
#> 77   39 2002-02-08 1.4303034407  9.139416 11.094398  3.641727
#> 79   40 2002-02-09 3.3022892803  9.201226 11.310918  4.857455
#> 82   41 2002-02-10 2.6962817025  9.263572 11.531528  4.561082
#> 83   42 2002-02-11 2.7492635022  9.326432 11.756182  4.671800
#> 86   43 2002-02-12 1.5283690128  9.389784 11.984831  3.971589
#> 87   44 2002-02-13 0.0504625840  9.453606 12.217425  3.086964
#> 89   45 2002-02-14 2.5459850021  9.517878 12.453910  4.779157
#> 91   46 2002-02-15 3.6827471091  9.582579 12.694229  5.612861
#> 94   47 2002-02-16 5.9091437536  9.647690 12.938322  7.196897
#> 95   48 2002-02-17 0.9519381954  9.713192 13.186128  3.942683
#> 97   49 2002-02-18 2.6352713159  9.779067 13.437583  5.169981
#> 99   50 2002-02-19 1.8431605394  9.845297 13.692618  4.704868
#> 102  51 2002-02-20 3.4849707878  9.911864 13.951164  5.940377
#> 103  52 2002-02-21 0.1366767546  9.978754 14.213148  3.650624
#> 105  53 2002-02-22 0.6102866689 10.045948 14.478496  4.059405
#> 107  54 2002-02-23 6.9717386514 10.113433 14.747130  8.769782
#> 110  55 2002-02-24 6.5949254383 10.181192 15.018971  8.619055
#> 112  56 2002-02-25 0.8706141578 10.249211 15.293936  4.473052
#> 113  57 2002-02-26 0.8063251157 10.317477 15.571940  4.501469
#> 115  58 2002-02-27 2.5668167255 10.385975 15.852898  5.922188
#> 117  59 2002-02-28 3.2972802492 10.454693 16.136720  6.578840
#> 120  60 2002-03-01 5.4892071187 10.523618 16.423316  8.389098
#> 121  61 2002-03-02 3.4327496893 10.592738 16.712592  6.886142
#> 123  62 2002-03-03 0.7136566842 10.662040 17.004454  4.820204
#> 125  63 2002-03-04 6.6042180702 10.731513 17.298805  9.647579
#> 128  64 2002-03-05 4.9429765164 10.801147 17.595546  8.425050
#> 129  65 2002-03-06 3.5158496560 10.870930 17.894578  7.367355
#> 131  66 2002-03-07 3.7095434354 10.940852 18.195799  7.633632
#> 133  67 2002-03-08 1.8089844391 11.010902 18.499104  6.144388
#> 136  68 2002-03-09 4.9822739193 11.081072 18.804391  8.928515
#> 138  69 2002-03-10 6.7792378489 11.151350 19.111552 10.587129
#> 139  70 2002-03-11 4.6275445281 11.221729 19.420480  8.859367
#> 141  71 2002-03-12 5.5737541325 11.292198 19.731067  9.802329
#> 143  72 2002-03-13 0.5124639245 11.362749 20.043203  5.462779
#> 146  73 2002-03-14 6.1932001910 11.433374 20.356778 10.602598
#> 147  74 2002-03-15 5.7869026377 11.504062 20.671679 10.367170
#> 149  75 2002-03-16 5.0803477154 11.574807 20.987796  9.852870
#> 151  76 2002-03-17 2.5713031518 11.645600 21.305014  7.678285
#> 154  77 2002-03-18 3.6639989160 11.716433 21.623220  8.786845
#> 155  78 2002-03-19 3.2891523887 11.787297 21.942300  8.546988
#> 157  79 2002-03-20 0.6899646998 11.858185 22.262140  6.213193
#> 159  80 2002-03-21 3.6122443716 11.929088 22.582624  9.064775
#> 162  81 2002-03-22 3.1481367925 12.000000 22.903638  8.730234
#> 163  82 2002-03-23 3.8462308343 12.070912 23.225065  9.506441
#> 166  83 2002-03-24 4.5465433290 12.141815 23.546791 10.295285
#> 167  84 2002-03-25 1.1132437827 12.212703 23.868700  7.055046
#> 169  85 2002-03-26 2.0562555443 12.283567 24.190677  8.072416
#> 171  86 2002-03-27 6.2389940687 12.354400 24.512606 12.317607
#> 174  87 2002-03-28 4.1165503818 12.425193 24.834373 10.322491
#> 175  88 2002-03-29 4.4326209428 12.495938 25.155863 10.750672
#> 177  89 2002-03-30 1.8165663346 12.566626 25.476961  8.210649
#> 179  90 2002-03-31 2.2274890896 12.637251 25.797553  8.722975
#> 182  91 2002-04-01 2.3747771671 12.707802 26.117528  8.969745
#> 183  92 2002-04-02 5.3253279696 12.778271 26.436771 12.117938
#> 185  93 2002-04-03 0.0761072491 12.848650 26.755170  6.768033
#> 187  94 2002-04-04 0.4419613727 12.918928 27.072615  7.231236
#> 190  95 2002-04-05 1.8468925457 12.989098 27.388994  8.794440
#> 192  96 2002-04-06 4.1568561199 13.059148 27.704199 11.335310
#> 193  97 2002-04-07 2.0829202044 13.129070 28.018120  9.227060
#> 195  98 2002-04-08 0.6623594053 13.198853 28.330649  7.793522
#> 197  99 2002-04-09 5.3287875215 13.268487 28.641680 12.911845
#> 200 100 2002-04-10 5.4073379326 13.337960 28.951106 13.106305
#> 201 101 2002-04-11 2.8139327522 13.407262 29.258822 10.385144
#> 203 102 2002-04-12 0.4135806712 13.476382 29.564726  7.844842
#> 205 103 2002-04-13 5.2517351902 13.545307 29.868714 13.257471
#> 208 104 2002-04-14 3.6982564875 13.614025 30.170685 11.640612
#> 209 105 2002-04-15 1.9306358229 13.682523 30.470539  9.767367
#> 211 106 2002-04-16 1.8286628504 13.750789 30.768176  9.737913
#> 213 107 2002-04-17 3.8068181823 13.818808 31.063500 12.044576
#> 216 108 2002-04-18 1.3930054551 13.886567 31.356413  9.411834
#> 218 109 2002-04-19 4.5153834075 13.954052 31.646820 13.031993
#> 219 110 2002-04-20 5.2516789224 14.021246 31.934628 13.964238
#> 221 111 2002-04-21 1.8968448413 14.088136 32.219743 10.223990
#> 223 112 2002-04-22 6.1293533649 14.154703 32.502075 15.162639
#> 226 113 2002-04-23 5.9622022181 14.220933 32.781534 15.067300
#> 227 114 2002-04-24 0.3209451749 14.286808 33.058030  8.635823
#> 229 115 2002-04-25 2.6673980125 14.352310 33.331477 11.430221
#> 231 116 2002-04-26 1.2831843798 14.417421 33.601789  9.895766
#> 234 117 2002-04-27 6.4525974893 14.482122 33.868881 16.012463
#> 235 118 2002-04-28 6.3433079422 14.546394 34.132670 15.975357
#> 237 119 2002-04-29 1.3506032247 14.610216 34.393074 10.187957
#> 239 120 2002-04-30 6.0506269371 14.673568 34.650012 15.806447
#> 242 121 2002-05-01 0.0929082131 14.736428 34.903405  8.835878
#> 243 122 2002-05-02 6.1996223773 14.798774 35.153175 16.151620
#> 245 123 2002-05-03 2.7407947497 14.860584 35.399245 12.114221
#> 248 124 2002-05-04 0.5332341893 14.921835 35.641540  9.547213
#> 249 125 2002-05-05 4.5274691340 14.982502 35.879987 14.391172
#> 251 126 2002-05-06 6.4165772784 15.042562 36.114511 16.731157
#> 253 127 2002-05-07 5.8415603414 15.101989 36.345041 16.115525
#> 256 128 2002-05-08 3.5087294918 15.160757 36.571508 13.374840
#> 257 129 2002-05-09 4.2150133890 15.218840 36.793843 14.293676
#> 259 130 2002-05-10 3.2827132393 15.276211 37.011977 13.229756
#> 261 131 2002-05-11 6.2046196673 15.332843 37.225844 16.838404
#> 264 132 2002-05-12 4.6938394464 15.388707 37.435378 15.068085
#> 265 133 2002-05-13 2.7564440363 15.443775 37.640517 12.769217
#> 267 134 2002-05-14 3.8590347059 15.498017 37.841197 14.171563
#> 269 135 2002-05-15 1.6974487305 15.551405 38.037357 11.585244
#> 272 136 2002-05-16 1.4224618033 15.603907 38.228936 11.299721
#> 274 137 2002-05-17 0.4783919237 15.655493 38.415875 10.190914
#> 275 138 2002-05-18 2.1509161601 15.706133 38.598118 12.292488
#> 277 139 2002-05-19 6.9518266600 15.755795 38.775606 18.248256
#> 279 140 2002-05-20 0.8141642399 15.804447 38.948286 10.740280
#> 282 141 2002-05-21 5.0079615365 15.852058 39.116103 15.957780
#> 283 142 2002-05-22 1.5827611203 15.898595 39.279005 11.774933
#> 285 143 2002-05-23 0.9334398333 15.944026 39.436940 11.013649
#> 287 144 2002-05-24 6.4909723308 15.988319 39.589859 17.933853
#> 290 145 2002-05-25 6.2588585999 16.031441 39.737713 17.691458
#> 291 146 2002-05-26 1.4256507456 16.073361 39.880455 11.738742
#> 293 147 2002-05-27 1.7988804334 16.114046 40.018039 12.238203
#> 295 148 2002-05-28 4.2980392252 16.153463 40.150421 15.379125
#> 298 149 2002-05-29 3.0904943133 16.191581 40.277558 13.913287
#> 300 150 2002-05-30 2.2091627968 16.228369 40.399408 12.849631
#> 301 151 2002-05-31 0.7091963268 16.263796 40.515933 11.012349
#> 303 152 2002-06-01 1.9085233700 16.297830 40.627093 12.535549
#> 305 153 2002-06-02 4.5758957798 16.330443 40.732853 15.890018
#> 308 154 2002-06-03 6.4950907156 16.361604 40.833176 18.313098
#> 309 155 2002-06-04 0.1863726901 16.391286 40.928031 10.464688
#> 311 156 2002-06-05 3.9164383798 16.419460 41.017385 15.146166
#> 313 157 2002-06-06 5.9792155649 16.446100 41.101209 17.746770
#> 316 158 2002-06-07 0.5311813466 16.471180 41.179475 10.958870
#> 317 159 2002-06-08 1.4408019935 16.494677 41.252156 12.114717
#> 319 160 2002-06-09 3.4712275269 16.516565 41.319229 14.671765
#> 321 161 2002-06-10 1.8329986085 16.536824 41.380671 12.638556
#> 324 162 2002-06-11 2.9117301547 16.555433 41.436461 14.002988
#> 325 163 2002-06-12 3.0740499967 16.572373 41.486582 14.219369
#> 328 164 2002-06-13 1.2766454353 16.587625 41.531016 11.980945
#> 329 165 2002-06-14 0.5482978623 16.601175 41.569750 11.078913
#> 331 166 2002-06-15 0.1243896922 16.613007 41.602770 10.556443
#> 333 167 2002-06-16 5.0972598132 16.623108 41.630068 16.790176
#> 336 168 2002-06-17 6.7959151736 16.631469 41.651634 18.922709
#> 337 169 2002-06-18 3.7465231165 16.638080 41.667464 15.108156
#> 339 170 2002-06-19 1.1654353749 16.642934 41.677552 11.878641
#> 341 171 2002-06-20 0.4072295546 16.646025 41.681897 10.930329
#> 344 172 2002-06-21 3.2139731331 16.647351 41.680500 14.443588
#> 345 173 2002-06-22 4.3595917500 16.646909 41.673363 15.875175
#> 347 174 2002-06-23 4.1637344097 16.644700 41.660491 15.625887
#> 349 175 2002-06-24 5.2011276619 16.640727 41.641891 16.918145
#> 352 176 2002-06-25 3.6699609405 16.634994 41.617572 14.995163
#> 354 177 2002-06-26 3.4777767523 16.627507 41.587545 14.746071
#> 355 178 2002-06-27 0.1030316185 16.618275 41.551823 10.516764
#> 357 179 2002-06-28 6.7054338716 16.607306 41.510423 18.757815
#> 359 180 2002-06-29 3.1716246458 16.594614 41.463360 14.328156
#> 362 181 2002-06-30 6.8707229439 16.580211 41.410655 18.932806
#> 363 182 2002-07-01 3.8129361621 16.564113 41.352328 15.097582
#> 365 183 2002-07-02 2.5202968079 16.546336 41.288405 13.466575
#> 367 184 2002-07-03 3.5220227316 16.526900 41.218909 14.696777
#> 370 185 2002-07-04 0.5224354262 16.505823 41.143868 10.937101
#> 371 186 2002-07-05 6.7885432325 16.483128 41.063311 18.721749
#> 373 187 2002-07-06 4.5395091963 16.458837 40.977269 15.895284
#> 375 188 2002-07-07 6.8919582665 16.432973 40.885775 18.795153
#> 378 189 2002-07-08 3.7529327988 16.405563 40.788864 14.862641
#> 380 190 2002-07-09 5.3274158353 16.376632 40.686570 16.789434
#> 381 191 2002-07-10 3.9507000893 16.346207 40.578932 15.048476
#> 383 192 2002-07-11 5.4491801246 16.314316 40.465990 16.874564
#> 385 193 2002-07-12 4.4734269199 16.280989 40.347784 15.630002
#> 388 194 2002-07-13 2.6527251494 16.246255 40.224356 13.340051
#> 389 195 2002-07-14 2.5167146940 16.210144 40.095751 13.136482
#> 391 196 2002-07-15 5.9096199137 16.172686 39.962013 17.291711
#> 393 197 2002-07-16 3.1972442656 16.133915 39.823189 13.901661
#> 396 198 2002-07-17 1.2532400342 16.093860 39.679327 11.464760
#> 397 199 2002-07-18 6.7634643705 16.052554 39.530476 18.210359
#> 399 200 2002-07-19 4.9938836617 16.010029 39.376687 15.985404
#> 401 201 2002-07-20 0.1742022182 15.966317 39.218012 10.018449
#> 404 202 2002-07-21 4.3851954595 15.921451 39.054503 15.141956
#> 405 203 2002-07-22 5.7954492546 15.875462 38.886214 16.819396
#> 408 204 2002-07-23 4.5014832215 15.828384 38.713200 15.183183
#> 409 205 2002-07-24 4.9593898323 15.780249 38.535519 15.689318
#> 411 206 2002-07-25 6.0950900745 15.731088 38.353226 17.018383
#> 413 207 2002-07-26 1.9726880719 15.680933 38.166382 11.942293
#> 416 208 2002-07-27 1.2125738936 15.629816 37.975044 10.966828
#> 417 209 2002-07-28 0.7676901331 15.577768 37.779274 10.375721
#> 419 210 2002-07-29 4.1035553410 15.524820 37.579134 14.361284
#> 421 211 2002-07-30 5.9550492915 15.471001 37.374685 16.536745
#> 424 212 2002-07-31 6.4506832347 15.416342 37.165992 17.067209
#> 425 213 2002-08-01 4.7296333136 15.360873 36.953119 14.927237
#> 427 214 2002-08-02 2.4541171370 15.304621 36.736132 12.129378
#> 429 215 2002-08-03 0.0606065234 15.247616 36.515097  9.201345
#> 432 216 2002-08-04 3.0162915480 15.189886 36.290082 12.675624
#> 434 217 2002-08-05 5.0745456801 15.131457 36.061156 15.062095
#> 435 218 2002-08-06 2.8857947094 15.072356 35.828388 12.386998
#> 437 219 2002-08-07 3.8560367664 15.012610 35.591848 13.468902
#> 439 220 2002-08-08 1.0945239908 14.952243 35.351609 10.131795
#> 442 221 2002-08-09 0.3010090524 14.891281 35.107744  9.131766
#> 443 222 2002-08-10 4.7925087914 14.829748 34.860325 14.347963
#> 445 223 2002-08-11 3.6008174797 14.767666 34.609429 12.871786
#> 447 224 2002-08-12 6.6198149386 14.705060 34.355129 16.321651
#> 450 225 2002-08-13 5.9016785952 14.641952 34.097505 15.396155
#> 451 226 2002-08-14 0.4541816534 14.578362 33.836632  8.986240
#> 453 227 2002-08-15 5.3347764020 14.514313 33.572591 14.562998
#> 455 228 2002-08-16 0.5486897659 14.449824 33.305462  8.958704
#> 458 229 2002-08-17 0.9345536388 14.384916 33.035326  9.331944
#> 460 230 2002-08-18 5.6678734070 14.319607 32.762264 14.674416
#> 461 231 2002-08-19 3.9156826755 14.253916 32.486361 12.583742
#> 463 232 2002-08-20 5.1361078029 14.187862 32.207701 13.881635
#> 465 233 2002-08-21 2.4478683304 14.121461 31.926370 10.748713
#> 468 234 2002-08-22 6.8786806664 14.054730 31.642455 15.653855
#> 469 235 2002-08-23 2.8619771593 13.987686 31.356043 11.046842
#> 471 236 2002-08-24 2.5261158552 13.920345 31.067223 10.585680
#> 473 237 2002-08-25 4.5675702749 13.852721 30.776085 12.767823
#> 476 238 2002-08-26 1.0049145808 13.784830 30.482721  8.731776
#> 477 239 2002-08-27 4.7655903073 13.716686 30.187223 12.790782
#> 479 240 2002-08-28 6.0150012553 13.648302 29.889684 14.058826
#> 481 241 2002-08-29 5.8704099276 13.579692 29.590199 13.793372
#> 484 242 2002-08-30 6.1691904743 13.510869 29.288864 14.009002
#> 485 243 2002-08-31 1.6638841627 13.441846 28.985774  9.040430
#> 487 244 2002-09-01 4.0504159203 13.372633 28.681027 11.513832
#> 490 245 2002-09-02 6.2853175399 13.303244 28.374722 13.796712
#> 491 246 2002-09-03 2.4110032138 13.233689 28.066958  9.573454
#> 493 247 2002-09-04 5.3260076088 13.163980 27.757836 12.554723
#> 495 248 2002-09-05 5.1177255162 13.094126 27.447457 12.225665
#> 498 249 2002-09-06 5.9670266353 13.024138 27.135924 13.000161
#> 499 250 2002-09-07 6.8307262855 12.954027 26.823338 13.777878
#> 501 251 2002-09-08 0.7913325059 12.883802 26.509805  7.441577
#> 503 252 2002-09-09 6.7918359581 12.813472 26.195428 13.491357
#> 506 253 2002-09-10 4.5365794727 12.743047 25.880314 11.076830
#> 507 254 2002-09-11 0.3925113028 12.672536 25.564567  6.787052
#> 509 255 2002-09-12 3.3119596706 12.601947 25.248295  9.629868
#> 511 256 2002-09-13 2.0623750919 12.531289 24.931604  8.284498
#> 514 257 2002-09-14 4.2720973536 12.460572 24.614602 10.373199
#> 516 258 2002-09-15 0.8477621619 12.389802 24.297397  6.905614
#> 517 259 2002-09-16 4.4059155553 12.318988 23.980097 10.283294
#> 519 260 2002-09-17 4.9836871200 12.248139 23.662812 10.729824
#> 521 261 2002-09-18 4.2845795064 12.177262 23.345649  9.943505
#> 524 262 2002-09-19 0.2410046314 12.106365 23.028718  5.986399
#> 525 263 2002-09-20 4.6612355092 12.035456 22.712128 10.076145
#> 527 264 2002-09-21 4.5707709356 11.964544 22.395989  9.876926
#> 529 265 2002-09-22 3.1630090780 11.893635 22.080409  8.456149
#> 532 266 2002-09-23 3.6176287131 11.822738 21.765497  8.771376
#> 533 267 2002-09-24 4.7362755802 11.751861 21.451363  9.685542
#> 535 268 2002-09-25 6.1665316585 11.681012 21.138114 10.864047
#> 537 269 2002-09-26 5.1186040824 11.610198 20.825859  9.797227
#> 540 270 2002-09-27 2.7903387067 11.539428 20.514705  7.608997
#> 542 271 2002-09-28 3.0136445761 11.468711 20.204761  7.705803
#> 543 272 2002-09-29 1.0178514274 11.398053 19.896131  5.862400
#> 545 273 2002-09-30 3.1578423053 11.327464 19.588923  7.627706
#> 547 274 2002-10-01 4.0514704413 11.256953 19.283241  8.290909
#> 550 275 2002-10-02 0.4929198441 11.186528 18.979189  5.162944
#> 551 276 2002-10-03 5.1962832615 11.116198 18.676871  9.034483
#> 553 277 2002-10-04 3.8445449444 11.045973 18.376389  7.792043
#> 555 278 2002-10-05 4.4418099055 10.975862 18.077844  8.177413
#> 558 279 2002-10-06 0.9564000864 10.905874 17.781337  5.225009
#> 559 280 2002-10-07 6.5116344769 10.836020 17.486965  9.625917
#> 561 281 2002-10-08 4.5578304774 10.766311 17.194826  7.938352
#> 563 282 2002-10-09 3.7009649067 10.696756 16.905016  7.150733
#> 566 283 2002-10-10 1.5758077204 10.627367 16.617629  5.386424
#> 567 284 2002-10-11 6.8828680869 10.558154 16.332759  9.406858
#> 570 285 2002-10-12 0.7633414995 10.489131 16.050495  4.596657
#> 571 286 2002-10-13 0.4896186206 10.420308 15.770928  4.313246
#> 573 287 2002-10-14 5.4435185483 10.351698 15.494145  7.947393
#> 575 288 2002-10-15 4.6082556229 10.283314 15.220231  7.215375
#> 578 289 2002-10-16 1.8144192055 10.215170 14.949271  5.064963
#> 579 290 2002-10-17 6.4283649134 10.147279 14.681345  8.320699
#> 581 291 2002-10-18 1.0409724275 10.079655 14.416534  4.348564
#> 583 292 2002-10-19 4.4847633946 10.012314 14.154915  6.708897
#> 586 293 2002-10-20 1.4086689684  9.945270 13.896562  4.458310
#> 587 294 2002-10-21 5.2053492058  9.878539 13.641550  7.004493
#> 589 295 2002-10-22 5.9334619991  9.812138 13.389948  7.395980
#> 591 296 2002-10-23 5.9754511106  9.746084 13.141825  7.314169
#> 594 297 2002-10-24 5.6238880090  9.680393 12.897247  6.970682
#> 596 298 2002-10-25 0.1671373537  9.615084 12.656277  3.274070
#> 597 299 2002-10-26 5.5676440035  9.550176 12.418977  6.724805
#> 599 300 2002-10-27 6.2058535737  9.485687 12.185404  7.032400
#> 601 301 2002-10-28 4.2002613691  9.421638 11.955615  5.653871
#> 604 302 2002-10-29 3.0295171652  9.358048 11.729663  4.831060
#> 605 303 2002-10-30 6.4419306223  9.294940 11.507600  6.864616
#> 607 304 2002-10-31 4.3121009958  9.232334 11.289473  5.458828
#> 609 305 2002-11-01 3.0500271111  9.170252 11.075328  4.610660
#> 612 306 2002-11-02 6.2811308708  9.108719 10.865209  6.462482
#> 613 307 2002-11-03 5.3830946553  9.047757 10.659156  5.835699
#> 615 308 2002-11-04 0.8968103081  8.987390 10.457208  3.136040
#> 617 309 2002-11-05 3.7704318434  8.927644 10.259398  4.731287
#> 620 310 2002-11-06 0.9031135815  8.868543 10.065762  3.028956
#> 622 311 2002-11-07 3.5767541877  8.810114  9.876328  4.473891
#> 623 312 2002-11-08 3.8065488376  8.752384  9.691126  4.530193
#> 625 313 2002-11-09 1.6924326292  8.695379  9.510180  3.303056
#> 627 314 2002-11-10 5.6220088068  8.639127  9.333513  5.370323
#> 630 315 2002-11-11 6.6353960712  8.583658  9.161147  5.831192
#> 631 316 2002-11-12 2.8515294176  8.528999  8.993100  3.751622
#> 633 317 2002-11-13 3.6032119396  8.475180  8.829388  4.084249
#> 635 318 2002-11-14 6.8707929526  8.422232  8.670025  5.703977
#> 638 319 2002-11-15 2.4793070969  8.370184  8.515023  3.389861
#> 639 320 2002-11-16 3.1908810346  8.319067  8.364392  3.695231
#> 641 321 2002-11-17 1.5401729443  8.268912  8.218139  2.819893
#> 643 322 2002-11-18 0.5728311650  8.219751  8.076271  2.300484
#> 646 323 2002-11-19 2.5846524800  8.171616  7.938791  3.240203
#> 647 324 2002-11-20 0.8866660867  8.124538  7.805702  2.377361
#> 650 325 2002-11-21 5.1791075380  8.078549  7.677004  4.380091
#> 651 326 2002-11-22 6.6910640008  8.033683  7.552698  5.033406
#> 653 327 2002-11-23 4.6807215177  7.989971  7.432780  4.035348
#> 655 328 2002-11-24 3.2766254421  7.947446  7.317248  3.337713
#> 658 329 2002-11-25 3.7673471260  7.906140  7.206095  3.518409
#> 659 330 2002-11-26 2.3166759131  7.866085  7.099317  2.820255
#> 661 331 2002-11-27 4.4007747553  7.827314  6.996907  3.716173
#> 663 332 2002-11-28 3.1678673008  7.789856  6.898857  3.127478
#> 666 333 2002-11-29 6.2161062430  7.753745  6.805157  4.429105
#> 667 334 2002-11-30 5.4894127378  7.719011  6.715800  4.066937
#> 669 335 2002-12-01 0.9664168563  7.685684  6.630775  2.074579
#> 671 336 2002-12-02 1.9691914173  7.653793  6.550072  2.480129
#> 674 337 2002-12-03 4.9486254861  7.623368  6.473680  3.719579
#> 676 338 2002-12-04 3.0953084414  7.594437  6.401590  2.904964
#> 677 339 2002-12-05 4.7438896613  7.567027  6.333789  3.568824
#> 679 340 2002-12-06 4.9050048993  7.541163  6.270268  3.606755
#> 681 341 2002-12-07 0.4788449355  7.516872  6.211016  1.750583
#> 684 342 2002-12-08 6.7095246441  7.494177  6.156023  4.294745
#> 685 343 2002-12-09 1.7113846468  7.473100  6.105278  2.225392
#> 687 344 2002-12-10 6.4839833423  7.453664  6.058772  4.149973
#> 689 345 2002-12-11 3.5192295376  7.435887  6.016497  2.927857
#> 692 346 2002-12-12 6.5486199944  7.419789  5.978443  4.132863
#> 693 347 2002-12-13 3.0715974274  7.405386  5.944603  2.718998
#> 695 348 2002-12-14 2.4534057572  7.392694  5.914970  2.460240
#> 697 349 2002-12-15 3.3355286187  7.381725  5.889539  2.803017
#> 700 350 2002-12-16 6.7631166757  7.372493  5.868303  4.158704
#> 702 351 2002-12-17 2.0801334889  7.365006  5.851259  2.289114
#> 703 352 2002-12-18 2.0345381584  7.359273  5.838404  2.266641
#> 705 353 2002-12-19 6.0180557331  7.355300  5.829736  3.842359
#> 707 354 2002-12-20 1.3531886679  7.353091  5.825253  1.992324
#> 710 355 2002-12-21 5.6637585997  7.352649  5.824956  3.699726
#> 711 356 2002-12-22 2.5584055870  7.353975  5.828845  2.471122
#> 713 357 2002-12-23 1.7688975721  7.357066  5.836924  2.160932
#> 715 358 2002-12-24 5.9925670966  7.361920  5.849194  3.842907
#> 718 359 2002-12-25 0.9101705870  7.368531  5.865662  1.828683
#> 719 360 2002-12-26 5.5433536230  7.376892  5.886331  3.683220
#> 721 361 2002-12-27 2.4913105716  7.386993  5.911209  2.474599
#> 723 362 2002-12-28 6.7504068639  7.398825  5.940303  4.194929
#> 726 363 2002-12-29 0.8307824642  7.412375  5.973622  1.828169
#> 727 364 2002-12-30 3.2071245466  7.427627  6.011174  2.800556
#> 729 365 2002-12-31 3.6655530566  7.444567  6.052971  3.003422
#> 2     1 2003-01-01 1.1298642773  7.463176  6.099033  1.986430
#> 4     2 2003-01-02 6.5554119824  7.483435  6.149345  4.230717
#> 5     3 2003-01-03 6.7947695362  7.505323  6.203934  4.359277
#> 8     4 2003-01-04 4.9719183680  7.528820  6.262814  3.633637
#> 10    5 2003-01-05 6.1956365127  7.553900  6.325996  4.175760
#> 12    6 2003-01-06 6.8381272161  7.580540  6.393496  4.482043
#> 13    7 2003-01-07 0.2442975054  7.608714  6.465324  1.720124
#> 16    8 2003-01-08 3.0713008842  7.638396  6.541496  2.950500
#> 18    9 2003-01-09 4.5112883241  7.669557  6.622023  3.603066
#> 20   10 2003-01-10 6.9638268326  7.702170  6.706920  4.708721
#> 21   11 2003-01-11 2.2026383446  7.736204  6.796198  2.666550
#> 24   12 2003-01-12 5.9911737633  7.771631  6.889871  4.378179
#> 26   13 2003-01-13 3.7823637708  7.808419  6.987950  3.439454
#> 28   14 2003-01-14 6.1139214663  7.846537  7.090446  4.535004
#> 29   15 2003-01-15 3.6090993364  7.885954  7.197371  3.446323
#> 31   16 2003-01-16 6.0862150858  7.926639  7.308732  4.633071
#> 34   17 2003-01-17 5.9956581620  7.968559  7.424539  4.649300
#> 36   18 2003-01-18 2.4197822879  8.011681  7.544799  3.025584
#> 38   19 2003-01-19 0.0007258395  8.055974  7.669519  1.917725
#> 39   20 2003-01-20 1.4362582173  8.101405  7.798703  2.640972
#> 42   21 2003-01-21 6.6165434348  8.147942  7.932354  5.203826
#> 44   22 2003-01-22 1.9693345916  8.195553  8.070474  2.987258
#> 46   23 2003-01-23 6.1665395207  8.244205  8.213063  5.124888
#> 47   24 2003-01-24 6.7278139368  8.293867  8.360119  5.480808
#> 50   25 2003-01-25 0.0281291772  8.344507  8.511639  2.142256
#> 52   26 2003-01-26 3.8703014392  8.396093  8.667618  4.164637
#> 54   27 2003-01-27 1.5346518091  8.448595  8.828046  3.008800
#> 55   28 2003-01-28 4.5270900032  8.501983  8.992915  4.642478
#> 57   29 2003-01-29 5.0939143272  8.556225  9.162213  5.017897
#> 60   30 2003-01-30 6.1081120968  8.611293  9.335925  5.645033
#> 62   31 2003-01-31 2.6738708224  8.667157  9.514035  3.846078
#> 64   32 2003-02-01 6.2492199291  8.723789  9.696523  5.897146
#> 65   33 2003-02-02 5.9070919997  8.781160  9.883368  5.795116
#> 68   34 2003-02-03 5.1083284605  8.839243 10.074546  5.429751
#> 70   35 2003-02-04 0.2689048359  8.898011 10.270030  2.722692
#> 72   36 2003-02-05 2.3367632220  8.957438 10.469790  3.983096
#> 73   37 2003-02-06 5.2449085654  9.017498 10.673795  5.772585
#> 76   38 2003-02-07 6.3921060748  9.078165 10.882010  6.551616
#> 78   39 2003-02-08 1.4303034407  9.139416 11.094398  3.641727
#> 80   40 2003-02-09 3.3022892803  9.201226 11.310918  4.857455
#> 81   41 2003-02-10 2.6962817025  9.263572 11.531528  4.561082
#> 84   42 2003-02-11 2.7492635022  9.326432 11.756182  4.671800
#> 85   43 2003-02-12 1.5283690128  9.389784 11.984831  3.971589
#> 88   44 2003-02-13 0.0504625840  9.453606 12.217425  3.086964
#> 90   45 2003-02-14 2.5459850021  9.517878 12.453910  4.779157
#> 92   46 2003-02-15 3.6827471091  9.582579 12.694229  5.612861
#> 93   47 2003-02-16 5.9091437536  9.647690 12.938322  7.196897
#> 96   48 2003-02-17 0.9519381954  9.713192 13.186128  3.942683
#> 98   49 2003-02-18 2.6352713159  9.779067 13.437583  5.169981
#> 100  50 2003-02-19 1.8431605394  9.845297 13.692618  4.704868
#> 101  51 2003-02-20 3.4849707878  9.911864 13.951164  5.940377
#> 104  52 2003-02-21 0.1366767546  9.978754 14.213148  3.650624
#> 106  53 2003-02-22 0.6102866689 10.045948 14.478496  4.059405
#> 108  54 2003-02-23 6.9717386514 10.113433 14.747130  8.769782
#> 109  55 2003-02-24 6.5949254383 10.181192 15.018971  8.619055
#> 111  56 2003-02-25 0.8706141578 10.249211 15.293936  4.473052
#> 114  57 2003-02-26 0.8063251157 10.317477 15.571940  4.501469
#> 116  58 2003-02-27 2.5668167255 10.385975 15.852898  5.922188
#> 118  59 2003-02-28 3.2972802492 10.454693 16.136720  6.578840
#> 119  60 2003-03-01 5.4892071187 10.523618 16.423316  8.389098
#> 122  61 2003-03-02 3.4327496893 10.592738 16.712592  6.886142
#> 124  62 2003-03-03 0.7136566842 10.662040 17.004454  4.820204
#> 126  63 2003-03-04 6.6042180702 10.731513 17.298805  9.647579
#> 127  64 2003-03-05 4.9429765164 10.801147 17.595546  8.425050
#> 130  65 2003-03-06 3.5158496560 10.870930 17.894578  7.367355
#> 132  66 2003-03-07 3.7095434354 10.940852 18.195799  7.633632
#> 134  67 2003-03-08 1.8089844391 11.010902 18.499104  6.144388
#> 135  68 2003-03-09 4.9822739193 11.081072 18.804391  8.928515
#> 137  69 2003-03-10 6.7792378489 11.151350 19.111552 10.587129
#> 140  70 2003-03-11 4.6275445281 11.221729 19.420480  8.859367
#> 142  71 2003-03-12 5.5737541325 11.292198 19.731067  9.802329
#> 144  72 2003-03-13 0.5124639245 11.362749 20.043203  5.462779
#> 145  73 2003-03-14 6.1932001910 11.433374 20.356778 10.602598
#> 148  74 2003-03-15 5.7869026377 11.504062 20.671679 10.367170
#> 150  75 2003-03-16 5.0803477154 11.574807 20.987796  9.852870
#> 152  76 2003-03-17 2.5713031518 11.645600 21.305014  7.678285
#> 153  77 2003-03-18 3.6639989160 11.716433 21.623220  8.786845
#> 156  78 2003-03-19 3.2891523887 11.787297 21.942300  8.546988
#> 158  79 2003-03-20 0.6899646998 11.858185 22.262140  6.213193
#> 160  80 2003-03-21 3.6122443716 11.929088 22.582624  9.064775
#> 161  81 2003-03-22 3.1481367925 12.000000 22.903638  8.730234
#> 164  82 2003-03-23 3.8462308343 12.070912 23.225065  9.506441
#> 165  83 2003-03-24 4.5465433290 12.141815 23.546791 10.295285
#> 168  84 2003-03-25 1.1132437827 12.212703 23.868700  7.055046
#> 170  85 2003-03-26 2.0562555443 12.283567 24.190677  8.072416
#> 172  86 2003-03-27 6.2389940687 12.354400 24.512606 12.317607
#> 173  87 2003-03-28 4.1165503818 12.425193 24.834373 10.322491
#> 176  88 2003-03-29 4.4326209428 12.495938 25.155863 10.750672
#> 178  89 2003-03-30 1.8165663346 12.566626 25.476961  8.210649
#> 180  90 2003-03-31 2.2274890896 12.637251 25.797553  8.722975
#> 181  91 2003-04-01 2.3747771671 12.707802 26.117528  8.969745
#> 184  92 2003-04-02 5.3253279696 12.778271 26.436771 12.117938
#> 186  93 2003-04-03 0.0761072491 12.848650 26.755170  6.768033
#> 188  94 2003-04-04 0.4419613727 12.918928 27.072615  7.231236
#> 189  95 2003-04-05 1.8468925457 12.989098 27.388994  8.794440
#> 191  96 2003-04-06 4.1568561199 13.059148 27.704199 11.335310
#> 194  97 2003-04-07 2.0829202044 13.129070 28.018120  9.227060
#> 196  98 2003-04-08 0.6623594053 13.198853 28.330649  7.793522
#> 198  99 2003-04-09 5.3287875215 13.268487 28.641680 12.911845
#> 199 100 2003-04-10 5.4073379326 13.337960 28.951106 13.106305
#> 202 101 2003-04-11 2.8139327522 13.407262 29.258822 10.385144
#> 204 102 2003-04-12 0.4135806712 13.476382 29.564726  7.844842
#> 206 103 2003-04-13 5.2517351902 13.545307 29.868714 13.257471
#> 207 104 2003-04-14 3.6982564875 13.614025 30.170685 11.640612
#> 210 105 2003-04-15 1.9306358229 13.682523 30.470539  9.767367
#> 212 106 2003-04-16 1.8286628504 13.750789 30.768176  9.737913
#> 214 107 2003-04-17 3.8068181823 13.818808 31.063500 12.044576
#> 215 108 2003-04-18 1.3930054551 13.886567 31.356413  9.411834
#> 217 109 2003-04-19 4.5153834075 13.954052 31.646820 13.031993
#> 220 110 2003-04-20 5.2516789224 14.021246 31.934628 13.964238
#> 222 111 2003-04-21 1.8968448413 14.088136 32.219743 10.223990
#> 224 112 2003-04-22 6.1293533649 14.154703 32.502075 15.162639
#> 225 113 2003-04-23 5.9622022181 14.220933 32.781534 15.067300
#> 228 114 2003-04-24 0.3209451749 14.286808 33.058030  8.635823
#> 230 115 2003-04-25 2.6673980125 14.352310 33.331477 11.430221
#> 232 116 2003-04-26 1.2831843798 14.417421 33.601789  9.895766
#> 233 117 2003-04-27 6.4525974893 14.482122 33.868881 16.012463
#> 236 118 2003-04-28 6.3433079422 14.546394 34.132670 15.975357
#> 238 119 2003-04-29 1.3506032247 14.610216 34.393074 10.187957
#> 240 120 2003-04-30 6.0506269371 14.673568 34.650012 15.806447
#> 241 121 2003-05-01 0.0929082131 14.736428 34.903405  8.835878
#> 244 122 2003-05-02 6.1996223773 14.798774 35.153175 16.151620
#> 246 123 2003-05-03 2.7407947497 14.860584 35.399245 12.114221
#> 247 124 2003-05-04 0.5332341893 14.921835 35.641540  9.547213
#> 250 125 2003-05-05 4.5274691340 14.982502 35.879987 14.391172
#> 252 126 2003-05-06 6.4165772784 15.042562 36.114511 16.731157
#> 254 127 2003-05-07 5.8415603414 15.101989 36.345041 16.115525
#> 255 128 2003-05-08 3.5087294918 15.160757 36.571508 13.374840
#> 258 129 2003-05-09 4.2150133890 15.218840 36.793843 14.293676
#> 260 130 2003-05-10 3.2827132393 15.276211 37.011977 13.229756
#> 262 131 2003-05-11 6.2046196673 15.332843 37.225844 16.838404
#> 263 132 2003-05-12 4.6938394464 15.388707 37.435378 15.068085
#> 266 133 2003-05-13 2.7564440363 15.443775 37.640517 12.769217
#> 268 134 2003-05-14 3.8590347059 15.498017 37.841197 14.171563
#> 270 135 2003-05-15 1.6974487305 15.551405 38.037357 11.585244
#> 271 136 2003-05-16 1.4224618033 15.603907 38.228936 11.299721
#> 273 137 2003-05-17 0.4783919237 15.655493 38.415875 10.190914
#> 276 138 2003-05-18 2.1509161601 15.706133 38.598118 12.292488
#> 278 139 2003-05-19 6.9518266600 15.755795 38.775606 18.248256
#> 280 140 2003-05-20 0.8141642399 15.804447 38.948286 10.740280
#> 281 141 2003-05-21 5.0079615365 15.852058 39.116103 15.957780
#> 284 142 2003-05-22 1.5827611203 15.898595 39.279005 11.774933
#> 286 143 2003-05-23 0.9334398333 15.944026 39.436940 11.013649
#> 288 144 2003-05-24 6.4909723308 15.988319 39.589859 17.933853
#> 289 145 2003-05-25 6.2588585999 16.031441 39.737713 17.691458
#> 292 146 2003-05-26 1.4256507456 16.073361 39.880455 11.738742
#> 294 147 2003-05-27 1.7988804334 16.114046 40.018039 12.238203
#> 296 148 2003-05-28 4.2980392252 16.153463 40.150421 15.379125
#> 297 149 2003-05-29 3.0904943133 16.191581 40.277558 13.913287
#> 299 150 2003-05-30 2.2091627968 16.228369 40.399408 12.849631
#> 302 151 2003-05-31 0.7091963268 16.263796 40.515933 11.012349
#> 304 152 2003-06-01 1.9085233700 16.297830 40.627093 12.535549
#> 306 153 2003-06-02 4.5758957798 16.330443 40.732853 15.890018
#> 307 154 2003-06-03 6.4950907156 16.361604 40.833176 18.313098
#> 310 155 2003-06-04 0.1863726901 16.391286 40.928031 10.464688
#> 312 156 2003-06-05 3.9164383798 16.419460 41.017385 15.146166
#> 314 157 2003-06-06 5.9792155649 16.446100 41.101209 17.746770
#> 315 158 2003-06-07 0.5311813466 16.471180 41.179475 10.958870
#> 318 159 2003-06-08 1.4408019935 16.494677 41.252156 12.114717
#> 320 160 2003-06-09 3.4712275269 16.516565 41.319229 14.671765
#> 322 161 2003-06-10 1.8329986085 16.536824 41.380671 12.638556
#> 323 162 2003-06-11 2.9117301547 16.555433 41.436461 14.002988
#> 326 163 2003-06-12 3.0740499967 16.572373 41.486582 14.219369
#> 327 164 2003-06-13 1.2766454353 16.587625 41.531016 11.980945
#> 330 165 2003-06-14 0.5482978623 16.601175 41.569750 11.078913
#> 332 166 2003-06-15 0.1243896922 16.613007 41.602770 10.556443
#> 334 167 2003-06-16 5.0972598132 16.623108 41.630068 16.790176
#> 335 168 2003-06-17 6.7959151736 16.631469 41.651634 18.922709
#> 338 169 2003-06-18 3.7465231165 16.638080 41.667464 15.108156
#> 340 170 2003-06-19 1.1654353749 16.642934 41.677552 11.878641
#> 342 171 2003-06-20 0.4072295546 16.646025 41.681897 10.930329
#> 343 172 2003-06-21 3.2139731331 16.647351 41.680500 14.443588
#> 346 173 2003-06-22 4.3595917500 16.646909 41.673363 15.875175
#> 348 174 2003-06-23 4.1637344097 16.644700 41.660491 15.625887
#> 350 175 2003-06-24 5.2011276619 16.640727 41.641891 16.918145
#> 351 176 2003-06-25 3.6699609405 16.634994 41.617572 14.995163
#> 353 177 2003-06-26 3.4777767523 16.627507 41.587545 14.746071
#> 356 178 2003-06-27 0.1030316185 16.618275 41.551823 10.516764
#> 358 179 2003-06-28 6.7054338716 16.607306 41.510423 18.757815
#> 360 180 2003-06-29 3.1716246458 16.594614 41.463360 14.328156
#> 361 181 2003-06-30 6.8707229439 16.580211 41.410655 18.932806
#> 364 182 2003-07-01 3.8129361621 16.564113 41.352328 15.097582
#> 366 183 2003-07-02 2.5202968079 16.546336 41.288405 13.466575
#> 368 184 2003-07-03 3.5220227316 16.526900 41.218909 14.696777
#> 369 185 2003-07-04 0.5224354262 16.505823 41.143868 10.937101
#> 372 186 2003-07-05 6.7885432325 16.483128 41.063311 18.721749
#> 374 187 2003-07-06 4.5395091963 16.458837 40.977269 15.895284
#> 376 188 2003-07-07 6.8919582665 16.432973 40.885775 18.795153
#> 377 189 2003-07-08 3.7529327988 16.405563 40.788864 14.862641
#> 379 190 2003-07-09 5.3274158353 16.376632 40.686570 16.789434
#> 382 191 2003-07-10 3.9507000893 16.346207 40.578932 15.048476
#> 384 192 2003-07-11 5.4491801246 16.314316 40.465990 16.874564
#> 386 193 2003-07-12 4.4734269199 16.280989 40.347784 15.630002
#> 387 194 2003-07-13 2.6527251494 16.246255 40.224356 13.340051
#> 390 195 2003-07-14 2.5167146940 16.210144 40.095751 13.136482
#> 392 196 2003-07-15 5.9096199137 16.172686 39.962013 17.291711
#> 394 197 2003-07-16 3.1972442656 16.133915 39.823189 13.901661
#> 395 198 2003-07-17 1.2532400342 16.093860 39.679327 11.464760
#> 398 199 2003-07-18 6.7634643705 16.052554 39.530476 18.210359
#> 400 200 2003-07-19 4.9938836617 16.010029 39.376687 15.985404
#> 402 201 2003-07-20 0.1742022182 15.966317 39.218012 10.018449
#> 403 202 2003-07-21 4.3851954595 15.921451 39.054503 15.141956
#> 406 203 2003-07-22 5.7954492546 15.875462 38.886214 16.819396
#> 407 204 2003-07-23 4.5014832215 15.828384 38.713200 15.183183
#> 410 205 2003-07-24 4.9593898323 15.780249 38.535519 15.689318
#> 412 206 2003-07-25 6.0950900745 15.731088 38.353226 17.018383
#> 414 207 2003-07-26 1.9726880719 15.680933 38.166382 11.942293
#> 415 208 2003-07-27 1.2125738936 15.629816 37.975044 10.966828
#> 418 209 2003-07-28 0.7676901331 15.577768 37.779274 10.375721
#> 420 210 2003-07-29 4.1035553410 15.524820 37.579134 14.361284
#> 422 211 2003-07-30 5.9550492915 15.471001 37.374685 16.536745
#> 423 212 2003-07-31 6.4506832347 15.416342 37.165992 17.067209
#> 426 213 2003-08-01 4.7296333136 15.360873 36.953119 14.927237
#> 428 214 2003-08-02 2.4541171370 15.304621 36.736132 12.129378
#> 430 215 2003-08-03 0.0606065234 15.247616 36.515097  9.201345
#> 431 216 2003-08-04 3.0162915480 15.189886 36.290082 12.675624
#> 433 217 2003-08-05 5.0745456801 15.131457 36.061156 15.062095
#> 436 218 2003-08-06 2.8857947094 15.072356 35.828388 12.386998
#> 438 219 2003-08-07 3.8560367664 15.012610 35.591848 13.468902
#> 440 220 2003-08-08 1.0945239908 14.952243 35.351609 10.131795
#> 441 221 2003-08-09 0.3010090524 14.891281 35.107744  9.131766
#> 444 222 2003-08-10 4.7925087914 14.829748 34.860325 14.347963
#> 446 223 2003-08-11 3.6008174797 14.767666 34.609429 12.871786
#> 448 224 2003-08-12 6.6198149386 14.705060 34.355129 16.321651
#> 449 225 2003-08-13 5.9016785952 14.641952 34.097505 15.396155
#> 452 226 2003-08-14 0.4541816534 14.578362 33.836632  8.986240
#> 454 227 2003-08-15 5.3347764020 14.514313 33.572591 14.562998
#> 456 228 2003-08-16 0.5486897659 14.449824 33.305462  8.958704
#> 457 229 2003-08-17 0.9345536388 14.384916 33.035326  9.331944
#> 459 230 2003-08-18 5.6678734070 14.319607 32.762264 14.674416
#> 462 231 2003-08-19 3.9156826755 14.253916 32.486361 12.583742
#> 464 232 2003-08-20 5.1361078029 14.187862 32.207701 13.881635
#> 466 233 2003-08-21 2.4478683304 14.121461 31.926370 10.748713
#> 467 234 2003-08-22 6.8786806664 14.054730 31.642455 15.653855
#> 470 235 2003-08-23 2.8619771593 13.987686 31.356043 11.046842
#> 472 236 2003-08-24 2.5261158552 13.920345 31.067223 10.585680
#> 474 237 2003-08-25 4.5675702749 13.852721 30.776085 12.767823
#> 475 238 2003-08-26 1.0049145808 13.784830 30.482721  8.731776
#> 478 239 2003-08-27 4.7655903073 13.716686 30.187223 12.790782
#> 480 240 2003-08-28 6.0150012553 13.648302 29.889684 14.058826
#> 482 241 2003-08-29 5.8704099276 13.579692 29.590199 13.793372
#> 483 242 2003-08-30 6.1691904743 13.510869 29.288864 14.009002
#> 486 243 2003-08-31 1.6638841627 13.441846 28.985774  9.040430
#> 488 244 2003-09-01 4.0504159203 13.372633 28.681027 11.513832
#> 489 245 2003-09-02 6.2853175399 13.303244 28.374722 13.796712
#> 492 246 2003-09-03 2.4110032138 13.233689 28.066958  9.573454
#> 494 247 2003-09-04 5.3260076088 13.163980 27.757836 12.554723
#> 496 248 2003-09-05 5.1177255162 13.094126 27.447457 12.225665
#> 497 249 2003-09-06 5.9670266353 13.024138 27.135924 13.000161
#> 500 250 2003-09-07 6.8307262855 12.954027 26.823338 13.777878
#> 502 251 2003-09-08 0.7913325059 12.883802 26.509805  7.441577
#> 504 252 2003-09-09 6.7918359581 12.813472 26.195428 13.491357
#> 505 253 2003-09-10 4.5365794727 12.743047 25.880314 11.076830
#> 508 254 2003-09-11 0.3925113028 12.672536 25.564567  6.787052
#> 510 255 2003-09-12 3.3119596706 12.601947 25.248295  9.629868
#> 512 256 2003-09-13 2.0623750919 12.531289 24.931604  8.284498
#> 513 257 2003-09-14 4.2720973536 12.460572 24.614602 10.373199
#> 515 258 2003-09-15 0.8477621619 12.389802 24.297397  6.905614
#> 518 259 2003-09-16 4.4059155553 12.318988 23.980097 10.283294
#> 520 260 2003-09-17 4.9836871200 12.248139 23.662812 10.729824
#> 522 261 2003-09-18 4.2845795064 12.177262 23.345649  9.943505
#> 523 262 2003-09-19 0.2410046314 12.106365 23.028718  5.986399
#> 526 263 2003-09-20 4.6612355092 12.035456 22.712128 10.076145
#> 528 264 2003-09-21 4.5707709356 11.964544 22.395989  9.876926
#> 530 265 2003-09-22 3.1630090780 11.893635 22.080409  8.456149
#> 531 266 2003-09-23 3.6176287131 11.822738 21.765497  8.771376
#> 534 267 2003-09-24 4.7362755802 11.751861 21.451363  9.685542
#> 536 268 2003-09-25 6.1665316585 11.681012 21.138114 10.864047
#> 538 269 2003-09-26 5.1186040824 11.610198 20.825859  9.797227
#> 539 270 2003-09-27 2.7903387067 11.539428 20.514705  7.608997
#> 541 271 2003-09-28 3.0136445761 11.468711 20.204761  7.705803
#> 544 272 2003-09-29 1.0178514274 11.398053 19.896131  5.862400
#> 546 273 2003-09-30 3.1578423053 11.327464 19.588923  7.627706
#> 548 274 2003-10-01 4.0514704413 11.256953 19.283241  8.290909
#> 549 275 2003-10-02 0.4929198441 11.186528 18.979189  5.162944
#> 552 276 2003-10-03 5.1962832615 11.116198 18.676871  9.034483
#> 554 277 2003-10-04 3.8445449444 11.045973 18.376389  7.792043
#> 556 278 2003-10-05 4.4418099055 10.975862 18.077844  8.177413
#> 557 279 2003-10-06 0.9564000864 10.905874 17.781337  5.225009
#> 560 280 2003-10-07 6.5116344769 10.836020 17.486965  9.625917
#> 562 281 2003-10-08 4.5578304774 10.766311 17.194826  7.938352
#> 564 282 2003-10-09 3.7009649067 10.696756 16.905016  7.150733
#> 565 283 2003-10-10 1.5758077204 10.627367 16.617629  5.386424
#> 568 284 2003-10-11 6.8828680869 10.558154 16.332759  9.406858
#> 569 285 2003-10-12 0.7633414995 10.489131 16.050495  4.596657
#> 572 286 2003-10-13 0.4896186206 10.420308 15.770928  4.313246
#> 574 287 2003-10-14 5.4435185483 10.351698 15.494145  7.947393
#> 576 288 2003-10-15 4.6082556229 10.283314 15.220231  7.215375
#> 577 289 2003-10-16 1.8144192055 10.215170 14.949271  5.064963
#> 580 290 2003-10-17 6.4283649134 10.147279 14.681345  8.320699
#> 582 291 2003-10-18 1.0409724275 10.079655 14.416534  4.348564
#> 584 292 2003-10-19 4.4847633946 10.012314 14.154915  6.708897
#> 585 293 2003-10-20 1.4086689684  9.945270 13.896562  4.458310
#> 588 294 2003-10-21 5.2053492058  9.878539 13.641550  7.004493
#> 590 295 2003-10-22 5.9334619991  9.812138 13.389948  7.395980
#> 592 296 2003-10-23 5.9754511106  9.746084 13.141825  7.314169
#> 593 297 2003-10-24 5.6238880090  9.680393 12.897247  6.970682
#> 595 298 2003-10-25 0.1671373537  9.615084 12.656277  3.274070
#> 598 299 2003-10-26 5.5676440035  9.550176 12.418977  6.724805
#> 600 300 2003-10-27 6.2058535737  9.485687 12.185404  7.032400
#> 602 301 2003-10-28 4.2002613691  9.421638 11.955615  5.653871
#> 603 302 2003-10-29 3.0295171652  9.358048 11.729663  4.831060
#> 606 303 2003-10-30 6.4419306223  9.294940 11.507600  6.864616
#> 608 304 2003-10-31 4.3121009958  9.232334 11.289473  5.458828
#> 610 305 2003-11-01 3.0500271111  9.170252 11.075328  4.610660
#> 611 306 2003-11-02 6.2811308708  9.108719 10.865209  6.462482
#> 614 307 2003-11-03 5.3830946553  9.047757 10.659156  5.835699
#> 616 308 2003-11-04 0.8968103081  8.987390 10.457208  3.136040
#> 618 309 2003-11-05 3.7704318434  8.927644 10.259398  4.731287
#> 619 310 2003-11-06 0.9031135815  8.868543 10.065762  3.028956
#> 621 311 2003-11-07 3.5767541877  8.810114  9.876328  4.473891
#> 624 312 2003-11-08 3.8065488376  8.752384  9.691126  4.530193
#> 626 313 2003-11-09 1.6924326292  8.695379  9.510180  3.303056
#> 628 314 2003-11-10 5.6220088068  8.639127  9.333513  5.370323
#> 629 315 2003-11-11 6.6353960712  8.583658  9.161147  5.831192
#> 632 316 2003-11-12 2.8515294176  8.528999  8.993100  3.751622
#> 634 317 2003-11-13 3.6032119396  8.475180  8.829388  4.084249
#> 636 318 2003-11-14 6.8707929526  8.422232  8.670025  5.703977
#> 637 319 2003-11-15 2.4793070969  8.370184  8.515023  3.389861
#> 640 320 2003-11-16 3.1908810346  8.319067  8.364392  3.695231
#> 642 321 2003-11-17 1.5401729443  8.268912  8.218139  2.819893
#> 644 322 2003-11-18 0.5728311650  8.219751  8.076271  2.300484
#> 645 323 2003-11-19 2.5846524800  8.171616  7.938791  3.240203
#> 648 324 2003-11-20 0.8866660867  8.124538  7.805702  2.377361
#> 649 325 2003-11-21 5.1791075380  8.078549  7.677004  4.380091
#> 652 326 2003-11-22 6.6910640008  8.033683  7.552698  5.033406
#> 654 327 2003-11-23 4.6807215177  7.989971  7.432780  4.035348
#> 656 328 2003-11-24 3.2766254421  7.947446  7.317248  3.337713
#> 657 329 2003-11-25 3.7673471260  7.906140  7.206095  3.518409
#> 660 330 2003-11-26 2.3166759131  7.866085  7.099317  2.820255
#> 662 331 2003-11-27 4.4007747553  7.827314  6.996907  3.716173
#> 664 332 2003-11-28 3.1678673008  7.789856  6.898857  3.127478
#> 665 333 2003-11-29 6.2161062430  7.753745  6.805157  4.429105
#> 668 334 2003-11-30 5.4894127378  7.719011  6.715800  4.066937
#> 670 335 2003-12-01 0.9664168563  7.685684  6.630775  2.074579
#> 672 336 2003-12-02 1.9691914173  7.653793  6.550072  2.480129
#> 673 337 2003-12-03 4.9486254861  7.623368  6.473680  3.719579
#> 675 338 2003-12-04 3.0953084414  7.594437  6.401590  2.904964
#> 678 339 2003-12-05 4.7438896613  7.567027  6.333789  3.568824
#> 680 340 2003-12-06 4.9050048993  7.541163  6.270268  3.606755
#> 682 341 2003-12-07 0.4788449355  7.516872  6.211016  1.750583
#> 683 342 2003-12-08 6.7095246441  7.494177  6.156023  4.294745
#> 686 343 2003-12-09 1.7113846468  7.473100  6.105278  2.225392
#> 688 344 2003-12-10 6.4839833423  7.453664  6.058772  4.149973
#> 690 345 2003-12-11 3.5192295376  7.435887  6.016497  2.927857
#> 691 346 2003-12-12 6.5486199944  7.419789  5.978443  4.132863
#> 694 347 2003-12-13 3.0715974274  7.405386  5.944603  2.718998
#> 696 348 2003-12-14 2.4534057572  7.392694  5.914970  2.460240
#> 698 349 2003-12-15 3.3355286187  7.381725  5.889539  2.803017
#> 699 350 2003-12-16 6.7631166757  7.372493  5.868303  4.158704
#> 701 351 2003-12-17 2.0801334889  7.365006  5.851259  2.289114
#> 704 352 2003-12-18 2.0345381584  7.359273  5.838404  2.266641
#> 706 353 2003-12-19 6.0180557331  7.355300  5.829736  3.842359
#> 708 354 2003-12-20 1.3531886679  7.353091  5.825253  1.992324
#> 709 355 2003-12-21 5.6637585997  7.352649  5.824956  3.699726
#> 712 356 2003-12-22 2.5584055870  7.353975  5.828845  2.471122
#> 714 357 2003-12-23 1.7688975721  7.357066  5.836924  2.160932
#> 716 358 2003-12-24 5.9925670966  7.361920  5.849194  3.842907
#> 717 359 2003-12-25 0.9101705870  7.368531  5.865662  1.828683
#> 720 360 2003-12-26 5.5433536230  7.376892  5.886331  3.683220
#> 722 361 2003-12-27 2.4913105716  7.386993  5.911209  2.474599
#> 724 362 2003-12-28 6.7504068639  7.398825  5.940303  4.194929
#> 725 363 2003-12-29 0.8307824642  7.412375  5.973622  1.828169
#> 728 364 2003-12-30 3.2071245466  7.427627  6.011174  2.800556
#> 730 365 2003-12-31 3.6655530566  7.444567  6.052971  3.003422