Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo       var1        var2        var3         var4        var5
#>  1: 2002  1 -0.1089230 -0.12683656 -0.71332329 -0.926218134 -0.12687530
#>  2: 2002  2 -0.4542524 -0.01994430 -0.72175526 -0.141797353 -0.28358228
#>  3: 2002  3 -0.8985413 -0.92311146 -0.07860949 -0.464287821 -0.25544338
#>  4: 2002  4 -0.1849121 -0.81999197 -0.40699732 -0.238775594 -0.23133707
#>  5: 2002  5 -0.6611485 -0.17297469 -0.87353669 -0.319918024 -0.07007722
#>  6: 2002  6 -0.3156762 -0.81323478 -0.51583334 -0.735572642 -0.57361340
#>  7: 2002  7 -0.3787310 -0.36689269 -0.01375242 -0.034080069 -0.87497964
#>  8: 2002  8 -0.1545400 -0.05741203 -0.55216112 -0.936572441 -0.34058723
#>  9: 2002  9 -0.1645757 -0.24422180 -0.00834414 -0.204712181 -0.99271111
#> 10: 2002 10 -0.2704048 -0.56261929 -0.75886701 -0.031393686 -0.07707505
#> 11: 2002 11 -0.9439778 -0.32656650 -0.93536470 -0.332011444 -0.88470518
#> 12: 2002 12 -0.4786304 -0.09531806 -0.18054049 -0.005397218 -0.68087570

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2  var_layer3   var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.12683656 -0.71332329 -0.926218134    1.341415    1.252165
#>  2: 2002  2 -0.01994430 -0.72175526 -0.141797353    1.879641    1.765478
#>  3: 2002  3 -0.92311146 -0.07860949 -0.464287821    1.284236    1.791992
#>  4: 2002  4 -0.81999197 -0.40699732 -0.238775594    1.993395    1.006807
#>  5: 2002  5 -0.17297469 -0.87353669 -0.319918024    1.872829    1.850527
#>  6: 2002  6 -0.81323478 -0.51583334 -0.735572642    1.140925    1.957716
#>  7: 2002  7 -0.36689269 -0.01375242 -0.034080069    1.111488    1.805285
#>  8: 2002  8 -0.05741203 -0.55216112 -0.936572441    1.048494    1.189647
#>  9: 2002  9 -0.24422180 -0.00834414 -0.204712181    1.911369    1.508904
#> 10: 2002 10 -0.56261929 -0.75886701 -0.031393686    1.927510    1.568911
#> 11: 2002 11 -0.32656650 -0.93536470 -0.332011444    1.941671    1.008031
#> 12: 2002 12 -0.09531806 -0.18054049 -0.005397218    1.168044    1.101214
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.005250    2.348870    2.564947    2.099435
#>  2:    1.133001    2.073483    2.951495    2.116272
#>  3:    1.256427    2.113431    2.005596    2.029438
#>  4:    1.776877    2.810495    2.809997    2.237570
#>  5:    1.209686    2.719109    2.635175    2.449837
#>  6:    1.676844    2.446262    2.047892    2.152877
#>  7:    1.309062    2.386225    2.008884    2.445017
#>  8:    1.868057    2.219054    2.610168    2.588025
#>  9:    1.179743    2.486580    2.090026    2.165966
#> 10:    1.824491    2.223472    2.333367    2.416141
#> 11:    1.102795    2.705023    2.816899    2.105623
#> 12:    1.070326    2.398398    2.601490    2.186711

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.341415    1.252165    1.005250    2.348870    2.564947
#>  2: 2002  2    1.879641    1.765478    1.133001    2.073483    2.951495
#>  3: 2002  3    1.284236    1.791992    1.256427    2.113431    2.005596
#>  4: 2002  4    1.993395    1.006807    1.776877    2.810495    2.809997
#>  5: 2002  5    1.872829    1.850527    1.209686    2.719109    2.635175
#>  6: 2002  6    1.140925    1.957716    1.676844    2.446262    2.047892
#>  7: 2002  7    1.111488    1.805285    1.309062    2.386225    2.008884
#>  8: 2002  8    1.048494    1.189647    1.868057    2.219054    2.610168
#>  9: 2002  9    1.911369    1.508904    1.179743    2.486580    2.090026
#> 10: 2002 10    1.927510    1.568911    1.824491    2.223472    2.333367
#> 11: 2002 11    1.941671    1.008031    1.102795    2.705023    2.816899
#> 12: 2002 12    1.168044    1.101214    1.070326    2.398398    2.601490
#>     var2_layer4
#>  1:    2.099435
#>  2:    2.116272
#>  3:    2.029438
#>  4:    2.237570
#>  5:    2.449837
#>  6:    2.152877
#>  7:    2.445017
#>  8:    2.588025
#>  9:    2.165966
#> 10:    2.416141
#> 11:    2.105623
#> 12:    2.186711