Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2        var3        var4         var5
#>  1: 2002  1 -0.44897764 -0.37660492 -0.15059931 -0.47860229 -0.240493204
#>  2: 2002  2 -0.49805025 -0.46477743 -0.72737334 -0.80620419 -0.051950375
#>  3: 2002  3 -0.16395649 -0.92551266 -0.32130314 -0.21548629 -0.972615028
#>  4: 2002  4 -0.33419862 -0.26607822 -0.72590009 -0.12301313 -0.611996501
#>  5: 2002  5 -0.19836734 -0.32519549 -0.24476217 -0.92945426 -0.757763445
#>  6: 2002  6 -0.73550828 -0.75331754 -0.89710200 -0.36523010 -0.999407202
#>  7: 2002  7 -0.09993743 -0.12211472 -0.59386298 -0.15306825 -0.007681782
#>  8: 2002  8 -0.22080250 -0.54723296 -0.37016166 -0.87046827 -0.877646178
#>  9: 2002  9 -0.81607281 -0.08360678 -0.60813686 -0.43680121 -0.996466602
#> 10: 2002 10 -0.73306003 -0.98305769 -0.82613195 -0.11453207 -0.390263633
#> 11: 2002 11 -0.97737311 -0.09472926 -0.22609836 -0.40437297 -0.106339972
#> 12: 2002 12 -0.54014201 -0.89796110 -0.02516413 -0.54247037 -0.736312124
#> 13: 2003  1 -0.72521717 -0.98570913 -0.18677290 -0.63825237 -0.201394848
#> 14: 2003  2 -0.52072571 -0.65992657 -0.27459977 -0.03082109 -0.989667955
#> 15: 2003  3 -0.93268371 -0.41166762 -0.67947340 -0.58560174 -0.874985086
#> 16: 2003  4 -0.79246676 -0.89504792 -0.76603756 -0.82023703 -0.487240401
#> 17: 2003  5 -0.56540324 -0.98695331 -0.40238465 -0.91194117 -0.225987845
#> 18: 2003  6 -0.20289190 -0.25814977 -0.70639727 -0.32212123 -0.651421033
#> 19: 2003  7 -0.22678705 -0.18383787 -0.32563352 -0.24332606 -0.638411850
#> 20: 2003  8 -0.98293401 -0.37808670 -0.55369279 -0.22287674 -0.567659975
#> 21: 2003  9 -0.60804640 -0.85290106 -0.92654153 -0.37370999 -0.039667384
#> 22: 2003 10 -0.59108243 -0.06530153 -0.67271105 -0.94586976 -0.454802746
#> 23: 2003 11 -0.36411976 -0.22236727 -0.09470145 -0.89719658 -0.916498876
#> 24: 2003 12 -0.79934986 -0.88990176 -0.69110354 -0.71384199 -0.765406125
#>       yr mo        var1        var2        var3        var4         var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2       var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.37660492 -0.15059931 -0.47860229 1.460419 1.951952 1.205009
#>  2: 2002  2 -0.46477743 -0.72737334 -0.80620419 1.049093 1.339700 1.978322
#>  3: 2002  3 -0.92551266 -0.32130314 -0.21548629 1.438752 1.020200 1.564335
#>  4: 2002  4 -0.26607822 -0.72590009 -0.12301313 1.529807 1.947752 1.348426
#>  5: 2002  5 -0.32519549 -0.24476217 -0.92945426 1.734701 1.764511 1.065803
#>  6: 2002  6 -0.75331754 -0.89710200 -0.36523010 1.203046 1.288093 1.603894
#>  7: 2002  7 -0.12211472 -0.59386298 -0.15306825 1.251501 1.312689 1.781779
#>  8: 2002  8 -0.54723296 -0.37016166 -0.87046827 1.567538 1.576153 1.518320
#>  9: 2002  9 -0.08360678 -0.60813686 -0.43680121 1.726439 1.800127 1.674904
#> 10: 2002 10 -0.98305769 -0.82613195 -0.11453207 1.824526 1.533261 1.522553
#> 11: 2002 11 -0.09472926 -0.22609836 -0.40437297 1.818358 1.818158 1.198743
#> 12: 2002 12 -0.89796110 -0.02516413 -0.54247037 1.799381 1.861885 1.486897
#> 13: 2003  1 -0.98570913 -0.18677290 -0.63825237 1.116154 1.658850 1.667812
#> 14: 2003  2 -0.65992657 -0.27459977 -0.03082109 1.070284 1.288228 1.804887
#> 15: 2003  3 -0.41166762 -0.67947340 -0.58560174 1.584835 1.547354 1.454023
#> 16: 2003  4 -0.89504792 -0.76603756 -0.82023703 1.742896 1.470827 1.438879
#> 17: 2003  5 -0.98695331 -0.40238465 -0.91194117 1.345381 1.319908 1.244588
#> 18: 2003  6 -0.25814977 -0.70639727 -0.32212123 1.675979 1.768841 1.086384
#> 19: 2003  7 -0.18383787 -0.32563352 -0.24332606 1.514907 1.670790 1.643173
#> 20: 2003  8 -0.37808670 -0.55369279 -0.22287674 1.815440 1.886363 1.733040
#> 21: 2003  9 -0.85290106 -0.92654153 -0.37370999 1.371817 1.603669 1.428188
#> 22: 2003 10 -0.06530153 -0.67271105 -0.94586976 1.978406 1.479205 1.954005
#> 23: 2003 11 -0.22236727 -0.09470145 -0.89719658 1.590885 1.345602 1.078431
#> 24: 2003 12 -0.88990176 -0.69110354 -0.71384199 1.040340 1.821849 1.606933
#>       yr mo       var_2       var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.933139 2.108315 2.093244
#>  2: 2.520152 2.978344 2.647001
#>  3: 2.483773 2.051730 2.290884
#>  4: 2.051866 2.232924 2.506146
#>  5: 2.294030 2.682963 2.698343
#>  6: 2.494412 2.986545 2.648124
#>  7: 2.000695 2.609230 2.201867
#>  8: 2.021513 2.610980 2.874459
#>  9: 2.719019 2.170964 2.172880
#> 10: 2.852650 2.648910 2.565032
#> 11: 2.914770 2.626348 2.781766
#> 12: 2.782673 2.810622 2.871206
#> 13: 2.102921 2.654548 2.640650
#> 14: 2.826115 2.854982 2.558962
#> 15: 2.844983 2.682363 2.946904
#> 16: 2.521965 2.100712 2.678434
#> 17: 2.462532 2.825689 2.179763
#> 18: 2.340779 2.788034 2.418476
#> 19: 2.900365 2.281242 2.930939
#> 20: 2.866463 2.927567 2.111690
#> 21: 2.208998 2.863447 2.717604
#> 22: 2.878576 2.900239 2.219654
#> 23: 2.811926 2.186555 2.916642
#> 24: 2.083980 2.889451 2.111464
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.460419 1.951952 1.205009 2.933139 2.108315 2.093244
#>  2: 2002  2 1.049093 1.339700 1.978322 2.520152 2.978344 2.647001
#>  3: 2002  3 1.438752 1.020200 1.564335 2.483773 2.051730 2.290884
#>  4: 2002  4 1.529807 1.947752 1.348426 2.051866 2.232924 2.506146
#>  5: 2002  5 1.734701 1.764511 1.065803 2.294030 2.682963 2.698343
#>  6: 2002  6 1.203046 1.288093 1.603894 2.494412 2.986545 2.648124
#>  7: 2002  7 1.251501 1.312689 1.781779 2.000695 2.609230 2.201867
#>  8: 2002  8 1.567538 1.576153 1.518320 2.021513 2.610980 2.874459
#>  9: 2002  9 1.726439 1.800127 1.674904 2.719019 2.170964 2.172880
#> 10: 2002 10 1.824526 1.533261 1.522553 2.852650 2.648910 2.565032
#> 11: 2002 11 1.818358 1.818158 1.198743 2.914770 2.626348 2.781766
#> 12: 2002 12 1.799381 1.861885 1.486897 2.782673 2.810622 2.871206
#> 13: 2003  1 1.116154 1.658850 1.667812 2.102921 2.654548 2.640650
#> 14: 2003  2 1.070284 1.288228 1.804887 2.826115 2.854982 2.558962
#> 15: 2003  3 1.584835 1.547354 1.454023 2.844983 2.682363 2.946904
#> 16: 2003  4 1.742896 1.470827 1.438879 2.521965 2.100712 2.678434
#> 17: 2003  5 1.345381 1.319908 1.244588 2.462532 2.825689 2.179763
#> 18: 2003  6 1.675979 1.768841 1.086384 2.340779 2.788034 2.418476
#> 19: 2003  7 1.514907 1.670790 1.643173 2.900365 2.281242 2.930939
#> 20: 2003  8 1.815440 1.886363 1.733040 2.866463 2.927567 2.111690
#> 21: 2003  9 1.371817 1.603669 1.428188 2.208998 2.863447 2.717604
#> 22: 2003 10 1.978406 1.479205 1.954005 2.878576 2.900239 2.219654
#> 23: 2003 11 1.590885 1.345602 1.078431 2.811926 2.186555 2.916642
#> 24: 2003 12 1.040340 1.821849 1.606933 2.083980 2.889451 2.111464
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4