Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo       var1        var2        var3        var4        var5
#>  1: 2002  1 -0.7434017 -0.94090327 -0.09972953 -0.98866030 -0.38378951
#>  2: 2002  2 -0.9619318 -0.73569150 -0.13925796 -0.98886452 -0.20665901
#>  3: 2002  3 -0.5534801 -0.83095517 -0.59342251 -0.78956877 -0.80159380
#>  4: 2002  4 -0.2220256 -0.09080104 -0.82138375 -0.69197422 -0.63593731
#>  5: 2002  5 -0.7427842 -0.58593574 -0.05421227 -0.62577996 -0.39986475
#>  6: 2002  6 -0.2700285 -0.62276375 -0.55914525 -0.87413962 -0.59620545
#>  7: 2002  7 -0.4202834 -0.53711401 -0.84598074 -0.44139231 -0.94952774
#>  8: 2002  8 -0.1657627 -0.91914855 -0.38363849 -0.09207014 -0.91277472
#>  9: 2002  9 -0.3578829 -0.94170192 -0.09130761 -0.31773445 -0.73952838
#> 10: 2002 10 -0.5780145 -0.42866336 -0.15598145 -0.64934039 -0.63627602
#> 11: 2002 11 -0.4739747 -0.73917613 -0.17308802 -0.49517418 -0.54209306
#> 12: 2002 12 -0.9886231 -0.63863544 -0.26829090 -0.99940636 -0.08513715

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2  var_layer3  var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.94090327 -0.09972953 -0.98866030    1.225538    1.275083
#>  2: 2002  2 -0.73569150 -0.13925796 -0.98886452    1.571989    1.815356
#>  3: 2002  3 -0.83095517 -0.59342251 -0.78956877    1.143150    1.558639
#>  4: 2002  4 -0.09080104 -0.82138375 -0.69197422    1.059290    1.327421
#>  5: 2002  5 -0.58593574 -0.05421227 -0.62577996    1.383705    1.241931
#>  6: 2002  6 -0.62276375 -0.55914525 -0.87413962    1.142245    1.743513
#>  7: 2002  7 -0.53711401 -0.84598074 -0.44139231    1.807724    1.113990
#>  8: 2002  8 -0.91914855 -0.38363849 -0.09207014    1.416764    1.092239
#>  9: 2002  9 -0.94170192 -0.09130761 -0.31773445    1.016381    1.987925
#> 10: 2002 10 -0.42866336 -0.15598145 -0.64934039    1.224603    1.446967
#> 11: 2002 11 -0.73917613 -0.17308802 -0.49517418    1.942686    1.872724
#> 12: 2002 12 -0.63863544 -0.26829090 -0.99940636    1.349285    1.466805
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.770275    2.073372    2.049970    2.355717
#>  2:    1.860005    2.746464    2.339358    2.467004
#>  3:    1.924350    2.253131    2.791357    2.387165
#>  4:    1.857609    2.861775    2.683877    2.622396
#>  5:    1.957873    2.538168    2.786486    2.046843
#>  6:    1.285621    2.974562    2.959139    2.228012
#>  7:    1.110312    2.225577    2.460315    2.798184
#>  8:    1.810652    2.510374    2.110844    2.491918
#>  9:    1.598431    2.904792    2.491498    2.601916
#> 10:    1.005153    2.283234    2.618845    2.552575
#> 11:    1.790258    2.671269    2.151562    2.144268
#> 12:    1.258938    2.894971    2.069579    2.802282

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.225538    1.275083    1.770275    2.073372    2.049970
#>  2: 2002  2    1.571989    1.815356    1.860005    2.746464    2.339358
#>  3: 2002  3    1.143150    1.558639    1.924350    2.253131    2.791357
#>  4: 2002  4    1.059290    1.327421    1.857609    2.861775    2.683877
#>  5: 2002  5    1.383705    1.241931    1.957873    2.538168    2.786486
#>  6: 2002  6    1.142245    1.743513    1.285621    2.974562    2.959139
#>  7: 2002  7    1.807724    1.113990    1.110312    2.225577    2.460315
#>  8: 2002  8    1.416764    1.092239    1.810652    2.510374    2.110844
#>  9: 2002  9    1.016381    1.987925    1.598431    2.904792    2.491498
#> 10: 2002 10    1.224603    1.446967    1.005153    2.283234    2.618845
#> 11: 2002 11    1.942686    1.872724    1.790258    2.671269    2.151562
#> 12: 2002 12    1.349285    1.466805    1.258938    2.894971    2.069579
#>     var2_layer4
#>  1:    2.355717
#>  2:    2.467004
#>  3:    2.387165
#>  4:    2.622396
#>  5:    2.046843
#>  6:    2.228012
#>  7:    2.798184
#>  8:    2.491918
#>  9:    2.601916
#> 10:    2.552575
#> 11:    2.144268
#> 12:    2.802282