Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2         var3        var4        var5
#>  1: 2002  1 -0.22404876 -0.79751061 -0.641569310 -0.93212371 -0.60756188
#>  2: 2002  2 -0.33290104 -0.45729867 -0.592939937 -0.15156410 -0.06982360
#>  3: 2002  3 -0.81371586 -0.06570054 -0.545790193 -0.55759609 -0.01352537
#>  4: 2002  4 -0.57469552 -0.91717923 -0.667486330 -0.53839611 -0.37920217
#>  5: 2002  5 -0.65174930 -0.27659798 -0.446928771 -0.62781638 -0.15356395
#>  6: 2002  6 -0.95158710 -0.18279883 -0.545668149 -0.89700498 -0.89179104
#>  7: 2002  7 -0.38140454 -0.78431524 -0.139700188 -0.71180425 -0.20947424
#>  8: 2002  8 -0.49619214 -0.97590783 -0.993635452 -0.93595584 -0.28737454
#>  9: 2002  9 -0.44801339 -0.51074699 -0.114201987 -0.16224425 -0.57095699
#> 10: 2002 10 -0.82807284 -0.86478701 -0.325640512 -0.78926491 -0.60015411
#> 11: 2002 11 -0.61254510 -0.54695338 -0.502486089 -0.08010910 -0.64777256
#> 12: 2002 12 -0.63332382 -0.25570041 -0.589023070 -0.84730753 -0.85028354
#> 13: 2003  1 -0.51099306 -0.05790353 -0.824685175 -0.15093030 -0.22835579
#> 14: 2003  2 -0.85242622 -0.64611303 -0.599159735 -0.70534077 -0.38956728
#> 15: 2003  3 -0.25277850 -0.80606770 -0.277942677 -0.88101120 -0.82664986
#> 16: 2003  4 -0.13708214 -0.46122541 -0.070894744 -0.54913644 -0.76408887
#> 17: 2003  5 -0.76644205 -0.37542400 -0.568124229 -0.04144784 -0.14806399
#> 18: 2003  6 -0.66065881 -0.23040789 -0.321627187 -0.43579574 -0.68460135
#> 19: 2003  7 -0.16600727 -0.98621735 -0.838982122 -0.36925759 -0.37918075
#> 20: 2003  8 -0.03685045 -0.27753308 -0.645199302 -0.55692897 -0.53616983
#> 21: 2003  9 -0.34725170 -0.46523664 -0.884442199 -0.39003448 -0.53130098
#> 22: 2003 10 -0.44216659 -0.89463770 -0.007846116 -0.18181832 -0.36967600
#> 23: 2003 11 -0.97905949 -0.33387085 -0.564957114 -0.51561783 -0.09709643
#> 24: 2003 12 -0.25513487 -0.96642380 -0.780812496 -0.57503938 -0.04268047
#>       yr mo        var1        var2         var3        var4        var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2        var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.79751061 -0.641569310 -0.93212371 1.941955 1.817528 1.984739
#>  2: 2002  2 -0.45729867 -0.592939937 -0.15156410 1.611453 1.701615 1.044794
#>  3: 2002  3 -0.06570054 -0.545790193 -0.55759609 1.622819 1.637244 1.708631
#>  4: 2002  4 -0.91717923 -0.667486330 -0.53839611 1.927684 1.305534 1.942401
#>  5: 2002  5 -0.27659798 -0.446928771 -0.62781638 1.699964 1.902064 1.754104
#>  6: 2002  6 -0.18279883 -0.545668149 -0.89700498 1.813832 1.593993 1.009804
#>  7: 2002  7 -0.78431524 -0.139700188 -0.71180425 1.926112 1.024778 1.350619
#>  8: 2002  8 -0.97590783 -0.993635452 -0.93595584 1.878446 1.131458 1.530700
#>  9: 2002  9 -0.51074699 -0.114201987 -0.16224425 1.153861 1.620738 1.866365
#> 10: 2002 10 -0.86478701 -0.325640512 -0.78926491 1.518820 1.772253 1.003092
#> 11: 2002 11 -0.54695338 -0.502486089 -0.08010910 1.430610 1.440046 1.083668
#> 12: 2002 12 -0.25570041 -0.589023070 -0.84730753 1.231028 1.857538 1.099920
#> 13: 2003  1 -0.05790353 -0.824685175 -0.15093030 1.155350 1.341100 1.055523
#> 14: 2003  2 -0.64611303 -0.599159735 -0.70534077 1.533299 1.756620 1.420749
#> 15: 2003  3 -0.80606770 -0.277942677 -0.88101120 1.447849 1.936159 1.162481
#> 16: 2003  4 -0.46122541 -0.070894744 -0.54913644 1.556260 1.760746 1.683806
#> 17: 2003  5 -0.37542400 -0.568124229 -0.04144784 1.066245 1.880542 1.655532
#> 18: 2003  6 -0.23040789 -0.321627187 -0.43579574 1.713322 1.524617 1.618793
#> 19: 2003  7 -0.98621735 -0.838982122 -0.36925759 1.111637 1.011722 1.841721
#> 20: 2003  8 -0.27753308 -0.645199302 -0.55692897 1.661875 1.163087 1.324571
#> 21: 2003  9 -0.46523664 -0.884442199 -0.39003448 1.469106 1.100746 1.259135
#> 22: 2003 10 -0.89463770 -0.007846116 -0.18181832 1.627567 1.099826 1.802501
#> 23: 2003 11 -0.33387085 -0.564957114 -0.51561783 1.686145 1.810945 1.092771
#> 24: 2003 12 -0.96642380 -0.780812496 -0.57503938 1.266061 1.180218 1.430368
#>       yr mo       var_2        var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.044226 2.404116 2.032613
#>  2: 2.810681 2.949812 2.217460
#>  3: 2.262255 2.917246 2.601911
#>  4: 2.552596 2.348060 2.984688
#>  5: 2.625445 2.901900 2.464157
#>  6: 2.049448 2.388929 2.486690
#>  7: 2.012380 2.137939 2.463177
#>  8: 2.593231 2.203088 2.932320
#>  9: 2.827738 2.534966 2.095260
#> 10: 2.219711 2.941112 2.702856
#> 11: 2.229289 2.875019 2.196921
#> 12: 2.915053 2.420692 2.687824
#> 13: 2.103330 2.593016 2.802396
#> 14: 2.063821 2.593532 2.059724
#> 15: 2.519755 2.412157 2.327812
#> 16: 2.320767 2.985247 2.359897
#> 17: 2.742032 2.573096 2.366606
#> 18: 2.644610 2.413492 2.790120
#> 19: 2.755845 2.157719 2.454586
#> 20: 2.238195 2.040886 2.320484
#> 21: 2.155509 2.902310 2.534481
#> 22: 2.109673 2.047235 2.577222
#> 23: 2.028893 2.043564 2.649150
#> 24: 2.522486 2.304749 2.966054
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.941955 1.817528 1.984739 2.044226 2.404116 2.032613
#>  2: 2002  2 1.611453 1.701615 1.044794 2.810681 2.949812 2.217460
#>  3: 2002  3 1.622819 1.637244 1.708631 2.262255 2.917246 2.601911
#>  4: 2002  4 1.927684 1.305534 1.942401 2.552596 2.348060 2.984688
#>  5: 2002  5 1.699964 1.902064 1.754104 2.625445 2.901900 2.464157
#>  6: 2002  6 1.813832 1.593993 1.009804 2.049448 2.388929 2.486690
#>  7: 2002  7 1.926112 1.024778 1.350619 2.012380 2.137939 2.463177
#>  8: 2002  8 1.878446 1.131458 1.530700 2.593231 2.203088 2.932320
#>  9: 2002  9 1.153861 1.620738 1.866365 2.827738 2.534966 2.095260
#> 10: 2002 10 1.518820 1.772253 1.003092 2.219711 2.941112 2.702856
#> 11: 2002 11 1.430610 1.440046 1.083668 2.229289 2.875019 2.196921
#> 12: 2002 12 1.231028 1.857538 1.099920 2.915053 2.420692 2.687824
#> 13: 2003  1 1.155350 1.341100 1.055523 2.103330 2.593016 2.802396
#> 14: 2003  2 1.533299 1.756620 1.420749 2.063821 2.593532 2.059724
#> 15: 2003  3 1.447849 1.936159 1.162481 2.519755 2.412157 2.327812
#> 16: 2003  4 1.556260 1.760746 1.683806 2.320767 2.985247 2.359897
#> 17: 2003  5 1.066245 1.880542 1.655532 2.742032 2.573096 2.366606
#> 18: 2003  6 1.713322 1.524617 1.618793 2.644610 2.413492 2.790120
#> 19: 2003  7 1.111637 1.011722 1.841721 2.755845 2.157719 2.454586
#> 20: 2003  8 1.661875 1.163087 1.324571 2.238195 2.040886 2.320484
#> 21: 2003  9 1.469106 1.100746 1.259135 2.155509 2.902310 2.534481
#> 22: 2003 10 1.627567 1.099826 1.802501 2.109673 2.047235 2.577222
#> 23: 2003 11 1.686145 1.810945 1.092771 2.028893 2.043564 2.649150
#> 24: 2003 12 1.266061 1.180218 1.430368 2.522486 2.304749 2.966054
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4