Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1       var2       var3       var4        var5
#>  1: 2002  1 -0.40814362 -0.4550859 -0.2379844 -0.6611023 -0.91198508
#>  2: 2002  2 -0.06547838 -0.3772388 -0.6570957 -0.4987791 -0.24499365
#>  3: 2002  3 -0.34346735 -0.8117283 -0.7835959 -0.5823135 -0.53547599
#>  4: 2002  4 -0.65163922 -0.1950602 -0.5297529 -0.6399475 -0.79897540
#>  5: 2002  5 -0.05164487 -0.5846479 -0.8168496 -0.5716687 -0.39800924
#>  6: 2002  6 -0.09906322 -0.5170203 -0.2250129 -0.9057687 -0.35242830
#>  7: 2002  7 -0.23193541 -0.2816569 -0.2857574 -0.6236203 -0.70304673
#>  8: 2002  8 -0.96500223 -0.7147437 -0.2084379 -0.3176740 -0.99847463
#>  9: 2002  9 -0.47493927 -0.5035428 -0.1575765 -0.1706670 -0.89708911
#> 10: 2002 10 -0.38527408 -0.9731980 -0.3400096 -0.4562763 -0.01003975
#> 11: 2002 11 -0.89549809 -0.7629609 -0.8440205 -0.7712967 -0.02618953
#> 12: 2002 12 -0.90401765 -0.3389585 -0.9351397 -0.9009164 -0.64835842

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo var_layer2 var_layer3 var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.4550859 -0.2379844 -0.6611023    1.172463    1.627149
#>  2: 2002  2 -0.3772388 -0.6570957 -0.4987791    1.848431    1.737130
#>  3: 2002  3 -0.8117283 -0.7835959 -0.5823135    1.268462    1.026273
#>  4: 2002  4 -0.1950602 -0.5297529 -0.6399475    1.587577    1.241460
#>  5: 2002  5 -0.5846479 -0.8168496 -0.5716687    1.938940    1.028317
#>  6: 2002  6 -0.5170203 -0.2250129 -0.9057687    1.569178    1.308218
#>  7: 2002  7 -0.2816569 -0.2857574 -0.6236203    1.276610    1.574669
#>  8: 2002  8 -0.7147437 -0.2084379 -0.3176740    1.630432    1.240031
#>  9: 2002  9 -0.5035428 -0.1575765 -0.1706670    1.580269    1.583668
#> 10: 2002 10 -0.9731980 -0.3400096 -0.4562763    1.608337    1.675639
#> 11: 2002 11 -0.7629609 -0.8440205 -0.7712967    1.776522    1.989003
#> 12: 2002 12 -0.3389585 -0.9351397 -0.9009164    1.524309    1.645208
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.352296    2.350820    2.092527    2.153701
#>  2:    1.084563    2.102116    2.146437    2.527242
#>  3:    1.296562    2.544753    2.280611    2.962569
#>  4:    1.564001    2.623988    2.496118    2.063943
#>  5:    1.033063    2.929074    2.795223    2.840934
#>  6:    1.304951    2.611583    2.396057    2.631103
#>  7:    1.268852    2.966901    2.082708    2.629411
#>  8:    1.893046    2.589235    2.377460    2.596409
#>  9:    1.647695    2.374588    2.119885    2.105098
#> 10:    1.873943    2.643102    2.539667    2.953611
#> 11:    1.988388    2.277114    2.521308    2.136417
#> 12:    1.837153    2.356933    2.374741    2.801885

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.172463    1.627149    1.352296    2.350820    2.092527
#>  2: 2002  2    1.848431    1.737130    1.084563    2.102116    2.146437
#>  3: 2002  3    1.268462    1.026273    1.296562    2.544753    2.280611
#>  4: 2002  4    1.587577    1.241460    1.564001    2.623988    2.496118
#>  5: 2002  5    1.938940    1.028317    1.033063    2.929074    2.795223
#>  6: 2002  6    1.569178    1.308218    1.304951    2.611583    2.396057
#>  7: 2002  7    1.276610    1.574669    1.268852    2.966901    2.082708
#>  8: 2002  8    1.630432    1.240031    1.893046    2.589235    2.377460
#>  9: 2002  9    1.580269    1.583668    1.647695    2.374588    2.119885
#> 10: 2002 10    1.608337    1.675639    1.873943    2.643102    2.539667
#> 11: 2002 11    1.776522    1.989003    1.988388    2.277114    2.521308
#> 12: 2002 12    1.524309    1.645208    1.837153    2.356933    2.374741
#>     var2_layer4
#>  1:    2.153701
#>  2:    2.527242
#>  3:    2.962569
#>  4:    2.063943
#>  5:    2.840934
#>  6:    2.631103
#>  7:    2.629411
#>  8:    2.596409
#>  9:    2.105098
#> 10:    2.953611
#> 11:    2.136417
#> 12:    2.801885