Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2         var3       var4        var5
#>  1: 2002  1 -0.44185224 -0.99337213 -0.951351363 -0.7105780 -0.70045864
#>  2: 2002  2 -0.94886826 -0.66610135 -0.723322562 -0.7442959 -0.64459011
#>  3: 2002  3 -0.18801147 -0.28656420 -0.885546875 -0.7120401 -0.31800917
#>  4: 2002  4 -0.70706237 -0.41573920 -0.132042025 -0.4989832 -0.51693009
#>  5: 2002  5 -0.46777075 -0.90635819 -0.168053937 -0.8049266 -0.43595260
#>  6: 2002  6 -0.98143113 -0.51938593 -0.847781910 -0.9187766 -0.40705454
#>  7: 2002  7 -0.25896054 -0.08676929 -0.625051663 -0.2921859 -0.51628986
#>  8: 2002  8 -0.82806131 -0.32285453 -0.583205231 -0.0437389 -0.12985311
#>  9: 2002  9 -0.48959166 -0.87106151 -0.376492939 -0.5620910 -0.73941558
#> 10: 2002 10 -0.86081584 -0.84680061 -0.928527780 -0.1355217 -0.95011934
#> 11: 2002 11 -0.30483365 -0.68510733 -0.798009776 -0.1002584 -0.40867361
#> 12: 2002 12 -0.53012455 -0.53258299 -0.928439948 -0.4609896 -0.83871640
#> 13: 2003  1 -0.55122061 -0.79993435 -0.248714328 -0.1825641 -0.80242531
#> 14: 2003  2 -0.78840183 -0.68843991 -0.639538835 -0.6910275 -0.19623065
#> 15: 2003  3 -0.33901009 -0.95347972 -0.540607607 -0.4243737 -0.27601657
#> 16: 2003  4 -0.95212943 -0.44753138 -0.693126653 -0.6648184 -0.28936575
#> 17: 2003  5 -0.82656975 -0.37830636 -0.034081736 -0.6092071 -0.05001137
#> 18: 2003  6 -0.96055995 -0.57765440 -0.640542171 -0.3726618 -0.28107423
#> 19: 2003  7 -0.78219572 -0.07937517 -0.387829824 -0.4662183 -0.72528907
#> 20: 2003  8 -0.40724769 -0.81033657 -0.956739668 -0.6446598 -0.66828831
#> 21: 2003  9 -0.19916402 -0.03181120 -0.341553529 -0.4626903 -0.66976070
#> 22: 2003 10 -0.81951839 -0.07862079 -0.762226023 -0.7694559 -0.59693231
#> 23: 2003 11 -0.13461559 -0.72949752 -0.263566474 -0.5579310 -0.88540462
#> 24: 2003 12 -0.03204333 -0.12109007 -0.008389655 -0.1423439 -0.35429359
#>       yr mo        var1        var2         var3       var4        var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2        var_3      var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.99337213 -0.951351363 -0.7105780 1.534748 1.564917 1.332330
#>  2: 2002  2 -0.66610135 -0.723322562 -0.7442959 1.748225 1.854544 1.429773
#>  3: 2002  3 -0.28656420 -0.885546875 -0.7120401 1.606936 1.994451 1.643857
#>  4: 2002  4 -0.41573920 -0.132042025 -0.4989832 1.617798 1.091708 1.603892
#>  5: 2002  5 -0.90635819 -0.168053937 -0.8049266 1.630905 1.136370 1.013376
#>  6: 2002  6 -0.51938593 -0.847781910 -0.9187766 1.057088 1.545894 1.747002
#>  7: 2002  7 -0.08676929 -0.625051663 -0.2921859 1.373436 1.816441 1.926416
#>  8: 2002  8 -0.32285453 -0.583205231 -0.0437389 1.476845 1.890898 1.029006
#>  9: 2002  9 -0.87106151 -0.376492939 -0.5620910 1.351720 1.674336 1.236270
#> 10: 2002 10 -0.84680061 -0.928527780 -0.1355217 1.888867 1.339109 1.372383
#> 11: 2002 11 -0.68510733 -0.798009776 -0.1002584 1.601524 1.357728 1.726824
#> 12: 2002 12 -0.53258299 -0.928439948 -0.4609896 1.956687 1.106301 1.630970
#> 13: 2003  1 -0.79993435 -0.248714328 -0.1825641 1.984424 1.681656 1.366496
#> 14: 2003  2 -0.68843991 -0.639538835 -0.6910275 1.428707 1.812605 1.988552
#> 15: 2003  3 -0.95347972 -0.540607607 -0.4243737 1.272681 1.712994 1.413446
#> 16: 2003  4 -0.44753138 -0.693126653 -0.6648184 1.583053 1.741310 1.435323
#> 17: 2003  5 -0.37830636 -0.034081736 -0.6092071 1.288060 1.232234 1.821573
#> 18: 2003  6 -0.57765440 -0.640542171 -0.3726618 1.960688 1.760492 1.346853
#> 19: 2003  7 -0.07937517 -0.387829824 -0.4662183 1.803285 1.615558 1.779977
#> 20: 2003  8 -0.81033657 -0.956739668 -0.6446598 1.106825 1.986852 1.695909
#> 21: 2003  9 -0.03181120 -0.341553529 -0.4626903 1.644605 1.046140 1.979762
#> 22: 2003 10 -0.07862079 -0.762226023 -0.7694559 1.075354 1.010623 1.496437
#> 23: 2003 11 -0.72949752 -0.263566474 -0.5579310 1.879148 1.409248 1.859771
#> 24: 2003 12 -0.12109007 -0.008389655 -0.1423439 1.589244 1.443581 1.670856
#>       yr mo       var_2        var_3      var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.276558 2.262789 2.426601
#>  2: 2.423900 2.842204 2.085967
#>  3: 2.710232 2.927665 2.391658
#>  4: 2.818661 2.910004 2.912244
#>  5: 2.191822 2.604021 2.266738
#>  6: 2.214650 2.152229 2.420270
#>  7: 2.222947 2.835091 2.201546
#>  8: 2.445994 2.319239 2.957838
#>  9: 2.589568 2.968052 2.325981
#> 10: 2.362809 2.601693 2.488180
#> 11: 2.828788 2.448650 2.148184
#> 12: 2.417608 2.859471 2.173602
#> 13: 2.124411 2.845273 2.685512
#> 14: 2.550287 2.290982 2.907974
#> 15: 2.559767 2.574790 2.913932
#> 16: 2.817973 2.371475 2.507169
#> 17: 2.615906 2.073872 2.352266
#> 18: 2.272155 2.815918 2.816507
#> 19: 2.576817 2.384400 2.351955
#> 20: 2.098830 2.191486 2.543270
#> 21: 2.272697 2.095257 2.327098
#> 22: 2.115754 2.483171 2.103621
#> 23: 2.051961 2.804577 2.605381
#> 24: 2.060958 2.882190 2.382227
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.534748 1.564917 1.332330 2.276558 2.262789 2.426601
#>  2: 2002  2 1.748225 1.854544 1.429773 2.423900 2.842204 2.085967
#>  3: 2002  3 1.606936 1.994451 1.643857 2.710232 2.927665 2.391658
#>  4: 2002  4 1.617798 1.091708 1.603892 2.818661 2.910004 2.912244
#>  5: 2002  5 1.630905 1.136370 1.013376 2.191822 2.604021 2.266738
#>  6: 2002  6 1.057088 1.545894 1.747002 2.214650 2.152229 2.420270
#>  7: 2002  7 1.373436 1.816441 1.926416 2.222947 2.835091 2.201546
#>  8: 2002  8 1.476845 1.890898 1.029006 2.445994 2.319239 2.957838
#>  9: 2002  9 1.351720 1.674336 1.236270 2.589568 2.968052 2.325981
#> 10: 2002 10 1.888867 1.339109 1.372383 2.362809 2.601693 2.488180
#> 11: 2002 11 1.601524 1.357728 1.726824 2.828788 2.448650 2.148184
#> 12: 2002 12 1.956687 1.106301 1.630970 2.417608 2.859471 2.173602
#> 13: 2003  1 1.984424 1.681656 1.366496 2.124411 2.845273 2.685512
#> 14: 2003  2 1.428707 1.812605 1.988552 2.550287 2.290982 2.907974
#> 15: 2003  3 1.272681 1.712994 1.413446 2.559767 2.574790 2.913932
#> 16: 2003  4 1.583053 1.741310 1.435323 2.817973 2.371475 2.507169
#> 17: 2003  5 1.288060 1.232234 1.821573 2.615906 2.073872 2.352266
#> 18: 2003  6 1.960688 1.760492 1.346853 2.272155 2.815918 2.816507
#> 19: 2003  7 1.803285 1.615558 1.779977 2.576817 2.384400 2.351955
#> 20: 2003  8 1.106825 1.986852 1.695909 2.098830 2.191486 2.543270
#> 21: 2003  9 1.644605 1.046140 1.979762 2.272697 2.095257 2.327098
#> 22: 2003 10 1.075354 1.010623 1.496437 2.115754 2.483171 2.103621
#> 23: 2003 11 1.879148 1.409248 1.859771 2.051961 2.804577 2.605381
#> 24: 2003 12 1.589244 1.443581 1.670856 2.060958 2.882190 2.382227
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4