Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

Usage

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo         var1        var2        var3        var4        var5
#>  1: 2002  1 -0.243331859 -0.48958732 -0.65363509 -0.56015395 -0.44192878
#>  2: 2002  2 -0.438372679 -0.06317285 -0.04641969 -0.67922531 -0.81349278
#>  3: 2002  3 -0.230392701 -0.62014515 -0.75518159 -0.74101617 -0.66662035
#>  4: 2002  4 -0.021260290 -0.74006486 -0.54133867 -0.88668079 -0.59308125
#>  5: 2002  5 -0.319293582 -0.01638061 -0.09943052 -0.60281740 -0.47824600
#>  6: 2002  6 -0.271970459 -0.59983649 -0.86862829 -0.86948155 -0.64790075
#>  7: 2002  7 -0.164681837 -0.06115361 -0.40061261 -0.81508869 -0.83360991
#>  8: 2002  8 -0.990357647 -0.04438513 -0.35812028 -0.74284171 -0.63087244
#>  9: 2002  9 -0.009689041 -0.53061454 -0.27270228 -0.46639095 -0.07244091
#> 10: 2002 10 -0.621708035 -0.27741446 -0.56894716 -0.83618156 -0.78654695
#> 11: 2002 11 -0.102111101 -0.45569973 -0.07846749 -0.05427461 -0.04674087
#> 12: 2002 12 -0.395180358 -0.59556418 -0.02142390 -0.70101393 -0.28726770
#> 13: 2003  1 -0.829068860 -0.24810032 -0.32697237 -0.47778566 -0.75555701
#> 14: 2003  2 -0.191709620 -0.62522960 -0.58897053 -0.95867230 -0.11496167
#> 15: 2003  3 -0.551141095 -0.27839108 -0.89818762 -0.11441560 -0.15582517
#> 16: 2003  4 -0.538265174 -0.76962255 -0.65007792 -0.96089330 -0.92093032
#> 17: 2003  5 -0.497818080 -0.82604271 -0.25825769 -0.93830506 -0.29840919
#> 18: 2003  6 -0.371448783 -0.48551508 -0.52611850 -0.16714467 -0.20277695
#> 19: 2003  7 -0.792524689 -0.26611052 -0.48905519 -0.56785742 -0.66028709
#> 20: 2003  8 -0.713809568 -0.37719029 -0.67154591 -0.90632593 -0.68492677
#> 21: 2003  9 -0.301510160 -0.38525595 -0.15722781 -0.29366713 -0.30703779
#> 22: 2003 10 -0.740963046 -0.67125592 -0.80969026 -0.11409127 -0.64306289
#> 23: 2003 11 -0.732674474 -0.05540308 -0.31371902 -0.07513979 -0.73054161
#> 24: 2003 12 -0.910924197 -0.40925822 -0.88798295 -0.44452784 -0.55113969
#>       yr mo         var1        var2        var3        var4        var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2       var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.48958732 -0.65363509 -0.56015395 1.992725 1.898619 1.276535
#>  2: 2002  2 -0.06317285 -0.04641969 -0.67922531 1.540241 1.410586 1.703964
#>  3: 2002  3 -0.62014515 -0.75518159 -0.74101617 1.401946 1.882720 1.330390
#>  4: 2002  4 -0.74006486 -0.54133867 -0.88668079 1.677714 1.857649 1.621806
#>  5: 2002  5 -0.01638061 -0.09943052 -0.60281740 1.024563 1.479387 1.395754
#>  6: 2002  6 -0.59983649 -0.86862829 -0.86948155 1.363209 1.770790 1.773004
#>  7: 2002  7 -0.06115361 -0.40061261 -0.81508869 1.575870 1.852304 1.263989
#>  8: 2002  8 -0.04438513 -0.35812028 -0.74284171 1.715160 1.735310 1.337777
#>  9: 2002  9 -0.53061454 -0.27270228 -0.46639095 1.038681 1.671862 1.126383
#> 10: 2002 10 -0.27741446 -0.56894716 -0.83618156 1.392441 1.172624 1.008280
#> 11: 2002 11 -0.45569973 -0.07846749 -0.05427461 1.010842 1.533776 1.757230
#> 12: 2002 12 -0.59556418 -0.02142390 -0.70101393 1.765293 1.734969 1.303691
#> 13: 2003  1 -0.24810032 -0.32697237 -0.47778566 1.417057 1.819183 1.706479
#> 14: 2003  2 -0.62522960 -0.58897053 -0.95867230 1.302392 1.831115 1.195964
#> 15: 2003  3 -0.27839108 -0.89818762 -0.11441560 1.150242 1.439591 1.919003
#> 16: 2003  4 -0.76962255 -0.65007792 -0.96089330 1.362625 1.199014 1.111401
#> 17: 2003  5 -0.82604271 -0.25825769 -0.93830506 1.658189 1.293499 1.337033
#> 18: 2003  6 -0.48551508 -0.52611850 -0.16714467 1.008488 1.423004 1.986645
#> 19: 2003  7 -0.26611052 -0.48905519 -0.56785742 1.664063 1.329539 1.927547
#> 20: 2003  8 -0.37719029 -0.67154591 -0.90632593 1.209961 1.458562 1.922416
#> 21: 2003  9 -0.38525595 -0.15722781 -0.29366713 1.603686 1.180378 1.996860
#> 22: 2003 10 -0.67125592 -0.80969026 -0.11409127 1.585206 1.354479 1.772963
#> 23: 2003 11 -0.05540308 -0.31371902 -0.07513979 1.987702 1.624418 1.330343
#> 24: 2003 12 -0.40925822 -0.88798295 -0.44452784 1.548774 1.488554 1.641966
#>       yr mo       var_2       var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.598227 2.113478 2.347718
#>  2: 2.084351 2.773950 2.014232
#>  3: 2.909508 2.058525 2.759857
#>  4: 2.798273 2.032155 2.075983
#>  5: 2.970301 2.835867 2.855572
#>  6: 2.897173 2.487262 2.094360
#>  7: 2.278247 2.605619 2.297215
#>  8: 2.318390 2.100863 2.866309
#>  9: 2.392503 2.632842 2.197862
#> 10: 2.322433 2.316267 2.033486
#> 11: 2.548403 2.478963 2.258006
#> 12: 2.370746 2.944154 2.625965
#> 13: 2.450882 2.174669 2.055707
#> 14: 2.854040 2.030032 2.291017
#> 15: 2.756335 2.995703 2.219827
#> 16: 2.807108 2.981463 2.783901
#> 17: 2.288905 2.583192 2.795649
#> 18: 2.917944 2.912164 2.014992
#> 19: 2.487654 2.853991 2.096508
#> 20: 2.165203 2.247521 2.177960
#> 21: 2.088270 2.368485 2.340644
#> 22: 2.763280 2.952236 2.555862
#> 23: 2.330023 2.020801 2.087705
#> 24: 2.248931 2.289872 2.945923
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.992725 1.898619 1.276535 2.598227 2.113478 2.347718
#>  2: 2002  2 1.540241 1.410586 1.703964 2.084351 2.773950 2.014232
#>  3: 2002  3 1.401946 1.882720 1.330390 2.909508 2.058525 2.759857
#>  4: 2002  4 1.677714 1.857649 1.621806 2.798273 2.032155 2.075983
#>  5: 2002  5 1.024563 1.479387 1.395754 2.970301 2.835867 2.855572
#>  6: 2002  6 1.363209 1.770790 1.773004 2.897173 2.487262 2.094360
#>  7: 2002  7 1.575870 1.852304 1.263989 2.278247 2.605619 2.297215
#>  8: 2002  8 1.715160 1.735310 1.337777 2.318390 2.100863 2.866309
#>  9: 2002  9 1.038681 1.671862 1.126383 2.392503 2.632842 2.197862
#> 10: 2002 10 1.392441 1.172624 1.008280 2.322433 2.316267 2.033486
#> 11: 2002 11 1.010842 1.533776 1.757230 2.548403 2.478963 2.258006
#> 12: 2002 12 1.765293 1.734969 1.303691 2.370746 2.944154 2.625965
#> 13: 2003  1 1.417057 1.819183 1.706479 2.450882 2.174669 2.055707
#> 14: 2003  2 1.302392 1.831115 1.195964 2.854040 2.030032 2.291017
#> 15: 2003  3 1.150242 1.439591 1.919003 2.756335 2.995703 2.219827
#> 16: 2003  4 1.362625 1.199014 1.111401 2.807108 2.981463 2.783901
#> 17: 2003  5 1.658189 1.293499 1.337033 2.288905 2.583192 2.795649
#> 18: 2003  6 1.008488 1.423004 1.986645 2.917944 2.912164 2.014992
#> 19: 2003  7 1.664063 1.329539 1.927547 2.487654 2.853991 2.096508
#> 20: 2003  8 1.209961 1.458562 1.922416 2.165203 2.247521 2.177960
#> 21: 2003  9 1.603686 1.180378 1.996860 2.088270 2.368485 2.340644
#> 22: 2003 10 1.585206 1.354479 1.772963 2.763280 2.952236 2.555862
#> 23: 2003 11 1.987702 1.624418 1.330343 2.330023 2.020801 2.087705
#> 24: 2003 12 1.548774 1.488554 1.641966 2.248931 2.289872 2.945923
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4