Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2        var3       var4       var5
#>  1: 2002  1 -0.13596026 -0.87771510 -0.90503198 -0.3539921 -0.8236663
#>  2: 2002  2 -0.32472273 -0.49414320 -0.82962457 -0.7784864 -0.1815269
#>  3: 2002  3 -0.17733668 -0.67738110 -0.38750657 -0.8359277 -0.6414941
#>  4: 2002  4 -0.03559688 -0.48184136 -0.55733361 -0.1632536 -0.2558712
#>  5: 2002  5 -0.98649169 -0.14264545 -0.03585662 -0.4010504 -0.3789358
#>  6: 2002  6 -0.89445238 -0.53921349 -0.78516630 -0.1073291 -0.5749774
#>  7: 2002  7 -0.57054947 -0.65859367 -0.70062185 -0.8360838 -0.5485583
#>  8: 2002  8 -0.65769649 -0.59399725 -0.61890057 -0.3208212 -0.7720391
#>  9: 2002  9 -0.69391617 -0.06136773 -0.86423957 -0.6279025 -0.8786885
#> 10: 2002 10 -0.54735605 -0.95412158 -0.91124993 -0.8277060 -0.9511393
#> 11: 2002 11 -0.50100489 -0.36339375 -0.61846232 -0.2374635 -0.7831085
#> 12: 2002 12 -0.43167819 -0.81732347 -0.52652713 -0.5576559 -0.5089422

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2  var_layer3 var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.87771510 -0.90503198 -0.3539921    1.824512    1.374156
#>  2: 2002  2 -0.49414320 -0.82962457 -0.7784864    1.375334    1.683216
#>  3: 2002  3 -0.67738110 -0.38750657 -0.8359277    1.765495    1.195441
#>  4: 2002  4 -0.48184136 -0.55733361 -0.1632536    1.895924    1.025250
#>  5: 2002  5 -0.14264545 -0.03585662 -0.4010504    1.572827    1.447568
#>  6: 2002  6 -0.53921349 -0.78516630 -0.1073291    1.883674    1.231547
#>  7: 2002  7 -0.65859367 -0.70062185 -0.8360838    1.696433    1.596747
#>  8: 2002  8 -0.59399725 -0.61890057 -0.3208212    1.752706    1.655207
#>  9: 2002  9 -0.06136773 -0.86423957 -0.6279025    1.812064    1.456508
#> 10: 2002 10 -0.95412158 -0.91124993 -0.8277060    1.358158    1.949911
#> 11: 2002 11 -0.36339375 -0.61846232 -0.2374635    1.422154    1.717726
#> 12: 2002 12 -0.81732347 -0.52652713 -0.5576559    1.368990    1.163671
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.521245    2.173286    2.962706    2.068881
#>  2:    1.691450    2.612645    2.769023    2.470610
#>  3:    1.005847    2.601435    2.346402    2.162799
#>  4:    1.433779    2.391432    2.673654    2.984497
#>  5:    1.439488    2.883005    2.190481    2.321776
#>  6:    1.887755    2.992208    2.696415    2.423396
#>  7:    1.653866    2.727383    2.172431    2.492947
#>  8:    1.024183    2.933654    2.309261    2.839651
#>  9:    1.296930    2.127674    2.570473    2.071333
#> 10:    1.092817    2.929386    2.224816    2.943135
#> 11:    1.820044    2.937486    2.817663    2.195861
#> 12:    1.823019    2.198518    2.250020    2.049733

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.824512    1.374156    1.521245    2.173286    2.962706
#>  2: 2002  2    1.375334    1.683216    1.691450    2.612645    2.769023
#>  3: 2002  3    1.765495    1.195441    1.005847    2.601435    2.346402
#>  4: 2002  4    1.895924    1.025250    1.433779    2.391432    2.673654
#>  5: 2002  5    1.572827    1.447568    1.439488    2.883005    2.190481
#>  6: 2002  6    1.883674    1.231547    1.887755    2.992208    2.696415
#>  7: 2002  7    1.696433    1.596747    1.653866    2.727383    2.172431
#>  8: 2002  8    1.752706    1.655207    1.024183    2.933654    2.309261
#>  9: 2002  9    1.812064    1.456508    1.296930    2.127674    2.570473
#> 10: 2002 10    1.358158    1.949911    1.092817    2.929386    2.224816
#> 11: 2002 11    1.422154    1.717726    1.820044    2.937486    2.817663
#> 12: 2002 12    1.368990    1.163671    1.823019    2.198518    2.250020
#>     var2_layer4
#>  1:    2.068881
#>  2:    2.470610
#>  3:    2.162799
#>  4:    2.984497
#>  5:    2.321776
#>  6:    2.423396
#>  7:    2.492947
#>  8:    2.839651
#>  9:    2.071333
#> 10:    2.943135
#> 11:    2.195861
#> 12:    2.049733