Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1       var2        var3        var4        var5
#>  1: 2002  1 -0.26044662 -0.5275225 -0.01765187 -0.55911791 -0.72114545
#>  2: 2002  2 -0.91818920 -0.4392953 -0.07913925 -0.79263684 -0.04388789
#>  3: 2002  3 -0.96662087 -0.1465437 -0.86395520 -0.51541360 -0.38144328
#>  4: 2002  4 -0.14519965 -0.2977820 -0.37900882 -0.45555255 -0.97248916
#>  5: 2002  5 -0.25316228 -0.8989728 -0.04061916 -0.50461444 -0.37718254
#>  6: 2002  6 -0.85602602 -0.2475825 -0.76999149 -0.33195017 -0.21393706
#>  7: 2002  7 -0.01868047 -0.9018402 -0.00714043 -0.46163265 -0.45644703
#>  8: 2002  8 -0.03571327 -0.7417072 -0.07655145 -0.95436456 -0.98887871
#>  9: 2002  9 -0.35261227 -0.2494264 -0.10750864 -0.01230859 -0.42571726
#> 10: 2002 10 -0.18181377 -0.2534076 -0.46266320 -0.23263262 -0.53367410
#> 11: 2002 11 -0.62488517 -0.6213107 -0.32646627 -0.80073161 -0.05052239
#> 12: 2002 12 -0.63926894 -0.4376853 -0.88677347 -0.41238449 -0.37700701

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo      var_2       var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.5275225 -0.01765187 -0.55911791 1.583381 1.516790 1.603359
#>  2: 2002  2 -0.4392953 -0.07913925 -0.79263684 1.677406 1.968752 1.618214
#>  3: 2002  3 -0.1465437 -0.86395520 -0.51541360 1.774900 1.586921 1.459746
#>  4: 2002  4 -0.2977820 -0.37900882 -0.45555255 1.907859 1.838462 1.185900
#>  5: 2002  5 -0.8989728 -0.04061916 -0.50461444 1.130133 1.609057 1.011706
#>  6: 2002  6 -0.2475825 -0.76999149 -0.33195017 1.694514 1.011169 1.213058
#>  7: 2002  7 -0.9018402 -0.00714043 -0.46163265 1.681584 1.385671 1.434968
#>  8: 2002  8 -0.7417072 -0.07655145 -0.95436456 1.014324 1.826376 1.631161
#>  9: 2002  9 -0.2494264 -0.10750864 -0.01230859 1.281610 1.817036 1.971949
#> 10: 2002 10 -0.2534076 -0.46266320 -0.23263262 1.454697 1.496553 1.302598
#> 11: 2002 11 -0.6213107 -0.32646627 -0.80073161 1.052666 1.864502 1.199080
#> 12: 2002 12 -0.4376853 -0.88677347 -0.41238449 1.611126 1.405800 1.279769
#>       var2_2   var2_3   var2_4
#>  1: 2.404431 2.034675 2.888916
#>  2: 2.969851 2.636369 2.164340
#>  3: 2.144542 2.381472 2.319157
#>  4: 2.817583 2.011659 2.222922
#>  5: 2.647659 2.309105 2.964435
#>  6: 2.071517 2.833819 2.865145
#>  7: 2.451566 2.825573 2.795343
#>  8: 2.806668 2.516375 2.507404
#>  9: 2.967896 2.597354 2.237342
#> 10: 2.420796 2.235249 2.076118
#> 11: 2.071585 2.162098 2.915111
#> 12: 2.189528 2.861301 2.796308

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.583381 1.516790 1.603359 2.404431 2.034675 2.888916
#>  2: 2002  2 1.677406 1.968752 1.618214 2.969851 2.636369 2.164340
#>  3: 2002  3 1.774900 1.586921 1.459746 2.144542 2.381472 2.319157
#>  4: 2002  4 1.907859 1.838462 1.185900 2.817583 2.011659 2.222922
#>  5: 2002  5 1.130133 1.609057 1.011706 2.647659 2.309105 2.964435
#>  6: 2002  6 1.694514 1.011169 1.213058 2.071517 2.833819 2.865145
#>  7: 2002  7 1.681584 1.385671 1.434968 2.451566 2.825573 2.795343
#>  8: 2002  8 1.014324 1.826376 1.631161 2.806668 2.516375 2.507404
#>  9: 2002  9 1.281610 1.817036 1.971949 2.967896 2.597354 2.237342
#> 10: 2002 10 1.454697 1.496553 1.302598 2.420796 2.235249 2.076118
#> 11: 2002 11 1.052666 1.864502 1.199080 2.071585 2.162098 2.915111
#> 12: 2002 12 1.611126 1.405800 1.279769 2.189528 2.861301 2.796308