Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1       var2       var3        var4        var5
#>  1: 2002  1 -0.97703539 -0.9413779 -0.8434001 -0.37020526 -0.01821506
#>  2: 2002  2 -0.33732417 -0.4280571 -0.3237525 -0.78036673 -0.69761624
#>  3: 2002  3 -0.62518934 -0.9445636 -0.2328844 -0.94222792 -0.06160635
#>  4: 2002  4 -0.82558488 -0.7678596 -0.4504922 -0.01348690 -0.68158339
#>  5: 2002  5 -0.53557572 -0.6476108 -0.9429244 -0.35328195 -0.48148052
#>  6: 2002  6 -0.04943712 -0.9804190 -0.5993990 -0.56538611 -0.77490333
#>  7: 2002  7 -0.61062442 -0.7437004 -0.8489526 -0.04734706 -0.27740617
#>  8: 2002  8 -0.75823232 -0.2864438 -0.8131531 -0.30850198 -0.10338135
#>  9: 2002  9 -0.58396019 -0.1889814 -0.7155412 -0.75114198 -0.91712437
#> 10: 2002 10 -0.51348999 -0.7281824 -0.1889986 -0.33148234 -0.21333112
#> 11: 2002 11 -0.18570808 -0.1069862 -0.7368987 -0.11429321 -0.27425899
#> 12: 2002 12 -0.66896053 -0.2088866 -0.7266292 -0.89127746 -0.28172337

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo var_layer2 var_layer3  var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.9413779 -0.8434001 -0.37020526    1.478425    1.279717
#>  2: 2002  2 -0.4280571 -0.3237525 -0.78036673    1.988507    1.933155
#>  3: 2002  3 -0.9445636 -0.2328844 -0.94222792    1.332864    1.537069
#>  4: 2002  4 -0.7678596 -0.4504922 -0.01348690    1.733004    1.561863
#>  5: 2002  5 -0.6476108 -0.9429244 -0.35328195    1.449843    1.377558
#>  6: 2002  6 -0.9804190 -0.5993990 -0.56538611    1.431282    1.548340
#>  7: 2002  7 -0.7437004 -0.8489526 -0.04734706    1.519408    1.562739
#>  8: 2002  8 -0.2864438 -0.8131531 -0.30850198    1.650152    1.049361
#>  9: 2002  9 -0.1889814 -0.7155412 -0.75114198    1.074298    1.495275
#> 10: 2002 10 -0.7281824 -0.1889986 -0.33148234    1.040675    1.890206
#> 11: 2002 11 -0.1069862 -0.7368987 -0.11429321    1.776337    1.393318
#> 12: 2002 12 -0.2088866 -0.7266292 -0.89127746    1.655498    1.929456
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.123509    2.153338    2.249860    2.839446
#>  2:    1.312202    2.439530    2.003869    2.292578
#>  3:    1.839208    2.084195    2.682351    2.770782
#>  4:    1.316315    2.722260    2.727531    2.299176
#>  5:    1.894242    2.231529    2.017751    2.141379
#>  6:    1.910482    2.433357    2.722342    2.212222
#>  7:    1.749261    2.838145    2.696788    2.961774
#>  8:    1.995209    2.569126    2.260332    2.418803
#>  9:    1.144337    2.735273    2.140422    2.625908
#> 10:    1.220146    2.593745    2.845197    2.979217
#> 11:    1.758436    2.840487    2.256765    2.051439
#> 12:    1.200573    2.995122    2.935997    2.897856

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.478425    1.279717    1.123509    2.153338    2.249860
#>  2: 2002  2    1.988507    1.933155    1.312202    2.439530    2.003869
#>  3: 2002  3    1.332864    1.537069    1.839208    2.084195    2.682351
#>  4: 2002  4    1.733004    1.561863    1.316315    2.722260    2.727531
#>  5: 2002  5    1.449843    1.377558    1.894242    2.231529    2.017751
#>  6: 2002  6    1.431282    1.548340    1.910482    2.433357    2.722342
#>  7: 2002  7    1.519408    1.562739    1.749261    2.838145    2.696788
#>  8: 2002  8    1.650152    1.049361    1.995209    2.569126    2.260332
#>  9: 2002  9    1.074298    1.495275    1.144337    2.735273    2.140422
#> 10: 2002 10    1.040675    1.890206    1.220146    2.593745    2.845197
#> 11: 2002 11    1.776337    1.393318    1.758436    2.840487    2.256765
#> 12: 2002 12    1.655498    1.929456    1.200573    2.995122    2.935997
#>     var2_layer4
#>  1:    2.839446
#>  2:    2.292578
#>  3:    2.770782
#>  4:    2.299176
#>  5:    2.141379
#>  6:    2.212222
#>  7:    2.961774
#>  8:    2.418803
#>  9:    2.625908
#> 10:    2.979217
#> 11:    2.051439
#> 12:    2.897856