Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

Usage

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1         var2       var3        var4       var5
#>  1: 2002  1 -0.40685905 -0.038913906 -0.3852714 -0.26633688 -0.5168815
#>  2: 2002  2 -0.50886970 -0.105135930 -0.2549385 -0.05218167 -0.4527945
#>  3: 2002  3 -0.86092451 -0.824709793 -0.3122264 -0.14357683 -0.9018557
#>  4: 2002  4 -0.71259206 -0.060311608 -0.3002745 -0.55562416 -0.6423637
#>  5: 2002  5 -0.94204540 -0.774500640 -0.9170300 -0.78384001 -0.2172175
#>  6: 2002  6 -0.66901342 -0.597407783 -0.4980525 -0.24248480 -0.8470037
#>  7: 2002  7 -0.93631666 -0.113984567 -0.7361860 -0.16562349 -0.1825540
#>  8: 2002  8 -0.53863810 -0.930654002 -0.7747696 -0.94643691 -0.3997954
#>  9: 2002  9 -0.92844412 -0.032967510 -0.7902685 -0.86354353 -0.6349416
#> 10: 2002 10 -0.82628854 -0.224697740 -0.9553497 -0.64632906 -0.2642625
#> 11: 2002 11 -0.82095213 -0.087306896 -0.9191578 -0.49692778 -0.1318929
#> 12: 2002 12 -0.29543414 -0.295173273 -0.4760566 -0.56501497 -0.4847976
#> 13: 2003  1 -0.86845002 -0.434536646 -0.9733818 -0.89042727 -0.7959809
#> 14: 2003  2 -0.59642296 -0.168235459 -0.9559031 -0.19860696 -0.7217191
#> 15: 2003  3 -0.04650102 -0.308043568 -0.5299162 -0.81305203 -0.6010901
#> 16: 2003  4 -0.83875668 -0.986482642 -0.6891765 -0.42850444 -0.8457587
#> 17: 2003  5 -0.96865164 -0.711635594 -0.1407419 -0.48109510 -0.0403142
#> 18: 2003  6 -0.59461746 -0.566559017 -0.8302460 -0.54548501 -0.8176859
#> 19: 2003  7 -0.38635034 -0.214379919 -0.3737967 -0.86155452 -0.2286807
#> 20: 2003  8 -0.47746300 -0.005087654 -0.7567358 -0.63514429 -0.4786681
#> 21: 2003  9 -0.27470823 -0.034814110 -0.9008383 -0.86501732 -0.5056770
#> 22: 2003 10 -0.92262470 -0.710392124 -0.3658929 -0.41770128 -0.5598135
#> 23: 2003 11 -0.35959468 -0.532414733 -0.1692651 -0.81088944 -0.3651311
#> 24: 2003 12 -0.21311232 -0.872090190 -0.3194298 -0.27005285 -0.9456216
#>       yr mo        var1         var2       var3        var4       var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo        var_2      var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.038913906 -0.3852714 -0.26633688 1.335861 1.477668 1.412423
#>  2: 2002  2 -0.105135930 -0.2549385 -0.05218167 1.238991 1.240586 1.791613
#>  3: 2002  3 -0.824709793 -0.3122264 -0.14357683 1.509617 1.196654 1.841452
#>  4: 2002  4 -0.060311608 -0.3002745 -0.55562416 1.590409 1.888206 1.895991
#>  5: 2002  5 -0.774500640 -0.9170300 -0.78384001 1.317580 1.834977 1.862947
#>  6: 2002  6 -0.597407783 -0.4980525 -0.24248480 1.158176 1.550282 1.480269
#>  7: 2002  7 -0.113984567 -0.7361860 -0.16562349 1.179500 1.130821 1.215004
#>  8: 2002  8 -0.930654002 -0.7747696 -0.94643691 1.819336 1.818990 1.511231
#>  9: 2002  9 -0.032967510 -0.7902685 -0.86354353 1.889579 1.204125 1.913624
#> 10: 2002 10 -0.224697740 -0.9553497 -0.64632906 1.089115 1.673294 1.616726
#> 11: 2002 11 -0.087306896 -0.9191578 -0.49692778 1.110711 1.527792 1.529122
#> 12: 2002 12 -0.295173273 -0.4760566 -0.56501497 1.731076 1.845856 1.556299
#> 13: 2003  1 -0.434536646 -0.9733818 -0.89042727 1.814556 1.094652 1.853577
#> 14: 2003  2 -0.168235459 -0.9559031 -0.19860696 1.973840 1.871996 1.075858
#> 15: 2003  3 -0.308043568 -0.5299162 -0.81305203 1.436379 1.597699 1.656042
#> 16: 2003  4 -0.986482642 -0.6891765 -0.42850444 1.486116 1.081863 1.563821
#> 17: 2003  5 -0.711635594 -0.1407419 -0.48109510 1.593008 1.213152 1.150371
#> 18: 2003  6 -0.566559017 -0.8302460 -0.54548501 1.537872 1.089407 1.731099
#> 19: 2003  7 -0.214379919 -0.3737967 -0.86155452 1.711629 1.791993 1.777396
#> 20: 2003  8 -0.005087654 -0.7567358 -0.63514429 1.394687 1.105290 1.047350
#> 21: 2003  9 -0.034814110 -0.9008383 -0.86501732 1.990224 1.359402 1.883173
#> 22: 2003 10 -0.710392124 -0.3658929 -0.41770128 1.874368 1.549509 1.566684
#> 23: 2003 11 -0.532414733 -0.1692651 -0.81088944 1.826964 1.927582 1.999647
#> 24: 2003 12 -0.872090190 -0.3194298 -0.27005285 1.935455 1.867442 1.295032
#>       yr mo        var_2      var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.560905 2.133399 2.183682
#>  2: 2.556886 2.533793 2.721709
#>  3: 2.001040 2.927942 2.921249
#>  4: 2.864435 2.033169 2.799432
#>  5: 2.368465 2.382388 2.780307
#>  6: 2.274346 2.892653 2.232555
#>  7: 2.105118 2.368538 2.918231
#>  8: 2.365046 2.392459 2.829918
#>  9: 2.321946 2.288981 2.397935
#> 10: 2.734245 2.347418 2.229702
#> 11: 2.162709 2.060793 2.873150
#> 12: 2.076458 2.752116 2.966973
#> 13: 2.065566 2.046434 2.554750
#> 14: 2.905622 2.821441 2.316869
#> 15: 2.967143 2.913402 2.791652
#> 16: 2.954751 2.164022 2.997123
#> 17: 2.492454 2.559997 2.811139
#> 18: 2.593306 2.730551 2.811825
#> 19: 2.337210 2.435749 2.995223
#> 20: 2.245196 2.446053 2.478141
#> 21: 2.219633 2.089151 2.385219
#> 22: 2.642573 2.778177 2.354184
#> 23: 2.757709 2.448637 2.576258
#> 24: 2.296336 2.624157 2.123923
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.335861 1.477668 1.412423 2.560905 2.133399 2.183682
#>  2: 2002  2 1.238991 1.240586 1.791613 2.556886 2.533793 2.721709
#>  3: 2002  3 1.509617 1.196654 1.841452 2.001040 2.927942 2.921249
#>  4: 2002  4 1.590409 1.888206 1.895991 2.864435 2.033169 2.799432
#>  5: 2002  5 1.317580 1.834977 1.862947 2.368465 2.382388 2.780307
#>  6: 2002  6 1.158176 1.550282 1.480269 2.274346 2.892653 2.232555
#>  7: 2002  7 1.179500 1.130821 1.215004 2.105118 2.368538 2.918231
#>  8: 2002  8 1.819336 1.818990 1.511231 2.365046 2.392459 2.829918
#>  9: 2002  9 1.889579 1.204125 1.913624 2.321946 2.288981 2.397935
#> 10: 2002 10 1.089115 1.673294 1.616726 2.734245 2.347418 2.229702
#> 11: 2002 11 1.110711 1.527792 1.529122 2.162709 2.060793 2.873150
#> 12: 2002 12 1.731076 1.845856 1.556299 2.076458 2.752116 2.966973
#> 13: 2003  1 1.814556 1.094652 1.853577 2.065566 2.046434 2.554750
#> 14: 2003  2 1.973840 1.871996 1.075858 2.905622 2.821441 2.316869
#> 15: 2003  3 1.436379 1.597699 1.656042 2.967143 2.913402 2.791652
#> 16: 2003  4 1.486116 1.081863 1.563821 2.954751 2.164022 2.997123
#> 17: 2003  5 1.593008 1.213152 1.150371 2.492454 2.559997 2.811139
#> 18: 2003  6 1.537872 1.089407 1.731099 2.593306 2.730551 2.811825
#> 19: 2003  7 1.711629 1.791993 1.777396 2.337210 2.435749 2.995223
#> 20: 2003  8 1.394687 1.105290 1.047350 2.245196 2.446053 2.478141
#> 21: 2003  9 1.990224 1.359402 1.883173 2.219633 2.089151 2.385219
#> 22: 2003 10 1.874368 1.549509 1.566684 2.642573 2.778177 2.354184
#> 23: 2003 11 1.826964 1.927582 1.999647 2.757709 2.448637 2.576258
#> 24: 2003 12 1.935455 1.867442 1.295032 2.296336 2.624157 2.123923
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4