Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2        var3       var4         var5
#>  1: 2002  1 -0.86969661 -0.54078348 -0.52734168 -0.4667840 -0.131930229
#>  2: 2002  2 -0.38401501 -0.36558686 -0.56665031 -0.9563372 -0.680717258
#>  3: 2002  3 -0.77800524 -0.82251896 -0.57102408 -0.4973475 -0.896393009
#>  4: 2002  4 -0.72339668 -0.49987441 -0.65766115 -0.9768596 -0.869730518
#>  5: 2002  5 -0.17288089 -0.88585868 -0.22807998 -0.8807026 -0.003620361
#>  6: 2002  6 -0.18375683 -0.40290508 -0.24933051 -0.3915416 -0.032770021
#>  7: 2002  7 -0.36102281 -0.81843464 -0.48294337 -0.3248965 -0.554456603
#>  8: 2002  8 -0.54066236 -0.58500848 -0.68698469 -0.5246223 -0.896034032
#>  9: 2002  9 -0.77785416 -0.94609506 -0.02255536 -0.6331841 -0.999752474
#> 10: 2002 10 -0.29975263 -0.90103345 -0.29822680 -0.2719952 -0.651191115
#> 11: 2002 11 -0.41655972 -0.61805674 -0.26534165 -0.8666004 -0.597849296
#> 12: 2002 12 -0.52575428 -0.14515569 -0.59547115 -0.8684022 -0.095060781
#> 13: 2003  1 -0.69734229 -0.13135864 -0.70028420 -0.1376382 -0.796548988
#> 14: 2003  2 -0.72711604 -0.25932878 -0.49136894 -0.9336707 -0.025436422
#> 15: 2003  3 -0.15779138 -0.37021672 -0.34373278 -0.2515193 -0.509703287
#> 16: 2003  4 -0.21392263 -0.63501435 -0.69206219 -0.3835354 -0.485532760
#> 17: 2003  5 -0.37548112 -0.88090377 -0.68056653 -0.6496853 -0.975926238
#> 18: 2003  6 -0.24422248 -0.96425741 -0.70866328 -0.3339929 -0.350090812
#> 19: 2003  7 -0.45109657 -0.77395029 -0.81316979 -0.7814495 -0.574916796
#> 20: 2003  8 -0.87646461 -0.08580645 -0.64266507 -0.6655363 -0.302516399
#> 21: 2003  9 -0.93850287 -0.53754274 -0.99354537 -0.7679302 -0.067720170
#> 22: 2003 10 -0.77156602 -0.76652982 -0.35699902 -0.2846181 -0.447809315
#> 23: 2003 11 -0.99614905 -0.67625040 -0.76559048 -0.9851964 -0.308803104
#> 24: 2003 12 -0.06582934 -0.16293926 -0.04050749 -0.7263386 -0.395117771
#>       yr mo        var1        var2        var3       var4         var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2       var_3      var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.54078348 -0.52734168 -0.4667840 1.476485 1.593314 1.253228
#>  2: 2002  2 -0.36558686 -0.56665031 -0.9563372 1.541840 1.519322 1.236318
#>  3: 2002  3 -0.82251896 -0.57102408 -0.4973475 1.354051 1.210179 1.166614
#>  4: 2002  4 -0.49987441 -0.65766115 -0.9768596 1.180545 1.146108 1.445649
#>  5: 2002  5 -0.88585868 -0.22807998 -0.8807026 1.511709 1.992721 1.937265
#>  6: 2002  6 -0.40290508 -0.24933051 -0.3915416 1.499667 1.577960 1.263161
#>  7: 2002  7 -0.81843464 -0.48294337 -0.3248965 1.397755 1.540378 1.981024
#>  8: 2002  8 -0.58500848 -0.68698469 -0.5246223 1.635037 1.729032 1.058206
#>  9: 2002  9 -0.94609506 -0.02255536 -0.6331841 1.557350 1.263481 1.798018
#> 10: 2002 10 -0.90103345 -0.29822680 -0.2719952 1.765524 1.076773 1.662369
#> 11: 2002 11 -0.61805674 -0.26534165 -0.8666004 1.658617 1.634670 1.025054
#> 12: 2002 12 -0.14515569 -0.59547115 -0.8684022 1.918892 1.749588 1.270514
#> 13: 2003  1 -0.13135864 -0.70028420 -0.1376382 1.814626 1.713136 1.514975
#> 14: 2003  2 -0.25932878 -0.49136894 -0.9336707 1.277210 1.598514 1.679247
#> 15: 2003  3 -0.37021672 -0.34373278 -0.2515193 1.368924 1.271611 1.068180
#> 16: 2003  4 -0.63501435 -0.69206219 -0.3835354 1.355147 1.400672 1.491239
#> 17: 2003  5 -0.88090377 -0.68056653 -0.6496853 1.128321 1.513581 1.908753
#> 18: 2003  6 -0.96425741 -0.70866328 -0.3339929 1.097041 1.517362 1.562956
#> 19: 2003  7 -0.77395029 -0.81316979 -0.7814495 1.161269 1.782915 1.191585
#> 20: 2003  8 -0.08580645 -0.64266507 -0.6655363 1.502778 1.544497 1.319261
#> 21: 2003  9 -0.53754274 -0.99354537 -0.7679302 1.939364 1.014016 1.950623
#> 22: 2003 10 -0.76652982 -0.35699902 -0.2846181 1.742558 1.902810 1.922345
#> 23: 2003 11 -0.67625040 -0.76559048 -0.9851964 1.606705 1.930201 1.054447
#> 24: 2003 12 -0.16293926 -0.04050749 -0.7263386 1.734071 1.424714 1.473773
#>       yr mo       var_2       var_3      var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.145309 2.130636 2.500791
#>  2: 2.482202 2.588128 2.969682
#>  3: 2.834604 2.368363 2.403937
#>  4: 2.239416 2.417825 2.645936
#>  5: 2.940831 2.619017 2.735928
#>  6: 2.753210 2.497795 2.934200
#>  7: 2.061153 2.319316 2.624179
#>  8: 2.564606 2.468053 2.600294
#>  9: 2.480262 2.856493 2.896933
#> 10: 2.832704 2.896897 2.929113
#> 11: 2.284255 2.688944 2.736084
#> 12: 2.031023 2.290368 2.495161
#> 13: 2.103059 2.934760 2.045936
#> 14: 2.395230 2.828026 2.112630
#> 15: 2.257111 2.746852 2.964942
#> 16: 2.046754 2.330640 2.359968
#> 17: 2.830403 2.149903 2.735407
#> 18: 2.377129 2.836918 2.145497
#> 19: 2.367658 2.786422 2.841707
#> 20: 2.866738 2.685022 2.018441
#> 21: 2.497198 2.389556 2.541163
#> 22: 2.593819 2.475217 2.990393
#> 23: 2.700801 2.122056 2.239540
#> 24: 2.603878 2.864239 2.797813
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.476485 1.593314 1.253228 2.145309 2.130636 2.500791
#>  2: 2002  2 1.541840 1.519322 1.236318 2.482202 2.588128 2.969682
#>  3: 2002  3 1.354051 1.210179 1.166614 2.834604 2.368363 2.403937
#>  4: 2002  4 1.180545 1.146108 1.445649 2.239416 2.417825 2.645936
#>  5: 2002  5 1.511709 1.992721 1.937265 2.940831 2.619017 2.735928
#>  6: 2002  6 1.499667 1.577960 1.263161 2.753210 2.497795 2.934200
#>  7: 2002  7 1.397755 1.540378 1.981024 2.061153 2.319316 2.624179
#>  8: 2002  8 1.635037 1.729032 1.058206 2.564606 2.468053 2.600294
#>  9: 2002  9 1.557350 1.263481 1.798018 2.480262 2.856493 2.896933
#> 10: 2002 10 1.765524 1.076773 1.662369 2.832704 2.896897 2.929113
#> 11: 2002 11 1.658617 1.634670 1.025054 2.284255 2.688944 2.736084
#> 12: 2002 12 1.918892 1.749588 1.270514 2.031023 2.290368 2.495161
#> 13: 2003  1 1.814626 1.713136 1.514975 2.103059 2.934760 2.045936
#> 14: 2003  2 1.277210 1.598514 1.679247 2.395230 2.828026 2.112630
#> 15: 2003  3 1.368924 1.271611 1.068180 2.257111 2.746852 2.964942
#> 16: 2003  4 1.355147 1.400672 1.491239 2.046754 2.330640 2.359968
#> 17: 2003  5 1.128321 1.513581 1.908753 2.830403 2.149903 2.735407
#> 18: 2003  6 1.097041 1.517362 1.562956 2.377129 2.836918 2.145497
#> 19: 2003  7 1.161269 1.782915 1.191585 2.367658 2.786422 2.841707
#> 20: 2003  8 1.502778 1.544497 1.319261 2.866738 2.685022 2.018441
#> 21: 2003  9 1.939364 1.014016 1.950623 2.497198 2.389556 2.541163
#> 22: 2003 10 1.742558 1.902810 1.922345 2.593819 2.475217 2.990393
#> 23: 2003 11 1.606705 1.930201 1.054447 2.700801 2.122056 2.239540
#> 24: 2003 12 1.734071 1.424714 1.473773 2.603878 2.864239 2.797813
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4