Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1        var2         var3        var4          var5
#>  1: 2002  1 -0.68353073 -0.02046564 -0.226743111 -0.33497459 -0.5271687480
#>  2: 2002  2 -0.13427638 -0.44390561 -0.480796539 -0.64225464 -0.2124742123
#>  3: 2002  3 -0.20760209 -0.84097232 -0.403308697 -0.17072046 -0.4601025286
#>  4: 2002  4 -0.72809585 -0.11590278 -0.631586418 -0.79016241 -0.0982104659
#>  5: 2002  5 -0.63678080 -0.22577691 -0.271274145 -0.45725755 -0.3419588048
#>  6: 2002  6 -0.55302206 -0.16202456 -0.632931006 -0.52142478 -0.2576362109
#>  7: 2002  7 -0.19020816 -0.51669899 -0.088889445 -0.44155177 -0.8333449038
#>  8: 2002  8 -0.80372852 -0.65228732 -0.015624921 -0.35091482 -0.2376582832
#>  9: 2002  9 -0.86931498 -0.66945362 -0.967959885 -0.62194425 -0.3460274807
#> 10: 2002 10 -0.09999642 -0.68711007 -0.938179615 -0.64827951 -0.2652717971
#> 11: 2002 11 -0.79378008 -0.39252266 -0.257069263 -0.99311949 -0.1151843844
#> 12: 2002 12 -0.24676869 -0.42664199 -0.047590100 -0.77377615 -0.0001451657
#> 13: 2003  1 -0.64957683 -0.19335221 -0.806583222 -0.63986809 -0.2213571146
#> 14: 2003  2 -0.68708773 -0.72424614 -0.666102357 -0.96851481 -0.2839590993
#> 15: 2003  3 -0.45201771 -0.09781093 -0.596232178 -0.03019744 -0.7013076795
#> 16: 2003  4 -0.83927608 -0.38223658 -0.192089236 -0.59923623 -0.7972200916
#> 17: 2003  5 -0.83851919 -0.09064635 -0.309692563 -0.32887522 -0.2442236342
#> 18: 2003  6 -0.45312643 -0.57596988 -0.002639699 -0.90569361 -0.8657086371
#> 19: 2003  7 -0.44401542 -0.97223038 -0.721072662 -0.68494784 -0.4835595412
#> 20: 2003  8 -0.04930826 -0.10758523 -0.178271192 -0.78589829 -0.1326485209
#> 21: 2003  9 -0.97554034 -0.86125942 -0.014995577 -0.82912879 -0.1861350089
#> 22: 2003 10 -0.19477701 -0.60036913 -0.084137138 -0.50838438 -0.1339634953
#> 23: 2003 11 -0.26485150 -0.51476858 -0.432465339 -0.80931193 -0.5565804932
#> 24: 2003 12 -0.38997933 -0.84495758 -0.120312874 -0.07675393 -0.7697060411
#>       yr mo        var1        var2         var3        var4          var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2        var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.02046564 -0.226743111 -0.33497459 1.463204 1.497727 1.028561
#>  2: 2002  2 -0.44390561 -0.480796539 -0.64225464 1.154034 1.227548 1.303073
#>  3: 2002  3 -0.84097232 -0.403308697 -0.17072046 1.412202 1.654816 1.000157
#>  4: 2002  4 -0.11590278 -0.631586418 -0.79016241 1.396698 1.590065 1.423953
#>  5: 2002  5 -0.22577691 -0.271274145 -0.45725755 1.302196 1.842167 1.320766
#>  6: 2002  6 -0.16202456 -0.632931006 -0.52142478 1.361079 1.011964 1.051439
#>  7: 2002  7 -0.51669899 -0.088889445 -0.44155177 1.791406 1.595571 1.122169
#>  8: 2002  8 -0.65228732 -0.015624921 -0.35091482 1.592752 1.775895 1.763662
#>  9: 2002  9 -0.66945362 -0.967959885 -0.62194425 1.717225 1.128470 1.611670
#> 10: 2002 10 -0.68711007 -0.938179615 -0.64827951 1.724157 1.495751 1.843449
#> 11: 2002 11 -0.39252266 -0.257069263 -0.99311949 1.335883 1.845153 1.749842
#> 12: 2002 12 -0.42664199 -0.047590100 -0.77377615 1.796257 1.432386 1.937484
#> 13: 2003  1 -0.19335221 -0.806583222 -0.63986809 1.938686 1.982389 1.988466
#> 14: 2003  2 -0.72424614 -0.666102357 -0.96851481 1.238623 1.389799 1.091605
#> 15: 2003  3 -0.09781093 -0.596232178 -0.03019744 1.574106 1.652591 1.868491
#> 16: 2003  4 -0.38223658 -0.192089236 -0.59923623 1.981939 1.099368 1.922260
#> 17: 2003  5 -0.09064635 -0.309692563 -0.32887522 1.374449 1.570735 1.373958
#> 18: 2003  6 -0.57596988 -0.002639699 -0.90569361 1.353300 1.296332 1.316667
#> 19: 2003  7 -0.97223038 -0.721072662 -0.68494784 1.465041 1.428813 1.736018
#> 20: 2003  8 -0.10758523 -0.178271192 -0.78589829 1.903003 1.449097 1.053354
#> 21: 2003  9 -0.86125942 -0.014995577 -0.82912879 1.445536 1.357616 1.969728
#> 22: 2003 10 -0.60036913 -0.084137138 -0.50838438 1.564656 1.987375 1.193712
#> 23: 2003 11 -0.51476858 -0.432465339 -0.80931193 1.651363 1.634821 1.582804
#> 24: 2003 12 -0.84495758 -0.120312874 -0.07675393 1.000243 1.759455 1.105044
#>       yr mo       var_2        var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.038877 2.458432 2.051585
#>  2: 2.245711 2.176852 2.341487
#>  3: 2.512658 2.777687 2.406076
#>  4: 2.847737 2.081283 2.094257
#>  5: 2.053888 2.142482 2.665443
#>  6: 2.064751 2.588360 2.688297
#>  7: 2.767073 2.387010 2.058234
#>  8: 2.784932 2.788199 2.299703
#>  9: 2.157991 2.080123 2.265899
#> 10: 2.291986 2.314208 2.273091
#> 11: 2.015505 2.314651 2.866303
#> 12: 2.718275 2.649475 2.671075
#> 13: 2.534348 2.572811 2.244068
#> 14: 2.001733 2.019339 2.706007
#> 15: 2.231553 2.833341 2.461190
#> 16: 2.313197 2.391012 2.610992
#> 17: 2.760320 2.702738 2.813674
#> 18: 2.958027 2.554922 2.567050
#> 19: 2.045359 2.742072 2.131186
#> 20: 2.351658 2.020600 2.289502
#> 21: 2.203889 2.388362 2.408145
#> 22: 2.969459 2.850681 2.579798
#> 23: 2.562900 2.661407 2.962852
#> 24: 2.159005 2.016768 2.598586
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.463204 1.497727 1.028561 2.038877 2.458432 2.051585
#>  2: 2002  2 1.154034 1.227548 1.303073 2.245711 2.176852 2.341487
#>  3: 2002  3 1.412202 1.654816 1.000157 2.512658 2.777687 2.406076
#>  4: 2002  4 1.396698 1.590065 1.423953 2.847737 2.081283 2.094257
#>  5: 2002  5 1.302196 1.842167 1.320766 2.053888 2.142482 2.665443
#>  6: 2002  6 1.361079 1.011964 1.051439 2.064751 2.588360 2.688297
#>  7: 2002  7 1.791406 1.595571 1.122169 2.767073 2.387010 2.058234
#>  8: 2002  8 1.592752 1.775895 1.763662 2.784932 2.788199 2.299703
#>  9: 2002  9 1.717225 1.128470 1.611670 2.157991 2.080123 2.265899
#> 10: 2002 10 1.724157 1.495751 1.843449 2.291986 2.314208 2.273091
#> 11: 2002 11 1.335883 1.845153 1.749842 2.015505 2.314651 2.866303
#> 12: 2002 12 1.796257 1.432386 1.937484 2.718275 2.649475 2.671075
#> 13: 2003  1 1.938686 1.982389 1.988466 2.534348 2.572811 2.244068
#> 14: 2003  2 1.238623 1.389799 1.091605 2.001733 2.019339 2.706007
#> 15: 2003  3 1.574106 1.652591 1.868491 2.231553 2.833341 2.461190
#> 16: 2003  4 1.981939 1.099368 1.922260 2.313197 2.391012 2.610992
#> 17: 2003  5 1.374449 1.570735 1.373958 2.760320 2.702738 2.813674
#> 18: 2003  6 1.353300 1.296332 1.316667 2.958027 2.554922 2.567050
#> 19: 2003  7 1.465041 1.428813 1.736018 2.045359 2.742072 2.131186
#> 20: 2003  8 1.903003 1.449097 1.053354 2.351658 2.020600 2.289502
#> 21: 2003  9 1.445536 1.357616 1.969728 2.203889 2.388362 2.408145
#> 22: 2003 10 1.564656 1.987375 1.193712 2.969459 2.850681 2.579798
#> 23: 2003 11 1.651363 1.634821 1.582804 2.562900 2.661407 2.962852
#> 24: 2003 12 1.000243 1.759455 1.105044 2.159005 2.016768 2.598586
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4