Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo       var1        var2       var3       var4        var5
#>  1: 2002  1 -0.2118593 -0.29932824 -0.5948099 -0.7323255 -0.02031751
#>  2: 2002  2 -0.5387224 -0.18832581 -0.2859579 -0.5484089 -0.74074526
#>  3: 2002  3 -0.3744613 -0.49342411 -0.8125183 -0.4164660 -0.80806873
#>  4: 2002  4 -0.8575960 -0.62265705 -0.5402118 -0.1109491 -0.37569101
#>  5: 2002  5 -0.5342633 -0.77032669 -0.3833859 -0.5655071 -0.07855638
#>  6: 2002  6 -0.8217070 -0.16712835 -0.1777715 -0.8364016 -0.54911940
#>  7: 2002  7 -0.9978921 -0.97781267 -0.4177440 -0.4204505 -0.82872015
#>  8: 2002  8 -0.3492861 -0.09274686 -0.3617363 -0.1111100 -0.04421507
#>  9: 2002  9 -0.8017680 -0.03824025 -0.5330707 -0.6860849 -0.11181242
#> 10: 2002 10 -0.8942224 -0.26756161 -0.2579290 -0.9131151 -0.91766187
#> 11: 2002 11 -0.3428663 -0.78466081 -0.5776727 -0.5335214 -0.89263370
#> 12: 2002 12 -0.2616384 -0.81483370 -0.4393205 -0.6498511 -0.96687562

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2 var_layer3 var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.29932824 -0.5948099 -0.7323255    1.064259    1.199897
#>  2: 2002  2 -0.18832581 -0.2859579 -0.5484089    1.765244    1.368143
#>  3: 2002  3 -0.49342411 -0.8125183 -0.4164660    1.315047    1.493588
#>  4: 2002  4 -0.62265705 -0.5402118 -0.1109491    1.271508    1.266522
#>  5: 2002  5 -0.77032669 -0.3833859 -0.5655071    1.042282    1.824477
#>  6: 2002  6 -0.16712835 -0.1777715 -0.8364016    1.823405    1.823799
#>  7: 2002  7 -0.97781267 -0.4177440 -0.4204505    1.470159    1.720856
#>  8: 2002  8 -0.09274686 -0.3617363 -0.1111100    1.684639    1.952371
#>  9: 2002  9 -0.03824025 -0.5330707 -0.6860849    1.534870    1.280043
#> 10: 2002 10 -0.26756161 -0.2579290 -0.9131151    1.043609    1.146872
#> 11: 2002 11 -0.78466081 -0.5776727 -0.5335214    1.432480    1.831964
#> 12: 2002 12 -0.81483370 -0.4393205 -0.6498511    1.234793    1.786910
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.296844    2.419908    2.583391    2.058306
#>  2:    1.478482    2.207059    2.937808    2.700760
#>  3:    1.216606    2.677962    2.915139    2.760962
#>  4:    1.063482    2.996806    2.095905    2.838038
#>  5:    1.756991    2.779549    2.898714    2.453384
#>  6:    1.611425    2.564931    2.378179    2.134767
#>  7:    1.598440    2.038909    2.941414    2.733585
#>  8:    1.897778    2.636976    2.458872    2.735987
#>  9:    1.285804    2.100394    2.504835    2.637030
#> 10:    1.010307    2.049500    2.478056    2.992778
#> 11:    1.542762    2.786852    2.711784    2.562305
#> 12:    1.668075    2.503200    2.967620    2.483570

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.064259    1.199897    1.296844    2.419908    2.583391
#>  2: 2002  2    1.765244    1.368143    1.478482    2.207059    2.937808
#>  3: 2002  3    1.315047    1.493588    1.216606    2.677962    2.915139
#>  4: 2002  4    1.271508    1.266522    1.063482    2.996806    2.095905
#>  5: 2002  5    1.042282    1.824477    1.756991    2.779549    2.898714
#>  6: 2002  6    1.823405    1.823799    1.611425    2.564931    2.378179
#>  7: 2002  7    1.470159    1.720856    1.598440    2.038909    2.941414
#>  8: 2002  8    1.684639    1.952371    1.897778    2.636976    2.458872
#>  9: 2002  9    1.534870    1.280043    1.285804    2.100394    2.504835
#> 10: 2002 10    1.043609    1.146872    1.010307    2.049500    2.478056
#> 11: 2002 11    1.432480    1.831964    1.542762    2.786852    2.711784
#> 12: 2002 12    1.234793    1.786910    1.668075    2.503200    2.967620
#>     var2_layer4
#>  1:    2.058306
#>  2:    2.700760
#>  3:    2.760962
#>  4:    2.838038
#>  5:    2.453384
#>  6:    2.134767
#>  7:    2.733585
#>  8:    2.735987
#>  9:    2.637030
#> 10:    2.992778
#> 11:    2.562305
#> 12:    2.483570