Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer data such as BELO and SWAT to the wide format, by casting variables with the layer number, using data.table's dcast-function.

extract_layer_output(dat, layers = NULL, value_vars = NULL, sep = "")

Arguments

dat

Data.frame with layer data organized in rows and identified by a layer number column nl.

layers

Integer vector addressing the layer numbers (nl) to be extracted from dat. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from dat. If not supplied, value_vars will be guessed.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns (wide format) with the names being made up of the variable name and layer number.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002:2003,
                  mo = 1:12)
df
#>     nl   yr mo
#> 1    1 2002  1
#> 2    2 2002  1
#> 3    3 2002  1
#> 4    4 2002  1
#> 5    5 2002  1
#> 6    1 2003  1
#> 7    2 2003  1
#> 8    3 2003  1
#> 9    4 2003  1
#> 10   5 2003  1
#> 11   1 2002  2
#> 12   2 2002  2
#> 13   3 2002  2
#> 14   4 2002  2
#> 15   5 2002  2
#> 16   1 2003  2
#> 17   2 2003  2
#> 18   3 2003  2
#> 19   4 2003  2
#> 20   5 2003  2
#> 21   1 2002  3
#> 22   2 2002  3
#> 23   3 2002  3
#> 24   4 2002  3
#> 25   5 2002  3
#> 26   1 2003  3
#> 27   2 2003  3
#> 28   3 2003  3
#> 29   4 2003  3
#> 30   5 2003  3
#> 31   1 2002  4
#> 32   2 2002  4
#> 33   3 2002  4
#> 34   4 2002  4
#> 35   5 2002  4
#> 36   1 2003  4
#> 37   2 2003  4
#> 38   3 2003  4
#> 39   4 2003  4
#> 40   5 2003  4
#> 41   1 2002  5
#> 42   2 2002  5
#> 43   3 2002  5
#> 44   4 2002  5
#> 45   5 2002  5
#> 46   1 2003  5
#> 47   2 2003  5
#> 48   3 2003  5
#> 49   4 2003  5
#> 50   5 2003  5
#> 51   1 2002  6
#> 52   2 2002  6
#> 53   3 2002  6
#> 54   4 2002  6
#> 55   5 2002  6
#> 56   1 2003  6
#> 57   2 2003  6
#> 58   3 2003  6
#> 59   4 2003  6
#> 60   5 2003  6
#> 61   1 2002  7
#> 62   2 2002  7
#> 63   3 2002  7
#> 64   4 2002  7
#> 65   5 2002  7
#> 66   1 2003  7
#> 67   2 2003  7
#> 68   3 2003  7
#> 69   4 2003  7
#> 70   5 2003  7
#> 71   1 2002  8
#> 72   2 2002  8
#> 73   3 2002  8
#> 74   4 2002  8
#> 75   5 2002  8
#> 76   1 2003  8
#> 77   2 2003  8
#> 78   3 2003  8
#> 79   4 2003  8
#> 80   5 2003  8
#> 81   1 2002  9
#> 82   2 2002  9
#> 83   3 2002  9
#> 84   4 2002  9
#> 85   5 2002  9
#> 86   1 2003  9
#> 87   2 2003  9
#> 88   3 2003  9
#> 89   4 2003  9
#> 90   5 2003  9
#> 91   1 2002 10
#> 92   2 2002 10
#> 93   3 2002 10
#> 94   4 2002 10
#> 95   5 2002 10
#> 96   1 2003 10
#> 97   2 2003 10
#> 98   3 2003 10
#> 99   4 2003 10
#> 100  5 2003 10
#> 101  1 2002 11
#> 102  2 2002 11
#> 103  3 2002 11
#> 104  4 2002 11
#> 105  5 2002 11
#> 106  1 2003 11
#> 107  2 2003 11
#> 108  3 2003 11
#> 109  4 2003 11
#> 110  5 2003 11
#> 111  1 2002 12
#> 112  2 2002 12
#> 113  3 2002 12
#> 114  4 2002 12
#> 115  5 2002 12
#> 116  1 2003 12
#> 117  2 2003 12
#> 118  3 2003 12
#> 119  4 2003 12
#> 120  5 2003 12
#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo         var1        var2         var3        var4        var5
#>  1: 2002  1 -0.913901179 -0.25373737 -0.751530027 -0.41462052 -0.49210719
#>  2: 2002  2 -0.720907563 -0.21667610 -0.269724905 -0.02952195 -0.22639328
#>  3: 2002  3 -0.512565416 -0.96468602 -0.155682014 -0.14861563 -0.64357107
#>  4: 2002  4 -0.005373652 -0.14022265 -0.228237594 -0.57616403 -0.29621195
#>  5: 2002  5 -0.363567965 -0.17585145 -0.705417418 -0.90550720 -0.72776164
#>  6: 2002  6 -0.302781584 -0.78673720 -0.550126317 -0.53078134 -0.39541476
#>  7: 2002  7 -0.653606292 -0.66464816 -0.111801561 -0.09446409 -0.69677896
#>  8: 2002  8 -0.446186471 -0.21343255 -0.772734583 -0.84033410 -0.46043708
#>  9: 2002  9 -0.042654745 -0.09477074 -0.527405295 -0.04210799 -0.60113837
#> 10: 2002 10 -0.796196142 -0.97585543 -0.347116683 -0.48127517 -0.39786852
#> 11: 2002 11 -0.012899099 -0.92810898 -0.010186991 -0.20374992 -0.01214628
#> 12: 2002 12 -0.703926966 -0.58542699 -0.972570712 -0.28748413 -0.56183228
#> 13: 2003  1 -0.795394553 -0.92537870 -0.550501984 -0.53561432 -0.92750386
#> 14: 2003  2 -0.273712795 -0.62927773 -0.033054444 -0.28465850 -0.55766399
#> 15: 2003  3 -0.296493914 -0.71208385 -0.808002063 -0.11108694 -0.80944435
#> 16: 2003  4 -0.987096082 -0.28943235 -0.879799565 -0.61721491 -0.45059885
#> 17: 2003  5 -0.138837438 -0.90524109 -0.115473514 -0.95783673 -0.98925679
#> 18: 2003  6 -0.028267038 -0.77186885 -0.480009071 -0.25681064 -0.41451923
#> 19: 2003  7 -0.685070493 -0.69904555 -0.788692027 -0.79982379 -0.48743016
#> 20: 2003  8 -0.806977815 -0.32225098 -0.826934459 -0.40040893 -0.92943897
#> 21: 2003  9 -0.361387828 -0.70233857 -0.475773742 -0.86658860 -0.91780603
#> 22: 2003 10 -0.263818174 -0.87944885 -0.005045217 -0.66390004 -0.43347132
#> 23: 2003 11 -0.509523719 -0.64567717 -0.346969666 -0.68313461 -0.10287057
#> 24: 2003 12 -0.899489063 -0.20923293 -0.103538865 -0.54753267 -0.39522965
#>       yr mo         var1        var2         var3        var4        var5

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_")
#>       yr mo       var_2        var_3       var_4   var1_2   var1_3   var1_4
#>  1: 2002  1 -0.25373737 -0.751530027 -0.41462052 1.350438 1.981349 1.936024
#>  2: 2002  2 -0.21667610 -0.269724905 -0.02952195 1.492739 1.350970 1.159983
#>  3: 2002  3 -0.96468602 -0.155682014 -0.14861563 1.298846 1.104672 1.167104
#>  4: 2002  4 -0.14022265 -0.228237594 -0.57616403 1.411519 1.046192 1.412485
#>  5: 2002  5 -0.17585145 -0.705417418 -0.90550720 1.388897 1.267265 1.897805
#>  6: 2002  6 -0.78673720 -0.550126317 -0.53078134 1.297350 1.444676 1.197344
#>  7: 2002  7 -0.66464816 -0.111801561 -0.09446409 1.923844 1.683948 1.411640
#>  8: 2002  8 -0.21343255 -0.772734583 -0.84033410 1.422102 1.593385 1.464799
#>  9: 2002  9 -0.09477074 -0.527405295 -0.04210799 1.754983 1.536324 1.429917
#> 10: 2002 10 -0.97585543 -0.347116683 -0.48127517 1.680927 1.448215 1.421773
#> 11: 2002 11 -0.92810898 -0.010186991 -0.20374992 1.083070 1.579589 1.619588
#> 12: 2002 12 -0.58542699 -0.972570712 -0.28748413 1.829618 1.034556 1.192925
#> 13: 2003  1 -0.92537870 -0.550501984 -0.53561432 1.436897 1.696474 1.930282
#> 14: 2003  2 -0.62927773 -0.033054444 -0.28465850 1.487851 1.080738 1.358886
#> 15: 2003  3 -0.71208385 -0.808002063 -0.11108694 1.322812 1.084785 1.821004
#> 16: 2003  4 -0.28943235 -0.879799565 -0.61721491 1.796348 1.597510 1.169244
#> 17: 2003  5 -0.90524109 -0.115473514 -0.95783673 1.570869 1.098589 1.092345
#> 18: 2003  6 -0.77186885 -0.480009071 -0.25681064 1.740975 1.302164 1.368859
#> 19: 2003  7 -0.69904555 -0.788692027 -0.79982379 1.025381 1.561611 1.674633
#> 20: 2003  8 -0.32225098 -0.826934459 -0.40040893 1.486325 1.094717 1.243843
#> 21: 2003  9 -0.70233857 -0.475773742 -0.86658860 1.020346 1.618679 1.104560
#> 22: 2003 10 -0.87944885 -0.005045217 -0.66390004 1.833020 1.257450 1.439095
#> 23: 2003 11 -0.64567717 -0.346969666 -0.68313461 1.117858 1.200279 1.524534
#> 24: 2003 12 -0.20923293 -0.103538865 -0.54753267 1.419126 1.663722 1.927500
#>       yr mo       var_2        var_3       var_4   var1_2   var1_3   var1_4
#>       var2_2   var2_3   var2_4
#>  1: 2.472342 2.730792 2.341587
#>  2: 2.160918 2.563985 2.570305
#>  3: 2.031401 2.105766 2.091140
#>  4: 2.892888 2.217825 2.966465
#>  5: 2.103678 2.979370 2.067494
#>  6: 2.719355 2.754350 2.708119
#>  7: 2.459561 2.729910 2.861077
#>  8: 2.717262 2.445831 2.779517
#>  9: 2.964570 2.279131 2.082704
#> 10: 2.114089 2.629657 2.246851
#> 11: 2.914536 2.055513 2.259677
#> 12: 2.368455 2.042550 2.266656
#> 13: 2.563363 2.506427 2.540248
#> 14: 2.392411 2.729039 2.554449
#> 15: 2.785615 2.897548 2.820803
#> 16: 2.435137 2.936072 2.462264
#> 17: 2.887732 2.658381 2.537434
#> 18: 2.467002 2.180027 2.559081
#> 19: 2.243486 2.636274 2.125802
#> 20: 2.557867 2.353654 2.612203
#> 21: 2.822640 2.285103 2.099415
#> 22: 2.849808 2.458766 2.638727
#> 23: 2.211180 2.485202 2.635466
#> 24: 2.670068 2.166714 2.256195
#>       var2_2   var2_3   var2_4

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_")
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4
#>  1: 2002  1 1.350438 1.981349 1.936024 2.472342 2.730792 2.341587
#>  2: 2002  2 1.492739 1.350970 1.159983 2.160918 2.563985 2.570305
#>  3: 2002  3 1.298846 1.104672 1.167104 2.031401 2.105766 2.091140
#>  4: 2002  4 1.411519 1.046192 1.412485 2.892888 2.217825 2.966465
#>  5: 2002  5 1.388897 1.267265 1.897805 2.103678 2.979370 2.067494
#>  6: 2002  6 1.297350 1.444676 1.197344 2.719355 2.754350 2.708119
#>  7: 2002  7 1.923844 1.683948 1.411640 2.459561 2.729910 2.861077
#>  8: 2002  8 1.422102 1.593385 1.464799 2.717262 2.445831 2.779517
#>  9: 2002  9 1.754983 1.536324 1.429917 2.964570 2.279131 2.082704
#> 10: 2002 10 1.680927 1.448215 1.421773 2.114089 2.629657 2.246851
#> 11: 2002 11 1.083070 1.579589 1.619588 2.914536 2.055513 2.259677
#> 12: 2002 12 1.829618 1.034556 1.192925 2.368455 2.042550 2.266656
#> 13: 2003  1 1.436897 1.696474 1.930282 2.563363 2.506427 2.540248
#> 14: 2003  2 1.487851 1.080738 1.358886 2.392411 2.729039 2.554449
#> 15: 2003  3 1.322812 1.084785 1.821004 2.785615 2.897548 2.820803
#> 16: 2003  4 1.796348 1.597510 1.169244 2.435137 2.936072 2.462264
#> 17: 2003  5 1.570869 1.098589 1.092345 2.887732 2.658381 2.537434
#> 18: 2003  6 1.740975 1.302164 1.368859 2.467002 2.180027 2.559081
#> 19: 2003  7 1.025381 1.561611 1.674633 2.243486 2.636274 2.125802
#> 20: 2003  8 1.486325 1.094717 1.243843 2.557867 2.353654 2.612203
#> 21: 2003  9 1.020346 1.618679 1.104560 2.822640 2.285103 2.099415
#> 22: 2003 10 1.833020 1.257450 1.439095 2.849808 2.458766 2.638727
#> 23: 2003 11 1.117858 1.200279 1.524534 2.211180 2.485202 2.635466
#> 24: 2003 12 1.419126 1.663722 1.927500 2.670068 2.166714 2.256195
#>       yr mo   var1_2   var1_3   var1_4   var2_2   var2_3   var2_4