Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo          var1        var2        var3       var4         var5
#>  1: 2002  1 -0.4135673677 -0.96009588 -0.32643218 -0.6120136 -0.118759838
#>  2: 2002  2 -0.8467281263 -0.01528153 -0.50394480 -0.0325648 -0.438527689
#>  3: 2002  3 -0.0329617211 -0.97413160 -0.34223736 -0.1145068 -0.266417497
#>  4: 2002  4 -0.2651976787 -0.04670706 -0.56709631 -0.2258767 -0.023513007
#>  5: 2002  5 -0.5209833193 -0.39888740 -0.28363384 -0.5335113 -0.006722308
#>  6: 2002  6 -0.8179895175 -0.07823792 -0.15481055 -0.7082780 -0.432383735
#>  7: 2002  7 -0.0275926013 -0.53246101 -0.93044851 -0.4122834 -0.827327112
#>  8: 2002  8 -0.2559759985 -0.65050086 -0.91084249 -0.2957744 -0.582448203
#>  9: 2002  9 -0.1745366636 -0.12195388 -0.38333313 -0.2563645 -0.385736363
#> 10: 2002 10 -0.0007970312 -0.46805278 -0.09763191 -0.2487100 -0.587454155
#> 11: 2002 11 -0.9179340149 -0.43274598 -0.51891095 -0.5593840 -0.663946804
#> 12: 2002 12 -0.0918606834 -0.81465764 -0.80535945 -0.4272562 -0.152243155

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2  var_layer3 var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.96009588 -0.32643218 -0.6120136    1.261716    1.417775
#>  2: 2002  2 -0.01528153 -0.50394480 -0.0325648    1.796636    1.578693
#>  3: 2002  3 -0.97413160 -0.34223736 -0.1145068    1.935628    1.462667
#>  4: 2002  4 -0.04670706 -0.56709631 -0.2258767    1.236622    1.252308
#>  5: 2002  5 -0.39888740 -0.28363384 -0.5335113    1.565784    1.601041
#>  6: 2002  6 -0.07823792 -0.15481055 -0.7082780    1.847716    1.691420
#>  7: 2002  7 -0.53246101 -0.93044851 -0.4122834    1.720255    1.335875
#>  8: 2002  8 -0.65050086 -0.91084249 -0.2957744    1.285377    1.225429
#>  9: 2002  9 -0.12195388 -0.38333313 -0.2563645    1.466998    1.193031
#> 10: 2002 10 -0.46805278 -0.09763191 -0.2487100    1.655224    1.051527
#> 11: 2002 11 -0.43274598 -0.51891095 -0.5593840    1.309867    1.497080
#> 12: 2002 12 -0.81465764 -0.80535945 -0.4272562    1.787468    1.439479
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.323011    2.162999    2.185878    2.342679
#>  2:    1.200922    2.547784    2.907242    2.900582
#>  3:    1.794252    2.083244    2.677404    2.672339
#>  4:    1.580896    2.220291    2.107146    2.000287
#>  5:    1.945250    2.106253    2.535105    2.891725
#>  6:    1.117167    2.298911    2.654931    2.483483
#>  7:    1.966116    2.680728    2.695124    2.955306
#>  8:    1.425930    2.433197    2.482897    2.223226
#>  9:    1.986560    2.666331    2.862298    2.599057
#> 10:    1.296916    2.254465    2.596028    2.395402
#> 11:    1.952892    2.718696    2.241792    2.634385
#> 12:    1.003479    2.998474    2.428933    2.493335

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.261716    1.417775    1.323011    2.162999    2.185878
#>  2: 2002  2    1.796636    1.578693    1.200922    2.547784    2.907242
#>  3: 2002  3    1.935628    1.462667    1.794252    2.083244    2.677404
#>  4: 2002  4    1.236622    1.252308    1.580896    2.220291    2.107146
#>  5: 2002  5    1.565784    1.601041    1.945250    2.106253    2.535105
#>  6: 2002  6    1.847716    1.691420    1.117167    2.298911    2.654931
#>  7: 2002  7    1.720255    1.335875    1.966116    2.680728    2.695124
#>  8: 2002  8    1.285377    1.225429    1.425930    2.433197    2.482897
#>  9: 2002  9    1.466998    1.193031    1.986560    2.666331    2.862298
#> 10: 2002 10    1.655224    1.051527    1.296916    2.254465    2.596028
#> 11: 2002 11    1.309867    1.497080    1.952892    2.718696    2.241792
#> 12: 2002 12    1.787468    1.439479    1.003479    2.998474    2.428933
#>     var2_layer4
#>  1:    2.342679
#>  2:    2.900582
#>  3:    2.672339
#>  4:    2.000287
#>  5:    2.891725
#>  6:    2.483483
#>  7:    2.955306
#>  8:    2.223226
#>  9:    2.599057
#> 10:    2.395402
#> 11:    2.634385
#> 12:    2.493335