Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo         var1        var2        var3        var4        var5
#>  1: 2002  1 -0.336818505 -0.36909811 -0.65971190 -0.06342649 -0.49168047
#>  2: 2002  2 -0.253301972 -0.72581741 -0.52702563 -0.36068689 -0.74949163
#>  3: 2002  3 -0.143766352 -0.53397767 -0.81028573 -0.34336572 -0.07157943
#>  4: 2002  4 -0.007010954 -0.87921467 -0.21174370 -0.92373418 -0.58941281
#>  5: 2002  5 -0.457950778 -0.28190663 -0.01266695 -0.42446964 -0.12335777
#>  6: 2002  6 -0.951783375 -0.04649649 -0.23699734 -0.48849788 -0.89613841
#>  7: 2002  7 -0.769450925 -0.70666046 -0.40515365 -0.95272400 -0.15974340
#>  8: 2002  8 -0.416125996 -0.74070654 -0.39998485 -0.47042154 -0.44497261
#>  9: 2002  9 -0.534482449 -0.13929612 -0.76540843 -0.17836992 -0.58259275
#> 10: 2002 10 -0.506970044 -0.44155483 -0.21888562 -0.02021812 -0.42604320
#> 11: 2002 11 -0.528658996 -0.51172570 -0.47686554 -0.44785125 -0.73299642
#> 12: 2002 12 -0.423409782 -0.48046515 -0.91771998 -0.90834142 -0.35718401

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2  var_layer3  var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.36909811 -0.65971190 -0.06342649    1.924569    1.941745
#>  2: 2002  2 -0.72581741 -0.52702563 -0.36068689    1.000232    1.222381
#>  3: 2002  3 -0.53397767 -0.81028573 -0.34336572    1.471333    1.988905
#>  4: 2002  4 -0.87921467 -0.21174370 -0.92373418    1.652847    1.044028
#>  5: 2002  5 -0.28190663 -0.01266695 -0.42446964    1.390528    1.632881
#>  6: 2002  6 -0.04649649 -0.23699734 -0.48849788    1.124452    1.717662
#>  7: 2002  7 -0.70666046 -0.40515365 -0.95272400    1.885854    1.161043
#>  8: 2002  8 -0.74070654 -0.39998485 -0.47042154    1.908507    1.116193
#>  9: 2002  9 -0.13929612 -0.76540843 -0.17836992    1.248646    1.808866
#> 10: 2002 10 -0.44155483 -0.21888562 -0.02021812    1.230987    1.923927
#> 11: 2002 11 -0.51172570 -0.47686554 -0.44785125    1.912855    1.893770
#> 12: 2002 12 -0.48046515 -0.91771998 -0.90834142    1.881524    1.167862
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.276061    2.516966    2.510373    2.321324
#>  2:    1.934037    2.504285    2.162928    2.859466
#>  3:    1.939020    2.482028    2.507910    2.717999
#>  4:    1.594671    2.656929    2.087110    2.333519
#>  5:    1.092258    2.867389    2.988447    2.956141
#>  6:    1.683083    2.471367    2.427731    2.355958
#>  7:    1.138505    2.032666    2.326190    2.229937
#>  8:    1.973749    2.551005    2.060043    2.778866
#>  9:    1.047277    2.555287    2.259749    2.689684
#> 10:    1.896727    2.495733    2.298930    2.032461
#> 11:    1.375554    2.280392    2.840037    2.584206
#> 12:    1.078329    2.022133    2.951638    2.482264

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.924569    1.941745    1.276061    2.516966    2.510373
#>  2: 2002  2    1.000232    1.222381    1.934037    2.504285    2.162928
#>  3: 2002  3    1.471333    1.988905    1.939020    2.482028    2.507910
#>  4: 2002  4    1.652847    1.044028    1.594671    2.656929    2.087110
#>  5: 2002  5    1.390528    1.632881    1.092258    2.867389    2.988447
#>  6: 2002  6    1.124452    1.717662    1.683083    2.471367    2.427731
#>  7: 2002  7    1.885854    1.161043    1.138505    2.032666    2.326190
#>  8: 2002  8    1.908507    1.116193    1.973749    2.551005    2.060043
#>  9: 2002  9    1.248646    1.808866    1.047277    2.555287    2.259749
#> 10: 2002 10    1.230987    1.923927    1.896727    2.495733    2.298930
#> 11: 2002 11    1.912855    1.893770    1.375554    2.280392    2.840037
#> 12: 2002 12    1.881524    1.167862    1.078329    2.022133    2.951638
#>     var2_layer4
#>  1:    2.321324
#>  2:    2.859466
#>  3:    2.717999
#>  4:    2.333519
#>  5:    2.956141
#>  6:    2.355958
#>  7:    2.229937
#>  8:    2.778866
#>  9:    2.689684
#> 10:    2.032461
#> 11:    2.584206
#> 12:    2.482264