Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo         var1         var2        var3        var4        var5
#>  1: 2002  1 -0.173125647 -0.367529958 -0.34685067 -0.05545805 -0.01172361
#>  2: 2002  2 -0.907189094 -0.155464616 -0.49986306 -0.27134766 -0.01833097
#>  3: 2002  3 -0.271007744 -0.476819921 -0.07858414 -0.19711345 -0.74647189
#>  4: 2002  4 -0.589328040 -0.766517996 -0.56668491 -0.39768050 -0.35056049
#>  5: 2002  5 -0.350839383 -0.594871407 -0.80873775 -0.92400051 -0.39494317
#>  6: 2002  6 -0.577889988 -0.033565959 -0.72989266 -0.05152700 -0.30864477
#>  7: 2002  7 -0.288797129 -0.007830178 -0.50481028 -0.10543815 -0.99333833
#>  8: 2002  8 -0.495665051 -0.132381617 -0.96127366 -0.75786438 -0.79795597
#>  9: 2002  9 -0.109841787 -0.247694770 -0.53166084 -0.09254927 -0.60575811
#> 10: 2002 10 -0.002762827 -0.455582800 -0.25881550 -0.98494851 -0.58480284
#> 11: 2002 11 -0.663811780 -0.518289055 -0.98145338 -0.82182841 -0.80218083
#> 12: 2002 12 -0.225958351 -0.914514696 -0.93240260 -0.69307689 -0.09264643

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo   var_layer2  var_layer3  var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.367529958 -0.34685067 -0.05545805    1.676948    1.653497
#>  2: 2002  2 -0.155464616 -0.49986306 -0.27134766    1.918878    1.871176
#>  3: 2002  3 -0.476819921 -0.07858414 -0.19711345    1.922192    1.664189
#>  4: 2002  4 -0.766517996 -0.56668491 -0.39768050    1.661016    1.313470
#>  5: 2002  5 -0.594871407 -0.80873775 -0.92400051    1.226099    1.992517
#>  6: 2002  6 -0.033565959 -0.72989266 -0.05152700    1.577804    1.008462
#>  7: 2002  7 -0.007830178 -0.50481028 -0.10543815    1.849526    1.207103
#>  8: 2002  8 -0.132381617 -0.96127366 -0.75786438    1.349621    1.378517
#>  9: 2002  9 -0.247694770 -0.53166084 -0.09254927    1.310588    1.259505
#> 10: 2002 10 -0.455582800 -0.25881550 -0.98494851    1.185289    1.088173
#> 11: 2002 11 -0.518289055 -0.98145338 -0.82182841    1.411778    1.365645
#> 12: 2002 12 -0.914514696 -0.93240260 -0.69307689    1.420198    1.413176
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.618897    2.218759    2.522140    2.646704
#>  2:    1.002782    2.406694    2.233398    2.426142
#>  3:    1.642728    2.197002    2.299611    2.870008
#>  4:    1.739194    2.672769    2.901854    2.968811
#>  5:    1.321737    2.969755    2.190086    2.683987
#>  6:    1.233278    2.149313    2.173318    2.949461
#>  7:    1.705385    2.110700    2.221018    2.692276
#>  8:    1.348512    2.928863    2.413399    2.161807
#>  9:    1.360928    2.256303    2.094501    2.398246
#> 10:    1.437666    2.566423    2.961413    2.434551
#> 11:    1.748611    2.770414    2.249259    2.748615
#> 12:    1.143528    2.520291    2.508789    2.950405

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.676948    1.653497    1.618897    2.218759    2.522140
#>  2: 2002  2    1.918878    1.871176    1.002782    2.406694    2.233398
#>  3: 2002  3    1.922192    1.664189    1.642728    2.197002    2.299611
#>  4: 2002  4    1.661016    1.313470    1.739194    2.672769    2.901854
#>  5: 2002  5    1.226099    1.992517    1.321737    2.969755    2.190086
#>  6: 2002  6    1.577804    1.008462    1.233278    2.149313    2.173318
#>  7: 2002  7    1.849526    1.207103    1.705385    2.110700    2.221018
#>  8: 2002  8    1.349621    1.378517    1.348512    2.928863    2.413399
#>  9: 2002  9    1.310588    1.259505    1.360928    2.256303    2.094501
#> 10: 2002 10    1.185289    1.088173    1.437666    2.566423    2.961413
#> 11: 2002 11    1.411778    1.365645    1.748611    2.770414    2.249259
#> 12: 2002 12    1.420198    1.413176    1.143528    2.520291    2.508789
#>     var2_layer4
#>  1:    2.646704
#>  2:    2.426142
#>  3:    2.870008
#>  4:    2.968811
#>  5:    2.683987
#>  6:    2.949461
#>  7:    2.692276
#>  8:    2.161807
#>  9:    2.398246
#> 10:    2.434551
#> 11:    2.748615
#> 12:    2.950405