Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo        var1       var2        var3        var4        var5
#>  1: 2002  1 -0.85709129 -0.6516407 -0.08968456 -0.63431156 -0.32715751
#>  2: 2002  2 -0.98141967 -0.2106557 -0.47649885 -0.97905250 -0.33223924
#>  3: 2002  3 -0.51946207 -0.1600381 -0.85521323 -0.75326276 -0.41538435
#>  4: 2002  4 -0.16927767 -0.8692244 -0.45400801 -0.25229769 -0.18442971
#>  5: 2002  5 -0.47193124 -0.5265122 -0.32100454 -0.76534523 -0.65740855
#>  6: 2002  6 -0.04444347 -0.3011462 -0.51997630 -0.40906165 -0.32713937
#>  7: 2002  7 -0.87690578 -0.5854233 -0.36691370 -0.08914108 -0.69385181
#>  8: 2002  8 -0.95349823 -0.5766676 -0.47639108 -0.84106565 -0.01312328
#>  9: 2002  9 -0.03381374 -0.6342258 -0.59812534 -0.28625827 -0.84107402
#> 10: 2002 10 -0.59028667 -0.2827605 -0.99566729 -0.40442988 -0.30287734
#> 11: 2002 11 -0.20249016 -0.1862441 -0.71417252 -0.06189185 -0.84569602
#> 12: 2002 12 -0.84909016 -0.6775067 -0.62836582 -0.32973494 -0.55687978

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo var_layer2  var_layer3  var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.6516407 -0.08968456 -0.63431156    1.647819    1.076045
#>  2: 2002  2 -0.2106557 -0.47649885 -0.97905250    1.098335    1.836307
#>  3: 2002  3 -0.1600381 -0.85521323 -0.75326276    1.140967    1.998121
#>  4: 2002  4 -0.8692244 -0.45400801 -0.25229769    1.717052    1.042723
#>  5: 2002  5 -0.5265122 -0.32100454 -0.76534523    1.091328    1.263413
#>  6: 2002  6 -0.3011462 -0.51997630 -0.40906165    1.037309    1.180383
#>  7: 2002  7 -0.5854233 -0.36691370 -0.08914108    1.961755    1.728066
#>  8: 2002  8 -0.5766676 -0.47639108 -0.84106565    1.229039    1.369605
#>  9: 2002  9 -0.6342258 -0.59812534 -0.28625827    1.002996    1.390109
#> 10: 2002 10 -0.2827605 -0.99566729 -0.40442988    1.610525    1.252294
#> 11: 2002 11 -0.1862441 -0.71417252 -0.06189185    1.792650    1.109289
#> 12: 2002 12 -0.6775067 -0.62836582 -0.32973494    1.394429    1.093603
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.111727    2.794325    2.591444    2.478506
#>  2:    1.729801    2.678857    2.368519    2.032363
#>  3:    1.477712    2.359637    2.321427    2.903578
#>  4:    1.265662    2.920660    2.658500    2.311350
#>  5:    1.052161    2.865732    2.622299    2.708951
#>  6:    1.470971    2.093817    2.788056    2.846333
#>  7:    1.212881    2.162095    2.036023    2.095447
#>  8:    1.159989    2.385761    2.285223    2.625660
#>  9:    1.076330    2.383877    2.247564    2.114802
#> 10:    1.980936    2.174806    2.163364    2.104594
#> 11:    1.330637    2.177044    2.606335    2.704508
#> 12:    1.998513    2.865594    2.774274    2.972557

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.647819    1.076045    1.111727    2.794325    2.591444
#>  2: 2002  2    1.098335    1.836307    1.729801    2.678857    2.368519
#>  3: 2002  3    1.140967    1.998121    1.477712    2.359637    2.321427
#>  4: 2002  4    1.717052    1.042723    1.265662    2.920660    2.658500
#>  5: 2002  5    1.091328    1.263413    1.052161    2.865732    2.622299
#>  6: 2002  6    1.037309    1.180383    1.470971    2.093817    2.788056
#>  7: 2002  7    1.961755    1.728066    1.212881    2.162095    2.036023
#>  8: 2002  8    1.229039    1.369605    1.159989    2.385761    2.285223
#>  9: 2002  9    1.002996    1.390109    1.076330    2.383877    2.247564
#> 10: 2002 10    1.610525    1.252294    1.980936    2.174806    2.163364
#> 11: 2002 11    1.792650    1.109289    1.330637    2.177044    2.606335
#> 12: 2002 12    1.394429    1.093603    1.998513    2.865594    2.774274
#>     var2_layer4
#>  1:    2.478506
#>  2:    2.032363
#>  3:    2.903578
#>  4:    2.311350
#>  5:    2.708951
#>  6:    2.846333
#>  7:    2.095447
#>  8:    2.625660
#>  9:    2.114802
#> 10:    2.104594
#> 11:    2.704508
#> 12:    2.972557