Extracts values from layer data and organizes layer-wise variables in columns

Convenience function to reorganize soil layer time series data from layer_output list entry produced with run_LWFB90. The data is tansformed to a wide format, by casting the variables with the layer number using data.table's dcast-function.

extract_layer_output(
  x,
  layers = NULL,
  value_vars = NULL,
  layer_index_name = "nl",
  sep = ""
)

Arguments

x

Data.frame or data.table with layer data organized in rows and identified by a layer index column named layer_index_nm.

layers

Integer vector to select a subset of layers. If not supplied, values from all layers will be returned.

value_vars

Character vector containing names of value-variables to be extracted from x. If not supplied, value_vars will be guessed.

layer_index_name

Column containing layer index. Defaults to 'nl' as in layer_output.

sep

Separation character for constructig names from variable name and layer index.

Value

A data.table with the layers' values of the variables organized in columns with the names being made up of the variable name and the layer index.

Examples

# create a data.frame with monthly values
# identifiers: layer number, yr and mo
df <- expand.grid(nl = 1:5,
                  yr = 2002,
                  mo = 1:12)
df
#>    nl   yr mo
#> 1   1 2002  1
#> 2   2 2002  1
#> 3   3 2002  1
#> 4   4 2002  1
#> 5   5 2002  1
#> 6   1 2002  2
#> 7   2 2002  2
#> 8   3 2002  2
#> 9   4 2002  2
#> 10  5 2002  2
#> 11  1 2002  3
#> 12  2 2002  3
#> 13  3 2002  3
#> 14  4 2002  3
#> 15  5 2002  3
#> 16  1 2002  4
#> 17  2 2002  4
#> 18  3 2002  4
#> 19  4 2002  4
#> 20  5 2002  4
#> 21  1 2002  5
#> 22  2 2002  5
#> 23  3 2002  5
#> 24  4 2002  5
#> 25  5 2002  5
#> 26  1 2002  6
#> 27  2 2002  6
#> 28  3 2002  6
#> 29  4 2002  6
#> 30  5 2002  6
#> 31  1 2002  7
#> 32  2 2002  7
#> 33  3 2002  7
#> 34  4 2002  7
#> 35  5 2002  7
#> 36  1 2002  8
#> 37  2 2002  8
#> 38  3 2002  8
#> 39  4 2002  8
#> 40  5 2002  8
#> 41  1 2002  9
#> 42  2 2002  9
#> 43  3 2002  9
#> 44  4 2002  9
#> 45  5 2002  9
#> 46  1 2002 10
#> 47  2 2002 10
#> 48  3 2002 10
#> 49  4 2002 10
#> 50  5 2002 10
#> 51  1 2002 11
#> 52  2 2002 11
#> 53  3 2002 11
#> 54  4 2002 11
#> 55  5 2002 11
#> 56  1 2002 12
#> 57  2 2002 12
#> 58  3 2002 12
#> 59  4 2002 12
#> 60  5 2002 12

#add a value variable
df$var <- runif(nrow(df), -1,0)

extract_layer_output(df)
#>       yr mo          var1        var2        var3        var4       var5
#>  1: 2002  1 -0.6604087350 -0.49599468 -0.38924788 -0.24016617 -0.9943165
#>  2: 2002  2 -0.8157186965 -0.11900442 -0.86373865 -0.04891902 -0.3404958
#>  3: 2002  3 -0.1440986381 -0.46896233 -0.74097196 -0.13928286 -0.6522067
#>  4: 2002  4 -0.7516575304 -0.12386228 -0.63821652 -0.24975252 -0.6327502
#>  5: 2002  5 -0.5687884816 -0.36484161 -0.85685200 -0.02806434 -0.8338161
#>  6: 2002  6 -0.3887608903 -0.88754760 -0.78725627 -0.91715769 -0.5309347
#>  7: 2002  7 -0.2263173803 -0.92962676 -0.63281406 -0.10599313 -0.3874370
#>  8: 2002  8 -0.6940298409 -0.79865368 -0.02825094 -0.01461289 -0.7664521
#>  9: 2002  9 -0.0006663974 -0.83874300 -0.57202021 -0.01268145 -0.5771520
#> 10: 2002 10 -0.2108226879 -0.50112467 -0.88471162 -0.66060718 -0.7526447
#> 11: 2002 11 -0.0840416148 -0.45591559 -0.71723979 -0.89118256 -0.7201205
#> 12: 2002 12 -0.1872406271 -0.01004734 -0.61241404 -0.61275118 -0.1688927

# add more variables
df$var1 <- runif(nrow(df), 1,2)
df$var2 <- runif(nrow(df), 2,3)

# extract specific layers
extract_layer_output(df,layers = 2:4, sep = "_layer")
#>       yr mo  var_layer2  var_layer3  var_layer4 var1_layer2 var1_layer3
#>  1: 2002  1 -0.49599468 -0.38924788 -0.24016617    1.034072    1.366088
#>  2: 2002  2 -0.11900442 -0.86373865 -0.04891902    1.877693    1.543911
#>  3: 2002  3 -0.46896233 -0.74097196 -0.13928286    1.787305    1.157732
#>  4: 2002  4 -0.12386228 -0.63821652 -0.24975252    1.728979    1.214924
#>  5: 2002  5 -0.36484161 -0.85685200 -0.02806434    1.708412    1.647308
#>  6: 2002  6 -0.88754760 -0.78725627 -0.91715769    1.369074    1.907781
#>  7: 2002  7 -0.92962676 -0.63281406 -0.10599313    1.985935    1.649099
#>  8: 2002  8 -0.79865368 -0.02825094 -0.01461289    1.685857    1.803241
#>  9: 2002  9 -0.83874300 -0.57202021 -0.01268145    1.753132    1.467901
#> 10: 2002 10 -0.50112467 -0.88471162 -0.66060718    1.538525    1.628515
#> 11: 2002 11 -0.45591559 -0.71723979 -0.89118256    1.697251    1.297335
#> 12: 2002 12 -0.01004734 -0.61241404 -0.61275118    1.294390    1.083163
#>     var1_layer4 var2_layer2 var2_layer3 var2_layer4
#>  1:    1.135266    2.690744    2.745203    2.661621
#>  2:    1.075998    2.505973    2.500467    2.335621
#>  3:    1.913960    2.712678    2.397027    2.155924
#>  4:    1.332644    2.285411    2.651411    2.132374
#>  5:    1.573728    2.141527    2.844493    2.682987
#>  6:    1.384039    2.674670    2.880955    2.544448
#>  7:    1.717799    2.319094    2.331887    2.933384
#>  8:    1.173337    2.196105    2.329777    2.146152
#>  9:    1.162189    2.581448    2.326244    2.484829
#> 10:    1.283339    2.557142    2.293662    2.672061
#> 11:    1.284306    2.256852    2.264511    2.663328
#> 12:    1.743540    2.672641    2.722269    2.039990

#extract specific variables
extract_layer_output(df, layers = 2:4, value_vars = c("var1", "var2"), sep = "_layer")
#>       yr mo var1_layer2 var1_layer3 var1_layer4 var2_layer2 var2_layer3
#>  1: 2002  1    1.034072    1.366088    1.135266    2.690744    2.745203
#>  2: 2002  2    1.877693    1.543911    1.075998    2.505973    2.500467
#>  3: 2002  3    1.787305    1.157732    1.913960    2.712678    2.397027
#>  4: 2002  4    1.728979    1.214924    1.332644    2.285411    2.651411
#>  5: 2002  5    1.708412    1.647308    1.573728    2.141527    2.844493
#>  6: 2002  6    1.369074    1.907781    1.384039    2.674670    2.880955
#>  7: 2002  7    1.985935    1.649099    1.717799    2.319094    2.331887
#>  8: 2002  8    1.685857    1.803241    1.173337    2.196105    2.329777
#>  9: 2002  9    1.753132    1.467901    1.162189    2.581448    2.326244
#> 10: 2002 10    1.538525    1.628515    1.283339    2.557142    2.293662
#> 11: 2002 11    1.697251    1.297335    1.284306    2.256852    2.264511
#> 12: 2002 12    1.294390    1.083163    1.743540    2.672641    2.722269
#>     var2_layer4
#>  1:    2.661621
#>  2:    2.335621
#>  3:    2.155924
#>  4:    2.132374
#>  5:    2.682987
#>  6:    2.544448
#>  7:    2.933384
#>  8:    2.146152
#>  9:    2.484829
#> 10:    2.672061
#> 11:    2.663328
#> 12:    2.039990